Enum qt_widgets::abstract_item_delegate::EndEditHint

pub enum EndEditHint {
    NoHint,
    EditNextItem,
    EditPreviousItem,
    SubmitModelCache,
    RevertModelCache,
}

C++ type: QAbstractItemDelegate::EndEditHint

C++ documentation:

This enum describes the different hints that the delegate can give to the model and view components to make editing data in a model a comfortable experience for the user.

These hints let the delegate influence the behavior of the view:

Note that custom views may interpret the concepts of next and previous differently.

The following hints are most useful when models are used that cache data, such as those that manipulate data locally in order to increase performance or conserve network bandwidth.

Although models and views should respond to these hints in appropriate ways, custom components may ignore any or all of them if they are not relevant.

Variants
NoHint

There is no recommended action to be performed. (C++ enum variant: NoHint = 0)

EditNextItem

The view should use the delegate to open an editor on the next item in the view. (C++ enum variant: EditNextItem = 1)

EditPreviousItem

The view should use the delegate to open an editor on the previous item in the view. (C++ enum variant: EditPreviousItem = 2)

SubmitModelCache

If the model caches data, it should write out cached data to the underlying data store. (C++ enum variant: SubmitModelCache = 3)

RevertModelCache

If the model caches data, it should discard cached data and replace it with data from the underlying data store. (C++ enum variant: RevertModelCache = 4)