Module qt_widgets::abstract_item_delegate::overloading

Traits

AbstractItemDelegateCloseEditorArgs
This trait represents a set of arguments accepted by AbstractItemDelegate::close_editor method.

AbstractItemDelegateConnectArgs
This trait represents a set of arguments accepted by AbstractItemDelegate::connect method.

AbstractItemDelegateConnectStaticArgs
This trait represents a set of arguments accepted by AbstractItemDelegate::connect_static method.

AbstractItemDelegateDestroyedArgs
This trait represents a set of arguments accepted by AbstractItemDelegate::destroyed method.

AbstractItemDelegateDisconnectArgs
This trait represents a set of arguments accepted by AbstractItemDelegate::disconnect method.

AbstractItemDelegateDisconnectStaticArgs
This trait represents a set of arguments accepted by AbstractItemDelegate::disconnect_static method.

AbstractItemDelegateDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by AbstractItemDelegate::dynamic_property_names method.

AbstractItemDelegateElidedTextArgs
This trait represents a set of arguments accepted by AbstractItemDelegate::elided_text method.

AbstractItemDelegateObjectNameArgs
This trait represents a set of arguments accepted by AbstractItemDelegate::object_name method.

AbstractItemDelegatePaintingRolesArgs
This trait represents a set of arguments accepted by AbstractItemDelegate::painting_roles method.

AbstractItemDelegatePropertyArgs
This trait represents a set of arguments accepted by AbstractItemDelegate::property method.

AbstractItemDelegateSizeHintArgs
This trait represents a set of arguments accepted by AbstractItemDelegate::size_hint method.

AbstractItemDelegateStartTimerArgs
This trait represents a set of arguments accepted by AbstractItemDelegate::start_timer method.

AbstractItemDelegateTrArgs
This trait represents a set of arguments accepted by AbstractItemDelegate::tr method.

AbstractItemDelegateTrUtf8Args
This trait represents a set of arguments accepted by AbstractItemDelegate::tr_utf8 method.