Enum qt_widgets::abstract_item_view::SelectionMode

pub enum SelectionMode {
    No,
    Single,
    Multi,
    Extended,
    Contiguous,
}

C++ type: QAbstractItemView::SelectionMode

C++ documentation:

This enum indicates how the view responds to user selections:

The most commonly used modes are SingleSelection and ExtendedSelection.

Variants
No

Items cannot be selected. (C++ enum variant: NoSelection = 0)

Single

When the user selects an item, any already-selected item becomes unselected, and the user cannot unselect the selected item by clicking on it. (C++ enum variant: SingleSelection = 1)

Multi

When the user selects an item in the usual way, the selection status of that item is toggled and the other items are left alone. Multiple items can be toggled by dragging the mouse over them. (C++ enum variant: MultiSelection = 2)

Extended

When the user selects an item in the usual way, the selection is cleared and the new item selected. However, if the user presses the Ctrl key when clicking on an item, the clicked item gets toggled and all other items are left untouched. If the user presses the Shift key while clicking on an item, all items between the current item and the clicked item are selected or unselected, depending on the state of the clicked item. Multiple items can be selected by dragging the mouse over them. (C++ enum variant: ExtendedSelection = 3)

Contiguous

When the user selects an item in the usual way, the selection is cleared and the new item selected. However, if the user presses the Shift key while clicking on an item, all items between the current item and the clicked item are selected or unselected, depending on the state of the clicked item. (C++ enum variant: ContiguousSelection = 4)