Enum qt_widgets::abstract_item_view::State

pub enum State {
    No,
    Dragging,
    DragSelecting,
    Editing,
    Expanding,
    Collapsing,
    Animating,
}

C++ type: QAbstractItemView::State

C++ documentation:

Describes the different states the view can be in. This is usually only interesting when reimplementing your own view.

Variants
No

The is the default state. (C++ enum variant: NoState = 0)

Dragging

The user is dragging items. (C++ enum variant: DraggingState = 1)

DragSelecting

The user is selecting items. (C++ enum variant: DragSelectingState = 2)

Editing

The user is editing an item in a widget editor. (C++ enum variant: EditingState = 3)

Expanding

The user is opening a branch of items. (C++ enum variant: ExpandingState = 4)

Collapsing

The user is closing a branch of items. (C++ enum variant: CollapsingState = 5)

Animating

The item view is performing an animation. (C++ enum variant: AnimatingState = 6)