Module qt_widgets::accessible_abstract_scroll_area::overloading

Traits

AccessibleAbstractScrollAreaActionNamesArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::action_names method.

AccessibleAbstractScrollAreaBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::background_color method.

AccessibleAbstractScrollAreaForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::foreground_color method.

AccessibleAbstractScrollAreaKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::key_bindings_for_action method.

AccessibleAbstractScrollAreaLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::localized_action_description method.

AccessibleAbstractScrollAreaLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::localized_action_name method.

AccessibleAbstractScrollAreaNewArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::new method.

AccessibleAbstractScrollAreaNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::next_page_action method.

AccessibleAbstractScrollAreaPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::previous_page_action method.

AccessibleAbstractScrollAreaRectArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::rect method.

AccessibleAbstractScrollAreaRelationsArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::relations method.

AccessibleAbstractScrollAreaScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::scroll_down_action method.

AccessibleAbstractScrollAreaScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::scroll_left_action method.

AccessibleAbstractScrollAreaScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::scroll_right_action method.

AccessibleAbstractScrollAreaScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::scroll_up_action method.

AccessibleAbstractScrollAreaStateArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::state method.

AccessibleAbstractScrollAreaTextArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::text method.

AccessibleAbstractScrollAreaTrArgs
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::tr method.

AccessibleAbstractScrollAreaTrUtf8Args
This trait represents a set of arguments accepted by AccessibleAbstractScrollArea::tr_utf8 method.