Module qt_widgets::accessible_abstract_slider::overloading

Traits

AccessibleAbstractSliderActionNamesArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::action_names method.

AccessibleAbstractSliderBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::background_color method.

AccessibleAbstractSliderCurrentValueArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::current_value method.

AccessibleAbstractSliderForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::foreground_color method.

AccessibleAbstractSliderKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::key_bindings_for_action method.

AccessibleAbstractSliderLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::localized_action_description method.

AccessibleAbstractSliderLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::localized_action_name method.

AccessibleAbstractSliderMaximumValueArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::maximum_value method.

AccessibleAbstractSliderMinimumStepSizeArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::minimum_step_size method.

AccessibleAbstractSliderMinimumValueArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::minimum_value method.

AccessibleAbstractSliderNewArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::new method.

AccessibleAbstractSliderNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::next_page_action method.

AccessibleAbstractSliderPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::previous_page_action method.

AccessibleAbstractSliderRectArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::rect method.

AccessibleAbstractSliderRelationsArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::relations method.

AccessibleAbstractSliderScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::scroll_down_action method.

AccessibleAbstractSliderScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::scroll_left_action method.

AccessibleAbstractSliderScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::scroll_right_action method.

AccessibleAbstractSliderScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::scroll_up_action method.

AccessibleAbstractSliderStateArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::state method.

AccessibleAbstractSliderTextArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::text method.

AccessibleAbstractSliderTrArgs
This trait represents a set of arguments accepted by AccessibleAbstractSlider::tr method.

AccessibleAbstractSliderTrUtf8Args
This trait represents a set of arguments accepted by AccessibleAbstractSlider::tr_utf8 method.