Module qt_widgets::accessible_abstract_spin_box::overloading

Traits

AccessibleAbstractSpinBoxActionNamesArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::action_names method.

AccessibleAbstractSpinBoxAttributesArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::attributes method.

AccessibleAbstractSpinBoxBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::background_color method.

AccessibleAbstractSpinBoxCharacterRectArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::character_rect method.

AccessibleAbstractSpinBoxCurrentValueArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::current_value method.

AccessibleAbstractSpinBoxForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::foreground_color method.

AccessibleAbstractSpinBoxKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::key_bindings_for_action method.

AccessibleAbstractSpinBoxLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::localized_action_description method.

AccessibleAbstractSpinBoxLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::localized_action_name method.

AccessibleAbstractSpinBoxMaximumValueArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::maximum_value method.

AccessibleAbstractSpinBoxMinimumStepSizeArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::minimum_step_size method.

AccessibleAbstractSpinBoxMinimumValueArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::minimum_value method.

AccessibleAbstractSpinBoxNewArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::new method.

AccessibleAbstractSpinBoxNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::next_page_action method.

AccessibleAbstractSpinBoxPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::previous_page_action method.

AccessibleAbstractSpinBoxRectArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::rect method.

AccessibleAbstractSpinBoxRelationsArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::relations method.

AccessibleAbstractSpinBoxScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::scroll_down_action method.

AccessibleAbstractSpinBoxScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::scroll_left_action method.

AccessibleAbstractSpinBoxScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::scroll_right_action method.

AccessibleAbstractSpinBoxScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::scroll_up_action method.

AccessibleAbstractSpinBoxStateArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::state method.

AccessibleAbstractSpinBoxTextAfterOffsetArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::text_after_offset method.

AccessibleAbstractSpinBoxTextArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::text method.

AccessibleAbstractSpinBoxTextAtOffsetArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::text_at_offset method.

AccessibleAbstractSpinBoxTextBeforeOffsetArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::text_before_offset method.

AccessibleAbstractSpinBoxTrArgs
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::tr method.

AccessibleAbstractSpinBoxTrUtf8Args
This trait represents a set of arguments accepted by AccessibleAbstractSpinBox::tr_utf8 method.