Module qt_widgets::accessible_button::overloading

Traits

AccessibleButtonActionNamesArgs
This trait represents a set of arguments accepted by AccessibleButton::action_names method.

AccessibleButtonBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleButton::background_color method.

AccessibleButtonForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleButton::foreground_color method.

AccessibleButtonKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleButton::key_bindings_for_action method.

AccessibleButtonLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleButton::localized_action_description method.

AccessibleButtonLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleButton::localized_action_name method.

AccessibleButtonNewArgs
This trait represents a set of arguments accepted by AccessibleButton::new method.

AccessibleButtonNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleButton::next_page_action method.

AccessibleButtonPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleButton::previous_page_action method.

AccessibleButtonRectArgs
This trait represents a set of arguments accepted by AccessibleButton::rect method.

AccessibleButtonRelationsArgs
This trait represents a set of arguments accepted by AccessibleButton::relations method.

AccessibleButtonScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleButton::scroll_down_action method.

AccessibleButtonScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleButton::scroll_left_action method.

AccessibleButtonScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleButton::scroll_right_action method.

AccessibleButtonScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleButton::scroll_up_action method.

AccessibleButtonStateArgs
This trait represents a set of arguments accepted by AccessibleButton::state method.

AccessibleButtonTextArgs
This trait represents a set of arguments accepted by AccessibleButton::text method.

AccessibleButtonTrArgs
This trait represents a set of arguments accepted by AccessibleButton::tr method.

AccessibleButtonTrUtf8Args
This trait represents a set of arguments accepted by AccessibleButton::tr_utf8 method.