Module qt_widgets::accessible_calendar_widget::overloading

Traits

AccessibleCalendarWidgetActionNamesArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::action_names method.

AccessibleCalendarWidgetBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::background_color method.

AccessibleCalendarWidgetForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::foreground_color method.

AccessibleCalendarWidgetKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::key_bindings_for_action method.

AccessibleCalendarWidgetLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::localized_action_description method.

AccessibleCalendarWidgetLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::localized_action_name method.

AccessibleCalendarWidgetNewArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::new method.

AccessibleCalendarWidgetNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::next_page_action method.

AccessibleCalendarWidgetPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::previous_page_action method.

AccessibleCalendarWidgetRectArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::rect method.

AccessibleCalendarWidgetRelationsArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::relations method.

AccessibleCalendarWidgetScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::scroll_down_action method.

AccessibleCalendarWidgetScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::scroll_left_action method.

AccessibleCalendarWidgetScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::scroll_right_action method.

AccessibleCalendarWidgetScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::scroll_up_action method.

AccessibleCalendarWidgetStateArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::state method.

AccessibleCalendarWidgetTextArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::text method.

AccessibleCalendarWidgetTrArgs
This trait represents a set of arguments accepted by AccessibleCalendarWidget::tr method.

AccessibleCalendarWidgetTrUtf8Args
This trait represents a set of arguments accepted by AccessibleCalendarWidget::tr_utf8 method.