Module qt_widgets::accessible_combo_box::overloading

Traits

AccessibleComboBoxActionNamesArgs
This trait represents a set of arguments accepted by AccessibleComboBox::action_names method.

AccessibleComboBoxBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleComboBox::background_color method.

AccessibleComboBoxForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleComboBox::foreground_color method.

AccessibleComboBoxKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleComboBox::key_bindings_for_action method.

AccessibleComboBoxLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleComboBox::localized_action_description method.

AccessibleComboBoxLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleComboBox::localized_action_name method.

AccessibleComboBoxNewArgs
This trait represents a set of arguments accepted by AccessibleComboBox::new method.

AccessibleComboBoxNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleComboBox::next_page_action method.

AccessibleComboBoxPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleComboBox::previous_page_action method.

AccessibleComboBoxRectArgs
This trait represents a set of arguments accepted by AccessibleComboBox::rect method.

AccessibleComboBoxRelationsArgs
This trait represents a set of arguments accepted by AccessibleComboBox::relations method.

AccessibleComboBoxScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleComboBox::scroll_down_action method.

AccessibleComboBoxScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleComboBox::scroll_left_action method.

AccessibleComboBoxScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleComboBox::scroll_right_action method.

AccessibleComboBoxScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleComboBox::scroll_up_action method.

AccessibleComboBoxStateArgs
This trait represents a set of arguments accepted by AccessibleComboBox::state method.

AccessibleComboBoxTextArgs
This trait represents a set of arguments accepted by AccessibleComboBox::text method.

AccessibleComboBoxTrArgs
This trait represents a set of arguments accepted by AccessibleComboBox::tr method.

AccessibleComboBoxTrUtf8Args
This trait represents a set of arguments accepted by AccessibleComboBox::tr_utf8 method.