Module qt_widgets::accessible_dial::overloading

Traits

AccessibleDialActionNamesArgs
This trait represents a set of arguments accepted by AccessibleDial::action_names method.

AccessibleDialBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleDial::background_color method.

AccessibleDialCurrentValueArgs
This trait represents a set of arguments accepted by AccessibleDial::current_value method.

AccessibleDialForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleDial::foreground_color method.

AccessibleDialKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleDial::key_bindings_for_action method.

AccessibleDialLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleDial::localized_action_description method.

AccessibleDialLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleDial::localized_action_name method.

AccessibleDialMaximumValueArgs
This trait represents a set of arguments accepted by AccessibleDial::maximum_value method.

AccessibleDialMinimumStepSizeArgs
This trait represents a set of arguments accepted by AccessibleDial::minimum_step_size method.

AccessibleDialMinimumValueArgs
This trait represents a set of arguments accepted by AccessibleDial::minimum_value method.

AccessibleDialNewArgs
This trait represents a set of arguments accepted by AccessibleDial::new method.

AccessibleDialNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleDial::next_page_action method.

AccessibleDialPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleDial::previous_page_action method.

AccessibleDialRectArgs
This trait represents a set of arguments accepted by AccessibleDial::rect method.

AccessibleDialRelationsArgs
This trait represents a set of arguments accepted by AccessibleDial::relations method.

AccessibleDialScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleDial::scroll_down_action method.

AccessibleDialScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleDial::scroll_left_action method.

AccessibleDialScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleDial::scroll_right_action method.

AccessibleDialScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleDial::scroll_up_action method.

AccessibleDialStateArgs
This trait represents a set of arguments accepted by AccessibleDial::state method.

AccessibleDialTextArgs
This trait represents a set of arguments accepted by AccessibleDial::text method.

AccessibleDialTrArgs
This trait represents a set of arguments accepted by AccessibleDial::tr method.

AccessibleDialTrUtf8Args
This trait represents a set of arguments accepted by AccessibleDial::tr_utf8 method.