Module qt_widgets::accessible_dialog_button_box::overloading

Traits

AccessibleDialogButtonBoxActionNamesArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::action_names method.

AccessibleDialogButtonBoxBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::background_color method.

AccessibleDialogButtonBoxForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::foreground_color method.

AccessibleDialogButtonBoxKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::key_bindings_for_action method.

AccessibleDialogButtonBoxLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::localized_action_description method.

AccessibleDialogButtonBoxLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::localized_action_name method.

AccessibleDialogButtonBoxNewArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::new method.

AccessibleDialogButtonBoxNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::next_page_action method.

AccessibleDialogButtonBoxPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::previous_page_action method.

AccessibleDialogButtonBoxRectArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::rect method.

AccessibleDialogButtonBoxRelationsArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::relations method.

AccessibleDialogButtonBoxScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::scroll_down_action method.

AccessibleDialogButtonBoxScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::scroll_left_action method.

AccessibleDialogButtonBoxScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::scroll_right_action method.

AccessibleDialogButtonBoxScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::scroll_up_action method.

AccessibleDialogButtonBoxStateArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::state method.

AccessibleDialogButtonBoxTextArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::text method.

AccessibleDialogButtonBoxTrArgs
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::tr method.

AccessibleDialogButtonBoxTrUtf8Args
This trait represents a set of arguments accepted by AccessibleDialogButtonBox::tr_utf8 method.