Module qt_widgets::accessible_display::overloading

Traits

AccessibleDisplayActionNamesArgs
This trait represents a set of arguments accepted by AccessibleDisplay::action_names method.

AccessibleDisplayBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleDisplay::background_color method.

AccessibleDisplayForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleDisplay::foreground_color method.

AccessibleDisplayImageDescriptionArgs
This trait represents a set of arguments accepted by AccessibleDisplay::image_description method.

AccessibleDisplayImagePositionArgs
This trait represents a set of arguments accepted by AccessibleDisplay::image_position method.

AccessibleDisplayImageSizeArgs
This trait represents a set of arguments accepted by AccessibleDisplay::image_size method.

AccessibleDisplayKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleDisplay::key_bindings_for_action method.

AccessibleDisplayLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleDisplay::localized_action_description method.

AccessibleDisplayLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleDisplay::localized_action_name method.

AccessibleDisplayNewArgs
This trait represents a set of arguments accepted by AccessibleDisplay::new method.

AccessibleDisplayNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleDisplay::next_page_action method.

AccessibleDisplayPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleDisplay::previous_page_action method.

AccessibleDisplayRectArgs
This trait represents a set of arguments accepted by AccessibleDisplay::rect method.

AccessibleDisplayRelationsArgs
This trait represents a set of arguments accepted by AccessibleDisplay::relations method.

AccessibleDisplayScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleDisplay::scroll_down_action method.

AccessibleDisplayScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleDisplay::scroll_left_action method.

AccessibleDisplayScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleDisplay::scroll_right_action method.

AccessibleDisplayScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleDisplay::scroll_up_action method.

AccessibleDisplayStateArgs
This trait represents a set of arguments accepted by AccessibleDisplay::state method.

AccessibleDisplayTextArgs
This trait represents a set of arguments accepted by AccessibleDisplay::text method.

AccessibleDisplayTrArgs
This trait represents a set of arguments accepted by AccessibleDisplay::tr method.

AccessibleDisplayTrUtf8Args
This trait represents a set of arguments accepted by AccessibleDisplay::tr_utf8 method.