Module qt_widgets::accessible_dock_widget::overloading

Traits

AccessibleDockWidgetActionNamesArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::action_names method.

AccessibleDockWidgetBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::background_color method.

AccessibleDockWidgetForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::foreground_color method.

AccessibleDockWidgetKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::key_bindings_for_action method.

AccessibleDockWidgetLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::localized_action_description method.

AccessibleDockWidgetLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::localized_action_name method.

AccessibleDockWidgetNewArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::new method.

AccessibleDockWidgetNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::next_page_action method.

AccessibleDockWidgetPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::previous_page_action method.

AccessibleDockWidgetRectArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::rect method.

AccessibleDockWidgetRelationsArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::relations method.

AccessibleDockWidgetScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::scroll_down_action method.

AccessibleDockWidgetScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::scroll_left_action method.

AccessibleDockWidgetScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::scroll_right_action method.

AccessibleDockWidgetScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::scroll_up_action method.

AccessibleDockWidgetStateArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::state method.

AccessibleDockWidgetTextArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::text method.

AccessibleDockWidgetTrArgs
This trait represents a set of arguments accepted by AccessibleDockWidget::tr method.

AccessibleDockWidgetTrUtf8Args
This trait represents a set of arguments accepted by AccessibleDockWidget::tr_utf8 method.