Module qt_widgets::accessible_double_spin_box::overloading

Traits

AccessibleDoubleSpinBoxActionNamesArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::action_names method.

AccessibleDoubleSpinBoxAttributesArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::attributes method.

AccessibleDoubleSpinBoxBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::background_color method.

AccessibleDoubleSpinBoxCharacterRectArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::character_rect method.

AccessibleDoubleSpinBoxCurrentValueArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::current_value method.

AccessibleDoubleSpinBoxForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::foreground_color method.

AccessibleDoubleSpinBoxKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::key_bindings_for_action method.

AccessibleDoubleSpinBoxLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::localized_action_description method.

AccessibleDoubleSpinBoxLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::localized_action_name method.

AccessibleDoubleSpinBoxMaximumValueArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::maximum_value method.

AccessibleDoubleSpinBoxMinimumStepSizeArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::minimum_step_size method.

AccessibleDoubleSpinBoxMinimumValueArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::minimum_value method.

AccessibleDoubleSpinBoxNewArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::new method.

AccessibleDoubleSpinBoxNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::next_page_action method.

AccessibleDoubleSpinBoxPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::previous_page_action method.

AccessibleDoubleSpinBoxRectArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::rect method.

AccessibleDoubleSpinBoxRelationsArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::relations method.

AccessibleDoubleSpinBoxScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::scroll_down_action method.

AccessibleDoubleSpinBoxScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::scroll_left_action method.

AccessibleDoubleSpinBoxScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::scroll_right_action method.

AccessibleDoubleSpinBoxScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::scroll_up_action method.

AccessibleDoubleSpinBoxStateArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::state method.

AccessibleDoubleSpinBoxTextAfterOffsetArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::text_after_offset method.

AccessibleDoubleSpinBoxTextArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::text method.

AccessibleDoubleSpinBoxTextAtOffsetArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::text_at_offset method.

AccessibleDoubleSpinBoxTextBeforeOffsetArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::text_before_offset method.

AccessibleDoubleSpinBoxTrArgs
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::tr method.

AccessibleDoubleSpinBoxTrUtf8Args
This trait represents a set of arguments accepted by AccessibleDoubleSpinBox::tr_utf8 method.