Struct qt_widgets::accessible_double_spin_box::AccessibleDoubleSpinBox

pub struct AccessibleDoubleSpinBox {
    // some fields omitted
}

C++ type: QAccessibleDoubleSpinBox

Methods

impl AccessibleDoubleSpinBox

fn action_interface(&mut self) -> *mut AccessibleActionInterface

C++ method: QAccessibleActionInterface* QAccessibleDoubleSpinBox::actionInterface()

fn action_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxActionNamesArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::actionNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn action_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn action_names(&self, cpp_utils::AsStruct) -> qt_core::string_list::StringList
C++ method: virtual QStringList QAccessibleDoubleSpinBox::actionNames() const

C++ documentation:

Reimplemented from QAccessibleActionInterface::actionNames().

fn add_selection(&mut self, start_offset: c_int, end_offset: c_int)

C++ method: virtual void QAccessibleDoubleSpinBox::addSelection(int startOffset, int endOffset)

fn attributes<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxAttributesArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::attributes

This is an overloaded function. Available variants:

Rust arguments:
1) fn attributes(&self, (libc::c_int, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn attributes(&self, (libc::c_int, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleDoubleSpinBox::attributes(int offset, int* startOffset, int* endOffset) const

fn background_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxBackgroundColorArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::backgroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn background_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn background_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: virtual QColor QAccessibleDoubleSpinBox::backgroundColor() const

C++ documentation:

Reimplemented from QAccessibleInterface::backgroundColor().

fn character_count(&self) -> c_int

C++ method: virtual int QAccessibleDoubleSpinBox::characterCount() const

fn character_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxCharacterRectArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::characterRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn character_rect(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn character_rect(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::rect::Rect
C++ method: virtual QRect QAccessibleDoubleSpinBox::characterRect(int offset) const

fn child(&self, index: c_int) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleDoubleSpinBox::child(int index) const

C++ documentation:

Reimplemented from QAccessibleInterface::child().

fn child_at(&self, x: c_int, y: c_int) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleDoubleSpinBox::childAt(int x, int y) const

C++ documentation:

Reimplemented from QAccessibleInterface::childAt().

fn child_count(&self) -> c_int

C++ method: virtual int QAccessibleDoubleSpinBox::childCount() const

C++ documentation:

Reimplemented from QAccessibleInterface::childCount().

fn current_value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxCurrentValueArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::currentValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn current_value(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: virtual QVariant QAccessibleDoubleSpinBox::currentValue() const

fn cursor_position(&self) -> c_int

C++ method: virtual int QAccessibleDoubleSpinBox::cursorPosition() const

fn decrease_action() -> &'static String

C++ method: static const QString& QAccessibleDoubleSpinBox::decreaseAction()

C++ documentation:

Returns the name of the decrease default action.

See also actionNames() and localizedActionName().

fn delete_text(&mut self, start_offset: c_int, end_offset: c_int)

C++ method: virtual void QAccessibleDoubleSpinBox::deleteText(int startOffset, int endOffset)

fn do_action(&mut self, action_name: &String)

C++ method: virtual void QAccessibleDoubleSpinBox::doAction(const QString& actionName)

C++ documentation:

Reimplemented from QAccessibleActionInterface::doAction().

fn editable_text_interface(&mut self) -> *mut AccessibleEditableTextInterface

C++ method: QAccessibleEditableTextInterface* QAccessibleDoubleSpinBox::editableTextInterface()

fn focus_child(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleDoubleSpinBox::focusChild() const

C++ documentation:

Reimplemented from QAccessibleInterface::focusChild().

fn foreground_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxForegroundColorArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::foregroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn foreground_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn foreground_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: virtual QColor QAccessibleDoubleSpinBox::foregroundColor() const

C++ documentation:

Reimplemented from QAccessibleInterface::foregroundColor().

fn image_interface(&mut self) -> *mut AccessibleImageInterface

C++ method: QAccessibleImageInterface* QAccessibleDoubleSpinBox::imageInterface()

fn increase_action() -> &'static String

C++ method: static const QString& QAccessibleDoubleSpinBox::increaseAction()

C++ documentation:

Returns the name of the increase default action.

See also actionNames() and localizedActionName().

fn index_of_child(&self, child: *const AccessibleInterface) -> c_int

C++ method: virtual int QAccessibleDoubleSpinBox::indexOfChild(const QAccessibleInterface* child) const

C++ documentation:

Reimplemented from QAccessibleInterface::indexOfChild().

fn insert_text(&mut self, offset: c_int, text: &String)

C++ method: virtual void QAccessibleDoubleSpinBox::insertText(int offset, const QString& text)

fn interface_cast(&mut self, t: InterfaceType) -> *mut c_void

C++ method: virtual void* QAccessibleDoubleSpinBox::interface_cast(QAccessible::InterfaceType t)

fn is_valid(&self) -> bool

C++ method: virtual bool QAccessibleDoubleSpinBox::isValid() const

C++ documentation:

Reimplemented from QAccessibleInterface::isValid().

fn key_bindings_for_action<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxKeyBindingsForActionArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::keyBindingsForAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn key_bindings_for_action(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn key_bindings_for_action(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string_list::StringList
C++ method: virtual QStringList QAccessibleDoubleSpinBox::keyBindingsForAction(const QString& actionName) const

C++ documentation:

Reimplemented from QAccessibleActionInterface::keyBindingsForAction().

fn localized_action_description<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxLocalizedActionDescriptionArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::localizedActionDescription

This is an overloaded function. Available variants:

Rust arguments:
1) fn localized_action_description(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn localized_action_description(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleDoubleSpinBox::localizedActionDescription(const QString& name) const

C++ documentation:

Returns a localized action description of the action actionName.

When using one of the default names, you can call this function in QAccessibleActionInterface to get the localized string.

See also actionNames() and localizedActionName().

fn localized_action_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxLocalizedActionNameArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::localizedActionName

This is an overloaded function. Available variants:

Rust arguments:
1) fn localized_action_name(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn localized_action_name(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleDoubleSpinBox::localizedActionName(const QString& name) const

C++ documentation:

Returns a localized action name of actionName.

For custom actions this function has to be re-implemented. When using one of the default names, you can call this function in QAccessibleActionInterface to get the localized string.

See also actionNames() and localizedActionDescription().

fn maximum_value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxMaximumValueArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::maximumValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn maximum_value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn maximum_value(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: virtual QVariant QAccessibleDoubleSpinBox::maximumValue() const

fn minimum_step_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxMinimumStepSizeArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::minimumStepSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn minimum_step_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn minimum_step_size(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: virtual QVariant QAccessibleDoubleSpinBox::minimumStepSize() const

fn minimum_value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxMinimumValueArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::minimumValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn minimum_value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn minimum_value(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: virtual QVariant QAccessibleDoubleSpinBox::minimumValue() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxNewArgs

C++ method: QAccessibleDoubleSpinBox::QAccessibleDoubleSpinBox

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((*mut ::widget::Widget, cpp_utils::AsStruct)) -> ::accessible_double_spin_box::AccessibleDoubleSpinBox
2) fn new((*mut ::widget::Widget, cpp_utils::AsBox)) -> cpp_utils::CppBox<::accessible_double_spin_box::AccessibleDoubleSpinBox>
C++ method: [constructor] void QAccessibleDoubleSpinBox::QAccessibleDoubleSpinBox(QWidget* widget)

fn next_page_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxNextPageActionArgs

C++ method: QAccessibleDoubleSpinBox::nextPageAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn next_page_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn next_page_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleDoubleSpinBox::nextPageAction()

C++ documentation:

Returns the name of the next page default action.

See also actionNames() and localizedActionName().

fn object(&self) -> *mut Object

C++ method: virtual QObject* QAccessibleDoubleSpinBox::object() const

C++ documentation:

Reimplemented from QAccessibleInterface::object().

fn offset_at_point(&self, point: &Point) -> c_int

C++ method: virtual int QAccessibleDoubleSpinBox::offsetAtPoint(const QPoint& point) const

fn parent(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleDoubleSpinBox::parent() const

C++ documentation:

Reimplemented from QAccessibleInterface::parent().

fn press_action() -> &'static String

C++ method: static const QString& QAccessibleDoubleSpinBox::pressAction()

C++ documentation:

Returns the name of the press default action.

See also actionNames() and localizedActionName().

fn previous_page_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxPreviousPageActionArgs

C++ method: QAccessibleDoubleSpinBox::previousPageAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn previous_page_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn previous_page_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleDoubleSpinBox::previousPageAction()

C++ documentation:

Returns the name of the previous page default action.

See also actionNames() and localizedActionName().

fn rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxRectArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::rect

This is an overloaded function. Available variants:

Rust arguments:
1) fn rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: virtual QRect QAccessibleDoubleSpinBox::rect() const

C++ documentation:

Reimplemented from QAccessibleInterface::rect().

fn relations<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxRelationsArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::relations

This is an overloaded function. Available variants:

Rust arguments:
1) fn relations(&self, (qt_core::flags::Flags<qt_gui::accessible::RelationFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
2) fn relations(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
3) fn relations(&self, (qt_core::flags::Flags<qt_gui::accessible::RelationFlag>, cpp_utils::AsStruct)) -> qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
4) fn relations(&self, cpp_utils::AsStruct) -> qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
C++ method: virtual QVector<QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>> QAccessibleDoubleSpinBox::relations(QFlags<QAccessible::RelationFlag> match = ?) const

C++ documentation:

Reimplemented from QAccessibleInterface::relations().

fn remove_selection(&mut self, selection_index: c_int)

C++ method: virtual void QAccessibleDoubleSpinBox::removeSelection(int selectionIndex)

fn replace_text(&mut self, start_offset: c_int, end_offset: c_int, text: &String)

C++ method: virtual void QAccessibleDoubleSpinBox::replaceText(int startOffset, int endOffset, const QString& text)

fn role(&self) -> Role

C++ method: virtual QAccessible::Role QAccessibleDoubleSpinBox::role() const

C++ documentation:

Reimplemented from QAccessibleInterface::role().

fn scroll_down_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxScrollDownActionArgs

C++ method: QAccessibleDoubleSpinBox::scrollDownAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_down_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_down_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleDoubleSpinBox::scrollDownAction()

C++ documentation:

Returns the name of the scroll down default action.

See also actionNames() and localizedActionName().

fn scroll_left_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxScrollLeftActionArgs

C++ method: QAccessibleDoubleSpinBox::scrollLeftAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_left_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_left_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleDoubleSpinBox::scrollLeftAction()

C++ documentation:

Returns the name of the scroll left default action.

See also actionNames() and localizedActionName().

fn scroll_right_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxScrollRightActionArgs

C++ method: QAccessibleDoubleSpinBox::scrollRightAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_right_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_right_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleDoubleSpinBox::scrollRightAction()

C++ documentation:

Returns the name of the scroll right default action.

See also actionNames() and localizedActionName().

fn scroll_to_substring(&mut self, start_index: c_int, end_index: c_int)

C++ method: virtual void QAccessibleDoubleSpinBox::scrollToSubstring(int startIndex, int endIndex)

fn scroll_up_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxScrollUpActionArgs

C++ method: QAccessibleDoubleSpinBox::scrollUpAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_up_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_up_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleDoubleSpinBox::scrollUpAction()

C++ documentation:

Returns the name of the scroll up default action.

See also actionNames() and localizedActionName().

fn selection(&self, selection_index: c_int, start_offset: *mut c_int, end_offset: *mut c_int)

C++ method: virtual void QAccessibleDoubleSpinBox::selection(int selectionIndex, int* startOffset, int* endOffset) const

fn selection_count(&self) -> c_int

C++ method: virtual int QAccessibleDoubleSpinBox::selectionCount() const

fn set_current_value(&mut self, value: &Variant)

C++ method: virtual void QAccessibleDoubleSpinBox::setCurrentValue(const QVariant& value)

fn set_cursor_position(&mut self, position: c_int)

C++ method: virtual void QAccessibleDoubleSpinBox::setCursorPosition(int position)

fn set_focus_action() -> &'static String

C++ method: static const QString& QAccessibleDoubleSpinBox::setFocusAction()

C++ documentation:

Returns the name of the set focus default action.

See also actionNames() and localizedActionName().

fn set_selection(&mut self, selection_index: c_int, start_offset: c_int, end_offset: c_int)

C++ method: virtual void QAccessibleDoubleSpinBox::setSelection(int selectionIndex, int startOffset, int endOffset)

fn set_text(&mut self, t: Text, text: &String)

C++ method: virtual void QAccessibleDoubleSpinBox::setText(QAccessible::Text t, const QString& text)

C++ documentation:

Reimplemented from QAccessibleInterface::setText().

fn show_menu_action() -> &'static String

C++ method: static const QString& QAccessibleDoubleSpinBox::showMenuAction()

C++ documentation:

Returns the name of the show menu default action.

See also actionNames() and localizedActionName().

fn state<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxStateArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::state

This is an overloaded function. Available variants:

Rust arguments:
1) fn state(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::accessible::State>
2) fn state(&self, cpp_utils::AsStruct) -> qt_gui::accessible::State
C++ method: virtual QAccessible::State QAccessibleDoubleSpinBox::state() const

C++ documentation:

Reimplemented from QAccessibleInterface::state().

fn table_cell_interface(&mut self) -> *mut AccessibleTableCellInterface

C++ method: QAccessibleTableCellInterface* QAccessibleDoubleSpinBox::tableCellInterface()

fn table_interface(&mut self) -> *mut AccessibleTableInterface

C++ method: QAccessibleTableInterface* QAccessibleDoubleSpinBox::tableInterface()

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxTextArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::text

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn text(&self, (qt_gui::accessible::Text, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, (qt_gui::accessible::Text, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleDoubleSpinBox::text(QAccessible::Text t) const

Variant 2

Rust arguments:
1) fn text(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleDoubleSpinBox::text(int startOffset, int endOffset) const

fn text_after_offset<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxTextAfterOffsetArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::textAfterOffset

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_after_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text_after_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleDoubleSpinBox::textAfterOffset(int offset, QAccessible::TextBoundaryType boundaryType, int* startOffset, int* endOffset) const

fn text_at_offset<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxTextAtOffsetArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::textAtOffset

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_at_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text_at_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleDoubleSpinBox::textAtOffset(int offset, QAccessible::TextBoundaryType boundaryType, int* startOffset, int* endOffset) const

fn text_before_offset<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxTextBeforeOffsetArgs<'largs>

C++ method: QAccessibleDoubleSpinBox::textBeforeOffset

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_before_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text_before_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleDoubleSpinBox::textBeforeOffset(int offset, QAccessible::TextBoundaryType boundaryType, int* endOffset, int* startOffset) const

fn text_interface(&mut self) -> *mut AccessibleTextInterface

C++ method: QAccessibleTextInterface* QAccessibleDoubleSpinBox::textInterface()

fn toggle_action() -> &'static String

C++ method: static const QString& QAccessibleDoubleSpinBox::toggleAction()

C++ documentation:

Returns the name of the toggle default action.

See also actionNames() and localizedActionName().

fn tr<Args>(args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxTrArgs

C++ method: QAccessibleDoubleSpinBox::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QAccessibleDoubleSpinBox::tr(const char* sourceText, const char* disambiguation, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: AccessibleDoubleSpinBoxTrUtf8Args

C++ method: QAccessibleDoubleSpinBox::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QAccessibleDoubleSpinBox::trUtf8(const char* sourceText, const char* disambiguation, int n)

fn value_interface(&mut self) -> *mut AccessibleValueInterface

C++ method: QAccessibleValueInterface* QAccessibleDoubleSpinBox::valueInterface()

fn virtual_hook(&mut self, id: c_int, data: *mut c_void)

C++ method: virtual void QAccessibleDoubleSpinBox::virtual_hook(int id, void* data)

fn window(&self) -> *mut Window

C++ method: virtual QWindow* QAccessibleDoubleSpinBox::window() const

C++ documentation:

Reimplemented from QAccessibleInterface::window().