Module qt_widgets::accessible_group_box::overloading

Traits

AccessibleGroupBoxActionNamesArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::action_names method.

AccessibleGroupBoxBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::background_color method.

AccessibleGroupBoxForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::foreground_color method.

AccessibleGroupBoxKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::key_bindings_for_action method.

AccessibleGroupBoxLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::localized_action_description method.

AccessibleGroupBoxLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::localized_action_name method.

AccessibleGroupBoxNewArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::new method.

AccessibleGroupBoxNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::next_page_action method.

AccessibleGroupBoxPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::previous_page_action method.

AccessibleGroupBoxRectArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::rect method.

AccessibleGroupBoxRelationsArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::relations method.

AccessibleGroupBoxScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::scroll_down_action method.

AccessibleGroupBoxScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::scroll_left_action method.

AccessibleGroupBoxScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::scroll_right_action method.

AccessibleGroupBoxScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::scroll_up_action method.

AccessibleGroupBoxStateArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::state method.

AccessibleGroupBoxTextArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::text method.

AccessibleGroupBoxTrArgs
This trait represents a set of arguments accepted by AccessibleGroupBox::tr method.

AccessibleGroupBoxTrUtf8Args
This trait represents a set of arguments accepted by AccessibleGroupBox::tr_utf8 method.