Struct qt_widgets::accessible_line_edit::AccessibleLineEdit

pub struct AccessibleLineEdit {
    // some fields omitted
}

C++ type: QAccessibleLineEdit

Methods

impl AccessibleLineEdit

fn action_interface(&mut self) -> *mut AccessibleActionInterface

C++ method: QAccessibleActionInterface* QAccessibleLineEdit::actionInterface()

fn action_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleLineEditActionNamesArgs<'largs>

C++ method: QAccessibleLineEdit::actionNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn action_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn action_names(&self, cpp_utils::AsStruct) -> qt_core::string_list::StringList
C++ method: virtual QStringList QAccessibleLineEdit::actionNames() const

C++ documentation:

Reimplemented from QAccessibleActionInterface::actionNames().

fn add_selection(&mut self, start_offset: c_int, end_offset: c_int)

C++ method: virtual void QAccessibleLineEdit::addSelection(int startOffset, int endOffset)

fn attributes<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleLineEditAttributesArgs<'largs>

C++ method: QAccessibleLineEdit::attributes

This is an overloaded function. Available variants:

Rust arguments:
1) fn attributes(&self, (libc::c_int, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn attributes(&self, (libc::c_int, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleLineEdit::attributes(int offset, int* startOffset, int* endOffset) const

fn background_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleLineEditBackgroundColorArgs<'largs>

C++ method: QAccessibleLineEdit::backgroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn background_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn background_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: virtual QColor QAccessibleLineEdit::backgroundColor() const

C++ documentation:

Reimplemented from QAccessibleInterface::backgroundColor().

fn character_count(&self) -> c_int

C++ method: virtual int QAccessibleLineEdit::characterCount() const

fn character_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleLineEditCharacterRectArgs<'largs>

C++ method: QAccessibleLineEdit::characterRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn character_rect(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn character_rect(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::rect::Rect
C++ method: virtual QRect QAccessibleLineEdit::characterRect(int offset) const

fn child(&self, index: c_int) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleLineEdit::child(int index) const

C++ documentation:

Reimplemented from QAccessibleInterface::child().

fn child_at(&self, x: c_int, y: c_int) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleLineEdit::childAt(int x, int y) const

C++ documentation:

Reimplemented from QAccessibleInterface::childAt().

fn child_count(&self) -> c_int

C++ method: virtual int QAccessibleLineEdit::childCount() const

C++ documentation:

Reimplemented from QAccessibleInterface::childCount().

fn cursor_position(&self) -> c_int

C++ method: virtual int QAccessibleLineEdit::cursorPosition() const

fn decrease_action() -> &'static String

C++ method: static const QString& QAccessibleLineEdit::decreaseAction()

C++ documentation:

Returns the name of the decrease default action.

See also actionNames() and localizedActionName().

fn delete_text(&mut self, start_offset: c_int, end_offset: c_int)

C++ method: virtual void QAccessibleLineEdit::deleteText(int startOffset, int endOffset)

fn do_action(&mut self, action_name: &String)

C++ method: virtual void QAccessibleLineEdit::doAction(const QString& actionName)

C++ documentation:

Reimplemented from QAccessibleActionInterface::doAction().

fn editable_text_interface(&mut self) -> *mut AccessibleEditableTextInterface

C++ method: QAccessibleEditableTextInterface* QAccessibleLineEdit::editableTextInterface()

fn focus_child(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleLineEdit::focusChild() const

C++ documentation:

Reimplemented from QAccessibleInterface::focusChild().

fn foreground_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleLineEditForegroundColorArgs<'largs>

C++ method: QAccessibleLineEdit::foregroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn foreground_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn foreground_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: virtual QColor QAccessibleLineEdit::foregroundColor() const

C++ documentation:

Reimplemented from QAccessibleInterface::foregroundColor().

fn image_interface(&mut self) -> *mut AccessibleImageInterface

C++ method: QAccessibleImageInterface* QAccessibleLineEdit::imageInterface()

fn increase_action() -> &'static String

C++ method: static const QString& QAccessibleLineEdit::increaseAction()

C++ documentation:

Returns the name of the increase default action.

See also actionNames() and localizedActionName().

fn index_of_child(&self, child: *const AccessibleInterface) -> c_int

C++ method: virtual int QAccessibleLineEdit::indexOfChild(const QAccessibleInterface* child) const

C++ documentation:

Reimplemented from QAccessibleInterface::indexOfChild().

fn insert_text(&mut self, offset: c_int, text: &String)

C++ method: virtual void QAccessibleLineEdit::insertText(int offset, const QString& text)

fn interface_cast(&mut self, t: InterfaceType) -> *mut c_void

C++ method: virtual void* QAccessibleLineEdit::interface_cast(QAccessible::InterfaceType t)

fn is_valid(&self) -> bool

C++ method: virtual bool QAccessibleLineEdit::isValid() const

C++ documentation:

Reimplemented from QAccessibleInterface::isValid().

fn key_bindings_for_action<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleLineEditKeyBindingsForActionArgs<'largs>

C++ method: QAccessibleLineEdit::keyBindingsForAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn key_bindings_for_action(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn key_bindings_for_action(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string_list::StringList
C++ method: virtual QStringList QAccessibleLineEdit::keyBindingsForAction(const QString& actionName) const

C++ documentation:

Reimplemented from QAccessibleActionInterface::keyBindingsForAction().

fn localized_action_description<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleLineEditLocalizedActionDescriptionArgs<'largs>

C++ method: QAccessibleLineEdit::localizedActionDescription

This is an overloaded function. Available variants:

Rust arguments:
1) fn localized_action_description(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn localized_action_description(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleLineEdit::localizedActionDescription(const QString& name) const

C++ documentation:

Returns a localized action description of the action actionName.

When using one of the default names, you can call this function in QAccessibleActionInterface to get the localized string.

See also actionNames() and localizedActionName().

fn localized_action_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleLineEditLocalizedActionNameArgs<'largs>

C++ method: QAccessibleLineEdit::localizedActionName

This is an overloaded function. Available variants:

Rust arguments:
1) fn localized_action_name(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn localized_action_name(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleLineEdit::localizedActionName(const QString& name) const

C++ documentation:

Returns a localized action name of actionName.

For custom actions this function has to be re-implemented. When using one of the default names, you can call this function in QAccessibleActionInterface to get the localized string.

See also actionNames() and localizedActionDescription().

fn new<Args>(args: Args) -> Args::ReturnType where Args: AccessibleLineEditNewArgs

C++ method: QAccessibleLineEdit::QAccessibleLineEdit

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((*mut ::widget::Widget, cpp_utils::AsStruct)) -> ::accessible_line_edit::AccessibleLineEdit
2) fn new((*mut ::widget::Widget, &qt_core::string::String, cpp_utils::AsStruct)) -> ::accessible_line_edit::AccessibleLineEdit
3) fn new((*mut ::widget::Widget, cpp_utils::AsBox)) -> cpp_utils::CppBox<::accessible_line_edit::AccessibleLineEdit>
4) fn new((*mut ::widget::Widget, &qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::accessible_line_edit::AccessibleLineEdit>
C++ method: [constructor] void QAccessibleLineEdit::QAccessibleLineEdit(QWidget* o, const QString& name = ?)

fn next_page_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleLineEditNextPageActionArgs

C++ method: QAccessibleLineEdit::nextPageAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn next_page_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn next_page_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleLineEdit::nextPageAction()

C++ documentation:

Returns the name of the next page default action.

See also actionNames() and localizedActionName().

fn object(&self) -> *mut Object

C++ method: virtual QObject* QAccessibleLineEdit::object() const

C++ documentation:

Reimplemented from QAccessibleInterface::object().

fn offset_at_point(&self, point: &Point) -> c_int

C++ method: virtual int QAccessibleLineEdit::offsetAtPoint(const QPoint& point) const

fn parent(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleLineEdit::parent() const

C++ documentation:

Reimplemented from QAccessibleInterface::parent().

fn press_action() -> &'static String

C++ method: static const QString& QAccessibleLineEdit::pressAction()

C++ documentation:

Returns the name of the press default action.

See also actionNames() and localizedActionName().

fn previous_page_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleLineEditPreviousPageActionArgs

C++ method: QAccessibleLineEdit::previousPageAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn previous_page_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn previous_page_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleLineEdit::previousPageAction()

C++ documentation:

Returns the name of the previous page default action.

See also actionNames() and localizedActionName().

fn rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleLineEditRectArgs<'largs>

C++ method: QAccessibleLineEdit::rect

This is an overloaded function. Available variants:

Rust arguments:
1) fn rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: virtual QRect QAccessibleLineEdit::rect() const

C++ documentation:

Reimplemented from QAccessibleInterface::rect().

fn relations<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleLineEditRelationsArgs<'largs>

C++ method: QAccessibleLineEdit::relations

This is an overloaded function. Available variants:

Rust arguments:
1) fn relations(&self, (qt_core::flags::Flags<qt_gui::accessible::RelationFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
2) fn relations(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
3) fn relations(&self, (qt_core::flags::Flags<qt_gui::accessible::RelationFlag>, cpp_utils::AsStruct)) -> qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
4) fn relations(&self, cpp_utils::AsStruct) -> qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
C++ method: virtual QVector<QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>> QAccessibleLineEdit::relations(QFlags<QAccessible::RelationFlag> match = ?) const

C++ documentation:

Reimplemented from QAccessibleInterface::relations().

fn remove_selection(&mut self, selection_index: c_int)

C++ method: virtual void QAccessibleLineEdit::removeSelection(int selectionIndex)

fn replace_text(&mut self, start_offset: c_int, end_offset: c_int, text: &String)

C++ method: virtual void QAccessibleLineEdit::replaceText(int startOffset, int endOffset, const QString& text)

fn role(&self) -> Role

C++ method: virtual QAccessible::Role QAccessibleLineEdit::role() const

C++ documentation:

Reimplemented from QAccessibleInterface::role().

fn scroll_down_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleLineEditScrollDownActionArgs

C++ method: QAccessibleLineEdit::scrollDownAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_down_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_down_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleLineEdit::scrollDownAction()

C++ documentation:

Returns the name of the scroll down default action.

See also actionNames() and localizedActionName().

fn scroll_left_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleLineEditScrollLeftActionArgs

C++ method: QAccessibleLineEdit::scrollLeftAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_left_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_left_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleLineEdit::scrollLeftAction()

C++ documentation:

Returns the name of the scroll left default action.

See also actionNames() and localizedActionName().

fn scroll_right_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleLineEditScrollRightActionArgs

C++ method: QAccessibleLineEdit::scrollRightAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_right_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_right_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleLineEdit::scrollRightAction()

C++ documentation:

Returns the name of the scroll right default action.

See also actionNames() and localizedActionName().

fn scroll_to_substring(&mut self, start_index: c_int, end_index: c_int)

C++ method: virtual void QAccessibleLineEdit::scrollToSubstring(int startIndex, int endIndex)

fn scroll_up_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleLineEditScrollUpActionArgs

C++ method: QAccessibleLineEdit::scrollUpAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_up_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_up_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleLineEdit::scrollUpAction()

C++ documentation:

Returns the name of the scroll up default action.

See also actionNames() and localizedActionName().

fn selection(&self, selection_index: c_int, start_offset: *mut c_int, end_offset: *mut c_int)

C++ method: virtual void QAccessibleLineEdit::selection(int selectionIndex, int* startOffset, int* endOffset) const

fn selection_count(&self) -> c_int

C++ method: virtual int QAccessibleLineEdit::selectionCount() const

fn set_cursor_position(&mut self, position: c_int)

C++ method: virtual void QAccessibleLineEdit::setCursorPosition(int position)

fn set_focus_action() -> &'static String

C++ method: static const QString& QAccessibleLineEdit::setFocusAction()

C++ documentation:

Returns the name of the set focus default action.

See also actionNames() and localizedActionName().

fn set_selection(&mut self, selection_index: c_int, start_offset: c_int, end_offset: c_int)

C++ method: virtual void QAccessibleLineEdit::setSelection(int selectionIndex, int startOffset, int endOffset)

fn set_text(&mut self, t: Text, text: &String)

C++ method: virtual void QAccessibleLineEdit::setText(QAccessible::Text t, const QString& text)

fn show_menu_action() -> &'static String

C++ method: static const QString& QAccessibleLineEdit::showMenuAction()

C++ documentation:

Returns the name of the show menu default action.

See also actionNames() and localizedActionName().

fn state<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleLineEditStateArgs<'largs>

C++ method: QAccessibleLineEdit::state

This is an overloaded function. Available variants:

Rust arguments:
1) fn state(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::accessible::State>
2) fn state(&self, cpp_utils::AsStruct) -> qt_gui::accessible::State
C++ method: virtual QAccessible::State QAccessibleLineEdit::state() const

fn table_cell_interface(&mut self) -> *mut AccessibleTableCellInterface

C++ method: QAccessibleTableCellInterface* QAccessibleLineEdit::tableCellInterface()

fn table_interface(&mut self) -> *mut AccessibleTableInterface

C++ method: QAccessibleTableInterface* QAccessibleLineEdit::tableInterface()

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleLineEditTextArgs<'largs>

C++ method: QAccessibleLineEdit::text

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn text(&self, (qt_gui::accessible::Text, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, (qt_gui::accessible::Text, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleLineEdit::text(QAccessible::Text t) const

Variant 2

Rust arguments:
1) fn text(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleLineEdit::text(int startOffset, int endOffset) const

fn text_after_offset<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleLineEditTextAfterOffsetArgs<'largs>

C++ method: QAccessibleLineEdit::textAfterOffset

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_after_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text_after_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleLineEdit::textAfterOffset(int offset, QAccessible::TextBoundaryType boundaryType, int* startOffset, int* endOffset) const

fn text_at_offset<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleLineEditTextAtOffsetArgs<'largs>

C++ method: QAccessibleLineEdit::textAtOffset

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_at_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text_at_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleLineEdit::textAtOffset(int offset, QAccessible::TextBoundaryType boundaryType, int* startOffset, int* endOffset) const

fn text_before_offset<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleLineEditTextBeforeOffsetArgs<'largs>

C++ method: QAccessibleLineEdit::textBeforeOffset

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_before_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text_before_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleLineEdit::textBeforeOffset(int offset, QAccessible::TextBoundaryType boundaryType, int* startOffset, int* endOffset) const

fn text_interface(&mut self) -> *mut AccessibleTextInterface

C++ method: QAccessibleTextInterface* QAccessibleLineEdit::textInterface()

fn toggle_action() -> &'static String

C++ method: static const QString& QAccessibleLineEdit::toggleAction()

C++ documentation:

Returns the name of the toggle default action.

See also actionNames() and localizedActionName().

fn tr<Args>(args: Args) -> Args::ReturnType where Args: AccessibleLineEditTrArgs

C++ method: QAccessibleLineEdit::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QAccessibleLineEdit::tr(const char* sourceText, const char* disambiguation, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: AccessibleLineEditTrUtf8Args

C++ method: QAccessibleLineEdit::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QAccessibleLineEdit::trUtf8(const char* sourceText, const char* disambiguation, int n)

fn value_interface(&mut self) -> *mut AccessibleValueInterface

C++ method: QAccessibleValueInterface* QAccessibleLineEdit::valueInterface()

fn virtual_hook(&mut self, id: c_int, data: *mut c_void)

C++ method: virtual void QAccessibleLineEdit::virtual_hook(int id, void* data)

fn window(&self) -> *mut Window

C++ method: virtual QWindow* QAccessibleLineEdit::window() const

C++ documentation:

Reimplemented from QAccessibleInterface::window().