Module qt_widgets::accessible_main_window::overloading

Traits

AccessibleMainWindowActionNamesArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::action_names method.

AccessibleMainWindowBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::background_color method.

AccessibleMainWindowForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::foreground_color method.

AccessibleMainWindowKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::key_bindings_for_action method.

AccessibleMainWindowLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::localized_action_description method.

AccessibleMainWindowLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::localized_action_name method.

AccessibleMainWindowNewArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::new method.

AccessibleMainWindowNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::next_page_action method.

AccessibleMainWindowPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::previous_page_action method.

AccessibleMainWindowRectArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::rect method.

AccessibleMainWindowRelationsArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::relations method.

AccessibleMainWindowScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::scroll_down_action method.

AccessibleMainWindowScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::scroll_left_action method.

AccessibleMainWindowScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::scroll_right_action method.

AccessibleMainWindowScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::scroll_up_action method.

AccessibleMainWindowStateArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::state method.

AccessibleMainWindowTextArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::text method.

AccessibleMainWindowTrArgs
This trait represents a set of arguments accepted by AccessibleMainWindow::tr method.

AccessibleMainWindowTrUtf8Args
This trait represents a set of arguments accepted by AccessibleMainWindow::tr_utf8 method.