Module qt_widgets::accessible_mdi_area::overloading

Traits

AccessibleMdiAreaActionNamesArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::action_names method.

AccessibleMdiAreaBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::background_color method.

AccessibleMdiAreaForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::foreground_color method.

AccessibleMdiAreaKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::key_bindings_for_action method.

AccessibleMdiAreaLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::localized_action_description method.

AccessibleMdiAreaLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::localized_action_name method.

AccessibleMdiAreaNewArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::new method.

AccessibleMdiAreaNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::next_page_action method.

AccessibleMdiAreaPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::previous_page_action method.

AccessibleMdiAreaRectArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::rect method.

AccessibleMdiAreaRelationsArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::relations method.

AccessibleMdiAreaScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::scroll_down_action method.

AccessibleMdiAreaScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::scroll_left_action method.

AccessibleMdiAreaScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::scroll_right_action method.

AccessibleMdiAreaScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::scroll_up_action method.

AccessibleMdiAreaStateArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::state method.

AccessibleMdiAreaTextArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::text method.

AccessibleMdiAreaTrArgs
This trait represents a set of arguments accepted by AccessibleMdiArea::tr method.

AccessibleMdiAreaTrUtf8Args
This trait represents a set of arguments accepted by AccessibleMdiArea::tr_utf8 method.