Module qt_widgets::accessible_mdi_sub_window::overloading

Traits

AccessibleMdiSubWindowActionNamesArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::action_names method.

AccessibleMdiSubWindowBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::background_color method.

AccessibleMdiSubWindowForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::foreground_color method.

AccessibleMdiSubWindowKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::key_bindings_for_action method.

AccessibleMdiSubWindowLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::localized_action_description method.

AccessibleMdiSubWindowLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::localized_action_name method.

AccessibleMdiSubWindowNewArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::new method.

AccessibleMdiSubWindowNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::next_page_action method.

AccessibleMdiSubWindowPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::previous_page_action method.

AccessibleMdiSubWindowRectArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::rect method.

AccessibleMdiSubWindowRelationsArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::relations method.

AccessibleMdiSubWindowScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::scroll_down_action method.

AccessibleMdiSubWindowScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::scroll_left_action method.

AccessibleMdiSubWindowScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::scroll_right_action method.

AccessibleMdiSubWindowScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::scroll_up_action method.

AccessibleMdiSubWindowStateArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::state method.

AccessibleMdiSubWindowTextArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::text method.

AccessibleMdiSubWindowTrArgs
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::tr method.

AccessibleMdiSubWindowTrUtf8Args
This trait represents a set of arguments accepted by AccessibleMdiSubWindow::tr_utf8 method.