Module qt_widgets::accessible_menu_bar::overloading

Traits

AccessibleMenuBarActionNamesArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::action_names method.

AccessibleMenuBarBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::background_color method.

AccessibleMenuBarForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::foreground_color method.

AccessibleMenuBarKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::key_bindings_for_action method.

AccessibleMenuBarLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::localized_action_description method.

AccessibleMenuBarLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::localized_action_name method.

AccessibleMenuBarNewArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::new method.

AccessibleMenuBarNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::next_page_action method.

AccessibleMenuBarPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::previous_page_action method.

AccessibleMenuBarRectArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::rect method.

AccessibleMenuBarRelationsArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::relations method.

AccessibleMenuBarScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::scroll_down_action method.

AccessibleMenuBarScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::scroll_left_action method.

AccessibleMenuBarScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::scroll_right_action method.

AccessibleMenuBarScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::scroll_up_action method.

AccessibleMenuBarStateArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::state method.

AccessibleMenuBarTextArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::text method.

AccessibleMenuBarTrArgs
This trait represents a set of arguments accepted by AccessibleMenuBar::tr method.

AccessibleMenuBarTrUtf8Args
This trait represents a set of arguments accepted by AccessibleMenuBar::tr_utf8 method.