Module qt_widgets::accessible_menu_item::overloading

Traits

AccessibleMenuItemActionNamesArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::action_names method.

AccessibleMenuItemBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::background_color method.

AccessibleMenuItemForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::foreground_color method.

AccessibleMenuItemKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::key_bindings_for_action method.

AccessibleMenuItemLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::localized_action_description method.

AccessibleMenuItemLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::localized_action_name method.

AccessibleMenuItemNewArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::new method.

AccessibleMenuItemNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::next_page_action method.

AccessibleMenuItemPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::previous_page_action method.

AccessibleMenuItemRectArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::rect method.

AccessibleMenuItemRelationsArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::relations method.

AccessibleMenuItemScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::scroll_down_action method.

AccessibleMenuItemScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::scroll_left_action method.

AccessibleMenuItemScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::scroll_right_action method.

AccessibleMenuItemScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::scroll_up_action method.

AccessibleMenuItemStateArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::state method.

AccessibleMenuItemTextArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::text method.

AccessibleMenuItemTrArgs
This trait represents a set of arguments accepted by AccessibleMenuItem::tr method.

AccessibleMenuItemTrUtf8Args
This trait represents a set of arguments accepted by AccessibleMenuItem::tr_utf8 method.