Module qt_widgets::accessible_plain_text_edit::overloading

Traits

AccessiblePlainTextEditActionNamesArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::action_names method.

AccessiblePlainTextEditAttributesArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::attributes method.

AccessiblePlainTextEditBackgroundColorArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::background_color method.

AccessiblePlainTextEditCharacterRectArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::character_rect method.

AccessiblePlainTextEditForegroundColorArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::foreground_color method.

AccessiblePlainTextEditKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::key_bindings_for_action method.

AccessiblePlainTextEditLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::localized_action_description method.

AccessiblePlainTextEditLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::localized_action_name method.

AccessiblePlainTextEditNewArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::new method.

AccessiblePlainTextEditNextPageActionArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::next_page_action method.

AccessiblePlainTextEditPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::previous_page_action method.

AccessiblePlainTextEditRectArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::rect method.

AccessiblePlainTextEditRelationsArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::relations method.

AccessiblePlainTextEditScrollDownActionArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::scroll_down_action method.

AccessiblePlainTextEditScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::scroll_left_action method.

AccessiblePlainTextEditScrollRightActionArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::scroll_right_action method.

AccessiblePlainTextEditScrollUpActionArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::scroll_up_action method.

AccessiblePlainTextEditStateArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::state method.

AccessiblePlainTextEditTextAfterOffsetArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::text_after_offset method.

AccessiblePlainTextEditTextArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::text method.

AccessiblePlainTextEditTextAtOffsetArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::text_at_offset method.

AccessiblePlainTextEditTextBeforeOffsetArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::text_before_offset method.

AccessiblePlainTextEditTrArgs
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::tr method.

AccessiblePlainTextEditTrUtf8Args
This trait represents a set of arguments accepted by AccessiblePlainTextEdit::tr_utf8 method.