Module qt_widgets::accessible_scroll_area::overloading

Traits

AccessibleScrollAreaActionNamesArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::action_names method.

AccessibleScrollAreaBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::background_color method.

AccessibleScrollAreaForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::foreground_color method.

AccessibleScrollAreaKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::key_bindings_for_action method.

AccessibleScrollAreaLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::localized_action_description method.

AccessibleScrollAreaLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::localized_action_name method.

AccessibleScrollAreaNewArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::new method.

AccessibleScrollAreaNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::next_page_action method.

AccessibleScrollAreaPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::previous_page_action method.

AccessibleScrollAreaRectArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::rect method.

AccessibleScrollAreaRelationsArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::relations method.

AccessibleScrollAreaScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::scroll_down_action method.

AccessibleScrollAreaScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::scroll_left_action method.

AccessibleScrollAreaScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::scroll_right_action method.

AccessibleScrollAreaScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::scroll_up_action method.

AccessibleScrollAreaStateArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::state method.

AccessibleScrollAreaTextArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::text method.

AccessibleScrollAreaTrArgs
This trait represents a set of arguments accepted by AccessibleScrollArea::tr method.

AccessibleScrollAreaTrUtf8Args
This trait represents a set of arguments accepted by AccessibleScrollArea::tr_utf8 method.