Module qt_widgets::accessible_scroll_bar::overloading

Traits

AccessibleScrollBarActionNamesArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::action_names method.

AccessibleScrollBarBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::background_color method.

AccessibleScrollBarCurrentValueArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::current_value method.

AccessibleScrollBarForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::foreground_color method.

AccessibleScrollBarKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::key_bindings_for_action method.

AccessibleScrollBarLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::localized_action_description method.

AccessibleScrollBarLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::localized_action_name method.

AccessibleScrollBarMaximumValueArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::maximum_value method.

AccessibleScrollBarMinimumStepSizeArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::minimum_step_size method.

AccessibleScrollBarMinimumValueArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::minimum_value method.

AccessibleScrollBarNewArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::new method.

AccessibleScrollBarNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::next_page_action method.

AccessibleScrollBarPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::previous_page_action method.

AccessibleScrollBarRectArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::rect method.

AccessibleScrollBarRelationsArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::relations method.

AccessibleScrollBarScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::scroll_down_action method.

AccessibleScrollBarScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::scroll_left_action method.

AccessibleScrollBarScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::scroll_right_action method.

AccessibleScrollBarScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::scroll_up_action method.

AccessibleScrollBarStateArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::state method.

AccessibleScrollBarTextArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::text method.

AccessibleScrollBarTrArgs
This trait represents a set of arguments accepted by AccessibleScrollBar::tr method.

AccessibleScrollBarTrUtf8Args
This trait represents a set of arguments accepted by AccessibleScrollBar::tr_utf8 method.