Struct qt_widgets::accessible_scroll_bar::AccessibleScrollBar

pub struct AccessibleScrollBar {
    // some fields omitted
}

C++ type: QAccessibleScrollBar

Methods

impl AccessibleScrollBar

fn action_interface(&mut self) -> *mut AccessibleActionInterface

C++ method: QAccessibleActionInterface* QAccessibleScrollBar::actionInterface()

fn action_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleScrollBarActionNamesArgs<'largs>

C++ method: QAccessibleScrollBar::actionNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn action_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn action_names(&self, cpp_utils::AsStruct) -> qt_core::string_list::StringList
C++ method: virtual QStringList QAccessibleScrollBar::actionNames() const

C++ documentation:

Reimplemented from QAccessibleActionInterface::actionNames().

fn background_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleScrollBarBackgroundColorArgs<'largs>

C++ method: QAccessibleScrollBar::backgroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn background_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn background_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: virtual QColor QAccessibleScrollBar::backgroundColor() const

C++ documentation:

Reimplemented from QAccessibleInterface::backgroundColor().

fn child(&self, index: c_int) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleScrollBar::child(int index) const

C++ documentation:

Reimplemented from QAccessibleInterface::child().

fn child_at(&self, x: c_int, y: c_int) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleScrollBar::childAt(int x, int y) const

C++ documentation:

Reimplemented from QAccessibleInterface::childAt().

fn child_count(&self) -> c_int

C++ method: virtual int QAccessibleScrollBar::childCount() const

C++ documentation:

Reimplemented from QAccessibleInterface::childCount().

fn current_value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleScrollBarCurrentValueArgs<'largs>

C++ method: QAccessibleScrollBar::currentValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn current_value(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: virtual QVariant QAccessibleScrollBar::currentValue() const

fn decrease_action() -> &'static String

C++ method: static const QString& QAccessibleScrollBar::decreaseAction()

C++ documentation:

Returns the name of the decrease default action.

See also actionNames() and localizedActionName().

fn do_action(&mut self, action_name: &String)

C++ method: virtual void QAccessibleScrollBar::doAction(const QString& actionName)

C++ documentation:

Reimplemented from QAccessibleActionInterface::doAction().

fn editable_text_interface(&mut self) -> *mut AccessibleEditableTextInterface

C++ method: QAccessibleEditableTextInterface* QAccessibleScrollBar::editableTextInterface()

fn focus_child(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleScrollBar::focusChild() const

C++ documentation:

Reimplemented from QAccessibleInterface::focusChild().

fn foreground_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleScrollBarForegroundColorArgs<'largs>

C++ method: QAccessibleScrollBar::foregroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn foreground_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn foreground_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: virtual QColor QAccessibleScrollBar::foregroundColor() const

C++ documentation:

Reimplemented from QAccessibleInterface::foregroundColor().

fn image_interface(&mut self) -> *mut AccessibleImageInterface

C++ method: QAccessibleImageInterface* QAccessibleScrollBar::imageInterface()

fn increase_action() -> &'static String

C++ method: static const QString& QAccessibleScrollBar::increaseAction()

C++ documentation:

Returns the name of the increase default action.

See also actionNames() and localizedActionName().

fn index_of_child(&self, child: *const AccessibleInterface) -> c_int

C++ method: virtual int QAccessibleScrollBar::indexOfChild(const QAccessibleInterface* child) const

C++ documentation:

Reimplemented from QAccessibleInterface::indexOfChild().

fn interface_cast(&mut self, t: InterfaceType) -> *mut c_void

C++ method: virtual void* QAccessibleScrollBar::interface_cast(QAccessible::InterfaceType t)

fn is_valid(&self) -> bool

C++ method: virtual bool QAccessibleScrollBar::isValid() const

C++ documentation:

Reimplemented from QAccessibleInterface::isValid().

fn key_bindings_for_action<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleScrollBarKeyBindingsForActionArgs<'largs>

C++ method: QAccessibleScrollBar::keyBindingsForAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn key_bindings_for_action(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn key_bindings_for_action(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string_list::StringList
C++ method: virtual QStringList QAccessibleScrollBar::keyBindingsForAction(const QString& actionName) const

C++ documentation:

Reimplemented from QAccessibleActionInterface::keyBindingsForAction().

fn localized_action_description<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleScrollBarLocalizedActionDescriptionArgs<'largs>

C++ method: QAccessibleScrollBar::localizedActionDescription

This is an overloaded function. Available variants:

Rust arguments:
1) fn localized_action_description(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn localized_action_description(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleScrollBar::localizedActionDescription(const QString& name) const

C++ documentation:

Returns a localized action description of the action actionName.

When using one of the default names, you can call this function in QAccessibleActionInterface to get the localized string.

See also actionNames() and localizedActionName().

fn localized_action_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleScrollBarLocalizedActionNameArgs<'largs>

C++ method: QAccessibleScrollBar::localizedActionName

This is an overloaded function. Available variants:

Rust arguments:
1) fn localized_action_name(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn localized_action_name(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleScrollBar::localizedActionName(const QString& name) const

C++ documentation:

Returns a localized action name of actionName.

For custom actions this function has to be re-implemented. When using one of the default names, you can call this function in QAccessibleActionInterface to get the localized string.

See also actionNames() and localizedActionDescription().

fn maximum_value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleScrollBarMaximumValueArgs<'largs>

C++ method: QAccessibleScrollBar::maximumValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn maximum_value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn maximum_value(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: virtual QVariant QAccessibleScrollBar::maximumValue() const

fn minimum_step_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleScrollBarMinimumStepSizeArgs<'largs>

C++ method: QAccessibleScrollBar::minimumStepSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn minimum_step_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn minimum_step_size(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: virtual QVariant QAccessibleScrollBar::minimumStepSize() const

fn minimum_value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleScrollBarMinimumValueArgs<'largs>

C++ method: QAccessibleScrollBar::minimumValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn minimum_value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn minimum_value(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: virtual QVariant QAccessibleScrollBar::minimumValue() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: AccessibleScrollBarNewArgs

C++ method: QAccessibleScrollBar::QAccessibleScrollBar

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((*mut ::widget::Widget, cpp_utils::AsStruct)) -> ::accessible_scroll_bar::AccessibleScrollBar
2) fn new((*mut ::widget::Widget, cpp_utils::AsBox)) -> cpp_utils::CppBox<::accessible_scroll_bar::AccessibleScrollBar>
C++ method: [constructor] void QAccessibleScrollBar::QAccessibleScrollBar(QWidget* w)

fn next_page_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleScrollBarNextPageActionArgs

C++ method: QAccessibleScrollBar::nextPageAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn next_page_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn next_page_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleScrollBar::nextPageAction()

C++ documentation:

Returns the name of the next page default action.

See also actionNames() and localizedActionName().

fn object(&self) -> *mut Object

C++ method: virtual QObject* QAccessibleScrollBar::object() const

C++ documentation:

Reimplemented from QAccessibleInterface::object().

fn parent(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleScrollBar::parent() const

C++ documentation:

Reimplemented from QAccessibleInterface::parent().

fn press_action() -> &'static String

C++ method: static const QString& QAccessibleScrollBar::pressAction()

C++ documentation:

Returns the name of the press default action.

See also actionNames() and localizedActionName().

fn previous_page_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleScrollBarPreviousPageActionArgs

C++ method: QAccessibleScrollBar::previousPageAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn previous_page_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn previous_page_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleScrollBar::previousPageAction()

C++ documentation:

Returns the name of the previous page default action.

See also actionNames() and localizedActionName().

fn rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleScrollBarRectArgs<'largs>

C++ method: QAccessibleScrollBar::rect

This is an overloaded function. Available variants:

Rust arguments:
1) fn rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: virtual QRect QAccessibleScrollBar::rect() const

C++ documentation:

Reimplemented from QAccessibleInterface::rect().

fn relations<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleScrollBarRelationsArgs<'largs>

C++ method: QAccessibleScrollBar::relations

This is an overloaded function. Available variants:

Rust arguments:
1) fn relations(&self, (qt_core::flags::Flags<qt_gui::accessible::RelationFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
2) fn relations(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
3) fn relations(&self, (qt_core::flags::Flags<qt_gui::accessible::RelationFlag>, cpp_utils::AsStruct)) -> qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
4) fn relations(&self, cpp_utils::AsStruct) -> qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
C++ method: virtual QVector<QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>> QAccessibleScrollBar::relations(QFlags<QAccessible::RelationFlag> match = ?) const

C++ documentation:

Reimplemented from QAccessibleInterface::relations().

fn role(&self) -> Role

C++ method: virtual QAccessible::Role QAccessibleScrollBar::role() const

C++ documentation:

Reimplemented from QAccessibleInterface::role().

fn scroll_down_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleScrollBarScrollDownActionArgs

C++ method: QAccessibleScrollBar::scrollDownAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_down_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_down_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleScrollBar::scrollDownAction()

C++ documentation:

Returns the name of the scroll down default action.

See also actionNames() and localizedActionName().

fn scroll_left_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleScrollBarScrollLeftActionArgs

C++ method: QAccessibleScrollBar::scrollLeftAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_left_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_left_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleScrollBar::scrollLeftAction()

C++ documentation:

Returns the name of the scroll left default action.

See also actionNames() and localizedActionName().

fn scroll_right_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleScrollBarScrollRightActionArgs

C++ method: QAccessibleScrollBar::scrollRightAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_right_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_right_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleScrollBar::scrollRightAction()

C++ documentation:

Returns the name of the scroll right default action.

See also actionNames() and localizedActionName().

fn scroll_up_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleScrollBarScrollUpActionArgs

C++ method: QAccessibleScrollBar::scrollUpAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_up_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_up_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleScrollBar::scrollUpAction()

C++ documentation:

Returns the name of the scroll up default action.

See also actionNames() and localizedActionName().

fn set_current_value(&mut self, value: &Variant)

C++ method: virtual void QAccessibleScrollBar::setCurrentValue(const QVariant& value)

fn set_focus_action() -> &'static String

C++ method: static const QString& QAccessibleScrollBar::setFocusAction()

C++ documentation:

Returns the name of the set focus default action.

See also actionNames() and localizedActionName().

fn set_text(&mut self, t: Text, text: &String)

C++ method: virtual void QAccessibleScrollBar::setText(QAccessible::Text t, const QString& text)

C++ documentation:

Reimplemented from QAccessibleInterface::setText().

fn show_menu_action() -> &'static String

C++ method: static const QString& QAccessibleScrollBar::showMenuAction()

C++ documentation:

Returns the name of the show menu default action.

See also actionNames() and localizedActionName().

fn state<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleScrollBarStateArgs<'largs>

C++ method: QAccessibleScrollBar::state

This is an overloaded function. Available variants:

Rust arguments:
1) fn state(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::accessible::State>
2) fn state(&self, cpp_utils::AsStruct) -> qt_gui::accessible::State
C++ method: virtual QAccessible::State QAccessibleScrollBar::state() const

C++ documentation:

Reimplemented from QAccessibleInterface::state().

fn table_cell_interface(&mut self) -> *mut AccessibleTableCellInterface

C++ method: QAccessibleTableCellInterface* QAccessibleScrollBar::tableCellInterface()

fn table_interface(&mut self) -> *mut AccessibleTableInterface

C++ method: QAccessibleTableInterface* QAccessibleScrollBar::tableInterface()

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleScrollBarTextArgs<'largs>

C++ method: QAccessibleScrollBar::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(&self, (qt_gui::accessible::Text, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, (qt_gui::accessible::Text, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleScrollBar::text(QAccessible::Text t) const

fn text_interface(&mut self) -> *mut AccessibleTextInterface

C++ method: QAccessibleTextInterface* QAccessibleScrollBar::textInterface()

fn toggle_action() -> &'static String

C++ method: static const QString& QAccessibleScrollBar::toggleAction()

C++ documentation:

Returns the name of the toggle default action.

See also actionNames() and localizedActionName().

fn tr<Args>(args: Args) -> Args::ReturnType where Args: AccessibleScrollBarTrArgs

C++ method: QAccessibleScrollBar::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QAccessibleScrollBar::tr(const char* sourceText, const char* disambiguation, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: AccessibleScrollBarTrUtf8Args

C++ method: QAccessibleScrollBar::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QAccessibleScrollBar::trUtf8(const char* sourceText, const char* disambiguation, int n)

fn value_interface(&mut self) -> *mut AccessibleValueInterface

C++ method: QAccessibleValueInterface* QAccessibleScrollBar::valueInterface()

fn virtual_hook(&mut self, id: c_int, data: *mut c_void)

C++ method: virtual void QAccessibleScrollBar::virtual_hook(int id, void* data)

fn window(&self) -> *mut Window

C++ method: virtual QWindow* QAccessibleScrollBar::window() const

C++ documentation:

Reimplemented from QAccessibleInterface::window().