Module qt_widgets::accessible_slider::overloading

Traits

AccessibleSliderActionNamesArgs
This trait represents a set of arguments accepted by AccessibleSlider::action_names method.

AccessibleSliderBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleSlider::background_color method.

AccessibleSliderCurrentValueArgs
This trait represents a set of arguments accepted by AccessibleSlider::current_value method.

AccessibleSliderForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleSlider::foreground_color method.

AccessibleSliderKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleSlider::key_bindings_for_action method.

AccessibleSliderLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleSlider::localized_action_description method.

AccessibleSliderLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleSlider::localized_action_name method.

AccessibleSliderMaximumValueArgs
This trait represents a set of arguments accepted by AccessibleSlider::maximum_value method.

AccessibleSliderMinimumStepSizeArgs
This trait represents a set of arguments accepted by AccessibleSlider::minimum_step_size method.

AccessibleSliderMinimumValueArgs
This trait represents a set of arguments accepted by AccessibleSlider::minimum_value method.

AccessibleSliderNewArgs
This trait represents a set of arguments accepted by AccessibleSlider::new method.

AccessibleSliderNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleSlider::next_page_action method.

AccessibleSliderPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleSlider::previous_page_action method.

AccessibleSliderRectArgs
This trait represents a set of arguments accepted by AccessibleSlider::rect method.

AccessibleSliderRelationsArgs
This trait represents a set of arguments accepted by AccessibleSlider::relations method.

AccessibleSliderScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleSlider::scroll_down_action method.

AccessibleSliderScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleSlider::scroll_left_action method.

AccessibleSliderScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleSlider::scroll_right_action method.

AccessibleSliderScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleSlider::scroll_up_action method.

AccessibleSliderStateArgs
This trait represents a set of arguments accepted by AccessibleSlider::state method.

AccessibleSliderTextArgs
This trait represents a set of arguments accepted by AccessibleSlider::text method.

AccessibleSliderTrArgs
This trait represents a set of arguments accepted by AccessibleSlider::tr method.

AccessibleSliderTrUtf8Args
This trait represents a set of arguments accepted by AccessibleSlider::tr_utf8 method.