Module qt_widgets::accessible_spin_box::overloading

Traits

AccessibleSpinBoxActionNamesArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::action_names method.

AccessibleSpinBoxAttributesArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::attributes method.

AccessibleSpinBoxBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::background_color method.

AccessibleSpinBoxCharacterRectArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::character_rect method.

AccessibleSpinBoxCurrentValueArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::current_value method.

AccessibleSpinBoxForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::foreground_color method.

AccessibleSpinBoxKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::key_bindings_for_action method.

AccessibleSpinBoxLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::localized_action_description method.

AccessibleSpinBoxLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::localized_action_name method.

AccessibleSpinBoxMaximumValueArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::maximum_value method.

AccessibleSpinBoxMinimumStepSizeArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::minimum_step_size method.

AccessibleSpinBoxMinimumValueArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::minimum_value method.

AccessibleSpinBoxNewArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::new method.

AccessibleSpinBoxNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::next_page_action method.

AccessibleSpinBoxPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::previous_page_action method.

AccessibleSpinBoxRectArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::rect method.

AccessibleSpinBoxRelationsArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::relations method.

AccessibleSpinBoxScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::scroll_down_action method.

AccessibleSpinBoxScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::scroll_left_action method.

AccessibleSpinBoxScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::scroll_right_action method.

AccessibleSpinBoxScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::scroll_up_action method.

AccessibleSpinBoxStateArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::state method.

AccessibleSpinBoxTextAfterOffsetArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::text_after_offset method.

AccessibleSpinBoxTextArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::text method.

AccessibleSpinBoxTextAtOffsetArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::text_at_offset method.

AccessibleSpinBoxTextBeforeOffsetArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::text_before_offset method.

AccessibleSpinBoxTrArgs
This trait represents a set of arguments accepted by AccessibleSpinBox::tr method.

AccessibleSpinBoxTrUtf8Args
This trait represents a set of arguments accepted by AccessibleSpinBox::tr_utf8 method.