Struct qt_widgets::accessible_spin_box::AccessibleSpinBox

pub struct AccessibleSpinBox {
    // some fields omitted
}

C++ type: QAccessibleSpinBox

Methods

impl AccessibleSpinBox

fn action_interface(&mut self) -> *mut AccessibleActionInterface

C++ method: QAccessibleActionInterface* QAccessibleSpinBox::actionInterface()

fn action_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxActionNamesArgs<'largs>

C++ method: QAccessibleSpinBox::actionNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn action_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn action_names(&self, cpp_utils::AsStruct) -> qt_core::string_list::StringList
C++ method: virtual QStringList QAccessibleSpinBox::actionNames() const

C++ documentation:

Reimplemented from QAccessibleActionInterface::actionNames().

fn add_selection(&mut self, start_offset: c_int, end_offset: c_int)

C++ method: virtual void QAccessibleSpinBox::addSelection(int startOffset, int endOffset)

fn attributes<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxAttributesArgs<'largs>

C++ method: QAccessibleSpinBox::attributes

This is an overloaded function. Available variants:

Rust arguments:
1) fn attributes(&self, (libc::c_int, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn attributes(&self, (libc::c_int, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleSpinBox::attributes(int offset, int* startOffset, int* endOffset) const

fn background_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxBackgroundColorArgs<'largs>

C++ method: QAccessibleSpinBox::backgroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn background_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn background_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: virtual QColor QAccessibleSpinBox::backgroundColor() const

C++ documentation:

Reimplemented from QAccessibleInterface::backgroundColor().

fn character_count(&self) -> c_int

C++ method: virtual int QAccessibleSpinBox::characterCount() const

fn character_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxCharacterRectArgs<'largs>

C++ method: QAccessibleSpinBox::characterRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn character_rect(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn character_rect(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::rect::Rect
C++ method: virtual QRect QAccessibleSpinBox::characterRect(int offset) const

fn child(&self, index: c_int) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleSpinBox::child(int index) const

C++ documentation:

Reimplemented from QAccessibleInterface::child().

fn child_at(&self, x: c_int, y: c_int) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleSpinBox::childAt(int x, int y) const

C++ documentation:

Reimplemented from QAccessibleInterface::childAt().

fn child_count(&self) -> c_int

C++ method: virtual int QAccessibleSpinBox::childCount() const

C++ documentation:

Reimplemented from QAccessibleInterface::childCount().

fn current_value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxCurrentValueArgs<'largs>

C++ method: QAccessibleSpinBox::currentValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn current_value(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: virtual QVariant QAccessibleSpinBox::currentValue() const

fn cursor_position(&self) -> c_int

C++ method: virtual int QAccessibleSpinBox::cursorPosition() const

fn decrease_action() -> &'static String

C++ method: static const QString& QAccessibleSpinBox::decreaseAction()

C++ documentation:

Returns the name of the decrease default action.

See also actionNames() and localizedActionName().

fn delete_text(&mut self, start_offset: c_int, end_offset: c_int)

C++ method: virtual void QAccessibleSpinBox::deleteText(int startOffset, int endOffset)

fn do_action(&mut self, action_name: &String)

C++ method: virtual void QAccessibleSpinBox::doAction(const QString& actionName)

C++ documentation:

Reimplemented from QAccessibleActionInterface::doAction().

fn editable_text_interface(&mut self) -> *mut AccessibleEditableTextInterface

C++ method: QAccessibleEditableTextInterface* QAccessibleSpinBox::editableTextInterface()

fn focus_child(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleSpinBox::focusChild() const

C++ documentation:

Reimplemented from QAccessibleInterface::focusChild().

fn foreground_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxForegroundColorArgs<'largs>

C++ method: QAccessibleSpinBox::foregroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn foreground_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn foreground_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: virtual QColor QAccessibleSpinBox::foregroundColor() const

C++ documentation:

Reimplemented from QAccessibleInterface::foregroundColor().

fn image_interface(&mut self) -> *mut AccessibleImageInterface

C++ method: QAccessibleImageInterface* QAccessibleSpinBox::imageInterface()

fn increase_action() -> &'static String

C++ method: static const QString& QAccessibleSpinBox::increaseAction()

C++ documentation:

Returns the name of the increase default action.

See also actionNames() and localizedActionName().

fn index_of_child(&self, child: *const AccessibleInterface) -> c_int

C++ method: virtual int QAccessibleSpinBox::indexOfChild(const QAccessibleInterface* child) const

C++ documentation:

Reimplemented from QAccessibleInterface::indexOfChild().

fn insert_text(&mut self, offset: c_int, text: &String)

C++ method: virtual void QAccessibleSpinBox::insertText(int offset, const QString& text)

fn interface_cast(&mut self, t: InterfaceType) -> *mut c_void

C++ method: virtual void* QAccessibleSpinBox::interface_cast(QAccessible::InterfaceType t)

fn is_valid(&self) -> bool

C++ method: virtual bool QAccessibleSpinBox::isValid() const

C++ documentation:

Reimplemented from QAccessibleInterface::isValid().

fn key_bindings_for_action<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxKeyBindingsForActionArgs<'largs>

C++ method: QAccessibleSpinBox::keyBindingsForAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn key_bindings_for_action(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn key_bindings_for_action(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string_list::StringList
C++ method: virtual QStringList QAccessibleSpinBox::keyBindingsForAction(const QString& actionName) const

C++ documentation:

Reimplemented from QAccessibleActionInterface::keyBindingsForAction().

fn localized_action_description<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxLocalizedActionDescriptionArgs<'largs>

C++ method: QAccessibleSpinBox::localizedActionDescription

This is an overloaded function. Available variants:

Rust arguments:
1) fn localized_action_description(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn localized_action_description(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleSpinBox::localizedActionDescription(const QString& name) const

C++ documentation:

Returns a localized action description of the action actionName.

When using one of the default names, you can call this function in QAccessibleActionInterface to get the localized string.

See also actionNames() and localizedActionName().

fn localized_action_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxLocalizedActionNameArgs<'largs>

C++ method: QAccessibleSpinBox::localizedActionName

This is an overloaded function. Available variants:

Rust arguments:
1) fn localized_action_name(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn localized_action_name(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleSpinBox::localizedActionName(const QString& name) const

C++ documentation:

Returns a localized action name of actionName.

For custom actions this function has to be re-implemented. When using one of the default names, you can call this function in QAccessibleActionInterface to get the localized string.

See also actionNames() and localizedActionDescription().

fn maximum_value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxMaximumValueArgs<'largs>

C++ method: QAccessibleSpinBox::maximumValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn maximum_value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn maximum_value(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: virtual QVariant QAccessibleSpinBox::maximumValue() const

fn minimum_step_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxMinimumStepSizeArgs<'largs>

C++ method: QAccessibleSpinBox::minimumStepSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn minimum_step_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn minimum_step_size(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: virtual QVariant QAccessibleSpinBox::minimumStepSize() const

fn minimum_value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxMinimumValueArgs<'largs>

C++ method: QAccessibleSpinBox::minimumValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn minimum_value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn minimum_value(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: virtual QVariant QAccessibleSpinBox::minimumValue() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxNewArgs

C++ method: QAccessibleSpinBox::QAccessibleSpinBox

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((*mut ::widget::Widget, cpp_utils::AsStruct)) -> ::accessible_spin_box::AccessibleSpinBox
2) fn new((*mut ::widget::Widget, cpp_utils::AsBox)) -> cpp_utils::CppBox<::accessible_spin_box::AccessibleSpinBox>
C++ method: [constructor] void QAccessibleSpinBox::QAccessibleSpinBox(QWidget* w)

fn next_page_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxNextPageActionArgs

C++ method: QAccessibleSpinBox::nextPageAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn next_page_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn next_page_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleSpinBox::nextPageAction()

C++ documentation:

Returns the name of the next page default action.

See also actionNames() and localizedActionName().

fn object(&self) -> *mut Object

C++ method: virtual QObject* QAccessibleSpinBox::object() const

C++ documentation:

Reimplemented from QAccessibleInterface::object().

fn offset_at_point(&self, point: &Point) -> c_int

C++ method: virtual int QAccessibleSpinBox::offsetAtPoint(const QPoint& point) const

fn parent(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleSpinBox::parent() const

C++ documentation:

Reimplemented from QAccessibleInterface::parent().

fn press_action() -> &'static String

C++ method: static const QString& QAccessibleSpinBox::pressAction()

C++ documentation:

Returns the name of the press default action.

See also actionNames() and localizedActionName().

fn previous_page_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxPreviousPageActionArgs

C++ method: QAccessibleSpinBox::previousPageAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn previous_page_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn previous_page_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleSpinBox::previousPageAction()

C++ documentation:

Returns the name of the previous page default action.

See also actionNames() and localizedActionName().

fn rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxRectArgs<'largs>

C++ method: QAccessibleSpinBox::rect

This is an overloaded function. Available variants:

Rust arguments:
1) fn rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: virtual QRect QAccessibleSpinBox::rect() const

C++ documentation:

Reimplemented from QAccessibleInterface::rect().

fn relations<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxRelationsArgs<'largs>

C++ method: QAccessibleSpinBox::relations

This is an overloaded function. Available variants:

Rust arguments:
1) fn relations(&self, (qt_core::flags::Flags<qt_gui::accessible::RelationFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
2) fn relations(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
3) fn relations(&self, (qt_core::flags::Flags<qt_gui::accessible::RelationFlag>, cpp_utils::AsStruct)) -> qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
4) fn relations(&self, cpp_utils::AsStruct) -> qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
C++ method: virtual QVector<QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>> QAccessibleSpinBox::relations(QFlags<QAccessible::RelationFlag> match = ?) const

C++ documentation:

Reimplemented from QAccessibleInterface::relations().

fn remove_selection(&mut self, selection_index: c_int)

C++ method: virtual void QAccessibleSpinBox::removeSelection(int selectionIndex)

fn replace_text(&mut self, start_offset: c_int, end_offset: c_int, text: &String)

C++ method: virtual void QAccessibleSpinBox::replaceText(int startOffset, int endOffset, const QString& text)

fn role(&self) -> Role

C++ method: virtual QAccessible::Role QAccessibleSpinBox::role() const

C++ documentation:

Reimplemented from QAccessibleInterface::role().

fn scroll_down_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxScrollDownActionArgs

C++ method: QAccessibleSpinBox::scrollDownAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_down_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_down_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleSpinBox::scrollDownAction()

C++ documentation:

Returns the name of the scroll down default action.

See also actionNames() and localizedActionName().

fn scroll_left_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxScrollLeftActionArgs

C++ method: QAccessibleSpinBox::scrollLeftAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_left_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_left_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleSpinBox::scrollLeftAction()

C++ documentation:

Returns the name of the scroll left default action.

See also actionNames() and localizedActionName().

fn scroll_right_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxScrollRightActionArgs

C++ method: QAccessibleSpinBox::scrollRightAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_right_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_right_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleSpinBox::scrollRightAction()

C++ documentation:

Returns the name of the scroll right default action.

See also actionNames() and localizedActionName().

fn scroll_to_substring(&mut self, start_index: c_int, end_index: c_int)

C++ method: virtual void QAccessibleSpinBox::scrollToSubstring(int startIndex, int endIndex)

fn scroll_up_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxScrollUpActionArgs

C++ method: QAccessibleSpinBox::scrollUpAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_up_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_up_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleSpinBox::scrollUpAction()

C++ documentation:

Returns the name of the scroll up default action.

See also actionNames() and localizedActionName().

fn selection(&self, selection_index: c_int, start_offset: *mut c_int, end_offset: *mut c_int)

C++ method: virtual void QAccessibleSpinBox::selection(int selectionIndex, int* startOffset, int* endOffset) const

fn selection_count(&self) -> c_int

C++ method: virtual int QAccessibleSpinBox::selectionCount() const

fn set_current_value(&mut self, value: &Variant)

C++ method: virtual void QAccessibleSpinBox::setCurrentValue(const QVariant& value)

fn set_cursor_position(&mut self, position: c_int)

C++ method: virtual void QAccessibleSpinBox::setCursorPosition(int position)

fn set_focus_action() -> &'static String

C++ method: static const QString& QAccessibleSpinBox::setFocusAction()

C++ documentation:

Returns the name of the set focus default action.

See also actionNames() and localizedActionName().

fn set_selection(&mut self, selection_index: c_int, start_offset: c_int, end_offset: c_int)

C++ method: virtual void QAccessibleSpinBox::setSelection(int selectionIndex, int startOffset, int endOffset)

fn set_text(&mut self, t: Text, text: &String)

C++ method: virtual void QAccessibleSpinBox::setText(QAccessible::Text t, const QString& text)

C++ documentation:

Reimplemented from QAccessibleInterface::setText().

fn show_menu_action() -> &'static String

C++ method: static const QString& QAccessibleSpinBox::showMenuAction()

C++ documentation:

Returns the name of the show menu default action.

See also actionNames() and localizedActionName().

fn state<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxStateArgs<'largs>

C++ method: QAccessibleSpinBox::state

This is an overloaded function. Available variants:

Rust arguments:
1) fn state(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::accessible::State>
2) fn state(&self, cpp_utils::AsStruct) -> qt_gui::accessible::State
C++ method: virtual QAccessible::State QAccessibleSpinBox::state() const

C++ documentation:

Reimplemented from QAccessibleInterface::state().

fn table_cell_interface(&mut self) -> *mut AccessibleTableCellInterface

C++ method: QAccessibleTableCellInterface* QAccessibleSpinBox::tableCellInterface()

fn table_interface(&mut self) -> *mut AccessibleTableInterface

C++ method: QAccessibleTableInterface* QAccessibleSpinBox::tableInterface()

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxTextArgs<'largs>

C++ method: QAccessibleSpinBox::text

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn text(&self, (qt_gui::accessible::Text, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, (qt_gui::accessible::Text, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleSpinBox::text(QAccessible::Text t) const

Variant 2

Rust arguments:
1) fn text(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleSpinBox::text(int startOffset, int endOffset) const

fn text_after_offset<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxTextAfterOffsetArgs<'largs>

C++ method: QAccessibleSpinBox::textAfterOffset

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_after_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text_after_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleSpinBox::textAfterOffset(int offset, QAccessible::TextBoundaryType boundaryType, int* startOffset, int* endOffset) const

fn text_at_offset<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxTextAtOffsetArgs<'largs>

C++ method: QAccessibleSpinBox::textAtOffset

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_at_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text_at_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleSpinBox::textAtOffset(int offset, QAccessible::TextBoundaryType boundaryType, int* startOffset, int* endOffset) const

fn text_before_offset<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxTextBeforeOffsetArgs<'largs>

C++ method: QAccessibleSpinBox::textBeforeOffset

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_before_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text_before_offset(&self, (libc::c_int, qt_gui::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleSpinBox::textBeforeOffset(int offset, QAccessible::TextBoundaryType boundaryType, int* endOffset, int* startOffset) const

fn text_interface(&mut self) -> *mut AccessibleTextInterface

C++ method: QAccessibleTextInterface* QAccessibleSpinBox::textInterface()

fn toggle_action() -> &'static String

C++ method: static const QString& QAccessibleSpinBox::toggleAction()

C++ documentation:

Returns the name of the toggle default action.

See also actionNames() and localizedActionName().

fn tr<Args>(args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxTrArgs

C++ method: QAccessibleSpinBox::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QAccessibleSpinBox::tr(const char* sourceText, const char* disambiguation, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: AccessibleSpinBoxTrUtf8Args

C++ method: QAccessibleSpinBox::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QAccessibleSpinBox::trUtf8(const char* sourceText, const char* disambiguation, int n)

fn value_interface(&mut self) -> *mut AccessibleValueInterface

C++ method: QAccessibleValueInterface* QAccessibleSpinBox::valueInterface()

fn virtual_hook(&mut self, id: c_int, data: *mut c_void)

C++ method: virtual void QAccessibleSpinBox::virtual_hook(int id, void* data)

fn window(&self) -> *mut Window

C++ method: virtual QWindow* QAccessibleSpinBox::window() const

C++ documentation:

Reimplemented from QAccessibleInterface::window().