Module qt_widgets::accessible_stacked_widget::overloading

Traits

AccessibleStackedWidgetActionNamesArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::action_names method.

AccessibleStackedWidgetBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::background_color method.

AccessibleStackedWidgetForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::foreground_color method.

AccessibleStackedWidgetKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::key_bindings_for_action method.

AccessibleStackedWidgetLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::localized_action_description method.

AccessibleStackedWidgetLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::localized_action_name method.

AccessibleStackedWidgetNewArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::new method.

AccessibleStackedWidgetNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::next_page_action method.

AccessibleStackedWidgetPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::previous_page_action method.

AccessibleStackedWidgetRectArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::rect method.

AccessibleStackedWidgetRelationsArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::relations method.

AccessibleStackedWidgetScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::scroll_down_action method.

AccessibleStackedWidgetScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::scroll_left_action method.

AccessibleStackedWidgetScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::scroll_right_action method.

AccessibleStackedWidgetScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::scroll_up_action method.

AccessibleStackedWidgetStateArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::state method.

AccessibleStackedWidgetTextArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::text method.

AccessibleStackedWidgetTrArgs
This trait represents a set of arguments accepted by AccessibleStackedWidget::tr method.

AccessibleStackedWidgetTrUtf8Args
This trait represents a set of arguments accepted by AccessibleStackedWidget::tr_utf8 method.