Module qt_widgets::accessible_tab_bar::overloading

Traits

AccessibleTabBarActionNamesArgs
This trait represents a set of arguments accepted by AccessibleTabBar::action_names method.

AccessibleTabBarBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleTabBar::background_color method.

AccessibleTabBarForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleTabBar::foreground_color method.

AccessibleTabBarKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleTabBar::key_bindings_for_action method.

AccessibleTabBarLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleTabBar::localized_action_description method.

AccessibleTabBarLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleTabBar::localized_action_name method.

AccessibleTabBarNewArgs
This trait represents a set of arguments accepted by AccessibleTabBar::new method.

AccessibleTabBarNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleTabBar::next_page_action method.

AccessibleTabBarPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleTabBar::previous_page_action method.

AccessibleTabBarRectArgs
This trait represents a set of arguments accepted by AccessibleTabBar::rect method.

AccessibleTabBarRelationsArgs
This trait represents a set of arguments accepted by AccessibleTabBar::relations method.

AccessibleTabBarScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleTabBar::scroll_down_action method.

AccessibleTabBarScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleTabBar::scroll_left_action method.

AccessibleTabBarScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleTabBar::scroll_right_action method.

AccessibleTabBarScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleTabBar::scroll_up_action method.

AccessibleTabBarStateArgs
This trait represents a set of arguments accepted by AccessibleTabBar::state method.

AccessibleTabBarTextArgs
This trait represents a set of arguments accepted by AccessibleTabBar::text method.

AccessibleTabBarTrArgs
This trait represents a set of arguments accepted by AccessibleTabBar::tr method.

AccessibleTabBarTrUtf8Args
This trait represents a set of arguments accepted by AccessibleTabBar::tr_utf8 method.