Module qt_widgets::accessible_table::overloading

Traits

AccessibleTableBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleTable::background_color method.

AccessibleTableColumnDescriptionArgs
This trait represents a set of arguments accepted by AccessibleTable::column_description method.

AccessibleTableForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleTable::foreground_color method.

AccessibleTableRectArgs
This trait represents a set of arguments accepted by AccessibleTable::rect method.

AccessibleTableRelationsArgs
This trait represents a set of arguments accepted by AccessibleTable::relations method.

AccessibleTableRowDescriptionArgs
This trait represents a set of arguments accepted by AccessibleTable::row_description method.

AccessibleTableSelectedCellsArgs
This trait represents a set of arguments accepted by AccessibleTable::selected_cells method.

AccessibleTableSelectedColumnsArgs
This trait represents a set of arguments accepted by AccessibleTable::selected_columns method.

AccessibleTableSelectedRowsArgs
This trait represents a set of arguments accepted by AccessibleTable::selected_rows method.

AccessibleTableStateArgs
This trait represents a set of arguments accepted by AccessibleTable::state method.

AccessibleTableTextArgs
This trait represents a set of arguments accepted by AccessibleTable::text method.