Struct qt_widgets::accessible_table::AccessibleTable

pub struct AccessibleTable {
    // some fields omitted
}

C++ type: QAccessibleTable

Methods

impl AccessibleTable

fn action_interface(&mut self) -> *mut AccessibleActionInterface

C++ method: QAccessibleActionInterface* QAccessibleTable::actionInterface()

fn background_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableBackgroundColorArgs<'largs>

C++ method: QAccessibleTable::backgroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn background_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn background_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: virtual QColor QAccessibleTable::backgroundColor() const

C++ documentation:

Returns the accessible's background color if applicable or an invalid QColor.

See also foregroundColor().

fn caption(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleTable::caption() const

fn cell_at(&self, row: c_int, column: c_int) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleTable::cellAt(int row, int column) const

fn child(&self, index: c_int) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleTable::child(int index) const

fn child_at(&self, x: c_int, y: c_int) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleTable::childAt(int x, int y) const

fn child_count(&self) -> c_int

C++ method: virtual int QAccessibleTable::childCount() const

fn column_count(&self) -> c_int

C++ method: virtual int QAccessibleTable::columnCount() const

fn column_description<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableColumnDescriptionArgs<'largs>

C++ method: QAccessibleTable::columnDescription

This is an overloaded function. Available variants:

Rust arguments:
1) fn column_description(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn column_description(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleTable::columnDescription(int column) const

fn editable_text_interface(&mut self) -> *mut AccessibleEditableTextInterface

C++ method: QAccessibleEditableTextInterface* QAccessibleTable::editableTextInterface()

fn focus_child(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleTable::focusChild() const

C++ documentation:

Returns the object that has the keyboard focus.

The object returned can be any descendant, including itself.

fn foreground_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableForegroundColorArgs<'largs>

C++ method: QAccessibleTable::foregroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn foreground_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn foreground_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: virtual QColor QAccessibleTable::foregroundColor() const

C++ documentation:

Returns the accessible's foreground color if applicable or an invalid QColor.

See also backgroundColor().

fn image_interface(&mut self) -> *mut AccessibleImageInterface

C++ method: QAccessibleImageInterface* QAccessibleTable::imageInterface()

fn index_of_child(&self, arg1: *const AccessibleInterface) -> c_int

C++ method: virtual int QAccessibleTable::indexOfChild(const QAccessibleInterface* arg1) const

fn interface_cast(&mut self, t: InterfaceType) -> *mut c_void

C++ method: virtual void* QAccessibleTable::interface_cast(QAccessible::InterfaceType t)

fn is_column_selected(&self, column: c_int) -> bool

C++ method: virtual bool QAccessibleTable::isColumnSelected(int column) const

fn is_row_selected(&self, row: c_int) -> bool

C++ method: virtual bool QAccessibleTable::isRowSelected(int row) const

fn is_valid(&self) -> bool

C++ method: virtual bool QAccessibleTable::isValid() const

fn model_change(&mut self, event: *mut AccessibleTableModelChangeEvent)

C++ method: virtual void QAccessibleTable::modelChange(QAccessibleTableModelChangeEvent* event)

fn object(&self) -> *mut Object

C++ method: virtual QObject* QAccessibleTable::object() const

C++ documentation:

Reimplemented from QAccessibleInterface::object().

fn parent(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleTable::parent() const

fn rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableRectArgs<'largs>

C++ method: QAccessibleTable::rect

This is an overloaded function. Available variants:

Rust arguments:
1) fn rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: virtual QRect QAccessibleTable::rect() const

fn relations<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableRelationsArgs<'largs>

C++ method: QAccessibleTable::relations

This is an overloaded function. Available variants:

Rust arguments:
1) fn relations(&self, (qt_core::flags::Flags<qt_gui::accessible::RelationFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
2) fn relations(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
3) fn relations(&self, (qt_core::flags::Flags<qt_gui::accessible::RelationFlag>, cpp_utils::AsStruct)) -> qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
4) fn relations(&self, cpp_utils::AsStruct) -> qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
C++ method: virtual QVector<QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>> QAccessibleTable::relations(QFlags<QAccessible::RelationFlag> match = ?) const

C++ documentation:

Returns the meaningful relations to other widgets. Usually this will not return parent/child relations, unless they are handled in a specific way such as in tree views. It will typically return the labelled-by and label relations.

It is possible to filter the relations by using match. It should never return itself.

See also parent() and child().

fn role(&self) -> Role

C++ method: virtual QAccessible::Role QAccessibleTable::role() const

fn row_count(&self) -> c_int

C++ method: virtual int QAccessibleTable::rowCount() const

fn row_description<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableRowDescriptionArgs<'largs>

C++ method: QAccessibleTable::rowDescription

This is an overloaded function. Available variants:

Rust arguments:
1) fn row_description(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn row_description(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleTable::rowDescription(int row) const

fn select_column(&mut self, column: c_int) -> bool

C++ method: virtual bool QAccessibleTable::selectColumn(int column)

fn select_row(&mut self, row: c_int) -> bool

C++ method: virtual bool QAccessibleTable::selectRow(int row)

fn selected_cell_count(&self) -> c_int

C++ method: virtual int QAccessibleTable::selectedCellCount() const

fn selected_cells<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableSelectedCellsArgs<'largs>

C++ method: QAccessibleTable::selectedCells

This is an overloaded function. Available variants:

Rust arguments:
1) fn selected_cells(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::list::ListAccessibleInterfaceMutPtr>
2) fn selected_cells(&self, cpp_utils::AsStruct) -> qt_gui::list::ListAccessibleInterfaceMutPtr
C++ method: virtual QList<QAccessibleInterface*> QAccessibleTable::selectedCells() const

fn selected_column_count(&self) -> c_int

C++ method: virtual int QAccessibleTable::selectedColumnCount() const

fn selected_columns<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableSelectedColumnsArgs<'largs>

C++ method: QAccessibleTable::selectedColumns

This is an overloaded function. Available variants:

Rust arguments:
1) fn selected_columns(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListCInt>
2) fn selected_columns(&self, cpp_utils::AsStruct) -> qt_core::list::ListCInt
C++ method: virtual QList<int> QAccessibleTable::selectedColumns() const

fn selected_row_count(&self) -> c_int

C++ method: virtual int QAccessibleTable::selectedRowCount() const

fn selected_rows<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableSelectedRowsArgs<'largs>

C++ method: QAccessibleTable::selectedRows

This is an overloaded function. Available variants:

Rust arguments:
1) fn selected_rows(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListCInt>
2) fn selected_rows(&self, cpp_utils::AsStruct) -> qt_core::list::ListCInt
C++ method: virtual QList<int> QAccessibleTable::selectedRows() const

fn set_text(&mut self, t: Text, text: &String)

C++ method: virtual void QAccessibleTable::setText(QAccessible::Text t, const QString& text)

C++ documentation:

Reimplemented from QAccessibleInterface::setText().

fn state<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableStateArgs<'largs>

C++ method: QAccessibleTable::state

This is an overloaded function. Available variants:

Rust arguments:
1) fn state(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::accessible::State>
2) fn state(&self, cpp_utils::AsStruct) -> qt_gui::accessible::State
C++ method: virtual QAccessible::State QAccessibleTable::state() const

fn summary(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleTable::summary() const

fn table_cell_interface(&mut self) -> *mut AccessibleTableCellInterface

C++ method: QAccessibleTableCellInterface* QAccessibleTable::tableCellInterface()

fn table_interface(&mut self) -> *mut AccessibleTableInterface

C++ method: QAccessibleTableInterface* QAccessibleTable::tableInterface()

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableTextArgs<'largs>

C++ method: QAccessibleTable::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(&self, (qt_gui::accessible::Text, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, (qt_gui::accessible::Text, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleTable::text(QAccessible::Text t) const

fn text_interface(&mut self) -> *mut AccessibleTextInterface

C++ method: QAccessibleTextInterface* QAccessibleTable::textInterface()

fn unselect_column(&mut self, column: c_int) -> bool

C++ method: virtual bool QAccessibleTable::unselectColumn(int column)

fn unselect_row(&mut self, row: c_int) -> bool

C++ method: virtual bool QAccessibleTable::unselectRow(int row)

fn value_interface(&mut self) -> *mut AccessibleValueInterface

C++ method: QAccessibleValueInterface* QAccessibleTable::valueInterface()

fn view(&self) -> *mut AbstractItemView

C++ method: QAbstractItemView* QAccessibleTable::view() const

fn virtual_hook(&mut self, id: c_int, data: *mut c_void)

C++ method: virtual void QAccessibleTable::virtual_hook(int id, void* data)

fn window(&self) -> *mut Window

C++ method: virtual QWindow* QAccessibleTable::window() const

C++ documentation:

Returns the window associated with the underlying object. For instance, QAccessibleWidget reimplements this and returns the windowHandle() of the QWidget.

It is used on some platforms to be able to notify the AT client about state changes. The backend will traverse up all ancestors until it finds a window. (This means that at least one interface among the ancestors should return a valid QWindow pointer).

The default implementation returns 0.