Module qt_widgets::accessible_table_cell::overloading

Traits

AccessibleTableCellActionNamesArgs
This trait represents a set of arguments accepted by AccessibleTableCell::action_names method.

AccessibleTableCellBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleTableCell::background_color method.

AccessibleTableCellColumnHeaderCellsArgs
This trait represents a set of arguments accepted by AccessibleTableCell::column_header_cells method.

AccessibleTableCellForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleTableCell::foreground_color method.

AccessibleTableCellKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleTableCell::key_bindings_for_action method.

AccessibleTableCellLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleTableCell::localized_action_description method.

AccessibleTableCellLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleTableCell::localized_action_name method.

AccessibleTableCellNewArgs
This trait represents a set of arguments accepted by AccessibleTableCell::new method.

AccessibleTableCellNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleTableCell::next_page_action method.

AccessibleTableCellPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleTableCell::previous_page_action method.

AccessibleTableCellRectArgs
This trait represents a set of arguments accepted by AccessibleTableCell::rect method.

AccessibleTableCellRelationsArgs
This trait represents a set of arguments accepted by AccessibleTableCell::relations method.

AccessibleTableCellRowHeaderCellsArgs
This trait represents a set of arguments accepted by AccessibleTableCell::row_header_cells method.

AccessibleTableCellScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleTableCell::scroll_down_action method.

AccessibleTableCellScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleTableCell::scroll_left_action method.

AccessibleTableCellScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleTableCell::scroll_right_action method.

AccessibleTableCellScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleTableCell::scroll_up_action method.

AccessibleTableCellStateArgs
This trait represents a set of arguments accepted by AccessibleTableCell::state method.

AccessibleTableCellTextArgs
This trait represents a set of arguments accepted by AccessibleTableCell::text method.

AccessibleTableCellTrArgs
This trait represents a set of arguments accepted by AccessibleTableCell::tr method.

AccessibleTableCellTrUtf8Args
This trait represents a set of arguments accepted by AccessibleTableCell::tr_utf8 method.