Module qt_widgets::accessible_table_corner_button::overloading

Traits

AccessibleTableCornerButtonBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleTableCornerButton::background_color method.

AccessibleTableCornerButtonForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleTableCornerButton::foreground_color method.

AccessibleTableCornerButtonNewArgs
This trait represents a set of arguments accepted by AccessibleTableCornerButton::new method.

AccessibleTableCornerButtonRectArgs
This trait represents a set of arguments accepted by AccessibleTableCornerButton::rect method.

AccessibleTableCornerButtonRelationsArgs
This trait represents a set of arguments accepted by AccessibleTableCornerButton::relations method.

AccessibleTableCornerButtonStateArgs
This trait represents a set of arguments accepted by AccessibleTableCornerButton::state method.

AccessibleTableCornerButtonTextArgs
This trait represents a set of arguments accepted by AccessibleTableCornerButton::text method.