Struct qt_widgets::accessible_table_corner_button::AccessibleTableCornerButton

pub struct AccessibleTableCornerButton {
    // some fields omitted
}

C++ type: QAccessibleTableCornerButton

Methods

impl AccessibleTableCornerButton

fn action_interface(&mut self) -> *mut AccessibleActionInterface

C++ method: QAccessibleActionInterface* QAccessibleTableCornerButton::actionInterface()

fn background_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableCornerButtonBackgroundColorArgs<'largs>

C++ method: QAccessibleTableCornerButton::backgroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn background_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn background_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: virtual QColor QAccessibleTableCornerButton::backgroundColor() const

C++ documentation:

Returns the accessible's background color if applicable or an invalid QColor.

See also foregroundColor().

fn child(&self, arg1: c_int) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleTableCornerButton::child(int arg1) const

fn child_at(&self, arg1: c_int, arg2: c_int) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleTableCornerButton::childAt(int arg1, int arg2) const

fn child_count(&self) -> c_int

C++ method: virtual int QAccessibleTableCornerButton::childCount() const

fn editable_text_interface(&mut self) -> *mut AccessibleEditableTextInterface

C++ method: QAccessibleEditableTextInterface* QAccessibleTableCornerButton::editableTextInterface()

fn focus_child(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleTableCornerButton::focusChild() const

C++ documentation:

Returns the object that has the keyboard focus.

The object returned can be any descendant, including itself.

fn foreground_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableCornerButtonForegroundColorArgs<'largs>

C++ method: QAccessibleTableCornerButton::foregroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn foreground_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn foreground_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: virtual QColor QAccessibleTableCornerButton::foregroundColor() const

C++ documentation:

Returns the accessible's foreground color if applicable or an invalid QColor.

See also backgroundColor().

fn image_interface(&mut self) -> *mut AccessibleImageInterface

C++ method: QAccessibleImageInterface* QAccessibleTableCornerButton::imageInterface()

fn index_of_child(&self, arg1: *const AccessibleInterface) -> c_int

C++ method: virtual int QAccessibleTableCornerButton::indexOfChild(const QAccessibleInterface* arg1) const

fn interface_cast(&mut self, arg1: InterfaceType) -> *mut c_void

C++ method: virtual void* QAccessibleTableCornerButton::interface_cast(QAccessible::InterfaceType arg1)

C++ documentation:

Returns a specialized accessibility interface type from the generic QAccessibleInterface.

This function must be reimplemented when providing more information about a widget or object through the specialized interfaces. For example a line edit should implement the QAccessibleTextInterface.

Qt's QLineEdit for example has its accessibility support implemented in QAccessibleLineEdit.

void QAccessibleLineEdit::interface_cast(QAccessible::InterfaceType t)
{
    if (t == QAccessible::TextInterface)
        return static_cast<QAccessibleTextInterface>(this);
    return QAccessibleWidget::interface_cast(t);
}

See also QAccessible::InterfaceType, QAccessibleTextInterface, QAccessibleValueInterface, QAccessibleActionInterface, QAccessibleTableInterface, and QAccessibleTableCellInterface.

fn is_valid(&self) -> bool

C++ method: virtual bool QAccessibleTableCornerButton::isValid() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: AccessibleTableCornerButtonNewArgs

C++ method: QAccessibleTableCornerButton::QAccessibleTableCornerButton

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((*mut ::abstract_item_view::AbstractItemView, cpp_utils::AsStruct)) -> ::accessible_table_corner_button::AccessibleTableCornerButton
2) fn new((*mut ::abstract_item_view::AbstractItemView, cpp_utils::AsBox)) -> cpp_utils::CppBox<::accessible_table_corner_button::AccessibleTableCornerButton>
C++ method: [constructor] void QAccessibleTableCornerButton::QAccessibleTableCornerButton(QAbstractItemView* view_)

fn object(&self) -> *mut Object

C++ method: virtual QObject* QAccessibleTableCornerButton::object() const

fn parent(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleTableCornerButton::parent() const

fn rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableCornerButtonRectArgs<'largs>

C++ method: QAccessibleTableCornerButton::rect

This is an overloaded function. Available variants:

Rust arguments:
1) fn rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: virtual QRect QAccessibleTableCornerButton::rect() const

fn relations<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableCornerButtonRelationsArgs<'largs>

C++ method: QAccessibleTableCornerButton::relations

This is an overloaded function. Available variants:

Rust arguments:
1) fn relations(&self, (qt_core::flags::Flags<qt_gui::accessible::RelationFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
2) fn relations(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
3) fn relations(&self, (qt_core::flags::Flags<qt_gui::accessible::RelationFlag>, cpp_utils::AsStruct)) -> qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
4) fn relations(&self, cpp_utils::AsStruct) -> qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
C++ method: virtual QVector<QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>> QAccessibleTableCornerButton::relations(QFlags<QAccessible::RelationFlag> match = ?) const

C++ documentation:

Returns the meaningful relations to other widgets. Usually this will not return parent/child relations, unless they are handled in a specific way such as in tree views. It will typically return the labelled-by and label relations.

It is possible to filter the relations by using match. It should never return itself.

See also parent() and child().

fn role(&self) -> Role

C++ method: virtual QAccessible::Role QAccessibleTableCornerButton::role() const

fn set_text(&mut self, arg1: Text, arg2: &String)

C++ method: virtual void QAccessibleTableCornerButton::setText(QAccessible::Text arg1, const QString& arg2)

fn state<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableCornerButtonStateArgs<'largs>

C++ method: QAccessibleTableCornerButton::state

This is an overloaded function. Available variants:

Rust arguments:
1) fn state(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::accessible::State>
2) fn state(&self, cpp_utils::AsStruct) -> qt_gui::accessible::State
C++ method: virtual QAccessible::State QAccessibleTableCornerButton::state() const

fn table_cell_interface(&mut self) -> *mut AccessibleTableCellInterface

C++ method: QAccessibleTableCellInterface* QAccessibleTableCornerButton::tableCellInterface()

fn table_interface(&mut self) -> *mut AccessibleTableInterface

C++ method: QAccessibleTableInterface* QAccessibleTableCornerButton::tableInterface()

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableCornerButtonTextArgs<'largs>

C++ method: QAccessibleTableCornerButton::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(&self, (qt_gui::accessible::Text, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, (qt_gui::accessible::Text, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleTableCornerButton::text(QAccessible::Text arg1) const

fn text_interface(&mut self) -> *mut AccessibleTextInterface

C++ method: QAccessibleTextInterface* QAccessibleTableCornerButton::textInterface()

fn value_interface(&mut self) -> *mut AccessibleValueInterface

C++ method: QAccessibleValueInterface* QAccessibleTableCornerButton::valueInterface()

fn virtual_hook(&mut self, id: c_int, data: *mut c_void)

C++ method: virtual void QAccessibleTableCornerButton::virtual_hook(int id, void* data)

fn window(&self) -> *mut Window

C++ method: virtual QWindow* QAccessibleTableCornerButton::window() const

C++ documentation:

Returns the window associated with the underlying object. For instance, QAccessibleWidget reimplements this and returns the windowHandle() of the QWidget.

It is used on some platforms to be able to notify the AT client about state changes. The backend will traverse up all ancestors until it finds a window. (This means that at least one interface among the ancestors should return a valid QWindow pointer).

The default implementation returns 0.