Module qt_widgets::accessible_text_browser::overloading

Traits

AccessibleTextBrowserActionNamesArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::action_names method.

AccessibleTextBrowserAttributesArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::attributes method.

AccessibleTextBrowserBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::background_color method.

AccessibleTextBrowserCharacterRectArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::character_rect method.

AccessibleTextBrowserForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::foreground_color method.

AccessibleTextBrowserKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::key_bindings_for_action method.

AccessibleTextBrowserLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::localized_action_description method.

AccessibleTextBrowserLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::localized_action_name method.

AccessibleTextBrowserNewArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::new method.

AccessibleTextBrowserNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::next_page_action method.

AccessibleTextBrowserPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::previous_page_action method.

AccessibleTextBrowserRectArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::rect method.

AccessibleTextBrowserRelationsArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::relations method.

AccessibleTextBrowserScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::scroll_down_action method.

AccessibleTextBrowserScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::scroll_left_action method.

AccessibleTextBrowserScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::scroll_right_action method.

AccessibleTextBrowserScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::scroll_up_action method.

AccessibleTextBrowserStateArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::state method.

AccessibleTextBrowserTextAfterOffsetArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::text_after_offset method.

AccessibleTextBrowserTextArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::text method.

AccessibleTextBrowserTextAtOffsetArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::text_at_offset method.

AccessibleTextBrowserTextBeforeOffsetArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::text_before_offset method.

AccessibleTextBrowserTrArgs
This trait represents a set of arguments accepted by AccessibleTextBrowser::tr method.

AccessibleTextBrowserTrUtf8Args
This trait represents a set of arguments accepted by AccessibleTextBrowser::tr_utf8 method.