Module qt_widgets::accessible_text_widget::overloading

Traits

AccessibleTextWidgetActionNamesArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::action_names method.

AccessibleTextWidgetAttributesArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::attributes method.

AccessibleTextWidgetBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::background_color method.

AccessibleTextWidgetCharacterRectArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::character_rect method.

AccessibleTextWidgetForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::foreground_color method.

AccessibleTextWidgetKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::key_bindings_for_action method.

AccessibleTextWidgetLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::localized_action_description method.

AccessibleTextWidgetLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::localized_action_name method.

AccessibleTextWidgetNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::next_page_action method.

AccessibleTextWidgetPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::previous_page_action method.

AccessibleTextWidgetRectArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::rect method.

AccessibleTextWidgetRelationsArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::relations method.

AccessibleTextWidgetScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::scroll_down_action method.

AccessibleTextWidgetScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::scroll_left_action method.

AccessibleTextWidgetScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::scroll_right_action method.

AccessibleTextWidgetScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::scroll_up_action method.

AccessibleTextWidgetStateArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::state method.

AccessibleTextWidgetTextAfterOffsetArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::text_after_offset method.

AccessibleTextWidgetTextArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::text method.

AccessibleTextWidgetTextAtOffsetArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::text_at_offset method.

AccessibleTextWidgetTextBeforeOffsetArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::text_before_offset method.

AccessibleTextWidgetTrArgs
This trait represents a set of arguments accepted by AccessibleTextWidget::tr method.

AccessibleTextWidgetTrUtf8Args
This trait represents a set of arguments accepted by AccessibleTextWidget::tr_utf8 method.