Module qt_widgets::accessible_tool_box::overloading

Traits

AccessibleToolBoxActionNamesArgs
This trait represents a set of arguments accepted by AccessibleToolBox::action_names method.

AccessibleToolBoxBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleToolBox::background_color method.

AccessibleToolBoxForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleToolBox::foreground_color method.

AccessibleToolBoxKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleToolBox::key_bindings_for_action method.

AccessibleToolBoxLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleToolBox::localized_action_description method.

AccessibleToolBoxLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleToolBox::localized_action_name method.

AccessibleToolBoxNewArgs
This trait represents a set of arguments accepted by AccessibleToolBox::new method.

AccessibleToolBoxNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleToolBox::next_page_action method.

AccessibleToolBoxPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleToolBox::previous_page_action method.

AccessibleToolBoxRectArgs
This trait represents a set of arguments accepted by AccessibleToolBox::rect method.

AccessibleToolBoxRelationsArgs
This trait represents a set of arguments accepted by AccessibleToolBox::relations method.

AccessibleToolBoxScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleToolBox::scroll_down_action method.

AccessibleToolBoxScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleToolBox::scroll_left_action method.

AccessibleToolBoxScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleToolBox::scroll_right_action method.

AccessibleToolBoxScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleToolBox::scroll_up_action method.

AccessibleToolBoxStateArgs
This trait represents a set of arguments accepted by AccessibleToolBox::state method.

AccessibleToolBoxTextArgs
This trait represents a set of arguments accepted by AccessibleToolBox::text method.

AccessibleToolBoxTrArgs
This trait represents a set of arguments accepted by AccessibleToolBox::tr method.

AccessibleToolBoxTrUtf8Args
This trait represents a set of arguments accepted by AccessibleToolBox::tr_utf8 method.