Module qt_widgets::accessible_tool_button::overloading

Traits

AccessibleToolButtonActionNamesArgs
This trait represents a set of arguments accepted by AccessibleToolButton::action_names method.

AccessibleToolButtonBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleToolButton::background_color method.

AccessibleToolButtonForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleToolButton::foreground_color method.

AccessibleToolButtonKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleToolButton::key_bindings_for_action method.

AccessibleToolButtonLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleToolButton::localized_action_description method.

AccessibleToolButtonLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleToolButton::localized_action_name method.

AccessibleToolButtonNewArgs
This trait represents a set of arguments accepted by AccessibleToolButton::new method.

AccessibleToolButtonNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleToolButton::next_page_action method.

AccessibleToolButtonPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleToolButton::previous_page_action method.

AccessibleToolButtonRectArgs
This trait represents a set of arguments accepted by AccessibleToolButton::rect method.

AccessibleToolButtonRelationsArgs
This trait represents a set of arguments accepted by AccessibleToolButton::relations method.

AccessibleToolButtonScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleToolButton::scroll_down_action method.

AccessibleToolButtonScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleToolButton::scroll_left_action method.

AccessibleToolButtonScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleToolButton::scroll_right_action method.

AccessibleToolButtonScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleToolButton::scroll_up_action method.

AccessibleToolButtonStateArgs
This trait represents a set of arguments accepted by AccessibleToolButton::state method.

AccessibleToolButtonTextArgs
This trait represents a set of arguments accepted by AccessibleToolButton::text method.

AccessibleToolButtonTrArgs
This trait represents a set of arguments accepted by AccessibleToolButton::tr method.

AccessibleToolButtonTrUtf8Args
This trait represents a set of arguments accepted by AccessibleToolButton::tr_utf8 method.