Module qt_widgets::accessible_tree::overloading

Traits

AccessibleTreeBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleTree::background_color method.

AccessibleTreeColumnDescriptionArgs
This trait represents a set of arguments accepted by AccessibleTree::column_description method.

AccessibleTreeForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleTree::foreground_color method.

AccessibleTreeNewArgs
This trait represents a set of arguments accepted by AccessibleTree::new method.

AccessibleTreeRectArgs
This trait represents a set of arguments accepted by AccessibleTree::rect method.

AccessibleTreeRelationsArgs
This trait represents a set of arguments accepted by AccessibleTree::relations method.

AccessibleTreeRowDescriptionArgs
This trait represents a set of arguments accepted by AccessibleTree::row_description method.

AccessibleTreeSelectedCellsArgs
This trait represents a set of arguments accepted by AccessibleTree::selected_cells method.

AccessibleTreeSelectedColumnsArgs
This trait represents a set of arguments accepted by AccessibleTree::selected_columns method.

AccessibleTreeSelectedRowsArgs
This trait represents a set of arguments accepted by AccessibleTree::selected_rows method.

AccessibleTreeStateArgs
This trait represents a set of arguments accepted by AccessibleTree::state method.

AccessibleTreeTextArgs
This trait represents a set of arguments accepted by AccessibleTree::text method.