Module qt_widgets::accessible_widget::overloading

Traits

AccessibleWidgetActionNamesArgs
This trait represents a set of arguments accepted by AccessibleWidget::action_names method.

AccessibleWidgetBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleWidget::background_color method.

AccessibleWidgetForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleWidget::foreground_color method.

AccessibleWidgetKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleWidget::key_bindings_for_action method.

AccessibleWidgetLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleWidget::localized_action_description method.

AccessibleWidgetLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleWidget::localized_action_name method.

AccessibleWidgetNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleWidget::next_page_action method.

AccessibleWidgetPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleWidget::previous_page_action method.

AccessibleWidgetRectArgs
This trait represents a set of arguments accepted by AccessibleWidget::rect method.

AccessibleWidgetRelationsArgs
This trait represents a set of arguments accepted by AccessibleWidget::relations method.

AccessibleWidgetScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleWidget::scroll_down_action method.

AccessibleWidgetScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleWidget::scroll_left_action method.

AccessibleWidgetScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleWidget::scroll_right_action method.

AccessibleWidgetScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleWidget::scroll_up_action method.

AccessibleWidgetStateArgs
This trait represents a set of arguments accepted by AccessibleWidget::state method.

AccessibleWidgetTextArgs
This trait represents a set of arguments accepted by AccessibleWidget::text method.

AccessibleWidgetTrArgs
This trait represents a set of arguments accepted by AccessibleWidget::tr method.

AccessibleWidgetTrUtf8Args
This trait represents a set of arguments accepted by AccessibleWidget::tr_utf8 method.