Struct qt_widgets::accessible_widget::AccessibleWidget

pub struct AccessibleWidget {
    // some fields omitted
}

C++ type: QAccessibleWidget

C++ documentation:

The QAccessibleWidget class implements the QAccessibleInterface for QWidgets.

This class is part of Accessibility for QWidget Applications.

This class is convenient to use as a base class for custom implementations of QAccessibleInterfaces that provide information about widget objects.

The class provides functions to retrieve the parentObject() (the widget's parent widget), and the associated widget(). Controlling signals can be added with addControllingSignal(), and setters are provided for various aspects of the interface implementation, for example setValue(), setDescription(), setAccelerator(), and setHelp().

Methods

impl AccessibleWidget

fn action_interface(&mut self) -> *mut AccessibleActionInterface

C++ method: QAccessibleActionInterface* QAccessibleWidget::actionInterface()

fn action_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleWidgetActionNamesArgs<'largs>

C++ method: QAccessibleWidget::actionNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn action_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn action_names(&self, cpp_utils::AsStruct) -> qt_core::string_list::StringList
C++ method: virtual QStringList QAccessibleWidget::actionNames() const

C++ documentation:

Reimplemented from QAccessibleActionInterface::actionNames().

fn background_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleWidgetBackgroundColorArgs<'largs>

C++ method: QAccessibleWidget::backgroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn background_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn background_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: virtual QColor QAccessibleWidget::backgroundColor() const

C++ documentation:

Reimplemented from QAccessibleInterface::backgroundColor().

fn child(&self, index: c_int) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleWidget::child(int index) const

C++ documentation:

Reimplemented from QAccessibleInterface::child().

fn child_at(&self, x: c_int, y: c_int) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleWidget::childAt(int x, int y) const

C++ documentation:

Reimplemented from QAccessibleInterface::childAt().

fn child_count(&self) -> c_int

C++ method: virtual int QAccessibleWidget::childCount() const

C++ documentation:

Reimplemented from QAccessibleInterface::childCount().

fn decrease_action() -> &'static String

C++ method: static const QString& QAccessibleWidget::decreaseAction()

C++ documentation:

Returns the name of the decrease default action.

See also actionNames() and localizedActionName().

fn do_action(&mut self, action_name: &String)

C++ method: virtual void QAccessibleWidget::doAction(const QString& actionName)

C++ documentation:

Reimplemented from QAccessibleActionInterface::doAction().

fn editable_text_interface(&mut self) -> *mut AccessibleEditableTextInterface

C++ method: QAccessibleEditableTextInterface* QAccessibleWidget::editableTextInterface()

fn focus_child(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleWidget::focusChild() const

C++ documentation:

Reimplemented from QAccessibleInterface::focusChild().

fn foreground_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleWidgetForegroundColorArgs<'largs>

C++ method: QAccessibleWidget::foregroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn foreground_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn foreground_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: virtual QColor QAccessibleWidget::foregroundColor() const

C++ documentation:

Reimplemented from QAccessibleInterface::foregroundColor().

fn image_interface(&mut self) -> *mut AccessibleImageInterface

C++ method: QAccessibleImageInterface* QAccessibleWidget::imageInterface()

fn increase_action() -> &'static String

C++ method: static const QString& QAccessibleWidget::increaseAction()

C++ documentation:

Returns the name of the increase default action.

See also actionNames() and localizedActionName().

fn index_of_child(&self, child: *const AccessibleInterface) -> c_int

C++ method: virtual int QAccessibleWidget::indexOfChild(const QAccessibleInterface* child) const

C++ documentation:

Reimplemented from QAccessibleInterface::indexOfChild().

fn interface_cast(&mut self, t: InterfaceType) -> *mut c_void

C++ method: virtual void* QAccessibleWidget::interface_cast(QAccessible::InterfaceType t)

C++ documentation:

Reimplemented from QAccessibleInterface::interface_cast().

fn is_valid(&self) -> bool

C++ method: virtual bool QAccessibleWidget::isValid() const

C++ documentation:

Reimplemented from QAccessibleInterface::isValid().

fn key_bindings_for_action<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleWidgetKeyBindingsForActionArgs<'largs>

C++ method: QAccessibleWidget::keyBindingsForAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn key_bindings_for_action(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn key_bindings_for_action(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string_list::StringList
C++ method: virtual QStringList QAccessibleWidget::keyBindingsForAction(const QString& actionName) const

C++ documentation:

Reimplemented from QAccessibleActionInterface::keyBindingsForAction().

fn localized_action_description<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleWidgetLocalizedActionDescriptionArgs<'largs>

C++ method: QAccessibleWidget::localizedActionDescription

This is an overloaded function. Available variants:

Rust arguments:
1) fn localized_action_description(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn localized_action_description(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleWidget::localizedActionDescription(const QString& name) const

C++ documentation:

Returns a localized action description of the action actionName.

When using one of the default names, you can call this function in QAccessibleActionInterface to get the localized string.

See also actionNames() and localizedActionName().

fn localized_action_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleWidgetLocalizedActionNameArgs<'largs>

C++ method: QAccessibleWidget::localizedActionName

This is an overloaded function. Available variants:

Rust arguments:
1) fn localized_action_name(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn localized_action_name(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleWidget::localizedActionName(const QString& name) const

C++ documentation:

Returns a localized action name of actionName.

For custom actions this function has to be re-implemented. When using one of the default names, you can call this function in QAccessibleActionInterface to get the localized string.

See also actionNames() and localizedActionDescription().

fn next_page_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleWidgetNextPageActionArgs

C++ method: QAccessibleWidget::nextPageAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn next_page_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn next_page_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleWidget::nextPageAction()

C++ documentation:

Returns the name of the next page default action.

See also actionNames() and localizedActionName().

fn object(&self) -> *mut Object

C++ method: virtual QObject* QAccessibleWidget::object() const

C++ documentation:

Reimplemented from QAccessibleInterface::object().

fn parent(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleWidget::parent() const

C++ documentation:

Reimplemented from QAccessibleInterface::parent().

fn press_action() -> &'static String

C++ method: static const QString& QAccessibleWidget::pressAction()

C++ documentation:

Returns the name of the press default action.

See also actionNames() and localizedActionName().

fn previous_page_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleWidgetPreviousPageActionArgs

C++ method: QAccessibleWidget::previousPageAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn previous_page_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn previous_page_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleWidget::previousPageAction()

C++ documentation:

Returns the name of the previous page default action.

See also actionNames() and localizedActionName().

fn rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleWidgetRectArgs<'largs>

C++ method: QAccessibleWidget::rect

This is an overloaded function. Available variants:

Rust arguments:
1) fn rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: virtual QRect QAccessibleWidget::rect() const

C++ documentation:

Reimplemented from QAccessibleInterface::rect().

fn relations<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleWidgetRelationsArgs<'largs>

C++ method: QAccessibleWidget::relations

This is an overloaded function. Available variants:

Rust arguments:
1) fn relations(&self, (qt_core::flags::Flags<qt_gui::accessible::RelationFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
2) fn relations(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
3) fn relations(&self, (qt_core::flags::Flags<qt_gui::accessible::RelationFlag>, cpp_utils::AsStruct)) -> qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
4) fn relations(&self, cpp_utils::AsStruct) -> qt_gui::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
C++ method: virtual QVector<QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>> QAccessibleWidget::relations(QFlags<QAccessible::RelationFlag> match = ?) const

C++ documentation:

Reimplemented from QAccessibleInterface::relations().

fn role(&self) -> Role

C++ method: virtual QAccessible::Role QAccessibleWidget::role() const

C++ documentation:

Reimplemented from QAccessibleInterface::role().

fn scroll_down_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleWidgetScrollDownActionArgs

C++ method: QAccessibleWidget::scrollDownAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_down_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_down_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleWidget::scrollDownAction()

C++ documentation:

Returns the name of the scroll down default action.

See also actionNames() and localizedActionName().

fn scroll_left_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleWidgetScrollLeftActionArgs

C++ method: QAccessibleWidget::scrollLeftAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_left_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_left_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleWidget::scrollLeftAction()

C++ documentation:

Returns the name of the scroll left default action.

See also actionNames() and localizedActionName().

fn scroll_right_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleWidgetScrollRightActionArgs

C++ method: QAccessibleWidget::scrollRightAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_right_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_right_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleWidget::scrollRightAction()

C++ documentation:

Returns the name of the scroll right default action.

See also actionNames() and localizedActionName().

fn scroll_up_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleWidgetScrollUpActionArgs

C++ method: QAccessibleWidget::scrollUpAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_up_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_up_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleWidget::scrollUpAction()

C++ documentation:

Returns the name of the scroll up default action.

See also actionNames() and localizedActionName().

fn set_focus_action() -> &'static String

C++ method: static const QString& QAccessibleWidget::setFocusAction()

C++ documentation:

Returns the name of the set focus default action.

See also actionNames() and localizedActionName().

fn set_text(&mut self, t: Text, text: &String)

C++ method: virtual void QAccessibleWidget::setText(QAccessible::Text t, const QString& text)

C++ documentation:

Reimplemented from QAccessibleInterface::setText().

fn show_menu_action() -> &'static String

C++ method: static const QString& QAccessibleWidget::showMenuAction()

C++ documentation:

Returns the name of the show menu default action.

See also actionNames() and localizedActionName().

fn state<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleWidgetStateArgs<'largs>

C++ method: QAccessibleWidget::state

This is an overloaded function. Available variants:

Rust arguments:
1) fn state(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::accessible::State>
2) fn state(&self, cpp_utils::AsStruct) -> qt_gui::accessible::State
C++ method: virtual QAccessible::State QAccessibleWidget::state() const

C++ documentation:

Reimplemented from QAccessibleInterface::state().

fn table_cell_interface(&mut self) -> *mut AccessibleTableCellInterface

C++ method: QAccessibleTableCellInterface* QAccessibleWidget::tableCellInterface()

fn table_interface(&mut self) -> *mut AccessibleTableInterface

C++ method: QAccessibleTableInterface* QAccessibleWidget::tableInterface()

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleWidgetTextArgs<'largs>

C++ method: QAccessibleWidget::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(&self, (qt_gui::accessible::Text, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, (qt_gui::accessible::Text, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleWidget::text(QAccessible::Text t) const

C++ documentation:

Reimplemented from QAccessibleInterface::text().

fn text_interface(&mut self) -> *mut AccessibleTextInterface

C++ method: QAccessibleTextInterface* QAccessibleWidget::textInterface()

fn toggle_action() -> &'static String

C++ method: static const QString& QAccessibleWidget::toggleAction()

C++ documentation:

Returns the name of the toggle default action.

See also actionNames() and localizedActionName().

fn tr<Args>(args: Args) -> Args::ReturnType where Args: AccessibleWidgetTrArgs

C++ method: QAccessibleWidget::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QAccessibleWidget::tr(const char* sourceText, const char* disambiguation, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: AccessibleWidgetTrUtf8Args

C++ method: QAccessibleWidget::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QAccessibleWidget::trUtf8(const char* sourceText, const char* disambiguation, int n)

fn value_interface(&mut self) -> *mut AccessibleValueInterface

C++ method: QAccessibleValueInterface* QAccessibleWidget::valueInterface()

fn virtual_hook(&mut self, id: c_int, data: *mut c_void)

C++ method: virtual void QAccessibleWidget::virtual_hook(int id, void* data)

fn window(&self) -> *mut Window

C++ method: virtual QWindow* QAccessibleWidget::window() const

C++ documentation:

Reimplemented from QAccessibleInterface::window().