Module qt_widgets::accessible_window_container::overloading

Traits

AccessibleWindowContainerActionNamesArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::action_names method.

AccessibleWindowContainerBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::background_color method.

AccessibleWindowContainerForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::foreground_color method.

AccessibleWindowContainerKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::key_bindings_for_action method.

AccessibleWindowContainerLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::localized_action_description method.

AccessibleWindowContainerLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::localized_action_name method.

AccessibleWindowContainerNewArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::new method.

AccessibleWindowContainerNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::next_page_action method.

AccessibleWindowContainerPreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::previous_page_action method.

AccessibleWindowContainerRectArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::rect method.

AccessibleWindowContainerRelationsArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::relations method.

AccessibleWindowContainerScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::scroll_down_action method.

AccessibleWindowContainerScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::scroll_left_action method.

AccessibleWindowContainerScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::scroll_right_action method.

AccessibleWindowContainerScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::scroll_up_action method.

AccessibleWindowContainerStateArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::state method.

AccessibleWindowContainerTextArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::text method.

AccessibleWindowContainerTrArgs
This trait represents a set of arguments accepted by AccessibleWindowContainer::tr method.

AccessibleWindowContainerTrUtf8Args
This trait represents a set of arguments accepted by AccessibleWindowContainer::tr_utf8 method.