Enum qt_widgets::action::MenuRole

pub enum MenuRole {
    No,
    TextHeuristic,
    ApplicationSpecific,
    AboutQt,
    About,
    Preferences,
    Quit,
}

C++ type: QAction::MenuRole

C++ documentation:

This enum describes how an action should be moved into the application menu on OS X.

Setting this value only has effect on items that are in the immediate menus of the menubar, not the submenus of those menus. For example, if you have File menu in your menubar and the File menu has a submenu, setting the MenuRole for the actions in that submenu have no effect. They will never be moved.

Variants
No

This action should not be put into the application menu (C++ enum variant: NoRole = 0)

TextHeuristic

This action should be put in the application menu based on the action's text as described in the QMenuBar documentation. (C++ enum variant: TextHeuristicRole = 1)

ApplicationSpecific

This action should be put in the application menu with an application specific role (C++ enum variant: ApplicationSpecificRole = 2)

AboutQt

This action matches handles the "About Qt" menu item. (C++ enum variant: AboutQtRole = 3)

About

This action should be placed where the "About" menu item is in the application menu. The text of the menu item will be set to "About <application name>". The application name is fetched from the Info.plist file in the application's bundle (See Qt for OS X - Deployment). (C++ enum variant: AboutRole = 4)

Preferences

This action should be placed where the "Preferences..." menu item is in the application menu. (C++ enum variant: PreferencesRole = 5)

Quit

This action should be placed where the Quit menu item is in the application menu. (C++ enum variant: QuitRole = 6)