Module qt_widgets::action::overloading

Traits

ActionAssociatedGraphicsWidgetsArgs
This trait represents a set of arguments accepted by Action::associated_graphics_widgets method.

ActionAssociatedWidgetsArgs
This trait represents a set of arguments accepted by Action::associated_widgets method.

ActionConnectArgs
This trait represents a set of arguments accepted by Action::connect method.

ActionConnectStaticArgs
This trait represents a set of arguments accepted by Action::connect_static method.

ActionDataArgs
This trait represents a set of arguments accepted by Action::data method.

ActionDestroyedArgs
This trait represents a set of arguments accepted by Action::destroyed method.

ActionDisconnectArgs
This trait represents a set of arguments accepted by Action::disconnect method.

ActionDisconnectStaticArgs
This trait represents a set of arguments accepted by Action::disconnect_static method.

ActionDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Action::dynamic_property_names method.

ActionFontArgs
This trait represents a set of arguments accepted by Action::font method.

ActionIconArgs
This trait represents a set of arguments accepted by Action::icon method.

ActionIconTextArgs
This trait represents a set of arguments accepted by Action::icon_text method.

ActionNewArgs
This trait represents a set of arguments accepted by Action::new method.

ActionObjectNameArgs
This trait represents a set of arguments accepted by Action::object_name method.

ActionPropertyArgs
This trait represents a set of arguments accepted by Action::property method.

ActionSetShortcutsArgs
This trait represents a set of arguments accepted by Action::set_shortcuts method.

ActionShortcutArgs
This trait represents a set of arguments accepted by Action::shortcut method.

ActionShortcutsArgs
This trait represents a set of arguments accepted by Action::shortcuts method.

ActionShowStatusTextArgs
This trait represents a set of arguments accepted by Action::show_status_text method.

ActionStartTimerArgs
This trait represents a set of arguments accepted by Action::start_timer method.

ActionStatusTipArgs
This trait represents a set of arguments accepted by Action::status_tip method.

ActionTextArgs
This trait represents a set of arguments accepted by Action::text method.

ActionToolTipArgs
This trait represents a set of arguments accepted by Action::tool_tip method.

ActionTrArgs
This trait represents a set of arguments accepted by Action::tr method.

ActionTrUtf8Args
This trait represents a set of arguments accepted by Action::tr_utf8 method.

ActionTriggeredArgs
This trait represents a set of arguments accepted by Action::triggered method.

ActionWhatsThisArgs
This trait represents a set of arguments accepted by Action::whats_this method.