Module qt_widgets::action_group::overloading

Traits

ActionGroupActionsArgs
This trait represents a set of arguments accepted by ActionGroup::actions method.

ActionGroupAddActionArgs
This trait represents a set of arguments accepted by ActionGroup::add_action method.

ActionGroupConnectArgs
This trait represents a set of arguments accepted by ActionGroup::connect method.

ActionGroupConnectStaticArgs
This trait represents a set of arguments accepted by ActionGroup::connect_static method.

ActionGroupDestroyedArgs
This trait represents a set of arguments accepted by ActionGroup::destroyed method.

ActionGroupDisconnectArgs
This trait represents a set of arguments accepted by ActionGroup::disconnect method.

ActionGroupDisconnectStaticArgs
This trait represents a set of arguments accepted by ActionGroup::disconnect_static method.

ActionGroupDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by ActionGroup::dynamic_property_names method.

ActionGroupNewArgs
This trait represents a set of arguments accepted by ActionGroup::new method.

ActionGroupObjectNameArgs
This trait represents a set of arguments accepted by ActionGroup::object_name method.

ActionGroupPropertyArgs
This trait represents a set of arguments accepted by ActionGroup::property method.

ActionGroupStartTimerArgs
This trait represents a set of arguments accepted by ActionGroup::start_timer method.

ActionGroupTrArgs
This trait represents a set of arguments accepted by ActionGroup::tr method.

ActionGroupTrUtf8Args
This trait represents a set of arguments accepted by ActionGroup::tr_utf8 method.