Module qt_widgets::application::overloading

Traits

ApplicationAlertArgs
This trait represents a set of arguments accepted by Application::alert method.

ApplicationAllWidgetsArgs
This trait represents a set of arguments accepted by Application::all_widgets method.

ApplicationAllWindowsArgs
This trait represents a set of arguments accepted by Application::all_windows method.

ApplicationApplicationDirPathArgs
This trait represents a set of arguments accepted by Application::application_dir_path method.

ApplicationApplicationDisplayNameArgs
This trait represents a set of arguments accepted by Application::application_display_name method.

ApplicationApplicationFilePathArgs
This trait represents a set of arguments accepted by Application::application_file_path method.

ApplicationApplicationNameArgs
This trait represents a set of arguments accepted by Application::application_name method.

ApplicationApplicationVersionArgs
This trait represents a set of arguments accepted by Application::application_version method.

ApplicationArgumentsArgs
This trait represents a set of arguments accepted by Application::arguments method.

ApplicationConnectArgs
This trait represents a set of arguments accepted by Application::connect method.

ApplicationConnectStaticArgs
This trait represents a set of arguments accepted by Application::connect_static method.

ApplicationDestroyedArgs
This trait represents a set of arguments accepted by Application::destroyed method.

ApplicationDisconnectArgs
This trait represents a set of arguments accepted by Application::disconnect method.

ApplicationDisconnectStaticArgs
This trait represents a set of arguments accepted by Application::disconnect_static method.

ApplicationDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Application::dynamic_property_names method.

ApplicationExitArgs
This trait represents a set of arguments accepted by Application::exit method.

ApplicationFontArgs
This trait represents a set of arguments accepted by Application::font method.

ApplicationFontMetricsArgs
This trait represents a set of arguments accepted by Application::font_metrics method.

ApplicationGlobalStrutArgs
This trait represents a set of arguments accepted by Application::global_strut method.

ApplicationLibraryPathsArgs
This trait represents a set of arguments accepted by Application::library_paths method.

ApplicationNewArgs
This trait represents a set of arguments accepted by Application::new method.

ApplicationObjectNameArgs
This trait represents a set of arguments accepted by Application::object_name method.

ApplicationOrganizationDomainArgs
This trait represents a set of arguments accepted by Application::organization_domain method.

ApplicationOrganizationNameArgs
This trait represents a set of arguments accepted by Application::organization_name method.

ApplicationPaletteArgs
This trait represents a set of arguments accepted by Application::palette method.

ApplicationPlatformNameArgs
This trait represents a set of arguments accepted by Application::platform_name method.

ApplicationPostEventArgs
This trait represents a set of arguments accepted by Application::post_event method.

ApplicationProcessEventsArgs
This trait represents a set of arguments accepted by Application::process_events method.

ApplicationPropertyArgs
This trait represents a set of arguments accepted by Application::property method.

ApplicationRemovePostedEventsArgs
This trait represents a set of arguments accepted by Application::remove_posted_events method.

ApplicationScreensArgs
This trait represents a set of arguments accepted by Application::screens method.

ApplicationSendPostedEventsArgs
This trait represents a set of arguments accepted by Application::send_posted_events method.

ApplicationSessionIdArgs
This trait represents a set of arguments accepted by Application::session_id method.

ApplicationSessionKeyArgs
This trait represents a set of arguments accepted by Application::session_key method.

ApplicationSetAttributeArgs
This trait represents a set of arguments accepted by Application::set_attribute method.

ApplicationSetEffectEnabledArgs
This trait represents a set of arguments accepted by Application::set_effect_enabled method.

ApplicationSetFontArgs
This trait represents a set of arguments accepted by Application::set_font method.

ApplicationSetPaletteArgs
This trait represents a set of arguments accepted by Application::set_palette method.

ApplicationSetStyleArgs
This trait represents a set of arguments accepted by Application::set_style method.

ApplicationStartTimerArgs
This trait represents a set of arguments accepted by Application::start_timer method.

ApplicationStyleSheetArgs
This trait represents a set of arguments accepted by Application::style_sheet method.

ApplicationTopLevelAtArgs
This trait represents a set of arguments accepted by Application::top_level_at method.

ApplicationTopLevelWidgetsArgs
This trait represents a set of arguments accepted by Application::top_level_widgets method.

ApplicationTopLevelWindowsArgs
This trait represents a set of arguments accepted by Application::top_level_windows method.

ApplicationTrArgs
This trait represents a set of arguments accepted by Application::tr method.

ApplicationTrUtf8Args
This trait represents a set of arguments accepted by Application::tr_utf8 method.

ApplicationTranslateArgs
This trait represents a set of arguments accepted by Application::translate method.

ApplicationWidgetAtArgs
This trait represents a set of arguments accepted by Application::widget_at method.

ApplicationWindowIconArgs
This trait represents a set of arguments accepted by Application::window_icon method.