Struct qt_widgets::colormap::Colormap

pub struct Colormap {
    // some fields omitted
}

C++ type: QColormap

C++ documentation:

The QColormap class maps device independent QColors to device dependent pixel values.

Methods

impl Colormap

fn cleanup()

C++ method: static void QColormap::cleanup()

fn color_at<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ColormapColorAtArgs<'largs>

C++ method: QColormap::colorAt

This is an overloaded function. Available variants:

Rust arguments:
1) fn color_at(&self, (libc::c_uint, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn color_at(&self, (libc::c_uint, cpp_utils::AsStruct)) -> qt_gui::color::Color
C++ method: const QColor QColormap::colorAt(unsigned int pixel) const

C++ documentation:

Returns a QColor for the pixel.

See also pixel().

fn colormap<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ColormapColormapArgs<'largs>

C++ method: QColormap::colormap

This is an overloaded function. Available variants:

Rust arguments:
1) fn colormap(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::vector::VectorColorRef>
2) fn colormap(&self, cpp_utils::AsStruct) -> ::vector::VectorColorRef
C++ method: const QVector<QColor> QColormap::colormap() const

C++ documentation:

Returns a vector of colors which represents the devices colormap for Indexed and Gray modes. This function returns an empty vector for Direct mode.

See also size().

fn depth(&self) -> c_int

C++ method: int QColormap::depth() const

C++ documentation:

Returns the depth of the device.

See also size().

fn initialize()

C++ method: static void QColormap::initialize()

fn instance<Args>(args: Args) -> Args::ReturnType where Args: ColormapInstanceArgs

C++ method: QColormap::instance

This is an overloaded function. Available variants:

Rust arguments:
1) fn instance(cpp_utils::AsBox) -> cpp_utils::CppBox<::colormap::Colormap>
2) fn instance((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::colormap::Colormap>
3) fn instance(cpp_utils::AsStruct) -> ::colormap::Colormap
4) fn instance((libc::c_int, cpp_utils::AsStruct)) -> ::colormap::Colormap
C++ method: static QColormap QColormap::instance(int screen = ?)

C++ documentation:

Returns the colormap for the specified screen. If screen is -1, this function returns the colormap for the default screen.

fn mode(&self) -> Mode

C++ method: QColormap::Mode QColormap::mode() const

C++ documentation:

Returns the mode of this colormap.

See also QColormap::Mode.

fn new<Args>(args: Args) -> Args::ReturnType where Args: ColormapNewArgs

C++ method: QColormap::QColormap

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((&::colormap::Colormap, cpp_utils::AsStruct)) -> ::colormap::Colormap
2) fn new((&::colormap::Colormap, cpp_utils::AsBox)) -> cpp_utils::CppBox<::colormap::Colormap>
C++ method: [constructor] void QColormap::QColormap(const QColormap& colormap)

fn op_assign<'l0, 'l1>(&'l0 mut self, colormap: &'l1 Colormap) -> &'l0 mut Colormap

C++ method: QColormap& QColormap::operator=(const QColormap& colormap)

C++ documentation:

Assigns the given colormap to this color map and returns a reference to this color map.

This function was introduced in Qt 4.2.

fn pixel(&self, color: &Color) -> c_uint

C++ method: unsigned int QColormap::pixel(const QColor& color) const

C++ documentation:

Returns a device dependent pixel value for the color.

See also colorAt().

fn size(&self) -> c_int

C++ method: int QColormap::size() const

C++ documentation:

Returns the size of the colormap for Indexed and Gray modes; Returns -1 for Direct mode.

See also colormap().