Module qt_widgets::completer::overloading

Traits

CompleterActivatedArgs
This trait represents a set of arguments accepted by Completer::activated method.

CompleterCompleteArgs
This trait represents a set of arguments accepted by Completer::complete method.

CompleterCompletionPrefixArgs
This trait represents a set of arguments accepted by Completer::completion_prefix method.

CompleterConnectArgs
This trait represents a set of arguments accepted by Completer::connect method.

CompleterConnectStaticArgs
This trait represents a set of arguments accepted by Completer::connect_static method.

CompleterCurrentCompletionArgs
This trait represents a set of arguments accepted by Completer::current_completion method.

CompleterCurrentIndexArgs
This trait represents a set of arguments accepted by Completer::current_index method.

CompleterDestroyedArgs
This trait represents a set of arguments accepted by Completer::destroyed method.

CompleterDisconnectArgs
This trait represents a set of arguments accepted by Completer::disconnect method.

CompleterDisconnectStaticArgs
This trait represents a set of arguments accepted by Completer::disconnect_static method.

CompleterDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Completer::dynamic_property_names method.

CompleterHighlightedArgs
This trait represents a set of arguments accepted by Completer::highlighted method.

CompleterNewArgs
This trait represents a set of arguments accepted by Completer::new method.

CompleterObjectNameArgs
This trait represents a set of arguments accepted by Completer::object_name method.

CompleterPathFromIndexArgs
This trait represents a set of arguments accepted by Completer::path_from_index method.

CompleterPropertyArgs
This trait represents a set of arguments accepted by Completer::property method.

CompleterSplitPathArgs
This trait represents a set of arguments accepted by Completer::split_path method.

CompleterStartTimerArgs
This trait represents a set of arguments accepted by Completer::start_timer method.

CompleterTrArgs
This trait represents a set of arguments accepted by Completer::tr method.

CompleterTrUtf8Args
This trait represents a set of arguments accepted by Completer::tr_utf8 method.