Struct qt_widgets::completer::Completer

pub struct Completer {
    // some fields omitted
}

C++ type: QCompleter

C++ documentation:

The QCompleter class provides completions based on an item model.

You can use QCompleter to provide auto completions in any Qt widget, such as QLineEdit and QComboBox. When the user starts typing a word, QCompleter suggests possible ways of completing the word, based on a word list. The word list is provided as a QAbstractItemModel. (For simple applications, where the word list is static, you can pass a QStringList to QCompleter's constructor.)

Methods

impl Completer

fn activated<'largs, Args>(&'largs mut self, args: Args) where Args: CompleterActivatedArgs<'largs>

C++ method: QCompleter::activated

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn activated(&mut self, &qt_core::model_index::ModelIndex) -> ()
C++ method: void QCompleter::activated(const QModelIndex& index)

C++ documentation:

This signal is sent when an item in the popup() is activated by the user. (by clicking or pressing return). The item's index in the completionModel() is given.

Variant 2

Rust arguments: fn activated(&mut self, &qt_core::string::String) -> ()
C++ method: void QCompleter::activated(const QString& text)

C++ documentation:

This signal is sent when an item in the popup() is activated by the user (by clicking or pressing return). The item's text is given.

fn block_signals(&mut self, b: bool) -> bool

C++ method: bool QCompleter::blockSignals(bool b)

C++ documentation:

If block is true, signals emitted by this object are blocked (i.e., emitting a signal will not invoke anything connected to it). If block is false, no such blocking will occur.

The return value is the previous value of signalsBlocked().

Note that the destroyed() signal will be emitted even if the signals for this object have been blocked.

See also signalsBlocked().

fn case_sensitivity(&self) -> CaseSensitivity

C++ method: Qt::CaseSensitivity QCompleter::caseSensitivity() const

C++ documentation:

This property holds the case sensitivity of the matching.

The default is Qt::CaseSensitive.

Access functions:

Qt::CaseSensitivity 	caseSensitivity() const

void 	setCaseSensitivity(Qt::CaseSensitivity caseSensitivity)

See also completionColumn, completionRole, and modelSorting.

fn children<'l0>(&'l0 self) -> &'l0 ListObjectMutPtr

C++ method: const QList<QObject*>& QCompleter::children() const

C++ documentation:

Returns a list of child objects. The QObjectList class is defined in the <QObject> header file as the following:

typedef QList<QObject*> QObjectList;

The first child added is the first object in the list and the last child added is the last object in the list, i.e. new children are appended at the end.

Note that the list order changes when QWidget children are raised or lowered. A widget that is raised becomes the last object in the list, and a widget that is lowered becomes the first object in the list.

See also findChild(), findChildren(), parent(), and setParent().

fn complete<'largs, Args>(&'largs mut self, args: Args) where Args: CompleterCompleteArgs<'largs>

C++ method: QCompleter::complete

This is an overloaded function. Available variants:

Rust arguments:
1) fn complete(&mut self, ()) -> ()
2) fn complete(&mut self, &qt_core::rect::Rect) -> ()
C++ method: void QCompleter::complete(const QRect& rect = ?)

C++ documentation:

For QCompleter::PopupCompletion and QCompletion::UnfilteredPopupCompletion modes, calling this function displays the popup displaying the current completions. By default, if rect is not specified, the popup is displayed on the bottom of the widget(). If rect is specified the popup is displayed on the left edge of the rectangle.

For QCompleter::InlineCompletion mode, the highlighted() signal is fired with the current completion.

fn completion_column(&self) -> c_int

C++ method: int QCompleter::completionColumn() const

C++ documentation:

This property holds the column in the model in which completions are searched for.

If the popup() is a QListView, it is automatically setup to display this column.

By default, the match column is 0.

Access functions:

int 	completionColumn() const

void 	setCompletionColumn(int column)

See also completionRole and caseSensitivity.

fn completion_count(&self) -> c_int

C++ method: int QCompleter::completionCount() const

C++ documentation:

Returns the number of completions for the current prefix. For an unsorted model with a large number of items this can be expensive. Use setCurrentRow() and currentCompletion() to iterate through all the completions.

fn completion_mode(&self) -> CompletionMode

C++ method: QCompleter::CompletionMode QCompleter::completionMode() const

C++ documentation:

This property holds how the completions are provided to the user.

The default value is QCompleter::PopupCompletion.

Access functions:

CompletionMode 	completionMode() const

void 	setCompletionMode(CompletionMode mode)

fn completion_model(&self) -> *mut AbstractItemModel

C++ method: QAbstractItemModel* QCompleter::completionModel() const

C++ documentation:

Returns the completion model. The completion model is a read-only list model that contains all the possible matches for the current completion prefix. The completion model is auto-updated to reflect the current completions.

Note: The return value of this function is defined to be an QAbstractItemModel purely for generality. This actual kind of model returned is an instance of an QAbstractProxyModel subclass.

See also completionPrefix and model().

fn completion_prefix<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CompleterCompletionPrefixArgs<'largs>

C++ method: QCompleter::completionPrefix

This is an overloaded function. Available variants:

Rust arguments:
1) fn completion_prefix(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn completion_prefix(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QCompleter::completionPrefix() const

C++ documentation:

This property holds the completion prefix used to provide completions.

The completionModel() is updated to reflect the list of possible matches for prefix.

Access functions:

QString 	completionPrefix() const

void 	setCompletionPrefix(const QString & prefix)

fn completion_role(&self) -> c_int

C++ method: int QCompleter::completionRole() const

C++ documentation:

This property holds the item role to be used to query the contents of items for matching.

The default role is Qt::EditRole.

Access functions:

int 	completionRole() const

void 	setCompletionRole(int role)

See also completionColumn and caseSensitivity.

fn connect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CompleterConnectArgs<'largs>

C++ method: QCompleter::connect

This is an overloaded function. Available variants:

Rust arguments:
1) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: QMetaObject::Connection QCompleter::connect(const QObject* sender, const char* signal, const char* member, Qt::ConnectionType type = ?) const

C++ documentation:

This function overloads connect().

Connects signal from the sender object to this object's method.

Equivalent to connect(sender, signal, this, method, type).

Every connection you make emits a signal, so duplicate connections emit two signals. You can break a connection using disconnect().

Note: This function is thread-safe.

See also disconnect().

fn connect_static<Args>(args: Args) -> Args::ReturnType where Args: CompleterConnectStaticArgs

C++ method: QCompleter::connect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: static QMetaObject::Connection QCompleter::connect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& method, Qt::ConnectionType type = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

The Connection handle will be invalid if it cannot create the connection, for example, the parameters were invalid. You can check if the QMetaObject::Connection is valid by casting it to a bool.

This function works in the same way as connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type) but it uses QMetaMethod to specify signal and method.

This function was introduced in Qt 4.8.

See also connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type).

Variant 2

Rust arguments:
1) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: static QMetaObject::Connection QCompleter::connect(const QObject* sender, const char* signal, const QObject* receiver, const char* member, Qt::ConnectionType arg5 = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

You must use the SIGNAL() and SLOT() macros when specifying the signal and the method, for example:

QLabel label = new QLabel;
QScrollBar scrollBar = new QScrollBar;
QObject::connect(scrollBar, SIGNAL(valueChanged(int)),
                 label,  SLOT(setNum(int)));

This example ensures that the label always displays the current scroll bar value. Note that the signal and slots parameters must not contain any variable names, only the type. E.g. the following would not work and return false:

// WRONG
QObject::connect(scrollBar, SIGNAL(valueChanged(int value)),
                 label, SLOT(setNum(int value)));

A signal can also be connected to another signal:

class MyWidget : public QWidget
{
    Q_OBJECT

public: MyWidget();

signals: void buttonClicked();

private: QPushButton *myButton; };

MyWidget::MyWidget() { myButton = new QPushButton(this); connect(myButton, SIGNAL(clicked()), this, SIGNAL(buttonClicked())); }
In this example, the MyWidget constructor relays a signal from a private member variable, and makes it available under a name that relates to MyWidget.

A signal can be connected to many slots and signals. Many signals can be connected to one slot.

If a signal is connected to several slots, the slots are activated in the same order in which the connections were made, when the signal is emitted.

The function returns a QMetaObject::Connection that represents a handle to a connection if it successfully connects the signal to the slot. The connection handle will be invalid if it cannot create the connection, for example, if QObject is unable to verify the existence of either signal or method, or if their signatures aren't compatible. You can check if the handle is valid by casting it to a bool.

By default, a signal is emitted for every connection you make; two signals are emitted for duplicate connections. You can break all of these connections with a single disconnect() call. If you pass the Qt::UniqueConnection type, the connection will only be made if it is not a duplicate. If there is already a duplicate (exact same signal to the exact same slot on the same objects), the connection will fail and connect will return an invalid QMetaObject::Connection.

The optional type parameter describes the type of connection to establish. In particular, it determines whether a particular signal is delivered to a slot immediately or queued for delivery at a later time. If the signal is queued, the parameters must be of types that are known to Qt's meta-object system, because Qt needs to copy the arguments to store them in an event behind the scenes. If you try to use a queued connection and get the error message

QObject::connect: Cannot queue arguments of type 'MyType'
(Make sure 'MyType' is registered using qRegisterMetaType().)

call qRegisterMetaType() to register the data type before you establish the connection.

Note: This function is thread-safe.

See also disconnect(), sender(), qRegisterMetaType(), and Q_DECLARE_METATYPE().

fn current_completion<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CompleterCurrentCompletionArgs<'largs>

C++ method: QCompleter::currentCompletion

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_completion(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn current_completion(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QCompleter::currentCompletion() const

C++ documentation:

Returns the current completion string. This includes the completionPrefix. When used alongside setCurrentRow(), it can be used to iterate through all the matches.

See also setCurrentRow() and currentIndex().

fn current_index<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CompleterCurrentIndexArgs<'largs>

C++ method: QCompleter::currentIndex

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_index(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::model_index::ModelIndex>
2) fn current_index(&self, cpp_utils::AsStruct) -> qt_core::model_index::ModelIndex
C++ method: QModelIndex QCompleter::currentIndex() const

C++ documentation:

Returns the model index of the current completion in the completionModel().

See also setCurrentRow(), currentCompletion(), and model().

fn current_row(&self) -> c_int

C++ method: int QCompleter::currentRow() const

C++ documentation:

Returns the current row.

See also setCurrentRow().

fn delete_later(&mut self)

C++ method: void QCompleter::deleteLater()

C++ documentation:

Schedules this object for deletion.

The object will be deleted when control returns to the event loop. If the event loop is not running when this function is called (e.g. deleteLater() is called on an object before QCoreApplication::exec()), the object will be deleted once the event loop is started. If deleteLater() is called after the main event loop has stopped, the object will not be deleted. Since Qt 4.8, if deleteLater() is called on an object that lives in a thread with no running event loop, the object will be destroyed when the thread finishes.

Note that entering and leaving a new event loop (e.g., by opening a modal dialog) will not perform the deferred deletion; for the object to be deleted, the control must return to the event loop from which deleteLater() was called.

Note: It is safe to call this function more than once; when the first deferred deletion event is delivered, any pending events for the object are removed from the event queue.

See also destroyed() and QPointer.

fn destroyed<'largs, Args>(&'largs mut self, args: Args) where Args: CompleterDestroyedArgs<'largs>

C++ method: QCompleter::destroyed

This is an overloaded function. Available variants:

Rust arguments:
1) fn destroyed(&mut self, *mut qt_core::object::Object) -> ()
2) fn destroyed(&mut self, ()) -> ()
C++ method: void QCompleter::destroyed(QObject* arg1 = ?)

C++ documentation:

This signal is emitted immediately before the object obj is destroyed, and can not be blocked.

All the objects's children are destroyed immediately after this signal is emitted.

See also deleteLater() and QPointer.

fn disconnect<'largs, Args>(&'largs self, args: Args) -> bool where Args: CompleterDisconnectArgs<'largs>

C++ method: QCompleter::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn disconnect(&self, *const qt_core::object::Object) -> bool
2) fn disconnect(&self, (*const qt_core::object::Object, *const libc::c_char)) -> bool
C++ method: bool QCompleter::disconnect(const QObject* receiver, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects all signals in this object from receiver's method.

A signal-slot connection is removed when either of the objects involved are destroyed.

Variant 2

Rust arguments:
1) fn disconnect(&self, *const libc::c_char) -> bool
2) fn disconnect(&self, (*const libc::c_char, *const qt_core::object::Object)) -> bool
3) fn disconnect(&self, (*const libc::c_char, *const qt_core::object::Object, *const libc::c_char)) -> bool
4) fn disconnect(&self, ()) -> bool
C++ method: bool QCompleter::disconnect(const char* signal = ?, const QObject* receiver = ?, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects signal from method of receiver.

A signal-slot connection is removed when either of the objects involved are destroyed.

Note: This function is thread-safe.

fn disconnect_static<Args>(args: Args) -> bool where Args: CompleterDisconnectStaticArgs

C++ method: QCompleter::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn disconnect_static(&qt_core::meta_object::Connection) -> bool
C++ method: static bool QCompleter::disconnect(const QMetaObject::Connection& arg1)

C++ documentation:

Disconnect a connection.

If the connection is invalid or has already been disconnected, do nothing and return false.

See also connect().

Variant 2

Rust arguments: fn disconnect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod)) -> bool
C++ method: static bool QCompleter::disconnect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

This function provides the same possibilities like disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method) but uses QMetaMethod to represent the signal and the method to be disconnected.

Additionally this function returnsfalse and no signals and slots disconnected if:

signal is not a member of sender class or one of its parent classes.

method is not a member of receiver class or one of its parent classes.

signal instance represents not a signal.

QMetaMethod() may be used as wildcard in the meaning "any signal" or "any slot in receiving object". In the same way 0 can be used for receiver in the meaning "any receiving object". In this case method should also be QMetaMethod(). sender parameter should be never 0.

This function was introduced in Qt 4.8.

See also disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method).

Variant 3

Rust arguments: fn disconnect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char)) -> bool
C++ method: static bool QCompleter::disconnect(const QObject* sender, const char* signal, const QObject* receiver, const char* member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

A signal-slot connection is removed when either of the objects involved are destroyed.

disconnect() is typically used in three ways, as the following examples demonstrate.

Disconnect everything connected to an object's signals:
disconnect(myObject, 0, 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect();

Disconnect everything connected to a specific signal:
disconnect(myObject, SIGNAL(mySignal()), 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect(SIGNAL(mySignal()));

Disconnect a specific receiver:
disconnect(myObject, 0, myReceiver, 0);

equivalent to the non-static overloaded function

myObject->disconnect(myReceiver);

0 may be used as a wildcard, meaning "any signal", "any receiving object", or "any slot in the receiving object", respectively.

The sender may never be 0. (You cannot disconnect signals from more than one object in a single call.)

If signal is 0, it disconnects receiver and method from any signal. If not, only the specified signal is disconnected.

If receiver is 0, it disconnects anything connected to signal. If not, slots in objects other than receiver are not disconnected.

If method is 0, it disconnects anything that is connected to receiver. If not, only slots named method will be disconnected, and all other slots are left alone. The method must be 0 if receiver is left out, so you cannot disconnect a specifically-named slot on all objects.

Note: This function is thread-safe.

See also connect().

fn dump_object_info(&mut self)

C++ method: void QCompleter::dumpObjectInfo()

C++ documentation:

Dumps information about signal connections, etc. for this object to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectTree().

fn dump_object_tree(&mut self)

C++ method: void QCompleter::dumpObjectTree()

C++ documentation:

Dumps a tree of children to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectInfo().

fn dynamic_property_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CompleterDynamicPropertyNamesArgs<'largs>

C++ method: QCompleter::dynamicPropertyNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn dynamic_property_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn dynamic_property_names(&self, cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: QList<QByteArray> QCompleter::dynamicPropertyNames() const

C++ documentation:

Returns the names of all properties that were dynamically added to the object using setProperty().

This function was introduced in Qt 4.2.

fn filter_mode(&self) -> Flags<MatchFlag>

C++ method: QFlags<Qt::MatchFlag> QCompleter::filterMode() const

C++ documentation:

This property holds how the filtering is performed.

If filterMode is set to Qt::MatchStartsWith, only those entries that start with the typed characters will be displayed. Qt::MatchContains will display the entries that contain the typed characters, and Qt::MatchEndsWith the ones that end with the typed characters.

Currently, only these three modes are implemented. Setting filterMode to any other Qt::MatchFlag will issue a warning, and no action will be performed.

The default mode is Qt::MatchStartsWith.

This property was introduced in Qt 5.2.

Access functions:

Qt::MatchFlags 	filterMode() const

void 	setFilterMode(Qt::MatchFlags filterMode)

fn highlighted<'largs, Args>(&'largs mut self, args: Args) where Args: CompleterHighlightedArgs<'largs>

C++ method: QCompleter::highlighted

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn highlighted(&mut self, &qt_core::model_index::ModelIndex) -> ()
C++ method: void QCompleter::highlighted(const QModelIndex& index)

C++ documentation:

This signal is sent when an item in the popup() is highlighted by the user. It is also sent if complete() is called with the completionMode() set to QCompleter::InlineCompletion. The item's index in the completionModel() is given.

Variant 2

Rust arguments: fn highlighted(&mut self, &qt_core::string::String) -> ()
C++ method: void QCompleter::highlighted(const QString& text)

C++ documentation:

This signal is sent when an item in the popup() is highlighted by the user. It is also sent if complete() is called with the completionMode() set to QCompleter::InlineCompletion. The item's text is given.

fn inherits(&self, classname: *const c_char) -> bool

C++ method: bool QCompleter::inherits(const char* classname) const

C++ documentation:

Returns true if this object is an instance of a class that inherits className or a QObject subclass that inherits className; otherwise returns false.

A class is considered to inherit itself.

Example:

QTimer *timer = new QTimer;         // QTimer inherits QObject
timer->inherits("QTimer");          // returns true
timer->inherits("QObject");         // returns true
timer->inherits("QAbstractButton"); // returns false

// QVBoxLayout inherits QObject and QLayoutItem QVBoxLayout *layout = new QVBoxLayout; layout->inherits("QObject"); // returns true layout->inherits("QLayoutItem"); // returns true (even though QLayoutItem is not a QObject)
If you need to determine whether an object is an instance of a particular class for the purpose of casting it, consider using qobject_cast<Type *>(object) instead.

See also metaObject() and qobject_cast().

fn install_event_filter(&mut self, arg1: *mut Object)

C++ method: void QCompleter::installEventFilter(QObject* arg1)

C++ documentation:

Installs an event filter filterObj on this object. For example:

monitoredObj->installEventFilter(filterObj);

An event filter is an object that receives all events that are sent to this object. The filter can either stop the event or forward it to this object. The event filter filterObj receives events via its eventFilter() function. The eventFilter() function must return true if the event should be filtered, (i.e. stopped); otherwise it must return false.

If multiple event filters are installed on a single object, the filter that was installed last is activated first.

Here's a KeyPressEater class that eats the key presses of its monitored objects:

class KeyPressEater : public QObject
{
    Q_OBJECT
    ...

protected: bool eventFilter(QObject obj, QEvent event); };

bool KeyPressEater::eventFilter(QObject obj, QEvent event) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent >(event); qDebug("Ate key press %d", keyEvent->key()); return true; } else { // standard event processing return QObject::eventFilter(obj, event); } }
And here's how to install it on two widgets:

KeyPressEater keyPressEater = new KeyPressEater(this);
QPushButton pushButton = new QPushButton(this);
QListView *listView = new QListView(this);

pushButton->installEventFilter(keyPressEater); listView->installEventFilter(keyPressEater);
The QShortcut class, for example, uses this technique to intercept shortcut key presses.

Warning: If you delete the receiver object in your eventFilter() function, be sure to return true. If you return false, Qt sends the event to the deleted object and the program will crash.

Note that the filtering object must be in the same thread as this object. If filterObj is in a different thread, this function does nothing. If either filterObj or this object are moved to a different thread after calling this function, the event filter will not be called until both objects have the same thread affinity again (it is not removed).

See also removeEventFilter(), eventFilter(), and event().

fn is_widget_type(&self) -> bool

C++ method: bool QCompleter::isWidgetType() const

C++ documentation:

Returns true if the object is a widget; otherwise returns false.

Calling this function is equivalent to calling inherits("QWidget"), except that it is much faster.

fn is_window_type(&self) -> bool

C++ method: bool QCompleter::isWindowType() const

C++ documentation:

Returns true if the object is a window; otherwise returns false.

Calling this function is equivalent to calling inherits("QWindow"), except that it is much faster.

fn kill_timer(&mut self, id: c_int)

C++ method: void QCompleter::killTimer(int id)

C++ documentation:

Kills the timer with timer identifier, id.

The timer identifier is returned by startTimer() when a timer event is started.

See also timerEvent() and startTimer().

fn max_visible_items(&self) -> c_int

C++ method: int QCompleter::maxVisibleItems() const

C++ documentation:

This property holds the maximum allowed size on screen of the completer, measured in items.

By default, this property has a value of 7.

This property was introduced in Qt 4.6.

Access functions:

int 	maxVisibleItems() const

void 	setMaxVisibleItems(int maxItems)

fn meta_object(&self) -> *const MetaObject

C++ method: virtual const QMetaObject* QCompleter::metaObject() const

fn model(&self) -> *mut AbstractItemModel

C++ method: QAbstractItemModel* QCompleter::model() const

C++ documentation:

Returns the model that provides completion strings.

See also setModel() and completionModel().

fn model_sorting(&self) -> ModelSorting

C++ method: QCompleter::ModelSorting QCompleter::modelSorting() const

C++ documentation:

This property holds the way the model is sorted.

By default, no assumptions are made about the order of the items in the model that provides the completions.

If the model's data for the completionColumn() and completionRole() is sorted in ascending order, you can set this property to CaseSensitivelySortedModel or CaseInsensitivelySortedModel. On large models, this can lead to significant performance improvements because the completer object can then use a binary search algorithm instead of linear search algorithm.

The sort order (i.e ascending or descending order) of the model is determined dynamically by inspecting the contents of the model.

Note: The performance improvements described above cannot take place when the completer's caseSensitivity is different to the case sensitivity used by the model's when sorting.

Access functions:

ModelSorting 	modelSorting() const

void 	setModelSorting(ModelSorting sorting)

See also setCaseSensitivity() and QCompleter::ModelSorting.

fn move_to_thread(&mut self, thread: *mut Thread)

C++ method: void QCompleter::moveToThread(QThread* thread)

C++ documentation:

Changes the thread affinity for this object and its children. The object cannot be moved if it has a parent. Event processing will continue in the targetThread.

To move an object to the main thread, use QApplication::instance() to retrieve a pointer to the current application, and then use QApplication::thread() to retrieve the thread in which the application lives. For example:

myObject->moveToThread(QApplication::instance()->thread());

If targetThread is zero, all event processing for this object and its children stops.

Note that all active timers for the object will be reset. The timers are first stopped in the current thread and restarted (with the same interval) in the targetThread. As a result, constantly moving an object between threads can postpone timer events indefinitely.

A QEvent::ThreadChange event is sent to this object just before the thread affinity is changed. You can handle this event to perform any special processing. Note that any new events that are posted to this object will be handled in the targetThread.

Warning: This function is not thread-safe; the current thread must be same as the current thread affinity. In other words, this function can only "push" an object from the current thread to another thread, it cannot "pull" an object from any arbitrary thread to the current thread.

See also thread().

fn new<Args>(args: Args) -> Args::ReturnType where Args: CompleterNewArgs

C++ method: QCompleter::QCompleter

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((*mut qt_core::abstract_item_model::AbstractItemModel, cpp_utils::AsStruct)) -> ::completer::Completer
2) fn new((*mut qt_core::abstract_item_model::AbstractItemModel, *mut qt_core::object::Object, cpp_utils::AsStruct)) -> ::completer::Completer
3) fn new((*mut qt_core::abstract_item_model::AbstractItemModel, cpp_utils::AsBox)) -> cpp_utils::CppBox<::completer::Completer>
4) fn new((*mut qt_core::abstract_item_model::AbstractItemModel, *mut qt_core::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::completer::Completer>
C++ method: [constructor] void QCompleter::QCompleter(QAbstractItemModel* model, QObject* parent = ?)

Variant 2

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::completer::Completer
2) fn new((*mut qt_core::object::Object, cpp_utils::AsStruct)) -> ::completer::Completer
3) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::completer::Completer>
4) fn new((*mut qt_core::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::completer::Completer>
C++ method: [constructor] void QCompleter::QCompleter(QObject* parent = ?)

Variant 3

Rust arguments:
1) fn new((&qt_core::string_list::StringList, cpp_utils::AsStruct)) -> ::completer::Completer
2) fn new((&qt_core::string_list::StringList, *mut qt_core::object::Object, cpp_utils::AsStruct)) -> ::completer::Completer
3) fn new((&qt_core::string_list::StringList, cpp_utils::AsBox)) -> cpp_utils::CppBox<::completer::Completer>
4) fn new((&qt_core::string_list::StringList, *mut qt_core::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::completer::Completer>
C++ method: [constructor] void QCompleter::QCompleter(const QStringList& completions, QObject* parent = ?)

fn object_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CompleterObjectNameArgs<'largs>

C++ method: QCompleter::objectName

This is an overloaded function. Available variants:

Rust arguments:
1) fn object_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn object_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QCompleter::objectName() const

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn parent(&self) -> *mut Object

C++ method: QObject* QCompleter::parent() const

C++ documentation:

Returns a pointer to the parent object.

See also setParent() and children().

fn path_from_index<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CompleterPathFromIndexArgs<'largs>

C++ method: QCompleter::pathFromIndex

This is an overloaded function. Available variants:

Rust arguments:
1) fn path_from_index(&self, (&qt_core::model_index::ModelIndex, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn path_from_index(&self, (&qt_core::model_index::ModelIndex, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QCompleter::pathFromIndex(const QModelIndex& index) const

C++ documentation:

Returns the path for the given index. The completer object uses this to obtain the completion text from the underlying model.

The default implementation returns the edit role of the item for list models. It returns the absolute file path if the model is a QFileSystemModel.

See also splitPath().

fn popup(&self) -> *mut AbstractItemView

C++ method: QAbstractItemView* QCompleter::popup() const

C++ documentation:

Returns the popup used to display completions.

See also setPopup().

fn property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CompleterPropertyArgs<'largs>

C++ method: QCompleter::property

This is an overloaded function. Available variants:

Rust arguments:
1) fn property(&self, (*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn property(&self, (*const libc::c_char, cpp_utils::AsStruct)) -> qt_core::variant::Variant
C++ method: QVariant QCompleter::property(const char* name) const

C++ documentation:

Returns the value of the object's name property.

If no such property exists, the returned variant is invalid.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

See also setProperty(), QVariant::isValid(), metaObject(), and dynamicPropertyNames().

fn qt_metacall(&mut self, arg1: Call, arg2: c_int, arg3: *mut *mut c_void) -> c_int

C++ method: virtual int QCompleter::qt_metacall(QMetaObject::Call arg1, int arg2, void** arg3)

fn qt_metacast(&mut self, arg1: *const c_char) -> *mut c_void

C++ method: virtual void* QCompleter::qt_metacast(const char* arg1)

fn register_user_data() -> c_uint

C++ method: static unsigned int QCompleter::registerUserData()

fn remove_event_filter(&mut self, arg1: *mut Object)

C++ method: void QCompleter::removeEventFilter(QObject* arg1)

C++ documentation:

Removes an event filter object obj from this object. The request is ignored if such an event filter has not been installed.

All event filters for this object are automatically removed when this object is destroyed.

It is always safe to remove an event filter, even during event filter activation (i.e. from the eventFilter() function).

See also installEventFilter(), eventFilter(), and event().

fn set_case_sensitivity(&mut self, case_sensitivity: CaseSensitivity)

C++ method: void QCompleter::setCaseSensitivity(Qt::CaseSensitivity caseSensitivity)

C++ documentation:

This property holds the case sensitivity of the matching.

The default is Qt::CaseSensitive.

Access functions:

Qt::CaseSensitivity 	caseSensitivity() const

void 	setCaseSensitivity(Qt::CaseSensitivity caseSensitivity)

See also completionColumn, completionRole, and modelSorting.

fn set_completion_column(&mut self, column: c_int)

C++ method: void QCompleter::setCompletionColumn(int column)

C++ documentation:

This property holds the column in the model in which completions are searched for.

If the popup() is a QListView, it is automatically setup to display this column.

By default, the match column is 0.

Access functions:

int 	completionColumn() const

void 	setCompletionColumn(int column)

See also completionRole and caseSensitivity.

fn set_completion_mode(&mut self, mode: CompletionMode)

C++ method: void QCompleter::setCompletionMode(QCompleter::CompletionMode mode)

C++ documentation:

This property holds how the completions are provided to the user.

The default value is QCompleter::PopupCompletion.

Access functions:

CompletionMode 	completionMode() const

void 	setCompletionMode(CompletionMode mode)

fn set_completion_prefix(&mut self, prefix: &String)

C++ method: void QCompleter::setCompletionPrefix(const QString& prefix)

C++ documentation:

This property holds the completion prefix used to provide completions.

The completionModel() is updated to reflect the list of possible matches for prefix.

Access functions:

QString 	completionPrefix() const

void 	setCompletionPrefix(const QString & prefix)

fn set_completion_role(&mut self, role: c_int)

C++ method: void QCompleter::setCompletionRole(int role)

C++ documentation:

This property holds the item role to be used to query the contents of items for matching.

The default role is Qt::EditRole.

Access functions:

int 	completionRole() const

void 	setCompletionRole(int role)

See also completionColumn and caseSensitivity.

fn set_current_row(&mut self, row: c_int) -> bool

C++ method: bool QCompleter::setCurrentRow(int row)

C++ documentation:

Sets the current row to the row specified. Returns true if successful; otherwise returns false.

This function may be used along with currentCompletion() to iterate through all the possible completions.

See also currentRow(), currentCompletion(), and completionCount().

fn set_filter_mode(&mut self, filter_mode: Flags<MatchFlag>)

C++ method: void QCompleter::setFilterMode(QFlags<Qt::MatchFlag> filterMode)

C++ documentation:

This property holds how the filtering is performed.

If filterMode is set to Qt::MatchStartsWith, only those entries that start with the typed characters will be displayed. Qt::MatchContains will display the entries that contain the typed characters, and Qt::MatchEndsWith the ones that end with the typed characters.

Currently, only these three modes are implemented. Setting filterMode to any other Qt::MatchFlag will issue a warning, and no action will be performed.

The default mode is Qt::MatchStartsWith.

This property was introduced in Qt 5.2.

Access functions:

Qt::MatchFlags 	filterMode() const

void 	setFilterMode(Qt::MatchFlags filterMode)

fn set_max_visible_items(&mut self, max_items: c_int)

C++ method: void QCompleter::setMaxVisibleItems(int maxItems)

C++ documentation:

This property holds the maximum allowed size on screen of the completer, measured in items.

By default, this property has a value of 7.

This property was introduced in Qt 4.6.

Access functions:

int 	maxVisibleItems() const

void 	setMaxVisibleItems(int maxItems)

fn set_model(&mut self, c: *mut AbstractItemModel)

C++ method: void QCompleter::setModel(QAbstractItemModel* c)

C++ documentation:

Sets the model which provides completions to model. The model can be list model or a tree model. If a model has been already previously set and it has the QCompleter as its parent, it is deleted.

For convenience, if model is a QFileSystemModel, QCompleter switches its caseSensitivity to Qt::CaseInsensitive on Windows and Qt::CaseSensitive on other platforms.

See also completionModel(), modelSorting, and Handling Tree Models.

fn set_model_sorting(&mut self, sorting: ModelSorting)

C++ method: void QCompleter::setModelSorting(QCompleter::ModelSorting sorting)

C++ documentation:

This property holds the way the model is sorted.

By default, no assumptions are made about the order of the items in the model that provides the completions.

If the model's data for the completionColumn() and completionRole() is sorted in ascending order, you can set this property to CaseSensitivelySortedModel or CaseInsensitivelySortedModel. On large models, this can lead to significant performance improvements because the completer object can then use a binary search algorithm instead of linear search algorithm.

The sort order (i.e ascending or descending order) of the model is determined dynamically by inspecting the contents of the model.

Note: The performance improvements described above cannot take place when the completer's caseSensitivity is different to the case sensitivity used by the model's when sorting.

Access functions:

ModelSorting 	modelSorting() const

void 	setModelSorting(ModelSorting sorting)

See also setCaseSensitivity() and QCompleter::ModelSorting.

fn set_object_name(&mut self, name: &String)

C++ method: void QCompleter::setObjectName(const QString& name)

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn set_parent(&mut self, arg1: *mut Object)

C++ method: void QCompleter::setParent(QObject* arg1)

C++ documentation:

Makes the object a child of parent.

See also parent() and children().

fn set_popup(&mut self, popup: *mut AbstractItemView)

C++ method: void QCompleter::setPopup(QAbstractItemView* popup)

C++ documentation:

Sets the popup used to display completions to popup. QCompleter takes ownership of the view.

A QListView is automatically created when the completionMode() is set to QCompleter::PopupCompletion or QCompleter::UnfilteredPopupCompletion. The default popup displays the completionColumn().

Ensure that this function is called before the view settings are modified. This is required since view's properties may require that a model has been set on the view (for example, hiding columns in the view requires a model to be set on the view).

See also popup().

fn set_property(&mut self, name: *const c_char, value: &Variant) -> bool

C++ method: bool QCompleter::setProperty(const char* name, const QVariant& value)

C++ documentation:

Sets the value of the object's name property to value.

If the property is defined in the class using Q_PROPERTY then true is returned on success and false otherwise. If the property is not defined using Q_PROPERTY, and therefore not listed in the meta-object, it is added as a dynamic property and false is returned.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

Dynamic properties can be queried again using property() and can be removed by setting the property value to an invalid QVariant. Changing the value of a dynamic property causes a QDynamicPropertyChangeEvent to be sent to the object.

Note: Dynamic properties starting with "_q_" are reserved for internal purposes.

See also property(), metaObject(), and dynamicPropertyNames().

fn set_user_data(&mut self, id: c_uint, data: *mut ObjectUserData)

C++ method: void QCompleter::setUserData(unsigned int id, QObjectUserData* data)

fn set_widget(&mut self, widget: *mut Widget)

C++ method: void QCompleter::setWidget(QWidget* widget)

C++ documentation:

Sets the widget for which completion are provided for to widget. This function is automatically called when a QCompleter is set on a QLineEdit using QLineEdit::setCompleter() or on a QComboBox using QComboBox::setCompleter(). The widget needs to be set explicitly when providing completions for custom widgets.

See also widget(), setModel(), and setPopup().

fn set_wrap_around(&mut self, wrap: bool)

C++ method: void QCompleter::setWrapAround(bool wrap)

C++ documentation:

This property holds the completions wrap around when navigating through items.

The default is true.

This property was introduced in Qt 4.3.

Access functions:

bool 	wrapAround() const

void 	setWrapAround(bool wrap)

fn signals_blocked(&self) -> bool

C++ method: bool QCompleter::signalsBlocked() const

C++ documentation:

Returns true if signals are blocked; otherwise returns false.

Signals are not blocked by default.

See also blockSignals().

fn split_path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CompleterSplitPathArgs<'largs>

C++ method: QCompleter::splitPath

This is an overloaded function. Available variants:

Rust arguments:
1) fn split_path(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn split_path(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string_list::StringList
C++ method: virtual QStringList QCompleter::splitPath(const QString& path) const

C++ documentation:

Splits the given path into strings that are used to match at each level in the model().

The default implementation of splitPath() splits a file system path based on QDir::separator() when the sourceModel() is a QFileSystemModel.

When used with list models, the first item in the returned list is used for matching.

See also pathFromIndex() and Handling Tree Models.

fn start_timer<'largs, Args>(&'largs mut self, args: Args) -> c_int where Args: CompleterStartTimerArgs<'largs>

C++ method: QCompleter::startTimer

This is an overloaded function. Available variants:

Rust arguments:
1) fn start_timer(&mut self, libc::c_int) -> libc::c_int
2) fn start_timer(&mut self, (libc::c_int, qt_core::qt::TimerType)) -> libc::c_int
C++ method: int QCompleter::startTimer(int interval, Qt::TimerType timerType = ?)

C++ documentation:

Starts a timer and returns a timer identifier, or returns zero if it could not start a timer.

A timer event will occur every interval milliseconds until killTimer() is called. If interval is 0, then the timer event occurs once every time there are no more window system events to process.

The virtual timerEvent() function is called with the QTimerEvent event parameter class when a timer event occurs. Reimplement this function to get timer events.

If multiple timers are running, the QTimerEvent::timerId() can be used to find out which timer was activated.

Example:

class MyObject : public QObject
{
    Q_OBJECT

public: MyObject(QObject *parent = 0);

protected: void timerEvent(QTimerEvent *event); };

MyObject::MyObject(QObject *parent) : QObject(parent) { startTimer(50); // 50-millisecond timer startTimer(1000); // 1-second timer startTimer(60000); // 1-minute timer }

void MyObject::timerEvent(QTimerEvent *event) { qDebug() << "Timer ID:" << event->timerId(); }
Note that QTimer's accuracy depends on the underlying operating system and hardware. The timerType argument allows you to customize the accuracy of the timer. See Qt::TimerType for information on the different timer types. Most platforms support an accuracy of 20 milliseconds; some provide more. If Qt is unable to deliver the requested number of timer events, it will silently discard some.

The QTimer class provides a high-level programming interface with single-shot timers and timer signals instead of events. There is also a QBasicTimer class that is more lightweight than QTimer and less clumsy than using timer IDs directly.

See also timerEvent(), killTimer(), and QTimer::singleShot().

fn thread(&self) -> *mut Thread

C++ method: QThread* QCompleter::thread() const

C++ documentation:

Returns the thread in which the object lives.

See also moveToThread().

fn tr<Args>(args: Args) -> Args::ReturnType where Args: CompleterTrArgs

C++ method: QCompleter::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QCompleter::tr(const char* s, const char* c, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: CompleterTrUtf8Args

C++ method: QCompleter::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QCompleter::trUtf8(const char* s, const char* c, int n)

fn user_data(&self, id: c_uint) -> *mut ObjectUserData

C++ method: QObjectUserData* QCompleter::userData(unsigned int id) const

fn widget(&self) -> *mut Widget

C++ method: QWidget* QCompleter::widget() const

C++ documentation:

Returns the widget for which the completer object is providing completions.

See also setWidget().

fn wrap_around(&self) -> bool

C++ method: bool QCompleter::wrapAround() const

C++ documentation:

This property holds the completions wrap around when navigating through items.

The default is true.

This property was introduced in Qt 4.3.

Access functions:

bool 	wrapAround() const

void 	setWrapAround(bool wrap)