Module qt_widgets::data_widget_mapper::overloading

Traits

DataWidgetMapperAddMappingArgs
This trait represents a set of arguments accepted by DataWidgetMapper::add_mapping method.

DataWidgetMapperConnectArgs
This trait represents a set of arguments accepted by DataWidgetMapper::connect method.

DataWidgetMapperConnectStaticArgs
This trait represents a set of arguments accepted by DataWidgetMapper::connect_static method.

DataWidgetMapperDestroyedArgs
This trait represents a set of arguments accepted by DataWidgetMapper::destroyed method.

DataWidgetMapperDisconnectArgs
This trait represents a set of arguments accepted by DataWidgetMapper::disconnect method.

DataWidgetMapperDisconnectStaticArgs
This trait represents a set of arguments accepted by DataWidgetMapper::disconnect_static method.

DataWidgetMapperDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by DataWidgetMapper::dynamic_property_names method.

DataWidgetMapperMappedPropertyNameArgs
This trait represents a set of arguments accepted by DataWidgetMapper::mapped_property_name method.

DataWidgetMapperNewArgs
This trait represents a set of arguments accepted by DataWidgetMapper::new method.

DataWidgetMapperObjectNameArgs
This trait represents a set of arguments accepted by DataWidgetMapper::object_name method.

DataWidgetMapperPropertyArgs
This trait represents a set of arguments accepted by DataWidgetMapper::property method.

DataWidgetMapperRootIndexArgs
This trait represents a set of arguments accepted by DataWidgetMapper::root_index method.

DataWidgetMapperStartTimerArgs
This trait represents a set of arguments accepted by DataWidgetMapper::start_timer method.

DataWidgetMapperTrArgs
This trait represents a set of arguments accepted by DataWidgetMapper::tr method.

DataWidgetMapperTrUtf8Args
This trait represents a set of arguments accepted by DataWidgetMapper::tr_utf8 method.