Module qt_widgets::desktop_widget::overloading

Traits

DesktopWidgetAccessibleDescriptionArgs
This trait represents a set of arguments accepted by DesktopWidget::accessible_description method.

DesktopWidgetAccessibleNameArgs
This trait represents a set of arguments accepted by DesktopWidget::accessible_name method.

DesktopWidgetActionsArgs
This trait represents a set of arguments accepted by DesktopWidget::actions method.

DesktopWidgetAvailableGeometryArgs
This trait represents a set of arguments accepted by DesktopWidget::available_geometry method.

DesktopWidgetBaseSizeArgs
This trait represents a set of arguments accepted by DesktopWidget::base_size method.

DesktopWidgetChildAtArgs
This trait represents a set of arguments accepted by DesktopWidget::child_at method.

DesktopWidgetChildrenRectArgs
This trait represents a set of arguments accepted by DesktopWidget::children_rect method.

DesktopWidgetChildrenRegionArgs
This trait represents a set of arguments accepted by DesktopWidget::children_region method.

DesktopWidgetConnectArgs
This trait represents a set of arguments accepted by DesktopWidget::connect method.

DesktopWidgetConnectStaticArgs
This trait represents a set of arguments accepted by DesktopWidget::connect_static method.

DesktopWidgetContentsMarginsArgs
This trait represents a set of arguments accepted by DesktopWidget::contents_margins method.

DesktopWidgetContentsRectArgs
This trait represents a set of arguments accepted by DesktopWidget::contents_rect method.

DesktopWidgetCreateWindowContainerArgs
This trait represents a set of arguments accepted by DesktopWidget::create_window_container method.

DesktopWidgetCursorArgs
This trait represents a set of arguments accepted by DesktopWidget::cursor method.

DesktopWidgetDestroyedArgs
This trait represents a set of arguments accepted by DesktopWidget::destroyed method.

DesktopWidgetDisconnectArgs
This trait represents a set of arguments accepted by DesktopWidget::disconnect method.

DesktopWidgetDisconnectStaticArgs
This trait represents a set of arguments accepted by DesktopWidget::disconnect_static method.

DesktopWidgetDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by DesktopWidget::dynamic_property_names method.

DesktopWidgetFontInfoArgs
This trait represents a set of arguments accepted by DesktopWidget::font_info method.

DesktopWidgetFontMetricsArgs
This trait represents a set of arguments accepted by DesktopWidget::font_metrics method.

DesktopWidgetFrameGeometryArgs
This trait represents a set of arguments accepted by DesktopWidget::frame_geometry method.

DesktopWidgetFrameSizeArgs
This trait represents a set of arguments accepted by DesktopWidget::frame_size method.

DesktopWidgetGrabArgs
This trait represents a set of arguments accepted by DesktopWidget::grab method.

DesktopWidgetGrabGestureArgs
This trait represents a set of arguments accepted by DesktopWidget::grab_gesture method.

DesktopWidgetGrabMouseArgs
This trait represents a set of arguments accepted by DesktopWidget::grab_mouse method.

DesktopWidgetGrabShortcutArgs
This trait represents a set of arguments accepted by DesktopWidget::grab_shortcut method.

DesktopWidgetInputMethodQueryArgs
This trait represents a set of arguments accepted by DesktopWidget::input_method_query method.

DesktopWidgetLocaleArgs
This trait represents a set of arguments accepted by DesktopWidget::locale method.

DesktopWidgetMapFromArgs
This trait represents a set of arguments accepted by DesktopWidget::map_from method.

DesktopWidgetMapFromGlobalArgs
This trait represents a set of arguments accepted by DesktopWidget::map_from_global method.

DesktopWidgetMapFromParentArgs
This trait represents a set of arguments accepted by DesktopWidget::map_from_parent method.

DesktopWidgetMapToArgs
This trait represents a set of arguments accepted by DesktopWidget::map_to method.

DesktopWidgetMapToGlobalArgs
This trait represents a set of arguments accepted by DesktopWidget::map_to_global method.

DesktopWidgetMapToParentArgs
This trait represents a set of arguments accepted by DesktopWidget::map_to_parent method.

DesktopWidgetMaskArgs
This trait represents a set of arguments accepted by DesktopWidget::mask method.

DesktopWidgetMaximumSizeArgs
This trait represents a set of arguments accepted by DesktopWidget::maximum_size method.

DesktopWidgetMinimumSizeArgs
This trait represents a set of arguments accepted by DesktopWidget::minimum_size method.

DesktopWidgetMinimumSizeHintArgs
This trait represents a set of arguments accepted by DesktopWidget::minimum_size_hint method.

DesktopWidgetMoveArgs
This trait represents a set of arguments accepted by DesktopWidget::move_ method.

DesktopWidgetNewArgs
This trait represents a set of arguments accepted by DesktopWidget::new method.

DesktopWidgetNormalGeometryArgs
This trait represents a set of arguments accepted by DesktopWidget::normal_geometry method.

DesktopWidgetObjectNameArgs
This trait represents a set of arguments accepted by DesktopWidget::object_name method.

DesktopWidgetPosArgs
This trait represents a set of arguments accepted by DesktopWidget::pos method.

DesktopWidgetPropertyArgs
This trait represents a set of arguments accepted by DesktopWidget::property method.

DesktopWidgetRectArgs
This trait represents a set of arguments accepted by DesktopWidget::rect method.

DesktopWidgetRenderArgs
This trait represents a set of arguments accepted by DesktopWidget::render method.

DesktopWidgetRepaintArgs
This trait represents a set of arguments accepted by DesktopWidget::repaint method.

DesktopWidgetResizeArgs
This trait represents a set of arguments accepted by DesktopWidget::resize method.

DesktopWidgetSaveGeometryArgs
This trait represents a set of arguments accepted by DesktopWidget::save_geometry method.

DesktopWidgetScreenArgs
This trait represents a set of arguments accepted by DesktopWidget::screen method.

DesktopWidgetScreenGeometryArgs
This trait represents a set of arguments accepted by DesktopWidget::screen_geometry method.

DesktopWidgetScreenNumberArgs
This trait represents a set of arguments accepted by DesktopWidget::screen_number method.

DesktopWidgetScrollArgs
This trait represents a set of arguments accepted by DesktopWidget::scroll method.

DesktopWidgetSetAttributeArgs
This trait represents a set of arguments accepted by DesktopWidget::set_attribute method.

DesktopWidgetSetBaseSizeArgs
This trait represents a set of arguments accepted by DesktopWidget::set_base_size method.

DesktopWidgetSetContentsMarginsArgs
This trait represents a set of arguments accepted by DesktopWidget::set_contents_margins method.

DesktopWidgetSetFixedSizeArgs
This trait represents a set of arguments accepted by DesktopWidget::set_fixed_size method.

DesktopWidgetSetFocusArgs
This trait represents a set of arguments accepted by DesktopWidget::set_focus method.

DesktopWidgetSetGeometryArgs
This trait represents a set of arguments accepted by DesktopWidget::set_geometry method.

DesktopWidgetSetMaskArgs
This trait represents a set of arguments accepted by DesktopWidget::set_mask method.

DesktopWidgetSetMaximumSizeArgs
This trait represents a set of arguments accepted by DesktopWidget::set_maximum_size method.

DesktopWidgetSetMinimumSizeArgs
This trait represents a set of arguments accepted by DesktopWidget::set_minimum_size method.

DesktopWidgetSetParentArgs
This trait represents a set of arguments accepted by DesktopWidget::set_parent method.

DesktopWidgetSetShortcutAutoRepeatArgs
This trait represents a set of arguments accepted by DesktopWidget::set_shortcut_auto_repeat method.

DesktopWidgetSetShortcutEnabledArgs
This trait represents a set of arguments accepted by DesktopWidget::set_shortcut_enabled method.

DesktopWidgetSetSizeIncrementArgs
This trait represents a set of arguments accepted by DesktopWidget::set_size_increment method.

DesktopWidgetSetSizePolicyArgs
This trait represents a set of arguments accepted by DesktopWidget::set_size_policy method.

DesktopWidgetSizeArgs
This trait represents a set of arguments accepted by DesktopWidget::size method.

DesktopWidgetSizeHintArgs
This trait represents a set of arguments accepted by DesktopWidget::size_hint method.

DesktopWidgetSizeIncrementArgs
This trait represents a set of arguments accepted by DesktopWidget::size_increment method.

DesktopWidgetSizePolicyArgs
This trait represents a set of arguments accepted by DesktopWidget::size_policy method.

DesktopWidgetStartTimerArgs
This trait represents a set of arguments accepted by DesktopWidget::start_timer method.

DesktopWidgetStatusTipArgs
This trait represents a set of arguments accepted by DesktopWidget::status_tip method.

DesktopWidgetStyleSheetArgs
This trait represents a set of arguments accepted by DesktopWidget::style_sheet method.

DesktopWidgetToolTipArgs
This trait represents a set of arguments accepted by DesktopWidget::tool_tip method.

DesktopWidgetTrArgs
This trait represents a set of arguments accepted by DesktopWidget::tr method.

DesktopWidgetTrUtf8Args
This trait represents a set of arguments accepted by DesktopWidget::tr_utf8 method.

DesktopWidgetUpdateArgs
This trait represents a set of arguments accepted by DesktopWidget::update method.

DesktopWidgetVisibleRegionArgs
This trait represents a set of arguments accepted by DesktopWidget::visible_region method.

DesktopWidgetWhatsThisArgs
This trait represents a set of arguments accepted by DesktopWidget::whats_this method.

DesktopWidgetWindowFilePathArgs
This trait represents a set of arguments accepted by DesktopWidget::window_file_path method.

DesktopWidgetWindowIconArgs
This trait represents a set of arguments accepted by DesktopWidget::window_icon method.

DesktopWidgetWindowIconTextArgs
This trait represents a set of arguments accepted by DesktopWidget::window_icon_text method.

DesktopWidgetWindowRoleArgs
This trait represents a set of arguments accepted by DesktopWidget::window_role method.

DesktopWidgetWindowTitleArgs
This trait represents a set of arguments accepted by DesktopWidget::window_title method.