Enum qt_widgets::file_dialog::Option

pub enum Option {
    ShowDirsOnly,
    DontResolveSymlinks,
    DontConfirmOverwrite,
    DontUseSheet,
    DontUseNativeDialog,
    ReadOnly,
    HideNameFilterDetails,
    DontUseCustomDirectoryIcons,
}

C++ type: QFileDialog::Option

C++ documentation:

The Options type is a typedef for QFlags<Option>. It stores an OR combination of Option values.

Variants
ShowDirsOnly

Only show directories in the file dialog. By default both files and directories are shown. (Valid only in the Directory file mode.) (C++ enum variant: ShowDirsOnly = 1)

DontResolveSymlinks

Don't resolve symlinks in the file dialog. By default symlinks are resolved. (C++ enum variant: DontResolveSymlinks = 2)

DontConfirmOverwrite

Don't ask for confirmation if an existing file is selected. By default confirmation is requested. (C++ enum variant: DontConfirmOverwrite = 4)

DontUseSheet

In previous versions of Qt, the static functions would create a sheet by default if the static function was given a parent. This is no longer supported and does nothing in Qt 4.5, The static functions will always be an application modal dialog. If you want to use sheets, use QFileDialog::open() instead. (C++ enum variant: DontUseSheet = 8)

DontUseNativeDialog

Don't use the native file dialog. By default, the native file dialog is used unless you use a subclass of QFileDialog that contains the Q_OBJECT macro, or the platform does not have a native dialog of the type that you require. (C++ enum variant: DontUseNativeDialog = 16)

ReadOnly

Indicates that the model is readonly. (C++ enum variant: ReadOnly = 32)

HideNameFilterDetails

Indicates if the file name filter details are hidden or not. (C++ enum variant: HideNameFilterDetails = 64)

DontUseCustomDirectoryIcons

Always use the default directory icon. Some platforms allow the user to set a different icon. Custom icon lookup cause a big performance impact over network or removable drives. Setting this will enable the QFileIconProvider::DontUseCustomDirectoryIcons option in the icon provider. This enum value was added in Qt 5.2. (C++ enum variant: DontUseCustomDirectoryIcons = 128)