Struct qt_widgets::file_icon_provider::FileIconProvider

pub struct FileIconProvider {
    // some fields omitted
}

C++ type: QFileIconProvider

C++ documentation:

The QFileIconProvider class provides file icons for the QDirModel and the QFileSystemModel classes.

Methods

impl FileIconProvider

fn icon<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileIconProviderIconArgs<'largs>

C++ method: QFileIconProvider::icon

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn icon(&self, (::file_icon_provider::IconType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_gui::icon::Icon>
2) fn icon(&self, (::file_icon_provider::IconType, cpp_utils::AsStruct)) -> qt_gui::icon::Icon
C++ method: virtual QIcon QFileIconProvider::icon(QFileIconProvider::IconType type) const

C++ documentation:

Returns an icon set for the given type.

Variant 2

Rust arguments:
1) fn icon(&self, (&qt_core::file_info::FileInfo, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_gui::icon::Icon>
2) fn icon(&self, (&qt_core::file_info::FileInfo, cpp_utils::AsStruct)) -> qt_gui::icon::Icon
C++ method: virtual QIcon QFileIconProvider::icon(const QFileInfo& info) const

C++ documentation:

Returns an icon for the file described by info.

fn new<Args>(args: Args) -> Args::ReturnType where Args: FileIconProviderNewArgs

C++ method: QFileIconProvider::QFileIconProvider

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::file_icon_provider::FileIconProvider
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::file_icon_provider::FileIconProvider>
C++ method: [constructor] void QFileIconProvider::QFileIconProvider()

fn options(&self) -> Flags<Option>

C++ method: QFlags<QFileIconProvider::Option> QFileIconProvider::options() const

C++ documentation:

Returns all the options that affect the icon provider. By default, all options are disabled.

This function was introduced in Qt 5.2.

See also setOptions().

fn set_options(&mut self, options: Flags<Option>)

C++ method: void QFileIconProvider::setOptions(QFlags<QFileIconProvider::Option> options)

C++ documentation:

Sets options that affect the icon provider.

This function was introduced in Qt 5.2.

See also options().

fn type_<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileIconProviderTypeArgs<'largs>

C++ method: QFileIconProvider::type

This is an overloaded function. Available variants:

Rust arguments:
1) fn type_(&self, (&qt_core::file_info::FileInfo, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn type_(&self, (&qt_core::file_info::FileInfo, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QFileIconProvider::type(const QFileInfo& info) const

C++ documentation:

Returns the type of the file described by info.