Enum qt_widgets::font_dialog::FontDialogOption

pub enum FontDialogOption {
    NoButtons,
    DontUseNativeDialog,
    ScalableFonts,
    NonScalableFonts,
    MonospacedFonts,
    ProportionalFonts,
}

C++ type: QFontDialog::FontDialogOption

C++ documentation:

This enum specifies various options that affect the look and feel of a font dialog.

For instance, it allows to specify which type of font should be displayed. If none are specified all fonts available will be listed.

Note that the font filtering options might not be supported on some platforms (e.g. Mac). They are always supported by the non native dialog (used on Windows or Linux).

This enum was introduced or modified in Qt 4.5.

The FontDialogOptions type is a typedef for QFlags<FontDialogOption>. It stores an OR combination of FontDialogOption values.

See also options, setOption(), and testOption().

Variants
NoButtons

Don't display OK and Cancel buttons. (Useful for "live dialogs".) (C++ enum variant: NoButtons = 1)

DontUseNativeDialog

Use Qt's standard font dialog on the Mac instead of Apple's native font panel. (Currently, the native dialog is never used, but this is likely to change in future Qt releases.) (C++ enum variant: DontUseNativeDialog = 2)

ScalableFonts

Show scalable fonts (C++ enum variant: ScalableFonts = 4)

NonScalableFonts

Show non scalable fonts (C++ enum variant: NonScalableFonts = 8)

MonospacedFonts

Show monospaced fonts (C++ enum variant: MonospacedFonts = 16)

ProportionalFonts

Show proportional fonts (C++ enum variant: ProportionalFonts = 32)