Module qt_widgets::gesture::overloading

Traits

GestureConnectArgs
This trait represents a set of arguments accepted by Gesture::connect method.

GestureConnectStaticArgs
This trait represents a set of arguments accepted by Gesture::connect_static method.

GestureDestroyedArgs
This trait represents a set of arguments accepted by Gesture::destroyed method.

GestureDisconnectArgs
This trait represents a set of arguments accepted by Gesture::disconnect method.

GestureDisconnectStaticArgs
This trait represents a set of arguments accepted by Gesture::disconnect_static method.

GestureDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Gesture::dynamic_property_names method.

GestureHotSpotArgs
This trait represents a set of arguments accepted by Gesture::hot_spot method.

GestureNewArgs
This trait represents a set of arguments accepted by Gesture::new method.

GestureObjectNameArgs
This trait represents a set of arguments accepted by Gesture::object_name method.

GesturePropertyArgs
This trait represents a set of arguments accepted by Gesture::property method.

GestureStartTimerArgs
This trait represents a set of arguments accepted by Gesture::start_timer method.

GestureTrArgs
This trait represents a set of arguments accepted by Gesture::tr method.

GestureTrUtf8Args
This trait represents a set of arguments accepted by Gesture::tr_utf8 method.

OpBitOrArgs
This trait represents a set of arguments accepted by op_bit_or method.

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.