Module qt_widgets::gesture_event::overloading

Traits

GestureEventAcceptArgs
This trait represents a set of arguments accepted by GestureEvent::accept method.

GestureEventActiveGesturesArgs
This trait represents a set of arguments accepted by GestureEvent::active_gestures method.

GestureEventCanceledGesturesArgs
This trait represents a set of arguments accepted by GestureEvent::canceled_gestures method.

GestureEventGesturesArgs
This trait represents a set of arguments accepted by GestureEvent::gestures method.

GestureEventIgnoreArgs
This trait represents a set of arguments accepted by GestureEvent::ignore method.

GestureEventIsAcceptedArgs
This trait represents a set of arguments accepted by GestureEvent::is_accepted method.

GestureEventMapToGraphicsSceneArgs
This trait represents a set of arguments accepted by GestureEvent::map_to_graphics_scene method.

GestureEventNewArgs
This trait represents a set of arguments accepted by GestureEvent::new method.

GestureEventRegisterEventTypeArgs
This trait represents a set of arguments accepted by GestureEvent::register_event_type method.

GestureEventSetAcceptedArgs
This trait represents a set of arguments accepted by GestureEvent::set_accepted method.