Struct qt_widgets::gesture_event::GestureEvent

pub struct GestureEvent {
    // some fields omitted
}

C++ type: QGestureEvent

C++ documentation:

The QGestureEvent class provides the description of triggered gestures.

The QGestureEvent class contains a list of gestures, which can be obtained using the gestures() function.

The gestures are either active or canceled. A list of those that are currently being executed can be obtained using the activeGestures() function. A list of those which were previously active and have been canceled can be accessed using the canceledGestures() function. A gesture might be canceled if the current window loses focus, for example, or because of a timeout, or for other reasons.

If the event handler does not accept the event by calling the generic QEvent::accept() function, all individual QGesture object that were not accepted and in the Qt::GestureStarted state will be propagated up the parent widget chain until a widget accepts them individually, by calling QGestureEvent::accept() for each of them, or an event filter consumes the event.

Methods

impl GestureEvent

fn accept<'largs, Args>(&'largs mut self, args: Args) where Args: GestureEventAcceptArgs<'largs>

C++ method: QGestureEvent::accept

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn accept(&mut self, ()) -> ()
C++ method: void QGestureEvent::accept()

C++ documentation:

Sets the accept flag of the event object, the equivalent of calling setAccepted(true).

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget.

See also ignore().

Variant 2

Rust arguments: fn accept(&mut self, *mut ::gesture::Gesture) -> ()
C++ method: void QGestureEvent::accept(QGesture* arg1)

C++ documentation:

Sets the accept flag of the given gesture object, the equivalent of calling setAccepted(gesture, true).

Setting the accept flag indicates that the event receiver wants the gesture. Unwanted gestures may be propagated to the parent widget.

See also QGestureEvent::ignore().

Variant 3

Rust arguments: fn accept(&mut self, qt_core::qt::GestureType) -> ()
C++ method: void QGestureEvent::accept(Qt::GestureType arg1)

C++ documentation:

Sets the accept flag of the given gestureType, the equivalent of calling setAccepted(gestureType, true).

Setting the accept flag indicates that the event receiver wants the gesture. Unwanted gestures may be propagated to the parent widget.

See also QGestureEvent::ignore().

fn active_gestures<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GestureEventActiveGesturesArgs<'largs>

C++ method: QGestureEvent::activeGestures

This is an overloaded function. Available variants:

Rust arguments:
1) fn active_gestures(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListGestureMutPtr>
2) fn active_gestures(&self, cpp_utils::AsStruct) -> ::list::ListGestureMutPtr
C++ method: QList<QGesture*> QGestureEvent::activeGestures() const

C++ documentation:

Returns a list of active (not canceled) gestures.

fn canceled_gestures<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GestureEventCanceledGesturesArgs<'largs>

C++ method: QGestureEvent::canceledGestures

This is an overloaded function. Available variants:

Rust arguments:
1) fn canceled_gestures(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListGestureMutPtr>
2) fn canceled_gestures(&self, cpp_utils::AsStruct) -> ::list::ListGestureMutPtr
C++ method: QList<QGesture*> QGestureEvent::canceledGestures() const

C++ documentation:

Returns a list of canceled gestures.

fn gesture(&self, type_: GestureType) -> *mut Gesture

C++ method: QGesture* QGestureEvent::gesture(Qt::GestureType type) const

C++ documentation:

Returns a gesture object by type.

fn gestures<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GestureEventGesturesArgs<'largs>

C++ method: QGestureEvent::gestures

This is an overloaded function. Available variants:

Rust arguments:
1) fn gestures(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListGestureMutPtr>
2) fn gestures(&self, cpp_utils::AsStruct) -> ::list::ListGestureMutPtr
C++ method: QList<QGesture*> QGestureEvent::gestures() const

C++ documentation:

Returns all gestures that are delivered in the event.

fn ignore<'largs, Args>(&'largs mut self, args: Args) where Args: GestureEventIgnoreArgs<'largs>

C++ method: QGestureEvent::ignore

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn ignore(&mut self, ()) -> ()
C++ method: void QGestureEvent::ignore()

C++ documentation:

Clears the accept flag parameter of the event object, the equivalent of calling setAccepted(false).

Clearing the accept parameter indicates that the event receiver does not want the event. Unwanted events might be propagated to the parent widget.

See also accept().

Variant 2

Rust arguments: fn ignore(&mut self, *mut ::gesture::Gesture) -> ()
C++ method: void QGestureEvent::ignore(QGesture* arg1)

C++ documentation:

Clears the accept flag parameter of the given gesture object, the equivalent of calling setAccepted(gesture, false).

Clearing the accept flag indicates that the event receiver does not want the gesture. Unwanted gestures may be propagated to the parent widget.

See also QGestureEvent::accept().

Variant 3

Rust arguments: fn ignore(&mut self, qt_core::qt::GestureType) -> ()
C++ method: void QGestureEvent::ignore(Qt::GestureType arg1)

C++ documentation:

Clears the accept flag parameter of the given gestureType, the equivalent of calling setAccepted(gesture, false).

Clearing the accept flag indicates that the event receiver does not want the gesture. Unwanted gestures may be propgated to the parent widget.

See also QGestureEvent::accept().

fn is_accepted<'largs, Args>(&'largs self, args: Args) -> bool where Args: GestureEventIsAcceptedArgs<'largs>

C++ method: QGestureEvent::isAccepted

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn is_accepted(&self, ()) -> bool
C++ method: bool QGestureEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

Variant 2

Rust arguments: fn is_accepted(&self, *mut ::gesture::Gesture) -> bool
C++ method: bool QGestureEvent::isAccepted(QGesture* arg1) const

C++ documentation:

Returns true if the gesture is accepted; otherwise returns false.

Variant 3

Rust arguments: fn is_accepted(&self, qt_core::qt::GestureType) -> bool
C++ method: bool QGestureEvent::isAccepted(Qt::GestureType arg1) const

C++ documentation:

Returns true if the gesture of type gestureType is accepted; otherwise returns false.

fn map_to_graphics_scene<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GestureEventMapToGraphicsSceneArgs<'largs>

C++ method: QGestureEvent::mapToGraphicsScene

This is an overloaded function. Available variants:

Rust arguments:
1) fn map_to_graphics_scene(&self, (&qt_core::point_f::PointF, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn map_to_graphics_scene(&self, (&qt_core::point_f::PointF, cpp_utils::AsStruct)) -> qt_core::point_f::PointF
C++ method: QPointF QGestureEvent::mapToGraphicsScene(const QPointF& gesturePoint) const

C++ documentation:

Returns the scene-local coordinates if the gesturePoint is inside a graphics view.

This functional might be useful when the gesture event is delivered to a QGraphicsObject to translate a point in screen coordinates to scene-local coordinates.

See also QPointF::isNull().

fn new<Args>(args: Args) -> Args::ReturnType where Args: GestureEventNewArgs

C++ method: QGestureEvent::QGestureEvent

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((&::list::ListGestureMutPtr, cpp_utils::AsStruct)) -> ::gesture_event::GestureEvent
2) fn new((&::list::ListGestureMutPtr, cpp_utils::AsBox)) -> cpp_utils::CppBox<::gesture_event::GestureEvent>
C++ method: [constructor] void QGestureEvent::QGestureEvent(const QList<QGesture*>& gestures)

fn register_event_type<Args>(args: Args) -> c_int where Args: GestureEventRegisterEventTypeArgs

C++ method: QGestureEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QGestureEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn set_accepted<'largs, Args>(&'largs mut self, args: Args) where Args: GestureEventSetAcceptedArgs<'largs>

C++ method: QGestureEvent::setAccepted

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_accepted(&mut self, (*mut ::gesture::Gesture, bool)) -> ()
C++ method: void QGestureEvent::setAccepted(QGesture* arg1, bool arg2)

C++ documentation:

Sets the accept flag of the given gesture object to the specified value.

Setting the accept flag indicates that the event receiver wants the gesture. Unwanted gestures may be propagated to the parent widget.

By default, gestures in events of type QEvent::Gesture are accepted, and gestures in QEvent::GestureOverride events are ignored.

For convenience, the accept flag can also be set with accept(gesture), and cleared with ignore(gesture).

See also isAccepted().

Variant 2

Rust arguments: fn set_accepted(&mut self, (qt_core::qt::GestureType, bool)) -> ()
C++ method: void QGestureEvent::setAccepted(Qt::GestureType arg1, bool arg2)

C++ documentation:

Sets the accept flag of the given gestureType object to the specified value.

Setting the accept flag indicates that the event receiver wants to receive gestures of the specified type, gestureType. Unwanted gestures may be propagated to the parent widget.

By default, gestures in events of type QEvent::Gesture are accepted, and gestures in QEvent::GestureOverride events are ignored.

For convenience, the accept flag can also be set with accept(gestureType), and cleared with ignore(gestureType).

Variant 3

Rust arguments: fn set_accepted(&mut self, bool) -> ()
C++ method: void QGestureEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn set_widget(&mut self, widget: *mut Widget)

C++ method: void QGestureEvent::setWidget(QWidget* widget)

fn spontaneous(&self) -> bool

C++ method: bool QGestureEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn type_(&self) -> Type

C++ method: QEvent::Type QGestureEvent::type() const

C++ documentation:

Returns the event type.

fn widget(&self) -> *mut Widget

C++ method: QWidget* QGestureEvent::widget() const

C++ documentation:

Returns the widget on which the event occurred.