Enum qt_widgets::gesture_recognizer::ResultFlag

pub enum ResultFlag {
    Ignore,
    MayBeGesture,
    TriggerGesture,
    FinishGesture,
    CancelGesture,
    ResultStateMask,
    ConsumeEventHint,
    ResultHintMask,
}

C++ type: QGestureRecognizer::ResultFlag

C++ documentation:

This enum describes the result of the current event filtering step in a gesture recognizer state machine.

The result consists of a state value (one of Ignore, MayBeGesture, TriggerGesture, FinishGesture, CancelGesture) and an optional hint (ConsumeEventHint).

The Result type is a typedef for QFlags<ResultFlag>. It stores an OR combination of ResultFlag values.

See also QGestureRecognizer::recognize().

Variants
Ignore

The event does not change the state of the recognizer. (C++ enum variant: Ignore = 1)

MayBeGesture

The event changed the internal state of the recognizer, but it isn't clear yet if it is a gesture or not. The recognizer needs to filter more events to decide. Gesture recognizers in the MayBeGesture state may be reset automatically if they take too long to recognize gestures. (C++ enum variant: MayBeGesture = 2)

TriggerGesture

The gesture has been triggered and the appropriate QGesture object will be delivered to the target as a part of a QGestureEvent. (C++ enum variant: TriggerGesture = 4)

FinishGesture

The gesture has been finished successfully and the appropriate QGesture object will be delivered to the target as a part of a QGestureEvent. (C++ enum variant: FinishGesture = 8)

CancelGesture

The event made it clear that it is not a gesture. If the gesture recognizer was in GestureTriggered state before, then the gesture is canceled and the appropriate QGesture object will be delivered to the target as a part of a QGestureEvent. (C++ enum variant: CancelGesture = 16)

ResultStateMask

C++ enum variant: ResultState_Mask = 255

ConsumeEventHint

This hint specifies that the gesture framework should consume the filtered event and not deliver it to the receiver. (C++ enum variant: ConsumeEventHint = 256)

ResultHintMask

C++ enum variant: ResultHint_Mask = 65280