Struct qt_widgets::gesture_recognizer::GestureRecognizer

pub struct GestureRecognizer {
    // some fields omitted
}

C++ type: QGestureRecognizer

C++ documentation:

The QGestureRecognizer class provides the infrastructure for gesture recognition.

Gesture recognizers are responsible for creating and managing QGesture objects and monitoring input events sent to QWidget and QGraphicsObject subclasses. QGestureRecognizer is the base class for implementing custom gestures.

Developers that only need to provide gesture recognition for standard gestures do not need to use this class directly. Instances will be created behind the scenes by the framework.

For an overview of gesture handling in Qt and information on using gestures in your applications, see the Gestures in Widgets and Graphics View document.

Methods

impl GestureRecognizer

fn create(&mut self, target: *mut Object) -> *mut Gesture

C++ method: virtual QGesture* QGestureRecognizer::create(QObject* target)

C++ documentation:

This function is called by Qt to create a new QGesture object for the given target (QWidget or QGraphicsObject).

Reimplement this function to create a custom QGesture-derived gesture object if necessary.

The application takes ownership of the created gesture object.

fn recognize(&mut self, state: *mut Gesture, watched: *mut Object, event: *mut Event) -> Flags<ResultFlag>

C++ method: pure virtual QFlags<QGestureRecognizer::ResultFlag> QGestureRecognizer::recognize(QGesture* state, QObject* watched, QEvent* event)

C++ documentation:

Handles the given event for the watched object, updating the state of the gesture object as required, and returns a suitable result for the current recognition step.

This function is called by the framework to allow the recognizer to filter input events dispatched to QWidget or QGraphicsObject instances that it is monitoring.

The result reflects how much of the gesture has been recognized. The state of the gesture object is set depending on the result.

See also QGestureRecognizer::Result.

fn register_recognizer(recognizer: *mut GestureRecognizer) -> GestureType

C++ method: static Qt::GestureType QGestureRecognizer::registerRecognizer(QGestureRecognizer* recognizer)

C++ documentation:

Registers the given recognizer in the gesture framework and returns a gesture ID for it.

The application takes ownership of the recognizer and returns the gesture type ID associated with it. For gesture recognizers which handle custom QGesture objects (i.e., those which return Qt::CustomGesture in a QGesture::gestureType() function) the return value is a generated gesture ID with the Qt::CustomGesture flag set.

See also unregisterRecognizer(), QGestureRecognizer::create(), and QGesture.

fn reset(&mut self, state: *mut Gesture)

C++ method: virtual void QGestureRecognizer::reset(QGesture* state)

C++ documentation:

This function is called by the framework to reset a given gesture.

Reimplement this function to implement additional requirements for custom QGesture objects. This may be necessary if you implement a custom QGesture whose properties need special handling when the gesture is reset.

fn unregister_recognizer(type_: GestureType)

C++ method: static void QGestureRecognizer::unregisterRecognizer(Qt::GestureType type)

C++ documentation:

Unregisters all gesture recognizers of the specified type.

See also registerRecognizer().