Module qt_widgets::graphics_colorize_effect::overloading

Traits

GraphicsColorizeEffectBoundingRectArgs
This trait represents a set of arguments accepted by GraphicsColorizeEffect::bounding_rect method.

GraphicsColorizeEffectBoundingRectForArgs
This trait represents a set of arguments accepted by GraphicsColorizeEffect::bounding_rect_for method.

GraphicsColorizeEffectColorArgs
This trait represents a set of arguments accepted by GraphicsColorizeEffect::color method.

GraphicsColorizeEffectConnectArgs
This trait represents a set of arguments accepted by GraphicsColorizeEffect::connect method.

GraphicsColorizeEffectConnectStaticArgs
This trait represents a set of arguments accepted by GraphicsColorizeEffect::connect_static method.

GraphicsColorizeEffectDestroyedArgs
This trait represents a set of arguments accepted by GraphicsColorizeEffect::destroyed method.

GraphicsColorizeEffectDisconnectArgs
This trait represents a set of arguments accepted by GraphicsColorizeEffect::disconnect method.

GraphicsColorizeEffectDisconnectStaticArgs
This trait represents a set of arguments accepted by GraphicsColorizeEffect::disconnect_static method.

GraphicsColorizeEffectDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by GraphicsColorizeEffect::dynamic_property_names method.

GraphicsColorizeEffectNewArgs
This trait represents a set of arguments accepted by GraphicsColorizeEffect::new method.

GraphicsColorizeEffectObjectNameArgs
This trait represents a set of arguments accepted by GraphicsColorizeEffect::object_name method.

GraphicsColorizeEffectPropertyArgs
This trait represents a set of arguments accepted by GraphicsColorizeEffect::property method.

GraphicsColorizeEffectStartTimerArgs
This trait represents a set of arguments accepted by GraphicsColorizeEffect::start_timer method.

GraphicsColorizeEffectTrArgs
This trait represents a set of arguments accepted by GraphicsColorizeEffect::tr method.

GraphicsColorizeEffectTrUtf8Args
This trait represents a set of arguments accepted by GraphicsColorizeEffect::tr_utf8 method.