Module qt_widgets::graphics_drop_shadow_effect::overloading

Traits

GraphicsDropShadowEffectBoundingRectArgs
This trait represents a set of arguments accepted by GraphicsDropShadowEffect::bounding_rect method.

GraphicsDropShadowEffectBoundingRectForArgs
This trait represents a set of arguments accepted by GraphicsDropShadowEffect::bounding_rect_for method.

GraphicsDropShadowEffectColorArgs
This trait represents a set of arguments accepted by GraphicsDropShadowEffect::color method.

GraphicsDropShadowEffectConnectArgs
This trait represents a set of arguments accepted by GraphicsDropShadowEffect::connect method.

GraphicsDropShadowEffectConnectStaticArgs
This trait represents a set of arguments accepted by GraphicsDropShadowEffect::connect_static method.

GraphicsDropShadowEffectDestroyedArgs
This trait represents a set of arguments accepted by GraphicsDropShadowEffect::destroyed method.

GraphicsDropShadowEffectDisconnectArgs
This trait represents a set of arguments accepted by GraphicsDropShadowEffect::disconnect method.

GraphicsDropShadowEffectDisconnectStaticArgs
This trait represents a set of arguments accepted by GraphicsDropShadowEffect::disconnect_static method.

GraphicsDropShadowEffectDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by GraphicsDropShadowEffect::dynamic_property_names method.

GraphicsDropShadowEffectNewArgs
This trait represents a set of arguments accepted by GraphicsDropShadowEffect::new method.

GraphicsDropShadowEffectObjectNameArgs
This trait represents a set of arguments accepted by GraphicsDropShadowEffect::object_name method.

GraphicsDropShadowEffectOffsetArgs
This trait represents a set of arguments accepted by GraphicsDropShadowEffect::offset method.

GraphicsDropShadowEffectPropertyArgs
This trait represents a set of arguments accepted by GraphicsDropShadowEffect::property method.

GraphicsDropShadowEffectSetOffsetArgs
This trait represents a set of arguments accepted by GraphicsDropShadowEffect::set_offset method.

GraphicsDropShadowEffectStartTimerArgs
This trait represents a set of arguments accepted by GraphicsDropShadowEffect::start_timer method.

GraphicsDropShadowEffectTrArgs
This trait represents a set of arguments accepted by GraphicsDropShadowEffect::tr method.

GraphicsDropShadowEffectTrUtf8Args
This trait represents a set of arguments accepted by GraphicsDropShadowEffect::tr_utf8 method.