Module qt_widgets::graphics_effect::overloading

Traits

GraphicsEffectBoundingRectArgs
This trait represents a set of arguments accepted by GraphicsEffect::bounding_rect method.

GraphicsEffectBoundingRectForArgs
This trait represents a set of arguments accepted by GraphicsEffect::bounding_rect_for method.

GraphicsEffectConnectArgs
This trait represents a set of arguments accepted by GraphicsEffect::connect method.

GraphicsEffectConnectStaticArgs
This trait represents a set of arguments accepted by GraphicsEffect::connect_static method.

GraphicsEffectDestroyedArgs
This trait represents a set of arguments accepted by GraphicsEffect::destroyed method.

GraphicsEffectDisconnectArgs
This trait represents a set of arguments accepted by GraphicsEffect::disconnect method.

GraphicsEffectDisconnectStaticArgs
This trait represents a set of arguments accepted by GraphicsEffect::disconnect_static method.

GraphicsEffectDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by GraphicsEffect::dynamic_property_names method.

GraphicsEffectObjectNameArgs
This trait represents a set of arguments accepted by GraphicsEffect::object_name method.

GraphicsEffectPropertyArgs
This trait represents a set of arguments accepted by GraphicsEffect::property method.

GraphicsEffectStartTimerArgs
This trait represents a set of arguments accepted by GraphicsEffect::start_timer method.

GraphicsEffectTrArgs
This trait represents a set of arguments accepted by GraphicsEffect::tr method.

GraphicsEffectTrUtf8Args
This trait represents a set of arguments accepted by GraphicsEffect::tr_utf8 method.

OpBitOrArgs
This trait represents a set of arguments accepted by op_bit_or method.