Struct qt_widgets::graphics_grid_layout::GraphicsGridLayout

pub struct GraphicsGridLayout {
    // some fields omitted
}

C++ type: QGraphicsGridLayout

C++ documentation:

The QGraphicsGridLayout class provides a grid layout for managing widgets in Graphics View.

The most common way to use QGraphicsGridLayout is to construct an object on the heap with no parent, add widgets and layouts by calling addItem(), and finally assign the layout to a widget by calling QGraphicsWidget::setLayout(). QGraphicsGridLayout automatically computes the dimensions of the grid as you add items.

QGraphicsScene scene;
QGraphicsWidget textEdit = scene.addWidget(new QTextEdit);
QGraphicsWidget pushButton = scene.addWidget(new QPushButton);

QGraphicsGridLayout *layout = new QGraphicsGridLayout; layout->addItem(textEdit, 0, 0); layout->addItem(pushButton, 0, 1);

QGraphicsWidget *form = new QGraphicsWidget; form->setLayout(layout); scene.addItem(form);
The layout takes ownership of the items. In some cases when the layout item also inherits from QGraphicsItem (such as QGraphicsWidget) there will be a ambiguity in ownership because the layout item belongs to two ownership hierarchies. See the documentation of QGraphicsLayoutItem::setOwnedByLayout() how to handle this. You can access each item in the layout by calling count() and itemAt(). Calling removeAt() will remove an item from the layout, without destroying it.

Methods

impl GraphicsGridLayout

fn activate(&mut self)

C++ method: void QGraphicsGridLayout::activate()

C++ documentation:

Activates the layout, causing all items in the layout to be immediately rearranged. This function is based on calling count() and itemAt(), and then calling setGeometry() on all items sequentially. When activated, the layout will adjust its geometry to its parent's contentsRect(). The parent will then invalidate any layout of its own.

If called in sequence or recursively, e.g., by one of the arranged items in response to being resized, this function will do nothing.

Note that the layout is free to use geometry caching to optimize this process. To forcefully invalidate any such cache, you can call invalidate() before calling activate().

See also invalidate().

fn add_item<'largs, Args>(&'largs mut self, args: Args) where Args: GraphicsGridLayoutAddItemArgs<'largs>

C++ method: QGraphicsGridLayout::addItem

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn add_item(&mut self, (*mut ::graphics_layout_item::GraphicsLayoutItem, libc::c_int, libc::c_int)) -> ()
2) fn add_item(&mut self, (*mut ::graphics_layout_item::GraphicsLayoutItem, libc::c_int, libc::c_int, qt_core::flags::Flags<qt_core::qt::AlignmentFlag>)) -> ()
C++ method: void QGraphicsGridLayout::addItem(QGraphicsLayoutItem* item, int row, int column, QFlags<Qt::AlignmentFlag> alignment = ?)

C++ documentation:

Adds item to the grid on row and column. You can specify an optional alignment for item.

Variant 2

Rust arguments:
1) fn add_item(&mut self, (*mut ::graphics_layout_item::GraphicsLayoutItem, libc::c_int, libc::c_int, libc::c_int, libc::c_int)) -> ()
2) fn add_item(&mut self, (*mut ::graphics_layout_item::GraphicsLayoutItem, libc::c_int, libc::c_int, libc::c_int, libc::c_int, qt_core::flags::Flags<qt_core::qt::AlignmentFlag>)) -> ()
C++ method: void QGraphicsGridLayout::addItem(QGraphicsLayoutItem* item, int row, int column, int rowSpan, int columnSpan, QFlags<Qt::AlignmentFlag> alignment = ?)

C++ documentation:

Adds item to the grid on row and column. You can specify a rowSpan and columnSpan and an optional alignment.

fn alignment(&self, item: *mut GraphicsLayoutItem) -> Flags<AlignmentFlag>

C++ method: QFlags<Qt::AlignmentFlag> QGraphicsGridLayout::alignment(QGraphicsLayoutItem* item) const

C++ documentation:

Returns the alignment for item.

See also setAlignment().

fn column_alignment(&self, column: c_int) -> Flags<AlignmentFlag>

C++ method: QFlags<Qt::AlignmentFlag> QGraphicsGridLayout::columnAlignment(int column) const

C++ documentation:

Returns the alignment for column.

See also setColumnAlignment().

fn column_count(&self) -> c_int

C++ method: int QGraphicsGridLayout::columnCount() const

C++ documentation:

Returns the number of columns in the grid layout. This is always one more than the index of the last column that is occupied by a layout item (empty columns are counted except for those at the end).

fn column_maximum_width(&self, column: c_int) -> c_double

C++ method: double QGraphicsGridLayout::columnMaximumWidth(int column) const

C++ documentation:

Returns the maximum width for column.

See also setColumnMaximumWidth().

fn column_minimum_width(&self, column: c_int) -> c_double

C++ method: double QGraphicsGridLayout::columnMinimumWidth(int column) const

C++ documentation:

Returns the minimum width for column.

See also setColumnMinimumWidth().

fn column_preferred_width(&self, column: c_int) -> c_double

C++ method: double QGraphicsGridLayout::columnPreferredWidth(int column) const

C++ documentation:

Returns the preferred width for column.

See also setColumnPreferredWidth().

fn column_spacing(&self, column: c_int) -> c_double

C++ method: double QGraphicsGridLayout::columnSpacing(int column) const

C++ documentation:

Returns the column spacing for column.

See also setColumnSpacing().

fn column_stretch_factor(&self, column: c_int) -> c_int

C++ method: int QGraphicsGridLayout::columnStretchFactor(int column) const

C++ documentation:

Returns the stretch factor for column.

See also setColumnStretchFactor().

fn contents_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsGridLayoutContentsRectArgs<'largs>

C++ method: QGraphicsGridLayout::contentsRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn contents_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn contents_rect(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QGraphicsGridLayout::contentsRect() const

C++ documentation:

Returns the contents rect in local coordinates.

The contents rect defines the subrectangle used by an associated layout when arranging subitems. This function is a convenience function that adjusts the item's geometry() by its contents margins. Note that getContentsMargins() is a virtual function that you can reimplement to return the item's contents margins.

See also getContentsMargins() and geometry().

fn count(&self) -> c_int

C++ method: virtual int QGraphicsGridLayout::count() const

C++ documentation:

Reimplemented from QGraphicsLayout::count().

Returns the number of layout items in this grid layout.

fn effective_size_hint<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsGridLayoutEffectiveSizeHintArgs<'largs>

C++ method: QGraphicsGridLayout::effectiveSizeHint

This is an overloaded function. Available variants:

Rust arguments:
1) fn effective_size_hint(&self, (qt_core::qt::SizeHint, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn effective_size_hint(&self, (qt_core::qt::SizeHint, &qt_core::size_f::SizeF, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
3) fn effective_size_hint(&self, (qt_core::qt::SizeHint, cpp_utils::AsStruct)) -> qt_core::size_f::SizeF
4) fn effective_size_hint(&self, (qt_core::qt::SizeHint, &qt_core::size_f::SizeF, cpp_utils::AsStruct)) -> qt_core::size_f::SizeF
C++ method: QSizeF QGraphicsGridLayout::effectiveSizeHint(Qt::SizeHint which, const QSizeF& constraint = ?) const

C++ documentation:

Returns the effective size hint for this QGraphicsLayoutItem.

which is the size hint in question. constraint is an optional argument that defines a special constrain when calculating the effective size hint. By default, constraint is QSizeF(-1, -1), which means there is no constraint to the size hint.

If you want to specify the widget's size hint for a given width or height, you can provide the fixed dimension in constraint. This is useful for widgets that can grow only either vertically or horizontally, and need to set either their width or their height to a special value.

For example, a text paragraph item fit into a column width of 200 may grow vertically. You can pass QSizeF(200, -1) as a constraint to get a suitable minimum, preferred and maximum height).

You can adjust the effective size hint by reimplementing sizeHint() in a QGraphicsLayoutItem subclass, or by calling one of the following functions: setMinimumSize(), setPreferredSize, or setMaximumSize() (or a combination of both).

This function caches each of the size hints and guarantees that sizeHint() will be called only once for each value of which - unless constraint is not specified and updateGeometry() has been called.

See also sizeHint().

fn geometry<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsGridLayoutGeometryArgs<'largs>

C++ method: QGraphicsGridLayout::geometry

This is an overloaded function. Available variants:

Rust arguments:
1) fn geometry(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn geometry(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QGraphicsGridLayout::geometry() const

C++ documentation:

Returns the item's geometry (e.g., position and size) as a QRectF. This function is equivalent to QRectF(pos(), size()).

See also setGeometry().

fn get_contents_margins(&self, left: *mut c_double, top: *mut c_double, right: *mut c_double, bottom: *mut c_double)

C++ method: virtual void QGraphicsGridLayout::getContentsMargins(double* left, double* top, double* right, double* bottom) const

C++ documentation:

Reimplemented from QGraphicsLayoutItem::getContentsMargins().

fn graphics_item(&self) -> *mut GraphicsItem

C++ method: QGraphicsItem* QGraphicsGridLayout::graphicsItem() const

C++ documentation:

Returns the QGraphicsItem that this layout item represents. For QGraphicsWidget it will return itself. For custom items it can return an aggregated value.

See also setGraphicsItem().

fn horizontal_spacing(&self) -> c_double

C++ method: double QGraphicsGridLayout::horizontalSpacing() const

C++ documentation:

Returns the default horizontal spacing for the grid layout.

See also setHorizontalSpacing().

fn instant_invalidate_propagation() -> bool

C++ method: static bool QGraphicsGridLayout::instantInvalidatePropagation()

fn invalidate(&mut self)

C++ method: virtual void QGraphicsGridLayout::invalidate()

C++ documentation:

Reimplemented from QGraphicsLayout::invalidate().

fn is_activated(&self) -> bool

C++ method: bool QGraphicsGridLayout::isActivated() const

C++ documentation:

Returns true if the layout is currently being activated; otherwise, returns false. If the layout is being activated, this means that it is currently in the process of rearranging its items (i.e., the activate() function has been called, and has not yet returned).

See also activate() and invalidate().

fn is_layout(&self) -> bool

C++ method: bool QGraphicsGridLayout::isLayout() const

C++ documentation:

Returns true if this QGraphicsLayoutItem is a layout (e.g., is inherited by an object that arranges other QGraphicsLayoutItem objects); otherwise returns false.

See also QGraphicsLayout.

fn item_at<'largs, Args>(&'largs self, args: Args) -> *mut GraphicsLayoutItem where Args: GraphicsGridLayoutItemAtArgs<'largs>

C++ method: QGraphicsGridLayout::itemAt

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn item_at(&self, (libc::c_int, libc::c_int)) -> *mut ::graphics_layout_item::GraphicsLayoutItem
C++ method: QGraphicsLayoutItem* QGraphicsGridLayout::itemAt(int row, int column) const

C++ documentation:

Returns a pointer to the layout item at (row, column).

Variant 2

Rust arguments: fn item_at(&self, libc::c_int) -> *mut ::graphics_layout_item::GraphicsLayoutItem
C++ method: virtual QGraphicsLayoutItem* QGraphicsGridLayout::itemAt(int index) const

C++ documentation:

Reimplemented from QGraphicsLayout::itemAt().

Returns the layout item at index, or 0 if there is no layout item at this index.

fn maximum_height(&self) -> c_double

C++ method: double QGraphicsGridLayout::maximumHeight() const

C++ documentation:

Returns the maximum height.

See also setMaximumHeight(), setMaximumSize(), and maximumSize().

fn maximum_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsGridLayoutMaximumSizeArgs<'largs>

C++ method: QGraphicsGridLayout::maximumSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn maximum_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn maximum_size(&self, cpp_utils::AsStruct) -> qt_core::size_f::SizeF
C++ method: QSizeF QGraphicsGridLayout::maximumSize() const

C++ documentation:

Returns the maximum size.

See also setMaximumSize(), minimumSize(), preferredSize(), Qt::MaximumSize, and sizeHint().

fn maximum_width(&self) -> c_double

C++ method: double QGraphicsGridLayout::maximumWidth() const

C++ documentation:

Returns the maximum width.

See also setMaximumWidth(), setMaximumSize(), and maximumSize().

fn minimum_height(&self) -> c_double

C++ method: double QGraphicsGridLayout::minimumHeight() const

C++ documentation:

Returns the minimum height.

See also setMinimumHeight(), setMinimumSize(), and minimumSize().

fn minimum_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsGridLayoutMinimumSizeArgs<'largs>

C++ method: QGraphicsGridLayout::minimumSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn minimum_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn minimum_size(&self, cpp_utils::AsStruct) -> qt_core::size_f::SizeF
C++ method: QSizeF QGraphicsGridLayout::minimumSize() const

C++ documentation:

Returns the minimum size.

See also setMinimumSize(), preferredSize(), maximumSize(), Qt::MinimumSize, and sizeHint().

fn minimum_width(&self) -> c_double

C++ method: double QGraphicsGridLayout::minimumWidth() const

C++ documentation:

Returns the minimum width.

See also setMinimumWidth(), setMinimumSize(), and minimumSize().

fn new<Args>(args: Args) -> Args::ReturnType where Args: GraphicsGridLayoutNewArgs

C++ method: QGraphicsGridLayout::QGraphicsGridLayout

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::graphics_grid_layout::GraphicsGridLayout
2) fn new((*mut ::graphics_layout_item::GraphicsLayoutItem, cpp_utils::AsStruct)) -> ::graphics_grid_layout::GraphicsGridLayout
3) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::graphics_grid_layout::GraphicsGridLayout>
4) fn new((*mut ::graphics_layout_item::GraphicsLayoutItem, cpp_utils::AsBox)) -> cpp_utils::CppBox<::graphics_grid_layout::GraphicsGridLayout>
C++ method: [constructor] void QGraphicsGridLayout::QGraphicsGridLayout(QGraphicsLayoutItem* parent = ?)

fn owned_by_layout(&self) -> bool

C++ method: bool QGraphicsGridLayout::ownedByLayout() const

C++ documentation:

Returns whether a layout should delete this item in its destructor. If its true, then the layout will delete it. If its false, then it is assumed that another object has the ownership of it, and the layout won't delete this item.

If the item inherits both QGraphicsItem and QGraphicsLayoutItem (such as QGraphicsWidget does) the item is really part of two ownership hierarchies. This property informs what the layout should do with its child items when it is destructed. In the case of QGraphicsWidget, it is preferred that when the layout is deleted it won't delete its children (since they are also part of the graphics item hierarchy).

By default this value is initialized to false in QGraphicsLayoutItem, but it is overridden by QGraphicsLayout to return true. This is because QGraphicsLayout is not normally part of the QGraphicsItem hierarchy, so the parent layout should delete it. Subclasses might override this default behaviour by calling setOwnedByLayout(true).

This function was introduced in Qt 4.6.

See also setOwnedByLayout().

fn parent_layout_item(&self) -> *mut GraphicsLayoutItem

C++ method: QGraphicsLayoutItem* QGraphicsGridLayout::parentLayoutItem() const

C++ documentation:

Returns the parent of this QGraphicsLayoutItem, or 0 if there is no parent, or if the parent does not inherit from QGraphicsLayoutItem (QGraphicsLayoutItem is often used through multiple inheritance with QObject-derived classes).

See also setParentLayoutItem().

fn preferred_height(&self) -> c_double

C++ method: double QGraphicsGridLayout::preferredHeight() const

C++ documentation:

Returns the preferred height.

See also setPreferredHeight(), setPreferredSize(), and preferredSize().

fn preferred_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsGridLayoutPreferredSizeArgs<'largs>

C++ method: QGraphicsGridLayout::preferredSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn preferred_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn preferred_size(&self, cpp_utils::AsStruct) -> qt_core::size_f::SizeF
C++ method: QSizeF QGraphicsGridLayout::preferredSize() const

C++ documentation:

Returns the preferred size.

See also setPreferredSize(), minimumSize(), maximumSize(), Qt::PreferredSize, and sizeHint().

fn preferred_width(&self) -> c_double

C++ method: double QGraphicsGridLayout::preferredWidth() const

C++ documentation:

Returns the preferred width.

See also setPreferredWidth(), setPreferredSize(), and preferredSize().

fn remove_at(&mut self, index: c_int)

C++ method: virtual void QGraphicsGridLayout::removeAt(int index)

C++ documentation:

Reimplemented from QGraphicsLayout::removeAt().

Removes the layout item at index without destroying it. Ownership of the item is transferred to the caller.

See also addItem().

fn remove_item(&mut self, item: *mut GraphicsLayoutItem)

C++ method: void QGraphicsGridLayout::removeItem(QGraphicsLayoutItem* item)

C++ documentation:

Removes the layout item item without destroying it. Ownership of the item is transferred to the caller.

See also addItem().

fn row_alignment(&self, row: c_int) -> Flags<AlignmentFlag>

C++ method: QFlags<Qt::AlignmentFlag> QGraphicsGridLayout::rowAlignment(int row) const

C++ documentation:

Returns the alignment of row.

See also setRowAlignment().

fn row_count(&self) -> c_int

C++ method: int QGraphicsGridLayout::rowCount() const

C++ documentation:

Returns the number of rows in the grid layout. This is always one more than the index of the last row that is occupied by a layout item (empty rows are counted except for those at the end).

fn row_maximum_height(&self, row: c_int) -> c_double

C++ method: double QGraphicsGridLayout::rowMaximumHeight(int row) const

C++ documentation:

Returns the maximum height for row, row.

See also setRowMaximumHeight().

fn row_minimum_height(&self, row: c_int) -> c_double

C++ method: double QGraphicsGridLayout::rowMinimumHeight(int row) const

C++ documentation:

Returns the minimum height for row, row.

See also setRowMinimumHeight().

fn row_preferred_height(&self, row: c_int) -> c_double

C++ method: double QGraphicsGridLayout::rowPreferredHeight(int row) const

C++ documentation:

Returns the preferred height for row, row.

See also setRowPreferredHeight().

fn row_spacing(&self, row: c_int) -> c_double

C++ method: double QGraphicsGridLayout::rowSpacing(int row) const

C++ documentation:

Returns the row spacing for row.

See also setRowSpacing().

fn row_stretch_factor(&self, row: c_int) -> c_int

C++ method: int QGraphicsGridLayout::rowStretchFactor(int row) const

C++ documentation:

Returns the stretch factor for row.

See also setRowStretchFactor().

fn set_alignment(&mut self, item: *mut GraphicsLayoutItem, alignment: Flags<AlignmentFlag>)

C++ method: void QGraphicsGridLayout::setAlignment(QGraphicsLayoutItem* item, QFlags<Qt::AlignmentFlag> alignment)

C++ documentation:

Sets the alignment for item to alignment.

See also alignment().

fn set_column_alignment(&mut self, column: c_int, alignment: Flags<AlignmentFlag>)

C++ method: void QGraphicsGridLayout::setColumnAlignment(int column, QFlags<Qt::AlignmentFlag> alignment)

C++ documentation:

Sets the alignment for column to alignment.

See also columnAlignment().

fn set_column_fixed_width(&mut self, column: c_int, width: c_double)

C++ method: void QGraphicsGridLayout::setColumnFixedWidth(int column, double width)

C++ documentation:

Sets the fixed width of column to width.

fn set_column_maximum_width(&mut self, column: c_int, width: c_double)

C++ method: void QGraphicsGridLayout::setColumnMaximumWidth(int column, double width)

C++ documentation:

Sets the maximum width of column to width.

See also columnMaximumWidth().

fn set_column_minimum_width(&mut self, column: c_int, width: c_double)

C++ method: void QGraphicsGridLayout::setColumnMinimumWidth(int column, double width)

C++ documentation:

Sets the minimum width for column to width.

See also columnMinimumWidth().

fn set_column_preferred_width(&mut self, column: c_int, width: c_double)

C++ method: void QGraphicsGridLayout::setColumnPreferredWidth(int column, double width)

C++ documentation:

Sets the preferred width for column to width.

See also columnPreferredWidth().

fn set_column_spacing(&mut self, column: c_int, spacing: c_double)

C++ method: void QGraphicsGridLayout::setColumnSpacing(int column, double spacing)

C++ documentation:

Sets the spacing for column to spacing.

See also columnSpacing().

fn set_column_stretch_factor(&mut self, column: c_int, stretch: c_int)

C++ method: void QGraphicsGridLayout::setColumnStretchFactor(int column, int stretch)

C++ documentation:

Sets the stretch factor for column to stretch.

See also columnStretchFactor().

fn set_contents_margins(&mut self, left: c_double, top: c_double, right: c_double, bottom: c_double)

C++ method: void QGraphicsGridLayout::setContentsMargins(double left, double top, double right, double bottom)

C++ documentation:

Sets the contents margins to left, top, right and bottom. The default contents margins for toplevel layouts are style dependent (by querying the pixelMetric for QStyle::PM_LayoutLeftMargin, QStyle::PM_LayoutTopMargin, QStyle::PM_LayoutRightMargin and QStyle::PM_LayoutBottomMargin).

For sublayouts the default margins are 0.

Changing the contents margins automatically invalidates the layout.

See also invalidate().

fn set_geometry(&mut self, rect: &RectF)

C++ method: virtual void QGraphicsGridLayout::setGeometry(const QRectF& rect)

C++ documentation:

Reimplemented from QGraphicsLayoutItem::setGeometry().

Sets the bounding geometry of the grid layout to rect.

fn set_horizontal_spacing(&mut self, spacing: c_double)

C++ method: void QGraphicsGridLayout::setHorizontalSpacing(double spacing)

C++ documentation:

Sets the default horizontal spacing for the grid layout to spacing.

See also horizontalSpacing().

fn set_instant_invalidate_propagation(enable: bool)

C++ method: static void QGraphicsGridLayout::setInstantInvalidatePropagation(bool enable)

fn set_maximum_height(&mut self, height: c_double)

C++ method: void QGraphicsGridLayout::setMaximumHeight(double height)

C++ documentation:

Sets the maximum height to height.

See also maximumHeight(), setMaximumSize(), and maximumSize().

fn set_maximum_size<'largs, Args>(&'largs mut self, args: Args) where Args: GraphicsGridLayoutSetMaximumSizeArgs<'largs>

C++ method: QGraphicsGridLayout::setMaximumSize

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_maximum_size(&mut self, &qt_core::size_f::SizeF) -> ()
C++ method: void QGraphicsGridLayout::setMaximumSize(const QSizeF& size)

C++ documentation:

Sets the maximum size to size. This property overrides sizeHint() for Qt::MaximumSize and ensures that effectiveSizeHint() will never return a size larger than size. In order to unset the maximum size, use an invalid size.

See also maximumSize(), minimumSize(), preferredSize(), Qt::MaximumSize, and sizeHint().

Variant 2

Rust arguments: fn set_maximum_size(&mut self, (libc::c_double, libc::c_double)) -> ()
C++ method: void QGraphicsGridLayout::setMaximumSize(double w, double h)

C++ documentation:

This convenience function is equivalent to calling setMaximumSize(QSizeF(w, h)).

See also maximumSize(), setMinimumSize(), setPreferredSize(), and sizeHint().

fn set_maximum_width(&mut self, width: c_double)

C++ method: void QGraphicsGridLayout::setMaximumWidth(double width)

C++ documentation:

Sets the maximum width to width.

See also maximumWidth(), setMaximumSize(), and maximumSize().

fn set_minimum_height(&mut self, height: c_double)

C++ method: void QGraphicsGridLayout::setMinimumHeight(double height)

C++ documentation:

Sets the minimum height to height.

See also minimumHeight(), setMinimumSize(), and minimumSize().

fn set_minimum_size<'largs, Args>(&'largs mut self, args: Args) where Args: GraphicsGridLayoutSetMinimumSizeArgs<'largs>

C++ method: QGraphicsGridLayout::setMinimumSize

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_minimum_size(&mut self, &qt_core::size_f::SizeF) -> ()
C++ method: void QGraphicsGridLayout::setMinimumSize(const QSizeF& size)

C++ documentation:

Sets the minimum size to size. This property overrides sizeHint() for Qt::MinimumSize and ensures that effectiveSizeHint() will never return a size smaller than size. In order to unset the minimum size, use an invalid size.

See also minimumSize(), maximumSize(), preferredSize(), Qt::MinimumSize, sizeHint(), setMinimumWidth(), and setMinimumHeight().

Variant 2

Rust arguments: fn set_minimum_size(&mut self, (libc::c_double, libc::c_double)) -> ()
C++ method: void QGraphicsGridLayout::setMinimumSize(double w, double h)

C++ documentation:

This convenience function is equivalent to calling setMinimumSize(QSizeF(w, h)).

See also minimumSize(), setMaximumSize(), setPreferredSize(), and sizeHint().

fn set_minimum_width(&mut self, width: c_double)

C++ method: void QGraphicsGridLayout::setMinimumWidth(double width)

C++ documentation:

Sets the minimum width to width.

See also minimumWidth(), setMinimumSize(), and minimumSize().

fn set_parent_layout_item(&mut self, parent: *mut GraphicsLayoutItem)

C++ method: void QGraphicsGridLayout::setParentLayoutItem(QGraphicsLayoutItem* parent)

C++ documentation:

Sets the parent of this QGraphicsLayoutItem to parent.

See also parentLayoutItem().

fn set_preferred_height(&mut self, height: c_double)

C++ method: void QGraphicsGridLayout::setPreferredHeight(double height)

C++ documentation:

Sets the preferred height to height.

See also preferredHeight(), preferredWidth(), setPreferredSize(), and preferredSize().

fn set_preferred_size<'largs, Args>(&'largs mut self, args: Args) where Args: GraphicsGridLayoutSetPreferredSizeArgs<'largs>

C++ method: QGraphicsGridLayout::setPreferredSize

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_preferred_size(&mut self, &qt_core::size_f::SizeF) -> ()
C++ method: void QGraphicsGridLayout::setPreferredSize(const QSizeF& size)

C++ documentation:

Sets the preferred size to size. This property overrides sizeHint() for Qt::PreferredSize and provides the default value for effectiveSizeHint(). In order to unset the preferred size, use an invalid size.

See also preferredSize(), minimumSize(), maximumSize(), Qt::PreferredSize, and sizeHint().

Variant 2

Rust arguments: fn set_preferred_size(&mut self, (libc::c_double, libc::c_double)) -> ()
C++ method: void QGraphicsGridLayout::setPreferredSize(double w, double h)

C++ documentation:

This convenience function is equivalent to calling setPreferredSize(QSizeF(w, h)).

See also preferredSize(), setMaximumSize(), setMinimumSize(), and sizeHint().

fn set_preferred_width(&mut self, width: c_double)

C++ method: void QGraphicsGridLayout::setPreferredWidth(double width)

C++ documentation:

Sets the preferred width to width.

See also preferredWidth(), preferredHeight(), setPreferredSize(), and preferredSize().

fn set_row_alignment(&mut self, row: c_int, alignment: Flags<AlignmentFlag>)

C++ method: void QGraphicsGridLayout::setRowAlignment(int row, QFlags<Qt::AlignmentFlag> alignment)

C++ documentation:

Sets the alignment of row to alignment.

See also rowAlignment().

fn set_row_fixed_height(&mut self, row: c_int, height: c_double)

C++ method: void QGraphicsGridLayout::setRowFixedHeight(int row, double height)

C++ documentation:

Sets the fixed height for row, row, to height.

fn set_row_maximum_height(&mut self, row: c_int, height: c_double)

C++ method: void QGraphicsGridLayout::setRowMaximumHeight(int row, double height)

C++ documentation:

Sets the maximum height for row, row, to height.

See also rowMaximumHeight().

fn set_row_minimum_height(&mut self, row: c_int, height: c_double)

C++ method: void QGraphicsGridLayout::setRowMinimumHeight(int row, double height)

C++ documentation:

Sets the minimum height for row, row, to height.

See also rowMinimumHeight().

fn set_row_preferred_height(&mut self, row: c_int, height: c_double)

C++ method: void QGraphicsGridLayout::setRowPreferredHeight(int row, double height)

C++ documentation:

Sets the preferred height for row, row, to height.

See also rowPreferredHeight().

fn set_row_spacing(&mut self, row: c_int, spacing: c_double)

C++ method: void QGraphicsGridLayout::setRowSpacing(int row, double spacing)

C++ documentation:

Sets the spacing for row to spacing.

See also rowSpacing().

fn set_row_stretch_factor(&mut self, row: c_int, stretch: c_int)

C++ method: void QGraphicsGridLayout::setRowStretchFactor(int row, int stretch)

C++ documentation:

Sets the stretch factor for row to stretch.

See also rowStretchFactor().

fn set_size_policy<'largs, Args>(&'largs mut self, args: Args) where Args: GraphicsGridLayoutSetSizePolicyArgs<'largs>

C++ method: QGraphicsGridLayout::setSizePolicy

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn set_size_policy(&mut self, (::size_policy::Policy, ::size_policy::Policy)) -> ()
2) fn set_size_policy(&mut self, (::size_policy::Policy, ::size_policy::Policy, ::size_policy::ControlType)) -> ()
C++ method: void QGraphicsGridLayout::setSizePolicy(QSizePolicy::Policy hPolicy, QSizePolicy::Policy vPolicy, QSizePolicy::ControlType controlType = ?)

C++ documentation:

This is an overloaded function.

This function is equivalent to calling setSizePolicy(QSizePolicy(hPolicy, vPolicy, controlType)).

See also sizePolicy() and QWidget::sizePolicy().

Variant 2

Rust arguments: fn set_size_policy(&mut self, &::size_policy::SizePolicy) -> ()
C++ method: void QGraphicsGridLayout::setSizePolicy(const QSizePolicy& policy)

C++ documentation:

Sets the size policy to policy. The size policy describes how the item should grow horizontally and vertically when arranged in a layout.

QGraphicsLayoutItem's default size policy is (QSizePolicy::Fixed, QSizePolicy::Fixed, QSizePolicy::DefaultType), but it is common for subclasses to change the default. For example, QGraphicsWidget defaults to (QSizePolicy::Preferred, QSizePolicy::Preferred, QSizePolicy::DefaultType).

See also sizePolicy() and QWidget::sizePolicy().

fn set_spacing(&mut self, spacing: c_double)

C++ method: void QGraphicsGridLayout::setSpacing(double spacing)

C++ documentation:

Sets the grid layout's default spacing, both vertical and horizontal, to spacing.

See also rowSpacing() and columnSpacing().

fn set_vertical_spacing(&mut self, spacing: c_double)

C++ method: void QGraphicsGridLayout::setVerticalSpacing(double spacing)

C++ documentation:

Sets the default vertical spacing for the grid layout to spacing.

See also verticalSpacing().

fn size_hint<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsGridLayoutSizeHintArgs<'largs>

C++ method: QGraphicsGridLayout::sizeHint

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_hint(&self, (qt_core::qt::SizeHint, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn size_hint(&self, (qt_core::qt::SizeHint, &qt_core::size_f::SizeF, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
3) fn size_hint(&self, (qt_core::qt::SizeHint, cpp_utils::AsStruct)) -> qt_core::size_f::SizeF
4) fn size_hint(&self, (qt_core::qt::SizeHint, &qt_core::size_f::SizeF, cpp_utils::AsStruct)) -> qt_core::size_f::SizeF
C++ method: virtual QSizeF QGraphicsGridLayout::sizeHint(Qt::SizeHint which, const QSizeF& constraint = ?) const

C++ documentation:

Reimplemented from QGraphicsLayoutItem::sizeHint().

fn size_policy<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsGridLayoutSizePolicyArgs<'largs>

C++ method: QGraphicsGridLayout::sizePolicy

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_policy(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::size_policy::SizePolicy>
2) fn size_policy(&self, cpp_utils::AsStruct) -> ::size_policy::SizePolicy
C++ method: QSizePolicy QGraphicsGridLayout::sizePolicy() const

C++ documentation:

Returns the current size policy.

See also setSizePolicy() and QWidget::sizePolicy().

fn update_geometry(&mut self)

C++ method: virtual void QGraphicsGridLayout::updateGeometry()

C++ documentation:

Reimplemented from QGraphicsLayoutItem::updateGeometry().

fn vertical_spacing(&self) -> c_double

C++ method: double QGraphicsGridLayout::verticalSpacing() const

C++ documentation:

Returns the default vertical spacing for the grid layout.

See also setVerticalSpacing().

fn widget_event(&mut self, e: *mut Event)

C++ method: virtual void QGraphicsGridLayout::widgetEvent(QEvent* e)

C++ documentation:

This virtual event handler receives all events for the managed widget. QGraphicsLayout uses this event handler to listen for layout related events such as geometry changes, layout changes or layout direction changes.

e is a pointer to the event.

You can reimplement this event handler to track similar events for your own custom layout.

See also QGraphicsWidget::event() and QGraphicsItem::sceneEvent().