Enum qt_widgets::graphics_item::CacheMode

pub enum CacheMode {
    No,
    ItemCoordinate,
    DeviceCoordinate,
}

C++ type: QGraphicsItem::CacheMode

C++ documentation:

This enum describes QGraphicsItem's cache modes. Caching is used to speed up rendering by allocating and rendering to an off-screen pixel buffer, which can be reused when the item requires redrawing. For some paint devices, the cache is stored directly in graphics memory, which makes rendering very quick.

This enum was introduced or modified in Qt 4.4.

See also QGraphicsItem::setCacheMode().

Variants
No

The default; all item caching is disabled. QGraphicsItem::paint() is called every time the item needs redrawing. (C++ enum variant: NoCache = 0)

ItemCoordinate

Caching is enabled for the item's logical (local) coordinate system. QGraphicsItem creates an off-screen pixel buffer with a configurable size / resolution that you can pass to QGraphicsItem::setCacheMode(). Rendering quality will typically degrade, depending on the resolution of the cache and the item transformation. The first time the item is redrawn, it will render itself into the cache, and the cache is then reused for every subsequent expose. The cache is also reused as the item is transformed. To adjust the resolution of the cache, you can call setCacheMode() again. (C++ enum variant: ItemCoordinateCache = 1)

DeviceCoordinate

Caching is enabled at the paint device level, in device coordinates. This mode is for items that can move, but are not rotated, scaled or sheared. If the item is transformed directly or indirectly, the cache will be regenerated automatically. Unlike ItemCoordinateCacheMode, DeviceCoordinateCache always renders at maximum quality. (C++ enum variant: DeviceCoordinateCache = 2)