Module qt_widgets::graphics_layout::overloading

Traits

GraphicsLayoutContentsRectArgs
This trait represents a set of arguments accepted by GraphicsLayout::contents_rect method.

GraphicsLayoutEffectiveSizeHintArgs
This trait represents a set of arguments accepted by GraphicsLayout::effective_size_hint method.

GraphicsLayoutGeometryArgs
This trait represents a set of arguments accepted by GraphicsLayout::geometry method.

GraphicsLayoutMaximumSizeArgs
This trait represents a set of arguments accepted by GraphicsLayout::maximum_size method.

GraphicsLayoutMinimumSizeArgs
This trait represents a set of arguments accepted by GraphicsLayout::minimum_size method.

GraphicsLayoutPreferredSizeArgs
This trait represents a set of arguments accepted by GraphicsLayout::preferred_size method.

GraphicsLayoutSetMaximumSizeArgs
This trait represents a set of arguments accepted by GraphicsLayout::set_maximum_size method.

GraphicsLayoutSetMinimumSizeArgs
This trait represents a set of arguments accepted by GraphicsLayout::set_minimum_size method.

GraphicsLayoutSetPreferredSizeArgs
This trait represents a set of arguments accepted by GraphicsLayout::set_preferred_size method.

GraphicsLayoutSetSizePolicyArgs
This trait represents a set of arguments accepted by GraphicsLayout::set_size_policy method.

GraphicsLayoutSizePolicyArgs
This trait represents a set of arguments accepted by GraphicsLayout::size_policy method.