Struct qt_widgets::graphics_layout::GraphicsLayout

pub struct GraphicsLayout {
    // some fields omitted
}

C++ type: QGraphicsLayout

C++ documentation:

The QGraphicsLayout class provides the base class for all layouts in Graphics View.

QGraphicsLayout is an abstract class that defines a virtual API for arranging QGraphicsWidget children and other QGraphicsLayoutItem objects for a QGraphicsWidget. QGraphicsWidget assigns responsibility to a QGraphicsLayout through QGraphicsWidget::setLayout(). As the widget is resized, the layout will automatically arrange the widget's children. QGraphicsLayout inherits QGraphicsLayoutItem, so, it can be managed by any layout, including its own subclasses.

Methods

impl GraphicsLayout

fn activate(&mut self)

C++ method: void QGraphicsLayout::activate()

C++ documentation:

Activates the layout, causing all items in the layout to be immediately rearranged. This function is based on calling count() and itemAt(), and then calling setGeometry() on all items sequentially. When activated, the layout will adjust its geometry to its parent's contentsRect(). The parent will then invalidate any layout of its own.

If called in sequence or recursively, e.g., by one of the arranged items in response to being resized, this function will do nothing.

Note that the layout is free to use geometry caching to optimize this process. To forcefully invalidate any such cache, you can call invalidate() before calling activate().

See also invalidate().

fn contents_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsLayoutContentsRectArgs<'largs>

C++ method: QGraphicsLayout::contentsRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn contents_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn contents_rect(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QGraphicsLayout::contentsRect() const

C++ documentation:

Returns the contents rect in local coordinates.

The contents rect defines the subrectangle used by an associated layout when arranging subitems. This function is a convenience function that adjusts the item's geometry() by its contents margins. Note that getContentsMargins() is a virtual function that you can reimplement to return the item's contents margins.

See also getContentsMargins() and geometry().

fn count(&self) -> c_int

C++ method: pure virtual int QGraphicsLayout::count() const

C++ documentation:

This pure virtual function must be reimplemented in a subclass of QGraphicsLayout to return the number of items in the layout.

The subclass is free to decide how to store the items.

See also itemAt() and removeAt().

fn effective_size_hint<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsLayoutEffectiveSizeHintArgs<'largs>

C++ method: QGraphicsLayout::effectiveSizeHint

This is an overloaded function. Available variants:

Rust arguments:
1) fn effective_size_hint(&self, (qt_core::qt::SizeHint, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn effective_size_hint(&self, (qt_core::qt::SizeHint, &qt_core::size_f::SizeF, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
3) fn effective_size_hint(&self, (qt_core::qt::SizeHint, cpp_utils::AsStruct)) -> qt_core::size_f::SizeF
4) fn effective_size_hint(&self, (qt_core::qt::SizeHint, &qt_core::size_f::SizeF, cpp_utils::AsStruct)) -> qt_core::size_f::SizeF
C++ method: QSizeF QGraphicsLayout::effectiveSizeHint(Qt::SizeHint which, const QSizeF& constraint = ?) const

C++ documentation:

Returns the effective size hint for this QGraphicsLayoutItem.

which is the size hint in question. constraint is an optional argument that defines a special constrain when calculating the effective size hint. By default, constraint is QSizeF(-1, -1), which means there is no constraint to the size hint.

If you want to specify the widget's size hint for a given width or height, you can provide the fixed dimension in constraint. This is useful for widgets that can grow only either vertically or horizontally, and need to set either their width or their height to a special value.

For example, a text paragraph item fit into a column width of 200 may grow vertically. You can pass QSizeF(200, -1) as a constraint to get a suitable minimum, preferred and maximum height).

You can adjust the effective size hint by reimplementing sizeHint() in a QGraphicsLayoutItem subclass, or by calling one of the following functions: setMinimumSize(), setPreferredSize, or setMaximumSize() (or a combination of both).

This function caches each of the size hints and guarantees that sizeHint() will be called only once for each value of which - unless constraint is not specified and updateGeometry() has been called.

See also sizeHint().

fn geometry<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsLayoutGeometryArgs<'largs>

C++ method: QGraphicsLayout::geometry

This is an overloaded function. Available variants:

Rust arguments:
1) fn geometry(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn geometry(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QGraphicsLayout::geometry() const

C++ documentation:

Returns the item's geometry (e.g., position and size) as a QRectF. This function is equivalent to QRectF(pos(), size()).

See also setGeometry().

fn get_contents_margins(&self, left: *mut c_double, top: *mut c_double, right: *mut c_double, bottom: *mut c_double)

C++ method: virtual void QGraphicsLayout::getContentsMargins(double* left, double* top, double* right, double* bottom) const

C++ documentation:

Reimplemented from QGraphicsLayoutItem::getContentsMargins().

fn graphics_item(&self) -> *mut GraphicsItem

C++ method: QGraphicsItem* QGraphicsLayout::graphicsItem() const

C++ documentation:

Returns the QGraphicsItem that this layout item represents. For QGraphicsWidget it will return itself. For custom items it can return an aggregated value.

See also setGraphicsItem().

fn instant_invalidate_propagation() -> bool

C++ method: static bool QGraphicsLayout::instantInvalidatePropagation()

fn invalidate(&mut self)

C++ method: virtual void QGraphicsLayout::invalidate()

C++ documentation:

Clears any cached geometry and size hint information in the layout, and posts a LayoutRequest event to the managed parent QGraphicsLayoutItem.

See also activate() and setGeometry().

fn is_activated(&self) -> bool

C++ method: bool QGraphicsLayout::isActivated() const

C++ documentation:

Returns true if the layout is currently being activated; otherwise, returns false. If the layout is being activated, this means that it is currently in the process of rearranging its items (i.e., the activate() function has been called, and has not yet returned).

See also activate() and invalidate().

fn is_layout(&self) -> bool

C++ method: bool QGraphicsLayout::isLayout() const

C++ documentation:

Returns true if this QGraphicsLayoutItem is a layout (e.g., is inherited by an object that arranges other QGraphicsLayoutItem objects); otherwise returns false.

See also QGraphicsLayout.

fn item_at(&self, i: c_int) -> *mut GraphicsLayoutItem

C++ method: pure virtual QGraphicsLayoutItem* QGraphicsLayout::itemAt(int i) const

C++ documentation:

This pure virtual function must be reimplemented in a subclass of QGraphicsLayout to return a pointer to the item at index i. The reimplementation can assume that i is valid (i.e., it respects the value of count()). Together with count(), it is provided as a means of iterating over all items in a layout.

The subclass is free to decide how to store the items, and the visual arrangement does not have to be reflected through this function.

See also count() and removeAt().

fn maximum_height(&self) -> c_double

C++ method: double QGraphicsLayout::maximumHeight() const

C++ documentation:

Returns the maximum height.

See also setMaximumHeight(), setMaximumSize(), and maximumSize().

fn maximum_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsLayoutMaximumSizeArgs<'largs>

C++ method: QGraphicsLayout::maximumSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn maximum_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn maximum_size(&self, cpp_utils::AsStruct) -> qt_core::size_f::SizeF
C++ method: QSizeF QGraphicsLayout::maximumSize() const

C++ documentation:

Returns the maximum size.

See also setMaximumSize(), minimumSize(), preferredSize(), Qt::MaximumSize, and sizeHint().

fn maximum_width(&self) -> c_double

C++ method: double QGraphicsLayout::maximumWidth() const

C++ documentation:

Returns the maximum width.

See also setMaximumWidth(), setMaximumSize(), and maximumSize().

fn minimum_height(&self) -> c_double

C++ method: double QGraphicsLayout::minimumHeight() const

C++ documentation:

Returns the minimum height.

See also setMinimumHeight(), setMinimumSize(), and minimumSize().

fn minimum_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsLayoutMinimumSizeArgs<'largs>

C++ method: QGraphicsLayout::minimumSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn minimum_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn minimum_size(&self, cpp_utils::AsStruct) -> qt_core::size_f::SizeF
C++ method: QSizeF QGraphicsLayout::minimumSize() const

C++ documentation:

Returns the minimum size.

See also setMinimumSize(), preferredSize(), maximumSize(), Qt::MinimumSize, and sizeHint().

fn minimum_width(&self) -> c_double

C++ method: double QGraphicsLayout::minimumWidth() const

C++ documentation:

Returns the minimum width.

See also setMinimumWidth(), setMinimumSize(), and minimumSize().

fn owned_by_layout(&self) -> bool

C++ method: bool QGraphicsLayout::ownedByLayout() const

C++ documentation:

Returns whether a layout should delete this item in its destructor. If its true, then the layout will delete it. If its false, then it is assumed that another object has the ownership of it, and the layout won't delete this item.

If the item inherits both QGraphicsItem and QGraphicsLayoutItem (such as QGraphicsWidget does) the item is really part of two ownership hierarchies. This property informs what the layout should do with its child items when it is destructed. In the case of QGraphicsWidget, it is preferred that when the layout is deleted it won't delete its children (since they are also part of the graphics item hierarchy).

By default this value is initialized to false in QGraphicsLayoutItem, but it is overridden by QGraphicsLayout to return true. This is because QGraphicsLayout is not normally part of the QGraphicsItem hierarchy, so the parent layout should delete it. Subclasses might override this default behaviour by calling setOwnedByLayout(true).

This function was introduced in Qt 4.6.

See also setOwnedByLayout().

fn parent_layout_item(&self) -> *mut GraphicsLayoutItem

C++ method: QGraphicsLayoutItem* QGraphicsLayout::parentLayoutItem() const

C++ documentation:

Returns the parent of this QGraphicsLayoutItem, or 0 if there is no parent, or if the parent does not inherit from QGraphicsLayoutItem (QGraphicsLayoutItem is often used through multiple inheritance with QObject-derived classes).

See also setParentLayoutItem().

fn preferred_height(&self) -> c_double

C++ method: double QGraphicsLayout::preferredHeight() const

C++ documentation:

Returns the preferred height.

See also setPreferredHeight(), setPreferredSize(), and preferredSize().

fn preferred_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsLayoutPreferredSizeArgs<'largs>

C++ method: QGraphicsLayout::preferredSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn preferred_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn preferred_size(&self, cpp_utils::AsStruct) -> qt_core::size_f::SizeF
C++ method: QSizeF QGraphicsLayout::preferredSize() const

C++ documentation:

Returns the preferred size.

See also setPreferredSize(), minimumSize(), maximumSize(), Qt::PreferredSize, and sizeHint().

fn preferred_width(&self) -> c_double

C++ method: double QGraphicsLayout::preferredWidth() const

C++ documentation:

Returns the preferred width.

See also setPreferredWidth(), setPreferredSize(), and preferredSize().

fn remove_at(&mut self, index: c_int)

C++ method: pure virtual void QGraphicsLayout::removeAt(int index)

C++ documentation:

This pure virtual function must be reimplemented in a subclass of QGraphicsLayout to remove the item at index. The reimplementation can assume that index is valid (i.e., it respects the value of count()).

The implementation must ensure that the parentLayoutItem() of the removed item does not point to this layout, since the item is considered to be removed from the layout hierarchy.

If the layout is to be reused between applications, we recommend that the layout deletes the item, but the graphics view framework does not depend on this.

The subclass is free to decide how to store the items.

See also itemAt() and count().

fn set_contents_margins(&mut self, left: c_double, top: c_double, right: c_double, bottom: c_double)

C++ method: void QGraphicsLayout::setContentsMargins(double left, double top, double right, double bottom)

C++ documentation:

Sets the contents margins to left, top, right and bottom. The default contents margins for toplevel layouts are style dependent (by querying the pixelMetric for QStyle::PM_LayoutLeftMargin, QStyle::PM_LayoutTopMargin, QStyle::PM_LayoutRightMargin and QStyle::PM_LayoutBottomMargin).

For sublayouts the default margins are 0.

Changing the contents margins automatically invalidates the layout.

See also invalidate().

fn set_geometry(&mut self, rect: &RectF)

C++ method: virtual void QGraphicsLayout::setGeometry(const QRectF& rect)

C++ documentation:

This virtual function sets the geometry of the QGraphicsLayoutItem to rect, which is in parent coordinates (e.g., the top-left corner of rect is equivalent to the item's position in parent coordinates).

You must reimplement this function in a subclass of QGraphicsLayoutItem to receive geometry updates. The layout will call this function when it does a rearrangement.

If rect is outside of the bounds of minimumSize and maximumSize, it will be adjusted to its closest size so that it is within the legal bounds.

See also geometry().

fn set_instant_invalidate_propagation(enable: bool)

C++ method: static void QGraphicsLayout::setInstantInvalidatePropagation(bool enable)

fn set_maximum_height(&mut self, height: c_double)

C++ method: void QGraphicsLayout::setMaximumHeight(double height)

C++ documentation:

Sets the maximum height to height.

See also maximumHeight(), setMaximumSize(), and maximumSize().

fn set_maximum_size<'largs, Args>(&'largs mut self, args: Args) where Args: GraphicsLayoutSetMaximumSizeArgs<'largs>

C++ method: QGraphicsLayout::setMaximumSize

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_maximum_size(&mut self, &qt_core::size_f::SizeF) -> ()
C++ method: void QGraphicsLayout::setMaximumSize(const QSizeF& size)

C++ documentation:

Sets the maximum size to size. This property overrides sizeHint() for Qt::MaximumSize and ensures that effectiveSizeHint() will never return a size larger than size. In order to unset the maximum size, use an invalid size.

See also maximumSize(), minimumSize(), preferredSize(), Qt::MaximumSize, and sizeHint().

Variant 2

Rust arguments: fn set_maximum_size(&mut self, (libc::c_double, libc::c_double)) -> ()
C++ method: void QGraphicsLayout::setMaximumSize(double w, double h)

C++ documentation:

This convenience function is equivalent to calling setMaximumSize(QSizeF(w, h)).

See also maximumSize(), setMinimumSize(), setPreferredSize(), and sizeHint().

fn set_maximum_width(&mut self, width: c_double)

C++ method: void QGraphicsLayout::setMaximumWidth(double width)

C++ documentation:

Sets the maximum width to width.

See also maximumWidth(), setMaximumSize(), and maximumSize().

fn set_minimum_height(&mut self, height: c_double)

C++ method: void QGraphicsLayout::setMinimumHeight(double height)

C++ documentation:

Sets the minimum height to height.

See also minimumHeight(), setMinimumSize(), and minimumSize().

fn set_minimum_size<'largs, Args>(&'largs mut self, args: Args) where Args: GraphicsLayoutSetMinimumSizeArgs<'largs>

C++ method: QGraphicsLayout::setMinimumSize

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_minimum_size(&mut self, &qt_core::size_f::SizeF) -> ()
C++ method: void QGraphicsLayout::setMinimumSize(const QSizeF& size)

C++ documentation:

Sets the minimum size to size. This property overrides sizeHint() for Qt::MinimumSize and ensures that effectiveSizeHint() will never return a size smaller than size. In order to unset the minimum size, use an invalid size.

See also minimumSize(), maximumSize(), preferredSize(), Qt::MinimumSize, sizeHint(), setMinimumWidth(), and setMinimumHeight().

Variant 2

Rust arguments: fn set_minimum_size(&mut self, (libc::c_double, libc::c_double)) -> ()
C++ method: void QGraphicsLayout::setMinimumSize(double w, double h)

C++ documentation:

This convenience function is equivalent to calling setMinimumSize(QSizeF(w, h)).

See also minimumSize(), setMaximumSize(), setPreferredSize(), and sizeHint().

fn set_minimum_width(&mut self, width: c_double)

C++ method: void QGraphicsLayout::setMinimumWidth(double width)

C++ documentation:

Sets the minimum width to width.

See also minimumWidth(), setMinimumSize(), and minimumSize().

fn set_parent_layout_item(&mut self, parent: *mut GraphicsLayoutItem)

C++ method: void QGraphicsLayout::setParentLayoutItem(QGraphicsLayoutItem* parent)

C++ documentation:

Sets the parent of this QGraphicsLayoutItem to parent.

See also parentLayoutItem().

fn set_preferred_height(&mut self, height: c_double)

C++ method: void QGraphicsLayout::setPreferredHeight(double height)

C++ documentation:

Sets the preferred height to height.

See also preferredHeight(), preferredWidth(), setPreferredSize(), and preferredSize().

fn set_preferred_size<'largs, Args>(&'largs mut self, args: Args) where Args: GraphicsLayoutSetPreferredSizeArgs<'largs>

C++ method: QGraphicsLayout::setPreferredSize

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_preferred_size(&mut self, &qt_core::size_f::SizeF) -> ()
C++ method: void QGraphicsLayout::setPreferredSize(const QSizeF& size)

C++ documentation:

Sets the preferred size to size. This property overrides sizeHint() for Qt::PreferredSize and provides the default value for effectiveSizeHint(). In order to unset the preferred size, use an invalid size.

See also preferredSize(), minimumSize(), maximumSize(), Qt::PreferredSize, and sizeHint().

Variant 2

Rust arguments: fn set_preferred_size(&mut self, (libc::c_double, libc::c_double)) -> ()
C++ method: void QGraphicsLayout::setPreferredSize(double w, double h)

C++ documentation:

This convenience function is equivalent to calling setPreferredSize(QSizeF(w, h)).

See also preferredSize(), setMaximumSize(), setMinimumSize(), and sizeHint().

fn set_preferred_width(&mut self, width: c_double)

C++ method: void QGraphicsLayout::setPreferredWidth(double width)

C++ documentation:

Sets the preferred width to width.

See also preferredWidth(), preferredHeight(), setPreferredSize(), and preferredSize().

fn set_size_policy<'largs, Args>(&'largs mut self, args: Args) where Args: GraphicsLayoutSetSizePolicyArgs<'largs>

C++ method: QGraphicsLayout::setSizePolicy

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn set_size_policy(&mut self, (::size_policy::Policy, ::size_policy::Policy)) -> ()
2) fn set_size_policy(&mut self, (::size_policy::Policy, ::size_policy::Policy, ::size_policy::ControlType)) -> ()
C++ method: void QGraphicsLayout::setSizePolicy(QSizePolicy::Policy hPolicy, QSizePolicy::Policy vPolicy, QSizePolicy::ControlType controlType = ?)

C++ documentation:

This is an overloaded function.

This function is equivalent to calling setSizePolicy(QSizePolicy(hPolicy, vPolicy, controlType)).

See also sizePolicy() and QWidget::sizePolicy().

Variant 2

Rust arguments: fn set_size_policy(&mut self, &::size_policy::SizePolicy) -> ()
C++ method: void QGraphicsLayout::setSizePolicy(const QSizePolicy& policy)

C++ documentation:

Sets the size policy to policy. The size policy describes how the item should grow horizontally and vertically when arranged in a layout.

QGraphicsLayoutItem's default size policy is (QSizePolicy::Fixed, QSizePolicy::Fixed, QSizePolicy::DefaultType), but it is common for subclasses to change the default. For example, QGraphicsWidget defaults to (QSizePolicy::Preferred, QSizePolicy::Preferred, QSizePolicy::DefaultType).

See also sizePolicy() and QWidget::sizePolicy().

fn size_policy<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsLayoutSizePolicyArgs<'largs>

C++ method: QGraphicsLayout::sizePolicy

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_policy(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::size_policy::SizePolicy>
2) fn size_policy(&self, cpp_utils::AsStruct) -> ::size_policy::SizePolicy
C++ method: QSizePolicy QGraphicsLayout::sizePolicy() const

C++ documentation:

Returns the current size policy.

See also setSizePolicy() and QWidget::sizePolicy().

fn update_geometry(&mut self)

C++ method: virtual void QGraphicsLayout::updateGeometry()

C++ documentation:

Reimplemented from QGraphicsLayoutItem::updateGeometry().

fn widget_event(&mut self, e: *mut Event)

C++ method: virtual void QGraphicsLayout::widgetEvent(QEvent* e)

C++ documentation:

This virtual event handler receives all events for the managed widget. QGraphicsLayout uses this event handler to listen for layout related events such as geometry changes, layout changes or layout direction changes.

e is a pointer to the event.

You can reimplement this event handler to track similar events for your own custom layout.

See also QGraphicsWidget::event() and QGraphicsItem::sceneEvent().