Module qt_widgets::graphics_layout_item::overloading

Traits

GraphicsLayoutItemContentsRectArgs
This trait represents a set of arguments accepted by GraphicsLayoutItem::contents_rect method.

GraphicsLayoutItemEffectiveSizeHintArgs
This trait represents a set of arguments accepted by GraphicsLayoutItem::effective_size_hint method.

GraphicsLayoutItemGeometryArgs
This trait represents a set of arguments accepted by GraphicsLayoutItem::geometry method.

GraphicsLayoutItemMaximumSizeArgs
This trait represents a set of arguments accepted by GraphicsLayoutItem::maximum_size method.

GraphicsLayoutItemMinimumSizeArgs
This trait represents a set of arguments accepted by GraphicsLayoutItem::minimum_size method.

GraphicsLayoutItemPreferredSizeArgs
This trait represents a set of arguments accepted by GraphicsLayoutItem::preferred_size method.

GraphicsLayoutItemSetMaximumSizeArgs
This trait represents a set of arguments accepted by GraphicsLayoutItem::set_maximum_size method.

GraphicsLayoutItemSetMinimumSizeArgs
This trait represents a set of arguments accepted by GraphicsLayoutItem::set_minimum_size method.

GraphicsLayoutItemSetPreferredSizeArgs
This trait represents a set of arguments accepted by GraphicsLayoutItem::set_preferred_size method.

GraphicsLayoutItemSetSizePolicyArgs
This trait represents a set of arguments accepted by GraphicsLayoutItem::set_size_policy method.

GraphicsLayoutItemSizePolicyArgs
This trait represents a set of arguments accepted by GraphicsLayoutItem::size_policy method.