Struct qt_widgets::graphics_linear_layout::GraphicsLinearLayout

pub struct GraphicsLinearLayout {
    // some fields omitted
}

C++ type: QGraphicsLinearLayout

C++ documentation:

The QGraphicsLinearLayout class provides a horizontal or vertical layout for managing widgets in Graphics View.

The default orientation for a linear layout is Qt::Horizontal. You can choose a vertical orientation either by calling setOrientation(), or by passing Qt::Vertical to QGraphicsLinearLayout's constructor.

The most common way to use QGraphicsLinearLayout is to construct an object on the heap with no parent, add widgets and layouts by calling addItem(), and finally assign the layout to a widget by calling QGraphicsWidget::setLayout().

QGraphicsScene scene;
QGraphicsWidget textEdit = scene.addWidget(new QTextEdit);
QGraphicsWidget pushButton = scene.addWidget(new QPushButton);

QGraphicsLinearLayout *layout = new QGraphicsLinearLayout; layout->addItem(textEdit); layout->addItem(pushButton);

QGraphicsWidget *form = new QGraphicsWidget; form->setLayout(layout); scene.addItem(form);
You can add widgets, layouts, stretches (addStretch(), insertStretch() or setStretchFactor()), and spacings (setItemSpacing()) to a linear layout. The layout takes ownership of the items. In some cases when the layout item also inherits from QGraphicsItem (such as QGraphicsWidget) there will be a ambiguity in ownership because the layout item belongs to two ownership hierarchies. See the documentation of QGraphicsLayoutItem::setOwnedByLayout() how to handle this. You can access each item in the layout by calling count() and itemAt(). Calling removeAt() or removeItem() will remove an item from the layout, without destroying it.

Methods

impl GraphicsLinearLayout

fn activate(&mut self)

C++ method: void QGraphicsLinearLayout::activate()

C++ documentation:

Activates the layout, causing all items in the layout to be immediately rearranged. This function is based on calling count() and itemAt(), and then calling setGeometry() on all items sequentially. When activated, the layout will adjust its geometry to its parent's contentsRect(). The parent will then invalidate any layout of its own.

If called in sequence or recursively, e.g., by one of the arranged items in response to being resized, this function will do nothing.

Note that the layout is free to use geometry caching to optimize this process. To forcefully invalidate any such cache, you can call invalidate() before calling activate().

See also invalidate().

fn add_item(&mut self, item: *mut GraphicsLayoutItem)

C++ method: void QGraphicsLinearLayout::addItem(QGraphicsLayoutItem* item)

C++ documentation:

This convenience function is equivalent to calling insertItem(-1, item).

fn add_stretch<'largs, Args>(&'largs mut self, args: Args) where Args: GraphicsLinearLayoutAddStretchArgs<'largs>

C++ method: QGraphicsLinearLayout::addStretch

This is an overloaded function. Available variants:

Rust arguments:
1) fn add_stretch(&mut self, ()) -> ()
2) fn add_stretch(&mut self, libc::c_int) -> ()
C++ method: void QGraphicsLinearLayout::addStretch(int stretch = ?)

C++ documentation:

This convenience function is equivalent to calling insertStretch(-1, stretch).

fn alignment(&self, item: *mut GraphicsLayoutItem) -> Flags<AlignmentFlag>

C++ method: QFlags<Qt::AlignmentFlag> QGraphicsLinearLayout::alignment(QGraphicsLayoutItem* item) const

C++ documentation:

Returns the alignment for item. The default alignment is Qt::AlignTop | Qt::AlignLeft.

The alignment decides how the item is positioned within its assigned space in the case where there's more space available in the layout than the widgets can occupy.

See also setAlignment().

fn contents_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsLinearLayoutContentsRectArgs<'largs>

C++ method: QGraphicsLinearLayout::contentsRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn contents_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn contents_rect(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QGraphicsLinearLayout::contentsRect() const

C++ documentation:

Returns the contents rect in local coordinates.

The contents rect defines the subrectangle used by an associated layout when arranging subitems. This function is a convenience function that adjusts the item's geometry() by its contents margins. Note that getContentsMargins() is a virtual function that you can reimplement to return the item's contents margins.

See also getContentsMargins() and geometry().

fn count(&self) -> c_int

C++ method: virtual int QGraphicsLinearLayout::count() const

C++ documentation:

Reimplemented from QGraphicsLayout::count().

fn dump<'largs, Args>(&'largs self, args: Args) where Args: GraphicsLinearLayoutDumpArgs<'largs>

C++ method: QGraphicsLinearLayout::dump

This is an overloaded function. Available variants:

Rust arguments:
1) fn dump(&self, libc::c_int) -> ()
2) fn dump(&self, ()) -> ()
C++ method: void QGraphicsLinearLayout::dump(int indent = ?) const

fn effective_size_hint<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsLinearLayoutEffectiveSizeHintArgs<'largs>

C++ method: QGraphicsLinearLayout::effectiveSizeHint

This is an overloaded function. Available variants:

Rust arguments:
1) fn effective_size_hint(&self, (qt_core::qt::SizeHint, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn effective_size_hint(&self, (qt_core::qt::SizeHint, &qt_core::size_f::SizeF, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
3) fn effective_size_hint(&self, (qt_core::qt::SizeHint, cpp_utils::AsStruct)) -> qt_core::size_f::SizeF
4) fn effective_size_hint(&self, (qt_core::qt::SizeHint, &qt_core::size_f::SizeF, cpp_utils::AsStruct)) -> qt_core::size_f::SizeF
C++ method: QSizeF QGraphicsLinearLayout::effectiveSizeHint(Qt::SizeHint which, const QSizeF& constraint = ?) const

C++ documentation:

Returns the effective size hint for this QGraphicsLayoutItem.

which is the size hint in question. constraint is an optional argument that defines a special constrain when calculating the effective size hint. By default, constraint is QSizeF(-1, -1), which means there is no constraint to the size hint.

If you want to specify the widget's size hint for a given width or height, you can provide the fixed dimension in constraint. This is useful for widgets that can grow only either vertically or horizontally, and need to set either their width or their height to a special value.

For example, a text paragraph item fit into a column width of 200 may grow vertically. You can pass QSizeF(200, -1) as a constraint to get a suitable minimum, preferred and maximum height).

You can adjust the effective size hint by reimplementing sizeHint() in a QGraphicsLayoutItem subclass, or by calling one of the following functions: setMinimumSize(), setPreferredSize, or setMaximumSize() (or a combination of both).

This function caches each of the size hints and guarantees that sizeHint() will be called only once for each value of which - unless constraint is not specified and updateGeometry() has been called.

See also sizeHint().

fn geometry<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsLinearLayoutGeometryArgs<'largs>

C++ method: QGraphicsLinearLayout::geometry

This is an overloaded function. Available variants:

Rust arguments:
1) fn geometry(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn geometry(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QGraphicsLinearLayout::geometry() const

C++ documentation:

Returns the item's geometry (e.g., position and size) as a QRectF. This function is equivalent to QRectF(pos(), size()).

See also setGeometry().

fn get_contents_margins(&self, left: *mut c_double, top: *mut c_double, right: *mut c_double, bottom: *mut c_double)

C++ method: virtual void QGraphicsLinearLayout::getContentsMargins(double* left, double* top, double* right, double* bottom) const

C++ documentation:

Reimplemented from QGraphicsLayoutItem::getContentsMargins().

fn graphics_item(&self) -> *mut GraphicsItem

C++ method: QGraphicsItem* QGraphicsLinearLayout::graphicsItem() const

C++ documentation:

Returns the QGraphicsItem that this layout item represents. For QGraphicsWidget it will return itself. For custom items it can return an aggregated value.

See also setGraphicsItem().

fn insert_item(&mut self, index: c_int, item: *mut GraphicsLayoutItem)

C++ method: void QGraphicsLinearLayout::insertItem(int index, QGraphicsLayoutItem* item)

C++ documentation:

Inserts item into the layout at index, or before any item that is currently at index.

See also addItem(), itemAt(), insertStretch(), and setItemSpacing().

fn insert_stretch<'largs, Args>(&'largs mut self, args: Args) where Args: GraphicsLinearLayoutInsertStretchArgs<'largs>

C++ method: QGraphicsLinearLayout::insertStretch

This is an overloaded function. Available variants:

Rust arguments:
1) fn insert_stretch(&mut self, libc::c_int) -> ()
2) fn insert_stretch(&mut self, (libc::c_int, libc::c_int)) -> ()
C++ method: void QGraphicsLinearLayout::insertStretch(int index, int stretch = ?)

C++ documentation:

Inserts a stretch of stretch at index, or before any item that is currently at index.

See also addStretch(), setStretchFactor(), setItemSpacing(), and insertItem().

fn instant_invalidate_propagation() -> bool

C++ method: static bool QGraphicsLinearLayout::instantInvalidatePropagation()

fn invalidate(&mut self)

C++ method: virtual void QGraphicsLinearLayout::invalidate()

C++ documentation:

Reimplemented from QGraphicsLayout::invalidate().

fn is_activated(&self) -> bool

C++ method: bool QGraphicsLinearLayout::isActivated() const

C++ documentation:

Returns true if the layout is currently being activated; otherwise, returns false. If the layout is being activated, this means that it is currently in the process of rearranging its items (i.e., the activate() function has been called, and has not yet returned).

See also activate() and invalidate().

fn is_layout(&self) -> bool

C++ method: bool QGraphicsLinearLayout::isLayout() const

C++ documentation:

Returns true if this QGraphicsLayoutItem is a layout (e.g., is inherited by an object that arranges other QGraphicsLayoutItem objects); otherwise returns false.

See also QGraphicsLayout.

fn item_at(&self, index: c_int) -> *mut GraphicsLayoutItem

C++ method: virtual QGraphicsLayoutItem* QGraphicsLinearLayout::itemAt(int index) const

C++ documentation:

Reimplemented from QGraphicsLayout::itemAt().

When iterating from 0 and up, it will return the items in the visual arranged order.

fn item_spacing(&self, index: c_int) -> c_double

C++ method: double QGraphicsLinearLayout::itemSpacing(int index) const

C++ documentation:

Returns the spacing after item at index.

See also setItemSpacing().

fn maximum_height(&self) -> c_double

C++ method: double QGraphicsLinearLayout::maximumHeight() const

C++ documentation:

Returns the maximum height.

See also setMaximumHeight(), setMaximumSize(), and maximumSize().

fn maximum_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsLinearLayoutMaximumSizeArgs<'largs>

C++ method: QGraphicsLinearLayout::maximumSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn maximum_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn maximum_size(&self, cpp_utils::AsStruct) -> qt_core::size_f::SizeF
C++ method: QSizeF QGraphicsLinearLayout::maximumSize() const

C++ documentation:

Returns the maximum size.

See also setMaximumSize(), minimumSize(), preferredSize(), Qt::MaximumSize, and sizeHint().

fn maximum_width(&self) -> c_double

C++ method: double QGraphicsLinearLayout::maximumWidth() const

C++ documentation:

Returns the maximum width.

See also setMaximumWidth(), setMaximumSize(), and maximumSize().

fn minimum_height(&self) -> c_double

C++ method: double QGraphicsLinearLayout::minimumHeight() const

C++ documentation:

Returns the minimum height.

See also setMinimumHeight(), setMinimumSize(), and minimumSize().

fn minimum_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsLinearLayoutMinimumSizeArgs<'largs>

C++ method: QGraphicsLinearLayout::minimumSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn minimum_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn minimum_size(&self, cpp_utils::AsStruct) -> qt_core::size_f::SizeF
C++ method: QSizeF QGraphicsLinearLayout::minimumSize() const

C++ documentation:

Returns the minimum size.

See also setMinimumSize(), preferredSize(), maximumSize(), Qt::MinimumSize, and sizeHint().

fn minimum_width(&self) -> c_double

C++ method: double QGraphicsLinearLayout::minimumWidth() const

C++ documentation:

Returns the minimum width.

See also setMinimumWidth(), setMinimumSize(), and minimumSize().

fn new<Args>(args: Args) -> Args::ReturnType where Args: GraphicsLinearLayoutNewArgs

C++ method: QGraphicsLinearLayout::QGraphicsLinearLayout

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::graphics_linear_layout::GraphicsLinearLayout
2) fn new((*mut ::graphics_layout_item::GraphicsLayoutItem, cpp_utils::AsStruct)) -> ::graphics_linear_layout::GraphicsLinearLayout
3) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::graphics_linear_layout::GraphicsLinearLayout>
4) fn new((*mut ::graphics_layout_item::GraphicsLayoutItem, cpp_utils::AsBox)) -> cpp_utils::CppBox<::graphics_linear_layout::GraphicsLinearLayout>
C++ method: [constructor] void QGraphicsLinearLayout::QGraphicsLinearLayout(QGraphicsLayoutItem* parent = ?)

Variant 2

Rust arguments:
1) fn new((qt_core::qt::Orientation, cpp_utils::AsStruct)) -> ::graphics_linear_layout::GraphicsLinearLayout
2) fn new((qt_core::qt::Orientation, *mut ::graphics_layout_item::GraphicsLayoutItem, cpp_utils::AsStruct)) -> ::graphics_linear_layout::GraphicsLinearLayout
3) fn new((qt_core::qt::Orientation, cpp_utils::AsBox)) -> cpp_utils::CppBox<::graphics_linear_layout::GraphicsLinearLayout>
4) fn new((qt_core::qt::Orientation, *mut ::graphics_layout_item::GraphicsLayoutItem, cpp_utils::AsBox)) -> cpp_utils::CppBox<::graphics_linear_layout::GraphicsLinearLayout>
C++ method: [constructor] void QGraphicsLinearLayout::QGraphicsLinearLayout(Qt::Orientation orientation, QGraphicsLayoutItem* parent = ?)

fn orientation(&self) -> Orientation

C++ method: Qt::Orientation QGraphicsLinearLayout::orientation() const

C++ documentation:

Returns the layout orientation.

See also setOrientation().

fn owned_by_layout(&self) -> bool

C++ method: bool QGraphicsLinearLayout::ownedByLayout() const

C++ documentation:

Returns whether a layout should delete this item in its destructor. If its true, then the layout will delete it. If its false, then it is assumed that another object has the ownership of it, and the layout won't delete this item.

If the item inherits both QGraphicsItem and QGraphicsLayoutItem (such as QGraphicsWidget does) the item is really part of two ownership hierarchies. This property informs what the layout should do with its child items when it is destructed. In the case of QGraphicsWidget, it is preferred that when the layout is deleted it won't delete its children (since they are also part of the graphics item hierarchy).

By default this value is initialized to false in QGraphicsLayoutItem, but it is overridden by QGraphicsLayout to return true. This is because QGraphicsLayout is not normally part of the QGraphicsItem hierarchy, so the parent layout should delete it. Subclasses might override this default behaviour by calling setOwnedByLayout(true).

This function was introduced in Qt 4.6.

See also setOwnedByLayout().

fn parent_layout_item(&self) -> *mut GraphicsLayoutItem

C++ method: QGraphicsLayoutItem* QGraphicsLinearLayout::parentLayoutItem() const

C++ documentation:

Returns the parent of this QGraphicsLayoutItem, or 0 if there is no parent, or if the parent does not inherit from QGraphicsLayoutItem (QGraphicsLayoutItem is often used through multiple inheritance with QObject-derived classes).

See also setParentLayoutItem().

fn preferred_height(&self) -> c_double

C++ method: double QGraphicsLinearLayout::preferredHeight() const

C++ documentation:

Returns the preferred height.

See also setPreferredHeight(), setPreferredSize(), and preferredSize().

fn preferred_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsLinearLayoutPreferredSizeArgs<'largs>

C++ method: QGraphicsLinearLayout::preferredSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn preferred_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn preferred_size(&self, cpp_utils::AsStruct) -> qt_core::size_f::SizeF
C++ method: QSizeF QGraphicsLinearLayout::preferredSize() const

C++ documentation:

Returns the preferred size.

See also setPreferredSize(), minimumSize(), maximumSize(), Qt::PreferredSize, and sizeHint().

fn preferred_width(&self) -> c_double

C++ method: double QGraphicsLinearLayout::preferredWidth() const

C++ documentation:

Returns the preferred width.

See also setPreferredWidth(), setPreferredSize(), and preferredSize().

fn remove_at(&mut self, index: c_int)

C++ method: virtual void QGraphicsLinearLayout::removeAt(int index)

C++ documentation:

Reimplemented from QGraphicsLayout::removeAt().

Removes the item at index without destroying it. Ownership of the item is transferred to the caller.

See also removeItem() and insertItem().

fn remove_item(&mut self, item: *mut GraphicsLayoutItem)

C++ method: void QGraphicsLinearLayout::removeItem(QGraphicsLayoutItem* item)

C++ documentation:

Removes item from the layout without destroying it. Ownership of item is transferred to the caller.

See also removeAt() and insertItem().

fn set_alignment(&mut self, item: *mut GraphicsLayoutItem, alignment: Flags<AlignmentFlag>)

C++ method: void QGraphicsLinearLayout::setAlignment(QGraphicsLayoutItem* item, QFlags<Qt::AlignmentFlag> alignment)

C++ documentation:

Sets the alignment of item to alignment. If item's alignment changes, the layout is automatically invalidated.

See also alignment() and invalidate().

fn set_contents_margins(&mut self, left: c_double, top: c_double, right: c_double, bottom: c_double)

C++ method: void QGraphicsLinearLayout::setContentsMargins(double left, double top, double right, double bottom)

C++ documentation:

Sets the contents margins to left, top, right and bottom. The default contents margins for toplevel layouts are style dependent (by querying the pixelMetric for QStyle::PM_LayoutLeftMargin, QStyle::PM_LayoutTopMargin, QStyle::PM_LayoutRightMargin and QStyle::PM_LayoutBottomMargin).

For sublayouts the default margins are 0.

Changing the contents margins automatically invalidates the layout.

See also invalidate().

fn set_geometry(&mut self, rect: &RectF)

C++ method: virtual void QGraphicsLinearLayout::setGeometry(const QRectF& rect)

C++ documentation:

Reimplemented from QGraphicsLayoutItem::setGeometry().

fn set_instant_invalidate_propagation(enable: bool)

C++ method: static void QGraphicsLinearLayout::setInstantInvalidatePropagation(bool enable)

fn set_item_spacing(&mut self, index: c_int, spacing: c_double)

C++ method: void QGraphicsLinearLayout::setItemSpacing(int index, double spacing)

C++ documentation:

Sets the spacing after item at index to spacing.

See also itemSpacing().

fn set_maximum_height(&mut self, height: c_double)

C++ method: void QGraphicsLinearLayout::setMaximumHeight(double height)

C++ documentation:

Sets the maximum height to height.

See also maximumHeight(), setMaximumSize(), and maximumSize().

fn set_maximum_size<'largs, Args>(&'largs mut self, args: Args) where Args: GraphicsLinearLayoutSetMaximumSizeArgs<'largs>

C++ method: QGraphicsLinearLayout::setMaximumSize

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_maximum_size(&mut self, &qt_core::size_f::SizeF) -> ()
C++ method: void QGraphicsLinearLayout::setMaximumSize(const QSizeF& size)

C++ documentation:

Sets the maximum size to size. This property overrides sizeHint() for Qt::MaximumSize and ensures that effectiveSizeHint() will never return a size larger than size. In order to unset the maximum size, use an invalid size.

See also maximumSize(), minimumSize(), preferredSize(), Qt::MaximumSize, and sizeHint().

Variant 2

Rust arguments: fn set_maximum_size(&mut self, (libc::c_double, libc::c_double)) -> ()
C++ method: void QGraphicsLinearLayout::setMaximumSize(double w, double h)

C++ documentation:

This convenience function is equivalent to calling setMaximumSize(QSizeF(w, h)).

See also maximumSize(), setMinimumSize(), setPreferredSize(), and sizeHint().

fn set_maximum_width(&mut self, width: c_double)

C++ method: void QGraphicsLinearLayout::setMaximumWidth(double width)

C++ documentation:

Sets the maximum width to width.

See also maximumWidth(), setMaximumSize(), and maximumSize().

fn set_minimum_height(&mut self, height: c_double)

C++ method: void QGraphicsLinearLayout::setMinimumHeight(double height)

C++ documentation:

Sets the minimum height to height.

See also minimumHeight(), setMinimumSize(), and minimumSize().

fn set_minimum_size<'largs, Args>(&'largs mut self, args: Args) where Args: GraphicsLinearLayoutSetMinimumSizeArgs<'largs>

C++ method: QGraphicsLinearLayout::setMinimumSize

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_minimum_size(&mut self, &qt_core::size_f::SizeF) -> ()
C++ method: void QGraphicsLinearLayout::setMinimumSize(const QSizeF& size)

C++ documentation:

Sets the minimum size to size. This property overrides sizeHint() for Qt::MinimumSize and ensures that effectiveSizeHint() will never return a size smaller than size. In order to unset the minimum size, use an invalid size.

See also minimumSize(), maximumSize(), preferredSize(), Qt::MinimumSize, sizeHint(), setMinimumWidth(), and setMinimumHeight().

Variant 2

Rust arguments: fn set_minimum_size(&mut self, (libc::c_double, libc::c_double)) -> ()
C++ method: void QGraphicsLinearLayout::setMinimumSize(double w, double h)

C++ documentation:

This convenience function is equivalent to calling setMinimumSize(QSizeF(w, h)).

See also minimumSize(), setMaximumSize(), setPreferredSize(), and sizeHint().

fn set_minimum_width(&mut self, width: c_double)

C++ method: void QGraphicsLinearLayout::setMinimumWidth(double width)

C++ documentation:

Sets the minimum width to width.

See also minimumWidth(), setMinimumSize(), and minimumSize().

fn set_orientation(&mut self, orientation: Orientation)

C++ method: void QGraphicsLinearLayout::setOrientation(Qt::Orientation orientation)

C++ documentation:

Change the layout orientation to orientation. Changing the layout orientation will automatically invalidate the layout.

See also orientation().

fn set_parent_layout_item(&mut self, parent: *mut GraphicsLayoutItem)

C++ method: void QGraphicsLinearLayout::setParentLayoutItem(QGraphicsLayoutItem* parent)

C++ documentation:

Sets the parent of this QGraphicsLayoutItem to parent.

See also parentLayoutItem().

fn set_preferred_height(&mut self, height: c_double)

C++ method: void QGraphicsLinearLayout::setPreferredHeight(double height)

C++ documentation:

Sets the preferred height to height.

See also preferredHeight(), preferredWidth(), setPreferredSize(), and preferredSize().

fn set_preferred_size<'largs, Args>(&'largs mut self, args: Args) where Args: GraphicsLinearLayoutSetPreferredSizeArgs<'largs>

C++ method: QGraphicsLinearLayout::setPreferredSize

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_preferred_size(&mut self, &qt_core::size_f::SizeF) -> ()
C++ method: void QGraphicsLinearLayout::setPreferredSize(const QSizeF& size)

C++ documentation:

Sets the preferred size to size. This property overrides sizeHint() for Qt::PreferredSize and provides the default value for effectiveSizeHint(). In order to unset the preferred size, use an invalid size.

See also preferredSize(), minimumSize(), maximumSize(), Qt::PreferredSize, and sizeHint().

Variant 2

Rust arguments: fn set_preferred_size(&mut self, (libc::c_double, libc::c_double)) -> ()
C++ method: void QGraphicsLinearLayout::setPreferredSize(double w, double h)

C++ documentation:

This convenience function is equivalent to calling setPreferredSize(QSizeF(w, h)).

See also preferredSize(), setMaximumSize(), setMinimumSize(), and sizeHint().

fn set_preferred_width(&mut self, width: c_double)

C++ method: void QGraphicsLinearLayout::setPreferredWidth(double width)

C++ documentation:

Sets the preferred width to width.

See also preferredWidth(), preferredHeight(), setPreferredSize(), and preferredSize().

fn set_size_policy<'largs, Args>(&'largs mut self, args: Args) where Args: GraphicsLinearLayoutSetSizePolicyArgs<'largs>

C++ method: QGraphicsLinearLayout::setSizePolicy

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn set_size_policy(&mut self, (::size_policy::Policy, ::size_policy::Policy)) -> ()
2) fn set_size_policy(&mut self, (::size_policy::Policy, ::size_policy::Policy, ::size_policy::ControlType)) -> ()
C++ method: void QGraphicsLinearLayout::setSizePolicy(QSizePolicy::Policy hPolicy, QSizePolicy::Policy vPolicy, QSizePolicy::ControlType controlType = ?)

C++ documentation:

This is an overloaded function.

This function is equivalent to calling setSizePolicy(QSizePolicy(hPolicy, vPolicy, controlType)).

See also sizePolicy() and QWidget::sizePolicy().

Variant 2

Rust arguments: fn set_size_policy(&mut self, &::size_policy::SizePolicy) -> ()
C++ method: void QGraphicsLinearLayout::setSizePolicy(const QSizePolicy& policy)

C++ documentation:

Sets the size policy to policy. The size policy describes how the item should grow horizontally and vertically when arranged in a layout.

QGraphicsLayoutItem's default size policy is (QSizePolicy::Fixed, QSizePolicy::Fixed, QSizePolicy::DefaultType), but it is common for subclasses to change the default. For example, QGraphicsWidget defaults to (QSizePolicy::Preferred, QSizePolicy::Preferred, QSizePolicy::DefaultType).

See also sizePolicy() and QWidget::sizePolicy().

fn set_spacing(&mut self, spacing: c_double)

C++ method: void QGraphicsLinearLayout::setSpacing(double spacing)

C++ documentation:

Sets the layout's spacing to spacing. Spacing refers to the vertical and horizontal distances between items.

See also spacing(), setItemSpacing(), setStretchFactor(), and QGraphicsGridLayout::setSpacing().

fn set_stretch_factor(&mut self, item: *mut GraphicsLayoutItem, stretch: c_int)

C++ method: void QGraphicsLinearLayout::setStretchFactor(QGraphicsLayoutItem* item, int stretch)

C++ documentation:

Sets the stretch factor for item to stretch. If an item's stretch factor changes, this function will invalidate the layout.

Setting stretch to 0 removes the stretch factor from the item, and is effectively equivalent to setting stretch to 1.

See also stretchFactor().

fn size_hint<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsLinearLayoutSizeHintArgs<'largs>

C++ method: QGraphicsLinearLayout::sizeHint

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_hint(&self, (qt_core::qt::SizeHint, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn size_hint(&self, (qt_core::qt::SizeHint, &qt_core::size_f::SizeF, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
3) fn size_hint(&self, (qt_core::qt::SizeHint, cpp_utils::AsStruct)) -> qt_core::size_f::SizeF
4) fn size_hint(&self, (qt_core::qt::SizeHint, &qt_core::size_f::SizeF, cpp_utils::AsStruct)) -> qt_core::size_f::SizeF
C++ method: virtual QSizeF QGraphicsLinearLayout::sizeHint(Qt::SizeHint which, const QSizeF& constraint = ?) const

C++ documentation:

Reimplemented from QGraphicsLayoutItem::sizeHint().

fn size_policy<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsLinearLayoutSizePolicyArgs<'largs>

C++ method: QGraphicsLinearLayout::sizePolicy

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_policy(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::size_policy::SizePolicy>
2) fn size_policy(&self, cpp_utils::AsStruct) -> ::size_policy::SizePolicy
C++ method: QSizePolicy QGraphicsLinearLayout::sizePolicy() const

C++ documentation:

Returns the current size policy.

See also setSizePolicy() and QWidget::sizePolicy().

fn spacing(&self) -> c_double

C++ method: double QGraphicsLinearLayout::spacing() const

C++ documentation:

Returns the layout's spacing. Spacing refers to the vertical and horizontal distances between items.

See also setSpacing().

fn stretch_factor(&self, item: *mut GraphicsLayoutItem) -> c_int

C++ method: int QGraphicsLinearLayout::stretchFactor(QGraphicsLayoutItem* item) const

C++ documentation:

Returns the stretch factor for item. The default stretch factor is 0, meaning that the item has no assigned stretch factor.

See also setStretchFactor().

fn update_geometry(&mut self)

C++ method: virtual void QGraphicsLinearLayout::updateGeometry()

C++ documentation:

Reimplemented from QGraphicsLayoutItem::updateGeometry().

fn widget_event(&mut self, e: *mut Event)

C++ method: virtual void QGraphicsLinearLayout::widgetEvent(QEvent* e)

C++ documentation:

This virtual event handler receives all events for the managed widget. QGraphicsLayout uses this event handler to listen for layout related events such as geometry changes, layout changes or layout direction changes.

e is a pointer to the event.

You can reimplement this event handler to track similar events for your own custom layout.

See also QGraphicsWidget::event() and QGraphicsItem::sceneEvent().