Module qt_widgets::graphics_object::overloading

Traits

GraphicsObjectBoundingRectArgs
This trait represents a set of arguments accepted by GraphicsObject::bounding_rect method.

GraphicsObjectBoundingRegionArgs
This trait represents a set of arguments accepted by GraphicsObject::bounding_region method.

GraphicsObjectChildItemsArgs
This trait represents a set of arguments accepted by GraphicsObject::child_items method.

GraphicsObjectChildrenBoundingRectArgs
This trait represents a set of arguments accepted by GraphicsObject::children_bounding_rect method.

GraphicsObjectClipPathArgs
This trait represents a set of arguments accepted by GraphicsObject::clip_path method.

GraphicsObjectCollidesWithItemArgs
This trait represents a set of arguments accepted by GraphicsObject::collides_with_item method.

GraphicsObjectCollidesWithPathArgs
This trait represents a set of arguments accepted by GraphicsObject::collides_with_path method.

GraphicsObjectCollidingItemsArgs
This trait represents a set of arguments accepted by GraphicsObject::colliding_items method.

GraphicsObjectConnectArgs
This trait represents a set of arguments accepted by GraphicsObject::connect method.

GraphicsObjectConnectStaticArgs
This trait represents a set of arguments accepted by GraphicsObject::connect_static method.

GraphicsObjectCursorArgs
This trait represents a set of arguments accepted by GraphicsObject::cursor method.

GraphicsObjectDataArgs
This trait represents a set of arguments accepted by GraphicsObject::data method.

GraphicsObjectDestroyedArgs
This trait represents a set of arguments accepted by GraphicsObject::destroyed method.

GraphicsObjectDeviceTransformArgs
This trait represents a set of arguments accepted by GraphicsObject::device_transform method.

GraphicsObjectDisconnectArgs
This trait represents a set of arguments accepted by GraphicsObject::disconnect method.

GraphicsObjectDisconnectStaticArgs
This trait represents a set of arguments accepted by GraphicsObject::disconnect_static method.

GraphicsObjectDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by GraphicsObject::dynamic_property_names method.

GraphicsObjectEnsureVisibleArgs
This trait represents a set of arguments accepted by GraphicsObject::ensure_visible method.

GraphicsObjectGrabGestureArgs
This trait represents a set of arguments accepted by GraphicsObject::grab_gesture method.

GraphicsObjectIsBlockedByModalPanelArgs
This trait represents a set of arguments accepted by GraphicsObject::is_blocked_by_modal_panel method.

GraphicsObjectIsObscuredArgs
This trait represents a set of arguments accepted by GraphicsObject::is_obscured method.

GraphicsObjectItemTransformArgs
This trait represents a set of arguments accepted by GraphicsObject::item_transform method.

GraphicsObjectMapFromItemArgs
This trait represents a set of arguments accepted by GraphicsObject::map_from_item method.

GraphicsObjectMapFromParentArgs
This trait represents a set of arguments accepted by GraphicsObject::map_from_parent method.

GraphicsObjectMapFromSceneArgs
This trait represents a set of arguments accepted by GraphicsObject::map_from_scene method.

GraphicsObjectMapRectFromItemArgs
This trait represents a set of arguments accepted by GraphicsObject::map_rect_from_item method.

GraphicsObjectMapRectFromParentArgs
This trait represents a set of arguments accepted by GraphicsObject::map_rect_from_parent method.

GraphicsObjectMapRectFromSceneArgs
This trait represents a set of arguments accepted by GraphicsObject::map_rect_from_scene method.

GraphicsObjectMapRectToItemArgs
This trait represents a set of arguments accepted by GraphicsObject::map_rect_to_item method.

GraphicsObjectMapRectToParentArgs
This trait represents a set of arguments accepted by GraphicsObject::map_rect_to_parent method.

GraphicsObjectMapRectToSceneArgs
This trait represents a set of arguments accepted by GraphicsObject::map_rect_to_scene method.

GraphicsObjectMapToItemArgs
This trait represents a set of arguments accepted by GraphicsObject::map_to_item method.

GraphicsObjectMapToParentArgs
This trait represents a set of arguments accepted by GraphicsObject::map_to_parent method.

GraphicsObjectMapToSceneArgs
This trait represents a set of arguments accepted by GraphicsObject::map_to_scene method.

GraphicsObjectMatrixArgs
This trait represents a set of arguments accepted by GraphicsObject::matrix method.

GraphicsObjectObjectNameArgs
This trait represents a set of arguments accepted by GraphicsObject::object_name method.

GraphicsObjectOpaqueAreaArgs
This trait represents a set of arguments accepted by GraphicsObject::opaque_area method.

GraphicsObjectPaintArgs
This trait represents a set of arguments accepted by GraphicsObject::paint method.

GraphicsObjectPosArgs
This trait represents a set of arguments accepted by GraphicsObject::pos method.

GraphicsObjectPropertyArgs
This trait represents a set of arguments accepted by GraphicsObject::property method.

GraphicsObjectSceneBoundingRectArgs
This trait represents a set of arguments accepted by GraphicsObject::scene_bounding_rect method.

GraphicsObjectSceneMatrixArgs
This trait represents a set of arguments accepted by GraphicsObject::scene_matrix method.

GraphicsObjectScenePosArgs
This trait represents a set of arguments accepted by GraphicsObject::scene_pos method.

GraphicsObjectSceneTransformArgs
This trait represents a set of arguments accepted by GraphicsObject::scene_transform method.

GraphicsObjectScrollArgs
This trait represents a set of arguments accepted by GraphicsObject::scroll method.

GraphicsObjectSetCacheModeArgs
This trait represents a set of arguments accepted by GraphicsObject::set_cache_mode method.

GraphicsObjectSetFlagArgs
This trait represents a set of arguments accepted by GraphicsObject::set_flag method.

GraphicsObjectSetFocusArgs
This trait represents a set of arguments accepted by GraphicsObject::set_focus method.

GraphicsObjectSetMatrixArgs
This trait represents a set of arguments accepted by GraphicsObject::set_matrix method.

GraphicsObjectSetPosArgs
This trait represents a set of arguments accepted by GraphicsObject::set_pos method.

GraphicsObjectSetTransformArgs
This trait represents a set of arguments accepted by GraphicsObject::set_transform method.

GraphicsObjectSetTransformOriginPointArgs
This trait represents a set of arguments accepted by GraphicsObject::set_transform_origin_point method.

GraphicsObjectShapeArgs
This trait represents a set of arguments accepted by GraphicsObject::shape method.

GraphicsObjectStartTimerArgs
This trait represents a set of arguments accepted by GraphicsObject::start_timer method.

GraphicsObjectToolTipArgs
This trait represents a set of arguments accepted by GraphicsObject::tool_tip method.

GraphicsObjectTrArgs
This trait represents a set of arguments accepted by GraphicsObject::tr method.

GraphicsObjectTrUtf8Args
This trait represents a set of arguments accepted by GraphicsObject::tr_utf8 method.

GraphicsObjectTransformArgs
This trait represents a set of arguments accepted by GraphicsObject::transform method.

GraphicsObjectTransformOriginPointArgs
This trait represents a set of arguments accepted by GraphicsObject::transform_origin_point method.

GraphicsObjectTransformationsArgs
This trait represents a set of arguments accepted by GraphicsObject::transformations method.

GraphicsObjectUpdateArgs
This trait represents a set of arguments accepted by GraphicsObject::update method.