Module qt_widgets::graphics_opacity_effect::overloading

Traits

GraphicsOpacityEffectBoundingRectArgs
This trait represents a set of arguments accepted by GraphicsOpacityEffect::bounding_rect method.

GraphicsOpacityEffectBoundingRectForArgs
This trait represents a set of arguments accepted by GraphicsOpacityEffect::bounding_rect_for method.

GraphicsOpacityEffectConnectArgs
This trait represents a set of arguments accepted by GraphicsOpacityEffect::connect method.

GraphicsOpacityEffectConnectStaticArgs
This trait represents a set of arguments accepted by GraphicsOpacityEffect::connect_static method.

GraphicsOpacityEffectDestroyedArgs
This trait represents a set of arguments accepted by GraphicsOpacityEffect::destroyed method.

GraphicsOpacityEffectDisconnectArgs
This trait represents a set of arguments accepted by GraphicsOpacityEffect::disconnect method.

GraphicsOpacityEffectDisconnectStaticArgs
This trait represents a set of arguments accepted by GraphicsOpacityEffect::disconnect_static method.

GraphicsOpacityEffectDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by GraphicsOpacityEffect::dynamic_property_names method.

GraphicsOpacityEffectNewArgs
This trait represents a set of arguments accepted by GraphicsOpacityEffect::new method.

GraphicsOpacityEffectObjectNameArgs
This trait represents a set of arguments accepted by GraphicsOpacityEffect::object_name method.

GraphicsOpacityEffectOpacityMaskArgs
This trait represents a set of arguments accepted by GraphicsOpacityEffect::opacity_mask method.

GraphicsOpacityEffectPropertyArgs
This trait represents a set of arguments accepted by GraphicsOpacityEffect::property method.

GraphicsOpacityEffectStartTimerArgs
This trait represents a set of arguments accepted by GraphicsOpacityEffect::start_timer method.

GraphicsOpacityEffectTrArgs
This trait represents a set of arguments accepted by GraphicsOpacityEffect::tr method.

GraphicsOpacityEffectTrUtf8Args
This trait represents a set of arguments accepted by GraphicsOpacityEffect::tr_utf8 method.