Enum qt_widgets::graphics_pixmap_item::ShapeMode

pub enum ShapeMode {
    Mask,
    BoundingRect,
    HeuristicMask,
}

C++ type: QGraphicsPixmapItem::ShapeMode

C++ documentation:

This enum describes how QGraphicsPixmapItem calculates its shape and opaque area.

The default value is MaskShape.

Variants
Mask

The shape is determined by calling QPixmap::mask(). This shape includes only the opaque pixels of the pixmap. Because the shape is more complex, however, it can be slower than the other modes, and uses more memory. (C++ enum variant: MaskShape = 0)

BoundingRect

The shape is determined by tracing the outline of the pixmap. This is the fastest shape mode, but it does not take into account any transparent areas on the pixmap. (C++ enum variant: BoundingRectShape = 1)

HeuristicMask

The shape is determine by calling QPixmap::createHeuristicMask(). The performance and memory consumption is similar to MaskShape. (C++ enum variant: HeuristicMaskShape = 2)