Module qt_widgets::graphics_proxy_widget::overloading

Traits

GraphicsProxyWidgetActionsArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::actions method.

GraphicsProxyWidgetBoundingRectArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::bounding_rect method.

GraphicsProxyWidgetBoundingRegionArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::bounding_region method.

GraphicsProxyWidgetChildItemsArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::child_items method.

GraphicsProxyWidgetChildrenBoundingRectArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::children_bounding_rect method.

GraphicsProxyWidgetClipPathArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::clip_path method.

GraphicsProxyWidgetCollidesWithItemArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::collides_with_item method.

GraphicsProxyWidgetCollidesWithPathArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::collides_with_path method.

GraphicsProxyWidgetCollidingItemsArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::colliding_items method.

GraphicsProxyWidgetConnectArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::connect method.

GraphicsProxyWidgetConnectStaticArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::connect_static method.

GraphicsProxyWidgetContentsRectArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::contents_rect method.

GraphicsProxyWidgetCursorArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::cursor method.

GraphicsProxyWidgetDataArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::data method.

GraphicsProxyWidgetDestroyedArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::destroyed method.

GraphicsProxyWidgetDeviceTransformArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::device_transform method.

GraphicsProxyWidgetDisconnectArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::disconnect method.

GraphicsProxyWidgetDisconnectStaticArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::disconnect_static method.

GraphicsProxyWidgetDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::dynamic_property_names method.

GraphicsProxyWidgetEffectiveSizeHintArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::effective_size_hint method.

GraphicsProxyWidgetEnsureVisibleArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::ensure_visible method.

GraphicsProxyWidgetFontArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::font method.

GraphicsProxyWidgetGeometryArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::geometry method.

GraphicsProxyWidgetGrabGestureArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::grab_gesture method.

GraphicsProxyWidgetGrabShortcutArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::grab_shortcut method.

GraphicsProxyWidgetIsBlockedByModalPanelArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::is_blocked_by_modal_panel method.

GraphicsProxyWidgetIsObscuredArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::is_obscured method.

GraphicsProxyWidgetItemTransformArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::item_transform method.

GraphicsProxyWidgetMapFromItemArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::map_from_item method.

GraphicsProxyWidgetMapFromParentArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::map_from_parent method.

GraphicsProxyWidgetMapFromSceneArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::map_from_scene method.

GraphicsProxyWidgetMapRectFromItemArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::map_rect_from_item method.

GraphicsProxyWidgetMapRectFromParentArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::map_rect_from_parent method.

GraphicsProxyWidgetMapRectFromSceneArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::map_rect_from_scene method.

GraphicsProxyWidgetMapRectToItemArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::map_rect_to_item method.

GraphicsProxyWidgetMapRectToParentArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::map_rect_to_parent method.

GraphicsProxyWidgetMapRectToSceneArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::map_rect_to_scene method.

GraphicsProxyWidgetMapToItemArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::map_to_item method.

GraphicsProxyWidgetMapToParentArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::map_to_parent method.

GraphicsProxyWidgetMapToSceneArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::map_to_scene method.

GraphicsProxyWidgetMatrixArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::matrix method.

GraphicsProxyWidgetMaximumSizeArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::maximum_size method.

GraphicsProxyWidgetMinimumSizeArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::minimum_size method.

GraphicsProxyWidgetNewArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::new method.

GraphicsProxyWidgetObjectNameArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::object_name method.

GraphicsProxyWidgetOpaqueAreaArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::opaque_area method.

GraphicsProxyWidgetPaintWindowFrameArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::paint_window_frame method.

GraphicsProxyWidgetPaletteArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::palette method.

GraphicsProxyWidgetPosArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::pos method.

GraphicsProxyWidgetPreferredSizeArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::preferred_size method.

GraphicsProxyWidgetPropertyArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::property method.

GraphicsProxyWidgetRectArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::rect method.

GraphicsProxyWidgetResizeArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::resize method.

GraphicsProxyWidgetSceneBoundingRectArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::scene_bounding_rect method.

GraphicsProxyWidgetSceneMatrixArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::scene_matrix method.

GraphicsProxyWidgetScenePosArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::scene_pos method.

GraphicsProxyWidgetSceneTransformArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::scene_transform method.

GraphicsProxyWidgetScrollArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::scroll method.

GraphicsProxyWidgetSetAttributeArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::set_attribute method.

GraphicsProxyWidgetSetCacheModeArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::set_cache_mode method.

GraphicsProxyWidgetSetFlagArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::set_flag method.

GraphicsProxyWidgetSetFocusArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::set_focus method.

GraphicsProxyWidgetSetMatrixArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::set_matrix method.

GraphicsProxyWidgetSetMaximumSizeArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::set_maximum_size method.

GraphicsProxyWidgetSetMinimumSizeArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::set_minimum_size method.

GraphicsProxyWidgetSetPosArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::set_pos method.

GraphicsProxyWidgetSetPreferredSizeArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::set_preferred_size method.

GraphicsProxyWidgetSetShortcutAutoRepeatArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::set_shortcut_auto_repeat method.

GraphicsProxyWidgetSetShortcutEnabledArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::set_shortcut_enabled method.

GraphicsProxyWidgetSetSizePolicyArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::set_size_policy method.

GraphicsProxyWidgetSetTransformArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::set_transform method.

GraphicsProxyWidgetSetTransformOriginPointArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::set_transform_origin_point method.

GraphicsProxyWidgetShapeArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::shape method.

GraphicsProxyWidgetSizeArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::size method.

GraphicsProxyWidgetSizePolicyArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::size_policy method.

GraphicsProxyWidgetStartTimerArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::start_timer method.

GraphicsProxyWidgetSubWidgetRectArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::sub_widget_rect method.

GraphicsProxyWidgetToolTipArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::tool_tip method.

GraphicsProxyWidgetTrArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::tr method.

GraphicsProxyWidgetTrUtf8Args
This trait represents a set of arguments accepted by GraphicsProxyWidget::tr_utf8 method.

GraphicsProxyWidgetTransformArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::transform method.

GraphicsProxyWidgetTransformOriginPointArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::transform_origin_point method.

GraphicsProxyWidgetTransformationsArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::transformations method.

GraphicsProxyWidgetUpdateArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::update method.

GraphicsProxyWidgetWindowFrameGeometryArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::window_frame_geometry method.

GraphicsProxyWidgetWindowFrameRectArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::window_frame_rect method.

GraphicsProxyWidgetWindowTitleArgs
This trait represents a set of arguments accepted by GraphicsProxyWidget::window_title method.