Module qt_widgets::graphics_rotation::overloading

Traits

GraphicsRotationAxisArgs
This trait represents a set of arguments accepted by GraphicsRotation::axis method.

GraphicsRotationConnectArgs
This trait represents a set of arguments accepted by GraphicsRotation::connect method.

GraphicsRotationConnectStaticArgs
This trait represents a set of arguments accepted by GraphicsRotation::connect_static method.

GraphicsRotationDestroyedArgs
This trait represents a set of arguments accepted by GraphicsRotation::destroyed method.

GraphicsRotationDisconnectArgs
This trait represents a set of arguments accepted by GraphicsRotation::disconnect method.

GraphicsRotationDisconnectStaticArgs
This trait represents a set of arguments accepted by GraphicsRotation::disconnect_static method.

GraphicsRotationDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by GraphicsRotation::dynamic_property_names method.

GraphicsRotationNewArgs
This trait represents a set of arguments accepted by GraphicsRotation::new method.

GraphicsRotationObjectNameArgs
This trait represents a set of arguments accepted by GraphicsRotation::object_name method.

GraphicsRotationOriginArgs
This trait represents a set of arguments accepted by GraphicsRotation::origin method.

GraphicsRotationPropertyArgs
This trait represents a set of arguments accepted by GraphicsRotation::property method.

GraphicsRotationSetAxisArgs
This trait represents a set of arguments accepted by GraphicsRotation::set_axis method.

GraphicsRotationStartTimerArgs
This trait represents a set of arguments accepted by GraphicsRotation::start_timer method.

GraphicsRotationTrArgs
This trait represents a set of arguments accepted by GraphicsRotation::tr method.

GraphicsRotationTrUtf8Args
This trait represents a set of arguments accepted by GraphicsRotation::tr_utf8 method.