Module qt_widgets::graphics_scale::overloading

Traits

GraphicsScaleConnectArgs
This trait represents a set of arguments accepted by GraphicsScale::connect method.

GraphicsScaleConnectStaticArgs
This trait represents a set of arguments accepted by GraphicsScale::connect_static method.

GraphicsScaleDestroyedArgs
This trait represents a set of arguments accepted by GraphicsScale::destroyed method.

GraphicsScaleDisconnectArgs
This trait represents a set of arguments accepted by GraphicsScale::disconnect method.

GraphicsScaleDisconnectStaticArgs
This trait represents a set of arguments accepted by GraphicsScale::disconnect_static method.

GraphicsScaleDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by GraphicsScale::dynamic_property_names method.

GraphicsScaleNewArgs
This trait represents a set of arguments accepted by GraphicsScale::new method.

GraphicsScaleObjectNameArgs
This trait represents a set of arguments accepted by GraphicsScale::object_name method.

GraphicsScaleOriginArgs
This trait represents a set of arguments accepted by GraphicsScale::origin method.

GraphicsScalePropertyArgs
This trait represents a set of arguments accepted by GraphicsScale::property method.

GraphicsScaleStartTimerArgs
This trait represents a set of arguments accepted by GraphicsScale::start_timer method.

GraphicsScaleTrArgs
This trait represents a set of arguments accepted by GraphicsScale::tr method.

GraphicsScaleTrUtf8Args
This trait represents a set of arguments accepted by GraphicsScale::tr_utf8 method.