Enum qt_widgets::graphics_scene::SceneLayer

pub enum SceneLayer {
    ItemLayer,
    BackgroundLayer,
    ForegroundLayer,
    AllLayers,
}

C++ type: QGraphicsScene::SceneLayer

C++ documentation:

This enum describes the rendering layers in a QGraphicsScene. When QGraphicsScene draws the scene contents, it renders each of these layers separately, in order.

Each layer represents a flag that can be OR'ed together when calling functions such as invalidate() or QGraphicsView::invalidateScene().

This enum was introduced or modified in Qt 4.3.

The SceneLayers type is a typedef for QFlags<SceneLayer>. It stores an OR combination of SceneLayer values.

See also invalidate() and QGraphicsView::invalidateScene().

Variants
ItemLayer

The item layer. QGraphicsScene renders all items are in this layer by calling the virtual function drawItems(). The item layer is drawn after the background layer, but before the foreground layer. (C++ enum variant: ItemLayer = 1)

BackgroundLayer

The background layer. QGraphicsScene renders the scene's background in this layer by calling the virtual function drawBackground(). The background layer is drawn first of all layers. (C++ enum variant: BackgroundLayer = 2)

ForegroundLayer

The foreground layer. QGraphicsScene renders the scene's foreground in this layer by calling the virtual function drawForeground(). The foreground layer is drawn last of all layers. (C++ enum variant: ForegroundLayer = 4)

AllLayers

All layers; this value represents a combination of all three layers. (C++ enum variant: AllLayers = 65535)