Struct qt_widgets::graphics_scene_drag_drop_event::GraphicsSceneDragDropEvent

pub struct GraphicsSceneDragDropEvent {
    // some fields omitted
}

C++ type: QGraphicsSceneDragDropEvent

C++ documentation:

The QGraphicsSceneDragDropEvent class provides events for drag and drop in the graphics view framework.

QGraphicsView inherits the drag and drop functionality provided by QWidget. When it receives a drag and drop event, it translates it to a QGraphicsSceneDragDropEvent.

QGraphicsSceneDragDropEvent stores events of type GraphicsSceneDragEnter, GraphicsSceneDragLeave, GraphicsSceneDragMove, or GraphicsSceneDrop.

QGraphicsSceneDragDropEvent contains the position of the mouse cursor in both item, scene, and screen coordinates; this can be retrieved with pos(), scenePos(), and screenPos().

The scene sends the event to the first QGraphicsItem under the mouse cursor that accepts drops; a graphics item is set to accept drops with setAcceptDrops().

Methods

impl GraphicsSceneDragDropEvent

fn accept(&mut self)

C++ method: void QGraphicsSceneDragDropEvent::accept()

C++ documentation:

Sets the accept flag of the event object, the equivalent of calling setAccepted(true).

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget.

See also ignore().

fn accept_proposed_action(&mut self)

C++ method: void QGraphicsSceneDragDropEvent::acceptProposedAction()

C++ documentation:

Sets the proposed action as accepted, i.e, the drop action is set to the proposed action. This is equal to:

setDropAction(proposedAction());

When using this function, one should not call accept().

See also dropAction(), setDropAction(), and proposedAction().

fn buttons(&self) -> Flags<MouseButton>

C++ method: QFlags<Qt::MouseButton> QGraphicsSceneDragDropEvent::buttons() const

C++ documentation:

Returns a Qt::MouseButtons value indicating which buttons were pressed on the mouse when this mouse event was generated.

See also Qt::MouseButtons.

fn drop_action(&self) -> DropAction

C++ method: Qt::DropAction QGraphicsSceneDragDropEvent::dropAction() const

C++ documentation:

Returns the action that was performed in this drag and drop. This should be set by the receiver of the drop and is returned by QDrag::exec().

See also setDropAction() and acceptProposedAction().

fn ignore(&mut self)

C++ method: void QGraphicsSceneDragDropEvent::ignore()

C++ documentation:

Clears the accept flag parameter of the event object, the equivalent of calling setAccepted(false).

Clearing the accept parameter indicates that the event receiver does not want the event. Unwanted events might be propagated to the parent widget.

See also accept().

fn is_accepted(&self) -> bool

C++ method: bool QGraphicsSceneDragDropEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn mime_data(&self) -> *const MimeData

C++ method: const QMimeData* QGraphicsSceneDragDropEvent::mimeData() const

C++ documentation:

This function returns the MIME data of the event.

fn modifiers(&self) -> Flags<KeyboardModifier>

C++ method: QFlags<Qt::KeyboardModifier> QGraphicsSceneDragDropEvent::modifiers() const

C++ documentation:

Returns the keyboard modifiers that were pressed when the drag and drop event was created.

See also Qt::KeyboardModifiers.

fn new<Args>(args: Args) -> Args::ReturnType where Args: GraphicsSceneDragDropEventNewArgs

C++ method: QGraphicsSceneDragDropEvent::QGraphicsSceneDragDropEvent

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::graphics_scene_drag_drop_event::GraphicsSceneDragDropEvent
2) fn new((qt_core::event::Type, cpp_utils::AsStruct)) -> ::graphics_scene_drag_drop_event::GraphicsSceneDragDropEvent
3) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::graphics_scene_drag_drop_event::GraphicsSceneDragDropEvent>
4) fn new((qt_core::event::Type, cpp_utils::AsBox)) -> cpp_utils::CppBox<::graphics_scene_drag_drop_event::GraphicsSceneDragDropEvent>
C++ method: [constructor] void QGraphicsSceneDragDropEvent::QGraphicsSceneDragDropEvent(QEvent::Type type = ?)

fn pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsSceneDragDropEventPosArgs<'largs>

C++ method: QGraphicsSceneDragDropEvent::pos

This is an overloaded function. Available variants:

Rust arguments:
1) fn pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QGraphicsSceneDragDropEvent::pos() const

C++ documentation:

Returns the mouse position of the event relative to the view that sent the event.

See also QGraphicsView, screenPos(), and scenePos().

fn possible_actions(&self) -> Flags<DropAction>

C++ method: QFlags<Qt::DropAction> QGraphicsSceneDragDropEvent::possibleActions() const

C++ documentation:

Returns the possible drop actions that the drag and drop can result in.

See also Qt::DropActions.

fn proposed_action(&self) -> DropAction

C++ method: Qt::DropAction QGraphicsSceneDragDropEvent::proposedAction() const

C++ documentation:

Returns the drop action that is proposed, i.e., preferred. The action must be one of the possible actions as defined by possibleActions().

See also Qt::DropAction and possibleActions().

fn register_event_type<Args>(args: Args) -> c_int where Args: GraphicsSceneDragDropEventRegisterEventTypeArgs

C++ method: QGraphicsSceneDragDropEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QGraphicsSceneDragDropEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn scene_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsSceneDragDropEventScenePosArgs<'largs>

C++ method: QGraphicsSceneDragDropEvent::scenePos

This is an overloaded function. Available variants:

Rust arguments:
1) fn scene_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn scene_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QGraphicsSceneDragDropEvent::scenePos() const

C++ documentation:

Returns the position of the mouse in scene coordinates.

See also pos() and screenPos().

fn screen_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsSceneDragDropEventScreenPosArgs<'largs>

C++ method: QGraphicsSceneDragDropEvent::screenPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn screen_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn screen_pos(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QGraphicsSceneDragDropEvent::screenPos() const

C++ documentation:

Returns the position of the mouse relative to the screen.

See also pos() and scenePos().

fn set_accepted(&mut self, accepted: bool)

C++ method: void QGraphicsSceneDragDropEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn set_buttons(&mut self, buttons: Flags<MouseButton>)

C++ method: void QGraphicsSceneDragDropEvent::setButtons(QFlags<Qt::MouseButton> buttons)

fn set_drop_action(&mut self, action: DropAction)

C++ method: void QGraphicsSceneDragDropEvent::setDropAction(Qt::DropAction action)

C++ documentation:

This function lets the receiver of the drop set the drop action that was performed to action, which should be one of the possible actions. Call accept() in stead of acceptProposedAction() if you use this function.

See also dropAction(), accept(), and possibleActions().

fn set_mime_data(&mut self, data: *const MimeData)

C++ method: void QGraphicsSceneDragDropEvent::setMimeData(const QMimeData* data)

fn set_modifiers(&mut self, modifiers: Flags<KeyboardModifier>)

C++ method: void QGraphicsSceneDragDropEvent::setModifiers(QFlags<Qt::KeyboardModifier> modifiers)

fn set_pos(&mut self, pos: &PointF)

C++ method: void QGraphicsSceneDragDropEvent::setPos(const QPointF& pos)

fn set_possible_actions(&mut self, actions: Flags<DropAction>)

C++ method: void QGraphicsSceneDragDropEvent::setPossibleActions(QFlags<Qt::DropAction> actions)

fn set_proposed_action(&mut self, action: DropAction)

C++ method: void QGraphicsSceneDragDropEvent::setProposedAction(Qt::DropAction action)

fn set_scene_pos(&mut self, pos: &PointF)

C++ method: void QGraphicsSceneDragDropEvent::setScenePos(const QPointF& pos)

fn set_screen_pos(&mut self, pos: &Point)

C++ method: void QGraphicsSceneDragDropEvent::setScreenPos(const QPoint& pos)

fn set_source(&mut self, source: *mut Widget)

C++ method: void QGraphicsSceneDragDropEvent::setSource(QWidget* source)

fn set_widget(&mut self, widget: *mut Widget)

C++ method: void QGraphicsSceneDragDropEvent::setWidget(QWidget* widget)

fn source(&self) -> *mut Widget

C++ method: QWidget* QGraphicsSceneDragDropEvent::source() const

C++ documentation:

This function returns the QGraphicsView that created the QGraphicsSceneDragDropEvent.

fn spontaneous(&self) -> bool

C++ method: bool QGraphicsSceneDragDropEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn type_(&self) -> Type

C++ method: QEvent::Type QGraphicsSceneDragDropEvent::type() const

C++ documentation:

Returns the event type.

fn widget(&self) -> *mut Widget

C++ method: QWidget* QGraphicsSceneDragDropEvent::widget() const

C++ documentation:

Returns the widget where the event originated, or 0 if the event originates from another application.