Struct qt_widgets::graphics_scene_help_event::GraphicsSceneHelpEvent

pub struct GraphicsSceneHelpEvent {
    // some fields omitted
}

C++ type: QGraphicsSceneHelpEvent

C++ documentation:

The QGraphicsSceneHelpEvent class provides events when a tooltip is requested.

When a QGraphicsView receives a QEvent of type QEvent::ToolTip, it creates a QGraphicsSceneHelpEvent, which is forwarded to the scene. You can set a tooltip on a QGraphicsItem with setToolTip(); by default QGraphicsScene displays the tooltip of the QGraphicsItem with the highest z-value (i.e, the top-most item) under the mouse position.

QGraphicsView does not forward events when "What's This" and status tip help is requested. If you need this, you can reimplement QGraphicsView::viewportEvent() and forward QStatusTipEvent events and QEvents of type QEvent::WhatsThis to the scene.

Methods

impl GraphicsSceneHelpEvent

fn accept(&mut self)

C++ method: void QGraphicsSceneHelpEvent::accept()

C++ documentation:

Sets the accept flag of the event object, the equivalent of calling setAccepted(true).

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget.

See also ignore().

fn ignore(&mut self)

C++ method: void QGraphicsSceneHelpEvent::ignore()

C++ documentation:

Clears the accept flag parameter of the event object, the equivalent of calling setAccepted(false).

Clearing the accept parameter indicates that the event receiver does not want the event. Unwanted events might be propagated to the parent widget.

See also accept().

fn is_accepted(&self) -> bool

C++ method: bool QGraphicsSceneHelpEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn new<Args>(args: Args) -> Args::ReturnType where Args: GraphicsSceneHelpEventNewArgs

C++ method: QGraphicsSceneHelpEvent::QGraphicsSceneHelpEvent

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::graphics_scene_help_event::GraphicsSceneHelpEvent
2) fn new((qt_core::event::Type, cpp_utils::AsStruct)) -> ::graphics_scene_help_event::GraphicsSceneHelpEvent
3) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::graphics_scene_help_event::GraphicsSceneHelpEvent>
4) fn new((qt_core::event::Type, cpp_utils::AsBox)) -> cpp_utils::CppBox<::graphics_scene_help_event::GraphicsSceneHelpEvent>
C++ method: [constructor] void QGraphicsSceneHelpEvent::QGraphicsSceneHelpEvent(QEvent::Type type = ?)

fn register_event_type<Args>(args: Args) -> c_int where Args: GraphicsSceneHelpEventRegisterEventTypeArgs

C++ method: QGraphicsSceneHelpEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QGraphicsSceneHelpEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn scene_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsSceneHelpEventScenePosArgs<'largs>

C++ method: QGraphicsSceneHelpEvent::scenePos

This is an overloaded function. Available variants:

Rust arguments:
1) fn scene_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn scene_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QGraphicsSceneHelpEvent::scenePos() const

C++ documentation:

Returns the position of the mouse cursor in scene coordinates at the moment the help event was sent.

See also screenPos().

fn screen_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsSceneHelpEventScreenPosArgs<'largs>

C++ method: QGraphicsSceneHelpEvent::screenPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn screen_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn screen_pos(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QGraphicsSceneHelpEvent::screenPos() const

C++ documentation:

Returns the position of the mouse cursor in screen coordinates at the moment the help event was sent.

See also scenePos().

fn set_accepted(&mut self, accepted: bool)

C++ method: void QGraphicsSceneHelpEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn set_scene_pos(&mut self, pos: &PointF)

C++ method: void QGraphicsSceneHelpEvent::setScenePos(const QPointF& pos)

fn set_screen_pos(&mut self, pos: &Point)

C++ method: void QGraphicsSceneHelpEvent::setScreenPos(const QPoint& pos)

fn set_widget(&mut self, widget: *mut Widget)

C++ method: void QGraphicsSceneHelpEvent::setWidget(QWidget* widget)

fn spontaneous(&self) -> bool

C++ method: bool QGraphicsSceneHelpEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn type_(&self) -> Type

C++ method: QEvent::Type QGraphicsSceneHelpEvent::type() const

C++ documentation:

Returns the event type.

fn widget(&self) -> *mut Widget

C++ method: QWidget* QGraphicsSceneHelpEvent::widget() const

C++ documentation:

Returns the widget where the event originated, or 0 if the event originates from another application.