Module qt_widgets::graphics_scene_hover_event::overloading

Traits

GraphicsSceneHoverEventLastPosArgs
This trait represents a set of arguments accepted by GraphicsSceneHoverEvent::last_pos method.

GraphicsSceneHoverEventLastScenePosArgs
This trait represents a set of arguments accepted by GraphicsSceneHoverEvent::last_scene_pos method.

GraphicsSceneHoverEventLastScreenPosArgs
This trait represents a set of arguments accepted by GraphicsSceneHoverEvent::last_screen_pos method.

GraphicsSceneHoverEventNewArgs
This trait represents a set of arguments accepted by GraphicsSceneHoverEvent::new method.

GraphicsSceneHoverEventPosArgs
This trait represents a set of arguments accepted by GraphicsSceneHoverEvent::pos method.

GraphicsSceneHoverEventRegisterEventTypeArgs
This trait represents a set of arguments accepted by GraphicsSceneHoverEvent::register_event_type method.

GraphicsSceneHoverEventScenePosArgs
This trait represents a set of arguments accepted by GraphicsSceneHoverEvent::scene_pos method.

GraphicsSceneHoverEventScreenPosArgs
This trait represents a set of arguments accepted by GraphicsSceneHoverEvent::screen_pos method.