Struct qt_widgets::graphics_scene_mouse_event::GraphicsSceneMouseEvent

pub struct GraphicsSceneMouseEvent {
    // some fields omitted
}

C++ type: QGraphicsSceneMouseEvent

C++ documentation:

The QGraphicsSceneMouseEvent class provides mouse events in the graphics view framework.

When a QGraphicsView receives a QMouseEvent, it translates it to a QGraphicsSceneMouseEvent. The event is then forwarded to the QGraphicsScene associated with the view. If the event is not handled by the scene, the view may use it, e.g., for the DragMode.

In addition to containing the item, scene, and screen coordinates of the event (as pos(), scenePos(), and screenPos()), mouse events also contain the coordinates of the previous mouse event received by the view. These can be retrieved with lastPos(), lastScreenPos(), and lastScenePos().

Methods

impl GraphicsSceneMouseEvent

fn accept(&mut self)

C++ method: void QGraphicsSceneMouseEvent::accept()

C++ documentation:

Sets the accept flag of the event object, the equivalent of calling setAccepted(true).

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget.

See also ignore().

fn button(&self) -> MouseButton

C++ method: Qt::MouseButton QGraphicsSceneMouseEvent::button() const

C++ documentation:

Returns the mouse button (if any) that caused the event.

See also buttons() and modifiers().

fn button_down_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsSceneMouseEventButtonDownPosArgs<'largs>

C++ method: QGraphicsSceneMouseEvent::buttonDownPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn button_down_pos(&self, (qt_core::qt::MouseButton, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn button_down_pos(&self, (qt_core::qt::MouseButton, cpp_utils::AsStruct)) -> qt_core::point_f::PointF
C++ method: QPointF QGraphicsSceneMouseEvent::buttonDownPos(Qt::MouseButton button) const

C++ documentation:

Returns the mouse cursor position in item coordinates where the specified button was clicked.

See also buttonDownScenePos(), buttonDownScreenPos(), and pos().

fn button_down_scene_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsSceneMouseEventButtonDownScenePosArgs<'largs>

C++ method: QGraphicsSceneMouseEvent::buttonDownScenePos

This is an overloaded function. Available variants:

Rust arguments:
1) fn button_down_scene_pos(&self, (qt_core::qt::MouseButton, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn button_down_scene_pos(&self, (qt_core::qt::MouseButton, cpp_utils::AsStruct)) -> qt_core::point_f::PointF
C++ method: QPointF QGraphicsSceneMouseEvent::buttonDownScenePos(Qt::MouseButton button) const

C++ documentation:

Returns the mouse cursor position in scene coordinates where the specified button was clicked.

See also buttonDownPos(), buttonDownScreenPos(), and scenePos().

fn button_down_screen_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsSceneMouseEventButtonDownScreenPosArgs<'largs>

C++ method: QGraphicsSceneMouseEvent::buttonDownScreenPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn button_down_screen_pos(&self, (qt_core::qt::MouseButton, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn button_down_screen_pos(&self, (qt_core::qt::MouseButton, cpp_utils::AsStruct)) -> qt_core::point::Point
C++ method: QPoint QGraphicsSceneMouseEvent::buttonDownScreenPos(Qt::MouseButton button) const

C++ documentation:

Returns the mouse cursor position in screen coordinates where the specified button was clicked.

See also screenPos(), buttonDownPos(), and buttonDownScenePos().

fn buttons(&self) -> Flags<MouseButton>

C++ method: QFlags<Qt::MouseButton> QGraphicsSceneMouseEvent::buttons() const

C++ documentation:

Returns the combination of mouse buttons that were pressed at the time the event was sent.

See also button() and modifiers().

fn flags(&self) -> Flags<MouseEventFlag>

C++ method: QFlags<Qt::MouseEventFlag> QGraphicsSceneMouseEvent::flags() const

C++ documentation:

Returns the mouse event flags.

The mouse event flags provide additional information about a mouse event.

This function was introduced in Qt 5.4.

See also Qt::MouseEventFlag and QMouseEvent::flags().

fn ignore(&mut self)

C++ method: void QGraphicsSceneMouseEvent::ignore()

C++ documentation:

Clears the accept flag parameter of the event object, the equivalent of calling setAccepted(false).

Clearing the accept parameter indicates that the event receiver does not want the event. Unwanted events might be propagated to the parent widget.

See also accept().

fn is_accepted(&self) -> bool

C++ method: bool QGraphicsSceneMouseEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn last_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsSceneMouseEventLastPosArgs<'largs>

C++ method: QGraphicsSceneMouseEvent::lastPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn last_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QGraphicsSceneMouseEvent::lastPos() const

C++ documentation:

Returns the last recorded mouse cursor position in item coordinates.

See also lastScenePos(), lastScreenPos(), and pos().

fn last_scene_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsSceneMouseEventLastScenePosArgs<'largs>

C++ method: QGraphicsSceneMouseEvent::lastScenePos

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_scene_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn last_scene_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QGraphicsSceneMouseEvent::lastScenePos() const

C++ documentation:

Returns the last recorded mouse cursor position in scene coordinates. The last recorded position is the position of the previous mouse event received by the view that created the event.

See also lastPos(), lastScreenPos(), and scenePos().

fn last_screen_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsSceneMouseEventLastScreenPosArgs<'largs>

C++ method: QGraphicsSceneMouseEvent::lastScreenPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_screen_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn last_screen_pos(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QGraphicsSceneMouseEvent::lastScreenPos() const

C++ documentation:

Returns the last recorded mouse cursor position in screen coordinates. The last recorded position is the position of the previous mouse event received by the view that created the event.

See also lastPos(), lastScenePos(), and screenPos().

fn modifiers(&self) -> Flags<KeyboardModifier>

C++ method: QFlags<Qt::KeyboardModifier> QGraphicsSceneMouseEvent::modifiers() const

C++ documentation:

Returns the keyboard modifiers in use at the time the event was sent.

See also buttons() and button().

fn new<Args>(args: Args) -> Args::ReturnType where Args: GraphicsSceneMouseEventNewArgs

C++ method: QGraphicsSceneMouseEvent::QGraphicsSceneMouseEvent

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::graphics_scene_mouse_event::GraphicsSceneMouseEvent
2) fn new((qt_core::event::Type, cpp_utils::AsStruct)) -> ::graphics_scene_mouse_event::GraphicsSceneMouseEvent
3) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::graphics_scene_mouse_event::GraphicsSceneMouseEvent>
4) fn new((qt_core::event::Type, cpp_utils::AsBox)) -> cpp_utils::CppBox<::graphics_scene_mouse_event::GraphicsSceneMouseEvent>
C++ method: [constructor] void QGraphicsSceneMouseEvent::QGraphicsSceneMouseEvent(QEvent::Type type = ?)

fn pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsSceneMouseEventPosArgs<'largs>

C++ method: QGraphicsSceneMouseEvent::pos

This is an overloaded function. Available variants:

Rust arguments:
1) fn pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QGraphicsSceneMouseEvent::pos() const

C++ documentation:

Returns the mouse cursor position in item coordinates.

See also scenePos(), screenPos(), and lastPos().

fn register_event_type<Args>(args: Args) -> c_int where Args: GraphicsSceneMouseEventRegisterEventTypeArgs

C++ method: QGraphicsSceneMouseEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QGraphicsSceneMouseEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn scene_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsSceneMouseEventScenePosArgs<'largs>

C++ method: QGraphicsSceneMouseEvent::scenePos

This is an overloaded function. Available variants:

Rust arguments:
1) fn scene_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn scene_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QGraphicsSceneMouseEvent::scenePos() const

C++ documentation:

Returns the mouse cursor position in scene coordinates.

See also pos(), screenPos(), and lastScenePos().

fn screen_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsSceneMouseEventScreenPosArgs<'largs>

C++ method: QGraphicsSceneMouseEvent::screenPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn screen_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn screen_pos(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QGraphicsSceneMouseEvent::screenPos() const

C++ documentation:

Returns the mouse cursor position in screen coordinates.

See also pos(), scenePos(), and lastScreenPos().

fn set_accepted(&mut self, accepted: bool)

C++ method: void QGraphicsSceneMouseEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn set_button(&mut self, button: MouseButton)

C++ method: void QGraphicsSceneMouseEvent::setButton(Qt::MouseButton button)

fn set_button_down_pos(&mut self, button: MouseButton, pos: &PointF)

C++ method: void QGraphicsSceneMouseEvent::setButtonDownPos(Qt::MouseButton button, const QPointF& pos)

fn set_button_down_scene_pos(&mut self, button: MouseButton, pos: &PointF)

C++ method: void QGraphicsSceneMouseEvent::setButtonDownScenePos(Qt::MouseButton button, const QPointF& pos)

fn set_button_down_screen_pos(&mut self, button: MouseButton, pos: &Point)

C++ method: void QGraphicsSceneMouseEvent::setButtonDownScreenPos(Qt::MouseButton button, const QPoint& pos)

fn set_buttons(&mut self, buttons: Flags<MouseButton>)

C++ method: void QGraphicsSceneMouseEvent::setButtons(QFlags<Qt::MouseButton> buttons)

fn set_flags(&mut self, arg1: Flags<MouseEventFlag>)

C++ method: void QGraphicsSceneMouseEvent::setFlags(QFlags<Qt::MouseEventFlag> arg1)

fn set_last_pos(&mut self, pos: &PointF)

C++ method: void QGraphicsSceneMouseEvent::setLastPos(const QPointF& pos)

fn set_last_scene_pos(&mut self, pos: &PointF)

C++ method: void QGraphicsSceneMouseEvent::setLastScenePos(const QPointF& pos)

fn set_last_screen_pos(&mut self, pos: &Point)

C++ method: void QGraphicsSceneMouseEvent::setLastScreenPos(const QPoint& pos)

fn set_modifiers(&mut self, modifiers: Flags<KeyboardModifier>)

C++ method: void QGraphicsSceneMouseEvent::setModifiers(QFlags<Qt::KeyboardModifier> modifiers)

fn set_pos(&mut self, pos: &PointF)

C++ method: void QGraphicsSceneMouseEvent::setPos(const QPointF& pos)

fn set_scene_pos(&mut self, pos: &PointF)

C++ method: void QGraphicsSceneMouseEvent::setScenePos(const QPointF& pos)

fn set_screen_pos(&mut self, pos: &Point)

C++ method: void QGraphicsSceneMouseEvent::setScreenPos(const QPoint& pos)

fn set_source(&mut self, source: MouseEventSource)

C++ method: void QGraphicsSceneMouseEvent::setSource(Qt::MouseEventSource source)

fn set_widget(&mut self, widget: *mut Widget)

C++ method: void QGraphicsSceneMouseEvent::setWidget(QWidget* widget)

fn source(&self) -> MouseEventSource

C++ method: Qt::MouseEventSource QGraphicsSceneMouseEvent::source() const

C++ documentation:

Returns information about the mouse event source.

The mouse event source can be used to distinguish between genuine and artificial mouse events. The latter are events that are synthesized from touch events by the operating system or Qt itself.

This function was introduced in Qt 5.4.

See also Qt::MouseEventSource and QMouseEvent::source().

fn spontaneous(&self) -> bool

C++ method: bool QGraphicsSceneMouseEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn type_(&self) -> Type

C++ method: QEvent::Type QGraphicsSceneMouseEvent::type() const

C++ documentation:

Returns the event type.

fn widget(&self) -> *mut Widget

C++ method: QWidget* QGraphicsSceneMouseEvent::widget() const

C++ documentation:

Returns the widget where the event originated, or 0 if the event originates from another application.