Module qt_widgets::graphics_scene_resize_event::overloading

Traits

GraphicsSceneResizeEventNewArgs
This trait represents a set of arguments accepted by GraphicsSceneResizeEvent::new method.

GraphicsSceneResizeEventNewSizeArgs
This trait represents a set of arguments accepted by GraphicsSceneResizeEvent::new_size method.

GraphicsSceneResizeEventOldSizeArgs
This trait represents a set of arguments accepted by GraphicsSceneResizeEvent::old_size method.

GraphicsSceneResizeEventRegisterEventTypeArgs
This trait represents a set of arguments accepted by GraphicsSceneResizeEvent::register_event_type method.