Struct qt_widgets::graphics_scene_resize_event::GraphicsSceneResizeEvent

pub struct GraphicsSceneResizeEvent {
    // some fields omitted
}

C++ type: QGraphicsSceneResizeEvent

C++ documentation:

The QGraphicsSceneResizeEvent class provides events for widget resizing in the graphics view framework.

A QGraphicsWidget sends itself a QGraphicsSceneResizeEvent immediately when its geometry changes.

It's similar to QResizeEvent, but its sizes, oldSize() and newSize(), use QSizeF instead of QSize.

Methods

impl GraphicsSceneResizeEvent

fn accept(&mut self)

C++ method: void QGraphicsSceneResizeEvent::accept()

C++ documentation:

Sets the accept flag of the event object, the equivalent of calling setAccepted(true).

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget.

See also ignore().

fn ignore(&mut self)

C++ method: void QGraphicsSceneResizeEvent::ignore()

C++ documentation:

Clears the accept flag parameter of the event object, the equivalent of calling setAccepted(false).

Clearing the accept parameter indicates that the event receiver does not want the event. Unwanted events might be propagated to the parent widget.

See also accept().

fn is_accepted(&self) -> bool

C++ method: bool QGraphicsSceneResizeEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn new<Args>(args: Args) -> Args::ReturnType where Args: GraphicsSceneResizeEventNewArgs

C++ method: QGraphicsSceneResizeEvent::QGraphicsSceneResizeEvent

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::graphics_scene_resize_event::GraphicsSceneResizeEvent
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::graphics_scene_resize_event::GraphicsSceneResizeEvent>
C++ method: [constructor] void QGraphicsSceneResizeEvent::QGraphicsSceneResizeEvent()

fn new_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsSceneResizeEventNewSizeArgs<'largs>

C++ method: QGraphicsSceneResizeEvent::newSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn new_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn new_size(&self, cpp_utils::AsStruct) -> qt_core::size_f::SizeF
C++ method: QSizeF QGraphicsSceneResizeEvent::newSize() const

C++ documentation:

Returns the new size (i.e., the current size).

See also oldSize() and QGraphicsWidget::resize().

fn old_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GraphicsSceneResizeEventOldSizeArgs<'largs>

C++ method: QGraphicsSceneResizeEvent::oldSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn old_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn old_size(&self, cpp_utils::AsStruct) -> qt_core::size_f::SizeF
C++ method: QSizeF QGraphicsSceneResizeEvent::oldSize() const

C++ documentation:

Returns the old size (i.e., the size immediately before the widget was resized).

See also newSize() and QGraphicsWidget::resize().

fn register_event_type<Args>(args: Args) -> c_int where Args: GraphicsSceneResizeEventRegisterEventTypeArgs

C++ method: QGraphicsSceneResizeEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QGraphicsSceneResizeEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn set_accepted(&mut self, accepted: bool)

C++ method: void QGraphicsSceneResizeEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn set_new_size(&mut self, size: &SizeF)

C++ method: void QGraphicsSceneResizeEvent::setNewSize(const QSizeF& size)

fn set_old_size(&mut self, size: &SizeF)

C++ method: void QGraphicsSceneResizeEvent::setOldSize(const QSizeF& size)

fn set_widget(&mut self, widget: *mut Widget)

C++ method: void QGraphicsSceneResizeEvent::setWidget(QWidget* widget)

fn spontaneous(&self) -> bool

C++ method: bool QGraphicsSceneResizeEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn type_(&self) -> Type

C++ method: QEvent::Type QGraphicsSceneResizeEvent::type() const

C++ documentation:

Returns the event type.

fn widget(&self) -> *mut Widget

C++ method: QWidget* QGraphicsSceneResizeEvent::widget() const

C++ documentation:

Returns the widget where the event originated, or 0 if the event originates from another application.