Enum qt_widgets::graphics_view::OptimizationFlag

pub enum OptimizationFlag {
    DontClipPainter,
    DontSavePainterState,
    DontAdjustForAntialiasing,
    IndirectPainting,
}

C++ type: QGraphicsView::OptimizationFlag

C++ documentation:

This enum describes flags that you can enable to improve rendering performance in QGraphicsView. By default, none of these flags are set. Note that setting a flag usually imposes a side effect, and this effect can vary between paint devices and platforms.

This enum was introduced or modified in Qt 4.3.

The OptimizationFlags type is a typedef for QFlags<OptimizationFlag>. It stores an OR combination of OptimizationFlag values.

Variants
DontClipPainter

This value is obsolete and has no effect. (C++ enum variant: DontClipPainter = 1)

DontSavePainterState

When rendering, QGraphicsView protects the painter state (see QPainter::save()) when rendering the background or foreground, and when rendering each item. This allows you to leave the painter in an altered state (i.e., you can call QPainter::setPen() or QPainter::setBrush() without restoring the state after painting). However, if the items consistently do restore the state, you should enable this flag to prevent QGraphicsView from doing the same. (C++ enum variant: DontSavePainterState = 2)

DontAdjustForAntialiasing

Disables QGraphicsView's antialiasing auto-adjustment of exposed areas. Items that render antialiased lines on the boundaries of their QGraphicsItem::boundingRect() can end up rendering parts of the line outside. To prevent rendering artifacts, QGraphicsView expands all exposed regions by 2 pixels in all directions. If you enable this flag, QGraphicsView will no longer perform these adjustments, minimizing the areas that require redrawing, which improves performance. A common side effect is that items that do draw with antialiasing can leave painting traces behind on the scene as they are moved. (C++ enum variant: DontAdjustForAntialiasing = 4)

IndirectPainting

Since Qt 4.6, restore the old painting algorithm that calls QGraphicsView::drawItems() and QGraphicsScene::drawItems(). To be used only for compatibility with old code. (C++ enum variant: IndirectPainting = 8)