Enum qt_widgets::graphics_view::ViewportUpdateMode

pub enum ViewportUpdateMode {
    Full,
    Minimal,
    Smart,
    No,
    BoundingRect,
}

C++ type: QGraphicsView::ViewportUpdateMode

C++ documentation:

This enum describes how QGraphicsView updates its viewport when the scene contents change or are exposed.

This enum was introduced or modified in Qt 4.3.

See also viewportUpdateMode.

Variants
Full

When any visible part of the scene changes or is reexposed, QGraphicsView will update the entire viewport. This approach is fastest when QGraphicsView spends more time figuring out what to draw than it would spend drawing (e.g., when very many small items are repeatedly updated). This is the preferred update mode for viewports that do not support partial updates, such as QGLWidget, and for viewports that need to disable scroll optimization. (C++ enum variant: FullViewportUpdate = 0)

Minimal

QGraphicsView will determine the minimal viewport region that requires a redraw, minimizing the time spent drawing by avoiding a redraw of areas that have not changed. This is QGraphicsView's default mode. Although this approach provides the best performance in general, if there are many small visible changes on the scene, QGraphicsView might end up spending more time finding the minimal approach than it will spend drawing. (C++ enum variant: MinimalViewportUpdate = 1)

Smart

QGraphicsView will attempt to find an optimal update mode by analyzing the areas that require a redraw. (C++ enum variant: SmartViewportUpdate = 2)

No

QGraphicsView will never update its viewport when the scene changes; the user is expected to control all updates. This mode disables all (potentially slow) item visibility testing in QGraphicsView, and is suitable for scenes that either require a fixed frame rate, or where the viewport is otherwise updated externally. (C++ enum variant: NoViewportUpdate = 3)

BoundingRect

The bounding rectangle of all changes in the viewport will be redrawn. This mode has the advantage that QGraphicsView searches only one region for changes, minimizing time spent determining what needs redrawing. The disadvantage is that areas that have not changed also need to be redrawn. (C++ enum variant: BoundingRectViewportUpdate = 4)