Module qt_widgets::graphics_widget::overloading

Traits

GraphicsWidgetActionsArgs
This trait represents a set of arguments accepted by GraphicsWidget::actions method.

GraphicsWidgetBoundingRectArgs
This trait represents a set of arguments accepted by GraphicsWidget::bounding_rect method.

GraphicsWidgetBoundingRegionArgs
This trait represents a set of arguments accepted by GraphicsWidget::bounding_region method.

GraphicsWidgetChildItemsArgs
This trait represents a set of arguments accepted by GraphicsWidget::child_items method.

GraphicsWidgetChildrenBoundingRectArgs
This trait represents a set of arguments accepted by GraphicsWidget::children_bounding_rect method.

GraphicsWidgetClipPathArgs
This trait represents a set of arguments accepted by GraphicsWidget::clip_path method.

GraphicsWidgetCollidesWithItemArgs
This trait represents a set of arguments accepted by GraphicsWidget::collides_with_item method.

GraphicsWidgetCollidesWithPathArgs
This trait represents a set of arguments accepted by GraphicsWidget::collides_with_path method.

GraphicsWidgetCollidingItemsArgs
This trait represents a set of arguments accepted by GraphicsWidget::colliding_items method.

GraphicsWidgetConnectArgs
This trait represents a set of arguments accepted by GraphicsWidget::connect method.

GraphicsWidgetConnectStaticArgs
This trait represents a set of arguments accepted by GraphicsWidget::connect_static method.

GraphicsWidgetContentsRectArgs
This trait represents a set of arguments accepted by GraphicsWidget::contents_rect method.

GraphicsWidgetCursorArgs
This trait represents a set of arguments accepted by GraphicsWidget::cursor method.

GraphicsWidgetDataArgs
This trait represents a set of arguments accepted by GraphicsWidget::data method.

GraphicsWidgetDestroyedArgs
This trait represents a set of arguments accepted by GraphicsWidget::destroyed method.

GraphicsWidgetDeviceTransformArgs
This trait represents a set of arguments accepted by GraphicsWidget::device_transform method.

GraphicsWidgetDisconnectArgs
This trait represents a set of arguments accepted by GraphicsWidget::disconnect method.

GraphicsWidgetDisconnectStaticArgs
This trait represents a set of arguments accepted by GraphicsWidget::disconnect_static method.

GraphicsWidgetDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by GraphicsWidget::dynamic_property_names method.

GraphicsWidgetEffectiveSizeHintArgs
This trait represents a set of arguments accepted by GraphicsWidget::effective_size_hint method.

GraphicsWidgetEnsureVisibleArgs
This trait represents a set of arguments accepted by GraphicsWidget::ensure_visible method.

GraphicsWidgetFontArgs
This trait represents a set of arguments accepted by GraphicsWidget::font method.

GraphicsWidgetGeometryArgs
This trait represents a set of arguments accepted by GraphicsWidget::geometry method.

GraphicsWidgetGrabGestureArgs
This trait represents a set of arguments accepted by GraphicsWidget::grab_gesture method.

GraphicsWidgetGrabShortcutArgs
This trait represents a set of arguments accepted by GraphicsWidget::grab_shortcut method.

GraphicsWidgetIsBlockedByModalPanelArgs
This trait represents a set of arguments accepted by GraphicsWidget::is_blocked_by_modal_panel method.

GraphicsWidgetIsObscuredArgs
This trait represents a set of arguments accepted by GraphicsWidget::is_obscured method.

GraphicsWidgetItemTransformArgs
This trait represents a set of arguments accepted by GraphicsWidget::item_transform method.

GraphicsWidgetMapFromItemArgs
This trait represents a set of arguments accepted by GraphicsWidget::map_from_item method.

GraphicsWidgetMapFromParentArgs
This trait represents a set of arguments accepted by GraphicsWidget::map_from_parent method.

GraphicsWidgetMapFromSceneArgs
This trait represents a set of arguments accepted by GraphicsWidget::map_from_scene method.

GraphicsWidgetMapRectFromItemArgs
This trait represents a set of arguments accepted by GraphicsWidget::map_rect_from_item method.

GraphicsWidgetMapRectFromParentArgs
This trait represents a set of arguments accepted by GraphicsWidget::map_rect_from_parent method.

GraphicsWidgetMapRectFromSceneArgs
This trait represents a set of arguments accepted by GraphicsWidget::map_rect_from_scene method.

GraphicsWidgetMapRectToItemArgs
This trait represents a set of arguments accepted by GraphicsWidget::map_rect_to_item method.

GraphicsWidgetMapRectToParentArgs
This trait represents a set of arguments accepted by GraphicsWidget::map_rect_to_parent method.

GraphicsWidgetMapRectToSceneArgs
This trait represents a set of arguments accepted by GraphicsWidget::map_rect_to_scene method.

GraphicsWidgetMapToItemArgs
This trait represents a set of arguments accepted by GraphicsWidget::map_to_item method.

GraphicsWidgetMapToParentArgs
This trait represents a set of arguments accepted by GraphicsWidget::map_to_parent method.

GraphicsWidgetMapToSceneArgs
This trait represents a set of arguments accepted by GraphicsWidget::map_to_scene method.

GraphicsWidgetMatrixArgs
This trait represents a set of arguments accepted by GraphicsWidget::matrix method.

GraphicsWidgetMaximumSizeArgs
This trait represents a set of arguments accepted by GraphicsWidget::maximum_size method.

GraphicsWidgetMinimumSizeArgs
This trait represents a set of arguments accepted by GraphicsWidget::minimum_size method.

GraphicsWidgetNewArgs
This trait represents a set of arguments accepted by GraphicsWidget::new method.

GraphicsWidgetObjectNameArgs
This trait represents a set of arguments accepted by GraphicsWidget::object_name method.

GraphicsWidgetOpaqueAreaArgs
This trait represents a set of arguments accepted by GraphicsWidget::opaque_area method.

GraphicsWidgetPaintArgs
This trait represents a set of arguments accepted by GraphicsWidget::paint method.

GraphicsWidgetPaintWindowFrameArgs
This trait represents a set of arguments accepted by GraphicsWidget::paint_window_frame method.

GraphicsWidgetPaletteArgs
This trait represents a set of arguments accepted by GraphicsWidget::palette method.

GraphicsWidgetPosArgs
This trait represents a set of arguments accepted by GraphicsWidget::pos method.

GraphicsWidgetPreferredSizeArgs
This trait represents a set of arguments accepted by GraphicsWidget::preferred_size method.

GraphicsWidgetPropertyArgs
This trait represents a set of arguments accepted by GraphicsWidget::property method.

GraphicsWidgetRectArgs
This trait represents a set of arguments accepted by GraphicsWidget::rect method.

GraphicsWidgetResizeArgs
This trait represents a set of arguments accepted by GraphicsWidget::resize method.

GraphicsWidgetSceneBoundingRectArgs
This trait represents a set of arguments accepted by GraphicsWidget::scene_bounding_rect method.

GraphicsWidgetSceneMatrixArgs
This trait represents a set of arguments accepted by GraphicsWidget::scene_matrix method.

GraphicsWidgetScenePosArgs
This trait represents a set of arguments accepted by GraphicsWidget::scene_pos method.

GraphicsWidgetSceneTransformArgs
This trait represents a set of arguments accepted by GraphicsWidget::scene_transform method.

GraphicsWidgetScrollArgs
This trait represents a set of arguments accepted by GraphicsWidget::scroll method.

GraphicsWidgetSetAttributeArgs
This trait represents a set of arguments accepted by GraphicsWidget::set_attribute method.

GraphicsWidgetSetCacheModeArgs
This trait represents a set of arguments accepted by GraphicsWidget::set_cache_mode method.

GraphicsWidgetSetFlagArgs
This trait represents a set of arguments accepted by GraphicsWidget::set_flag method.

GraphicsWidgetSetFocusArgs
This trait represents a set of arguments accepted by GraphicsWidget::set_focus method.

GraphicsWidgetSetGeometryArgs
This trait represents a set of arguments accepted by GraphicsWidget::set_geometry method.

GraphicsWidgetSetMatrixArgs
This trait represents a set of arguments accepted by GraphicsWidget::set_matrix method.

GraphicsWidgetSetMaximumSizeArgs
This trait represents a set of arguments accepted by GraphicsWidget::set_maximum_size method.

GraphicsWidgetSetMinimumSizeArgs
This trait represents a set of arguments accepted by GraphicsWidget::set_minimum_size method.

GraphicsWidgetSetPosArgs
This trait represents a set of arguments accepted by GraphicsWidget::set_pos method.

GraphicsWidgetSetPreferredSizeArgs
This trait represents a set of arguments accepted by GraphicsWidget::set_preferred_size method.

GraphicsWidgetSetShortcutAutoRepeatArgs
This trait represents a set of arguments accepted by GraphicsWidget::set_shortcut_auto_repeat method.

GraphicsWidgetSetShortcutEnabledArgs
This trait represents a set of arguments accepted by GraphicsWidget::set_shortcut_enabled method.

GraphicsWidgetSetSizePolicyArgs
This trait represents a set of arguments accepted by GraphicsWidget::set_size_policy method.

GraphicsWidgetSetTransformArgs
This trait represents a set of arguments accepted by GraphicsWidget::set_transform method.

GraphicsWidgetSetTransformOriginPointArgs
This trait represents a set of arguments accepted by GraphicsWidget::set_transform_origin_point method.

GraphicsWidgetShapeArgs
This trait represents a set of arguments accepted by GraphicsWidget::shape method.

GraphicsWidgetSizeArgs
This trait represents a set of arguments accepted by GraphicsWidget::size method.

GraphicsWidgetSizePolicyArgs
This trait represents a set of arguments accepted by GraphicsWidget::size_policy method.

GraphicsWidgetStartTimerArgs
This trait represents a set of arguments accepted by GraphicsWidget::start_timer method.

GraphicsWidgetToolTipArgs
This trait represents a set of arguments accepted by GraphicsWidget::tool_tip method.

GraphicsWidgetTrArgs
This trait represents a set of arguments accepted by GraphicsWidget::tr method.

GraphicsWidgetTrUtf8Args
This trait represents a set of arguments accepted by GraphicsWidget::tr_utf8 method.

GraphicsWidgetTransformArgs
This trait represents a set of arguments accepted by GraphicsWidget::transform method.

GraphicsWidgetTransformOriginPointArgs
This trait represents a set of arguments accepted by GraphicsWidget::transform_origin_point method.

GraphicsWidgetTransformationsArgs
This trait represents a set of arguments accepted by GraphicsWidget::transformations method.

GraphicsWidgetUpdateArgs
This trait represents a set of arguments accepted by GraphicsWidget::update method.

GraphicsWidgetWindowFrameGeometryArgs
This trait represents a set of arguments accepted by GraphicsWidget::window_frame_geometry method.

GraphicsWidgetWindowFrameRectArgs
This trait represents a set of arguments accepted by GraphicsWidget::window_frame_rect method.

GraphicsWidgetWindowTitleArgs
This trait represents a set of arguments accepted by GraphicsWidget::window_title method.