Struct qt_widgets::hash::HashCIntCUint

pub struct HashCIntCUint {
    // some fields omitted
}

C++ type: QHash<int, unsigned int>

C++ documentation:

The QHash class is a template class that provides a hash-table-based dictionary.

QHash<Key, T> is one of Qt's generic container classes. It stores (key, value) pairs and provides very fast lookup of the value associated with a key.

QHash provides very similar functionality to QMap. The differences are:

QHash provides faster lookups than QMap. (See Algorithmic Complexity for details.)

When iterating over a QMap, the items are always sorted by key. With QHash, the items are arbitrarily ordered.

The key type of a QMap must provide operator<(). The key type of a QHash must provide operator==() and a global hash function called qHash() (see qHash).

Here's an example QHash with QString keys and int values:

QHash<QString, int> hash;

To insert a (key, value) pair into the hash, you can use operator:

hash["one"] = 1;
hash["three"] = 3;
hash["seven"] = 7;

This inserts the following three (key, value) pairs into the QHash: ("one", 1), ("three", 3), and ("seven", 7). Another way to insert items into the hash is to use insert():

hash.insert("twelve", 12);

To look up a value, use operator or value():

int num1 = hash["thirteen"];
int num2 = hash.value("thirteen");

If there is no item with the specified key in the hash, these functions return a default-constructed value.

If you want to check whether the hash contains a particular key, use contains():

int timeout = 30;
if (hash.contains("TIMEOUT"))
    timeout = hash.value("TIMEOUT");

There is also a value() overload that uses its second argument as a default value if there is no item with the specified key:

int timeout = hash.value("TIMEOUT", 30);

In general, we recommend that you use contains() and value() rather than operator for looking up a key in a hash. The reason is that operator silently inserts an item into the hash if no item exists with the same key (unless the hash is const). For example, the following code snippet will create 1000 items in memory:

// WRONG
QHash<int, QWidget *> hash;
...
for (int i = 0; i < 1000; ++i) {
    if (hash[i] == okButton)
        cout << "Found button at index " << i << endl;
}

To avoid this problem, replace hash[i] with hash.value(i) in the code above.

Internally, QHash uses a hash table to perform lookups. Unlike Qt 3's QDict class, which needed to be initialized with a prime number, QHash's hash table automatically grows and shrinks to provide fast lookups without wasting too much memory. You can still control the size of the hash table by calling reserve() if you already know approximately how many items the QHash will contain, but this isn't necessary to obtain good performance. You can also call capacity() to retrieve the hash table's size.

If you want to navigate through all the (key, value) pairs stored in a QHash, you can use an iterator. QHash provides both Java-style iterators (QHashIterator and QMutableHashIterator) and STL-style iterators (QHash::const_iterator and QHash::iterator). Here's how to iterate over a QHash<QString, int> using a Java-style iterator:

QHashIterator<QString, int> i(hash);
while (i.hasNext()) {
    i.next();
    cout << i.key() << ": " << i.value() << endl;
}

Here's the same code, but using an STL-style iterator:

QHash<QString, int>::const_iterator i = hash.constBegin();
while (i != hash.constEnd()) {
    cout << i.key() << ": " << i.value() << endl;
    ++i;
}

QHash is unordered, so an iterator's sequence cannot be assumed to be predictable. If ordering by key is required, use a QMap.

Normally, a QHash allows only one value per key. If you call insert() with a key that already exists in the QHash, the previous value is erased. For example:

hash.insert("plenty", 100);
hash.insert("plenty", 2000);
// hash.value("plenty") == 2000

However, you can store multiple values per key by using insertMulti() instead of insert() (or using the convenience subclass QMultiHash). If you want to retrieve all the values for a single key, you can use values(const Key &key), which returns a QList<T>:

QList<int> values = hash.values("plenty");
for (int i = 0; i < values.size(); ++i)
    cout << values.at(i) << endl;

The items that share the same key are available from most recently to least recently inserted. A more efficient approach is to call find() to get the iterator for the first item with a key and iterate from there:

QHash<QString, int>::iterator i = hash.find("plenty");
while (i != hash.end() && i.key() == "plenty") {
    cout << i.value() << endl;
    ++i;
}

If you only need to extract the values from a hash (not the keys), you can also use foreach:

QHash<QString, int> hash;
...
foreach (int value, hash)
    cout << value << endl;

Items can be removed from the hash in several ways. One way is to call remove(); this will remove any item with the given key. Another way is to use QMutableHashIterator::remove(). In addition, you can clear the entire hash using clear().

QHash's key and value data types must be assignable data types. You cannot, for example, store a QWidget as a value; instead, store a QWidget *.

The qHash() hashing function

A QHash's key type has additional requirements other than being an assignable data type: it must provide operator==(), and there must also be a qHash() function in the type's namespace that returns a hash value for an argument of the key's type.

The qHash() function computes a numeric value based on a key. It can use any algorithm imaginable, as long as it always returns the same value if given the same argument. In other words, if e1 == e2, then qHash(e1) == qHash(e2) must hold as well. However, to obtain good performance, the qHash() function should attempt to return different hash values for different keys to the largest extent possible.

For a key type K, the qHash function must have one of these signatures:

uint qHash(K key);
uint qHash(const K &key);

uint qHash(K key, uint seed); uint qHash(const K &key, uint seed);
The two-arguments overloads take an unsigned integer that should be used to seed the calculation of the hash function. This seed is provided by QHash in order to prevent a family of algorithmic complexity attacks. If both a one-argument and a two-arguments overload are defined for a key type, the latter is used by QHash (note that you can simply define a two-arguments version, and use a default value for the seed parameter).

Here's a partial list of the C++ and Qt types that can serve as keys in a QHash: any integer type (char, unsigned long, etc.), any pointer type, QChar, QString, and QByteArray. For all of these, the <QHash> header defines a qHash() function that computes an adequate hash value. Many other Qt classes also declare a qHash overload for their type; please refer to the documentation of each class.

If you want to use other types as the key, make sure that you provide operator==() and a qHash() implementation.

Example:

#ifndef EMPLOYEE_H
#define EMPLOYEE_H

class Employee { public: Employee() {} Employee(const QString &name, const QDate &dateOfBirth); ...

private: QString myName; QDate myDateOfBirth; };

inline bool operator==(const Employee &e1, const Employee &e2) { return e1.name() == e2.name() && e1.dateOfBirth() == e2.dateOfBirth(); }

inline uint qHash(const Employee &key, uint seed) { return qHash(key.name(), seed) key.dateOfBirth().day(); }

#endif // EMPLOYEE_H
In the example above, we've relied on Qt's global qHash(const QString &, uint) to give us a hash value for the employee's name, and XOR'ed this with the day they were born to help produce unique hashes for people with the same name.

Note that the implementation of the qHash() overloads offered by Qt may change at any time. You must not rely on the fact that qHash() will give the same results (for the same inputs) across different Qt versions.

Algorithmic complexity attacks

All hash tables are vulnerable to a particular class of denial of service attacks, in which the attacker carefully pre-computes a set of different keys that are going to be hashed in the same bucket of a hash table (or even have the very same hash value). The attack aims at getting the worst-case algorithmic behavior (O(n) instead of amortized O(1), see Algorithmic Complexity for the details) when the data is fed into the table.

In order to avoid this worst-case behavior, the calculation of the hash value done by qHash() can be salted by a random seed, that nullifies the attack's extent. This seed is automatically generated by QHash once per process, and then passed by QHash as the second argument of the two-arguments overload of the qHash() function.

This randomization of QHash is enabled by default. Even though programs should never depend on a particular QHash ordering, there may be situations where you temporarily need deterministic behavior, for example for debugging or regression testing. To disable the randomization, define the environment variable QT_HASH_SEED. The contents of that variable, interpreted as a decimal value, will be used as the seed for qHash().

Methods

impl HashCIntCUint

fn capacity(&self) -> c_int

C++ method: int QHash<int, unsigned int>::capacity() const

C++ documentation:

Returns the number of buckets in the QHash's internal hash table.

The sole purpose of this function is to provide a means of fine tuning QHash's memory usage. In general, you will rarely ever need to call this function. If you want to know how many items are in the hash, call size().

See also reserve() and squeeze().

fn clear(&mut self)

C++ method: void QHash<int, unsigned int>::clear()

C++ documentation:

Removes all items from the hash.

See also remove().

fn contains(&self, key: &c_int) -> bool

C++ method: bool QHash<int, unsigned int>::contains(const int& key) const

C++ documentation:

Returns true if the hash contains an item with the key; otherwise returns false.

See also count() and QMultiHash::contains().

fn count<'largs, Args>(&'largs self, args: Args) -> c_int where Args: HashCIntCUintCountArgs<'largs>

C++ method: QHash<int, unsigned int>::count

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn count(&self, ()) -> libc::c_int
C++ method: int QHash<int, unsigned int>::count() const

C++ documentation:

This is an overloaded function.

Same as size().

Variant 2

Rust arguments: fn count(&self, &libc::c_int) -> libc::c_int
C++ method: int QHash<int, unsigned int>::count(const int& key) const

C++ documentation:

Returns the number of items associated with the key.

See also contains() and insertMulti().

fn detach(&mut self)

C++ method: void QHash<int, unsigned int>::detach()

fn empty(&self) -> bool

C++ method: bool QHash<int, unsigned int>::empty() const

C++ documentation:

This function is provided for STL compatibility. It is equivalent to isEmpty(), returning true if the hash is empty; otherwise returns false.

fn is_detached(&self) -> bool

C++ method: bool QHash<int, unsigned int>::isDetached() const

fn is_empty(&self) -> bool

C++ method: bool QHash<int, unsigned int>::isEmpty() const

C++ documentation:

Returns true if the hash contains no items; otherwise returns false.

See also size().

fn is_shared_with(&self, other: &HashCIntCUint) -> bool

C++ method: bool QHash<int, unsigned int>::isSharedWith(const QHash<int, unsigned int>& other) const

fn key<'largs, Args>(&'largs self, args: Args) -> c_int where Args: HashCIntCUintKeyArgs<'largs>

C++ method: QHash<int, unsigned int>::key

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn key(&self, &libc::c_uint) -> libc::c_int
C++ method: const int QHash<int, unsigned int>::key(const unsigned int& value) const

C++ documentation:

Returns the first key mapped to value.

If the hash contains no item with the value, the function returns a default-constructed key.

This function can be slow (linear time), because QHash's internal data structure is optimized for fast lookup by key, not by value.

See also value() and keys().

Variant 2

Rust arguments: fn key(&self, (&libc::c_uint, &libc::c_int)) -> libc::c_int
C++ method: const int QHash<int, unsigned int>::key(const unsigned int& value, const int& defaultKey) const

C++ documentation:

This is an overloaded function.

Returns the first key mapped to value, or defaultKey if the hash contains no item mapped to value.

This function can be slow (linear time), because QHash's internal data structure is optimized for fast lookup by key, not by value.

This function was introduced in Qt 4.3.

fn keys<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: HashCIntCUintKeysArgs<'largs>

C++ method: QHash<int, unsigned int>::keys

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn keys(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListCInt>
2) fn keys(&self, cpp_utils::AsStruct) -> qt_core::list::ListCInt
C++ method: QList<int> QHash<int, unsigned int>::keys() const

C++ documentation:

Returns a list containing all the keys in the hash, in an arbitrary order. Keys that occur multiple times in the hash (because items were inserted with insertMulti(), or unite() was used) also occur multiple times in the list.

To obtain a list of unique keys, where each key from the map only occurs once, use uniqueKeys().

The order is guaranteed to be the same as that used by values().

See also uniqueKeys(), values(), and key().

Variant 2

Rust arguments:
1) fn keys(&self, (&libc::c_uint, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::list::ListCInt>
2) fn keys(&self, (&libc::c_uint, cpp_utils::AsStruct)) -> qt_core::list::ListCInt
C++ method: QList<int> QHash<int, unsigned int>::keys(const unsigned int& value) const

C++ documentation:

This is an overloaded function.

Returns a list containing all the keys associated with value value, in an arbitrary order.

This function can be slow (linear time), because QHash's internal data structure is optimized for fast lookup by key, not by value.

fn new<Args>(args: Args) -> Args::ReturnType where Args: HashCIntCUintNewArgs

C++ method: QHash<int, unsigned int>::QHash

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::hash::HashCIntCUint
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::hash::HashCIntCUint>
C++ method: [constructor] void QHash<int, unsigned int>::QHash()

Variant 2

Rust arguments:
1) fn new((&::hash::HashCIntCUint, cpp_utils::AsStruct)) -> ::hash::HashCIntCUint
2) fn new((&::hash::HashCIntCUint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::hash::HashCIntCUint>
C++ method: [constructor] void QHash<int, unsigned int>::QHash(const QHash<int, unsigned int>& other)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 HashCIntCUint) -> &'l0 mut HashCIntCUint

C++ method: QHash<int, unsigned int>& QHash<int, unsigned int>::operator=(const QHash<int, unsigned int>& other)

C++ documentation:

Assigns other to this hash and returns a reference to this hash.

fn op_eq(&self, other: &HashCIntCUint) -> bool

C++ method: bool QHash<int, unsigned int>::operator==(const QHash<int, unsigned int>& other) const

C++ documentation:

Returns true if other is equal to this hash; otherwise returns false.

Two hashes are considered equal if they contain the same (key, value) pairs.

This function requires the value type to implement operator==().

See also operator!=().

fn op_index(&self, key: &c_int) -> c_uint

C++ method: const unsigned int QHash<int, unsigned int>::operator[](const int& key) const

C++ documentation:

This is an overloaded function.

Same as value().

fn op_index_mut<'l0, 'l1>(&'l0 mut self, key: &'l1 c_int) -> &'l0 mut c_uint

C++ method: unsigned int& QHash<int, unsigned int>::operator[](const int& key)

C++ documentation:

Returns the value associated with the key as a modifiable reference.

If the hash contains no item with the key, the function inserts a default-constructed value into the hash with the key, and returns a reference to it. If the hash contains multiple items with the key, this function returns a reference to the most recently inserted value.

See also insert() and value().

fn op_neq(&self, other: &HashCIntCUint) -> bool

C++ method: bool QHash<int, unsigned int>::operator!=(const QHash<int, unsigned int>& other) const

C++ documentation:

Returns true if other is not equal to this hash; otherwise returns false.

Two hashes are considered equal if they contain the same (key, value) pairs.

This function requires the value type to implement operator==().

See also operator==().

fn remove(&mut self, key: &c_int) -> c_int

C++ method: int QHash<int, unsigned int>::remove(const int& key)

C++ documentation:

Removes all the items that have the key from the hash. Returns the number of items removed which is usually 1 but will be 0 if the key isn't in the hash, or greater than 1 if insertMulti() has been used with the key.

See also clear(), take(), and QMultiHash::remove().

fn reserve(&mut self, size: c_int)

C++ method: void QHash<int, unsigned int>::reserve(int size)

C++ documentation:

Ensures that the QHash's internal hash table consists of at least size buckets.

This function is useful for code that needs to build a huge hash and wants to avoid repeated reallocation. For example:

QHash<QString, int> hash;
hash.reserve(20000);
for (int i = 0; i < 20000; ++i)
    hash.insert(keys[i], values[i]);

Ideally, size should be slightly more than the maximum number of items expected in the hash. size doesn't have to be prime, because QHash will use a prime number internally anyway. If size is an underestimate, the worst that will happen is that the QHash will be a bit slower.

In general, you will rarely ever need to call this function. QHash's internal hash table automatically shrinks or grows to provide good performance without wasting too much memory.

See also squeeze() and capacity().

fn set_sharable(&mut self, sharable: bool)

C++ method: void QHash<int, unsigned int>::setSharable(bool sharable)

fn size(&self) -> c_int

C++ method: int QHash<int, unsigned int>::size() const

C++ documentation:

Returns the number of items in the hash.

See also isEmpty() and count().

fn squeeze(&mut self)

C++ method: void QHash<int, unsigned int>::squeeze()

C++ documentation:

Reduces the size of the QHash's internal hash table to save memory.

The sole purpose of this function is to provide a means of fine tuning QHash's memory usage. In general, you will rarely ever need to call this function.

See also reserve() and capacity().

fn swap(&mut self, other: &mut HashCIntCUint)

C++ method: void QHash<int, unsigned int>::swap(QHash<int, unsigned int>& other)

C++ documentation:

Swaps hash other with this hash. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

fn take(&mut self, key: &c_int) -> c_uint

C++ method: unsigned int QHash<int, unsigned int>::take(const int& key)

C++ documentation:

Removes the item with the key from the hash and returns the value associated with it.

If the item does not exist in the hash, the function simply returns a default-constructed value. If there are multiple items for key in the hash, only the most recently inserted one is removed.

If you don't use the return value, remove() is more efficient.

See also remove().

fn unique_keys<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: HashCIntCUintUniqueKeysArgs<'largs>

C++ method: QHash<int, unsigned int>::uniqueKeys

This is an overloaded function. Available variants:

Rust arguments:
1) fn unique_keys(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListCInt>
2) fn unique_keys(&self, cpp_utils::AsStruct) -> qt_core::list::ListCInt
C++ method: QList<int> QHash<int, unsigned int>::uniqueKeys() const

C++ documentation:

Returns a list containing all the keys in the map. Keys that occur multiple times in the map (because items were inserted with insertMulti(), or unite() was used) occur only once in the returned list.

This function was introduced in Qt 4.2.

See also keys() and values().

fn unite<'l0, 'l1>(&'l0 mut self, other: &'l1 HashCIntCUint) -> &'l0 mut HashCIntCUint

C++ method: QHash<int, unsigned int>& QHash<int, unsigned int>::unite(const QHash<int, unsigned int>& other)

C++ documentation:

Inserts all the items in the other hash into this hash. If a key is common to both hashes, the resulting hash will contain the key multiple times.

See also insertMulti().

fn value<'largs, Args>(&'largs self, args: Args) -> c_uint where Args: HashCIntCUintValueArgs<'largs>

C++ method: QHash<int, unsigned int>::value

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn value(&self, &libc::c_int) -> libc::c_uint
C++ method: const unsigned int QHash<int, unsigned int>::value(const int& key) const

C++ documentation:

Returns the value associated with the key.

If the hash contains no item with the key, the function returns a default-constructed value. If there are multiple items for the key in the hash, the value of the most recently inserted one is returned.

See also key(), values(), contains(), and operator[]().

Variant 2

Rust arguments: fn value(&self, (&libc::c_int, &libc::c_uint)) -> libc::c_uint
C++ method: const unsigned int QHash<int, unsigned int>::value(const int& key, const unsigned int& defaultValue) const

C++ documentation:

This is an overloaded function.

If the hash contains no item with the given key, the function returns defaultValue.