Enum qt_widgets::header_view::ResizeMode

pub enum ResizeMode {
    Interactive,
    Stretch,
    Fixed,
    ResizeToContents,
}

C++ type: QHeaderView::ResizeMode

C++ documentation:

The resize mode specifies the behavior of the header sections. It can be set on the entire header view or on individual sections using setSectionResizeMode().

The following values are obsolete:

See also setResizeMode(), setSectionResizeMode(), stretchLastSection, and minimumSectionSize.

Variants
Interactive

The user can resize the section. The section can also be resized programmatically using resizeSection(). The section size defaults to defaultSectionSize. (See also cascadingSectionResizes.) (C++ enum variant: Interactive = 0)

Stretch

QHeaderView will automatically resize the section to fill the available space. The size cannot be changed by the user or programmatically. (C++ enum variant: Stretch = 1)

Fixed

This variant corresponds to multiple C++ enum variants with the same value:

Fixed = 2: The user cannot resize the section. The section can only be resized programmatically using resizeSection(). The section size defaults to defaultSectionSize.

Custom = 2: Use Fixed instead.

ResizeToContents

QHeaderView will automatically resize the section to its optimal size based on the contents of the entire column or row. The size cannot be changed by the user or programmatically. (This value was introduced in 4.2) (C++ enum variant: ResizeToContents = 3)