Module qt_widgets::item_delegate::overloading

Traits

ItemDelegateCloseEditorArgs
This trait represents a set of arguments accepted by ItemDelegate::close_editor method.

ItemDelegateConnectArgs
This trait represents a set of arguments accepted by ItemDelegate::connect method.

ItemDelegateConnectStaticArgs
This trait represents a set of arguments accepted by ItemDelegate::connect_static method.

ItemDelegateDestroyedArgs
This trait represents a set of arguments accepted by ItemDelegate::destroyed method.

ItemDelegateDisconnectArgs
This trait represents a set of arguments accepted by ItemDelegate::disconnect method.

ItemDelegateDisconnectStaticArgs
This trait represents a set of arguments accepted by ItemDelegate::disconnect_static method.

ItemDelegateDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by ItemDelegate::dynamic_property_names method.

ItemDelegateElidedTextArgs
This trait represents a set of arguments accepted by ItemDelegate::elided_text method.

ItemDelegateNewArgs
This trait represents a set of arguments accepted by ItemDelegate::new method.

ItemDelegateObjectNameArgs
This trait represents a set of arguments accepted by ItemDelegate::object_name method.

ItemDelegatePaintingRolesArgs
This trait represents a set of arguments accepted by ItemDelegate::painting_roles method.

ItemDelegatePropertyArgs
This trait represents a set of arguments accepted by ItemDelegate::property method.

ItemDelegateSizeHintArgs
This trait represents a set of arguments accepted by ItemDelegate::size_hint method.

ItemDelegateStartTimerArgs
This trait represents a set of arguments accepted by ItemDelegate::start_timer method.

ItemDelegateTrArgs
This trait represents a set of arguments accepted by ItemDelegate::tr method.

ItemDelegateTrUtf8Args
This trait represents a set of arguments accepted by ItemDelegate::tr_utf8 method.