Struct qt_widgets::item_editor_creator_base::ItemEditorCreatorBase

pub struct ItemEditorCreatorBase {
    // some fields omitted
}

C++ type: QItemEditorCreatorBase

C++ documentation:

The QItemEditorCreatorBase class provides an abstract base class that must be subclassed when implementing new item editor creators.

QItemEditorCreatorBase objects are specialized widget factories that provide editor widgets for one particular QVariant data type. They are used by QItemEditorFactory to create editors for QItemDelegates. Creator bases must be registered with QItemEditorFactory::registerEditor().

An editor should provide a user property for the data it edits. QItemDelagates can then access the property using Qt's meta-object system to set and retrieve the editing data. A property is set as the user property with the USER keyword:

Q_PROPERTY(QColor color READ color WRITE setColor USER true)

If the editor does not provide a user property, it must return the name of the property from valuePropertyName(); delegates will then use the name to access the property. If a user property exists, item delegates will not call valuePropertyName().

QStandardItemEditorCreator is a convenience template class that can be used to register widgets without the need to subclass QItemEditorCreatorBase.

Methods

impl ItemEditorCreatorBase

fn create_widget(&self, parent: *mut Widget) -> *mut Widget

C++ method: pure virtual QWidget* QItemEditorCreatorBase::createWidget(QWidget* parent) const

C++ documentation:

Returns an editor widget with the given parent.

When implementing this function in subclasses of this class, you must construct and return new editor widgets with the parent widget specified.

fn value_property_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ItemEditorCreatorBaseValuePropertyNameArgs<'largs>

C++ method: QItemEditorCreatorBase::valuePropertyName

This is an overloaded function. Available variants:

Rust arguments:
1) fn value_property_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::byte_array::ByteArray>
2) fn value_property_name(&self, cpp_utils::AsStruct) -> qt_core::byte_array::ByteArray
C++ method: pure virtual QByteArray QItemEditorCreatorBase::valuePropertyName() const

C++ documentation:

Returns the name of the property used to get and set values in the creator's editor widgets.

When implementing this function in subclasses, you must ensure that the editor widget's property specified by this function can accept the type the creator is registered for. For example, a creator which constructs QCheckBox widgets to edit boolean values would return the checkable property name from this function, and must be registered in the item editor factory for the QVariant::Bool type.

Note: Since Qt 4.2 the item delegates query the user property of widgets, and only call this function if the widget has no user property. You can override this behavior by reimplementing QAbstractItemDelegate::setModelData() and QAbstractItemDelegate::setEditorData().

See also QMetaObject::userProperty() and QItemEditorFactory::registerEditor().