Struct qt_widgets::item_editor_factory::ItemEditorFactory

pub struct ItemEditorFactory {
    // some fields omitted
}

C++ type: QItemEditorFactory

C++ documentation:

The QItemEditorFactory class provides widgets for editing item data in views and delegates.

When editing data in an item view, editors are created and displayed by a delegate. QItemDelegate, which is the delegate by default installed on Qt's item views, uses a QItemEditorFactory to create editors for it. A default unique instance provided by QItemEditorFactory is used by all item delegates. If you set a new default factory with setDefaultFactory(), the new factory will be used by existing and new delegates.

A factory keeps a collection of QItemEditorCreatorBase instances, which are specialized editors that produce editors for one particular QVariant data type (All Qt models store their data in QVariants).

Methods

impl ItemEditorFactory

fn create_editor(&self, user_type: c_int, parent: *mut Widget) -> *mut Widget

C++ method: virtual QWidget* QItemEditorFactory::createEditor(int userType, QWidget* parent) const

C++ documentation:

Creates an editor widget with the given parent for the specified userType of data, and returns it as a QWidget.

See also registerEditor().

fn default_factory() -> *const ItemEditorFactory

C++ method: static const QItemEditorFactory* QItemEditorFactory::defaultFactory()

C++ documentation:

Returns the default item editor factory.

See also setDefaultFactory().

fn new<Args>(args: Args) -> Args::ReturnType where Args: ItemEditorFactoryNewArgs

C++ method: QItemEditorFactory::QItemEditorFactory

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::item_editor_factory::ItemEditorFactory
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::item_editor_factory::ItemEditorFactory>
C++ method: [constructor] void QItemEditorFactory::QItemEditorFactory()

fn register_editor(&mut self, user_type: c_int, creator: *mut ItemEditorCreatorBase)

C++ method: void QItemEditorFactory::registerEditor(int userType, QItemEditorCreatorBase* creator)

C++ documentation:

Registers an item editor creator specified by creator for the given userType of data.

Note: The factory takes ownership of the item editor creator and will destroy it if a new creator for the same type is registered later.

See also createEditor().

fn set_default_factory(factory: *mut ItemEditorFactory)

C++ method: static void QItemEditorFactory::setDefaultFactory(QItemEditorFactory* factory)

C++ documentation:

Sets the default item editor factory to the given factory. Both new and existing delegates will use the new factory.

See also defaultFactory().

fn value_property_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ItemEditorFactoryValuePropertyNameArgs<'largs>

C++ method: QItemEditorFactory::valuePropertyName

This is an overloaded function. Available variants:

Rust arguments:
1) fn value_property_name(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::byte_array::ByteArray>
2) fn value_property_name(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::byte_array::ByteArray
C++ method: virtual QByteArray QItemEditorFactory::valuePropertyName(int userType) const

C++ documentation:

Returns the property name used to access data for the given userType of data.