Module qt_widgets::key_event_transition::overloading

Traits

KeyEventTransitionAnimationsArgs
This trait represents a set of arguments accepted by KeyEventTransition::animations method.

KeyEventTransitionConnectArgs
This trait represents a set of arguments accepted by KeyEventTransition::connect method.

KeyEventTransitionConnectStaticArgs
This trait represents a set of arguments accepted by KeyEventTransition::connect_static method.

KeyEventTransitionDestroyedArgs
This trait represents a set of arguments accepted by KeyEventTransition::destroyed method.

KeyEventTransitionDisconnectArgs
This trait represents a set of arguments accepted by KeyEventTransition::disconnect method.

KeyEventTransitionDisconnectStaticArgs
This trait represents a set of arguments accepted by KeyEventTransition::disconnect_static method.

KeyEventTransitionDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by KeyEventTransition::dynamic_property_names method.

KeyEventTransitionNewArgs
This trait represents a set of arguments accepted by KeyEventTransition::new method.

KeyEventTransitionObjectNameArgs
This trait represents a set of arguments accepted by KeyEventTransition::object_name method.

KeyEventTransitionPropertyArgs
This trait represents a set of arguments accepted by KeyEventTransition::property method.

KeyEventTransitionStartTimerArgs
This trait represents a set of arguments accepted by KeyEventTransition::start_timer method.

KeyEventTransitionTargetStatesArgs
This trait represents a set of arguments accepted by KeyEventTransition::target_states method.

KeyEventTransitionTrArgs
This trait represents a set of arguments accepted by KeyEventTransition::tr method.

KeyEventTransitionTrUtf8Args
This trait represents a set of arguments accepted by KeyEventTransition::tr_utf8 method.