Module qt_widgets::layout::overloading

Traits

LayoutClosestAcceptableSizeArgs
This trait represents a set of arguments accepted by Layout::closest_acceptable_size method.

LayoutConnectArgs
This trait represents a set of arguments accepted by Layout::connect method.

LayoutConnectStaticArgs
This trait represents a set of arguments accepted by Layout::connect_static method.

LayoutContentsMarginsArgs
This trait represents a set of arguments accepted by Layout::contents_margins method.

LayoutContentsRectArgs
This trait represents a set of arguments accepted by Layout::contents_rect method.

LayoutDestroyedArgs
This trait represents a set of arguments accepted by Layout::destroyed method.

LayoutDisconnectArgs
This trait represents a set of arguments accepted by Layout::disconnect method.

LayoutDisconnectStaticArgs
This trait represents a set of arguments accepted by Layout::disconnect_static method.

LayoutDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Layout::dynamic_property_names method.

LayoutGeometryArgs
This trait represents a set of arguments accepted by Layout::geometry method.

LayoutMaximumSizeArgs
This trait represents a set of arguments accepted by Layout::maximum_size method.

LayoutMinimumSizeArgs
This trait represents a set of arguments accepted by Layout::minimum_size method.

LayoutObjectNameArgs
This trait represents a set of arguments accepted by Layout::object_name method.

LayoutPropertyArgs
This trait represents a set of arguments accepted by Layout::property method.

LayoutReplaceWidgetArgs
This trait represents a set of arguments accepted by Layout::replace_widget method.

LayoutSetAlignmentArgs
This trait represents a set of arguments accepted by Layout::set_alignment method.

LayoutSetContentsMarginsArgs
This trait represents a set of arguments accepted by Layout::set_contents_margins method.

LayoutSizeHintArgs
This trait represents a set of arguments accepted by Layout::size_hint method.

LayoutStartTimerArgs
This trait represents a set of arguments accepted by Layout::start_timer method.

LayoutTotalMaximumSizeArgs
This trait represents a set of arguments accepted by Layout::total_maximum_size method.

LayoutTotalMinimumSizeArgs
This trait represents a set of arguments accepted by Layout::total_minimum_size method.

LayoutTotalSizeHintArgs
This trait represents a set of arguments accepted by Layout::total_size_hint method.

LayoutTrArgs
This trait represents a set of arguments accepted by Layout::tr method.

LayoutTrUtf8Args
This trait represents a set of arguments accepted by Layout::tr_utf8 method.