Struct qt_widgets::layout::Layout

pub struct Layout {
    // some fields omitted
}

C++ type: QLayout

C++ documentation:

The QLayout class is the base class of geometry managers.

This is an abstract base class inherited by the concrete classes QBoxLayout, QGridLayout, QFormLayout, and QStackedLayout.

For users of QLayout subclasses or of QMainWindow there is seldom any need to use the basic functions provided by QLayout, such as setSizeConstraint() or setMenuBar(). See Layout Management for more information.

To make your own layout manager, implement the functions addItem(), sizeHint(), setGeometry(), itemAt() and takeAt(). You should also implement minimumSize() to ensure your layout isn't resized to zero size if there is too little space. To support children whose heights depend on their widths, implement hasHeightForWidth() and heightForWidth(). See the Border Layout and Flow Layout examples for more information about implementing custom layout managers.

Geometry management stops when the layout manager is deleted.

Methods

impl Layout

fn activate(&mut self) -> bool

C++ method: bool QLayout::activate()

C++ documentation:

Redoes the layout for parentWidget() if necessary.

You should generally not need to call this because it is automatically called at the most appropriate times. It returns true if the layout was redone.

See also update() and QWidget::updateGeometry().

fn add_item(&mut self, arg1: *mut LayoutItem)

C++ method: pure virtual void QLayout::addItem(QLayoutItem* arg1)

C++ documentation:

Implemented in subclasses to add an item. How it is added is specific to each subclass.

This function is not usually called in application code. To add a widget to a layout, use the addWidget() function; to add a child layout, use the addLayout() function provided by the relevant QLayout subclass.

Note: The ownership of item is transferred to the layout, and it's the layout's responsibility to delete it.

See also addWidget(), QBoxLayout::addLayout(), and QGridLayout::addLayout().

fn add_widget(&mut self, w: *mut Widget)

C++ method: void QLayout::addWidget(QWidget* w)

C++ documentation:

Adds widget w to this layout in a manner specific to the layout. This function uses addItem().

fn alignment(&self) -> Flags<AlignmentFlag>

C++ method: QFlags<Qt::AlignmentFlag> QLayout::alignment() const

C++ documentation:

Returns the alignment of this item.

See also setAlignment().

fn block_signals(&mut self, b: bool) -> bool

C++ method: bool QLayout::blockSignals(bool b)

C++ documentation:

If block is true, signals emitted by this object are blocked (i.e., emitting a signal will not invoke anything connected to it). If block is false, no such blocking will occur.

The return value is the previous value of signalsBlocked().

Note that the destroyed() signal will be emitted even if the signals for this object have been blocked.

See also signalsBlocked().

fn children<'l0>(&'l0 self) -> &'l0 ListObjectMutPtr

C++ method: const QList<QObject*>& QLayout::children() const

C++ documentation:

Returns a list of child objects. The QObjectList class is defined in the <QObject> header file as the following:

typedef QList<QObject*> QObjectList;

The first child added is the first object in the list and the last child added is the last object in the list, i.e. new children are appended at the end.

Note that the list order changes when QWidget children are raised or lowered. A widget that is raised becomes the last object in the list, and a widget that is lowered becomes the first object in the list.

See also findChild(), findChildren(), parent(), and setParent().

fn closest_acceptable_size<Args>(args: Args) -> Args::ReturnType where Args: LayoutClosestAcceptableSizeArgs

C++ method: QLayout::closestAcceptableSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn closest_acceptable_size((*const ::widget::Widget, &qt_core::size::Size, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn closest_acceptable_size((*const ::widget::Widget, &qt_core::size::Size, cpp_utils::AsStruct)) -> qt_core::size::Size
C++ method: static QSize QLayout::closestAcceptableSize(const QWidget* w, const QSize& s)

C++ documentation:

Returns a size that satisfies all size constraints on widget, including heightForWidth() and that is as close as possible to size.

fn connect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LayoutConnectArgs<'largs>

C++ method: QLayout::connect

This is an overloaded function. Available variants:

Rust arguments:
1) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: QMetaObject::Connection QLayout::connect(const QObject* sender, const char* signal, const char* member, Qt::ConnectionType type = ?) const

C++ documentation:

This function overloads connect().

Connects signal from the sender object to this object's method.

Equivalent to connect(sender, signal, this, method, type).

Every connection you make emits a signal, so duplicate connections emit two signals. You can break a connection using disconnect().

Note: This function is thread-safe.

See also disconnect().

fn connect_static<Args>(args: Args) -> Args::ReturnType where Args: LayoutConnectStaticArgs

C++ method: QLayout::connect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: static QMetaObject::Connection QLayout::connect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& method, Qt::ConnectionType type = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

The Connection handle will be invalid if it cannot create the connection, for example, the parameters were invalid. You can check if the QMetaObject::Connection is valid by casting it to a bool.

This function works in the same way as connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type) but it uses QMetaMethod to specify signal and method.

This function was introduced in Qt 4.8.

See also connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type).

Variant 2

Rust arguments:
1) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: static QMetaObject::Connection QLayout::connect(const QObject* sender, const char* signal, const QObject* receiver, const char* member, Qt::ConnectionType arg5 = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

You must use the SIGNAL() and SLOT() macros when specifying the signal and the method, for example:

QLabel label = new QLabel;
QScrollBar scrollBar = new QScrollBar;
QObject::connect(scrollBar, SIGNAL(valueChanged(int)),
                 label,  SLOT(setNum(int)));

This example ensures that the label always displays the current scroll bar value. Note that the signal and slots parameters must not contain any variable names, only the type. E.g. the following would not work and return false:

// WRONG
QObject::connect(scrollBar, SIGNAL(valueChanged(int value)),
                 label, SLOT(setNum(int value)));

A signal can also be connected to another signal:

class MyWidget : public QWidget
{
    Q_OBJECT

public: MyWidget();

signals: void buttonClicked();

private: QPushButton *myButton; };

MyWidget::MyWidget() { myButton = new QPushButton(this); connect(myButton, SIGNAL(clicked()), this, SIGNAL(buttonClicked())); }
In this example, the MyWidget constructor relays a signal from a private member variable, and makes it available under a name that relates to MyWidget.

A signal can be connected to many slots and signals. Many signals can be connected to one slot.

If a signal is connected to several slots, the slots are activated in the same order in which the connections were made, when the signal is emitted.

The function returns a QMetaObject::Connection that represents a handle to a connection if it successfully connects the signal to the slot. The connection handle will be invalid if it cannot create the connection, for example, if QObject is unable to verify the existence of either signal or method, or if their signatures aren't compatible. You can check if the handle is valid by casting it to a bool.

By default, a signal is emitted for every connection you make; two signals are emitted for duplicate connections. You can break all of these connections with a single disconnect() call. If you pass the Qt::UniqueConnection type, the connection will only be made if it is not a duplicate. If there is already a duplicate (exact same signal to the exact same slot on the same objects), the connection will fail and connect will return an invalid QMetaObject::Connection.

The optional type parameter describes the type of connection to establish. In particular, it determines whether a particular signal is delivered to a slot immediately or queued for delivery at a later time. If the signal is queued, the parameters must be of types that are known to Qt's meta-object system, because Qt needs to copy the arguments to store them in an event behind the scenes. If you try to use a queued connection and get the error message

QObject::connect: Cannot queue arguments of type 'MyType'
(Make sure 'MyType' is registered using qRegisterMetaType().)

call qRegisterMetaType() to register the data type before you establish the connection.

Note: This function is thread-safe.

See also disconnect(), sender(), qRegisterMetaType(), and Q_DECLARE_METATYPE().

fn contents_margins<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LayoutContentsMarginsArgs<'largs>

C++ method: QLayout::contentsMargins

This is an overloaded function. Available variants:

Rust arguments:
1) fn contents_margins(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::margins::Margins>
2) fn contents_margins(&self, cpp_utils::AsStruct) -> qt_core::margins::Margins
C++ method: QMargins QLayout::contentsMargins() const

C++ documentation:

Returns the margins used around the layout.

By default, QLayout uses the values provided by the style. On most platforms, the margin is 11 pixels in all directions.

This function was introduced in Qt 4.6.

See also setContentsMargins().

fn contents_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LayoutContentsRectArgs<'largs>

C++ method: QLayout::contentsRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn contents_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn contents_rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: QRect QLayout::contentsRect() const

C++ documentation:

Returns the layout's geometry() rectangle, but taking into account the contents margins.

This function was introduced in Qt 4.3.

See also setContentsMargins() and getContentsMargins().

fn control_types(&self) -> Flags<ControlType>

C++ method: virtual QFlags<QSizePolicy::ControlType> QLayout::controlTypes() const

C++ documentation:

Reimplemented from QLayoutItem::controlTypes().

fn count(&self) -> c_int

C++ method: pure virtual int QLayout::count() const

C++ documentation:

Must be implemented in subclasses to return the number of items in the layout.

See also itemAt().

fn delete_later(&mut self)

C++ method: void QLayout::deleteLater()

C++ documentation:

Schedules this object for deletion.

The object will be deleted when control returns to the event loop. If the event loop is not running when this function is called (e.g. deleteLater() is called on an object before QCoreApplication::exec()), the object will be deleted once the event loop is started. If deleteLater() is called after the main event loop has stopped, the object will not be deleted. Since Qt 4.8, if deleteLater() is called on an object that lives in a thread with no running event loop, the object will be destroyed when the thread finishes.

Note that entering and leaving a new event loop (e.g., by opening a modal dialog) will not perform the deferred deletion; for the object to be deleted, the control must return to the event loop from which deleteLater() was called.

Note: It is safe to call this function more than once; when the first deferred deletion event is delivered, any pending events for the object are removed from the event queue.

See also destroyed() and QPointer.

fn destroyed<'largs, Args>(&'largs mut self, args: Args) where Args: LayoutDestroyedArgs<'largs>

C++ method: QLayout::destroyed

This is an overloaded function. Available variants:

Rust arguments:
1) fn destroyed(&mut self, *mut qt_core::object::Object) -> ()
2) fn destroyed(&mut self, ()) -> ()
C++ method: void QLayout::destroyed(QObject* arg1 = ?)

C++ documentation:

This signal is emitted immediately before the object obj is destroyed, and can not be blocked.

All the objects's children are destroyed immediately after this signal is emitted.

See also deleteLater() and QPointer.

fn disconnect<'largs, Args>(&'largs self, args: Args) -> bool where Args: LayoutDisconnectArgs<'largs>

C++ method: QLayout::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn disconnect(&self, *const qt_core::object::Object) -> bool
2) fn disconnect(&self, (*const qt_core::object::Object, *const libc::c_char)) -> bool
C++ method: bool QLayout::disconnect(const QObject* receiver, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects all signals in this object from receiver's method.

A signal-slot connection is removed when either of the objects involved are destroyed.

Variant 2

Rust arguments:
1) fn disconnect(&self, *const libc::c_char) -> bool
2) fn disconnect(&self, (*const libc::c_char, *const qt_core::object::Object)) -> bool
3) fn disconnect(&self, (*const libc::c_char, *const qt_core::object::Object, *const libc::c_char)) -> bool
4) fn disconnect(&self, ()) -> bool
C++ method: bool QLayout::disconnect(const char* signal = ?, const QObject* receiver = ?, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects signal from method of receiver.

A signal-slot connection is removed when either of the objects involved are destroyed.

Note: This function is thread-safe.

fn disconnect_static<Args>(args: Args) -> bool where Args: LayoutDisconnectStaticArgs

C++ method: QLayout::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn disconnect_static(&qt_core::meta_object::Connection) -> bool
C++ method: static bool QLayout::disconnect(const QMetaObject::Connection& arg1)

C++ documentation:

Disconnect a connection.

If the connection is invalid or has already been disconnected, do nothing and return false.

See also connect().

Variant 2

Rust arguments: fn disconnect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod)) -> bool
C++ method: static bool QLayout::disconnect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

This function provides the same possibilities like disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method) but uses QMetaMethod to represent the signal and the method to be disconnected.

Additionally this function returnsfalse and no signals and slots disconnected if:

signal is not a member of sender class or one of its parent classes.

method is not a member of receiver class or one of its parent classes.

signal instance represents not a signal.

QMetaMethod() may be used as wildcard in the meaning "any signal" or "any slot in receiving object". In the same way 0 can be used for receiver in the meaning "any receiving object". In this case method should also be QMetaMethod(). sender parameter should be never 0.

This function was introduced in Qt 4.8.

See also disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method).

Variant 3

Rust arguments: fn disconnect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char)) -> bool
C++ method: static bool QLayout::disconnect(const QObject* sender, const char* signal, const QObject* receiver, const char* member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

A signal-slot connection is removed when either of the objects involved are destroyed.

disconnect() is typically used in three ways, as the following examples demonstrate.

Disconnect everything connected to an object's signals:
disconnect(myObject, 0, 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect();

Disconnect everything connected to a specific signal:
disconnect(myObject, SIGNAL(mySignal()), 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect(SIGNAL(mySignal()));

Disconnect a specific receiver:
disconnect(myObject, 0, myReceiver, 0);

equivalent to the non-static overloaded function

myObject->disconnect(myReceiver);

0 may be used as a wildcard, meaning "any signal", "any receiving object", or "any slot in the receiving object", respectively.

The sender may never be 0. (You cannot disconnect signals from more than one object in a single call.)

If signal is 0, it disconnects receiver and method from any signal. If not, only the specified signal is disconnected.

If receiver is 0, it disconnects anything connected to signal. If not, slots in objects other than receiver are not disconnected.

If method is 0, it disconnects anything that is connected to receiver. If not, only slots named method will be disconnected, and all other slots are left alone. The method must be 0 if receiver is left out, so you cannot disconnect a specifically-named slot on all objects.

Note: This function is thread-safe.

See also connect().

fn dump_object_info(&mut self)

C++ method: void QLayout::dumpObjectInfo()

C++ documentation:

Dumps information about signal connections, etc. for this object to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectTree().

fn dump_object_tree(&mut self)

C++ method: void QLayout::dumpObjectTree()

C++ documentation:

Dumps a tree of children to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectInfo().

fn dynamic_property_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LayoutDynamicPropertyNamesArgs<'largs>

C++ method: QLayout::dynamicPropertyNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn dynamic_property_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn dynamic_property_names(&self, cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: QList<QByteArray> QLayout::dynamicPropertyNames() const

C++ documentation:

Returns the names of all properties that were dynamically added to the object using setProperty().

This function was introduced in Qt 4.2.

fn event(&mut self, arg1: *mut Event) -> bool

C++ method: virtual bool QLayout::event(QEvent* arg1)

C++ documentation:

This virtual function receives events to an object and should return true if the event e was recognized and processed.

The event() function can be reimplemented to customize the behavior of an object.

See also installEventFilter(), timerEvent(), QCoreApplication::sendEvent(), and QCoreApplication::postEvent().

fn event_filter(&mut self, arg1: *mut Object, arg2: *mut Event) -> bool

C++ method: virtual bool QLayout::eventFilter(QObject* arg1, QEvent* arg2)

C++ documentation:

Filters events if this object has been installed as an event filter for the watched object.

In your reimplementation of this function, if you want to filter the event out, i.e. stop it being handled further, return true; otherwise return false.

Example:

class MainWindow : public QMainWindow
{
public:
    MainWindow();

protected: bool eventFilter(QObject obj, QEvent ev);

private: QTextEdit *textEdit; };

MainWindow::MainWindow() { textEdit = new QTextEdit; setCentralWidget(textEdit);

textEdit->installEventFilter(this); }

bool MainWindow::eventFilter(QObject obj, QEvent event) { if (obj == textEdit) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent>(event); qDebug() << "Ate key press" << keyEvent->key(); return true; } else { return false; } } else { // pass the event on to the parent class return QMainWindow::eventFilter(obj, event); } }
Notice in the example above that unhandled events are passed to the base class's eventFilter() function, since the base class might have reimplemented eventFilter() for its own internal purposes.

Warning: If you delete the receiver object in this function, be sure to return true. Otherwise, Qt will forward the event to the deleted object and the program might crash.

See also installEventFilter().

fn expanding_directions(&self) -> Flags<Orientation>

C++ method: virtual QFlags<Qt::Orientation> QLayout::expandingDirections() const

C++ documentation:

Reimplemented from QLayoutItem::expandingDirections().

Returns whether this layout can make use of more space than sizeHint(). A value of Qt::Vertical or Qt::Horizontal means that it wants to grow in only one dimension, whereas Qt::Vertical | Qt::Horizontal means that it wants to grow in both dimensions.

The default implementation returns Qt::Horizontal | Qt::Vertical. Subclasses reimplement it to return a meaningful value based on their child widgets's size policies.

See also sizeHint().

fn geometry<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LayoutGeometryArgs<'largs>

C++ method: QLayout::geometry

This is an overloaded function. Available variants:

Rust arguments:
1) fn geometry(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn geometry(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: virtual QRect QLayout::geometry() const

C++ documentation:

Reimplemented from QLayoutItem::geometry().

See also setGeometry().

fn get_contents_margins(&self, left: *mut c_int, top: *mut c_int, right: *mut c_int, bottom: *mut c_int)

C++ method: void QLayout::getContentsMargins(int* left, int* top, int* right, int* bottom) const

C++ documentation:

Extracts the left, top, right, and bottom margins used around the layout, and assigns them to *left, *top, *right, and *bottom (unless they are null pointers).

By default, QLayout uses the values provided by the style. On most platforms, the margin is 11 pixels in all directions.

This function was introduced in Qt 4.3.

See also setContentsMargins(), QStyle::pixelMetric(), PM_LayoutLeftMargin, PM_LayoutTopMargin, PM_LayoutRightMargin, and PM_LayoutBottomMargin.

fn has_height_for_width(&self) -> bool

C++ method: virtual bool QLayout::hasHeightForWidth() const

C++ documentation:

Returns true if this layout's preferred height depends on its width; otherwise returns false. The default implementation returns false.

Reimplement this function in layout managers that support height for width.

See also heightForWidth() and QWidget::heightForWidth().

fn height_for_width(&self, arg1: c_int) -> c_int

C++ method: virtual int QLayout::heightForWidth(int arg1) const

C++ documentation:

Returns the preferred height for this layout item, given the width w.

The default implementation returns -1, indicating that the preferred height is independent of the width of the item. Using the function hasHeightForWidth() will typically be much faster than calling this function and testing for -1.

Reimplement this function in layout managers that support height for width. A typical implementation will look like this:

int MyLayout::heightForWidth(int w) const
{
    if (cache_dirty || cached_width != w) {
        // not all C++ compilers support "mutable"
        MyLayout that = (MyLayout)this;
        int h = calculateHeightForWidth(w);
        that->cached_hfw = h;
        return h;
    }
    return cached_hfw;
}

Caching is strongly recommended; without it layout will take exponential time.

See also hasHeightForWidth().

fn index_of(&self, arg1: *mut Widget) -> c_int

C++ method: virtual int QLayout::indexOf(QWidget* arg1) const

C++ documentation:

Searches for widget widget in this layout (not including child layouts).

Returns the index of widget, or -1 if widget is not found.

The default implementation iterates over all items using itemAt()

fn inherits(&self, classname: *const c_char) -> bool

C++ method: bool QLayout::inherits(const char* classname) const

C++ documentation:

Returns true if this object is an instance of a class that inherits className or a QObject subclass that inherits className; otherwise returns false.

A class is considered to inherit itself.

Example:

QTimer *timer = new QTimer;         // QTimer inherits QObject
timer->inherits("QTimer");          // returns true
timer->inherits("QObject");         // returns true
timer->inherits("QAbstractButton"); // returns false

// QVBoxLayout inherits QObject and QLayoutItem QVBoxLayout *layout = new QVBoxLayout; layout->inherits("QObject"); // returns true layout->inherits("QLayoutItem"); // returns true (even though QLayoutItem is not a QObject)
If you need to determine whether an object is an instance of a particular class for the purpose of casting it, consider using qobject_cast<Type *>(object) instead.

See also metaObject() and qobject_cast().

fn install_event_filter(&mut self, arg1: *mut Object)

C++ method: void QLayout::installEventFilter(QObject* arg1)

C++ documentation:

Installs an event filter filterObj on this object. For example:

monitoredObj->installEventFilter(filterObj);

An event filter is an object that receives all events that are sent to this object. The filter can either stop the event or forward it to this object. The event filter filterObj receives events via its eventFilter() function. The eventFilter() function must return true if the event should be filtered, (i.e. stopped); otherwise it must return false.

If multiple event filters are installed on a single object, the filter that was installed last is activated first.

Here's a KeyPressEater class that eats the key presses of its monitored objects:

class KeyPressEater : public QObject
{
    Q_OBJECT
    ...

protected: bool eventFilter(QObject obj, QEvent event); };

bool KeyPressEater::eventFilter(QObject obj, QEvent event) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent >(event); qDebug("Ate key press %d", keyEvent->key()); return true; } else { // standard event processing return QObject::eventFilter(obj, event); } }
And here's how to install it on two widgets:

KeyPressEater keyPressEater = new KeyPressEater(this);
QPushButton pushButton = new QPushButton(this);
QListView *listView = new QListView(this);

pushButton->installEventFilter(keyPressEater); listView->installEventFilter(keyPressEater);
The QShortcut class, for example, uses this technique to intercept shortcut key presses.

Warning: If you delete the receiver object in your eventFilter() function, be sure to return true. If you return false, Qt sends the event to the deleted object and the program will crash.

Note that the filtering object must be in the same thread as this object. If filterObj is in a different thread, this function does nothing. If either filterObj or this object are moved to a different thread after calling this function, the event filter will not be called until both objects have the same thread affinity again (it is not removed).

See also removeEventFilter(), eventFilter(), and event().

fn invalidate(&mut self)

C++ method: virtual void QLayout::invalidate()

C++ documentation:

Reimplemented from QLayoutItem::invalidate().

fn is_empty(&self) -> bool

C++ method: virtual bool QLayout::isEmpty() const

C++ documentation:

Reimplemented from QLayoutItem::isEmpty().

fn is_enabled(&self) -> bool

C++ method: bool QLayout::isEnabled() const

C++ documentation:

Returns true if the layout is enabled; otherwise returns false.

See also setEnabled().

fn is_widget_type(&self) -> bool

C++ method: bool QLayout::isWidgetType() const

C++ documentation:

Returns true if the object is a widget; otherwise returns false.

Calling this function is equivalent to calling inherits("QWidget"), except that it is much faster.

fn is_window_type(&self) -> bool

C++ method: bool QLayout::isWindowType() const

C++ documentation:

Returns true if the object is a window; otherwise returns false.

Calling this function is equivalent to calling inherits("QWindow"), except that it is much faster.

fn item_at(&self, index: c_int) -> *mut LayoutItem

C++ method: pure virtual QLayoutItem* QLayout::itemAt(int index) const

C++ documentation:

Must be implemented in subclasses to return the layout item at index. If there is no such item, the function must return 0. Items are numbered consecutively from 0. If an item is deleted, other items will be renumbered.

This function can be used to iterate over a layout. The following code will draw a rectangle for each layout item in the layout structure of the widget.

static void paintLayout(QPainter painter, QLayoutItem item)
{
    QLayout *layout = item->layout();
    if (layout) {
        for (int i = 0; i < layout->count(); ++i)
            paintLayout(painter, layout->itemAt(i));
    }
    painter->drawRect(item->geometry());
}

void MyWidget::paintEvent(QPaintEvent *) { QPainter painter(this); if (layout()) paintLayout(&painter, layout()); }
See also count() and takeAt().

fn kill_timer(&mut self, id: c_int)

C++ method: void QLayout::killTimer(int id)

C++ documentation:

Kills the timer with timer identifier, id.

The timer identifier is returned by startTimer() when a timer event is started.

See also timerEvent() and startTimer().

fn layout(&mut self) -> *mut Layout

C++ method: virtual QLayout* QLayout::layout()

C++ documentation:

Reimplemented from QLayoutItem::layout().

fn margin(&self) -> c_int

C++ method: int QLayout::margin() const

C++ documentation:

This property holds the width of the outside border of the layout.

Use setContentsMargins() and getContentsMargins() instead.

Access functions:

int 	margin() const

void 	setMargin(int margin)

See also contentsRect() and spacing.

fn maximum_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LayoutMaximumSizeArgs<'largs>

C++ method: QLayout::maximumSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn maximum_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn maximum_size(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: virtual QSize QLayout::maximumSize() const

C++ documentation:

Reimplemented from QLayoutItem::maximumSize().

Returns the maximum size of this layout. This is the largest size that the layout can have while still respecting the specifications.

The returned value doesn't include the space required by QWidget::setContentsMargins() or menuBar().

The default implementation allows unlimited resizing.

fn menu_bar(&self) -> *mut Widget

C++ method: QWidget* QLayout::menuBar() const

C++ documentation:

Returns the menu bar set for this layout, or 0 if no menu bar is set.

See also setMenuBar().

fn meta_object(&self) -> *const MetaObject

C++ method: virtual const QMetaObject* QLayout::metaObject() const

fn minimum_height_for_width(&self, arg1: c_int) -> c_int

C++ method: virtual int QLayout::minimumHeightForWidth(int arg1) const

C++ documentation:

Returns the minimum height this widget needs for the given width, w. The default implementation simply returns heightForWidth(w).

fn minimum_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LayoutMinimumSizeArgs<'largs>

C++ method: QLayout::minimumSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn minimum_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn minimum_size(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: virtual QSize QLayout::minimumSize() const

C++ documentation:

Reimplemented from QLayoutItem::minimumSize().

Returns the minimum size of this layout. This is the smallest size that the layout can have while still respecting the specifications.

The returned value doesn't include the space required by QWidget::setContentsMargins() or menuBar().

The default implementation allows unlimited resizing.

fn move_to_thread(&mut self, thread: *mut Thread)

C++ method: void QLayout::moveToThread(QThread* thread)

C++ documentation:

Changes the thread affinity for this object and its children. The object cannot be moved if it has a parent. Event processing will continue in the targetThread.

To move an object to the main thread, use QApplication::instance() to retrieve a pointer to the current application, and then use QApplication::thread() to retrieve the thread in which the application lives. For example:

myObject->moveToThread(QApplication::instance()->thread());

If targetThread is zero, all event processing for this object and its children stops.

Note that all active timers for the object will be reset. The timers are first stopped in the current thread and restarted (with the same interval) in the targetThread. As a result, constantly moving an object between threads can postpone timer events indefinitely.

A QEvent::ThreadChange event is sent to this object just before the thread affinity is changed. You can handle this event to perform any special processing. Note that any new events that are posted to this object will be handled in the targetThread.

Warning: This function is not thread-safe; the current thread must be same as the current thread affinity. In other words, this function can only "push" an object from the current thread to another thread, it cannot "pull" an object from any arbitrary thread to the current thread.

See also thread().

fn object_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LayoutObjectNameArgs<'largs>

C++ method: QLayout::objectName

This is an overloaded function. Available variants:

Rust arguments:
1) fn object_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn object_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QLayout::objectName() const

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn parent(&self) -> *mut Object

C++ method: QObject* QLayout::parent() const

C++ documentation:

Returns a pointer to the parent object.

See also setParent() and children().

fn parent_widget(&self) -> *mut Widget

C++ method: QWidget* QLayout::parentWidget() const

C++ documentation:

Returns the parent widget of this layout, or 0 if this layout is not installed on any widget.

If the layout is a sub-layout, this function returns the parent widget of the parent layout.

See also parent().

fn property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LayoutPropertyArgs<'largs>

C++ method: QLayout::property

This is an overloaded function. Available variants:

Rust arguments:
1) fn property(&self, (*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn property(&self, (*const libc::c_char, cpp_utils::AsStruct)) -> qt_core::variant::Variant
C++ method: QVariant QLayout::property(const char* name) const

C++ documentation:

Returns the value of the object's name property.

If no such property exists, the returned variant is invalid.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

See also setProperty(), QVariant::isValid(), metaObject(), and dynamicPropertyNames().

fn qt_metacall(&mut self, arg1: Call, arg2: c_int, arg3: *mut *mut c_void) -> c_int

C++ method: virtual int QLayout::qt_metacall(QMetaObject::Call arg1, int arg2, void** arg3)

fn qt_metacast(&mut self, arg1: *const c_char) -> *mut c_void

C++ method: virtual void* QLayout::qt_metacast(const char* arg1)

fn register_user_data() -> c_uint

C++ method: static unsigned int QLayout::registerUserData()

fn remove_event_filter(&mut self, arg1: *mut Object)

C++ method: void QLayout::removeEventFilter(QObject* arg1)

C++ documentation:

Removes an event filter object obj from this object. The request is ignored if such an event filter has not been installed.

All event filters for this object are automatically removed when this object is destroyed.

It is always safe to remove an event filter, even during event filter activation (i.e. from the eventFilter() function).

See also installEventFilter(), eventFilter(), and event().

fn remove_item(&mut self, arg1: *mut LayoutItem)

C++ method: void QLayout::removeItem(QLayoutItem* arg1)

C++ documentation:

Removes the layout item item from the layout. It is the caller's responsibility to delete the item.

Notice that item can be a layout (since QLayout inherits QLayoutItem).

See also removeWidget() and addItem().

fn remove_widget(&mut self, w: *mut Widget)

C++ method: void QLayout::removeWidget(QWidget* w)

C++ documentation:

Removes the widget widget from the layout. After this call, it is the caller's responsibility to give the widget a reasonable geometry or to put the widget back into a layout.

Note: The ownership of widget remains the same as when it was added.

See also removeItem(), QWidget::setGeometry(), and addWidget().

fn replace_widget<'largs, Args>(&'largs mut self, args: Args) -> *mut LayoutItem where Args: LayoutReplaceWidgetArgs<'largs>

C++ method: QLayout::replaceWidget

This is an overloaded function. Available variants:

Rust arguments:
1) fn replace_widget(&mut self, (*mut ::widget::Widget, *mut ::widget::Widget)) -> *mut ::layout_item::LayoutItem
2) fn replace_widget(&mut self, (*mut ::widget::Widget, *mut ::widget::Widget, qt_core::flags::Flags<qt_core::qt::FindChildOption>)) -> *mut ::layout_item::LayoutItem
C++ method: QLayoutItem* QLayout::replaceWidget(QWidget* from, QWidget* to, QFlags<Qt::FindChildOption> options = ?)

C++ documentation:

Searches for widget from and replaces it with widget to if found. Returns the layout item that contains the widget from on success. Otherwise 0 is returned. If options contains Qt::FindChildrenRecursively (the default), sub-layouts are searched for doing the replacement. Any other flag in options is ignored.

Notice that the returned item therefore might not belong to this layout, but to a sub-layout.

The returned layout item is no longer owned by the layout and should be either deleted or inserted to another layout. The widget from is no longer managed by the layout and may need to be deleted or hidden. The parent of widget from is left unchanged.

This function works for the built-in Qt layouts, but might not work for custom layouts.

This function was introduced in Qt 5.2.

See also indexOf().

fn set_alignment<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: LayoutSetAlignmentArgs<'largs>

C++ method: QLayout::setAlignment

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_alignment(&mut self, (*mut ::layout::Layout, qt_core::flags::Flags<qt_core::qt::AlignmentFlag>)) -> bool
C++ method: bool QLayout::setAlignment(QLayout* l, QFlags<Qt::AlignmentFlag> alignment)

C++ documentation:

This is an overloaded function.

Sets the alignment for the layout l to alignment and returns true if l is found in this layout (not including child layouts); otherwise returns false.

Variant 2

Rust arguments: fn set_alignment(&mut self, (*mut ::widget::Widget, qt_core::flags::Flags<qt_core::qt::AlignmentFlag>)) -> bool
C++ method: bool QLayout::setAlignment(QWidget* w, QFlags<Qt::AlignmentFlag> alignment)

C++ documentation:

Sets the alignment for widget w to alignment and returns true if w is found in this layout (not including child layouts); otherwise returns false.

Variant 3

Rust arguments: fn set_alignment(&mut self, qt_core::flags::Flags<qt_core::qt::AlignmentFlag>) -> ()
C++ method: void QLayout::setAlignment(QFlags<Qt::AlignmentFlag> a)

C++ documentation:

Sets the alignment of this item to alignment.

Note: Item alignment is only supported by QLayoutItem subclasses where it would have a visual effect. Except for QSpacerItem, which provides blank space for layouts, all public Qt classes that inherit QLayoutItem support item alignment.

See also alignment().

fn set_contents_margins<'largs, Args>(&'largs mut self, args: Args) where Args: LayoutSetContentsMarginsArgs<'largs>

C++ method: QLayout::setContentsMargins

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_contents_margins(&mut self, &qt_core::margins::Margins) -> ()
C++ method: void QLayout::setContentsMargins(const QMargins& margins)

C++ documentation:

Sets the margins to use around the layout.

By default, QLayout uses the values provided by the style. On most platforms, the margin is 11 pixels in all directions.

This function was introduced in Qt 4.6.

See also contentsMargins().

Variant 2

Rust arguments: fn set_contents_margins(&mut self, (libc::c_int, libc::c_int, libc::c_int, libc::c_int)) -> ()
C++ method: void QLayout::setContentsMargins(int left, int top, int right, int bottom)

C++ documentation:

Sets the left, top, right, and bottom margins to use around the layout.

By default, QLayout uses the values provided by the style. On most platforms, the margin is 11 pixels in all directions.

This function was introduced in Qt 4.3.

See also contentsMargins(), getContentsMargins(), QStyle::pixelMetric(), PM_LayoutLeftMargin, PM_LayoutTopMargin, PM_LayoutRightMargin, and PM_LayoutBottomMargin.

fn set_enabled(&mut self, arg1: bool)

C++ method: void QLayout::setEnabled(bool arg1)

C++ documentation:

Enables this layout if enable is true, otherwise disables it.

An enabled layout adjusts dynamically to changes; a disabled layout acts as if it did not exist.

By default all layouts are enabled.

See also isEnabled().

fn set_geometry(&mut self, arg1: &Rect)

C++ method: virtual void QLayout::setGeometry(const QRect& arg1)

C++ documentation:

Reimplemented from QLayoutItem::setGeometry().

See also geometry().

fn set_margin(&mut self, arg1: c_int)

C++ method: void QLayout::setMargin(int arg1)

C++ documentation:

This property holds the width of the outside border of the layout.

Use setContentsMargins() and getContentsMargins() instead.

Access functions:

int 	margin() const

void 	setMargin(int margin)

See also contentsRect() and spacing.

fn set_menu_bar(&mut self, w: *mut Widget)

C++ method: void QLayout::setMenuBar(QWidget* w)

C++ documentation:

Tells the geometry manager to place the menu bar widget at the top of parentWidget(), outside QWidget::contentsMargins(). All child widgets are placed below the bottom edge of the menu bar.

See also menuBar().

fn set_object_name(&mut self, name: &String)

C++ method: void QLayout::setObjectName(const QString& name)

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn set_parent(&mut self, arg1: *mut Object)

C++ method: void QLayout::setParent(QObject* arg1)

C++ documentation:

Makes the object a child of parent.

See also parent() and children().

fn set_property(&mut self, name: *const c_char, value: &Variant) -> bool

C++ method: bool QLayout::setProperty(const char* name, const QVariant& value)

C++ documentation:

Sets the value of the object's name property to value.

If the property is defined in the class using Q_PROPERTY then true is returned on success and false otherwise. If the property is not defined using Q_PROPERTY, and therefore not listed in the meta-object, it is added as a dynamic property and false is returned.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

Dynamic properties can be queried again using property() and can be removed by setting the property value to an invalid QVariant. Changing the value of a dynamic property causes a QDynamicPropertyChangeEvent to be sent to the object.

Note: Dynamic properties starting with "_q_" are reserved for internal purposes.

See also property(), metaObject(), and dynamicPropertyNames().

fn set_size_constraint(&mut self, arg1: SizeConstraint)

C++ method: void QLayout::setSizeConstraint(QLayout::SizeConstraint arg1)

C++ documentation:

This property holds the resize mode of the layout.

The default mode is SetDefaultConstraint.

Access functions:

SizeConstraint 	sizeConstraint() const

void 	setSizeConstraint(SizeConstraint)

fn set_spacing(&mut self, arg1: c_int)

C++ method: void QLayout::setSpacing(int arg1)

C++ documentation:

This property holds the spacing between widgets inside the layout.

If no value is explicitly set, the layout's spacing is inherited from the parent layout, or from the style settings for the parent widget.

For QGridLayout and QFormLayout, it is possible to set different horizontal and vertical spacings using setHorizontalSpacing() and setVerticalSpacing(). In that case, spacing() returns -1.

Access functions:

int 	spacing() const

void 	setSpacing(int)

See also contentsRect(), getContentsMargins(), QStyle::layoutSpacing(), and QStyle::pixelMetric().

fn set_user_data(&mut self, id: c_uint, data: *mut ObjectUserData)

C++ method: void QLayout::setUserData(unsigned int id, QObjectUserData* data)

fn signals_blocked(&self) -> bool

C++ method: bool QLayout::signalsBlocked() const

C++ documentation:

Returns true if signals are blocked; otherwise returns false.

Signals are not blocked by default.

See also blockSignals().

fn size_constraint(&self) -> SizeConstraint

C++ method: QLayout::SizeConstraint QLayout::sizeConstraint() const

C++ documentation:

This property holds the resize mode of the layout.

The default mode is SetDefaultConstraint.

Access functions:

SizeConstraint 	sizeConstraint() const

void 	setSizeConstraint(SizeConstraint)

fn size_hint<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LayoutSizeHintArgs<'largs>

C++ method: QLayout::sizeHint

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_hint(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size_hint(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: pure virtual QSize QLayout::sizeHint() const

C++ documentation:

Implemented in subclasses to return the preferred size of this item.

fn spacer_item(&mut self) -> *mut SpacerItem

C++ method: virtual QSpacerItem* QLayout::spacerItem()

C++ documentation:

If this item is a QSpacerItem, it is returned as a QSpacerItem; otherwise 0 is returned. This function provides type-safe casting.

fn spacing(&self) -> c_int

C++ method: int QLayout::spacing() const

C++ documentation:

This property holds the spacing between widgets inside the layout.

If no value is explicitly set, the layout's spacing is inherited from the parent layout, or from the style settings for the parent widget.

For QGridLayout and QFormLayout, it is possible to set different horizontal and vertical spacings using setHorizontalSpacing() and setVerticalSpacing(). In that case, spacing() returns -1.

Access functions:

int 	spacing() const

void 	setSpacing(int)

See also contentsRect(), getContentsMargins(), QStyle::layoutSpacing(), and QStyle::pixelMetric().

fn start_timer<'largs, Args>(&'largs mut self, args: Args) -> c_int where Args: LayoutStartTimerArgs<'largs>

C++ method: QLayout::startTimer

This is an overloaded function. Available variants:

Rust arguments:
1) fn start_timer(&mut self, libc::c_int) -> libc::c_int
2) fn start_timer(&mut self, (libc::c_int, qt_core::qt::TimerType)) -> libc::c_int
C++ method: int QLayout::startTimer(int interval, Qt::TimerType timerType = ?)

C++ documentation:

Starts a timer and returns a timer identifier, or returns zero if it could not start a timer.

A timer event will occur every interval milliseconds until killTimer() is called. If interval is 0, then the timer event occurs once every time there are no more window system events to process.

The virtual timerEvent() function is called with the QTimerEvent event parameter class when a timer event occurs. Reimplement this function to get timer events.

If multiple timers are running, the QTimerEvent::timerId() can be used to find out which timer was activated.

Example:

class MyObject : public QObject
{
    Q_OBJECT

public: MyObject(QObject *parent = 0);

protected: void timerEvent(QTimerEvent *event); };

MyObject::MyObject(QObject *parent) : QObject(parent) { startTimer(50); // 50-millisecond timer startTimer(1000); // 1-second timer startTimer(60000); // 1-minute timer }

void MyObject::timerEvent(QTimerEvent *event) { qDebug() << "Timer ID:" << event->timerId(); }
Note that QTimer's accuracy depends on the underlying operating system and hardware. The timerType argument allows you to customize the accuracy of the timer. See Qt::TimerType for information on the different timer types. Most platforms support an accuracy of 20 milliseconds; some provide more. If Qt is unable to deliver the requested number of timer events, it will silently discard some.

The QTimer class provides a high-level programming interface with single-shot timers and timer signals instead of events. There is also a QBasicTimer class that is more lightweight than QTimer and less clumsy than using timer IDs directly.

See also timerEvent(), killTimer(), and QTimer::singleShot().

fn take_at(&mut self, index: c_int) -> *mut LayoutItem

C++ method: pure virtual QLayoutItem* QLayout::takeAt(int index)

C++ documentation:

Must be implemented in subclasses to remove the layout item at index from the layout, and return the item. If there is no such item, the function must do nothing and return 0. Items are numbered consecutively from 0. If an item is removed, other items will be renumbered.

The following code fragment shows a safe way to remove all items from a layout:

QLayoutItem *child;
while ((child = layout->takeAt(0)) != 0) {
    ...
    delete child;
}

See also itemAt() and count().

fn thread(&self) -> *mut Thread

C++ method: QThread* QLayout::thread() const

C++ documentation:

Returns the thread in which the object lives.

See also moveToThread().

fn total_height_for_width(&self, w: c_int) -> c_int

C++ method: int QLayout::totalHeightForWidth(int w) const

fn total_maximum_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LayoutTotalMaximumSizeArgs<'largs>

C++ method: QLayout::totalMaximumSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn total_maximum_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn total_maximum_size(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: QSize QLayout::totalMaximumSize() const

fn total_minimum_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LayoutTotalMinimumSizeArgs<'largs>

C++ method: QLayout::totalMinimumSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn total_minimum_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn total_minimum_size(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: QSize QLayout::totalMinimumSize() const

fn total_size_hint<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LayoutTotalSizeHintArgs<'largs>

C++ method: QLayout::totalSizeHint

This is an overloaded function. Available variants:

Rust arguments:
1) fn total_size_hint(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn total_size_hint(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: QSize QLayout::totalSizeHint() const

fn tr<Args>(args: Args) -> Args::ReturnType where Args: LayoutTrArgs

C++ method: QLayout::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QLayout::tr(const char* s, const char* c, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: LayoutTrUtf8Args

C++ method: QLayout::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QLayout::trUtf8(const char* s, const char* c, int n)

fn update(&mut self)

C++ method: void QLayout::update()

C++ documentation:

Updates the layout for parentWidget().

You should generally not need to call this because it is automatically called at the most appropriate times.

See also activate() and invalidate().

fn user_data(&self, id: c_uint) -> *mut ObjectUserData

C++ method: QObjectUserData* QLayout::userData(unsigned int id) const

fn widget(&mut self) -> *mut Widget

C++ method: virtual QWidget* QLayout::widget()

C++ documentation:

If this item is a QWidget, it is returned as a QWidget; otherwise 0 is returned. This function provides type-safe casting.