Struct qt_widgets::layout_item::LayoutItem

pub struct LayoutItem {
    // some fields omitted
}

C++ type: QLayoutItem

C++ documentation:

The QLayoutItem class provides an abstract item that a QLayout manipulates.

This is used by custom layouts.

Pure virtual functions are provided to return information about the layout, including, sizeHint(), minimumSize(), maximumSize() and expanding().

The layout's geometry can be set and retrieved with setGeometry() and geometry(), and its alignment with setAlignment() and alignment().

isEmpty() returns whether the layout item is empty. If the concrete item is a QWidget, it can be retrieved using widget(). Similarly for layout() and spacerItem().

Some layouts have width and height interdependencies. These can be expressed using hasHeightForWidth(), heightForWidth(), and minimumHeightForWidth(). For more explanation see the Qt Quarterly article Trading Height for Width.

Methods

impl LayoutItem

fn alignment(&self) -> Flags<AlignmentFlag>

C++ method: QFlags<Qt::AlignmentFlag> QLayoutItem::alignment() const

C++ documentation:

Returns the alignment of this item.

See also setAlignment().

fn control_types(&self) -> Flags<ControlType>

C++ method: virtual QFlags<QSizePolicy::ControlType> QLayoutItem::controlTypes() const

C++ documentation:

Returns the control type(s) for the layout item. For a QWidgetItem, the control type comes from the widget's size policy; for a QLayoutItem, the control types is derived from the layout's contents.

See also QSizePolicy::controlType().

fn expanding_directions(&self) -> Flags<Orientation>

C++ method: pure virtual QFlags<Qt::Orientation> QLayoutItem::expandingDirections() const

C++ documentation:

Returns whether this layout item can make use of more space than sizeHint(). A value of Qt::Vertical or Qt::Horizontal means that it wants to grow in only one dimension, whereas Qt::Vertical | Qt::Horizontal means that it wants to grow in both dimensions.

fn geometry<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LayoutItemGeometryArgs<'largs>

C++ method: QLayoutItem::geometry

This is an overloaded function. Available variants:

Rust arguments:
1) fn geometry(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn geometry(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: pure virtual QRect QLayoutItem::geometry() const

C++ documentation:

Returns the rectangle covered by this layout item.

See also setGeometry().

fn has_height_for_width(&self) -> bool

C++ method: virtual bool QLayoutItem::hasHeightForWidth() const

C++ documentation:

Returns true if this layout's preferred height depends on its width; otherwise returns false. The default implementation returns false.

Reimplement this function in layout managers that support height for width.

See also heightForWidth() and QWidget::heightForWidth().

fn height_for_width(&self, arg1: c_int) -> c_int

C++ method: virtual int QLayoutItem::heightForWidth(int arg1) const

C++ documentation:

Returns the preferred height for this layout item, given the width w.

The default implementation returns -1, indicating that the preferred height is independent of the width of the item. Using the function hasHeightForWidth() will typically be much faster than calling this function and testing for -1.

Reimplement this function in layout managers that support height for width. A typical implementation will look like this:

int MyLayout::heightForWidth(int w) const
{
    if (cache_dirty || cached_width != w) {
        // not all C++ compilers support "mutable"
        MyLayout that = (MyLayout)this;
        int h = calculateHeightForWidth(w);
        that->cached_hfw = h;
        return h;
    }
    return cached_hfw;
}

Caching is strongly recommended; without it layout will take exponential time.

See also hasHeightForWidth().

fn invalidate(&mut self)

C++ method: virtual void QLayoutItem::invalidate()

C++ documentation:

Invalidates any cached information in this layout item.

fn is_empty(&self) -> bool

C++ method: pure virtual bool QLayoutItem::isEmpty() const

C++ documentation:

Implemented in subclasses to return whether this item is empty, i.e. whether it contains any widgets.

fn layout(&mut self) -> *mut Layout

C++ method: virtual QLayout* QLayoutItem::layout()

C++ documentation:

If this item is a QLayout, it is returned as a QLayout; otherwise 0 is returned. This function provides type-safe casting.

fn maximum_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LayoutItemMaximumSizeArgs<'largs>

C++ method: QLayoutItem::maximumSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn maximum_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn maximum_size(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: pure virtual QSize QLayoutItem::maximumSize() const

C++ documentation:

Implemented in subclasses to return the maximum size of this item.

fn minimum_height_for_width(&self, arg1: c_int) -> c_int

C++ method: virtual int QLayoutItem::minimumHeightForWidth(int arg1) const

C++ documentation:

Returns the minimum height this widget needs for the given width, w. The default implementation simply returns heightForWidth(w).

fn minimum_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LayoutItemMinimumSizeArgs<'largs>

C++ method: QLayoutItem::minimumSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn minimum_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn minimum_size(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: pure virtual QSize QLayoutItem::minimumSize() const

C++ documentation:

Implemented in subclasses to return the minimum size of this item.

fn set_alignment(&mut self, a: Flags<AlignmentFlag>)

C++ method: void QLayoutItem::setAlignment(QFlags<Qt::AlignmentFlag> a)

C++ documentation:

Sets the alignment of this item to alignment.

Note: Item alignment is only supported by QLayoutItem subclasses where it would have a visual effect. Except for QSpacerItem, which provides blank space for layouts, all public Qt classes that inherit QLayoutItem support item alignment.

See also alignment().

fn set_geometry(&mut self, arg1: &Rect)

C++ method: pure virtual void QLayoutItem::setGeometry(const QRect& arg1)

C++ documentation:

Implemented in subclasses to set this item's geometry to r.

See also geometry().

fn size_hint<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LayoutItemSizeHintArgs<'largs>

C++ method: QLayoutItem::sizeHint

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_hint(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size_hint(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: pure virtual QSize QLayoutItem::sizeHint() const

C++ documentation:

Implemented in subclasses to return the preferred size of this item.

fn spacer_item(&mut self) -> *mut SpacerItem

C++ method: virtual QSpacerItem* QLayoutItem::spacerItem()

C++ documentation:

If this item is a QSpacerItem, it is returned as a QSpacerItem; otherwise 0 is returned. This function provides type-safe casting.

fn widget(&mut self) -> *mut Widget

C++ method: virtual QWidget* QLayoutItem::widget()

C++ documentation:

If this item is a QWidget, it is returned as a QWidget; otherwise 0 is returned. This function provides type-safe casting.