Struct qt_widgets::list::ListMdiSubWindowMutPtr

pub struct ListMdiSubWindowMutPtr {
    // some fields omitted
}

C++ type: QList<QMdiSubWindow*>

C++ documentation:

The QList class is a template class that provides lists.

QList<T> is one of Qt's generic container classes. It stores a list of values and provides fast index-based access as well as fast insertions and removals.

QList<T>, QLinkedList<T>, and QVector<T> provide similar functionality. Here's an overview:

For most purposes, QList is the right class to use. Its index-based API is more convenient than QLinkedList's iterator-based API, and it is usually faster than QVector because of the way it stores its items in memory. It also expands to less code in your executable.

If you need a real linked list, with guarantees of constant time insertions in the middle of the list and iterators to items rather than indexes, use QLinkedList.

If you want the items to occupy adjacent memory positions, use QVector.

Internally, QList<T> is represented as an array of pointers to items of type T. If T is itself a pointer type or a basic type that is no larger than a pointer, or if T is one of Qt's shared classes, then QList<T> stores the items directly in the pointer array. For lists under a thousand items, this array representation allows for very fast insertions in the middle, and it allows index-based access. Furthermore, operations like prepend() and append() are very fast, because QList preallocates memory at both ends of its internal array. (See Algorithmic Complexity for details.) Note, however, that for unshared list items that are larger than a pointer, each append or insert of a new item requires allocating the new item on the heap, and this per item allocation might make QVector a better choice in cases that do lots of appending or inserting, since QVector allocates memory for its items in a single heap allocation.

Note that the internal array only ever gets bigger over the life of the list. It never shrinks. The internal array is deallocated by the destructor and by the assignment operator, when one list is assigned to another.

Here's an example of a QList that stores integers and a QList that stores QDate values:

QList<int> integerList;
QList<QDate> dateList;

Qt includes a QStringList class that inherits QList<QString> and adds a few convenience functions, such as QStringList::join() and QStringList::filter(). QString::split() creates QStringLists from strings.

QList stores a list of items. The default constructor creates an empty list. To insert items into the list, you can use operator<<():

QList<QString> list;
list << "one" << "two" << "three";
// list: ["one", "two", "three"]

QList provides these basic functions to add, move, and remove items: insert(), replace(), removeAt(), move(), and swap(). In addition, it provides the following convenience functions: append(), prepend(), removeFirst(), and removeLast().

QList uses 0-based indexes, just like C++ arrays. To access the item at a particular index position, you can use operator. On non-const lists, operator returns a reference to the item and can be used on the left side of an assignment:

if (list[0] == "Bob")
    list[0] = "Robert";

Because QList is implemented as an array of pointers, this operation is very fast (constant time). For read-only access, an alternative syntax is to use at():

for (int i = 0; i < list.size(); ++i) {
    if (list.at(i) == "Jane")
        cout << "Found Jane at position " << i << endl;
}

at() can be faster than operator, because it never causes a deep copy to occur.

A common requirement is to remove an item from a list and do something with it. For this, QList provides takeAt(), takeFirst(), and takeLast(). Here's a loop that removes the items from a list one at a time and calls delete on them:

QList<QWidget *> list;
...
while (!list.isEmpty())
    delete list.takeFirst();

Inserting and removing items at either ends of the list is very fast (constant time in most cases), because QList preallocates extra space on both sides of its internal buffer to allow for fast growth at both ends of the list.

If you want to find all occurrences of a particular value in a list, use indexOf() or lastIndexOf(). The former searches forward starting from a given index position, the latter searches backward. Both return the index of a matching item if they find it; otherwise, they return -1. For example:

int i = list.indexOf("Jane");
if (i != -1)
    cout << "First occurrence of Jane is at position " << i << endl;

If you simply want to check whether a list contains a particular value, use contains(). If you want to find out how many times a particular value occurs in the list, use count(). If you want to replace all occurrences of a particular value with another, use replace().

QList's value type must be an assignable data type. This covers most data types that are commonly used, but the compiler won't let you, for example, store a QWidget as a value; instead, store a QWidget *. A few functions have additional requirements; for example, indexOf() and lastIndexOf() expect the value type to support operator==(). These requirements are documented on a per-function basis.

Like the other container classes, QList provides Java-style iterators (QListIterator and QMutableListIterator) and STL-style iterators (QList::const_iterator and QList::iterator). In practice, these are rarely used, because you can use indexes into the QList. QList is implemented in such a way that direct index-based access is just as fast as using iterators.

QList does not support inserting, prepending, appending or replacing with references to its own values. Doing so will cause your application to abort with an error message.

To make QList as efficient as possible, its member functions don't validate their input before using it. Except for isEmpty(), member functions always assume the list is not empty. Member functions that take index values as parameters always assume their index value parameters are in the valid range. This means QList member functions can fail. If you define QT_NO_DEBUG when you compile, failures will not be detected. If you don't define QT_NO_DEBUG, failures will be detected using Q_ASSERT() or Q_ASSERT_X() with an appropriate message.

To avoid failures when your list can be empty, call isEmpty() before calling other member functions. If you must pass an index value that might not be in the valid range, check that it is less than the value returned by size() but not less than 0.

Methods

impl ListMdiSubWindowMutPtr

fn append<'largs, Args>(&'largs mut self, args: Args) where Args: ListMdiSubWindowMutPtrAppendArgs<'largs>

C++ method: QList<QMdiSubWindow*>::append

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn append(&mut self, &*mut ::mdi_sub_window::MdiSubWindow) -> ()
C++ method: void QList<QMdiSubWindow*>::append(QMdiSubWindow* const & t)

C++ documentation:

Inserts value at the end of the list.

Example:

QList<QString> list;
list.append("one");
list.append("two");
list.append("three");
// list: ["one", "two", "three"]

This is the same as list.insert(size(), value).

If this list is not shared, this operation is typically very fast (amortized constant time), because QList preallocates extra space on both sides of its internal buffer to allow for fast growth at both ends of the list.

See also operator<<(), prepend(), and insert().

Variant 2

Rust arguments: fn append(&mut self, &::list::ListMdiSubWindowMutPtr) -> ()
C++ method: void QList<QMdiSubWindow*>::append(const QList<QMdiSubWindow*>& t)

C++ documentation:

This is an overloaded function.

Appends the items of the value list to this list.

This function was introduced in Qt 4.5.

See also operator<<() and operator+=().

fn at<'l0>(&'l0 self, i: c_int) -> &'l0 *mut MdiSubWindow

C++ method: QMdiSubWindow* const & QList<QMdiSubWindow*>::at(int i) const

C++ documentation:

Returns the item at index position i in the list. i must be a valid index position in the list (i.e., 0 <= i < size()).

This function is very fast (constant time).

See also value() and operator[]().

fn back<'l0>(&'l0 self) -> &'l0 *mut MdiSubWindow

C++ method: QMdiSubWindow* const & QList<QMdiSubWindow*>::back() const

C++ documentation:

This is an overloaded function.

fn back_mut<'l0>(&'l0 mut self) -> &'l0 mut *mut MdiSubWindow

C++ method: QMdiSubWindow*& QList<QMdiSubWindow*>::back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to last(). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

fn clear(&mut self)

C++ method: void QList<QMdiSubWindow*>::clear()

C++ documentation:

Removes all items from the list.

See also removeAll().

fn contains(&self, t: &*mut MdiSubWindow) -> bool

C++ method: bool QList<QMdiSubWindow*>::contains(QMdiSubWindow* const & t) const

C++ documentation:

Returns true if the list contains an occurrence of value; otherwise returns false.

This function requires the value type to have an implementation of operator==().

See also indexOf() and count().

fn count<'largs, Args>(&'largs self, args: Args) -> c_int where Args: ListMdiSubWindowMutPtrCountArgs<'largs>

C++ method: QList<QMdiSubWindow*>::count

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn count(&self, ()) -> libc::c_int
C++ method: int QList<QMdiSubWindow*>::count() const

C++ documentation:

Returns the number of items in the list. This is effectively the same as size().

Variant 2

Rust arguments: fn count(&self, &*mut ::mdi_sub_window::MdiSubWindow) -> libc::c_int
C++ method: int QList<QMdiSubWindow*>::count(QMdiSubWindow* const & t) const

C++ documentation:

Returns the number of occurrences of value in the list.

This function requires the value type to have an implementation of operator==().

See also contains() and indexOf().

fn detach(&mut self)

C++ method: void QList<QMdiSubWindow*>::detach()

fn detach_shared(&mut self)

C++ method: void QList<QMdiSubWindow*>::detachShared()

fn empty(&self) -> bool

C++ method: bool QList<QMdiSubWindow*>::empty() const

C++ documentation:

This function is provided for STL compatibility. It is equivalent to isEmpty() and returns true if the list is empty.

fn ends_with(&self, t: &*mut MdiSubWindow) -> bool

C++ method: bool QList<QMdiSubWindow*>::endsWith(QMdiSubWindow* const & t) const

C++ documentation:

Returns true if this list is not empty and its last item is equal to value; otherwise returns false.

This function was introduced in Qt 4.5.

See also isEmpty() and contains().

fn first<'l0>(&'l0 self) -> &'l0 *mut MdiSubWindow

C++ method: QMdiSubWindow* const & QList<QMdiSubWindow*>::first() const

C++ documentation:

This is an overloaded function.

fn first_mut<'l0>(&'l0 mut self) -> &'l0 mut *mut MdiSubWindow

C++ method: QMdiSubWindow*& QList<QMdiSubWindow*>::first()

C++ documentation:

Returns a reference to the first item in the list. The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

See also last() and isEmpty().

fn front<'l0>(&'l0 self) -> &'l0 *mut MdiSubWindow

C++ method: QMdiSubWindow* const & QList<QMdiSubWindow*>::front() const

C++ documentation:

This is an overloaded function.

fn front_mut<'l0>(&'l0 mut self) -> &'l0 mut *mut MdiSubWindow

C++ method: QMdiSubWindow*& QList<QMdiSubWindow*>::front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to first(). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

fn index_of<'largs, Args>(&'largs self, args: Args) -> c_int where Args: ListMdiSubWindowMutPtrIndexOfArgs<'largs>

C++ method: QList<QMdiSubWindow*>::indexOf

This is an overloaded function. Available variants:

Rust arguments:
1) fn index_of(&self, &*mut ::mdi_sub_window::MdiSubWindow) -> libc::c_int
2) fn index_of(&self, (&*mut ::mdi_sub_window::MdiSubWindow, libc::c_int)) -> libc::c_int
C++ method: int QList<QMdiSubWindow*>::indexOf(QMdiSubWindow* const & t, int from = ?) const

C++ documentation:

Returns the index position of the first occurrence of value in the list, searching forward from index position from. Returns -1 if no item matched.

Example:

QList<QString> list;
list << "A" << "B" << "C" << "B" << "A";
list.indexOf("B");          // returns 1
list.indexOf("B", 1);       // returns 1
list.indexOf("B", 2);       // returns 3
list.indexOf("X");          // returns -1

This function requires the value type to have an implementation of operator==().

Note that QList uses 0-based indexes, just like C++ arrays. Negative indexes are not supported with the exception of the value mentioned above.

See also lastIndexOf() and contains().

fn insert(&mut self, i: c_int, t: &*mut MdiSubWindow)

C++ method: void QList<QMdiSubWindow*>::insert(int i, QMdiSubWindow* const & t)

C++ documentation:

Inserts value at index position i in the list. If i is 0, the value is prepended to the list. If i is size(), the value is appended to the list.

Example:

QList<QString> list;
list << "alpha" << "beta" << "delta";
list.insert(2, "gamma");
// list: ["alpha", "beta", "gamma", "delta"]

See also append(), prepend(), replace(), and removeAt().

fn is_detached(&self) -> bool

C++ method: bool QList<QMdiSubWindow*>::isDetached() const

fn is_empty(&self) -> bool

C++ method: bool QList<QMdiSubWindow*>::isEmpty() const

C++ documentation:

Returns true if the list contains no items; otherwise returns false.

See also size().

fn is_shared_with(&self, other: &ListMdiSubWindowMutPtr) -> bool

C++ method: bool QList<QMdiSubWindow*>::isSharedWith(const QList<QMdiSubWindow*>& other) const

fn last<'l0>(&'l0 self) -> &'l0 *mut MdiSubWindow

C++ method: QMdiSubWindow* const & QList<QMdiSubWindow*>::last() const

C++ documentation:

This is an overloaded function.

fn last_index_of<'largs, Args>(&'largs self, args: Args) -> c_int where Args: ListMdiSubWindowMutPtrLastIndexOfArgs<'largs>

C++ method: QList<QMdiSubWindow*>::lastIndexOf

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_index_of(&self, &*mut ::mdi_sub_window::MdiSubWindow) -> libc::c_int
2) fn last_index_of(&self, (&*mut ::mdi_sub_window::MdiSubWindow, libc::c_int)) -> libc::c_int
C++ method: int QList<QMdiSubWindow*>::lastIndexOf(QMdiSubWindow* const & t, int from = ?) const

C++ documentation:

Returns the index position of the last occurrence of value in the list, searching backward from index position from. If from is -1 (the default), the search starts at the last item. Returns -1 if no item matched.

Example:

QList<QString> list;
list << "A" << "B" << "C" << "B" << "A";
list.lastIndexOf("B");      // returns 3
list.lastIndexOf("B", 3);   // returns 3
list.lastIndexOf("B", 2);   // returns 1
list.lastIndexOf("X");      // returns -1

This function requires the value type to have an implementation of operator==().

Note that QList uses 0-based indexes, just like C++ arrays. Negative indexes are not supported with the exception of the value mentioned above.

See also indexOf().

fn last_mut<'l0>(&'l0 mut self) -> &'l0 mut *mut MdiSubWindow

C++ method: QMdiSubWindow*& QList<QMdiSubWindow*>::last()

C++ documentation:

Returns a reference to the last item in the list. The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

See also first() and isEmpty().

fn length(&self) -> c_int

C++ method: int QList<QMdiSubWindow*>::length() const

C++ documentation:

This function is identical to count().

This function was introduced in Qt 4.5.

See also count().

fn mid<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListMdiSubWindowMutPtrMidArgs<'largs>

C++ method: QList<QMdiSubWindow*>::mid

This is an overloaded function. Available variants:

Rust arguments:
1) fn mid(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListMdiSubWindowMutPtr>
2) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListMdiSubWindowMutPtr>
3) fn mid(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::list::ListMdiSubWindowMutPtr
4) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::list::ListMdiSubWindowMutPtr
C++ method: QList<QMdiSubWindow*> QList<QMdiSubWindow*>::mid(int pos, int length = ?) const

C++ documentation:

Returns a sub-list which includes elements from this list, starting at position pos. If length is -1 (the default), all elements from pos are included; otherwise length elements (or all remaining elements if there are less than length elements) are included.

fn move_(&mut self, from: c_int, to: c_int)

C++ method: void QList<QMdiSubWindow*>::move(int from, int to)

C++ documentation:

Moves the item at index position from to index position to.

Example:

QList<QString> list;
list << "A" << "B" << "C" << "D" << "E" << "F";
list.move(1, 4);
// list: ["A", "C", "D", "E", "B", "F"]

This is the same as insert(to, takeAt(from)).This function assumes that both from and to are at least 0 but less than size(). To avoid failure, test that both from and to are at least 0 and less than size().

See also swap(), insert(), and takeAt().

fn new<Args>(args: Args) -> Args::ReturnType where Args: ListMdiSubWindowMutPtrNewArgs

C++ method: QList<QMdiSubWindow*>::QList

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::list::ListMdiSubWindowMutPtr
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListMdiSubWindowMutPtr>
C++ method: [constructor] void QList<QMdiSubWindow*>::QList()

Variant 2

Rust arguments:
1) fn new((&::list::ListMdiSubWindowMutPtr, cpp_utils::AsStruct)) -> ::list::ListMdiSubWindowMutPtr
2) fn new((&::list::ListMdiSubWindowMutPtr, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListMdiSubWindowMutPtr>
C++ method: [constructor] void QList<QMdiSubWindow*>::QList(const QList<QMdiSubWindow*>& l)

fn op_add<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListMdiSubWindowMutPtrOpAddArgs<'largs>

C++ method: QList<QMdiSubWindow*>::operator+

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_add(&self, (&::list::ListMdiSubWindowMutPtr, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListMdiSubWindowMutPtr>
2) fn op_add(&self, (&::list::ListMdiSubWindowMutPtr, cpp_utils::AsStruct)) -> ::list::ListMdiSubWindowMutPtr
C++ method: QList<QMdiSubWindow*> QList<QMdiSubWindow*>::operator+(const QList<QMdiSubWindow*>& l) const

C++ documentation:

Returns a list that contains all the items in this list followed by all the items in the other list.

See also operator+=().

fn op_add_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut ListMdiSubWindowMutPtr where Args: ListMdiSubWindowMutPtrOpAddAssignArgs<'largs>

C++ method: QList<QMdiSubWindow*>::operator+=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_add_assign(&mut self, &'l1 *mut ::mdi_sub_window::MdiSubWindow) -> &'l0 mut ::list::ListMdiSubWindowMutPtr
C++ method: QList<QMdiSubWindow*>& QList<QMdiSubWindow*>::operator+=(QMdiSubWindow* const & t)

C++ documentation:

This is an overloaded function.

Appends value to the list.

See also append() and operator<<().

Variant 2

Rust arguments: fn op_add_assign(&mut self, &'l1 ::list::ListMdiSubWindowMutPtr) -> &'l0 mut ::list::ListMdiSubWindowMutPtr
C++ method: QList<QMdiSubWindow*>& QList<QMdiSubWindow*>::operator+=(const QList<QMdiSubWindow*>& l)

C++ documentation:

Appends the items of the other list to this list and returns a reference to this list.

See also operator+() and append().

fn op_assign<'l0, 'l1>(&'l0 mut self, l: &'l1 ListMdiSubWindowMutPtr) -> &'l0 mut ListMdiSubWindowMutPtr

C++ method: QList<QMdiSubWindow*>& QList<QMdiSubWindow*>::operator=(const QList<QMdiSubWindow*>& l)

C++ documentation:

Assigns other to this list and returns a reference to this list.

fn op_eq(&self, l: &ListMdiSubWindowMutPtr) -> bool

C++ method: bool QList<QMdiSubWindow*>::operator==(const QList<QMdiSubWindow*>& l) const

C++ documentation:

Returns true if other is equal to this list; otherwise returns false.

Two lists are considered equal if they contain the same values in the same order.

This function requires the value type to have an implementation of operator==().

See also operator!=().

fn op_index<'l0>(&'l0 self, i: c_int) -> &'l0 *mut MdiSubWindow

C++ method: QMdiSubWindow* const & QList<QMdiSubWindow*>::operator[](int i) const

C++ documentation:

This is an overloaded function.

Same as at(). This function runs in constant time.

fn op_index_mut<'l0>(&'l0 mut self, i: c_int) -> &'l0 mut *mut MdiSubWindow

C++ method: QMdiSubWindow*& QList<QMdiSubWindow*>::operator[](int i)

C++ documentation:

Returns the item at index position i as a modifiable reference. i must be a valid index position in the list (i.e., 0 <= i < size()).

If this function is called on a list that is currently being shared, it will trigger a copy of all elements. Otherwise, this function runs in constant time. If you do not want to modify the list you should use QList::at().

See also at() and value().

fn op_neq(&self, l: &ListMdiSubWindowMutPtr) -> bool

C++ method: bool QList<QMdiSubWindow*>::operator!=(const QList<QMdiSubWindow*>& l) const

C++ documentation:

Returns true if other is not equal to this list; otherwise returns false.

Two lists are considered equal if they contain the same values in the same order.

This function requires the value type to have an implementation of operator==().

See also operator==().

fn op_shl<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut ListMdiSubWindowMutPtr where Args: ListMdiSubWindowMutPtrOpShlArgs<'largs>

C++ method: QList<QMdiSubWindow*>::operator<<

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_shl(&mut self, &'l1 *mut ::mdi_sub_window::MdiSubWindow) -> &'l0 mut ::list::ListMdiSubWindowMutPtr
C++ method: QList<QMdiSubWindow*>& QList<QMdiSubWindow*>::operator<<(QMdiSubWindow* const & t)

C++ documentation:

This is an overloaded function.

Appends value to the list.

Variant 2

Rust arguments: fn op_shl(&mut self, &'l1 ::list::ListMdiSubWindowMutPtr) -> &'l0 mut ::list::ListMdiSubWindowMutPtr
C++ method: QList<QMdiSubWindow*>& QList<QMdiSubWindow*>::operator<<(const QList<QMdiSubWindow*>& l)

C++ documentation:

Appends the items of the other list to this list and returns a reference to this list.

See also operator+=() and append().

fn pop_back(&mut self)

C++ method: void QList<QMdiSubWindow*>::pop_back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeLast(). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

fn pop_front(&mut self)

C++ method: void QList<QMdiSubWindow*>::pop_front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeFirst(). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

fn prepend(&mut self, t: &*mut MdiSubWindow)

C++ method: void QList<QMdiSubWindow*>::prepend(QMdiSubWindow* const & t)

C++ documentation:

Inserts value at the beginning of the list.

Example:

QList<QString> list;
list.prepend("one");
list.prepend("two");
list.prepend("three");
// list: ["three", "two", "one"]

This is the same as list.insert(0, value).

If this list is not shared, this operation is typically very fast (amortized constant time), because QList preallocates extra space on both sides of its internal buffer to allow for fast growth at both ends of the list.

See also append() and insert().

fn push_back(&mut self, t: &*mut MdiSubWindow)

C++ method: void QList<QMdiSubWindow*>::push_back(QMdiSubWindow* const & t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to append(value).

fn push_front(&mut self, t: &*mut MdiSubWindow)

C++ method: void QList<QMdiSubWindow*>::push_front(QMdiSubWindow* const & t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to prepend(value).

fn remove_all(&mut self, t: &*mut MdiSubWindow) -> c_int

C++ method: int QList<QMdiSubWindow*>::removeAll(QMdiSubWindow* const & t)

C++ documentation:

Removes all occurrences of value in the list and returns the number of entries removed.

Example:

QList<QString> list;
list << "sun" << "cloud" << "sun" << "rain";
list.removeAll("sun");
// list: ["cloud", "rain"]

This function requires the value type to have an implementation of operator==().

See also removeOne(), removeAt(), takeAt(), and replace().

fn remove_at(&mut self, i: c_int)

C++ method: void QList<QMdiSubWindow*>::removeAt(int i)

C++ documentation:

Removes the item at index position i. i must be a valid index position in the list (i.e., 0 <= i < size()).

See also takeAt(), removeFirst(), removeLast(), and removeOne().

fn remove_first(&mut self)

C++ method: void QList<QMdiSubWindow*>::removeFirst()

C++ documentation:

Removes the first item in the list. Calling this function is equivalent to calling removeAt(0). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

See also removeAt() and takeFirst().

fn remove_last(&mut self)

C++ method: void QList<QMdiSubWindow*>::removeLast()

C++ documentation:

Removes the last item in the list. Calling this function is equivalent to calling removeAt(size() - 1). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

See also removeAt() and takeLast().

fn remove_one(&mut self, t: &*mut MdiSubWindow) -> bool

C++ method: bool QList<QMdiSubWindow*>::removeOne(QMdiSubWindow* const & t)

C++ documentation:

Removes the first occurrence of value in the list and returns true on success; otherwise returns false.

Example:

QList<QString> list;
list << "sun" << "cloud" << "sun" << "rain";
list.removeOne("sun");
// list: ["cloud", ,"sun", "rain"]

This function requires the value type to have an implementation of operator==().

This function was introduced in Qt 4.4.

See also removeAll(), removeAt(), takeAt(), and replace().

fn replace(&mut self, i: c_int, t: &*mut MdiSubWindow)

C++ method: void QList<QMdiSubWindow*>::replace(int i, QMdiSubWindow* const & t)

C++ documentation:

Replaces the item at index position i with value. i must be a valid index position in the list (i.e., 0 <= i < size()).

See also operator[]() and removeAt().

fn reserve(&mut self, size: c_int)

C++ method: void QList<QMdiSubWindow*>::reserve(int size)

C++ documentation:

Reserve space for alloc elements.

If alloc is smaller than the current size of the list, nothing will happen.

Use this function to avoid repetetive reallocation of QList's internal data if you can predict how many elements will be appended. Note that the reservation applies only to the internal pointer array.

This function was introduced in Qt 4.7.

fn set_sharable(&mut self, sharable: bool)

C++ method: void QList<QMdiSubWindow*>::setSharable(bool sharable)

fn size(&self) -> c_int

C++ method: int QList<QMdiSubWindow*>::size() const

C++ documentation:

Returns the number of items in the list.

See also isEmpty() and count().

fn starts_with(&self, t: &*mut MdiSubWindow) -> bool

C++ method: bool QList<QMdiSubWindow*>::startsWith(QMdiSubWindow* const & t) const

C++ documentation:

Returns true if this list is not empty and its first item is equal to value; otherwise returns false.

This function was introduced in Qt 4.5.

See also isEmpty() and contains().

fn swap<'largs, Args>(&'largs mut self, args: Args) where Args: ListMdiSubWindowMutPtrSwapArgs<'largs>

C++ method: QList<QMdiSubWindow*>::swap

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn swap(&mut self, &mut ::list::ListMdiSubWindowMutPtr) -> ()
C++ method: void QList<QMdiSubWindow*>::swap(QList<QMdiSubWindow*>& other)

C++ documentation:

Swaps list other with this list. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

Variant 2

Rust arguments: fn swap(&mut self, (libc::c_int, libc::c_int)) -> ()
C++ method: void QList<QMdiSubWindow*>::swap(int i, int j)

C++ documentation:

Exchange the item at index position i with the item at index position j. This function assumes that both i and j are at least 0 but less than size(). To avoid failure, test that both i and j are at least 0 and less than size().

Example:

QList<QString> list;
list << "A" << "B" << "C" << "D" << "E" << "F";
list.swap(1, 4);
// list: ["A", "E", "C", "D", "B", "F"]

See also move().

fn take_at(&mut self, i: c_int) -> *mut MdiSubWindow

C++ method: QMdiSubWindow* QList<QMdiSubWindow*>::takeAt(int i)

C++ documentation:

Removes the item at index position i and returns it. i must be a valid index position in the list (i.e., 0 <= i < size()).

If you don't use the return value, removeAt() is more efficient.

See also removeAt(), takeFirst(), and takeLast().

fn take_first(&mut self) -> *mut MdiSubWindow

C++ method: QMdiSubWindow* QList<QMdiSubWindow*>::takeFirst()

C++ documentation:

Removes the first item in the list and returns it. This is the same as takeAt(0). This function assumes the list is not empty. To avoid failure, call isEmpty() before calling this function.

If this list is not shared, this operation takes constant time.

If you don't use the return value, removeFirst() is more efficient.

See also takeLast(), takeAt(), and removeFirst().

fn take_last(&mut self) -> *mut MdiSubWindow

C++ method: QMdiSubWindow* QList<QMdiSubWindow*>::takeLast()

C++ documentation:

Removes the last item in the list and returns it. This is the same as takeAt(size() - 1). This function assumes the list is not empty. To avoid failure, call isEmpty() before calling this function.

If this list is not shared, this operation takes constant time.

If you don't use the return value, removeLast() is more efficient.

See also takeFirst(), takeAt(), and removeLast().

fn value<'largs, Args>(&'largs self, args: Args) -> *mut MdiSubWindow where Args: ListMdiSubWindowMutPtrValueArgs<'largs>

C++ method: QList<QMdiSubWindow*>::value

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn value(&self, libc::c_int) -> *mut ::mdi_sub_window::MdiSubWindow
C++ method: QMdiSubWindow* QList<QMdiSubWindow*>::value(int i) const

C++ documentation:

Returns the value at index position i in the list.

If the index i is out of bounds, the function returns a default-constructed value. If you are certain that the index is going to be within bounds, you can use at() instead, which is slightly faster.

See also at() and operator[]().

Variant 2

Rust arguments: fn value(&self, (libc::c_int, &*mut ::mdi_sub_window::MdiSubWindow)) -> *mut ::mdi_sub_window::MdiSubWindow
C++ method: QMdiSubWindow* QList<QMdiSubWindow*>::value(int i, QMdiSubWindow* const & defaultValue) const

C++ documentation:

This is an overloaded function.

If the index i is out of bounds, the function returns defaultValue.