Module qt_widgets::list_widget_item::overloading

Traits

ListWidgetItemBackgroundArgs
This trait represents a set of arguments accepted by ListWidgetItem::background method.

ListWidgetItemBackgroundColorArgs
This trait represents a set of arguments accepted by ListWidgetItem::background_color method.

ListWidgetItemDataArgs
This trait represents a set of arguments accepted by ListWidgetItem::data method.

ListWidgetItemFontArgs
This trait represents a set of arguments accepted by ListWidgetItem::font method.

ListWidgetItemForegroundArgs
This trait represents a set of arguments accepted by ListWidgetItem::foreground method.

ListWidgetItemIconArgs
This trait represents a set of arguments accepted by ListWidgetItem::icon method.

ListWidgetItemNewArgs
This trait represents a set of arguments accepted by ListWidgetItem::new method.

ListWidgetItemSizeHintArgs
This trait represents a set of arguments accepted by ListWidgetItem::size_hint method.

ListWidgetItemStatusTipArgs
This trait represents a set of arguments accepted by ListWidgetItem::status_tip method.

ListWidgetItemTextArgs
This trait represents a set of arguments accepted by ListWidgetItem::text method.

ListWidgetItemTextColorArgs
This trait represents a set of arguments accepted by ListWidgetItem::text_color method.

ListWidgetItemToolTipArgs
This trait represents a set of arguments accepted by ListWidgetItem::tool_tip method.

ListWidgetItemWhatsThisArgs
This trait represents a set of arguments accepted by ListWidgetItem::whats_this method.