Struct qt_widgets::list_widget_item::ListWidgetItem

pub struct ListWidgetItem {
    // some fields omitted
}

C++ type: QListWidgetItem

C++ documentation:

The QListWidgetItem class provides an item for use with the QListWidget item view class.

A QListWidgetItem represents a single item in a QListWidget. Each item can hold several pieces of information, and will display them appropriately.

The item view convenience classes use a classic item-based interface rather than a pure model/view approach. For a more flexible list view widget, consider using the QListView class with a standard model.

List items can be inserted automatically into a list, when they are constructed, by specifying the list widget:

    new QListWidgetItem(tr("Hazel"), listWidget);

Alternatively, list items can also be created without a parent widget, and later inserted into a list using QListWidget::insertItem().

List items are typically used to display text() and an icon(). These are set with the setText() and setIcon() functions. The appearance of the text can be customized with setFont(), setForeground(), and setBackground(). Text in list items can be aligned using the setTextAlignment() function. Tooltips, status tips and "What's This?" help can be added to list items with setToolTip(), setStatusTip(), and setWhatsThis().

By default, items are enabled, selectable, checkable, and can be the source of drag and drop operations.

Each item's flags can be changed by calling setFlags() with the appropriate value (see Qt::ItemFlags). Checkable items can be checked, unchecked and partially checked with the setCheckState() function. The corresponding checkState() function indicates the item's current check state.

The isHidden() function can be used to determine whether the item is hidden. To hide an item, use setHidden().

Methods

impl ListWidgetItem

fn background<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListWidgetItemBackgroundArgs<'largs>

C++ method: QListWidgetItem::background

This is an overloaded function. Available variants:

Rust arguments:
1) fn background(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::brush::Brush>
2) fn background(&self, cpp_utils::AsStruct) -> qt_gui::brush::Brush
C++ method: QBrush QListWidgetItem::background() const

C++ documentation:

Returns the brush used to display the list item's background.

This function was introduced in Qt 4.2.

See also setBackground() and foreground().

fn background_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListWidgetItemBackgroundColorArgs<'largs>

C++ method: QListWidgetItem::backgroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn background_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn background_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: QColor QListWidgetItem::backgroundColor() const

C++ documentation:

This function is deprecated. Use background() instead.

See also setBackgroundColor().

fn check_state(&self) -> CheckState

C++ method: Qt::CheckState QListWidgetItem::checkState() const

C++ documentation:

Returns the checked state of the list item (see Qt::CheckState).

See also setCheckState() and flags().

fn clone(&self) -> *mut ListWidgetItem

C++ method: virtual QListWidgetItem* QListWidgetItem::clone() const

C++ documentation:

Creates an exact copy of the item.

fn data<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListWidgetItemDataArgs<'largs>

C++ method: QListWidgetItem::data

This is an overloaded function. Available variants:

Rust arguments:
1) fn data(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn data(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::variant::Variant
C++ method: virtual QVariant QListWidgetItem::data(int role) const

C++ documentation:

Returns the item's data for a given role. Reimplement this function if you need extra roles or special behavior for certain roles.

See also Qt::ItemDataRole and setData().

fn flags(&self) -> Flags<ItemFlag>

C++ method: QFlags<Qt::ItemFlag> QListWidgetItem::flags() const

C++ documentation:

Returns the item flags for this item (see Qt::ItemFlags).

See also setFlags().

fn font<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListWidgetItemFontArgs<'largs>

C++ method: QListWidgetItem::font

This is an overloaded function. Available variants:

Rust arguments:
1) fn font(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::font::Font>
2) fn font(&self, cpp_utils::AsStruct) -> qt_gui::font::Font
C++ method: QFont QListWidgetItem::font() const

C++ documentation:

Returns the font used to display this list item's text.

See also setFont().

fn foreground<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListWidgetItemForegroundArgs<'largs>

C++ method: QListWidgetItem::foreground

This is an overloaded function. Available variants:

Rust arguments:
1) fn foreground(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::brush::Brush>
2) fn foreground(&self, cpp_utils::AsStruct) -> qt_gui::brush::Brush
C++ method: QBrush QListWidgetItem::foreground() const

C++ documentation:

Returns the brush used to display the list item's foreground (e.g. text).

This function was introduced in Qt 4.2.

See also setForeground() and background().

fn icon<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListWidgetItemIconArgs<'largs>

C++ method: QListWidgetItem::icon

This is an overloaded function. Available variants:

Rust arguments:
1) fn icon(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::icon::Icon>
2) fn icon(&self, cpp_utils::AsStruct) -> qt_gui::icon::Icon
C++ method: QIcon QListWidgetItem::icon() const

C++ documentation:

Returns the list item's icon.

See also setIcon() and iconSize.

fn is_hidden(&self) -> bool

C++ method: bool QListWidgetItem::isHidden() const

C++ documentation:

Returns true if the item is hidden; otherwise returns false.

This function was introduced in Qt 4.2.

See also setHidden().

fn is_selected(&self) -> bool

C++ method: bool QListWidgetItem::isSelected() const

C++ documentation:

Returns true if the item is selected; otherwise returns false.

This function was introduced in Qt 4.2.

See also setSelected().

fn list_widget(&self) -> *mut ListWidget

C++ method: QListWidget* QListWidgetItem::listWidget() const

C++ documentation:

Returns the list widget containing the item.

fn new<Args>(args: Args) -> Args::ReturnType where Args: ListWidgetItemNewArgs

C++ method: QListWidgetItem::QListWidgetItem

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::list_widget_item::ListWidgetItem
2) fn new((*mut ::list_widget::ListWidget, cpp_utils::AsStruct)) -> ::list_widget_item::ListWidgetItem
3) fn new((*mut ::list_widget::ListWidget, libc::c_int, cpp_utils::AsStruct)) -> ::list_widget_item::ListWidgetItem
4) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::list_widget_item::ListWidgetItem>
5) fn new((*mut ::list_widget::ListWidget, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list_widget_item::ListWidgetItem>
6) fn new((*mut ::list_widget::ListWidget, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list_widget_item::ListWidgetItem>
C++ method: [constructor] void QListWidgetItem::QListWidgetItem(QListWidget* view = ?, int type = ?)

Variant 2

Rust arguments:
1) fn new((&qt_gui::icon::Icon, &qt_core::string::String, cpp_utils::AsStruct)) -> ::list_widget_item::ListWidgetItem
2) fn new((&qt_gui::icon::Icon, &qt_core::string::String, *mut ::list_widget::ListWidget, cpp_utils::AsStruct)) -> ::list_widget_item::ListWidgetItem
3) fn new((&qt_gui::icon::Icon, &qt_core::string::String, *mut ::list_widget::ListWidget, libc::c_int, cpp_utils::AsStruct)) -> ::list_widget_item::ListWidgetItem
4) fn new((&qt_gui::icon::Icon, &qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list_widget_item::ListWidgetItem>
5) fn new((&qt_gui::icon::Icon, &qt_core::string::String, *mut ::list_widget::ListWidget, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list_widget_item::ListWidgetItem>
6) fn new((&qt_gui::icon::Icon, &qt_core::string::String, *mut ::list_widget::ListWidget, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list_widget_item::ListWidgetItem>
C++ method: [constructor] void QListWidgetItem::QListWidgetItem(const QIcon& icon, const QString& text, QListWidget* view = ?, int type = ?)

Variant 3

Rust arguments:
1) fn new((&::list_widget_item::ListWidgetItem, cpp_utils::AsStruct)) -> ::list_widget_item::ListWidgetItem
2) fn new((&::list_widget_item::ListWidgetItem, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list_widget_item::ListWidgetItem>
C++ method: [constructor] void QListWidgetItem::QListWidgetItem(const QListWidgetItem& other)

Variant 4

Rust arguments:
1) fn new((&qt_core::string::String, cpp_utils::AsStruct)) -> ::list_widget_item::ListWidgetItem
2) fn new((&qt_core::string::String, *mut ::list_widget::ListWidget, cpp_utils::AsStruct)) -> ::list_widget_item::ListWidgetItem
3) fn new((&qt_core::string::String, *mut ::list_widget::ListWidget, libc::c_int, cpp_utils::AsStruct)) -> ::list_widget_item::ListWidgetItem
4) fn new((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list_widget_item::ListWidgetItem>
5) fn new((&qt_core::string::String, *mut ::list_widget::ListWidget, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list_widget_item::ListWidgetItem>
6) fn new((&qt_core::string::String, *mut ::list_widget::ListWidget, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list_widget_item::ListWidgetItem>
C++ method: [constructor] void QListWidgetItem::QListWidgetItem(const QString& text, QListWidget* view = ?, int type = ?)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 ListWidgetItem) -> &'l0 mut ListWidgetItem

C++ method: QListWidgetItem& QListWidgetItem::operator=(const QListWidgetItem& other)

C++ documentation:

Assigns other's data and flags to this item. Note that type() and listWidget() are not copied.

This function is useful when reimplementing clone().

See also data() and flags().

fn op_lt(&self, other: &ListWidgetItem) -> bool

C++ method: virtual bool QListWidgetItem::operator<(const QListWidgetItem& other) const

C++ documentation:

Returns true if this item's text is less then other item's text; otherwise returns false.

fn read(&mut self, in_: &mut DataStream)

C++ method: virtual void QListWidgetItem::read(QDataStream& in)

C++ documentation:

Reads the item from stream in.

See also write().

fn set_background(&mut self, brush: &Brush)

C++ method: void QListWidgetItem::setBackground(const QBrush& brush)

C++ documentation:

Sets the background brush of the list item to the given brush.

This function was introduced in Qt 4.2.

See also background() and setForeground().

fn set_background_color(&mut self, color: &Color)

C++ method: virtual void QListWidgetItem::setBackgroundColor(const QColor& color)

C++ documentation:

This function is deprecated. Use setBackground() instead.

See also backgroundColor().

fn set_check_state(&mut self, state: CheckState)

C++ method: void QListWidgetItem::setCheckState(Qt::CheckState state)

C++ documentation:

Sets the check state of the list item to state.

See also checkState().

fn set_data(&mut self, role: c_int, value: &Variant)

C++ method: virtual void QListWidgetItem::setData(int role, const QVariant& value)

C++ documentation:

Sets the data for a given role to the given value. Reimplement this function if you need extra roles or special behavior for certain roles.

See also Qt::ItemDataRole and data().

fn set_flags(&mut self, flags: Flags<ItemFlag>)

C++ method: void QListWidgetItem::setFlags(QFlags<Qt::ItemFlag> flags)

C++ documentation:

Sets the item flags for the list item to flags.

See also flags() and Qt::ItemFlags.

fn set_font(&mut self, font: &Font)

C++ method: void QListWidgetItem::setFont(const QFont& font)

C++ documentation:

Sets the font used when painting the item to the given font.

See also font().

fn set_foreground(&mut self, brush: &Brush)

C++ method: void QListWidgetItem::setForeground(const QBrush& brush)

C++ documentation:

Sets the foreground brush of the list item to the given brush.

This function was introduced in Qt 4.2.

See also foreground() and setBackground().

fn set_hidden(&mut self, hide: bool)

C++ method: void QListWidgetItem::setHidden(bool hide)

C++ documentation:

Hides the item if hide is true; otherwise shows the item.

This function was introduced in Qt 4.2.

See also isHidden().

fn set_icon(&mut self, icon: &Icon)

C++ method: void QListWidgetItem::setIcon(const QIcon& icon)

C++ documentation:

Sets the icon for the list item to the given icon.

See also icon(), text(), and iconSize.

fn set_selected(&mut self, select: bool)

C++ method: void QListWidgetItem::setSelected(bool select)

C++ documentation:

Sets the selected state of the item to select.

This function was introduced in Qt 4.2.

See also isSelected().

fn set_size_hint(&mut self, size: &Size)

C++ method: void QListWidgetItem::setSizeHint(const QSize& size)

C++ documentation:

Sets the size hint for the list item to be size. If no size hint is set, the item delegate will compute the size hint based on the item data.

This function was introduced in Qt 4.1.

See also sizeHint().

fn set_status_tip(&mut self, status_tip: &String)

C++ method: void QListWidgetItem::setStatusTip(const QString& statusTip)

C++ documentation:

Sets the status tip for the list item to the text specified by statusTip. QListWidget mouseTracking needs to be enabled for this feature to work.

See also statusTip(), setToolTip(), setWhatsThis(), and QWidget::setMouseTracking().

fn set_text(&mut self, text: &String)

C++ method: void QListWidgetItem::setText(const QString& text)

C++ documentation:

Sets the text for the list widget item's to the given text.

See also text().

fn set_text_alignment(&mut self, alignment: c_int)

C++ method: void QListWidgetItem::setTextAlignment(int alignment)

C++ documentation:

Sets the list item's text alignment to alignment.

See also textAlignment() and Qt::AlignmentFlag.

fn set_text_color(&mut self, color: &Color)

C++ method: void QListWidgetItem::setTextColor(const QColor& color)

C++ documentation:

This function is deprecated. Use setForeground() instead.

See also textColor().

fn set_tool_tip(&mut self, tool_tip: &String)

C++ method: void QListWidgetItem::setToolTip(const QString& toolTip)

C++ documentation:

Sets the tooltip for the list item to the text specified by toolTip.

See also toolTip(), setStatusTip(), and setWhatsThis().

fn set_whats_this(&mut self, whats_this: &String)

C++ method: void QListWidgetItem::setWhatsThis(const QString& whatsThis)

C++ documentation:

Sets the "What's This?" help for the list item to the text specified by whatsThis.

See also whatsThis(), setStatusTip(), and setToolTip().

fn size_hint<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListWidgetItemSizeHintArgs<'largs>

C++ method: QListWidgetItem::sizeHint

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_hint(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size_hint(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: QSize QListWidgetItem::sizeHint() const

C++ documentation:

Returns the size hint set for the list item.

This function was introduced in Qt 4.1.

See also setSizeHint().

fn status_tip<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListWidgetItemStatusTipArgs<'largs>

C++ method: QListWidgetItem::statusTip

This is an overloaded function. Available variants:

Rust arguments:
1) fn status_tip(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn status_tip(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QListWidgetItem::statusTip() const

C++ documentation:

Returns the list item's status tip.

See also setStatusTip().

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListWidgetItemTextArgs<'largs>

C++ method: QListWidgetItem::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QListWidgetItem::text() const

C++ documentation:

Returns the list item's text.

See also setText().

fn text_alignment(&self) -> c_int

C++ method: int QListWidgetItem::textAlignment() const

C++ documentation:

Returns the text alignment for the list item.

See also setTextAlignment() and Qt::AlignmentFlag.

fn text_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListWidgetItemTextColorArgs<'largs>

C++ method: QListWidgetItem::textColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn text_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: QColor QListWidgetItem::textColor() const

C++ documentation:

Returns the color used to display the list item's text.

This function is deprecated. Use foreground() instead.

See also setTextColor().

fn tool_tip<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListWidgetItemToolTipArgs<'largs>

C++ method: QListWidgetItem::toolTip

This is an overloaded function. Available variants:

Rust arguments:
1) fn tool_tip(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tool_tip(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QListWidgetItem::toolTip() const

C++ documentation:

Returns the list item's tooltip.

See also setToolTip(), statusTip(), and whatsThis().

fn type_(&self) -> c_int

C++ method: int QListWidgetItem::type() const

C++ documentation:

Returns the type passed to the QListWidgetItem constructor.

fn whats_this<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListWidgetItemWhatsThisArgs<'largs>

C++ method: QListWidgetItem::whatsThis

This is an overloaded function. Available variants:

Rust arguments:
1) fn whats_this(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn whats_this(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QListWidgetItem::whatsThis() const

C++ documentation:

Returns the list item's "What's This?" help text.

See also setWhatsThis(), statusTip(), and toolTip().

fn write(&self, out: &mut DataStream)

C++ method: virtual void QListWidgetItem::write(QDataStream& out) const

C++ documentation:

Writes the item to stream out.

See also read().