Enum qt_widgets::main_window::DockOption

pub enum DockOption {
    AnimatedDocks,
    AllowNestedDocks,
    AllowTabbedDocks,
    ForceTabbedDocks,
    VerticalTabs,
}

C++ type: QMainWindow::DockOption

C++ documentation:

This enum contains flags that specify the docking behavior of QMainWindow.

These options only control how dock widgets may be dropped in a QMainWindow. They do not re-arrange the dock widgets to conform with the specified options. For this reason they should be set before any dock widgets are added to the main window. Exceptions to this are the AnimatedDocks and VerticalTabs options, which may be set at any time.

This enum was introduced or modified in Qt 4.3.

The DockOptions type is a typedef for QFlags<DockOption>. It stores an OR combination of DockOption values.

Variants
AnimatedDocks

Identical to the animated property. (C++ enum variant: AnimatedDocks = 1)

AllowNestedDocks

Identical to the dockNestingEnabled property. (C++ enum variant: AllowNestedDocks = 2)

AllowTabbedDocks

The user can drop one dock widget "on top" of another. The two widgets are stacked and a tab bar appears for selecting which one is visible. (C++ enum variant: AllowTabbedDocks = 4)

ForceTabbedDocks

Each dock area contains a single stack of tabbed dock widgets. In other words, dock widgets cannot be placed next to each other in a dock area. If this option is set, AllowNestedDocks has no effect. (C++ enum variant: ForceTabbedDocks = 8)

VerticalTabs

The two vertical dock areas on the sides of the main window show their tabs vertically. If this option is not set, all dock areas show their tabs at the bottom. Implies AllowTabbedDocks. See also setTabPosition(). (C++ enum variant: VerticalTabs = 16)