Struct qt_widgets::map::MapGestureTypeWidgetMutPtr

pub struct MapGestureTypeWidgetMutPtr {
    // some fields omitted
}

C++ type: QMap<Qt::GestureType, QWidget*>

C++ documentation:

The QMap class is a template class that provides a red-black-tree-based dictionary.

QMap<Key, T> is one of Qt's generic container classes. It stores (key, value) pairs and provides fast lookup of the value associated with a key.

QMap and QHash provide very similar functionality. The differences are:

QHash provides average faster lookups than QMap. (See Algorithmic Complexity for details.)

When iterating over a QHash, the items are arbitrarily ordered. With QMap, the items are always sorted by key.

The key type of a QHash must provide operator==() and a global qHash(Key) function. The key type of a QMap must provide operator<() specifying a total order.

Here's an example QMap with QString keys and int values:

QMap<QString, int> map;

To insert a (key, value) pair into the map, you can use operator:

map["one"] = 1;
map["three"] = 3;
map["seven"] = 7;

This inserts the following three (key, value) pairs into the QMap: ("one", 1), ("three", 3), and ("seven", 7). Another way to insert items into the map is to use insert():

map.insert("twelve", 12);

To look up a value, use operator or value():

int num1 = map["thirteen"];
int num2 = map.value("thirteen");

If there is no item with the specified key in the map, these functions return a default-constructed value.

If you want to check whether the map contains a certain key, use contains():

int timeout = 30;
if (map.contains("TIMEOUT"))
    timeout = map.value("TIMEOUT");

There is also a value() overload that uses its second argument as a default value if there is no item with the specified key:

int timeout = map.value("TIMEOUT", 30);

In general, we recommend that you use contains() and value() rather than operator for looking up a key in a map. The reason is that operator silently inserts an item into the map if no item exists with the same key (unless the map is const). For example, the following code snippet will create 1000 items in memory:

// WRONG
QMap<int, QWidget *> map;
...
for (int i = 0; i < 1000; ++i) {
    if (map[i] == okButton)
        cout << "Found button at index " << i << endl;
}

To avoid this problem, replace map[i] with map.value(i) in the code above.

If you want to navigate through all the (key, value) pairs stored in a QMap, you can use an iterator. QMap provides both Java-style iterators (QMapIterator and QMutableMapIterator) and STL-style iterators (QMap::const_iterator and QMap::iterator). Here's how to iterate over a QMap<QString, int> using a Java-style iterator:

QMapIterator<QString, int> i(map);
while (i.hasNext()) {
    i.next();
    cout << i.key() << ": " << i.value() << endl;
}

Here's the same code, but using an STL-style iterator this time:

QMap<QString, int>::const_iterator i = map.constBegin();
while (i != map.constEnd()) {
    cout << i.key() << ": " << i.value() << endl;
    ++i;
}

The items are traversed in ascending key order.

Normally, a QMap allows only one value per key. If you call insert() with a key that already exists in the QMap, the previous value will be erased. For example:

map.insert("plenty", 100);
map.insert("plenty", 2000);
// map.value("plenty") == 2000

However, you can store multiple values per key by using insertMulti() instead of insert() (or using the convenience subclass QMultiMap). If you want to retrieve all the values for a single key, you can use values(const Key &key), which returns a QList<T>:

QList<int> values = map.values("plenty");
for (int i = 0; i < values.size(); ++i)
    cout << values.at(i) << endl;

The items that share the same key are available from most recently to least recently inserted. Another approach is to call find() to get the STL-style iterator for the first item with a key and iterate from there:

QMap<QString, int>::iterator i = map.find("plenty");
while (i != map.end() && i.key() == "plenty") {
    cout << i.value() << endl;
    ++i;
}

If you only need to extract the values from a map (not the keys), you can also use foreach:

QMap<QString, int> map;
...
foreach (int value, map)
    cout << value << endl;

Items can be removed from the map in several ways. One way is to call remove(); this will remove any item with the given key. Another way is to use QMutableMapIterator::remove(). In addition, you can clear the entire map using clear().

QMap's key and value data types must be assignable data types. This covers most data types you are likely to encounter, but the compiler won't let you, for example, store a QWidget as a value; instead, store a QWidget *. In addition, QMap's key type must provide operator<(). QMap uses it to keep its items sorted, and assumes that two keys x and y are equal if neither x < y nor y < x is true.

Example:

#ifndef EMPLOYEE_H
#define EMPLOYEE_H

class Employee { public: Employee() {} Employee(const QString &name, const QDate &dateOfBirth); ...

private: QString myName; QDate myDateOfBirth; };

inline bool operator<(const Employee &e1, const Employee &e2) { if (e1.name() != e2.name()) return e1.name() < e2.name(); return e1.dateOfBirth() < e2.dateOfBirth(); }

#endif // EMPLOYEE_H
In the example, we start by comparing the employees' names. If they're equal, we compare their dates of birth to break the tie.

Methods

impl MapGestureTypeWidgetMutPtr

fn clear(&mut self)

C++ method: void QMap<Qt::GestureType, QWidget*>::clear()

C++ documentation:

Removes all items from the map.

See also remove().

fn contains(&self, key: &GestureType) -> bool

C++ method: bool QMap<Qt::GestureType, QWidget*>::contains(const Qt::GestureType& key) const

C++ documentation:

Returns true if the map contains an item with key key; otherwise returns false.

See also count() and QMultiMap::contains().

fn count<'largs, Args>(&'largs self, args: Args) -> c_int where Args: MapGestureTypeWidgetMutPtrCountArgs<'largs>

C++ method: QMap<Qt::GestureType, QWidget*>::count

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn count(&self, ()) -> libc::c_int
C++ method: int QMap<Qt::GestureType, QWidget*>::count() const

C++ documentation:

This is an overloaded function.

Same as size().

Variant 2

Rust arguments: fn count(&self, &qt_core::qt::GestureType) -> libc::c_int
C++ method: int QMap<Qt::GestureType, QWidget*>::count(const Qt::GestureType& key) const

C++ documentation:

Returns the number of items associated with key key.

See also contains(), insertMulti(), and QMultiMap::count().

fn detach(&mut self)

C++ method: void QMap<Qt::GestureType, QWidget*>::detach()

fn empty(&self) -> bool

C++ method: bool QMap<Qt::GestureType, QWidget*>::empty() const

C++ documentation:

This function is provided for STL compatibility. It is equivalent to isEmpty(), returning true if the map is empty; otherwise returning false.

fn first<'l0>(&'l0 self) -> &'l0 *mut Widget

C++ method: QWidget* const & QMap<Qt::GestureType, QWidget*>::first() const

C++ documentation:

This is an overloaded function.

This function was introduced in Qt 5.2.

fn first_key<'l0>(&'l0 self) -> &'l0 GestureType

C++ method: const Qt::GestureType& QMap<Qt::GestureType, QWidget*>::firstKey() const

C++ documentation:

Returns a reference to the smallest key in the map. This function assumes that the map is not empty.

This executes in constant time.

This function was introduced in Qt 5.2.

See also lastKey(), first(), and isEmpty().

fn first_mut<'l0>(&'l0 mut self) -> &'l0 mut *mut Widget

C++ method: QWidget*& QMap<Qt::GestureType, QWidget*>::first()

C++ documentation:

Returns a reference to the first value in the map, that is the value mapped to the smallest key. This function assumes that the map is not empty.

When unshared (or const version is called), this executes in constant time.

This function was introduced in Qt 5.2.

See also last(), firstKey(), and isEmpty().

fn is_detached(&self) -> bool

C++ method: bool QMap<Qt::GestureType, QWidget*>::isDetached() const

fn is_empty(&self) -> bool

C++ method: bool QMap<Qt::GestureType, QWidget*>::isEmpty() const

C++ documentation:

Returns true if the map contains no items; otherwise returns false.

See also size().

fn is_shared_with(&self, other: &MapGestureTypeWidgetMutPtr) -> bool

C++ method: bool QMap<Qt::GestureType, QWidget*>::isSharedWith(const QMap<Qt::GestureType, QWidget*>& other) const

fn key<'largs, Args>(&'largs self, args: Args) -> GestureType where Args: MapGestureTypeWidgetMutPtrKeyArgs<'largs>

C++ method: QMap<Qt::GestureType, QWidget*>::key

This is an overloaded function. Available variants:

Rust arguments:
1) fn key(&self, &*mut ::widget::Widget) -> qt_core::qt::GestureType
2) fn key(&self, (&*mut ::widget::Widget, &qt_core::qt::GestureType)) -> qt_core::qt::GestureType
C++ method: const Qt::GestureType QMap<Qt::GestureType, QWidget*>::key(QWidget* const & value, const Qt::GestureType& defaultKey = ?) const

C++ documentation:

This is an overloaded function.

Returns the first key with value value, or defaultKey if the map contains no item with value value. If no defaultKey is provided the function returns a default-constructed key.

This function can be slow (linear time), because QMap's internal data structure is optimized for fast lookup by key, not by value.

This function was introduced in Qt 4.3.

See also value() and keys().

fn last<'l0>(&'l0 self) -> &'l0 *mut Widget

C++ method: QWidget* const & QMap<Qt::GestureType, QWidget*>::last() const

C++ documentation:

This is an overloaded function.

This function was introduced in Qt 5.2.

fn last_key<'l0>(&'l0 self) -> &'l0 GestureType

C++ method: const Qt::GestureType& QMap<Qt::GestureType, QWidget*>::lastKey() const

C++ documentation:

Returns a reference to the largest key in the map. This function assumes that the map is not empty.

This executes in logarithmic time.

This function was introduced in Qt 5.2.

See also firstKey(), last(), and isEmpty().

fn last_mut<'l0>(&'l0 mut self) -> &'l0 mut *mut Widget

C++ method: QWidget*& QMap<Qt::GestureType, QWidget*>::last()

C++ documentation:

Returns a reference to the last value in the map, that is the value mapped to the largest key. This function assumes that the map is not empty.

When unshared (or const version is called), this executes in logarithmic time.

This function was introduced in Qt 5.2.

See also first(), lastKey(), and isEmpty().

fn new<Args>(args: Args) -> Args::ReturnType where Args: MapGestureTypeWidgetMutPtrNewArgs

C++ method: QMap<Qt::GestureType, QWidget*>::QMap

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::map::MapGestureTypeWidgetMutPtr
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::map::MapGestureTypeWidgetMutPtr>
C++ method: [constructor] void QMap<Qt::GestureType, QWidget*>::QMap()

Variant 2

Rust arguments:
1) fn new((&::map::MapGestureTypeWidgetMutPtr, cpp_utils::AsStruct)) -> ::map::MapGestureTypeWidgetMutPtr
2) fn new((&::map::MapGestureTypeWidgetMutPtr, cpp_utils::AsBox)) -> cpp_utils::CppBox<::map::MapGestureTypeWidgetMutPtr>
C++ method: [constructor] void QMap<Qt::GestureType, QWidget*>::QMap(const QMap<Qt::GestureType, QWidget*>& other)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 MapGestureTypeWidgetMutPtr) -> &'l0 mut MapGestureTypeWidgetMutPtr

C++ method: QMap<Qt::GestureType, QWidget*>& QMap<Qt::GestureType, QWidget*>::operator=(const QMap<Qt::GestureType, QWidget*>& other)

C++ documentation:

Assigns other to this map and returns a reference to this map.

fn op_eq(&self, other: &MapGestureTypeWidgetMutPtr) -> bool

C++ method: bool QMap<Qt::GestureType, QWidget*>::operator==(const QMap<Qt::GestureType, QWidget*>& other) const

C++ documentation:

Returns true if other is equal to this map; otherwise returns false.

Two maps are considered equal if they contain the same (key, value) pairs.

This function requires the value type to implement operator==().

See also operator!=().

fn op_index(&self, key: &GestureType) -> *mut Widget

C++ method: QWidget* QMap<Qt::GestureType, QWidget*>::operator[](const Qt::GestureType& key) const

C++ documentation:

This is an overloaded function.

Same as value().

fn op_index_mut<'l0, 'l1>(&'l0 mut self, key: &'l1 GestureType) -> &'l0 mut *mut Widget

C++ method: QWidget*& QMap<Qt::GestureType, QWidget*>::operator[](const Qt::GestureType& key)

C++ documentation:

Returns the value associated with the key key as a modifiable reference.

If the map contains no item with key key, the function inserts a default-constructed value into the map with key key, and returns a reference to it. If the map contains multiple items with key key, this function returns a reference to the most recently inserted value.

See also insert() and value().

fn op_neq(&self, other: &MapGestureTypeWidgetMutPtr) -> bool

C++ method: bool QMap<Qt::GestureType, QWidget*>::operator!=(const QMap<Qt::GestureType, QWidget*>& other) const

C++ documentation:

Returns true if other is not equal to this map; otherwise returns false.

Two maps are considered equal if they contain the same (key, value) pairs.

This function requires the value type to implement operator==().

See also operator==().

fn remove(&mut self, key: &GestureType) -> c_int

C++ method: int QMap<Qt::GestureType, QWidget*>::remove(const Qt::GestureType& key)

C++ documentation:

Removes all the items that have the key key from the map. Returns the number of items removed which is usually 1 but will be 0 if the key isn't in the map, or > 1 if insertMulti() has been used with the key.

See also clear(), take(), and QMultiMap::remove().

fn set_sharable(&mut self, sharable: bool)

C++ method: void QMap<Qt::GestureType, QWidget*>::setSharable(bool sharable)

fn size(&self) -> c_int

C++ method: int QMap<Qt::GestureType, QWidget*>::size() const

C++ documentation:

Returns the number of (key, value) pairs in the map.

See also isEmpty() and count().

fn swap(&mut self, other: &mut MapGestureTypeWidgetMutPtr)

C++ method: void QMap<Qt::GestureType, QWidget*>::swap(QMap<Qt::GestureType, QWidget*>& other)

C++ documentation:

Swaps map other with this map. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

fn take(&mut self, key: &GestureType) -> *mut Widget

C++ method: QWidget* QMap<Qt::GestureType, QWidget*>::take(const Qt::GestureType& key)

C++ documentation:

Removes the item with the key key from the map and returns the value associated with it.

If the item does not exist in the map, the function simply returns a default-constructed value. If there are multiple items for key in the map, only the most recently inserted one is removed and returned.

If you don't use the return value, remove() is more efficient.

See also remove().

fn unite<'l0, 'l1>(&'l0 mut self, other: &'l1 MapGestureTypeWidgetMutPtr) -> &'l0 mut MapGestureTypeWidgetMutPtr

C++ method: QMap<Qt::GestureType, QWidget*>& QMap<Qt::GestureType, QWidget*>::unite(const QMap<Qt::GestureType, QWidget*>& other)

C++ documentation:

Inserts all the items in the other map into this map. If a key is common to both maps, the resulting map will contain the key multiple times.

See also insertMulti().

fn value<'largs, Args>(&'largs self, args: Args) -> *mut Widget where Args: MapGestureTypeWidgetMutPtrValueArgs<'largs>

C++ method: QMap<Qt::GestureType, QWidget*>::value

This is an overloaded function. Available variants:

Rust arguments:
1) fn value(&self, &qt_core::qt::GestureType) -> *mut ::widget::Widget
2) fn value(&self, (&qt_core::qt::GestureType, &*mut ::widget::Widget)) -> *mut ::widget::Widget
C++ method: QWidget* QMap<Qt::GestureType, QWidget*>::value(const Qt::GestureType& key, QWidget* const & defaultValue = ?) const

C++ documentation:

Returns the value associated with the key key.

If the map contains no item with key key, the function returns defaultValue. If no defaultValue is specified, the function returns a default-constructed value. If there are multiple items for key in the map, the value of the most recently inserted one is returned.

See also key(), values(), contains(), and operator[]().

fn values<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MapGestureTypeWidgetMutPtrValuesArgs<'largs>

C++ method: QMap<Qt::GestureType, QWidget*>::values

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn values(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListWidgetMutPtr>
2) fn values(&self, cpp_utils::AsStruct) -> ::list::ListWidgetMutPtr
C++ method: QList<QWidget*> QMap<Qt::GestureType, QWidget*>::values() const

C++ documentation:

Returns a list containing all the values in the map, in ascending order of their keys. If a key is associated with multiple values, all of its values will be in the list, and not just the most recently inserted one.

See also keys() and value().

Variant 2

Rust arguments:
1) fn values(&self, (&qt_core::qt::GestureType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListWidgetMutPtr>
2) fn values(&self, (&qt_core::qt::GestureType, cpp_utils::AsStruct)) -> ::list::ListWidgetMutPtr
C++ method: QList<QWidget*> QMap<Qt::GestureType, QWidget*>::values(const Qt::GestureType& key) const

C++ documentation:

This is an overloaded function.

Returns a list containing all the values associated with key key, from the most recently inserted to the least recently inserted one.

See also count() and insertMulti().