Enum qt_widgets::mdi_sub_window::SubWindowOption

pub enum SubWindowOption {
    AllowOutsideAreaHorizontally,
    AllowOutsideAreaVertically,
    RubberBandResize,
    RubberBandMove,
}

C++ type: QMdiSubWindow::SubWindowOption

C++ documentation:

This enum describes options that customize the behavior of QMdiSubWindow.

The SubWindowOptions type is a typedef for QFlags<SubWindowOption>. It stores an OR combination of SubWindowOption values.

Variants
AllowOutsideAreaHorizontally

C++ enum variant: AllowOutsideAreaHorizontally = 1

AllowOutsideAreaVertically

C++ enum variant: AllowOutsideAreaVertically = 2

RubberBandResize

If you enable this option, a rubber band control is used to represent the subwindow's outline, and the user resizes this instead of the subwindow itself. As a result, the subwindow maintains its original position and size until the resize operation has been completed, at which time it will receive a single QResizeEvent. By default, this option is disabled. (C++ enum variant: RubberBandResize = 4)

RubberBandMove

If you enable this option, a rubber band control is used to represent the subwindow's outline, and the user moves this instead of the subwindow itself. As a result, the subwindow remains in its original position until the move operation has completed, at which time a QMoveEvent is sent to the window. By default, this option is disabled. (C++ enum variant: RubberBandMove = 8)