Module qt_widgets::menu::overloading

Traits

MenuAccessibleDescriptionArgs
This trait represents a set of arguments accepted by Menu::accessible_description method.

MenuAccessibleNameArgs
This trait represents a set of arguments accepted by Menu::accessible_name method.

MenuActionGeometryArgs
This trait represents a set of arguments accepted by Menu::action_geometry method.

MenuActionsArgs
This trait represents a set of arguments accepted by Menu::actions method.

MenuAddActionArgs
This trait represents a set of arguments accepted by Menu::add_action method.

MenuAddMenuArgs
This trait represents a set of arguments accepted by Menu::add_menu method.

MenuAddSectionArgs
This trait represents a set of arguments accepted by Menu::add_section method.

MenuBaseSizeArgs
This trait represents a set of arguments accepted by Menu::base_size method.

MenuChildAtArgs
This trait represents a set of arguments accepted by Menu::child_at method.

MenuChildrenRectArgs
This trait represents a set of arguments accepted by Menu::children_rect method.

MenuChildrenRegionArgs
This trait represents a set of arguments accepted by Menu::children_region method.

MenuConnectArgs
This trait represents a set of arguments accepted by Menu::connect method.

MenuConnectStaticArgs
This trait represents a set of arguments accepted by Menu::connect_static method.

MenuContentsMarginsArgs
This trait represents a set of arguments accepted by Menu::contents_margins method.

MenuContentsRectArgs
This trait represents a set of arguments accepted by Menu::contents_rect method.

MenuCreateWindowContainerArgs
This trait represents a set of arguments accepted by Menu::create_window_container method.

MenuCursorArgs
This trait represents a set of arguments accepted by Menu::cursor method.

MenuDestroyedArgs
This trait represents a set of arguments accepted by Menu::destroyed method.

MenuDisconnectArgs
This trait represents a set of arguments accepted by Menu::disconnect method.

MenuDisconnectStaticArgs
This trait represents a set of arguments accepted by Menu::disconnect_static method.

MenuDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Menu::dynamic_property_names method.

MenuExecArgs
This trait represents a set of arguments accepted by Menu::exec method.

MenuExecStaticArgs
This trait represents a set of arguments accepted by Menu::exec_static method.

MenuFontInfoArgs
This trait represents a set of arguments accepted by Menu::font_info method.

MenuFontMetricsArgs
This trait represents a set of arguments accepted by Menu::font_metrics method.

MenuFrameGeometryArgs
This trait represents a set of arguments accepted by Menu::frame_geometry method.

MenuFrameSizeArgs
This trait represents a set of arguments accepted by Menu::frame_size method.

MenuGrabArgs
This trait represents a set of arguments accepted by Menu::grab method.

MenuGrabGestureArgs
This trait represents a set of arguments accepted by Menu::grab_gesture method.

MenuGrabMouseArgs
This trait represents a set of arguments accepted by Menu::grab_mouse method.

MenuGrabShortcutArgs
This trait represents a set of arguments accepted by Menu::grab_shortcut method.

MenuIconArgs
This trait represents a set of arguments accepted by Menu::icon method.

MenuInputMethodQueryArgs
This trait represents a set of arguments accepted by Menu::input_method_query method.

MenuInsertSectionArgs
This trait represents a set of arguments accepted by Menu::insert_section method.

MenuLocaleArgs
This trait represents a set of arguments accepted by Menu::locale method.

MenuMapFromArgs
This trait represents a set of arguments accepted by Menu::map_from method.

MenuMapFromGlobalArgs
This trait represents a set of arguments accepted by Menu::map_from_global method.

MenuMapFromParentArgs
This trait represents a set of arguments accepted by Menu::map_from_parent method.

MenuMapToArgs
This trait represents a set of arguments accepted by Menu::map_to method.

MenuMapToGlobalArgs
This trait represents a set of arguments accepted by Menu::map_to_global method.

MenuMapToParentArgs
This trait represents a set of arguments accepted by Menu::map_to_parent method.

MenuMaskArgs
This trait represents a set of arguments accepted by Menu::mask method.

MenuMaximumSizeArgs
This trait represents a set of arguments accepted by Menu::maximum_size method.

MenuMinimumSizeArgs
This trait represents a set of arguments accepted by Menu::minimum_size method.

MenuMinimumSizeHintArgs
This trait represents a set of arguments accepted by Menu::minimum_size_hint method.

MenuMoveArgs
This trait represents a set of arguments accepted by Menu::move_ method.

MenuNewArgs
This trait represents a set of arguments accepted by Menu::new method.

MenuNormalGeometryArgs
This trait represents a set of arguments accepted by Menu::normal_geometry method.

MenuObjectNameArgs
This trait represents a set of arguments accepted by Menu::object_name method.

MenuPopupArgs
This trait represents a set of arguments accepted by Menu::popup method.

MenuPosArgs
This trait represents a set of arguments accepted by Menu::pos method.

MenuPropertyArgs
This trait represents a set of arguments accepted by Menu::property method.

MenuRectArgs
This trait represents a set of arguments accepted by Menu::rect method.

MenuRenderArgs
This trait represents a set of arguments accepted by Menu::render method.

MenuRepaintArgs
This trait represents a set of arguments accepted by Menu::repaint method.

MenuResizeArgs
This trait represents a set of arguments accepted by Menu::resize method.

MenuSaveGeometryArgs
This trait represents a set of arguments accepted by Menu::save_geometry method.

MenuScrollArgs
This trait represents a set of arguments accepted by Menu::scroll method.

MenuSetAttributeArgs
This trait represents a set of arguments accepted by Menu::set_attribute method.

MenuSetBaseSizeArgs
This trait represents a set of arguments accepted by Menu::set_base_size method.

MenuSetContentsMarginsArgs
This trait represents a set of arguments accepted by Menu::set_contents_margins method.

MenuSetFixedSizeArgs
This trait represents a set of arguments accepted by Menu::set_fixed_size method.

MenuSetFocusArgs
This trait represents a set of arguments accepted by Menu::set_focus method.

MenuSetGeometryArgs
This trait represents a set of arguments accepted by Menu::set_geometry method.

MenuSetMaskArgs
This trait represents a set of arguments accepted by Menu::set_mask method.

MenuSetMaximumSizeArgs
This trait represents a set of arguments accepted by Menu::set_maximum_size method.

MenuSetMinimumSizeArgs
This trait represents a set of arguments accepted by Menu::set_minimum_size method.

MenuSetParentArgs
This trait represents a set of arguments accepted by Menu::set_parent method.

MenuSetShortcutAutoRepeatArgs
This trait represents a set of arguments accepted by Menu::set_shortcut_auto_repeat method.

MenuSetShortcutEnabledArgs
This trait represents a set of arguments accepted by Menu::set_shortcut_enabled method.

MenuSetSizeIncrementArgs
This trait represents a set of arguments accepted by Menu::set_size_increment method.

MenuSetSizePolicyArgs
This trait represents a set of arguments accepted by Menu::set_size_policy method.

MenuSizeArgs
This trait represents a set of arguments accepted by Menu::size method.

MenuSizeHintArgs
This trait represents a set of arguments accepted by Menu::size_hint method.

MenuSizeIncrementArgs
This trait represents a set of arguments accepted by Menu::size_increment method.

MenuSizePolicyArgs
This trait represents a set of arguments accepted by Menu::size_policy method.

MenuStartTimerArgs
This trait represents a set of arguments accepted by Menu::start_timer method.

MenuStatusTipArgs
This trait represents a set of arguments accepted by Menu::status_tip method.

MenuStyleSheetArgs
This trait represents a set of arguments accepted by Menu::style_sheet method.

MenuTitleArgs
This trait represents a set of arguments accepted by Menu::title method.

MenuToolTipArgs
This trait represents a set of arguments accepted by Menu::tool_tip method.

MenuTrArgs
This trait represents a set of arguments accepted by Menu::tr method.

MenuTrUtf8Args
This trait represents a set of arguments accepted by Menu::tr_utf8 method.

MenuUpdateArgs
This trait represents a set of arguments accepted by Menu::update method.

MenuVisibleRegionArgs
This trait represents a set of arguments accepted by Menu::visible_region method.

MenuWhatsThisArgs
This trait represents a set of arguments accepted by Menu::whats_this method.

MenuWindowFilePathArgs
This trait represents a set of arguments accepted by Menu::window_file_path method.

MenuWindowIconArgs
This trait represents a set of arguments accepted by Menu::window_icon method.

MenuWindowIconTextArgs
This trait represents a set of arguments accepted by Menu::window_icon_text method.

MenuWindowRoleArgs
This trait represents a set of arguments accepted by Menu::window_role method.

MenuWindowTitleArgs
This trait represents a set of arguments accepted by Menu::window_title method.