Module qt_widgets::menu_bar::overloading

Traits

MenuBarAccessibleDescriptionArgs
This trait represents a set of arguments accepted by MenuBar::accessible_description method.

MenuBarAccessibleNameArgs
This trait represents a set of arguments accepted by MenuBar::accessible_name method.

MenuBarActionGeometryArgs
This trait represents a set of arguments accepted by MenuBar::action_geometry method.

MenuBarActionsArgs
This trait represents a set of arguments accepted by MenuBar::actions method.

MenuBarAddActionArgs
This trait represents a set of arguments accepted by MenuBar::add_action method.

MenuBarAddMenuArgs
This trait represents a set of arguments accepted by MenuBar::add_menu method.

MenuBarBaseSizeArgs
This trait represents a set of arguments accepted by MenuBar::base_size method.

MenuBarChildAtArgs
This trait represents a set of arguments accepted by MenuBar::child_at method.

MenuBarChildrenRectArgs
This trait represents a set of arguments accepted by MenuBar::children_rect method.

MenuBarChildrenRegionArgs
This trait represents a set of arguments accepted by MenuBar::children_region method.

MenuBarConnectArgs
This trait represents a set of arguments accepted by MenuBar::connect method.

MenuBarConnectStaticArgs
This trait represents a set of arguments accepted by MenuBar::connect_static method.

MenuBarContentsMarginsArgs
This trait represents a set of arguments accepted by MenuBar::contents_margins method.

MenuBarContentsRectArgs
This trait represents a set of arguments accepted by MenuBar::contents_rect method.

MenuBarCornerWidgetArgs
This trait represents a set of arguments accepted by MenuBar::corner_widget method.

MenuBarCreateWindowContainerArgs
This trait represents a set of arguments accepted by MenuBar::create_window_container method.

MenuBarCursorArgs
This trait represents a set of arguments accepted by MenuBar::cursor method.

MenuBarDestroyedArgs
This trait represents a set of arguments accepted by MenuBar::destroyed method.

MenuBarDisconnectArgs
This trait represents a set of arguments accepted by MenuBar::disconnect method.

MenuBarDisconnectStaticArgs
This trait represents a set of arguments accepted by MenuBar::disconnect_static method.

MenuBarDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by MenuBar::dynamic_property_names method.

MenuBarFontInfoArgs
This trait represents a set of arguments accepted by MenuBar::font_info method.

MenuBarFontMetricsArgs
This trait represents a set of arguments accepted by MenuBar::font_metrics method.

MenuBarFrameGeometryArgs
This trait represents a set of arguments accepted by MenuBar::frame_geometry method.

MenuBarFrameSizeArgs
This trait represents a set of arguments accepted by MenuBar::frame_size method.

MenuBarGrabArgs
This trait represents a set of arguments accepted by MenuBar::grab method.

MenuBarGrabGestureArgs
This trait represents a set of arguments accepted by MenuBar::grab_gesture method.

MenuBarGrabMouseArgs
This trait represents a set of arguments accepted by MenuBar::grab_mouse method.

MenuBarGrabShortcutArgs
This trait represents a set of arguments accepted by MenuBar::grab_shortcut method.

MenuBarInputMethodQueryArgs
This trait represents a set of arguments accepted by MenuBar::input_method_query method.

MenuBarLocaleArgs
This trait represents a set of arguments accepted by MenuBar::locale method.

MenuBarMapFromArgs
This trait represents a set of arguments accepted by MenuBar::map_from method.

MenuBarMapFromGlobalArgs
This trait represents a set of arguments accepted by MenuBar::map_from_global method.

MenuBarMapFromParentArgs
This trait represents a set of arguments accepted by MenuBar::map_from_parent method.

MenuBarMapToArgs
This trait represents a set of arguments accepted by MenuBar::map_to method.

MenuBarMapToGlobalArgs
This trait represents a set of arguments accepted by MenuBar::map_to_global method.

MenuBarMapToParentArgs
This trait represents a set of arguments accepted by MenuBar::map_to_parent method.

MenuBarMaskArgs
This trait represents a set of arguments accepted by MenuBar::mask method.

MenuBarMaximumSizeArgs
This trait represents a set of arguments accepted by MenuBar::maximum_size method.

MenuBarMinimumSizeArgs
This trait represents a set of arguments accepted by MenuBar::minimum_size method.

MenuBarMinimumSizeHintArgs
This trait represents a set of arguments accepted by MenuBar::minimum_size_hint method.

MenuBarMoveArgs
This trait represents a set of arguments accepted by MenuBar::move_ method.

MenuBarNewArgs
This trait represents a set of arguments accepted by MenuBar::new method.

MenuBarNormalGeometryArgs
This trait represents a set of arguments accepted by MenuBar::normal_geometry method.

MenuBarObjectNameArgs
This trait represents a set of arguments accepted by MenuBar::object_name method.

MenuBarPosArgs
This trait represents a set of arguments accepted by MenuBar::pos method.

MenuBarPropertyArgs
This trait represents a set of arguments accepted by MenuBar::property method.

MenuBarRectArgs
This trait represents a set of arguments accepted by MenuBar::rect method.

MenuBarRenderArgs
This trait represents a set of arguments accepted by MenuBar::render method.

MenuBarRepaintArgs
This trait represents a set of arguments accepted by MenuBar::repaint method.

MenuBarResizeArgs
This trait represents a set of arguments accepted by MenuBar::resize method.

MenuBarSaveGeometryArgs
This trait represents a set of arguments accepted by MenuBar::save_geometry method.

MenuBarScrollArgs
This trait represents a set of arguments accepted by MenuBar::scroll method.

MenuBarSetAttributeArgs
This trait represents a set of arguments accepted by MenuBar::set_attribute method.

MenuBarSetBaseSizeArgs
This trait represents a set of arguments accepted by MenuBar::set_base_size method.

MenuBarSetContentsMarginsArgs
This trait represents a set of arguments accepted by MenuBar::set_contents_margins method.

MenuBarSetCornerWidgetArgs
This trait represents a set of arguments accepted by MenuBar::set_corner_widget method.

MenuBarSetFixedSizeArgs
This trait represents a set of arguments accepted by MenuBar::set_fixed_size method.

MenuBarSetFocusArgs
This trait represents a set of arguments accepted by MenuBar::set_focus method.

MenuBarSetGeometryArgs
This trait represents a set of arguments accepted by MenuBar::set_geometry method.

MenuBarSetMaskArgs
This trait represents a set of arguments accepted by MenuBar::set_mask method.

MenuBarSetMaximumSizeArgs
This trait represents a set of arguments accepted by MenuBar::set_maximum_size method.

MenuBarSetMinimumSizeArgs
This trait represents a set of arguments accepted by MenuBar::set_minimum_size method.

MenuBarSetParentArgs
This trait represents a set of arguments accepted by MenuBar::set_parent method.

MenuBarSetShortcutAutoRepeatArgs
This trait represents a set of arguments accepted by MenuBar::set_shortcut_auto_repeat method.

MenuBarSetShortcutEnabledArgs
This trait represents a set of arguments accepted by MenuBar::set_shortcut_enabled method.

MenuBarSetSizeIncrementArgs
This trait represents a set of arguments accepted by MenuBar::set_size_increment method.

MenuBarSetSizePolicyArgs
This trait represents a set of arguments accepted by MenuBar::set_size_policy method.

MenuBarSizeArgs
This trait represents a set of arguments accepted by MenuBar::size method.

MenuBarSizeHintArgs
This trait represents a set of arguments accepted by MenuBar::size_hint method.

MenuBarSizeIncrementArgs
This trait represents a set of arguments accepted by MenuBar::size_increment method.

MenuBarSizePolicyArgs
This trait represents a set of arguments accepted by MenuBar::size_policy method.

MenuBarStartTimerArgs
This trait represents a set of arguments accepted by MenuBar::start_timer method.

MenuBarStatusTipArgs
This trait represents a set of arguments accepted by MenuBar::status_tip method.

MenuBarStyleSheetArgs
This trait represents a set of arguments accepted by MenuBar::style_sheet method.

MenuBarToolTipArgs
This trait represents a set of arguments accepted by MenuBar::tool_tip method.

MenuBarTrArgs
This trait represents a set of arguments accepted by MenuBar::tr method.

MenuBarTrUtf8Args
This trait represents a set of arguments accepted by MenuBar::tr_utf8 method.

MenuBarUpdateArgs
This trait represents a set of arguments accepted by MenuBar::update method.

MenuBarVisibleRegionArgs
This trait represents a set of arguments accepted by MenuBar::visible_region method.

MenuBarWhatsThisArgs
This trait represents a set of arguments accepted by MenuBar::whats_this method.

MenuBarWindowFilePathArgs
This trait represents a set of arguments accepted by MenuBar::window_file_path method.

MenuBarWindowIconArgs
This trait represents a set of arguments accepted by MenuBar::window_icon method.

MenuBarWindowIconTextArgs
This trait represents a set of arguments accepted by MenuBar::window_icon_text method.

MenuBarWindowRoleArgs
This trait represents a set of arguments accepted by MenuBar::window_role method.

MenuBarWindowTitleArgs
This trait represents a set of arguments accepted by MenuBar::window_title method.