Enum qt_widgets::message_box::StandardButton

pub enum StandardButton {
    ButtonMask,
    NoButton,
    Default,
    Escape,
    FlagMask,
    Ok,
    Save,
    SaveAll,
    Open,
    Yes,
    YesToAll,
    No,
    NoToAll,
    Abort,
    Retry,
    Ignore,
    Close,
    Cancel,
    Discard,
    Help,
    Apply,
    Reset,
    RestoreDefaults,
}

C++ type: QMessageBox::StandardButton

C++ documentation:

These enums describe flags for standard buttons. Each button has a defined ButtonRole.

The following values are obsolete:

This enum was introduced or modified in Qt 4.2.

The StandardButtons type is a typedef for QFlags<StandardButton>. It stores an OR combination of StandardButton values.

See also ButtonRole and standardButtons.

Variants
ButtonMask

(C++ enum variant: ButtonMask = -769)

NoButton

An invalid button. (C++ enum variant: NoButton = 0)

Default

Use the defaultButton argument of information(), warning(), etc. instead, or call setDefaultButton(). (C++ enum variant: Default = 256)

Escape

Call setEscapeButton() instead. (C++ enum variant: Escape = 512)

FlagMask

(C++ enum variant: FlagMask = 768)

Ok

This variant corresponds to multiple C++ enum variants with the same value:

Ok = 1024: An "OK" button defined with the AcceptRole.

FirstButton = 1024

Save

A "Save" button defined with the AcceptRole. (C++ enum variant: Save = 2048)

SaveAll

A "Save All" button defined with the AcceptRole. (C++ enum variant: SaveAll = 4096)

Open

An "Open" button defined with the AcceptRole. (C++ enum variant: Open = 8192)

Yes

A "Yes" button defined with the YesRole. (C++ enum variant: Yes = 16384)

YesToAll

This variant corresponds to multiple C++ enum variants with the same value:

YesToAll = 32768: A "Yes to All" button defined with the YesRole.

YesAll = 32768: Use YesToAll instead.

No

A "No" button defined with the NoRole. (C++ enum variant: No = 65536)

NoToAll

This variant corresponds to multiple C++ enum variants with the same value:

NoToAll = 131072: A "No to All" button defined with the NoRole.

NoAll = 131072: Use NoToAll instead.

Abort

An "Abort" button defined with the RejectRole. (C++ enum variant: Abort = 262144)

Retry

A "Retry" button defined with the AcceptRole. (C++ enum variant: Retry = 524288)

Ignore

An "Ignore" button defined with the AcceptRole. (C++ enum variant: Ignore = 1048576)

Close

A "Close" button defined with the RejectRole. (C++ enum variant: Close = 2097152)

Cancel

A "Cancel" button defined with the RejectRole. (C++ enum variant: Cancel = 4194304)

Discard

A "Discard" or "Don't Save" button, depending on the platform, defined with the DestructiveRole. (C++ enum variant: Discard = 8388608)

Help

A "Help" button defined with the HelpRole. (C++ enum variant: Help = 16777216)

Apply

An "Apply" button defined with the ApplyRole. (C++ enum variant: Apply = 33554432)

Reset

A "Reset" button defined with the ResetRole. (C++ enum variant: Reset = 67108864)

RestoreDefaults

This variant corresponds to multiple C++ enum variants with the same value:

RestoreDefaults = 134217728: A "Restore Defaults" button defined with the ResetRole.

LastButton = 134217728