Enum qt_widgets::open_g_l_widget::UpdateBehavior

pub enum UpdateBehavior {
    NoPartialUpdate,
    PartialUpdate,
}

C++ type: QOpenGLWidget::UpdateBehavior

C++ documentation:

This enum describes the update semantics of QOpenGLWidget.

This enum was introduced or modified in Qt 5.5.

See also updateBehavior() and setUpdateBehavior().

Variants
NoPartialUpdate

QOpenGLWidget will discard the contents of the color buffer and the ancillary buffers after the QOpenGLWidget is rendered to screen. This is the same behavior that can be expected by calling QOpenGLContext::swapBuffers with a default opengl enabled QWindow as the argument. NoPartialUpdate can have some performance benefits on certain hardware architectures common in the mobile and embedded space when a framebuffer object is used as the rendering target. The framebuffer object is invalidated between frames with glDiscardFramebufferEXT if supported or a glClear. Please see the documentation of EXT_discard_framebuffer for more information: https://www.khronos.org/registry/gles/extensions/EXT/EXT_discard_framebuffer.txt (C++ enum variant: NoPartialUpdate = 0)

PartialUpdate

The framebuffer objects color buffer and ancillary buffers are not invalidated between frames. (C++ enum variant: PartialUpdate = 1)