Module qt_widgets::pinch_gesture::overloading

Traits

PinchGestureCenterPointArgs
This trait represents a set of arguments accepted by PinchGesture::center_point method.

PinchGestureConnectArgs
This trait represents a set of arguments accepted by PinchGesture::connect method.

PinchGestureConnectStaticArgs
This trait represents a set of arguments accepted by PinchGesture::connect_static method.

PinchGestureDestroyedArgs
This trait represents a set of arguments accepted by PinchGesture::destroyed method.

PinchGestureDisconnectArgs
This trait represents a set of arguments accepted by PinchGesture::disconnect method.

PinchGestureDisconnectStaticArgs
This trait represents a set of arguments accepted by PinchGesture::disconnect_static method.

PinchGestureDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by PinchGesture::dynamic_property_names method.

PinchGestureHotSpotArgs
This trait represents a set of arguments accepted by PinchGesture::hot_spot method.

PinchGestureLastCenterPointArgs
This trait represents a set of arguments accepted by PinchGesture::last_center_point method.

PinchGestureNewArgs
This trait represents a set of arguments accepted by PinchGesture::new method.

PinchGestureObjectNameArgs
This trait represents a set of arguments accepted by PinchGesture::object_name method.

PinchGesturePropertyArgs
This trait represents a set of arguments accepted by PinchGesture::property method.

PinchGestureStartCenterPointArgs
This trait represents a set of arguments accepted by PinchGesture::start_center_point method.

PinchGestureStartTimerArgs
This trait represents a set of arguments accepted by PinchGesture::start_timer method.

PinchGestureTrArgs
This trait represents a set of arguments accepted by PinchGesture::tr method.

PinchGestureTrUtf8Args
This trait represents a set of arguments accepted by PinchGesture::tr_utf8 method.