Module qt_widgets::plain_text_document_layout::overloading

Traits

PlainTextDocumentLayoutAnchorAtArgs
This trait represents a set of arguments accepted by PlainTextDocumentLayout::anchor_at method.

PlainTextDocumentLayoutBlockBoundingRectArgs
This trait represents a set of arguments accepted by PlainTextDocumentLayout::block_bounding_rect method.

PlainTextDocumentLayoutConnectArgs
This trait represents a set of arguments accepted by PlainTextDocumentLayout::connect method.

PlainTextDocumentLayoutConnectStaticArgs
This trait represents a set of arguments accepted by PlainTextDocumentLayout::connect_static method.

PlainTextDocumentLayoutDestroyedArgs
This trait represents a set of arguments accepted by PlainTextDocumentLayout::destroyed method.

PlainTextDocumentLayoutDisconnectArgs
This trait represents a set of arguments accepted by PlainTextDocumentLayout::disconnect method.

PlainTextDocumentLayoutDisconnectStaticArgs
This trait represents a set of arguments accepted by PlainTextDocumentLayout::disconnect_static method.

PlainTextDocumentLayoutDocumentSizeArgs
This trait represents a set of arguments accepted by PlainTextDocumentLayout::document_size method.

PlainTextDocumentLayoutDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by PlainTextDocumentLayout::dynamic_property_names method.

PlainTextDocumentLayoutFrameBoundingRectArgs
This trait represents a set of arguments accepted by PlainTextDocumentLayout::frame_bounding_rect method.

PlainTextDocumentLayoutNewArgs
This trait represents a set of arguments accepted by PlainTextDocumentLayout::new method.

PlainTextDocumentLayoutObjectNameArgs
This trait represents a set of arguments accepted by PlainTextDocumentLayout::object_name method.

PlainTextDocumentLayoutPropertyArgs
This trait represents a set of arguments accepted by PlainTextDocumentLayout::property method.

PlainTextDocumentLayoutStartTimerArgs
This trait represents a set of arguments accepted by PlainTextDocumentLayout::start_timer method.

PlainTextDocumentLayoutTrArgs
This trait represents a set of arguments accepted by PlainTextDocumentLayout::tr method.

PlainTextDocumentLayoutTrUtf8Args
This trait represents a set of arguments accepted by PlainTextDocumentLayout::tr_utf8 method.

PlainTextDocumentLayoutUnregisterHandlerArgs
This trait represents a set of arguments accepted by PlainTextDocumentLayout::unregister_handler method.

PlainTextDocumentLayoutUpdateArgs
This trait represents a set of arguments accepted by PlainTextDocumentLayout::update method.