Struct qt_widgets::pointer::PointerWidgetRef

pub struct PointerWidgetRef {
    // some fields omitted
}

C++ type: QPointer<QWidget>

C++ documentation:

The QPointer class is a template class that provides guarded pointers to QObject.

A guarded pointer, QPointer<T>, behaves like a normal C++ pointer T , except that it is automatically set to 0 when the referenced object is destroyed (unlike normal C++ pointers, which become "dangling pointers" in such cases). T must be a subclass of QObject.

Guarded pointers are useful whenever you need to store a pointer to a QObject that is owned by someone else, and therefore might be destroyed while you still hold a reference to it. You can safely test the pointer for validity.

Note that Qt 5 introduces a slight change in behavior when using QPointer.

When using QPointer on a QWidget (or a subclass of QWidget), previously the QPointer would be cleared by the QWidget destructor. Now, the QPointer is cleared by the QObject destructor (since this is when QWeakPointer objects are cleared). Any QPointers tracking a widget will NOT be cleared before the QWidget destructor destroys the children for the widget being tracked.

Qt also provides QSharedPointer, an implementation of a reference-counted shared pointer object, which can be used to maintain a collection of references to an individual pointer.

Example:

    QPointer<QLabel> label = new QLabel;
    label->setText("&Status:");
    ...
    if (label)
        label->show();

If the QLabel is deleted in the meantime, the label variable will hold 0 instead of an invalid address, and the last line will never be executed.

The functions and operators available with a QPointer are the same as those available with a normal unguarded pointer, except the pointer arithmetic operators (+, -, ++, and --), which are normally used only with arrays of objects.

Use QPointers like normal pointers and you will not need to read this class documentation.

For creating guarded pointers, you can construct or assign to them from a T or from another guarded pointer of the same type. You can compare them with each other using operator==() and operator!=(), or test for 0 with isNull(). You can dereference them using either the *x or the x->member notation.

A guarded pointer will automatically cast to a T *, so you can freely mix guarded and unguarded pointers. This means that if you have a QPointer<QWidget>, you can pass it to a function that requires a QWidget *. For this reason, it is of little value to declare functions to take a QPointer as a parameter; just use normal pointers. Use a QPointer when you are storing a pointer over time.

Note that class T must inherit QObject, or a compilation or link error will result.

Methods

impl PointerWidgetRef

fn as_q_widget_ptr(&self) -> *mut Widget

C++ method: QWidget* QPointer<QWidget>::operator QWidget*() const

fn clear(&mut self)

C++ method: void QPointer<QWidget>::clear()

C++ documentation:

Clears this QPointer object.

This function was introduced in Qt 5.0.

See also isNull().

fn data(&self) -> *mut Widget

C++ method: QWidget* QPointer<QWidget>::data() const

C++ documentation:

Returns the pointer to the object being guarded.

This function was introduced in Qt 4.4.

fn is_null(&self) -> bool

C++ method: bool QPointer<QWidget>::isNull() const

C++ documentation:

Returns true if the referenced object has been destroyed or if there is no referenced object; otherwise returns false.

fn new<Args>(args: Args) -> Args::ReturnType where Args: PointerWidgetRefNewArgs

C++ method: QPointer<QWidget>::QPointer

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::pointer::PointerWidgetRef
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::pointer::PointerWidgetRef>
C++ method: [constructor] void QPointer<QWidget>::QPointer()

Variant 2

Rust arguments:
1) fn new((*mut ::widget::Widget, cpp_utils::AsStruct)) -> ::pointer::PointerWidgetRef
2) fn new((*mut ::widget::Widget, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pointer::PointerWidgetRef>
C++ method: [constructor] void QPointer<QWidget>::QPointer(QWidget* p)

fn op_assign<'l0>(&'l0 mut self, p: *mut Widget) -> &'l0 mut PointerWidgetRef

C++ method: QPointer<QWidget>& QPointer<QWidget>::operator=(QWidget* p)

C++ documentation:

Assignment operator. This guarded pointer will now point to the same object that p points to.

fn op_indirection<'l0>(&'l0 self) -> &'l0 mut Widget

C++ method: QWidget& QPointer<QWidget>::operator*() const

C++ documentation:

Dereference operator; implements pointer semantics. Just use this operator as you would with a normal C++ pointer.

fn op_struct_deref(&self) -> *mut Widget

C++ method: QWidget* QPointer<QWidget>::operator->() const

C++ documentation:

Overloaded arrow operator; implements pointer semantics. Just use this operator as you would with a normal C++ pointer.