Module qt_widgets::proxy_style::overloading

Traits

ProxyStyleAlignedRectArgs
This trait represents a set of arguments accepted by ProxyStyle::aligned_rect method.

ProxyStyleCombinedLayoutSpacingArgs
This trait represents a set of arguments accepted by ProxyStyle::combined_layout_spacing method.

ProxyStyleConnectArgs
This trait represents a set of arguments accepted by ProxyStyle::connect method.

ProxyStyleConnectStaticArgs
This trait represents a set of arguments accepted by ProxyStyle::connect_static method.

ProxyStyleDestroyedArgs
This trait represents a set of arguments accepted by ProxyStyle::destroyed method.

ProxyStyleDisconnectArgs
This trait represents a set of arguments accepted by ProxyStyle::disconnect method.

ProxyStyleDisconnectStaticArgs
This trait represents a set of arguments accepted by ProxyStyle::disconnect_static method.

ProxyStyleDrawComplexControlArgs
This trait represents a set of arguments accepted by ProxyStyle::draw_complex_control method.

ProxyStyleDrawControlArgs
This trait represents a set of arguments accepted by ProxyStyle::draw_control method.

ProxyStyleDrawItemTextArgs
This trait represents a set of arguments accepted by ProxyStyle::draw_item_text method.

ProxyStyleDrawPrimitiveArgs
This trait represents a set of arguments accepted by ProxyStyle::draw_primitive method.

ProxyStyleDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by ProxyStyle::dynamic_property_names method.

ProxyStyleGeneratedIconPixmapArgs
This trait represents a set of arguments accepted by ProxyStyle::generated_icon_pixmap method.

ProxyStyleHitTestComplexControlArgs
This trait represents a set of arguments accepted by ProxyStyle::hit_test_complex_control method.

ProxyStyleItemPixmapRectArgs
This trait represents a set of arguments accepted by ProxyStyle::item_pixmap_rect method.

ProxyStyleItemTextRectArgs
This trait represents a set of arguments accepted by ProxyStyle::item_text_rect method.

ProxyStyleLayoutSpacingArgs
This trait represents a set of arguments accepted by ProxyStyle::layout_spacing method.

ProxyStyleNewArgs
This trait represents a set of arguments accepted by ProxyStyle::new method.

ProxyStyleObjectNameArgs
This trait represents a set of arguments accepted by ProxyStyle::object_name method.

ProxyStylePixelMetricArgs
This trait represents a set of arguments accepted by ProxyStyle::pixel_metric method.

ProxyStylePolishArgs
This trait represents a set of arguments accepted by ProxyStyle::polish method.

ProxyStylePropertyArgs
This trait represents a set of arguments accepted by ProxyStyle::property method.

ProxyStyleSizeFromContentsArgs
This trait represents a set of arguments accepted by ProxyStyle::size_from_contents method.

ProxyStyleSliderPositionFromValueArgs
This trait represents a set of arguments accepted by ProxyStyle::slider_position_from_value method.

ProxyStyleSliderValueFromPositionArgs
This trait represents a set of arguments accepted by ProxyStyle::slider_value_from_position method.

ProxyStyleStandardIconArgs
This trait represents a set of arguments accepted by ProxyStyle::standard_icon method.

ProxyStyleStandardPaletteArgs
This trait represents a set of arguments accepted by ProxyStyle::standard_palette method.

ProxyStyleStandardPixmapArgs
This trait represents a set of arguments accepted by ProxyStyle::standard_pixmap method.

ProxyStyleStartTimerArgs
This trait represents a set of arguments accepted by ProxyStyle::start_timer method.

ProxyStyleStyleHintArgs
This trait represents a set of arguments accepted by ProxyStyle::style_hint method.

ProxyStyleSubControlRectArgs
This trait represents a set of arguments accepted by ProxyStyle::sub_control_rect method.

ProxyStyleSubElementRectArgs
This trait represents a set of arguments accepted by ProxyStyle::sub_element_rect method.

ProxyStyleTrArgs
This trait represents a set of arguments accepted by ProxyStyle::tr method.

ProxyStyleTrUtf8Args
This trait represents a set of arguments accepted by ProxyStyle::tr_utf8 method.

ProxyStyleUnpolishArgs
This trait represents a set of arguments accepted by ProxyStyle::unpolish method.

ProxyStyleVisualPosArgs
This trait represents a set of arguments accepted by ProxyStyle::visual_pos method.

ProxyStyleVisualRectArgs
This trait represents a set of arguments accepted by ProxyStyle::visual_rect method.