Enum qt_widgets::scroller::State

pub enum State {
    Inactive,
    Pressed,
    Dragging,
    Scrolling,
}

C++ type: QScroller::State

C++ documentation:

This enum contains the different QScroller states.

Variants
Inactive

The scroller is not scrolling and nothing is pressed. (C++ enum variant: Inactive = 0)

Pressed

A touch event was received or the mouse button was pressed but the scroll area is currently not dragged. (C++ enum variant: Pressed = 1)

Dragging

The scroll area is currently following the touch point or mouse. (C++ enum variant: Dragging = 2)

Scrolling

The scroll area is moving on it's own. (C++ enum variant: Scrolling = 3)