Module qt_widgets::scroller::overloading

Traits

ScrollerActiveScrollersArgs
This trait represents a set of arguments accepted by Scroller::active_scrollers method.

ScrollerConnectArgs
This trait represents a set of arguments accepted by Scroller::connect method.

ScrollerConnectStaticArgs
This trait represents a set of arguments accepted by Scroller::connect_static method.

ScrollerDestroyedArgs
This trait represents a set of arguments accepted by Scroller::destroyed method.

ScrollerDisconnectArgs
This trait represents a set of arguments accepted by Scroller::disconnect method.

ScrollerDisconnectStaticArgs
This trait represents a set of arguments accepted by Scroller::disconnect_static method.

ScrollerDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Scroller::dynamic_property_names method.

ScrollerEnsureVisibleArgs
This trait represents a set of arguments accepted by Scroller::ensure_visible method.

ScrollerFinalPositionArgs
This trait represents a set of arguments accepted by Scroller::final_position method.

ScrollerGrabGestureArgs
This trait represents a set of arguments accepted by Scroller::grab_gesture method.

ScrollerHandleInputArgs
This trait represents a set of arguments accepted by Scroller::handle_input method.

ScrollerObjectNameArgs
This trait represents a set of arguments accepted by Scroller::object_name method.

ScrollerPixelPerMeterArgs
This trait represents a set of arguments accepted by Scroller::pixel_per_meter method.

ScrollerPropertyArgs
This trait represents a set of arguments accepted by Scroller::property method.

ScrollerScrollToArgs
This trait represents a set of arguments accepted by Scroller::scroll_to method.

ScrollerScrollerArgs
This trait represents a set of arguments accepted by Scroller::scroller method.

ScrollerScrollerPropertiesArgs
This trait represents a set of arguments accepted by Scroller::scroller_properties method.

ScrollerSetSnapPositionsXArgs
This trait represents a set of arguments accepted by Scroller::set_snap_positions_x method.

ScrollerSetSnapPositionsYArgs
This trait represents a set of arguments accepted by Scroller::set_snap_positions_y method.

ScrollerStartTimerArgs
This trait represents a set of arguments accepted by Scroller::start_timer method.

ScrollerTrArgs
This trait represents a set of arguments accepted by Scroller::tr method.

ScrollerTrUtf8Args
This trait represents a set of arguments accepted by Scroller::tr_utf8 method.

ScrollerVelocityArgs
This trait represents a set of arguments accepted by Scroller::velocity method.