Struct qt_widgets::scroller::Scroller

pub struct Scroller {
    // some fields omitted
}

C++ type: QScroller

C++ documentation:

The QScroller class enables kinetic scrolling for any scrolling widget or graphics item.

With kinetic scrolling, the user can push the widget in a given direction and it will continue to scroll in this direction until it is stopped either by the user or by friction. Aspects of inertia, friction and other physical concepts can be changed in order to fine-tune an intuitive user experience.

The QScroller object is the object that stores the current position and scrolling speed and takes care of updates. QScroller can be triggered by a flick gesture

QWidget w = ...;
QScroller::grabGesture(w, QScroller::LeftMouseButtonGesture);

or directly like this:

QWidget w = ...;
QScroller *scroller = QScroller::scroller(w);
scroller->scrollTo(QPointF(100, 100));

The scrolled QObjects receive a QScrollPrepareEvent whenever the scroller needs to update its geometry information and a QScrollEvent whenever the content of the object should actually be scrolled.

The scroller uses the global QAbstractAnimation timer to generate its QScrollEvents. This can be changed with QScrollerProperties::FrameRate on a per-QScroller basis.

Several examples in the scroller examples directory show how QScroller, QScrollEvent and the scroller gesture can be used.

Even though this kinetic scroller has a large number of settings available via QScrollerProperties, we recommend that you leave them all at their default, platform optimized values. Before changing them you can experiment with the plot example in the scroller examples directory.

Methods

impl Scroller

fn active_scrollers<Args>(args: Args) -> Args::ReturnType where Args: ScrollerActiveScrollersArgs

C++ method: QScroller::activeScrollers

This is an overloaded function. Available variants:

Rust arguments:
1) fn active_scrollers(cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListScrollerMutPtr>
2) fn active_scrollers(cpp_utils::AsStruct) -> ::list::ListScrollerMutPtr
C++ method: static QList<QScroller*> QScroller::activeScrollers()

C++ documentation:

Returns an application wide list of currently active QScroller objects. Active QScroller objects are in a state() that is not QScroller::Inactive. This function is useful when writing your own gesture recognizer.

fn block_signals(&mut self, b: bool) -> bool

C++ method: bool QScroller::blockSignals(bool b)

C++ documentation:

If block is true, signals emitted by this object are blocked (i.e., emitting a signal will not invoke anything connected to it). If block is false, no such blocking will occur.

The return value is the previous value of signalsBlocked().

Note that the destroyed() signal will be emitted even if the signals for this object have been blocked.

See also signalsBlocked().

fn children<'l0>(&'l0 self) -> &'l0 ListObjectMutPtr

C++ method: const QList<QObject*>& QScroller::children() const

C++ documentation:

Returns a list of child objects. The QObjectList class is defined in the <QObject> header file as the following:

typedef QList<QObject*> QObjectList;

The first child added is the first object in the list and the last child added is the last object in the list, i.e. new children are appended at the end.

Note that the list order changes when QWidget children are raised or lowered. A widget that is raised becomes the last object in the list, and a widget that is lowered becomes the first object in the list.

See also findChild(), findChildren(), parent(), and setParent().

fn connect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ScrollerConnectArgs<'largs>

C++ method: QScroller::connect

This is an overloaded function. Available variants:

Rust arguments:
1) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: QMetaObject::Connection QScroller::connect(const QObject* sender, const char* signal, const char* member, Qt::ConnectionType type = ?) const

C++ documentation:

This function overloads connect().

Connects signal from the sender object to this object's method.

Equivalent to connect(sender, signal, this, method, type).

Every connection you make emits a signal, so duplicate connections emit two signals. You can break a connection using disconnect().

Note: This function is thread-safe.

See also disconnect().

fn connect_static<Args>(args: Args) -> Args::ReturnType where Args: ScrollerConnectStaticArgs

C++ method: QScroller::connect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: static QMetaObject::Connection QScroller::connect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& method, Qt::ConnectionType type = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

The Connection handle will be invalid if it cannot create the connection, for example, the parameters were invalid. You can check if the QMetaObject::Connection is valid by casting it to a bool.

This function works in the same way as connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type) but it uses QMetaMethod to specify signal and method.

This function was introduced in Qt 4.8.

See also connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type).

Variant 2

Rust arguments:
1) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: static QMetaObject::Connection QScroller::connect(const QObject* sender, const char* signal, const QObject* receiver, const char* member, Qt::ConnectionType arg5 = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

You must use the SIGNAL() and SLOT() macros when specifying the signal and the method, for example:

QLabel label = new QLabel;
QScrollBar scrollBar = new QScrollBar;
QObject::connect(scrollBar, SIGNAL(valueChanged(int)),
                 label,  SLOT(setNum(int)));

This example ensures that the label always displays the current scroll bar value. Note that the signal and slots parameters must not contain any variable names, only the type. E.g. the following would not work and return false:

// WRONG
QObject::connect(scrollBar, SIGNAL(valueChanged(int value)),
                 label, SLOT(setNum(int value)));

A signal can also be connected to another signal:

class MyWidget : public QWidget
{
    Q_OBJECT

public: MyWidget();

signals: void buttonClicked();

private: QPushButton *myButton; };

MyWidget::MyWidget() { myButton = new QPushButton(this); connect(myButton, SIGNAL(clicked()), this, SIGNAL(buttonClicked())); }
In this example, the MyWidget constructor relays a signal from a private member variable, and makes it available under a name that relates to MyWidget.

A signal can be connected to many slots and signals. Many signals can be connected to one slot.

If a signal is connected to several slots, the slots are activated in the same order in which the connections were made, when the signal is emitted.

The function returns a QMetaObject::Connection that represents a handle to a connection if it successfully connects the signal to the slot. The connection handle will be invalid if it cannot create the connection, for example, if QObject is unable to verify the existence of either signal or method, or if their signatures aren't compatible. You can check if the handle is valid by casting it to a bool.

By default, a signal is emitted for every connection you make; two signals are emitted for duplicate connections. You can break all of these connections with a single disconnect() call. If you pass the Qt::UniqueConnection type, the connection will only be made if it is not a duplicate. If there is already a duplicate (exact same signal to the exact same slot on the same objects), the connection will fail and connect will return an invalid QMetaObject::Connection.

The optional type parameter describes the type of connection to establish. In particular, it determines whether a particular signal is delivered to a slot immediately or queued for delivery at a later time. If the signal is queued, the parameters must be of types that are known to Qt's meta-object system, because Qt needs to copy the arguments to store them in an event behind the scenes. If you try to use a queued connection and get the error message

QObject::connect: Cannot queue arguments of type 'MyType'
(Make sure 'MyType' is registered using qRegisterMetaType().)

call qRegisterMetaType() to register the data type before you establish the connection.

Note: This function is thread-safe.

See also disconnect(), sender(), qRegisterMetaType(), and Q_DECLARE_METATYPE().

fn delete_later(&mut self)

C++ method: void QScroller::deleteLater()

C++ documentation:

Schedules this object for deletion.

The object will be deleted when control returns to the event loop. If the event loop is not running when this function is called (e.g. deleteLater() is called on an object before QCoreApplication::exec()), the object will be deleted once the event loop is started. If deleteLater() is called after the main event loop has stopped, the object will not be deleted. Since Qt 4.8, if deleteLater() is called on an object that lives in a thread with no running event loop, the object will be destroyed when the thread finishes.

Note that entering and leaving a new event loop (e.g., by opening a modal dialog) will not perform the deferred deletion; for the object to be deleted, the control must return to the event loop from which deleteLater() was called.

Note: It is safe to call this function more than once; when the first deferred deletion event is delivered, any pending events for the object are removed from the event queue.

See also destroyed() and QPointer.

fn destroyed<'largs, Args>(&'largs mut self, args: Args) where Args: ScrollerDestroyedArgs<'largs>

C++ method: QScroller::destroyed

This is an overloaded function. Available variants:

Rust arguments:
1) fn destroyed(&mut self, *mut qt_core::object::Object) -> ()
2) fn destroyed(&mut self, ()) -> ()
C++ method: void QScroller::destroyed(QObject* arg1 = ?)

C++ documentation:

This signal is emitted immediately before the object obj is destroyed, and can not be blocked.

All the objects's children are destroyed immediately after this signal is emitted.

See also deleteLater() and QPointer.

fn disconnect<'largs, Args>(&'largs self, args: Args) -> bool where Args: ScrollerDisconnectArgs<'largs>

C++ method: QScroller::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn disconnect(&self, *const qt_core::object::Object) -> bool
2) fn disconnect(&self, (*const qt_core::object::Object, *const libc::c_char)) -> bool
C++ method: bool QScroller::disconnect(const QObject* receiver, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects all signals in this object from receiver's method.

A signal-slot connection is removed when either of the objects involved are destroyed.

Variant 2

Rust arguments:
1) fn disconnect(&self, *const libc::c_char) -> bool
2) fn disconnect(&self, (*const libc::c_char, *const qt_core::object::Object)) -> bool
3) fn disconnect(&self, (*const libc::c_char, *const qt_core::object::Object, *const libc::c_char)) -> bool
4) fn disconnect(&self, ()) -> bool
C++ method: bool QScroller::disconnect(const char* signal = ?, const QObject* receiver = ?, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects signal from method of receiver.

A signal-slot connection is removed when either of the objects involved are destroyed.

Note: This function is thread-safe.

fn disconnect_static<Args>(args: Args) -> bool where Args: ScrollerDisconnectStaticArgs

C++ method: QScroller::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn disconnect_static(&qt_core::meta_object::Connection) -> bool
C++ method: static bool QScroller::disconnect(const QMetaObject::Connection& arg1)

C++ documentation:

Disconnect a connection.

If the connection is invalid or has already been disconnected, do nothing and return false.

See also connect().

Variant 2

Rust arguments: fn disconnect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod)) -> bool
C++ method: static bool QScroller::disconnect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

This function provides the same possibilities like disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method) but uses QMetaMethod to represent the signal and the method to be disconnected.

Additionally this function returnsfalse and no signals and slots disconnected if:

signal is not a member of sender class or one of its parent classes.

method is not a member of receiver class or one of its parent classes.

signal instance represents not a signal.

QMetaMethod() may be used as wildcard in the meaning "any signal" or "any slot in receiving object". In the same way 0 can be used for receiver in the meaning "any receiving object". In this case method should also be QMetaMethod(). sender parameter should be never 0.

This function was introduced in Qt 4.8.

See also disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method).

Variant 3

Rust arguments: fn disconnect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char)) -> bool
C++ method: static bool QScroller::disconnect(const QObject* sender, const char* signal, const QObject* receiver, const char* member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

A signal-slot connection is removed when either of the objects involved are destroyed.

disconnect() is typically used in three ways, as the following examples demonstrate.

Disconnect everything connected to an object's signals:
disconnect(myObject, 0, 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect();

Disconnect everything connected to a specific signal:
disconnect(myObject, SIGNAL(mySignal()), 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect(SIGNAL(mySignal()));

Disconnect a specific receiver:
disconnect(myObject, 0, myReceiver, 0);

equivalent to the non-static overloaded function

myObject->disconnect(myReceiver);

0 may be used as a wildcard, meaning "any signal", "any receiving object", or "any slot in the receiving object", respectively.

The sender may never be 0. (You cannot disconnect signals from more than one object in a single call.)

If signal is 0, it disconnects receiver and method from any signal. If not, only the specified signal is disconnected.

If receiver is 0, it disconnects anything connected to signal. If not, slots in objects other than receiver are not disconnected.

If method is 0, it disconnects anything that is connected to receiver. If not, only slots named method will be disconnected, and all other slots are left alone. The method must be 0 if receiver is left out, so you cannot disconnect a specifically-named slot on all objects.

Note: This function is thread-safe.

See also connect().

fn dump_object_info(&mut self)

C++ method: void QScroller::dumpObjectInfo()

C++ documentation:

Dumps information about signal connections, etc. for this object to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectTree().

fn dump_object_tree(&mut self)

C++ method: void QScroller::dumpObjectTree()

C++ documentation:

Dumps a tree of children to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectInfo().

fn dynamic_property_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ScrollerDynamicPropertyNamesArgs<'largs>

C++ method: QScroller::dynamicPropertyNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn dynamic_property_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn dynamic_property_names(&self, cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: QList<QByteArray> QScroller::dynamicPropertyNames() const

C++ documentation:

Returns the names of all properties that were dynamically added to the object using setProperty().

This function was introduced in Qt 4.2.

fn ensure_visible<'largs, Args>(&'largs mut self, args: Args) where Args: ScrollerEnsureVisibleArgs<'largs>

C++ method: QScroller::ensureVisible

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn ensure_visible(&mut self, (&qt_core::rect_f::RectF, libc::c_double, libc::c_double)) -> ()
C++ method: void QScroller::ensureVisible(const QRectF& rect, double xmargin, double ymargin)

C++ documentation:

Starts scrolling so that the rectangle rect is visible inside the viewport with additional margins specified in pixels by xmargin and ymargin around the rect.

In cases where it is not possible to fit the rect plus margins inside the viewport the contents are scrolled so that as much as possible is visible from rect.

The scrolling speed is calculated so that the given position is reached after a platform-defined time span.

This function performs the actual scrolling by calling scrollTo().

See also scrollTo().

Variant 2

Rust arguments: fn ensure_visible(&mut self, (&qt_core::rect_f::RectF, libc::c_double, libc::c_double, libc::c_int)) -> ()
C++ method: void QScroller::ensureVisible(const QRectF& rect, double xmargin, double ymargin, int scrollTime)

C++ documentation:

This is an overloaded function.

This version will reach its destination position in scrollTime milliseconds.

fn event(&mut self, arg1: *mut Event) -> bool

C++ method: virtual bool QScroller::event(QEvent* arg1)

C++ documentation:

This virtual function receives events to an object and should return true if the event e was recognized and processed.

The event() function can be reimplemented to customize the behavior of an object.

See also installEventFilter(), timerEvent(), QCoreApplication::sendEvent(), and QCoreApplication::postEvent().

fn event_filter(&mut self, arg1: *mut Object, arg2: *mut Event) -> bool

C++ method: virtual bool QScroller::eventFilter(QObject* arg1, QEvent* arg2)

C++ documentation:

Filters events if this object has been installed as an event filter for the watched object.

In your reimplementation of this function, if you want to filter the event out, i.e. stop it being handled further, return true; otherwise return false.

Example:

class MainWindow : public QMainWindow
{
public:
    MainWindow();

protected: bool eventFilter(QObject obj, QEvent ev);

private: QTextEdit *textEdit; };

MainWindow::MainWindow() { textEdit = new QTextEdit; setCentralWidget(textEdit);

textEdit->installEventFilter(this); }

bool MainWindow::eventFilter(QObject obj, QEvent event) { if (obj == textEdit) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent>(event); qDebug() << "Ate key press" << keyEvent->key(); return true; } else { return false; } } else { // pass the event on to the parent class return QMainWindow::eventFilter(obj, event); } }
Notice in the example above that unhandled events are passed to the base class's eventFilter() function, since the base class might have reimplemented eventFilter() for its own internal purposes.

Warning: If you delete the receiver object in this function, be sure to return true. Otherwise, Qt will forward the event to the deleted object and the program might crash.

See also installEventFilter().

fn final_position<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ScrollerFinalPositionArgs<'largs>

C++ method: QScroller::finalPosition

This is an overloaded function. Available variants:

Rust arguments:
1) fn final_position(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn final_position(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QScroller::finalPosition() const

C++ documentation:

Returns the estimated final position for the current scroll movement. Returns the current position if the scroller state is not Scrolling. The result is undefined when the scroller state is Inactive.

The target position is in pixel.

See also pixelPerMeter() and scrollTo().

fn grab_gesture<Args>(args: Args) -> GestureType where Args: ScrollerGrabGestureArgs

C++ method: QScroller::grabGesture

This is an overloaded function. Available variants:

Rust arguments:
1) fn grab_gesture(*mut qt_core::object::Object) -> qt_core::qt::GestureType
2) fn grab_gesture((*mut qt_core::object::Object, ::scroller::ScrollerGestureType)) -> qt_core::qt::GestureType
C++ method: static Qt::GestureType QScroller::grabGesture(QObject* target, QScroller::ScrollerGestureType gestureType = ?)

C++ documentation:

Registers a custom scroll gesture recognizer, grabs it for the target and returns the resulting gesture type. If scrollGestureType is set to TouchGesture the gesture triggers on touch events. If it is set to one of LeftMouseButtonGesture, RightMouseButtonGesture or MiddleMouseButtonGesture it triggers on mouse events of the corresponding button.

Only one scroll gesture can be active on a single object at the same time. If you call this function twice on the same object, it will ungrab the existing gesture before grabbing the new one.

Note: To avoid unwanted side-effects, mouse events are consumed while the gesture is triggered. Since the initial mouse press event is not consumed, the gesture sends a fake mouse release event at the global position (INT_MIN, INT_MIN). This ensures that internal states of the widget that received the original mouse press are consistent.

See also ungrabGesture() and grabbedGesture().

fn grabbed_gesture(target: *mut Object) -> GestureType

C++ method: static Qt::GestureType QScroller::grabbedGesture(QObject* target)

C++ documentation:

Returns the gesture type currently grabbed for the target or 0 if no gesture is grabbed.

See also grabGesture() and ungrabGesture().

fn handle_input<'largs, Args>(&'largs mut self, args: Args) -> bool where Args: ScrollerHandleInputArgs<'largs>

C++ method: QScroller::handleInput

This is an overloaded function. Available variants:

Rust arguments:
1) fn handle_input(&mut self, (::scroller::Input, &qt_core::point_f::PointF)) -> bool
2) fn handle_input(&mut self, (::scroller::Input, &qt_core::point_f::PointF, i64)) -> bool
C++ method: bool QScroller::handleInput(QScroller::Input input, const QPointF& position, qint64 timestamp = ?)

C++ documentation:

This function is used by gesture recognizers to inform the scroller about a new input event. The scroller changes its internal state() according to the input event and its attached scroller properties. The scroller doesn't distinguish between the kind of input device the event came from. Therefore the event needs to be split into the input type, a position and a milli-second timestamp. The position needs to be in the target's coordinate system.

The return value is true if the event should be consumed by the calling filter or false if the event should be forwarded to the control.

Note: Using grabGesture() should be sufficient for most use cases.

fn has_scroller(target: *mut Object) -> bool

C++ method: static bool QScroller::hasScroller(QObject* target)

C++ documentation:

Returns true if a QScroller object was already created for target; false otherwise.

See also scroller().

fn inherits(&self, classname: *const c_char) -> bool

C++ method: bool QScroller::inherits(const char* classname) const

C++ documentation:

Returns true if this object is an instance of a class that inherits className or a QObject subclass that inherits className; otherwise returns false.

A class is considered to inherit itself.

Example:

QTimer *timer = new QTimer;         // QTimer inherits QObject
timer->inherits("QTimer");          // returns true
timer->inherits("QObject");         // returns true
timer->inherits("QAbstractButton"); // returns false

// QVBoxLayout inherits QObject and QLayoutItem QVBoxLayout *layout = new QVBoxLayout; layout->inherits("QObject"); // returns true layout->inherits("QLayoutItem"); // returns true (even though QLayoutItem is not a QObject)
If you need to determine whether an object is an instance of a particular class for the purpose of casting it, consider using qobject_cast<Type *>(object) instead.

See also metaObject() and qobject_cast().

fn install_event_filter(&mut self, arg1: *mut Object)

C++ method: void QScroller::installEventFilter(QObject* arg1)

C++ documentation:

Installs an event filter filterObj on this object. For example:

monitoredObj->installEventFilter(filterObj);

An event filter is an object that receives all events that are sent to this object. The filter can either stop the event or forward it to this object. The event filter filterObj receives events via its eventFilter() function. The eventFilter() function must return true if the event should be filtered, (i.e. stopped); otherwise it must return false.

If multiple event filters are installed on a single object, the filter that was installed last is activated first.

Here's a KeyPressEater class that eats the key presses of its monitored objects:

class KeyPressEater : public QObject
{
    Q_OBJECT
    ...

protected: bool eventFilter(QObject obj, QEvent event); };

bool KeyPressEater::eventFilter(QObject obj, QEvent event) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent >(event); qDebug("Ate key press %d", keyEvent->key()); return true; } else { // standard event processing return QObject::eventFilter(obj, event); } }
And here's how to install it on two widgets:

KeyPressEater keyPressEater = new KeyPressEater(this);
QPushButton pushButton = new QPushButton(this);
QListView *listView = new QListView(this);

pushButton->installEventFilter(keyPressEater); listView->installEventFilter(keyPressEater);
The QShortcut class, for example, uses this technique to intercept shortcut key presses.

Warning: If you delete the receiver object in your eventFilter() function, be sure to return true. If you return false, Qt sends the event to the deleted object and the program will crash.

Note that the filtering object must be in the same thread as this object. If filterObj is in a different thread, this function does nothing. If either filterObj or this object are moved to a different thread after calling this function, the event filter will not be called until both objects have the same thread affinity again (it is not removed).

See also removeEventFilter(), eventFilter(), and event().

fn is_widget_type(&self) -> bool

C++ method: bool QScroller::isWidgetType() const

C++ documentation:

Returns true if the object is a widget; otherwise returns false.

Calling this function is equivalent to calling inherits("QWidget"), except that it is much faster.

fn is_window_type(&self) -> bool

C++ method: bool QScroller::isWindowType() const

C++ documentation:

Returns true if the object is a window; otherwise returns false.

Calling this function is equivalent to calling inherits("QWindow"), except that it is much faster.

fn kill_timer(&mut self, id: c_int)

C++ method: void QScroller::killTimer(int id)

C++ documentation:

Kills the timer with timer identifier, id.

The timer identifier is returned by startTimer() when a timer event is started.

See also timerEvent() and startTimer().

fn meta_object(&self) -> *const MetaObject

C++ method: virtual const QMetaObject* QScroller::metaObject() const

fn move_to_thread(&mut self, thread: *mut Thread)

C++ method: void QScroller::moveToThread(QThread* thread)

C++ documentation:

Changes the thread affinity for this object and its children. The object cannot be moved if it has a parent. Event processing will continue in the targetThread.

To move an object to the main thread, use QApplication::instance() to retrieve a pointer to the current application, and then use QApplication::thread() to retrieve the thread in which the application lives. For example:

myObject->moveToThread(QApplication::instance()->thread());

If targetThread is zero, all event processing for this object and its children stops.

Note that all active timers for the object will be reset. The timers are first stopped in the current thread and restarted (with the same interval) in the targetThread. As a result, constantly moving an object between threads can postpone timer events indefinitely.

A QEvent::ThreadChange event is sent to this object just before the thread affinity is changed. You can handle this event to perform any special processing. Note that any new events that are posted to this object will be handled in the targetThread.

Warning: This function is not thread-safe; the current thread must be same as the current thread affinity. In other words, this function can only "push" an object from the current thread to another thread, it cannot "pull" an object from any arbitrary thread to the current thread.

See also thread().

fn object_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ScrollerObjectNameArgs<'largs>

C++ method: QScroller::objectName

This is an overloaded function. Available variants:

Rust arguments:
1) fn object_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn object_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QScroller::objectName() const

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn parent(&self) -> *mut Object

C++ method: QObject* QScroller::parent() const

C++ documentation:

Returns a pointer to the parent object.

See also setParent() and children().

fn pixel_per_meter<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ScrollerPixelPerMeterArgs<'largs>

C++ method: QScroller::pixelPerMeter

This is an overloaded function. Available variants:

Rust arguments:
1) fn pixel_per_meter(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn pixel_per_meter(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QScroller::pixelPerMeter() const

C++ documentation:

Returns the pixel per meter metric for the scrolled widget.

The value is reported for both the x and y axis separately by using a QPointF.

Note: Please note that this value should be physically correct. The actual DPI settings that Qt returns for the display may be reported wrongly on purpose by the underlying windowing system, for example on Mac OS X.

fn property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ScrollerPropertyArgs<'largs>

C++ method: QScroller::property

This is an overloaded function. Available variants:

Rust arguments:
1) fn property(&self, (*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn property(&self, (*const libc::c_char, cpp_utils::AsStruct)) -> qt_core::variant::Variant
C++ method: QVariant QScroller::property(const char* name) const

C++ documentation:

Returns the value of the object's name property.

If no such property exists, the returned variant is invalid.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

See also setProperty(), QVariant::isValid(), metaObject(), and dynamicPropertyNames().

fn qt_metacall(&mut self, arg1: Call, arg2: c_int, arg3: *mut *mut c_void) -> c_int

C++ method: virtual int QScroller::qt_metacall(QMetaObject::Call arg1, int arg2, void** arg3)

fn qt_metacast(&mut self, arg1: *const c_char) -> *mut c_void

C++ method: virtual void* QScroller::qt_metacast(const char* arg1)

fn register_user_data() -> c_uint

C++ method: static unsigned int QScroller::registerUserData()

fn remove_event_filter(&mut self, arg1: *mut Object)

C++ method: void QScroller::removeEventFilter(QObject* arg1)

C++ documentation:

Removes an event filter object obj from this object. The request is ignored if such an event filter has not been installed.

All event filters for this object are automatically removed when this object is destroyed.

It is always safe to remove an event filter, even during event filter activation (i.e. from the eventFilter() function).

See also installEventFilter(), eventFilter(), and event().

fn resend_prepare_event(&mut self)

C++ method: void QScroller::resendPrepareEvent()

C++ documentation:

This function resends the QScrollPrepareEvent. Calling resendPrepareEvent triggers a QScrollPrepareEvent from the scroller. This allows the receiver to re-set content position and content size while scrolling. Calling this function while in the Inactive state is useless as the prepare event is sent again before scrolling starts.

fn scroll_to<'largs, Args>(&'largs mut self, args: Args) where Args: ScrollerScrollToArgs<'largs>

C++ method: QScroller::scrollTo

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn scroll_to(&mut self, &qt_core::point_f::PointF) -> ()
C++ method: void QScroller::scrollTo(const QPointF& pos)

C++ documentation:

Starts scrolling the widget so that point pos is at the top-left position in the viewport.

The behaviour when scrolling outside the valid scroll area is undefined. In this case the scroller might or might not overshoot.

The scrolling speed will be calculated so that the given position will be reached after a platform-defined time span.

pos is given in viewport coordinates.

See also ensureVisible().

Variant 2

Rust arguments: fn scroll_to(&mut self, (&qt_core::point_f::PointF, libc::c_int)) -> ()
C++ method: void QScroller::scrollTo(const QPointF& pos, int scrollTime)

C++ documentation:

This is an overloaded function.

This version will reach its destination position in scrollTime milliseconds.

fn scroller<Args>(args: Args) -> Args::ReturnType where Args: ScrollerScrollerArgs

C++ method: QScroller::scroller

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn scroller(*mut qt_core::object::Object) -> *mut ::scroller::Scroller
C++ method: static QScroller* QScroller::scroller(QObject* target)

C++ documentation:

Returns the scroller for the given target. As long as the object exists this function will always return the same QScroller instance. If no QScroller exists for the target, one will implicitly be created. At no point more than one QScroller will be active on an object.

See also hasScroller() and target().

Variant 2

Rust arguments: fn scroller(*const qt_core::object::Object) -> *const ::scroller::Scroller
C++ method: static const QScroller* QScroller::scroller(const QObject* target)

C++ documentation:

This is an overloaded function.

This is the const version of scroller().

fn scroller_properties<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ScrollerScrollerPropertiesArgs<'largs>

C++ method: QScroller::scrollerProperties

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroller_properties(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::scroller_properties::ScrollerProperties>
2) fn scroller_properties(&self, cpp_utils::AsStruct) -> ::scroller_properties::ScrollerProperties
C++ method: QScrollerProperties QScroller::scrollerProperties() const

C++ documentation:

This property holds the scroller properties of this scroller. The properties are used by the QScroller to determine its scrolling behavior.

Access functions:

QScrollerProperties 	scrollerProperties() const

void 	setScrollerProperties(const QScrollerProperties & prop)

Notifier signal:

void 	scrollerPropertiesChanged(const QScrollerProperties & newProperties)

fn scroller_properties_changed(&mut self, arg1: &ScrollerProperties)

C++ method: void QScroller::scrollerPropertiesChanged(const QScrollerProperties& arg1)

C++ documentation:

This property holds the scroller properties of this scroller. The properties are used by the QScroller to determine its scrolling behavior.

Access functions:

QScrollerProperties 	scrollerProperties() const

void 	setScrollerProperties(const QScrollerProperties & prop)

Notifier signal:

void 	scrollerPropertiesChanged(const QScrollerProperties & newProperties)

fn set_object_name(&mut self, name: &String)

C++ method: void QScroller::setObjectName(const QString& name)

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn set_parent(&mut self, arg1: *mut Object)

C++ method: void QScroller::setParent(QObject* arg1)

C++ documentation:

Makes the object a child of parent.

See also parent() and children().

fn set_property(&mut self, name: *const c_char, value: &Variant) -> bool

C++ method: bool QScroller::setProperty(const char* name, const QVariant& value)

C++ documentation:

Sets the value of the object's name property to value.

If the property is defined in the class using Q_PROPERTY then true is returned on success and false otherwise. If the property is not defined using Q_PROPERTY, and therefore not listed in the meta-object, it is added as a dynamic property and false is returned.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

Dynamic properties can be queried again using property() and can be removed by setting the property value to an invalid QVariant. Changing the value of a dynamic property causes a QDynamicPropertyChangeEvent to be sent to the object.

Note: Dynamic properties starting with "_q_" are reserved for internal purposes.

See also property(), metaObject(), and dynamicPropertyNames().

fn set_scroller_properties(&mut self, prop: &ScrollerProperties)

C++ method: void QScroller::setScrollerProperties(const QScrollerProperties& prop)

C++ documentation:

This property holds the scroller properties of this scroller. The properties are used by the QScroller to determine its scrolling behavior.

Access functions:

QScrollerProperties 	scrollerProperties() const

void 	setScrollerProperties(const QScrollerProperties & prop)

Notifier signal:

void 	scrollerPropertiesChanged(const QScrollerProperties & newProperties)

fn set_snap_positions_x<'largs, Args>(&'largs mut self, args: Args) where Args: ScrollerSetSnapPositionsXArgs<'largs>

C++ method: QScroller::setSnapPositionsX

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_snap_positions_x(&mut self, &qt_gui::list::ListCDouble) -> ()
C++ method: void QScroller::setSnapPositionsX(const QList<double>& positions)

C++ documentation:

Set the snap positions for the horizontal axis to a list of positions. This overwrites all previously set snap positions and also a previously set snapping interval. Snapping can be deactivated by setting an empty list of positions.

Variant 2

Rust arguments: fn set_snap_positions_x(&mut self, (libc::c_double, libc::c_double)) -> ()
C++ method: void QScroller::setSnapPositionsX(double first, double interval)

C++ documentation:

Set the snap positions for the horizontal axis to regular spaced intervals. The first snap position is at first. The next at first + interval. This can be used to implement a list header. This overwrites all previously set snap positions and also a previously set snapping interval. Snapping can be deactivated by setting an interval of 0.0

fn set_snap_positions_y<'largs, Args>(&'largs mut self, args: Args) where Args: ScrollerSetSnapPositionsYArgs<'largs>

C++ method: QScroller::setSnapPositionsY

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_snap_positions_y(&mut self, &qt_gui::list::ListCDouble) -> ()
C++ method: void QScroller::setSnapPositionsY(const QList<double>& positions)

C++ documentation:

Set the snap positions for the vertical axis to a list of positions. This overwrites all previously set snap positions and also a previously set snapping interval. Snapping can be deactivated by setting an empty list of positions.

Variant 2

Rust arguments: fn set_snap_positions_y(&mut self, (libc::c_double, libc::c_double)) -> ()
C++ method: void QScroller::setSnapPositionsY(double first, double interval)

C++ documentation:

Set the snap positions for the vertical axis to regular spaced intervals. The first snap position is at first. The next at first + interval. This overwrites all previously set snap positions and also a previously set snapping interval. Snapping can be deactivated by setting an interval of 0.0

fn set_user_data(&mut self, id: c_uint, data: *mut ObjectUserData)

C++ method: void QScroller::setUserData(unsigned int id, QObjectUserData* data)

fn signals_blocked(&self) -> bool

C++ method: bool QScroller::signalsBlocked() const

C++ documentation:

Returns true if signals are blocked; otherwise returns false.

Signals are not blocked by default.

See also blockSignals().

fn start_timer<'largs, Args>(&'largs mut self, args: Args) -> c_int where Args: ScrollerStartTimerArgs<'largs>

C++ method: QScroller::startTimer

This is an overloaded function. Available variants:

Rust arguments:
1) fn start_timer(&mut self, libc::c_int) -> libc::c_int
2) fn start_timer(&mut self, (libc::c_int, qt_core::qt::TimerType)) -> libc::c_int
C++ method: int QScroller::startTimer(int interval, Qt::TimerType timerType = ?)

C++ documentation:

Starts a timer and returns a timer identifier, or returns zero if it could not start a timer.

A timer event will occur every interval milliseconds until killTimer() is called. If interval is 0, then the timer event occurs once every time there are no more window system events to process.

The virtual timerEvent() function is called with the QTimerEvent event parameter class when a timer event occurs. Reimplement this function to get timer events.

If multiple timers are running, the QTimerEvent::timerId() can be used to find out which timer was activated.

Example:

class MyObject : public QObject
{
    Q_OBJECT

public: MyObject(QObject *parent = 0);

protected: void timerEvent(QTimerEvent *event); };

MyObject::MyObject(QObject *parent) : QObject(parent) { startTimer(50); // 50-millisecond timer startTimer(1000); // 1-second timer startTimer(60000); // 1-minute timer }

void MyObject::timerEvent(QTimerEvent *event) { qDebug() << "Timer ID:" << event->timerId(); }
Note that QTimer's accuracy depends on the underlying operating system and hardware. The timerType argument allows you to customize the accuracy of the timer. See Qt::TimerType for information on the different timer types. Most platforms support an accuracy of 20 milliseconds; some provide more. If Qt is unable to deliver the requested number of timer events, it will silently discard some.

The QTimer class provides a high-level programming interface with single-shot timers and timer signals instead of events. There is also a QBasicTimer class that is more lightweight than QTimer and less clumsy than using timer IDs directly.

See also timerEvent(), killTimer(), and QTimer::singleShot().

fn state(&self) -> State

C++ method: QScroller::State QScroller::state() const

C++ documentation:

This property holds the state of the scroller.

Access functions:

State 	state() const

Notifier signal:

void 	stateChanged(QScroller::State newState)

See also QScroller::State.

fn state_changed(&mut self, newstate: State)

C++ method: void QScroller::stateChanged(QScroller::State newstate)

C++ documentation:

This property holds the state of the scroller.

Access functions:

State 	state() const

Notifier signal:

void 	stateChanged(QScroller::State newState)

See also QScroller::State.

fn stop(&mut self)

C++ method: void QScroller::stop()

C++ documentation:

Stops the scroller and resets its state back to Inactive.

fn target(&self) -> *mut Object

C++ method: QObject* QScroller::target() const

C++ documentation:

Returns the target object of this scroller.

See also hasScroller() and scroller().

fn thread(&self) -> *mut Thread

C++ method: QThread* QScroller::thread() const

C++ documentation:

Returns the thread in which the object lives.

See also moveToThread().

fn tr<Args>(args: Args) -> Args::ReturnType where Args: ScrollerTrArgs

C++ method: QScroller::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QScroller::tr(const char* s, const char* c, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: ScrollerTrUtf8Args

C++ method: QScroller::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QScroller::trUtf8(const char* s, const char* c, int n)

fn ungrab_gesture(target: *mut Object)

C++ method: static void QScroller::ungrabGesture(QObject* target)

C++ documentation:

Ungrabs the gesture for the target. Does nothing if no gesture is grabbed.

See also grabGesture() and grabbedGesture().

fn user_data(&self, id: c_uint) -> *mut ObjectUserData

C++ method: QObjectUserData* QScroller::userData(unsigned int id) const

fn velocity<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ScrollerVelocityArgs<'largs>

C++ method: QScroller::velocity

This is an overloaded function. Available variants:

Rust arguments:
1) fn velocity(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn velocity(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QScroller::velocity() const

C++ documentation:

Returns the current scrolling velocity in meter per second when the state is Scrolling or Dragging. Returns a zero velocity otherwise.

The velocity is reported for both the x and y axis separately by using a QPointF.

See also pixelPerMeter().