Enum qt_widgets::scroller_properties::ScrollMetric

pub enum ScrollMetric {
    MousePressEventDelay,
    DragStartDistance,
    DragVelocitySmoothingFactor,
    AxisLockThreshold,
    ScrollingCurve,
    DecelerationFactor,
    MinimumVelocity,
    MaximumVelocity,
    MaximumClickThroughVelocity,
    AcceleratingFlickMaximumTime,
    AcceleratingFlickSpeedupFactor,
    SnapPositionRatio,
    SnapTime,
    OvershootDragResistanceFactor,
    OvershootDragDistanceFactor,
    OvershootScrollDistanceFactor,
    OvershootScrollTime,
    HorizontalOvershootPolicy,
    VerticalOvershootPolicy,
    FrameRate,
    ScrollMetricCount,
}

C++ type: QScrollerProperties::ScrollMetric

C++ documentation:

This enum contains the different scroll metric types. When not indicated otherwise the setScrollMetric function expects a QVariant of type qreal.

See the QScroller documentation for further details of the concepts behind the different values.

Variants
MousePressEventDelay

This is the time a mouse press event is delayed when starting a flick gesture in [s]. If the gesture is triggered within that time, no mouse press or release is sent to the scrolled object. If it triggers after that delay the delayed mouse press plus a faked release event at global position QPoint(-QWIDGETSIZE_MAX, -QWIDGETSIZE_MAX) is sent. If the gesture is canceled, then both the delayed mouse press plus the real release event are delivered. (C++ enum variant: MousePressEventDelay = 0)

DragStartDistance

This is the minimum distance the touch or mouse point needs to be moved before the flick gesture is triggered in m. (C++ enum variant: DragStartDistance = 1)

DragVelocitySmoothingFactor

A value that describes to which extent new drag velocities are included in the final scrolling velocity. This value should be in the range between 0 and 1. The lower the value, the more smoothing is applied to the dragging velocity. (C++ enum variant: DragVelocitySmoothingFactor = 2)

AxisLockThreshold

Restricts the movement to one axis if the movement is inside an angle around the axis. The threshold must be in the range 0 to 1. (C++ enum variant: AxisLockThreshold = 3)

ScrollingCurve

The QEasingCurve used when decelerating the scrolling velocity after an user initiated flick. Please note that this is the easing curve for the positions, not the velocity: the default is QEasingCurve::OutQuad, which results in a linear decrease in velocity (1st derivative) and a constant deceleration (2nd derivative). (C++ enum variant: ScrollingCurve = 4)

DecelerationFactor

This factor influences how long it takes the scroller to decelerate to 0 velocity. The actual value depends on the chosen ScrollingCurve. For most types the value should be in the range from 0.1 to 2.0 (C++ enum variant: DecelerationFactor = 5)

MinimumVelocity

The minimum velocity that is needed after ending the touch or releasing the mouse to start scrolling in m/s. (C++ enum variant: MinimumVelocity = 6)

MaximumVelocity

This is the maximum velocity that can be reached in m/s. (C++ enum variant: MaximumVelocity = 7)

MaximumClickThroughVelocity

This is the maximum allowed scroll speed for a click-through in m/s. This means that a click on a currently (slowly) scrolling object will not only stop the scrolling but the click event will also be delivered to the UI control. This is useful when using exponential-type scrolling curves. (C++ enum variant: MaximumClickThroughVelocity = 8)

AcceleratingFlickMaximumTime

This is the maximum time in seconds that a flick gesture can take to be recognized as an accelerating flick. If set to zero no such gesture is detected. An "accelerating flick" is a flick gesture executed on an already scrolling object. In such cases the scrolling speed is multiplied by AcceleratingFlickSpeedupFactor in order to accelerate it. (C++ enum variant: AcceleratingFlickMaximumTime = 9)

AcceleratingFlickSpeedupFactor

The current speed is multiplied by this number if an accelerating flick is detected. Should be >= 1. (C++ enum variant: AcceleratingFlickSpeedupFactor = 10)

SnapPositionRatio

This is the distance that the user must drag the area beween two snap points in order to snap it to the next position. 0.33 means that the scroll must only reach one third of the distance between two snap points to snap to the next one. The ratio must be between 0 and 1. (C++ enum variant: SnapPositionRatio = 11)

SnapTime

This is the time factor for the scrolling curve. A lower value means that the scrolling will take longer. The scrolling distance is independet of this value. (C++ enum variant: SnapTime = 12)

OvershootDragResistanceFactor

This value is the factor between the mouse dragging and the actual scroll area movement (during overshoot). The factor must be between 0 and 1. (C++ enum variant: OvershootDragResistanceFactor = 13)

OvershootDragDistanceFactor

This is the maximum distance for overshoot movements while dragging. The actual overshoot distance is calculated by multiplying this value with the viewport size of the scrolled object. The factor must be between 0 and 1. (C++ enum variant: OvershootDragDistanceFactor = 14)

OvershootScrollDistanceFactor

This is the maximum distance for overshoot movements while scrolling. The actual overshoot distance is calculated by multiplying this value with the viewport size of the scrolled object. The factor must be between 0 and 1. (C++ enum variant: OvershootScrollDistanceFactor = 15)

OvershootScrollTime

This is the time in seconds that is used to play the complete overshoot animation. (C++ enum variant: OvershootScrollTime = 16)

HorizontalOvershootPolicy

This is the horizontal overshooting policy (see OvershootPolicy). (C++ enum variant: HorizontalOvershootPolicy = 17)

VerticalOvershootPolicy

This is the horizontal overshooting policy (see OvershootPolicy). (C++ enum variant: VerticalOvershootPolicy = 18)

FrameRate

This is the frame rate which should be used while dragging or scrolling. QScroller uses a QAbstractAnimation timer internally to sync all scrolling operations to other animations that might be active at the same time. If the standard value of 60 frames per second is too fast, it can be lowered with this setting, while still being in-sync with QAbstractAnimation. Please note that only the values of the FrameRates enum are allowed here. (C++ enum variant: FrameRate = 19)

ScrollMetricCount

This is always the last entry. (C++ enum variant: ScrollMetricCount = 20)