Struct qt_widgets::scroller_properties::ScrollerProperties

pub struct ScrollerProperties {
    // some fields omitted
}

C++ type: QScrollerProperties

C++ documentation:

The QScrollerProperties class stores the settings for a QScroller.

The QScrollerProperties class stores the parameters used by QScroller.

The default settings are platform dependent so that Qt emulates the platform behaviour for kinetic scrolling.

As a convention the QScrollerProperties are in physical units (meter, seconds) and are converted by QScroller using the current DPI.

Methods

impl ScrollerProperties

fn new<Args>(args: Args) -> Args::ReturnType where Args: ScrollerPropertiesNewArgs

C++ method: QScrollerProperties::QScrollerProperties

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::scroller_properties::ScrollerProperties
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::scroller_properties::ScrollerProperties>
C++ method: [constructor] void QScrollerProperties::QScrollerProperties()

Variant 2

Rust arguments:
1) fn new((&::scroller_properties::ScrollerProperties, cpp_utils::AsStruct)) -> ::scroller_properties::ScrollerProperties
2) fn new((&::scroller_properties::ScrollerProperties, cpp_utils::AsBox)) -> cpp_utils::CppBox<::scroller_properties::ScrollerProperties>
C++ method: [constructor] void QScrollerProperties::QScrollerProperties(const QScrollerProperties& sp)

fn op_assign<'l0, 'l1>(&'l0 mut self, sp: &'l1 ScrollerProperties) -> &'l0 mut ScrollerProperties

C++ method: QScrollerProperties& QScrollerProperties::operator=(const QScrollerProperties& sp)

C++ documentation:

Assigns sp to these scroller properties and returns a reference to these scroller properties.

fn op_eq(&self, sp: &ScrollerProperties) -> bool

C++ method: bool QScrollerProperties::operator==(const QScrollerProperties& sp) const

C++ documentation:

Returns true if these scroller properties are equal to sp; otherwise returns false.

fn op_neq(&self, sp: &ScrollerProperties) -> bool

C++ method: bool QScrollerProperties::operator!=(const QScrollerProperties& sp) const

C++ documentation:

Returns true if these scroller properties are different from sp; otherwise returns false.

fn scroll_metric<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ScrollerPropertiesScrollMetricArgs<'largs>

C++ method: QScrollerProperties::scrollMetric

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_metric(&self, (::scroller_properties::ScrollMetric, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn scroll_metric(&self, (::scroller_properties::ScrollMetric, cpp_utils::AsStruct)) -> qt_core::variant::Variant
C++ method: QVariant QScrollerProperties::scrollMetric(QScrollerProperties::ScrollMetric metric) const

C++ documentation:

Query the metric value of the scroller properties.

See also setScrollMetric() and ScrollMetric.

fn set_default_scroller_properties(sp: &ScrollerProperties)

C++ method: static void QScrollerProperties::setDefaultScrollerProperties(const QScrollerProperties& sp)

C++ documentation:

Sets the scroller properties for all new QScrollerProperties objects to sp.

Use this function to override the platform default properties returned by the default constructor. If you only want to change the scroller properties of a single scroller, use QScroller::setScrollerProperties()

Note: Calling this function will not change the content of already existing QScrollerProperties objects.

See also unsetDefaultScrollerProperties().

fn set_scroll_metric(&mut self, metric: ScrollMetric, value: &Variant)

C++ method: void QScrollerProperties::setScrollMetric(QScrollerProperties::ScrollMetric metric, const QVariant& value)

C++ documentation:

Set a specific value of the metric ScrollerMetric to value.

See also scrollMetric() and ScrollMetric.

fn unset_default_scroller_properties()

C++ method: static void QScrollerProperties::unsetDefaultScrollerProperties()

C++ documentation:

Sets the scroller properties returned by the default constructor back to the platform default properties.

See also setDefaultScrollerProperties().