Module qt_widgets::shortcut::overloading

Traits

ShortcutConnectArgs
This trait represents a set of arguments accepted by Shortcut::connect method.

ShortcutConnectStaticArgs
This trait represents a set of arguments accepted by Shortcut::connect_static method.

ShortcutDestroyedArgs
This trait represents a set of arguments accepted by Shortcut::destroyed method.

ShortcutDisconnectArgs
This trait represents a set of arguments accepted by Shortcut::disconnect method.

ShortcutDisconnectStaticArgs
This trait represents a set of arguments accepted by Shortcut::disconnect_static method.

ShortcutDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Shortcut::dynamic_property_names method.

ShortcutKeyArgs
This trait represents a set of arguments accepted by Shortcut::key method.

ShortcutNewArgs
This trait represents a set of arguments accepted by Shortcut::new method.

ShortcutObjectNameArgs
This trait represents a set of arguments accepted by Shortcut::object_name method.

ShortcutPropertyArgs
This trait represents a set of arguments accepted by Shortcut::property method.

ShortcutStartTimerArgs
This trait represents a set of arguments accepted by Shortcut::start_timer method.

ShortcutTrArgs
This trait represents a set of arguments accepted by Shortcut::tr method.

ShortcutTrUtf8Args
This trait represents a set of arguments accepted by Shortcut::tr_utf8 method.

ShortcutWhatsThisArgs
This trait represents a set of arguments accepted by Shortcut::whats_this method.