Enum qt_widgets::size_policy::Policy

pub enum Policy {
    Fixed,
    Minimum,
    MinimumExpanding,
    Maximum,
    Preferred,
    Expanding,
    Ignored,
}

C++ type: QSizePolicy::Policy

C++ documentation:

This enum describes the various per-dimension sizing types used when constructing a QSizePolicy.

See also PolicyFlag, setHorizontalPolicy(), and setVerticalPolicy().

Variants
Fixed

The QWidget::sizeHint() is the only acceptable alternative, so the widget can never grow or shrink (e.g. the vertical direction of a push button). (C++ enum variant: Fixed = 0)

Minimum

The sizeHint() is minimal, and sufficient. The widget can be expanded, but there is no advantage to it being larger (e.g. the horizontal direction of a push button). It cannot be smaller than the size provided by sizeHint(). (C++ enum variant: Minimum = 1)

MinimumExpanding

The sizeHint() is minimal, and sufficient. The widget can make use of extra space, so it should get as much space as possible (e.g. the horizontal direction of a horizontal slider). (C++ enum variant: MinimumExpanding = 3)

Maximum

The sizeHint() is a maximum. The widget can be shrunk any amount without detriment if other widgets need the space (e.g. a separator line). It cannot be larger than the size provided by sizeHint(). (C++ enum variant: Maximum = 4)

Preferred

The sizeHint() is best, but the widget can be shrunk and still be useful. The widget can be expanded, but there is no advantage to it being larger than sizeHint() (the default QWidget policy). (C++ enum variant: Preferred = 5)

Expanding

The sizeHint() is a sensible size, but the widget can be shrunk and still be useful. The widget can make use of extra space, so it should get as much space as possible (e.g. the horizontal direction of a horizontal slider). (C++ enum variant: Expanding = 7)

Ignored

The sizeHint() is ignored. The widget will get as much space as possible. (C++ enum variant: Ignored = 13)