Enum qt_widgets::size_policy::PolicyFlag

pub enum PolicyFlag {
    Grow,
    Expand,
    Shrink,
    Ignore,
}

C++ type: QSizePolicy::PolicyFlag

C++ documentation:

These flags are combined together to form the various Policy values:

See also Policy.

Variants
Grow

The widget can grow beyond its size hint if necessary. (C++ enum variant: GrowFlag = 1)

Expand

The widget should get as much space as possible. (C++ enum variant: ExpandFlag = 2)

Shrink

The widget can shrink below its size hint if necessary. (C++ enum variant: ShrinkFlag = 4)

Ignore

The widget's size hint is ignored. The widget will get as much space as possible. (C++ enum variant: IgnoreFlag = 8)