Struct qt_widgets::size_policy::SizePolicy

pub struct SizePolicy {
    // some fields omitted
}

C++ type: QSizePolicy

C++ documentation:

The QSizePolicy class is a layout attribute describing horizontal and vertical resizing policy.

The size policy of a widget is an expression of its willingness to be resized in various ways, and affects how the widget is treated by the layout engine. Each widget returns a QSizePolicy that describes the horizontal and vertical resizing policy it prefers when being laid out. You can change this for a specific widget by changing its QWidget::sizePolicy property.

QSizePolicy contains two independent QSizePolicy::Policy values and two stretch factors; one describes the widgets's horizontal size policy, and the other describes its vertical size policy. It also contains a flag to indicate whether the height and width of its preferred size are related.

The horizontal and vertical policies can be set in the constructor, and altered using the setHorizontalPolicy() and setVerticalPolicy() functions. The stretch factors can be set using the setHorizontalStretch() and setVerticalStretch() functions. The flag indicating whether the widget's sizeHint() is width-dependent (such as a menu bar or a word-wrapping label) can be set using the setHeightForWidth() function.

The current size policies and stretch factors be retrieved using the horizontalPolicy(), verticalPolicy(), horizontalStretch() and verticalStretch() functions. Alternatively, use the transpose() function to swap the horizontal and vertical policies and stretches. The hasHeightForWidth() function returns the current status of the flag indicating the size hint dependencies.

Use the expandingDirections() function to determine whether the associated widget can make use of more space than its sizeHint() function indicates, as well as find out in which directions it can expand.

Finally, the QSizePolicy class provides operators comparing this size policy to a given policy, as well as a QVariant operator storing this QSizePolicy as a QVariant object.

Methods

impl SizePolicy

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SizePolicyAsQVariantArgs<'largs>

C++ method: QSizePolicy::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QSizePolicy::operator QVariant() const

C++ documentation:

Returns a QVariant storing this QSizePolicy.

fn control_type(&self) -> ControlType

C++ method: QSizePolicy::ControlType QSizePolicy::controlType() const

C++ documentation:

Returns the control type associated with the widget for which this size policy applies.

This function was introduced in Qt 4.3.

See also setControlType().

fn expanding_directions(&self) -> Flags<Orientation>

C++ method: QFlags<Qt::Orientation> QSizePolicy::expandingDirections() const

C++ documentation:

Returns whether a widget can make use of more space than the QWidget::sizeHint() function indicates.

A value of Qt::Horizontal or Qt::Vertical means that the widget can grow horizontally or vertically (i.e., the horizontal or vertical policy is Expanding or MinimumExpanding), whereas Qt::Horizontal | Qt::Vertical means that it can grow in both dimensions.

See also horizontalPolicy() and verticalPolicy().

fn has_height_for_width(&self) -> bool

C++ method: bool QSizePolicy::hasHeightForWidth() const

C++ documentation:

Returns true if the widget's preferred height depends on its width; otherwise returns false.

See also setHeightForWidth().

fn has_width_for_height(&self) -> bool

C++ method: bool QSizePolicy::hasWidthForHeight() const

C++ documentation:

Returns true if the widget's width depends on its height; otherwise returns false.

See also setWidthForHeight().

fn horizontal_policy(&self) -> Policy

C++ method: QSizePolicy::Policy QSizePolicy::horizontalPolicy() const

C++ documentation:

Returns the horizontal component of the size policy.

See also setHorizontalPolicy(), verticalPolicy(), and horizontalStretch().

fn horizontal_stretch(&self) -> c_int

C++ method: int QSizePolicy::horizontalStretch() const

C++ documentation:

Returns the horizontal stretch factor of the size policy.

See also setHorizontalStretch(), verticalStretch(), and horizontalPolicy().

fn new<Args>(args: Args) -> Args::ReturnType where Args: SizePolicyNewArgs

C++ method: QSizePolicy::QSizePolicy

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::size_policy::SizePolicy
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::size_policy::SizePolicy>
C++ method: [constructor] void QSizePolicy::QSizePolicy()

Variant 2

Rust arguments:
1) fn new((::size_policy::Policy, ::size_policy::Policy, cpp_utils::AsStruct)) -> ::size_policy::SizePolicy
2) fn new((::size_policy::Policy, ::size_policy::Policy, ::size_policy::ControlType, cpp_utils::AsStruct)) -> ::size_policy::SizePolicy
3) fn new((::size_policy::Policy, ::size_policy::Policy, cpp_utils::AsBox)) -> cpp_utils::CppBox<::size_policy::SizePolicy>
4) fn new((::size_policy::Policy, ::size_policy::Policy, ::size_policy::ControlType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::size_policy::SizePolicy>
C++ method: [constructor] void QSizePolicy::QSizePolicy(QSizePolicy::Policy horizontal, QSizePolicy::Policy vertical, QSizePolicy::ControlType type = ?)

fn op_eq(&self, s: &SizePolicy) -> bool

C++ method: bool QSizePolicy::operator==(const QSizePolicy& s) const

C++ documentation:

Returns true if this policy is equal to other; otherwise returns false.

See also operator!=().

fn op_neq(&self, s: &SizePolicy) -> bool

C++ method: bool QSizePolicy::operator!=(const QSizePolicy& s) const

C++ documentation:

Returns true if this policy is different from other; otherwise returns false.

See also operator==().

fn retain_size_when_hidden(&self) -> bool

C++ method: bool QSizePolicy::retainSizeWhenHidden() const

C++ documentation:

Returns whether the layout should retain the widget's size when it is hidden. This is false by default.

This function was introduced in Qt 5.2.

See also setRetainSizeWhenHidden().

fn set_control_type(&mut self, type_: ControlType)

C++ method: void QSizePolicy::setControlType(QSizePolicy::ControlType type)

C++ documentation:

Sets the control type associated with the widget for which this size policy applies to type.

The control type specifies the type of the widget for which this size policy applies. It is used by some styles, notably QMacStyle, to insert proper spacing between widgets. For example, the Mac OS X Aqua guidelines specify that push buttons should be separated by 12 pixels, whereas vertically stacked radio buttons only require 6 pixels.

This function was introduced in Qt 4.3.

See also controlType() and QStyle::layoutSpacing().

fn set_height_for_width(&mut self, b: bool)

C++ method: void QSizePolicy::setHeightForWidth(bool b)

C++ documentation:

Sets the flag determining whether the widget's preferred height depends on its width, to dependent.

See also hasHeightForWidth() and setWidthForHeight().

fn set_horizontal_policy(&mut self, d: Policy)

C++ method: void QSizePolicy::setHorizontalPolicy(QSizePolicy::Policy d)

C++ documentation:

Sets the horizontal component to the given policy.

See also horizontalPolicy(), setVerticalPolicy(), and setHorizontalStretch().

fn set_horizontal_stretch(&mut self, stretch_factor: c_int)

C++ method: void QSizePolicy::setHorizontalStretch(int stretchFactor)

C++ documentation:

Sets the horizontal stretch factor of the size policy to the given stretchFactor. stretchFactor must be in the range [0,255].

When two widgets are adjacent to each other in a horizontal layout, setting the horizontal stretch factor of the widget on the left to 2 and the factor of widget on the right to 1 will ensure that the widget on the left will always be twice the size of the one on the right.

See also horizontalStretch(), setVerticalStretch(), and setHorizontalPolicy().

fn set_retain_size_when_hidden(&mut self, retain_size: bool)

C++ method: void QSizePolicy::setRetainSizeWhenHidden(bool retainSize)

C++ documentation:

Sets whether a layout should retain the widget's size when it is hidden. If retainSize is true, the layout will not be changed by hiding the widget.

This function was introduced in Qt 5.2.

See also retainSizeWhenHidden().

fn set_vertical_policy(&mut self, d: Policy)

C++ method: void QSizePolicy::setVerticalPolicy(QSizePolicy::Policy d)

C++ documentation:

Sets the vertical component to the given policy.

See also verticalPolicy(), setHorizontalPolicy(), and setVerticalStretch().

fn set_vertical_stretch(&mut self, stretch_factor: c_int)

C++ method: void QSizePolicy::setVerticalStretch(int stretchFactor)

C++ documentation:

Sets the vertical stretch factor of the size policy to the given stretchFactor. stretchFactor must be in the range [0,255].

When two widgets are adjacent to each other in a vertical layout, setting the vertical stretch factor of the widget on the top to 2 and the factor of widget on the bottom to 1 will ensure that the widget on the top will always be twice the size of the one on the bottom.

See also verticalStretch(), setHorizontalStretch(), and setVerticalPolicy().

fn set_width_for_height(&mut self, b: bool)

C++ method: void QSizePolicy::setWidthForHeight(bool b)

C++ documentation:

Sets the flag determining whether the widget's width depends on its height, to dependent.

This is only supported for QGraphicsLayout's subclasses. It is not possible to have a layout with both height-for-width and width-for-height constraints at the same time.

See also hasWidthForHeight() and setHeightForWidth().

fn transpose(&mut self)

C++ method: void QSizePolicy::transpose()

C++ documentation:

Swaps the horizontal and vertical policies and stretches.

fn vertical_policy(&self) -> Policy

C++ method: QSizePolicy::Policy QSizePolicy::verticalPolicy() const

C++ documentation:

Returns the vertical component of the size policy.

See also setVerticalPolicy(), horizontalPolicy(), and verticalStretch().

fn vertical_stretch(&self) -> c_int

C++ method: int QSizePolicy::verticalStretch() const

C++ documentation:

Returns the vertical stretch factor of the size policy.

See also setVerticalStretch(), horizontalStretch(), and verticalPolicy().