Enum qt_widgets::slider::TickPosition

pub enum TickPosition {
    NoTicks,
    TicksAbove,
    TicksBelow,
    TicksBothSides,
}

C++ type: QSlider::TickPosition

C++ documentation:

This enum specifies where the tick marks are to be drawn relative to the slider's groove and the handle the user moves.

Variants
NoTicks

Do not draw any tick marks. (C++ enum variant: NoTicks = 0)

TicksAbove

This variant corresponds to multiple C++ enum variants with the same value:

TicksAbove = 1: Draw tick marks above the (horizontal) slider

TicksLeft = 1: Draw tick marks to the left of the (vertical) slider

TicksBelow

This variant corresponds to multiple C++ enum variants with the same value:

TicksBelow = 2: Draw tick marks below the (horizontal) slider

TicksRight = 2: Draw tick marks to the right of the (vertical) slider

TicksBothSides

Draw tick marks on both sides of the groove. (C++ enum variant: TicksBothSides = 3)