Module qt_widgets::spacer_item::overloading

Traits

SpacerItemChangeSizeArgs
This trait represents a set of arguments accepted by SpacerItem::change_size method.

SpacerItemGeometryArgs
This trait represents a set of arguments accepted by SpacerItem::geometry method.

SpacerItemMaximumSizeArgs
This trait represents a set of arguments accepted by SpacerItem::maximum_size method.

SpacerItemMinimumSizeArgs
This trait represents a set of arguments accepted by SpacerItem::minimum_size method.

SpacerItemNewArgs
This trait represents a set of arguments accepted by SpacerItem::new method.

SpacerItemSizeHintArgs
This trait represents a set of arguments accepted by SpacerItem::size_hint method.

SpacerItemSizePolicyArgs
This trait represents a set of arguments accepted by SpacerItem::size_policy method.