Struct qt_widgets::spacer_item::SpacerItem

pub struct SpacerItem {
    // some fields omitted
}

C++ type: QSpacerItem

C++ documentation:

The QSpacerItem class provides blank space in a layout.

Normally, you don't need to use this class directly. Qt's built-in layout managers provide the following functions for manipulating empty space in layouts:

Class	Functions

QHBoxLayout	addSpacing(), addStretch(), insertSpacing(), insertStretch()

QGridLayout	setRowMinimumHeight(), setRowStretch(), setColumnMinimumWidth(), setColumnStretch()

Methods

impl SpacerItem

fn alignment(&self) -> Flags<AlignmentFlag>

C++ method: QFlags<Qt::AlignmentFlag> QSpacerItem::alignment() const

C++ documentation:

Returns the alignment of this item.

See also setAlignment().

fn change_size<'largs, Args>(&'largs mut self, args: Args) where Args: SpacerItemChangeSizeArgs<'largs>

C++ method: QSpacerItem::changeSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn change_size(&mut self, (libc::c_int, libc::c_int)) -> ()
2) fn change_size(&mut self, (libc::c_int, libc::c_int, ::size_policy::Policy)) -> ()
3) fn change_size(&mut self, (libc::c_int, libc::c_int, ::size_policy::Policy, ::size_policy::Policy)) -> ()
C++ method: void QSpacerItem::changeSize(int w, int h, QSizePolicy::Policy hData = ?, QSizePolicy::Policy vData = ?)

C++ documentation:

Changes this spacer item to have preferred width w, preferred height h, horizontal size policy hPolicy and vertical size policy vPolicy.

The default values provide a gap that is able to stretch if nothing else wants the space.

Note that if changeSize() is called after the spacer item has been added to a layout, it is necessary to invalidate the layout in order for the spacer item's new size to take effect.

See also QSpacerItem::invalidate().

fn control_types(&self) -> Flags<ControlType>

C++ method: virtual QFlags<QSizePolicy::ControlType> QSpacerItem::controlTypes() const

C++ documentation:

Returns the control type(s) for the layout item. For a QWidgetItem, the control type comes from the widget's size policy; for a QLayoutItem, the control types is derived from the layout's contents.

See also QSizePolicy::controlType().

fn expanding_directions(&self) -> Flags<Orientation>

C++ method: virtual QFlags<Qt::Orientation> QSpacerItem::expandingDirections() const

C++ documentation:

Reimplemented from QLayoutItem::expandingDirections().

fn geometry<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SpacerItemGeometryArgs<'largs>

C++ method: QSpacerItem::geometry

This is an overloaded function. Available variants:

Rust arguments:
1) fn geometry(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn geometry(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: virtual QRect QSpacerItem::geometry() const

C++ documentation:

Reimplemented from QLayoutItem::geometry().

See also setGeometry().

fn has_height_for_width(&self) -> bool

C++ method: virtual bool QSpacerItem::hasHeightForWidth() const

C++ documentation:

Returns true if this layout's preferred height depends on its width; otherwise returns false. The default implementation returns false.

Reimplement this function in layout managers that support height for width.

See also heightForWidth() and QWidget::heightForWidth().

fn height_for_width(&self, arg1: c_int) -> c_int

C++ method: virtual int QSpacerItem::heightForWidth(int arg1) const

C++ documentation:

Returns the preferred height for this layout item, given the width w.

The default implementation returns -1, indicating that the preferred height is independent of the width of the item. Using the function hasHeightForWidth() will typically be much faster than calling this function and testing for -1.

Reimplement this function in layout managers that support height for width. A typical implementation will look like this:

int MyLayout::heightForWidth(int w) const
{
    if (cache_dirty || cached_width != w) {
        // not all C++ compilers support "mutable"
        MyLayout that = (MyLayout)this;
        int h = calculateHeightForWidth(w);
        that->cached_hfw = h;
        return h;
    }
    return cached_hfw;
}

Caching is strongly recommended; without it layout will take exponential time.

See also hasHeightForWidth().

fn invalidate(&mut self)

C++ method: virtual void QSpacerItem::invalidate()

C++ documentation:

Invalidates any cached information in this layout item.

fn is_empty(&self) -> bool

C++ method: virtual bool QSpacerItem::isEmpty() const

C++ documentation:

Reimplemented from QLayoutItem::isEmpty().

Returns true.

fn layout(&mut self) -> *mut Layout

C++ method: virtual QLayout* QSpacerItem::layout()

C++ documentation:

If this item is a QLayout, it is returned as a QLayout; otherwise 0 is returned. This function provides type-safe casting.

fn maximum_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SpacerItemMaximumSizeArgs<'largs>

C++ method: QSpacerItem::maximumSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn maximum_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn maximum_size(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: virtual QSize QSpacerItem::maximumSize() const

C++ documentation:

Reimplemented from QLayoutItem::maximumSize().

fn minimum_height_for_width(&self, arg1: c_int) -> c_int

C++ method: virtual int QSpacerItem::minimumHeightForWidth(int arg1) const

C++ documentation:

Returns the minimum height this widget needs for the given width, w. The default implementation simply returns heightForWidth(w).

fn minimum_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SpacerItemMinimumSizeArgs<'largs>

C++ method: QSpacerItem::minimumSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn minimum_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn minimum_size(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: virtual QSize QSpacerItem::minimumSize() const

C++ documentation:

Reimplemented from QLayoutItem::minimumSize().

fn new<Args>(args: Args) -> Args::ReturnType where Args: SpacerItemNewArgs

C++ method: QSpacerItem::QSpacerItem

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::spacer_item::SpacerItem
2) fn new((libc::c_int, libc::c_int, ::size_policy::Policy, cpp_utils::AsStruct)) -> ::spacer_item::SpacerItem
3) fn new((libc::c_int, libc::c_int, ::size_policy::Policy, ::size_policy::Policy, cpp_utils::AsStruct)) -> ::spacer_item::SpacerItem
4) fn new((libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::spacer_item::SpacerItem>
5) fn new((libc::c_int, libc::c_int, ::size_policy::Policy, cpp_utils::AsBox)) -> cpp_utils::CppBox<::spacer_item::SpacerItem>
6) fn new((libc::c_int, libc::c_int, ::size_policy::Policy, ::size_policy::Policy, cpp_utils::AsBox)) -> cpp_utils::CppBox<::spacer_item::SpacerItem>
C++ method: [constructor] void QSpacerItem::QSpacerItem(int w, int h, QSizePolicy::Policy hData = ?, QSizePolicy::Policy vData = ?)

fn set_alignment(&mut self, a: Flags<AlignmentFlag>)

C++ method: void QSpacerItem::setAlignment(QFlags<Qt::AlignmentFlag> a)

C++ documentation:

Sets the alignment of this item to alignment.

Note: Item alignment is only supported by QLayoutItem subclasses where it would have a visual effect. Except for QSpacerItem, which provides blank space for layouts, all public Qt classes that inherit QLayoutItem support item alignment.

See also alignment().

fn set_geometry(&mut self, arg1: &Rect)

C++ method: virtual void QSpacerItem::setGeometry(const QRect& arg1)

C++ documentation:

Reimplemented from QLayoutItem::setGeometry().

See also geometry().

fn size_hint<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SpacerItemSizeHintArgs<'largs>

C++ method: QSpacerItem::sizeHint

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_hint(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size_hint(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: virtual QSize QSpacerItem::sizeHint() const

C++ documentation:

Reimplemented from QLayoutItem::sizeHint().

fn size_policy<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SpacerItemSizePolicyArgs<'largs>

C++ method: QSpacerItem::sizePolicy

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_policy(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::size_policy::SizePolicy>
2) fn size_policy(&self, cpp_utils::AsStruct) -> ::size_policy::SizePolicy
C++ method: QSizePolicy QSpacerItem::sizePolicy() const

C++ documentation:

Returns the size policy of this item.

This function was introduced in Qt 5.5.

fn spacer_item(&mut self) -> *mut SpacerItem

C++ method: virtual QSpacerItem* QSpacerItem::spacerItem()

C++ documentation:

Reimplemented from QLayoutItem::spacerItem().

Returns a pointer to this object.

fn widget(&mut self) -> *mut Widget

C++ method: virtual QWidget* QSpacerItem::widget()

C++ documentation:

If this item is a QWidget, it is returned as a QWidget; otherwise 0 is returned. This function provides type-safe casting.