Enum qt_widgets::style::ContentsType

pub enum ContentsType {
    CTCustomBase,
    CTPushButton,
    CTCheckBox,
    CTRadioButton,
    CTToolButton,
    CTComboBox,
    CTSplitter,
    CTProgressBar,
    CTMenuItem,
    CTMenuBarItem,
    CTMenuBar,
    CTMenu,
    CTTabBarTab,
    CTSlider,
    CTScrollBar,
    CTLineEdit,
    CTSpinBox,
    CTSizeGrip,
    CTTabWidget,
    CTDialogButtons,
    CTHeaderSection,
    CTGroupBox,
    CTMdiControls,
    CTItemViewItem,
}

C++ type: QStyle::ContentsType

C++ documentation:

This enum describes the available contents types. These are used to calculate sizes for the contents of various widgets.

See also sizeFromContents().

Variants
CTCustomBase

Base value for custom contents types. Custom values must be greater than this value. (C++ enum variant: CT_CustomBase = -268435456)

CTPushButton

A push button, like QPushButton. (C++ enum variant: CT_PushButton = 0)

CTCheckBox

A check box, like QCheckBox. (C++ enum variant: CT_CheckBox = 1)

CTRadioButton

A radio button, like QRadioButton. (C++ enum variant: CT_RadioButton = 2)

CTToolButton

A tool button, like QToolButton. (C++ enum variant: CT_ToolButton = 3)

CTComboBox

A combo box, like QComboBox. (C++ enum variant: CT_ComboBox = 4)

CTSplitter

A splitter, like QSplitter. (C++ enum variant: CT_Splitter = 5)

CTProgressBar

A progress bar, like QProgressBar. (C++ enum variant: CT_ProgressBar = 6)

CTMenuItem

A menu item, like QMenuItem. (C++ enum variant: CT_MenuItem = 7)

CTMenuBarItem

A menu bar item, like the buttons in a QMenuBar. (C++ enum variant: CT_MenuBarItem = 8)

CTMenuBar

A menu bar, like QMenuBar. (C++ enum variant: CT_MenuBar = 9)

CTMenu

A menu, like QMenu. (C++ enum variant: CT_Menu = 10)

CTTabBarTab

A tab on a tab bar, like QTabBar. (C++ enum variant: CT_TabBarTab = 11)

CTSlider

A slider, like QSlider. (C++ enum variant: CT_Slider = 12)

CTScrollBar

A scroll bar, like QScrollBar. (C++ enum variant: CT_ScrollBar = 13)

CTLineEdit

A line edit, like QLineEdit. (C++ enum variant: CT_LineEdit = 14)

CTSpinBox

A spin box, like QSpinBox. (C++ enum variant: CT_SpinBox = 15)

CTSizeGrip

A size grip, like QSizeGrip. (C++ enum variant: CT_SizeGrip = 16)

CTTabWidget

A tab widget, like QTabWidget. (C++ enum variant: CT_TabWidget = 17)

CTDialogButtons

C++ enum variant: CT_DialogButtons = 18

CTHeaderSection

A header section, like QHeader. (C++ enum variant: CT_HeaderSection = 19)

CTGroupBox

A group box, like QGroupBox. (C++ enum variant: CT_GroupBox = 20)

CTMdiControls

The minimize, normal, and close button in the menu bar for a maximized MDI subwindow. (C++ enum variant: CT_MdiControls = 21)

CTItemViewItem

An item inside an item view. (C++ enum variant: CT_ItemViewItem = 22)