Enum qt_widgets::style::ControlElement

pub enum ControlElement {
    CECustomBase,
    CEPushButton,
    CEPushButtonBevel,
    CEPushButtonLabel,
    CECheckBox,
    CECheckBoxLabel,
    CERadioButton,
    CERadioButtonLabel,
    CETabBarTab,
    CETabBarTabShape,
    CETabBarTabLabel,
    CEProgressBar,
    CEProgressBarGroove,
    CEProgressBarContents,
    CEProgressBarLabel,
    CEMenuItem,
    CEMenuScroller,
    CEMenuVMargin,
    CEMenuHMargin,
    CEMenuTearoff,
    CEMenuEmptyArea,
    CEMenuBarItem,
    CEMenuBarEmptyArea,
    CEToolButtonLabel,
    CEHeader,
    CEHeaderSection,
    CEHeaderLabel,
    CEToolBoxTab,
    CESizeGrip,
    CESplitter,
    CERubberBand,
    CEDockWidgetTitle,
    CEScrollBarAddLine,
    CEScrollBarSubLine,
    CEScrollBarAddPage,
    CEScrollBarSubPage,
    CEScrollBarSlider,
    CEScrollBarFirst,
    CEScrollBarLast,
    CEFocusFrame,
    CEComboBoxLabel,
    CEToolBar,
    CEToolBoxTabShape,
    CEToolBoxTabLabel,
    CEHeaderEmptyArea,
    CEColumnViewGrip,
    CEItemViewItem,
    CEShapedFrame,
}

C++ type: QStyle::ControlElement

C++ documentation:

This enum represents a control element. A control element is a part of a widget that performs some action or displays information to the user.

See also drawControl().

Variants
CECustomBase

Base value for custom control elements; custom values must be greater than this value. (C++ enum variant: CE_CustomBase = -268435456)

CEPushButton

A QPushButton, draws CE_PushButtonBevel, CE_PushButtonLabel and PE_FrameFocusRect. (C++ enum variant: CE_PushButton = 0)

CEPushButtonBevel

The bevel and default indicator of a QPushButton. (C++ enum variant: CE_PushButtonBevel = 1)

CEPushButtonLabel

The label (an icon with text or pixmap) of a QPushButton. (C++ enum variant: CE_PushButtonLabel = 2)

CECheckBox

A QCheckBox, draws a PE_IndicatorCheckBox, a CE_CheckBoxLabel and a PE_FrameFocusRect. (C++ enum variant: CE_CheckBox = 3)

CECheckBoxLabel

The label (text or pixmap) of a QCheckBox. (C++ enum variant: CE_CheckBoxLabel = 4)

CERadioButton

A QRadioButton, draws a PE_IndicatorRadioButton, a CE_RadioButtonLabel and a PE_FrameFocusRect. (C++ enum variant: CE_RadioButton = 5)

CERadioButtonLabel

The label (text or pixmap) of a QRadioButton. (C++ enum variant: CE_RadioButtonLabel = 6)

CETabBarTab

The tab and label within a QTabBar. (C++ enum variant: CE_TabBarTab = 7)

CETabBarTabShape

The tab shape within a tab bar. (C++ enum variant: CE_TabBarTabShape = 8)

CETabBarTabLabel

The label within a tab. (C++ enum variant: CE_TabBarTabLabel = 9)

CEProgressBar

A QProgressBar, draws CE_ProgressBarGroove, CE_ProgressBarContents and CE_ProgressBarLabel. (C++ enum variant: CE_ProgressBar = 10)

CEProgressBarGroove

The groove where the progress indicator is drawn in a QProgressBar. (C++ enum variant: CE_ProgressBarGroove = 11)

CEProgressBarContents

The progress indicator of a QProgressBar. (C++ enum variant: CE_ProgressBarContents = 12)

CEProgressBarLabel

The text label of a QProgressBar. (C++ enum variant: CE_ProgressBarLabel = 13)

CEMenuItem

A menu item in a QMenu. (C++ enum variant: CE_MenuItem = 14)

CEMenuScroller

Scrolling areas in a QMenu when the style supports scrolling. (C++ enum variant: CE_MenuScroller = 15)

CEMenuVMargin

The vertical extra space on the top/bottom of a menu. (C++ enum variant: CE_MenuVMargin = 16)

CEMenuHMargin

The horizontal extra space on the left/right of a menu. (C++ enum variant: CE_MenuHMargin = 17)

CEMenuTearoff

A menu item representing the tear off section of a QMenu. (C++ enum variant: CE_MenuTearoff = 18)

CEMenuEmptyArea

The area in a menu without menu items. (C++ enum variant: CE_MenuEmptyArea = 19)

CEMenuBarItem

A menu item in a QMenuBar. (C++ enum variant: CE_MenuBarItem = 20)

CEMenuBarEmptyArea

The empty area of a QMenuBar. (C++ enum variant: CE_MenuBarEmptyArea = 21)

CEToolButtonLabel

A tool button's label. (C++ enum variant: CE_ToolButtonLabel = 22)

CEHeader

A header. (C++ enum variant: CE_Header = 23)

CEHeaderSection

A header section. (C++ enum variant: CE_HeaderSection = 24)

CEHeaderLabel

The header's label. (C++ enum variant: CE_HeaderLabel = 25)

CEToolBoxTab

The toolbox's tab and label within a QToolBox. (C++ enum variant: CE_ToolBoxTab = 26)

CESizeGrip

Window resize handle; see also QSizeGrip. (C++ enum variant: CE_SizeGrip = 27)

CESplitter

Splitter handle; see also QSplitter. (C++ enum variant: CE_Splitter = 28)

CERubberBand

Rubber band used in for example an icon view. (C++ enum variant: CE_RubberBand = 29)

CEDockWidgetTitle

Dock window title. (C++ enum variant: CE_DockWidgetTitle = 30)

CEScrollBarAddLine

Scroll bar line increase indicator. (i.e., scroll down); see also QScrollBar. (C++ enum variant: CE_ScrollBarAddLine = 31)

CEScrollBarSubLine

Scroll bar line decrease indicator (i.e., scroll up). (C++ enum variant: CE_ScrollBarSubLine = 32)

CEScrollBarAddPage

Scolllbar page increase indicator (i.e., page down). (C++ enum variant: CE_ScrollBarAddPage = 33)

CEScrollBarSubPage

Scroll bar page decrease indicator (i.e., page up). (C++ enum variant: CE_ScrollBarSubPage = 34)

CEScrollBarSlider

Scroll bar slider. (C++ enum variant: CE_ScrollBarSlider = 35)

CEScrollBarFirst

Scroll bar first line indicator (i.e., home). (C++ enum variant: CE_ScrollBarFirst = 36)

CEScrollBarLast

Scroll bar last line indicator (i.e., end). (C++ enum variant: CE_ScrollBarLast = 37)

CEFocusFrame

Focus frame that is style controlled. (C++ enum variant: CE_FocusFrame = 38)

CEComboBoxLabel

The label of a non-editable QComboBox. (C++ enum variant: CE_ComboBoxLabel = 39)

CEToolBar

A toolbar like QToolBar. (C++ enum variant: CE_ToolBar = 40)

CEToolBoxTabShape

The toolbox's tab shape. (C++ enum variant: CE_ToolBoxTabShape = 41)

CEToolBoxTabLabel

The toolbox's tab label. (C++ enum variant: CE_ToolBoxTabLabel = 42)

CEHeaderEmptyArea

The area of a header view where there are no header sections. (C++ enum variant: CE_HeaderEmptyArea = 43)

CEColumnViewGrip

C++ enum variant: CE_ColumnViewGrip = 44

CEItemViewItem

An item inside an item view. (C++ enum variant: CE_ItemViewItem = 45)

CEShapedFrame

The frame with the shape specified in the QStyleOptionFrameV3; see QFrame. (C++ enum variant: CE_ShapedFrame = 46)