Enum qt_widgets::style::PixelMetric

pub enum PixelMetric {
    PMCustomBase,
    PMButtonMargin,
    PMButtonDefaultIndicator,
    PMMenuButtonIndicator,
    PMButtonShiftHorizontal,
    PMButtonShiftVertical,
    PMDefaultFrameWidth,
    PMSpinBoxFrameWidth,
    PMComboBoxFrameWidth,
    PMMaximumDragDistance,
    PMScrollBarExtent,
    PMScrollBarSliderMin,
    PMSliderThickness,
    PMSliderControlThickness,
    PMSliderLength,
    PMSliderTickmarkOffset,
    PMSliderSpaceAvailable,
    PMDockWidgetSeparatorExtent,
    PMDockWidgetHandleExtent,
    PMDockWidgetFrameWidth,
    PMTabBarTabOverlap,
    PMTabBarTabHSpace,
    PMTabBarTabVSpace,
    PMTabBarBaseHeight,
    PMTabBarBaseOverlap,
    PMProgressBarChunkWidth,
    PMSplitterWidth,
    PMTitleBarHeight,
    PMMenuScrollerHeight,
    PMMenuHMargin,
    PMMenuVMargin,
    PMMenuPanelWidth,
    PMMenuTearoffHeight,
    PMMenuDesktopFrameWidth,
    PMMenuBarPanelWidth,
    PMMenuBarItemSpacing,
    PMMenuBarVMargin,
    PMMenuBarHMargin,
    PMIndicatorWidth,
    PMIndicatorHeight,
    PMExclusiveIndicatorWidth,
    PMExclusiveIndicatorHeight,
    PMDialogButtonsSeparator,
    PMDialogButtonsButtonWidth,
    PMDialogButtonsButtonHeight,
    PMMdiSubWindowFrameWidth,
    PMMdiSubWindowMinimizedWidth,
    PMHeaderMargin,
    PMHeaderMarkSize,
    PMHeaderGripMargin,
    PMTabBarTabShiftHorizontal,
    PMTabBarTabShiftVertical,
    PMTabBarScrollButtonWidth,
    PMToolBarFrameWidth,
    PMToolBarHandleExtent,
    PMToolBarItemSpacing,
    PMToolBarItemMargin,
    PMToolBarSeparatorExtent,
    PMToolBarExtensionExtent,
    PMSpinBoxSliderHeight,
    PMDefaultTopLevelMargin,
    PMDefaultChildMargin,
    PMDefaultLayoutSpacing,
    PMToolBarIconSize,
    PMListViewIconSize,
    PMIconViewIconSize,
    PMSmallIconSize,
    PMLargeIconSize,
    PMFocusFrameVMargin,
    PMFocusFrameHMargin,
    PMToolTipLabelFrameWidth,
    PMCheckBoxLabelSpacing,
    PMTabBarIconSize,
    PMSizeGripSize,
    PMDockWidgetTitleMargin,
    PMMessageBoxIconSize,
    PMButtonIconSize,
    PMDockWidgetTitleBarButtonMargin,
    PMRadioButtonLabelSpacing,
    PMLayoutLeftMargin,
    PMLayoutTopMargin,
    PMLayoutRightMargin,
    PMLayoutBottomMargin,
    PMLayoutHorizontalSpacing,
    PMLayoutVerticalSpacing,
    PMTabBarScrollButtonOverlap,
    PMTextCursorWidth,
    PMTabCloseIndicatorWidth,
    PMTabCloseIndicatorHeight,
    PMScrollViewScrollBarSpacing,
    PMScrollViewScrollBarOverlap,
    PMSubMenuOverlap,
    PMTreeViewIndentation,
    PMHeaderDefaultSectionSizeHorizontal,
    PMHeaderDefaultSectionSizeVertical,
}

C++ type: QStyle::PixelMetric

C++ documentation:

This enum describes the various available pixel metrics. A pixel metric is a style dependent size represented by a single pixel value.

The following values are obsolete:

See also pixelMetric().

Variants
PMCustomBase

Base value for custom pixel metrics. Custom values must be greater than this value. (C++ enum variant: PM_CustomBase = -268435456)

PMButtonMargin

Amount of whitespace between push button labels and the frame. (C++ enum variant: PM_ButtonMargin = 0)

PMButtonDefaultIndicator

Width of the default-button indicator frame. (C++ enum variant: PM_ButtonDefaultIndicator = 1)

PMMenuButtonIndicator

Width of the menu button indicator proportional to the widget height. (C++ enum variant: PM_MenuButtonIndicator = 2)

PMButtonShiftHorizontal

Horizontal contents shift of a button when the button is down. (C++ enum variant: PM_ButtonShiftHorizontal = 3)

PMButtonShiftVertical

Vertical contents shift of a button when the button is down. (C++ enum variant: PM_ButtonShiftVertical = 4)

PMDefaultFrameWidth

Default frame width (usually 2). (C++ enum variant: PM_DefaultFrameWidth = 5)

PMSpinBoxFrameWidth

Frame width of a spin box, defaults to PM_DefaultFrameWidth. (C++ enum variant: PM_SpinBoxFrameWidth = 6)

PMComboBoxFrameWidth

Frame width of a combo box, defaults to PM_DefaultFrameWidth. (C++ enum variant: PM_ComboBoxFrameWidth = 7)

PMMaximumDragDistance

The maximum allowed distance between the mouse and a scrollbar when dragging. Exceeding the specified distance will cause the slider to jump back to the original position; a value of -1 disables this behavior. (C++ enum variant: PM_MaximumDragDistance = 8)

PMScrollBarExtent

Width of a vertical scroll bar and the height of a horizontal scroll bar. (C++ enum variant: PM_ScrollBarExtent = 9)

PMScrollBarSliderMin

The minimum height of a vertical scroll bar's slider and the minimum width of a horizontal scroll bar's slider. (C++ enum variant: PM_ScrollBarSliderMin = 10)

PMSliderThickness

Total slider thickness. (C++ enum variant: PM_SliderThickness = 11)

PMSliderControlThickness

Thickness of the slider handle. (C++ enum variant: PM_SliderControlThickness = 12)

PMSliderLength

Length of the slider. (C++ enum variant: PM_SliderLength = 13)

PMSliderTickmarkOffset

The offset between the tickmarks and the slider. (C++ enum variant: PM_SliderTickmarkOffset = 14)

PMSliderSpaceAvailable

The available space for the slider to move. (C++ enum variant: PM_SliderSpaceAvailable = 15)

PMDockWidgetSeparatorExtent

Width of a separator in a horizontal dock window and the height of a separator in a vertical dock window. (C++ enum variant: PM_DockWidgetSeparatorExtent = 16)

PMDockWidgetHandleExtent

Width of the handle in a horizontal dock window and the height of the handle in a vertical dock window. (C++ enum variant: PM_DockWidgetHandleExtent = 17)

PMDockWidgetFrameWidth

Frame width of a dock window. (C++ enum variant: PM_DockWidgetFrameWidth = 18)

PMTabBarTabOverlap

Number of pixels the tabs should overlap. (Currently only used in styles, not inside of QTabBar) (C++ enum variant: PM_TabBarTabOverlap = 19)

PMTabBarTabHSpace

Extra space added to the tab width. (C++ enum variant: PM_TabBarTabHSpace = 20)

PMTabBarTabVSpace

Extra space added to the tab height. (C++ enum variant: PM_TabBarTabVSpace = 21)

PMTabBarBaseHeight

Height of the area between the tab bar and the tab pages. (C++ enum variant: PM_TabBarBaseHeight = 22)

PMTabBarBaseOverlap

Number of pixels the tab bar overlaps the tab bar base. (C++ enum variant: PM_TabBarBaseOverlap = 23)

PMProgressBarChunkWidth

Width of a chunk in a progress bar indicator. (C++ enum variant: PM_ProgressBarChunkWidth = 24)

PMSplitterWidth

Width of a splitter. (C++ enum variant: PM_SplitterWidth = 25)

PMTitleBarHeight

Height of the title bar. (C++ enum variant: PM_TitleBarHeight = 26)

PMMenuScrollerHeight

Height of the scroller area in a QMenu. (C++ enum variant: PM_MenuScrollerHeight = 27)

PMMenuHMargin

Additional border (used on left and right) for a QMenu. (C++ enum variant: PM_MenuHMargin = 28)

PMMenuVMargin

Additional border (used for bottom and top) for a QMenu. (C++ enum variant: PM_MenuVMargin = 29)

PMMenuPanelWidth

Border width (applied on all sides) for a QMenu. (C++ enum variant: PM_MenuPanelWidth = 30)

PMMenuTearoffHeight

Height of a tear off area in a QMenu. (C++ enum variant: PM_MenuTearoffHeight = 31)

PMMenuDesktopFrameWidth

The frame width for the menu on the desktop. (C++ enum variant: PM_MenuDesktopFrameWidth = 32)

PMMenuBarPanelWidth

Frame width of a menu bar, defaults to PM_DefaultFrameWidth. (C++ enum variant: PM_MenuBarPanelWidth = 33)

PMMenuBarItemSpacing

Spacing between menu bar items. (C++ enum variant: PM_MenuBarItemSpacing = 34)

PMMenuBarVMargin

Spacing between menu bar items and top/bottom of bar. (C++ enum variant: PM_MenuBarVMargin = 35)

PMMenuBarHMargin

Spacing between menu bar items and left/right of bar. (C++ enum variant: PM_MenuBarHMargin = 36)

PMIndicatorWidth

Width of a check box indicator. (C++ enum variant: PM_IndicatorWidth = 37)

PMIndicatorHeight

Height of a checkbox indicator. (C++ enum variant: PM_IndicatorHeight = 38)

PMExclusiveIndicatorWidth

Width of a radio button indicator. (C++ enum variant: PM_ExclusiveIndicatorWidth = 39)

PMExclusiveIndicatorHeight

Height of a radio button indicator. (C++ enum variant: PM_ExclusiveIndicatorHeight = 40)

PMDialogButtonsSeparator

C++ enum variant: PM_DialogButtonsSeparator = 41

PMDialogButtonsButtonWidth

C++ enum variant: PM_DialogButtonsButtonWidth = 42

PMDialogButtonsButtonHeight

C++ enum variant: PM_DialogButtonsButtonHeight = 43

PMMdiSubWindowFrameWidth

This variant corresponds to multiple C++ enum variants with the same value:

PM_MdiSubWindowFrameWidth = 44: Frame width of an MDI window.

PM_MDIFrameWidth = 44: Obsolete. Use PM_MdiSubWindowFrameWidth instead.

PMMdiSubWindowMinimizedWidth

This variant corresponds to multiple C++ enum variants with the same value:

PM_MdiSubWindowMinimizedWidth = 45: Width of a minimized MDI window.

PM_MDIMinimizedWidth = 45: Obsolete. Use PM_MdiSubWindowMinimizedWidth instead.

PMHeaderMargin

The size of the margin between the sort indicator and the text. (C++ enum variant: PM_HeaderMargin = 46)

PMHeaderMarkSize

The size of the sort indicator in a header. (C++ enum variant: PM_HeaderMarkSize = 47)

PMHeaderGripMargin

The size of the resize grip in a header. (C++ enum variant: PM_HeaderGripMargin = 48)

PMTabBarTabShiftHorizontal

Horizontal pixel shift when a tab is selected. (C++ enum variant: PM_TabBarTabShiftHorizontal = 49)

PMTabBarTabShiftVertical

Vertical pixel shift when a tab is selected. (C++ enum variant: PM_TabBarTabShiftVertical = 50)

PMTabBarScrollButtonWidth

(C++ enum variant: PM_TabBarScrollButtonWidth = 51)

PMToolBarFrameWidth

Width of the frame around toolbars. (C++ enum variant: PM_ToolBarFrameWidth = 52)

PMToolBarHandleExtent

Width of a toolbar handle in a horizontal toolbar and the height of the handle in a vertical toolbar. (C++ enum variant: PM_ToolBarHandleExtent = 53)

PMToolBarItemSpacing

Spacing between toolbar items. (C++ enum variant: PM_ToolBarItemSpacing = 54)

PMToolBarItemMargin

Spacing between the toolbar frame and the items. (C++ enum variant: PM_ToolBarItemMargin = 55)

PMToolBarSeparatorExtent

Width of a toolbar separator in a horizontal toolbar and the height of a separator in a vertical toolbar. (C++ enum variant: PM_ToolBarSeparatorExtent = 56)

PMToolBarExtensionExtent

Width of a toolbar extension button in a horizontal toolbar and the height of the button in a vertical toolbar. (C++ enum variant: PM_ToolBarExtensionExtent = 57)

PMSpinBoxSliderHeight

The height of the optional spin box slider. (C++ enum variant: PM_SpinBoxSliderHeight = 58)

PMDefaultTopLevelMargin

Use PM_LayoutLeftMargin, PM_LayoutTopMargin, PM_LayoutRightMargin, and PM_LayoutBottomMargin instead. (C++ enum variant: PM_DefaultTopLevelMargin = 59)

PMDefaultChildMargin

Use PM_LayoutLeftMargin, PM_LayoutTopMargin, PM_LayoutRightMargin, and PM_LayoutBottomMargin instead. (C++ enum variant: PM_DefaultChildMargin = 60)

PMDefaultLayoutSpacing

Use PM_LayoutHorizontalSpacing and PM_LayoutVerticalSpacing instead. (C++ enum variant: PM_DefaultLayoutSpacing = 61)

PMToolBarIconSize

Default tool bar icon size (C++ enum variant: PM_ToolBarIconSize = 62)

PMListViewIconSize

The default size for icons in a list view. (C++ enum variant: PM_ListViewIconSize = 63)

PMIconViewIconSize

The default size for icons in an icon view. (C++ enum variant: PM_IconViewIconSize = 64)

PMSmallIconSize

Default small icon size (C++ enum variant: PM_SmallIconSize = 65)

PMLargeIconSize

Default large icon size (C++ enum variant: PM_LargeIconSize = 66)

PMFocusFrameVMargin

Vertical margin that the focus frame will outset the widget by. (C++ enum variant: PM_FocusFrameVMargin = 67)

PMFocusFrameHMargin

Horizontal margin that the focus frame will outset the widget by. (C++ enum variant: PM_FocusFrameHMargin = 68)

PMToolTipLabelFrameWidth

The frame width for a tool tip label. (C++ enum variant: PM_ToolTipLabelFrameWidth = 69)

PMCheckBoxLabelSpacing

The spacing between a check box indicator and its label. (C++ enum variant: PM_CheckBoxLabelSpacing = 70)

PMTabBarIconSize

The default icon size for a tab bar. (C++ enum variant: PM_TabBarIconSize = 71)

PMSizeGripSize

The size of a size grip. (C++ enum variant: PM_SizeGripSize = 72)

PMDockWidgetTitleMargin

Margin of the dock window title. (C++ enum variant: PM_DockWidgetTitleMargin = 73)

PMMessageBoxIconSize

The size of the standard icons in a message box (C++ enum variant: PM_MessageBoxIconSize = 74)

PMButtonIconSize

The default size of button icons (C++ enum variant: PM_ButtonIconSize = 75)

PMDockWidgetTitleBarButtonMargin

Amount of whitespace between dock widget's title bar button labels and the frame. (C++ enum variant: PM_DockWidgetTitleBarButtonMargin = 76)

PMRadioButtonLabelSpacing

The spacing between a radio button indicator and its label. (C++ enum variant: PM_RadioButtonLabelSpacing = 77)

PMLayoutLeftMargin

Default left margin for a QLayout. (C++ enum variant: PM_LayoutLeftMargin = 78)

PMLayoutTopMargin

Default top margin for a QLayout. (C++ enum variant: PM_LayoutTopMargin = 79)

PMLayoutRightMargin

Default right margin for a QLayout. (C++ enum variant: PM_LayoutRightMargin = 80)

PMLayoutBottomMargin

Default bottom margin for a QLayout. (C++ enum variant: PM_LayoutBottomMargin = 81)

PMLayoutHorizontalSpacing

Default horizontal spacing for a QLayout. (C++ enum variant: PM_LayoutHorizontalSpacing = 82)

PMLayoutVerticalSpacing

Default vertical spacing for a QLayout. (C++ enum variant: PM_LayoutVerticalSpacing = 83)

PMTabBarScrollButtonOverlap

The distance between the left and right buttons in a tab bar. (C++ enum variant: PM_TabBar_ScrollButtonOverlap = 84)

PMTextCursorWidth

The width of the cursor in a line edit or text edit (C++ enum variant: PM_TextCursorWidth = 85)

PMTabCloseIndicatorWidth

The default width of a close button on a tab in a tab bar. (C++ enum variant: PM_TabCloseIndicatorWidth = 86)

PMTabCloseIndicatorHeight

The default height of a close button on a tab in a tab bar. (C++ enum variant: PM_TabCloseIndicatorHeight = 87)

PMScrollViewScrollBarSpacing

Distance between frame and scrollbar with SH_ScrollView_FrameOnlyAroundContents set. (C++ enum variant: PM_ScrollView_ScrollBarSpacing = 88)

PMScrollViewScrollBarOverlap

Overlap between scroll bars and scroll content (C++ enum variant: PM_ScrollView_ScrollBarOverlap = 89)

PMSubMenuOverlap

The horizontal overlap between a submenu and its parent. (C++ enum variant: PM_SubMenuOverlap = 90)

PMTreeViewIndentation

The indentation of items in a tree view. This enum value has been introduced in Qt 5.4. (C++ enum variant: PM_TreeViewIndentation = 91)

PMHeaderDefaultSectionSizeHorizontal

The default size of sections in a horizontal header. This enum value has been introduced in Qt 5.5. (C++ enum variant: PM_HeaderDefaultSectionSizeHorizontal = 92)

PMHeaderDefaultSectionSizeVertical

The default size of sections in a vertical header. This enum value has been introduced in Qt 5.5. (C++ enum variant: PM_HeaderDefaultSectionSizeVertical = 93)