Enum qt_widgets::style::PrimitiveElement

pub enum PrimitiveElement {
    PEFrame,
    PEFrameDefaultButton,
    PEFrameDockWidget,
    PEFrameFocusRect,
    PEFrameGroupBox,
    PEFrameLineEdit,
    PEFrameMenu,
    PEFrameStatusBar,
    PEFrameTabWidget,
    PEFrameWindow,
    PEFrameButtonBevel,
    PEFrameButtonTool,
    PEFrameTabBarBase,
    PEPanelButtonCommand,
    PEPanelButtonBevel,
    PEPanelButtonTool,
    PEPanelMenuBar,
    PEPanelToolBar,
    PEPanelLineEdit,
    PEIndicatorArrowDown,
    PEIndicatorArrowLeft,
    PEIndicatorArrowRight,
    PEIndicatorArrowUp,
    PEIndicatorBranch,
    PEIndicatorButtonDropDown,
    PEIndicatorViewItemCheck,
    PEIndicatorCheckBox,
    PEIndicatorDockWidgetResizeHandle,
    PEIndicatorHeaderArrow,
    PEIndicatorMenuCheckMark,
    PEIndicatorProgressChunk,
    PEIndicatorRadioButton,
    PEIndicatorSpinDown,
    PEIndicatorSpinMinus,
    PEIndicatorSpinPlus,
    PEIndicatorSpinUp,
    PEIndicatorToolBarHandle,
    PEIndicatorToolBarSeparator,
    PEPanelTipLabel,
    PEIndicatorTabTear,
    PEPanelScrollAreaCorner,
    PEWidget,
    PEIndicatorColumnViewArrow,
    PEIndicatorItemViewItemDrop,
    PEPanelItemViewItem,
    PEPanelItemViewRow,
    PEPanelStatusBar,
    PEIndicatorTabClose,
    PEPanelMenu,
    PECustomBase,
}

C++ type: QStyle::PrimitiveElement

C++ documentation:

This enum describes the various primitive elements. A primitive element is a common GUI element, such as a checkbox indicator or button bevel.

See also drawPrimitive().

Variants
PEFrame

Generic frame (C++ enum variant: PE_Frame = 0)

PEFrameDefaultButton

This frame around a default button, e.g. in a dialog. (C++ enum variant: PE_FrameDefaultButton = 1)

PEFrameDockWidget

Panel frame for dock windows and toolbars. (C++ enum variant: PE_FrameDockWidget = 2)

PEFrameFocusRect

Generic focus indicator. (C++ enum variant: PE_FrameFocusRect = 3)

PEFrameGroupBox

Panel frame around group boxes. (C++ enum variant: PE_FrameGroupBox = 4)

PEFrameLineEdit

Panel frame for line edits. (C++ enum variant: PE_FrameLineEdit = 5)

PEFrameMenu

Frame for popup windows/menus; see also QMenu. (C++ enum variant: PE_FrameMenu = 6)

PEFrameStatusBar

This variant corresponds to multiple C++ enum variants with the same value:

PE_FrameStatusBar = 7: Obsolete. Use PE_FrameStatusBarItem instead.

PE_FrameStatusBarItem = 7: Frame for an item of a status bar; see also QStatusBar.

PEFrameTabWidget

Frame for tab widgets. (C++ enum variant: PE_FrameTabWidget = 8)

PEFrameWindow

Frame around a MDI window or a docking window. (C++ enum variant: PE_FrameWindow = 9)

PEFrameButtonBevel

Panel frame for a button bevel. (C++ enum variant: PE_FrameButtonBevel = 10)

PEFrameButtonTool

Panel frame for a tool button. (C++ enum variant: PE_FrameButtonTool = 11)

PEFrameTabBarBase

The frame that is drawn for a tab bar, ususally drawn for a tab bar that isn't part of a tab widget. (C++ enum variant: PE_FrameTabBarBase = 12)

PEPanelButtonCommand

Button used to initiate an action, for example, a QPushButton. (C++ enum variant: PE_PanelButtonCommand = 13)

PEPanelButtonBevel

Generic panel with a button bevel. (C++ enum variant: PE_PanelButtonBevel = 14)

PEPanelButtonTool

Panel for a Tool button, used with QToolButton. (C++ enum variant: PE_PanelButtonTool = 15)

PEPanelMenuBar

Panel for menu bars. (C++ enum variant: PE_PanelMenuBar = 16)

PEPanelToolBar

The panel for a toolbar. (C++ enum variant: PE_PanelToolBar = 17)

PEPanelLineEdit

Panel for a QLineEdit. (C++ enum variant: PE_PanelLineEdit = 18)

PEIndicatorArrowDown

Generic Down arrow. (C++ enum variant: PE_IndicatorArrowDown = 19)

PEIndicatorArrowLeft

Generic Left arrow. (C++ enum variant: PE_IndicatorArrowLeft = 20)

PEIndicatorArrowRight

Generic Right arrow. (C++ enum variant: PE_IndicatorArrowRight = 21)

PEIndicatorArrowUp

Generic Up arrow. (C++ enum variant: PE_IndicatorArrowUp = 22)

PEIndicatorBranch

Lines used to represent the branch of a tree in a tree view. (C++ enum variant: PE_IndicatorBranch = 23)

PEIndicatorButtonDropDown

Indicator for a drop down button, for example, a tool button that displays a menu. (C++ enum variant: PE_IndicatorButtonDropDown = 24)

PEIndicatorViewItemCheck

This variant corresponds to multiple C++ enum variants with the same value:

PE_IndicatorViewItemCheck = 25

PE_IndicatorItemViewItemCheck = 25: On/off indicator for a view item.

PEIndicatorCheckBox

On/off indicator, for example, a QCheckBox. (C++ enum variant: PE_IndicatorCheckBox = 26)

PEIndicatorDockWidgetResizeHandle

Resize handle for dock windows. (C++ enum variant: PE_IndicatorDockWidgetResizeHandle = 27)

PEIndicatorHeaderArrow

Arrow used to indicate sorting on a list or table header. (C++ enum variant: PE_IndicatorHeaderArrow = 28)

PEIndicatorMenuCheckMark

Check mark used in a menu. (C++ enum variant: PE_IndicatorMenuCheckMark = 29)

PEIndicatorProgressChunk

Section of a progress bar indicator; see also QProgressBar. (C++ enum variant: PE_IndicatorProgressChunk = 30)

PEIndicatorRadioButton

Exclusive on/off indicator, for example, a QRadioButton. (C++ enum variant: PE_IndicatorRadioButton = 31)

PEIndicatorSpinDown

Down symbol for a spin widget. (C++ enum variant: PE_IndicatorSpinDown = 32)

PEIndicatorSpinMinus

Decrease symbol for a spin widget. (C++ enum variant: PE_IndicatorSpinMinus = 33)

PEIndicatorSpinPlus

Increase symbol for a spin widget. (C++ enum variant: PE_IndicatorSpinPlus = 34)

PEIndicatorSpinUp

Up symbol for a spin widget, for example a QSpinBox. (C++ enum variant: PE_IndicatorSpinUp = 35)

PEIndicatorToolBarHandle

The handle of a toolbar. (C++ enum variant: PE_IndicatorToolBarHandle = 36)

PEIndicatorToolBarSeparator

The separator in a toolbar. (C++ enum variant: PE_IndicatorToolBarSeparator = 37)

PEPanelTipLabel

The panel for a tip label. (C++ enum variant: PE_PanelTipLabel = 38)

PEIndicatorTabTear

An indicator that a tab is partially scrolled out of the visible tab bar when there are many tabs. (C++ enum variant: PE_IndicatorTabTear = 39)

PEPanelScrollAreaCorner

Panel at the bottom-right (or bottom-left) corner of a scroll area. (C++ enum variant: PE_PanelScrollAreaCorner = 40)

PEWidget

A plain QWidget. (C++ enum variant: PE_Widget = 41)

PEIndicatorColumnViewArrow

An arrow in a QColumnView. (C++ enum variant: PE_IndicatorColumnViewArrow = 42)

PEIndicatorItemViewItemDrop

An indicator that is drawn to show where an item in an item view is about to be dropped during a drag-and-drop operation in an item view. (C++ enum variant: PE_IndicatorItemViewItemDrop = 43)

PEPanelItemViewItem

The background for an item in an item view. (C++ enum variant: PE_PanelItemViewItem = 44)

PEPanelItemViewRow

The background of a row in an item view. (C++ enum variant: PE_PanelItemViewRow = 45)

PEPanelStatusBar

The panel for a status bar. (C++ enum variant: PE_PanelStatusBar = 46)

PEIndicatorTabClose

The close button on a tab bar. (C++ enum variant: PE_IndicatorTabClose = 47)

PEPanelMenu

The panel for a menu. (C++ enum variant: PE_PanelMenu = 48)

PECustomBase

Base value for custom primitive elements. All values above this are reserved for custom use. Custom values must be greater than this value. (C++ enum variant: PE_CustomBase = 251658240)