Enum qt_widgets::style::StandardPixmap

pub enum StandardPixmap {
    SPCustomBase,
    SPTitleBarMenuButton,
    SPTitleBarMinButton,
    SPTitleBarMaxButton,
    SPTitleBarCloseButton,
    SPTitleBarNormalButton,
    SPTitleBarShadeButton,
    SPTitleBarUnshadeButton,
    SPTitleBarContextHelpButton,
    SPDockWidgetCloseButton,
    SPMessageBoxInformation,
    SPMessageBoxWarning,
    SPMessageBoxCritical,
    SPMessageBoxQuestion,
    SPDesktopIcon,
    SPTrashIcon,
    SPComputerIcon,
    SPDriveFDIcon,
    SPDriveHDIcon,
    SPDriveCDIcon,
    SPDriveDVDIcon,
    SPDriveNetIcon,
    SPDirOpenIcon,
    SPDirClosedIcon,
    SPDirLinkIcon,
    SPDirLinkOpenIcon,
    SPFileIcon,
    SPFileLinkIcon,
    SPToolBarHorizontalExtensionButton,
    SPToolBarVerticalExtensionButton,
    SPFileDialogStart,
    SPFileDialogEnd,
    SPFileDialogToParent,
    SPFileDialogNewFolder,
    SPFileDialogDetailedView,
    SPFileDialogInfoView,
    SPFileDialogContentsView,
    SPFileDialogListView,
    SPFileDialogBack,
    SPDirIcon,
    SPDialogOkButton,
    SPDialogCancelButton,
    SPDialogHelpButton,
    SPDialogOpenButton,
    SPDialogSaveButton,
    SPDialogCloseButton,
    SPDialogApplyButton,
    SPDialogResetButton,
    SPDialogDiscardButton,
    SPDialogYesButton,
    SPDialogNoButton,
    SPArrowUp,
    SPArrowDown,
    SPArrowLeft,
    SPArrowRight,
    SPArrowBack,
    SPArrowForward,
    SPDirHomeIcon,
    SPCommandLink,
    SPVistaShield,
    SPBrowserReload,
    SPBrowserStop,
    SPMediaPlay,
    SPMediaStop,
    SPMediaPause,
    SPMediaSkipForward,
    SPMediaSkipBackward,
    SPMediaSeekForward,
    SPMediaSeekBackward,
    SPMediaVolume,
    SPMediaVolumeMuted,
    SPLineEditClearButton,
}

C++ type: QStyle::StandardPixmap

C++ documentation:

This enum describes the available standard pixmaps. A standard pixmap is a pixmap that can follow some existing GUI style or guideline.

See also standardIcon().

Variants
SPCustomBase

Base value for custom standard pixmaps; custom values must be greater than this value. (C++ enum variant: SP_CustomBase = -268435456)

SPTitleBarMenuButton

Menu button on a title bar. (C++ enum variant: SP_TitleBarMenuButton = 0)

SPTitleBarMinButton

Minimize button on title bars (e.g., in QMdiSubWindow). (C++ enum variant: SP_TitleBarMinButton = 1)

SPTitleBarMaxButton

Maximize button on title bars. (C++ enum variant: SP_TitleBarMaxButton = 2)

SPTitleBarCloseButton

Close button on title bars. (C++ enum variant: SP_TitleBarCloseButton = 3)

SPTitleBarNormalButton

Normal (restore) button on title bars. (C++ enum variant: SP_TitleBarNormalButton = 4)

SPTitleBarShadeButton

Shade button on title bars. (C++ enum variant: SP_TitleBarShadeButton = 5)

SPTitleBarUnshadeButton

Unshade button on title bars. (C++ enum variant: SP_TitleBarUnshadeButton = 6)

SPTitleBarContextHelpButton

The Context help button on title bars. (C++ enum variant: SP_TitleBarContextHelpButton = 7)

SPDockWidgetCloseButton

Close button on dock windows (see also QDockWidget). (C++ enum variant: SP_DockWidgetCloseButton = 8)

SPMessageBoxInformation

The "information" icon. (C++ enum variant: SP_MessageBoxInformation = 9)

SPMessageBoxWarning

The "warning" icon. (C++ enum variant: SP_MessageBoxWarning = 10)

SPMessageBoxCritical

The "critical" icon. (C++ enum variant: SP_MessageBoxCritical = 11)

SPMessageBoxQuestion

The "question" icon. (C++ enum variant: SP_MessageBoxQuestion = 12)

SPDesktopIcon

The "desktop" icon. (C++ enum variant: SP_DesktopIcon = 13)

SPTrashIcon

The "trash" icon. (C++ enum variant: SP_TrashIcon = 14)

SPComputerIcon

The "My computer" icon. (C++ enum variant: SP_ComputerIcon = 15)

SPDriveFDIcon

The floppy icon. (C++ enum variant: SP_DriveFDIcon = 16)

SPDriveHDIcon

The harddrive icon. (C++ enum variant: SP_DriveHDIcon = 17)

SPDriveCDIcon

The CD icon. (C++ enum variant: SP_DriveCDIcon = 18)

SPDriveDVDIcon

The DVD icon. (C++ enum variant: SP_DriveDVDIcon = 19)

SPDriveNetIcon

The network icon. (C++ enum variant: SP_DriveNetIcon = 20)

SPDirOpenIcon

The open directory icon. (C++ enum variant: SP_DirOpenIcon = 21)

SPDirClosedIcon

The closed directory icon. (C++ enum variant: SP_DirClosedIcon = 22)

SPDirLinkIcon

The link to directory icon. (C++ enum variant: SP_DirLinkIcon = 23)

SPDirLinkOpenIcon

The link to open directory icon. (C++ enum variant: SP_DirLinkOpenIcon = 24)

SPFileIcon

The file icon. (C++ enum variant: SP_FileIcon = 25)

SPFileLinkIcon

The link to file icon. (C++ enum variant: SP_FileLinkIcon = 26)

SPToolBarHorizontalExtensionButton

Extension button for horizontal toolbars. (C++ enum variant: SP_ToolBarHorizontalExtensionButton = 27)

SPToolBarVerticalExtensionButton

Extension button for vertical toolbars. (C++ enum variant: SP_ToolBarVerticalExtensionButton = 28)

SPFileDialogStart

The "start" icon in a file dialog. (C++ enum variant: SP_FileDialogStart = 29)

SPFileDialogEnd

The "end" icon in a file dialog. (C++ enum variant: SP_FileDialogEnd = 30)

SPFileDialogToParent

The "parent directory" icon in a file dialog. (C++ enum variant: SP_FileDialogToParent = 31)

SPFileDialogNewFolder

The "create new folder" icon in a file dialog. (C++ enum variant: SP_FileDialogNewFolder = 32)

SPFileDialogDetailedView

The detailed view icon in a file dialog. (C++ enum variant: SP_FileDialogDetailedView = 33)

SPFileDialogInfoView

The file info icon in a file dialog. (C++ enum variant: SP_FileDialogInfoView = 34)

SPFileDialogContentsView

The contents view icon in a file dialog. (C++ enum variant: SP_FileDialogContentsView = 35)

SPFileDialogListView

The list view icon in a file dialog. (C++ enum variant: SP_FileDialogListView = 36)

SPFileDialogBack

The back arrow in a file dialog. (C++ enum variant: SP_FileDialogBack = 37)

SPDirIcon

The directory icon. (C++ enum variant: SP_DirIcon = 38)

SPDialogOkButton

Icon for a standard OK button in a QDialogButtonBox. (C++ enum variant: SP_DialogOkButton = 39)

SPDialogCancelButton

Icon for a standard Cancel button in a QDialogButtonBox. (C++ enum variant: SP_DialogCancelButton = 40)

SPDialogHelpButton

Icon for a standard Help button in a QDialogButtonBox. (C++ enum variant: SP_DialogHelpButton = 41)

SPDialogOpenButton

Icon for a standard Open button in a QDialogButtonBox. (C++ enum variant: SP_DialogOpenButton = 42)

SPDialogSaveButton

Icon for a standard Save button in a QDialogButtonBox. (C++ enum variant: SP_DialogSaveButton = 43)

SPDialogCloseButton

Icon for a standard Close button in a QDialogButtonBox. (C++ enum variant: SP_DialogCloseButton = 44)

SPDialogApplyButton

Icon for a standard Apply button in a QDialogButtonBox. (C++ enum variant: SP_DialogApplyButton = 45)

SPDialogResetButton

Icon for a standard Reset button in a QDialogButtonBox. (C++ enum variant: SP_DialogResetButton = 46)

SPDialogDiscardButton

Icon for a standard Discard button in a QDialogButtonBox. (C++ enum variant: SP_DialogDiscardButton = 47)

SPDialogYesButton

Icon for a standard Yes button in a QDialogButtonBox. (C++ enum variant: SP_DialogYesButton = 48)

SPDialogNoButton

Icon for a standard No button in a QDialogButtonBox. (C++ enum variant: SP_DialogNoButton = 49)

SPArrowUp

Icon arrow pointing up. (C++ enum variant: SP_ArrowUp = 50)

SPArrowDown

Icon arrow pointing down. (C++ enum variant: SP_ArrowDown = 51)

SPArrowLeft

Icon arrow pointing left. (C++ enum variant: SP_ArrowLeft = 52)

SPArrowRight

Icon arrow pointing right. (C++ enum variant: SP_ArrowRight = 53)

SPArrowBack

Equivalent to SP_ArrowLeft when the current layout direction is Qt::LeftToRight, otherwise SP_ArrowRight. (C++ enum variant: SP_ArrowBack = 54)

SPArrowForward

Equivalent to SP_ArrowRight when the current layout direction is Qt::LeftToRight, otherwise SP_ArrowLeft. (C++ enum variant: SP_ArrowForward = 55)

SPDirHomeIcon

The home directory icon. (C++ enum variant: SP_DirHomeIcon = 56)

SPCommandLink

Icon used to indicate a Vista style command link glyph. (C++ enum variant: SP_CommandLink = 57)

SPVistaShield

Icon used to indicate UAC prompts on Windows Vista. This will return a null pixmap or icon on all other platforms. (C++ enum variant: SP_VistaShield = 58)

SPBrowserReload

Icon indicating that the current page should be reloaded. (C++ enum variant: SP_BrowserReload = 59)

SPBrowserStop

Icon indicating that the page loading should stop. (C++ enum variant: SP_BrowserStop = 60)

SPMediaPlay

Icon indicating that media should begin playback. (C++ enum variant: SP_MediaPlay = 61)

SPMediaStop

Icon indicating that media should stop playback. (C++ enum variant: SP_MediaStop = 62)

SPMediaPause

Icon indicating that media should pause playback. (C++ enum variant: SP_MediaPause = 63)

SPMediaSkipForward

Icon indicating that media should skip forward. (C++ enum variant: SP_MediaSkipForward = 64)

SPMediaSkipBackward

Icon indicating that media should skip backward. (C++ enum variant: SP_MediaSkipBackward = 65)

SPMediaSeekForward

Icon indicating that media should seek forward. (C++ enum variant: SP_MediaSeekForward = 66)

SPMediaSeekBackward

Icon indicating that media should seek backward. (C++ enum variant: SP_MediaSeekBackward = 67)

SPMediaVolume

Icon indicating a volume control. (C++ enum variant: SP_MediaVolume = 68)

SPMediaVolumeMuted

Icon indicating a muted volume control. (C++ enum variant: SP_MediaVolumeMuted = 69)

SPLineEditClearButton

Icon for a standard clear button in a QLineEdit. This enum value was added in Qt 5.2. (C++ enum variant: SP_LineEditClearButton = 70)