Enum qt_widgets::style::StateFlag

pub enum StateFlag {
    None,
    Enabled,
    Raised,
    Sunken,
    Off,
    NoChange,
    On,
    DownArrow,
    Horizontal,
    HasFocus,
    Top,
    Bottom,
    FocusAtBorder,
    AutoRaise,
    MouseOver,
    UpArrow,
    Selected,
    Active,
    Window,
    Open,
    Children,
    Item,
    Sibling,
    Editing,
    KeyboardFocusChange,
    ReadOnly,
    Small,
    Mini,
}

C++ type: QStyle::StateFlag

C++ documentation:

This enum describes flags that are used when drawing primitive elements.

Note that not all primitives use all of these flags, and that the flags may mean different things to different items.

The State type is a typedef for QFlags<StateFlag>. It stores an OR combination of StateFlag values.

See also drawPrimitive().

Variants
None

Indicates that the widget does not have a state. (C++ enum variant: State_None = 0)

Enabled

Used to indicate if the widget is enabled. (C++ enum variant: State_Enabled = 1)

Raised

Used to indicate if a button is raised. (C++ enum variant: State_Raised = 2)

Sunken

Used to indicate if the widget is sunken or pressed. (C++ enum variant: State_Sunken = 4)

Off

Used to indicate if the widget is not checked. (C++ enum variant: State_Off = 8)

NoChange

Used to indicate a tri-state checkbox. (C++ enum variant: State_NoChange = 16)

On

Used to indicate if the widget is checked. (C++ enum variant: State_On = 32)

DownArrow

Used to indicate if a down arrow should be visible on the widget. (C++ enum variant: State_DownArrow = 64)

Horizontal

Used to indicate if the widget is laid out horizontally, for example. a tool bar. (C++ enum variant: State_Horizontal = 128)

HasFocus

Used to indicate if the widget has focus. (C++ enum variant: State_HasFocus = 256)

Top

C++ enum variant: State_Top = 512

Bottom

C++ enum variant: State_Bottom = 1024

FocusAtBorder

C++ enum variant: State_FocusAtBorder = 2048

AutoRaise

Used to indicate if auto-raise appearance should be usd on a tool button. (C++ enum variant: State_AutoRaise = 4096)

MouseOver

Used to indicate if the widget is under the mouse. (C++ enum variant: State_MouseOver = 8192)

UpArrow

Used to indicate if an up arrow should be visible on the widget. (C++ enum variant: State_UpArrow = 16384)

Selected

Used to indicate if a widget is selected. (C++ enum variant: State_Selected = 32768)

Active

Indicates that the widget is active. (C++ enum variant: State_Active = 65536)

Window

C++ enum variant: State_Window = 131072

Open

Used by item views to indicate if the tree branch is open. (C++ enum variant: State_Open = 262144)

Children

Used to indicate if an item view branch has children. (C++ enum variant: State_Children = 524288)

Item

Used by item views to indicate if a horizontal branch should be drawn. (C++ enum variant: State_Item = 1048576)

Sibling

Used by item views to indicate if a vertical line needs to be drawn (for siblings). (C++ enum variant: State_Sibling = 2097152)

Editing

Used to indicate if an editor is opened on the widget. (C++ enum variant: State_Editing = 4194304)

KeyboardFocusChange

Used to indicate if the focus was changed with the keyboard, e.g., tab, backtab or shortcut. (C++ enum variant: State_KeyboardFocusChange = 8388608)

ReadOnly

Used to indicate if a widget is read-only. (C++ enum variant: State_ReadOnly = 33554432)

Small

Used to indicate a small style Mac widget or button. (C++ enum variant: State_Small = 67108864)

Mini

Used to indicate a mini style Mac widget or button. (C++ enum variant: State_Mini = 134217728)