Enum qt_widgets::style::StyleHint

pub enum StyleHint {
    SHCustomBase,
    SHEtchDisabledText,
    SHDitherDisabledText,
    SHScrollBarMiddleClickAbsolutePosition,
    SHScrollBarScrollWhenPointerLeavesControl,
    SHTabBarSelectMouseType,
    SHTabBarAlignment,
    SHHeaderArrowAlignment,
    SHSliderSnapToValue,
    SHSliderSloppyKeyEvents,
    SHProgressDialogCenterCancelButton,
    SHProgressDialogTextLabelAlignment,
    SHPrintDialogRightAlignButtons,
    SHMainWindowSpaceBelowMenuBar,
    SHFontDialogSelectAssociatedText,
    SHMenuAllowActiveAndDisabled,
    SHMenuSpaceActivatesItem,
    SHMenuSubMenuPopupDelay,
    SHScrollViewFrameOnlyAroundContents,
    SHMenuBarAltKeyNavigation,
    SHComboBoxListMouseTracking,
    SHMenuMouseTracking,
    SHMenuBarMouseTracking,
    SHItemViewChangeHighlightOnFocus,
    SHWidgetShareActivation,
    SHWorkspaceFillSpaceOnMaximize,
    SHComboBoxPopup,
    SHTitleBarNoBorder,
    SHSliderStopMouseOverSlider,
    SHBlinkCursorWhenTextSelected,
    SHRichTextFullWidthSelection,
    SHMenuScrollable,
    SHGroupBoxTextLabelVerticalAlignment,
    SHGroupBoxTextLabelColor,
    SHMenuSloppySubMenus,
    SHTableGridLineColor,
    SHLineEditPasswordCharacter,
    SHDialogButtonsDefaultButton,
    SHToolBoxSelectedPageTitleBold,
    SHTabBarPreferNoArrows,
    SHScrollBarLeftClickAbsolutePosition,
    SHListViewExpandSelectMouseType,
    SHUnderlineShortcut,
    SHSpinBoxAnimateButton,
    SHSpinBoxKeyPressAutoRepeatRate,
    SHSpinBoxClickAutoRepeatRate,
    SHMenuFillScreenWithScroll,
    SHToolTipLabelOpacity,
    SHDrawMenuBarSeparator,
    SHTitleBarModifyNotification,
    SHButtonFocusPolicy,
    SHMessageBoxUseBorderForButtonSpacing,
    SHTitleBarAutoRaise,
    SHToolButtonPopupDelay,
    SHFocusFrameMask,
    SHRubberBandMask,
    SHWindowFrameMask,
    SHSpinControlsDisableOnBounds,
    SHDialBackgroundRole,
    SHComboBoxLayoutDirection,
    SHItemViewEllipsisLocation,
    SHItemViewShowDecorationSelected,
    SHItemViewActivateItemOnSingleClick,
    SHScrollBarContextMenu,
    SHScrollBarRollBetweenButtons,
    SHSliderAbsoluteSetButtons,
    SHSliderPageSetButtons,
    SHMenuKeyboardSearch,
    SHTabBarElideMode,
    SHDialogButtonLayout,
    SHComboBoxPopupFrameStyle,
    SHMessageBoxTextInteractionFlags,
    SHDialogButtonBoxButtonsHaveIcons,
    SHSpellCheckUnderlineStyle,
    SHMessageBoxCenterButtons,
    SHMenuSelectionWrap,
    SHItemViewMovementWithoutUpdatingSelection,
    SHToolTipMask,
    SHFocusFrameAboveWidget,
    SHTextControlFocusIndicatorTextCharFormat,
    SHWizardStyle,
    SHItemViewArrowKeysNavigateIntoChildren,
    SHMenuMask,
    SHMenuFlashTriggeredItem,
    SHMenuFadeOutOnHide,
    SHSpinBoxClickAutoRepeatThreshold,
    SHItemViewPaintAlternatingRowColorsForEmptyArea,
    SHFormLayoutWrapPolicy,
    SHTabWidgetDefaultTabPosition,
    SHToolBarMovable,
    SHFormLayoutFieldGrowthPolicy,
    SHFormLayoutFormAlignment,
    SHFormLayoutLabelAlignment,
    SHItemViewDrawDelegateFrame,
    SHTabBarCloseButtonPosition,
    SHDockWidgetButtonsHaveFrame,
    SHToolButtonStyle,
    SHRequestSoftwareInputPanel,
    SHScrollBarTransient,
    SHMenuSupportsSections,
    SHToolTipWakeUpDelay,
    SHToolTipFallAsleepDelay,
    SHWidgetAnimate,
    SHSplitterOpaqueResize,
    SHComboBoxUseNativePopup,
    SHLineEditPasswordMaskDelay,
    SHTabBarChangeCurrentDelay,
    SHMenuSubMenuUniDirection,
    SHMenuSubMenuUniDirectionFailCount,
    SHMenuSubMenuSloppySelectOtherActions,
    SHMenuSubMenuSloppyCloseTimeout,
    SHMenuSubMenuResetWhenReenteringParent,
    SHMenuSubMenuDontStartSloppyOnLeave,
}

C++ type: QStyle::StyleHint

C++ documentation:

This enum describes the available style hints. A style hint is a general look and/or feel hint.

See also styleHint().

Variants
SHCustomBase

Base value for custom style hints. Custom values must be greater than this value. (C++ enum variant: SH_CustomBase = -268435456)

SHEtchDisabledText

Disabled text is "etched" as it is on Windows. (C++ enum variant: SH_EtchDisabledText = 0)

SHDitherDisabledText

Disabled text is dithered as it is on Motif. (C++ enum variant: SH_DitherDisabledText = 1)

SHScrollBarMiddleClickAbsolutePosition

A boolean value. If true, middle clicking on a scroll bar causes the slider to jump to that position. If false, middle clicking is ignored. (C++ enum variant: SH_ScrollBar_MiddleClickAbsolutePosition = 2)

SHScrollBarScrollWhenPointerLeavesControl

A boolean value. If true, when clicking a scroll bar SubControl, holding the mouse button down and moving the pointer outside the SubControl, the scroll bar continues to scroll. If false, the scollbar stops scrolling when the pointer leaves the SubControl. (C++ enum variant: SH_ScrollBar_ScrollWhenPointerLeavesControl = 3)

SHTabBarSelectMouseType

Which type of mouse event should cause a tab to be selected. (C++ enum variant: SH_TabBar_SelectMouseType = 4)

SHTabBarAlignment

The alignment for tabs in a QTabWidget. Possible values are Qt::AlignLeft, Qt::AlignCenter and Qt::AlignRight. (C++ enum variant: SH_TabBar_Alignment = 5)

SHHeaderArrowAlignment

The placement of the sorting indicator may appear in list or table headers. Possible values are Qt::Alignment values (that is, an OR combination of Qt::AlignmentFlag flags). (C++ enum variant: SH_Header_ArrowAlignment = 6)

SHSliderSnapToValue

Sliders snap to values while moving, as they do on Windows. (C++ enum variant: SH_Slider_SnapToValue = 7)

SHSliderSloppyKeyEvents

Key presses handled in a sloppy manner, i.e., left on a vertical slider subtracts a line. (C++ enum variant: SH_Slider_SloppyKeyEvents = 8)

SHProgressDialogCenterCancelButton

Center button on progress dialogs, otherwise right aligned. (C++ enum variant: SH_ProgressDialog_CenterCancelButton = 9)

SHProgressDialogTextLabelAlignment

The alignment for text labels in progress dialogs; Qt::AlignCenter on Windows, Qt::AlignVCenter otherwise. (C++ enum variant: SH_ProgressDialog_TextLabelAlignment = 10)

SHPrintDialogRightAlignButtons

Right align buttons in the print dialog, as done on Windows. (C++ enum variant: SH_PrintDialog_RightAlignButtons = 11)

SHMainWindowSpaceBelowMenuBar

One or two pixel space between the menu bar and the dockarea, as done on Windows. (C++ enum variant: SH_MainWindow_SpaceBelowMenuBar = 12)

SHFontDialogSelectAssociatedText

Select the text in the line edit, or when selecting an item from the listbox, or when the line edit receives focus, as done on Windows. (C++ enum variant: SH_FontDialog_SelectAssociatedText = 13)

SHMenuAllowActiveAndDisabled

Allows disabled menu items to be active. (C++ enum variant: SH_Menu_AllowActiveAndDisabled = 14)

SHMenuSpaceActivatesItem

Pressing the space bar activates the item, as done on Motif. (C++ enum variant: SH_Menu_SpaceActivatesItem = 15)

SHMenuSubMenuPopupDelay

The number of milliseconds to wait before opening a submenu (256 on Windows, 96 on Motif). (C++ enum variant: SH_Menu_SubMenuPopupDelay = 16)

SHScrollViewFrameOnlyAroundContents

Whether scrollviews draw their frame only around contents (like Motif), or around contents, scroll bars and corner widgets (like Windows). (C++ enum variant: SH_ScrollView_FrameOnlyAroundContents = 17)

SHMenuBarAltKeyNavigation

Menu bars items are navigable by pressing Alt, followed by using the arrow keys to select the desired item. (C++ enum variant: SH_MenuBar_AltKeyNavigation = 18)

SHComboBoxListMouseTracking

Mouse tracking in combobox drop-down lists. (C++ enum variant: SH_ComboBox_ListMouseTracking = 19)

SHMenuMouseTracking

Mouse tracking in popup menus. (C++ enum variant: SH_Menu_MouseTracking = 20)

SHMenuBarMouseTracking

Mouse tracking in menu bars. (C++ enum variant: SH_MenuBar_MouseTracking = 21)

SHItemViewChangeHighlightOnFocus

Gray out selected items when losing focus. (C++ enum variant: SH_ItemView_ChangeHighlightOnFocus = 22)

SHWidgetShareActivation

Turn on sharing activation with floating modeless dialogs. (C++ enum variant: SH_Widget_ShareActivation = 23)

SHWorkspaceFillSpaceOnMaximize

The workspace should maximize the client area. (C++ enum variant: SH_Workspace_FillSpaceOnMaximize = 24)

SHComboBoxPopup

Allows popups as a combobox drop-down menu. (C++ enum variant: SH_ComboBox_Popup = 25)

SHTitleBarNoBorder

The title bar has no border. (C++ enum variant: SH_TitleBar_NoBorder = 26)

SHSliderStopMouseOverSlider

This variant corresponds to multiple C++ enum variants with the same value:

SH_Slider_StopMouseOverSlider = 27: Stops auto-repeat when the slider reaches the mouse position.

SH_ScrollBar_StopMouseOverSlider = 27: Obsolete. Use SH_Slider_StopMouseOverSlider instead.

SHBlinkCursorWhenTextSelected

Whether cursor should blink when text is selected. (C++ enum variant: SH_BlinkCursorWhenTextSelected = 28)

SHRichTextFullWidthSelection

Whether richtext selections should extend to the full width of the document. (C++ enum variant: SH_RichText_FullWidthSelection = 29)

SHMenuScrollable

Whether popup menus must support scrolling. (C++ enum variant: SH_Menu_Scrollable = 30)

SHGroupBoxTextLabelVerticalAlignment

How to vertically align a group box's text label. (C++ enum variant: SH_GroupBox_TextLabelVerticalAlignment = 31)

SHGroupBoxTextLabelColor

How to paint a group box's text label. (C++ enum variant: SH_GroupBox_TextLabelColor = 32)

SHMenuSloppySubMenus

Whether popup menus must support the user moving the mouse cursor to a submenu while crossing other items of the menu. This is supported on most modern desktop platforms. (C++ enum variant: SH_Menu_SloppySubMenus = 33)

SHTableGridLineColor

The RGB value of the grid for a table. (C++ enum variant: SH_Table_GridLineColor = 34)

SHLineEditPasswordCharacter

The Unicode character to be used for passwords. (C++ enum variant: SH_LineEdit_PasswordCharacter = 35)

SHDialogButtonsDefaultButton

Which button gets the default status in a dialog's button widget. (C++ enum variant: SH_DialogButtons_DefaultButton = 36)

SHToolBoxSelectedPageTitleBold

Boldness of the selected page title in a QToolBox. (C++ enum variant: SH_ToolBox_SelectedPageTitleBold = 37)

SHTabBarPreferNoArrows

Whether a tab bar should suggest a size to prevent scoll arrows. (C++ enum variant: SH_TabBar_PreferNoArrows = 38)

SHScrollBarLeftClickAbsolutePosition

A boolean value. If true, left clicking on a scroll bar causes the slider to jump to that position. If false, left clicking will behave as appropriate for each control. (C++ enum variant: SH_ScrollBar_LeftClickAbsolutePosition = 39)

SHListViewExpandSelectMouseType

Which type of mouse event should cause a list view expansion to be selected. (C++ enum variant: SH_ListViewExpand_SelectMouseType = 40)

SHUnderlineShortcut

Whether shortcuts are underlined. (C++ enum variant: SH_UnderlineShortcut = 41)

SHSpinBoxAnimateButton

Animate a click when up or down is pressed in a spin box. (C++ enum variant: SH_SpinBox_AnimateButton = 42)

SHSpinBoxKeyPressAutoRepeatRate

Auto-repeat interval for spinbox key presses. (C++ enum variant: SH_SpinBox_KeyPressAutoRepeatRate = 43)

SHSpinBoxClickAutoRepeatRate

Auto-repeat interval for spinbox mouse clicks. (C++ enum variant: SH_SpinBox_ClickAutoRepeatRate = 44)

SHMenuFillScreenWithScroll

Whether scrolling popups should fill the screen as they are scrolled. (C++ enum variant: SH_Menu_FillScreenWithScroll = 45)

SHToolTipLabelOpacity

An integer indicating the opacity for the tip label, 0 is completely transparent, 255 is completely opaque. (C++ enum variant: SH_ToolTipLabel_Opacity = 46)

SHDrawMenuBarSeparator

Indicates whether or not the menu bar draws separators. (C++ enum variant: SH_DrawMenuBarSeparator = 47)

SHTitleBarModifyNotification

Indicates if the title bar should show a '*' for windows that are modified. (C++ enum variant: SH_TitleBar_ModifyNotification = 48)

SHButtonFocusPolicy

The default focus policy for buttons. (C++ enum variant: SH_Button_FocusPolicy = 49)

SHMessageBoxUseBorderForButtonSpacing

A boolean indicating what the to use the border of the buttons (computed as half the button height) for the spacing of the button in a message box. (C++ enum variant: SH_MessageBox_UseBorderForButtonSpacing = 50)

SHTitleBarAutoRaise

A boolean indicating whether controls on a title bar ought to update when the mouse is over them. (C++ enum variant: SH_TitleBar_AutoRaise = 51)

SHToolButtonPopupDelay

An int indicating the popup delay in milliseconds for menus attached to tool buttons. (C++ enum variant: SH_ToolButton_PopupDelay = 52)

SHFocusFrameMask

The mask of the focus frame. (C++ enum variant: SH_FocusFrame_Mask = 53)

SHRubberBandMask

The mask of the rubber band. (C++ enum variant: SH_RubberBand_Mask = 54)

SHWindowFrameMask

The mask of the window frame. (C++ enum variant: SH_WindowFrame_Mask = 55)

SHSpinControlsDisableOnBounds

Determines if the spin controls will shown as disabled when reaching the spin range boundary. (C++ enum variant: SH_SpinControls_DisableOnBounds = 56)

SHDialBackgroundRole

Defines the style's preferred background role (as QPalette::ColorRole) for a dial widget. (C++ enum variant: SH_Dial_BackgroundRole = 57)

SHComboBoxLayoutDirection

The layout direction for the combo box. By default it should be the same as indicated by the QStyleOption::direction variable. (C++ enum variant: SH_ComboBox_LayoutDirection = 58)

SHItemViewEllipsisLocation

The location where ellipses should be added for item text that is too long to fit in an view item. (C++ enum variant: SH_ItemView_EllipsisLocation = 59)

SHItemViewShowDecorationSelected

When an item in an item view is selected, also highlight the branch or other decoration. (C++ enum variant: SH_ItemView_ShowDecorationSelected = 60)

SHItemViewActivateItemOnSingleClick

Emit the activated signal when the user single clicks on an item in an item in an item view. Otherwise the signal is emitted when the user double clicks on an item. (C++ enum variant: SH_ItemView_ActivateItemOnSingleClick = 61)

SHScrollBarContextMenu

Whether or not a scroll bar has a context menu. (C++ enum variant: SH_ScrollBar_ContextMenu = 62)

SHScrollBarRollBetweenButtons

A boolean value. If true, when clicking a scroll bar button (SC_ScrollBarAddLine or SC_ScrollBarSubLine) and dragging over to the opposite button (rolling) will press the new button and release the old one. When it is false, the original button is released and nothing happens (like a push button). (C++ enum variant: SH_ScrollBar_RollBetweenButtons = 63)

SHSliderAbsoluteSetButtons

Which mouse buttons cause a slider to set the value to the position clicked on. (C++ enum variant: SH_Slider_AbsoluteSetButtons = 64)

SHSliderPageSetButtons

Which mouse buttons cause a slider to page step the value. (C++ enum variant: SH_Slider_PageSetButtons = 65)

SHMenuKeyboardSearch

Typing causes a menu to be search for relevant items, otherwise only mnemnonic is considered. (C++ enum variant: SH_Menu_KeyboardSearch = 66)

SHTabBarElideMode

The default eliding style for a tab bar. (C++ enum variant: SH_TabBar_ElideMode = 67)

SHDialogButtonLayout

Controls how buttons are laid out in a QDialogButtonBox, returns a QDialogButtonBox::ButtonLayout enum. (C++ enum variant: SH_DialogButtonLayout = 68)

SHComboBoxPopupFrameStyle

The frame style used when drawing a combobox popup menu. (C++ enum variant: SH_ComboBox_PopupFrameStyle = 69)

SHMessageBoxTextInteractionFlags

A boolean indicating if the text in a message box should allow user interfactions (e.g. selection) or not. (C++ enum variant: SH_MessageBox_TextInteractionFlags = 70)

SHDialogButtonBoxButtonsHaveIcons

Indicates whether or not StandardButtons in QDialogButtonBox should have icons or not. (C++ enum variant: SH_DialogButtonBox_ButtonsHaveIcons = 71)

SHSpellCheckUnderlineStyle

A QTextCharFormat::UnderlineStyle value that specifies the way misspelled words should be underlined. (C++ enum variant: SH_SpellCheckUnderlineStyle = 72)

SHMessageBoxCenterButtons

A boolean indicating whether the buttons in the message box should be centered or not (see QDialogButtonBox::setCentered()). (C++ enum variant: SH_MessageBox_CenterButtons = 73)

SHMenuSelectionWrap

Whether popups should allow the selections to wrap, that is when selection should the next item be the first item. (C++ enum variant: SH_Menu_SelectionWrap = 74)

SHItemViewMovementWithoutUpdatingSelection

The item view is able to indicate a current item without changing the selection. (C++ enum variant: SH_ItemView_MovementWithoutUpdatingSelection = 75)

SHToolTipMask

The mask of a tool tip. (C++ enum variant: SH_ToolTip_Mask = 76)

SHFocusFrameAboveWidget

The FocusFrame is stacked above the widget that it is "focusing on". (C++ enum variant: SH_FocusFrame_AboveWidget = 77)

SHTextControlFocusIndicatorTextCharFormat

Specifies the text format used to highlight focused anchors in rich text documents displayed for example in QTextBrowser. The format has to be a QTextCharFormat returned in the variant of the QStyleHintReturnVariant return value. The QTextFormat::OutlinePen property is used for the outline and QTextFormat::BackgroundBrush for the background of the highlighted area. (C++ enum variant: SH_TextControl_FocusIndicatorTextCharFormat = 78)

SHWizardStyle

Controls the look and feel of a QWizard. Returns a QWizard::WizardStyle enum. (C++ enum variant: SH_WizardStyle = 79)

SHItemViewArrowKeysNavigateIntoChildren

Controls whether the tree view will select the first child when it is exapanded and the right arrow key is pressed. (C++ enum variant: SH_ItemView_ArrowKeysNavigateIntoChildren = 80)

SHMenuMask

The mask for a popup menu. (C++ enum variant: SH_Menu_Mask = 81)

SHMenuFlashTriggeredItem

Flash triggered item. (C++ enum variant: SH_Menu_FlashTriggeredItem = 82)

SHMenuFadeOutOnHide

Fade out the menu instead of hiding it immediately. (C++ enum variant: SH_Menu_FadeOutOnHide = 83)

SHSpinBoxClickAutoRepeatThreshold

Auto-repeat threshold for spinbox mouse clicks. (C++ enum variant: SH_SpinBox_ClickAutoRepeatThreshold = 84)

SHItemViewPaintAlternatingRowColorsForEmptyArea

Whether QTreeView paints alternating row colors for the area that does not have any items. (C++ enum variant: SH_ItemView_PaintAlternatingRowColorsForEmptyArea = 85)

SHFormLayoutWrapPolicy

Provides a default for how rows are wrapped in a QFormLayout. Returns a QFormLayout::RowWrapPolicy enum. (C++ enum variant: SH_FormLayoutWrapPolicy = 86)

SHTabWidgetDefaultTabPosition

Default position of the tab bar in a tab widget. (C++ enum variant: SH_TabWidget_DefaultTabPosition = 87)

SHToolBarMovable

Determines if the tool bar is movable by default. (C++ enum variant: SH_ToolBar_Movable = 88)

SHFormLayoutFieldGrowthPolicy

Provides a default for how fields can grow in a QFormLayout. Returns a QFormLayout::FieldGrowthPolicy enum. (C++ enum variant: SH_FormLayoutFieldGrowthPolicy = 89)

SHFormLayoutFormAlignment

Provides a default for how a QFormLayout aligns its contents within the available space. Returns a Qt::Alignment enum. (C++ enum variant: SH_FormLayoutFormAlignment = 90)

SHFormLayoutLabelAlignment

Provides a default for how a QFormLayout aligns labels within the available space. Returns a Qt::Alignment enum. (C++ enum variant: SH_FormLayoutLabelAlignment = 91)

SHItemViewDrawDelegateFrame

Determines if there should be a frame for a delegate widget. (C++ enum variant: SH_ItemView_DrawDelegateFrame = 92)

SHTabBarCloseButtonPosition

Determines the position of the close button on a tab in a tab bar. (C++ enum variant: SH_TabBar_CloseButtonPosition = 93)

SHDockWidgetButtonsHaveFrame

Determines if dockwidget buttons should have frames. Default is true. (C++ enum variant: SH_DockWidget_ButtonsHaveFrame = 94)

SHToolButtonStyle

Determines the default system style for tool buttons that uses Qt::ToolButtonFollowStyle. (C++ enum variant: SH_ToolButtonStyle = 95)

SHRequestSoftwareInputPanel

Determines when a software input panel should be requested by input widgets. Returns an enum of type QStyle::RequestSoftwareInputPanel. (C++ enum variant: SH_RequestSoftwareInputPanel = 96)

SHScrollBarTransient

Determines if the style supports transient scroll bars. Transient scroll bars appear when the content is scrolled and disappear when they are no longer needed. (C++ enum variant: SH_ScrollBar_Transient = 97)

SHMenuSupportsSections

Determines if the style displays sections in menus or treat them as plain separators. Sections are separators with a text and icon hint. (C++ enum variant: SH_Menu_SupportsSections = 98)

SHToolTipWakeUpDelay

Determines the delay before a tooltip is shown, in milliseconds. (C++ enum variant: SH_ToolTip_WakeUpDelay = 99)

SHToolTipFallAsleepDelay

Determines the delay (in milliseconds) before a new wake time is needed when a tooltip is shown (notice: shown, not hidden). When a new wake isn't needed, a user-requested tooltip will be shown nearly instantly. (C++ enum variant: SH_ToolTip_FallAsleepDelay = 100)

SHWidgetAnimate

Determines if the widget should show animations or not, for example a transition between checked and unchecked statuses in a checkbox. This enum value has been introduced in Qt 5.2. (C++ enum variant: SH_Widget_Animate = 101)

SHSplitterOpaqueResize

Determines if resizing is opaque This enum value has been introduced in Qt 5.2 (C++ enum variant: SH_Splitter_OpaqueResize = 102)

SHComboBoxUseNativePopup

C++ enum variant: SH_ComboBox_UseNativePopup = 103

SHLineEditPasswordMaskDelay

Determines the delay before visible character is masked with password character, in milliseconds. This enum value was added in Qt 5.4. (C++ enum variant: SH_LineEdit_PasswordMaskDelay = 104)

SHTabBarChangeCurrentDelay

Determines the delay before the current tab is changed while dragging over the tabbar, in milliseconds. This enum value has been introduced in Qt 5.4 (C++ enum variant: SH_TabBar_ChangeCurrentDelay = 105)

SHMenuSubMenuUniDirection

Since Qt 5.5. If the cursor has to move towards the submenu (like it is on OS X), or if the cursor can move in any direction as long as it reaches the submenu before the sloppy timeout. (C++ enum variant: SH_Menu_SubMenuUniDirection = 106)

SHMenuSubMenuUniDirectionFailCount

Since Qt 5.5. When SH_Menu_SubMenuUniDirection is defined this enum defines the number of failed mouse moves before the sloppy submenu is discarded. This can be used to control the "strictness" of the uni direction algorithm. (C++ enum variant: SH_Menu_SubMenuUniDirectionFailCount = 107)

SHMenuSubMenuSloppySelectOtherActions

Since Qt 5.5. Should other action items be selected when the mouse moves towards a sloppy submenu. (C++ enum variant: SH_Menu_SubMenuSloppySelectOtherActions = 108)

SHMenuSubMenuSloppyCloseTimeout

Since Qt 5.5. The timeout used to close sloppy submenus. (C++ enum variant: SH_Menu_SubMenuSloppyCloseTimeout = 109)

SHMenuSubMenuResetWhenReenteringParent

Since Qt 5.5. When entering parent from child submenu, should the sloppy state be reset, effectively closing the child and making the current submenu active. (C++ enum variant: SH_Menu_SubMenuResetWhenReenteringParent = 110)

SHMenuSubMenuDontStartSloppyOnLeave

Since Qt 5.5. Do not start sloppy timers when the mouse leaves a sub-menu. (C++ enum variant: SH_Menu_SubMenuDontStartSloppyOnLeave = 111)