Enum qt_widgets::style::SubControl

pub enum SubControl {
    SCCustomBase,
    SCAll,
    SCNone,
    SCScrollBarAddLine,
    SCScrollBarSubLine,
    SCScrollBarAddPage,
    SCScrollBarSubPage,
    SCScrollBarFirst,
    SCScrollBarLast,
    SCScrollBarSlider,
    SCScrollBarGroove,
    SCTitleBarLabel,
}

C++ type: QStyle::SubControl

C++ documentation:

This enum describes the available sub controls. A subcontrol is a control element within a complex control (ComplexControl).

The SubControls type is a typedef for QFlags<SubControl>. It stores an OR combination of SubControl values.

See also ComplexControl.

Variants
SCCustomBase

C++ enum variant: SC_CustomBase = -268435456

SCAll

Special value that matches all sub-controls. (C++ enum variant: SC_All = -1)

SCNone

Special value that matches no other sub control. (C++ enum variant: SC_None = 0)

SCScrollBarAddLine

This variant corresponds to multiple C++ enum variants with the same value:

SC_ScrollBarAddLine = 1: Scroll bar add line (i.e., down/right arrow); see also QScrollBar.

SC_SpinBoxUp = 1: Spin widget up/increase; see also QSpinBox.

SC_ComboBoxFrame = 1: Combobox frame.

SC_SliderGroove = 1: Special sub-control which contains the area in which the slider handle may move.

SC_ToolButton = 1: Tool button (see also QToolButton).

SC_TitleBarSysMenu = 1: System menu button (i.e., restore, close, etc.).

SC_DialGroove = 1: The groove for the dial.

SC_GroupBoxCheckBox = 1: The optional check box of a group box.

SC_MdiMinButton = 1: The minimize button for a MDI subwindow in the menu bar.

SCScrollBarSubLine

This variant corresponds to multiple C++ enum variants with the same value:

SC_ScrollBarSubLine = 2: Scroll bar sub line (i.e., up/left arrow).

SC_SpinBoxDown = 2: Spin widget down/decrease.

SC_ComboBoxEditField = 2: Combobox edit field; see also QComboBox.

SC_SliderHandle = 2: Slider handle.

SC_ToolButtonMenu = 2: Sub-control for opening a popup menu in a tool button.

SC_TitleBarMinButton = 2: Minimize button.

SC_DialHandle = 2: The handle of the dial (i.e. what you use to control the dial).

SC_GroupBoxLabel = 2: The title of a group box.

SC_MdiNormalButton = 2: The normal button for a MDI subwindow in the menu bar.

SCScrollBarAddPage

This variant corresponds to multiple C++ enum variants with the same value:

SC_ScrollBarAddPage = 4: Scroll bar add page (i.e., page down).

SC_SpinBoxFrame = 4: Spin widget frame.

SC_ComboBoxArrow = 4: Combobox arrow button.

SC_SliderTickmarks = 4: Slider tickmarks.

SC_TitleBarMaxButton = 4: Maximize button.

SC_DialTickmarks = 4: The tickmarks for the dial.

SC_GroupBoxContents = 4: The group box contents.

SC_MdiCloseButton = 4: The close button for a MDI subwindow in the menu bar.

SCScrollBarSubPage

This variant corresponds to multiple C++ enum variants with the same value:

SC_ScrollBarSubPage = 8: Scroll bar sub page (i.e., page up).

SC_SpinBoxEditField = 8: Spin widget edit field.

SC_ComboBoxListBoxPopup = 8: The reference rectangle for the combobox popup. Used to calculate the position of the popup.

SC_TitleBarCloseButton = 8: Close button.

SC_GroupBoxFrame = 8: The frame of a group box.

SCScrollBarFirst

This variant corresponds to multiple C++ enum variants with the same value:

SC_ScrollBarFirst = 16: Scroll bar first line (i.e., home).

SC_TitleBarNormalButton = 16: Normal (restore) button.

SCScrollBarLast

This variant corresponds to multiple C++ enum variants with the same value:

SC_ScrollBarLast = 32: Scroll bar last line (i.e., end).

SC_TitleBarShadeButton = 32: Shade button.

SCScrollBarSlider

This variant corresponds to multiple C++ enum variants with the same value:

SC_ScrollBarSlider = 64: Scroll bar slider handle.

SC_TitleBarUnshadeButton = 64: Unshade button.

SCScrollBarGroove

This variant corresponds to multiple C++ enum variants with the same value:

SC_ScrollBarGroove = 128: Special sub-control which contains the area in which the slider handle may move.

SC_TitleBarContextHelpButton = 128: Context Help button.

SCTitleBarLabel

Window title label. (C++ enum variant: SC_TitleBarLabel = 256)