Enum qt_widgets::style::SubElement

pub enum SubElement {
    SECustomBase,
    SEPushButtonContents,
    SEPushButtonFocusRect,
    SECheckBoxIndicator,
    SECheckBoxContents,
    SECheckBoxFocusRect,
    SECheckBoxClickRect,
    SERadioButtonIndicator,
    SERadioButtonContents,
    SERadioButtonFocusRect,
    SERadioButtonClickRect,
    SEComboBoxFocusRect,
    SESliderFocusRect,
    SEProgressBarGroove,
    SEProgressBarContents,
    SEProgressBarLabel,
    SEToolBoxTabContents,
    SEHeaderLabel,
    SEHeaderArrow,
    SETabWidgetTabBar,
    SETabWidgetTabPane,
    SETabWidgetTabContents,
    SETabWidgetLeftCorner,
    SETabWidgetRightCorner,
    SEViewItemCheckIndicator,
    SETabBarTearIndicator,
    SETreeViewDisclosureItem,
    SELineEditContents,
    SEFrameContents,
    SEDockWidgetCloseButton,
    SEDockWidgetFloatButton,
    SEDockWidgetTitleBarText,
    SEDockWidgetIcon,
    SECheckBoxLayoutItem,
    SEComboBoxLayoutItem,
    SEDateTimeEditLayoutItem,
    SEDialogButtonBoxLayoutItem,
    SELabelLayoutItem,
    SEProgressBarLayoutItem,
    SEPushButtonLayoutItem,
    SERadioButtonLayoutItem,
    SESliderLayoutItem,
    SESpinBoxLayoutItem,
    SEToolButtonLayoutItem,
    SEFrameLayoutItem,
    SEGroupBoxLayoutItem,
    SETabWidgetLayoutItem,
    SEItemViewItemDecoration,
    SEItemViewItemText,
    SEItemViewItemFocusRect,
    SETabBarTabLeftButton,
    SETabBarTabRightButton,
    SETabBarTabText,
    SEShapedFrameContents,
    SEToolBarHandle,
}

C++ type: QStyle::SubElement

C++ documentation:

This enum represents a sub-area of a widget. Style implementations use these areas to draw the different parts of a widget.

See also subElementRect().

Variants
SECustomBase

Base value for custom sub-elements. Custom values must be greater than this value. (C++ enum variant: SE_CustomBase = -268435456)

SEPushButtonContents

Area containing the label (icon with text or pixmap). (C++ enum variant: SE_PushButtonContents = 0)

SEPushButtonFocusRect

Area for the focus rect (usually larger than the contents rect). (C++ enum variant: SE_PushButtonFocusRect = 1)

SECheckBoxIndicator

Area for the state indicator (e.g., check mark). (C++ enum variant: SE_CheckBoxIndicator = 2)

SECheckBoxContents

Area for the label (text or pixmap). (C++ enum variant: SE_CheckBoxContents = 3)

SECheckBoxFocusRect

Area for the focus indicator. (C++ enum variant: SE_CheckBoxFocusRect = 4)

SECheckBoxClickRect

Clickable area, defaults to SE_CheckBoxFocusRect. (C++ enum variant: SE_CheckBoxClickRect = 5)

SERadioButtonIndicator

Area for the state indicator. (C++ enum variant: SE_RadioButtonIndicator = 6)

SERadioButtonContents

Area for the label. (C++ enum variant: SE_RadioButtonContents = 7)

SERadioButtonFocusRect

Area for the focus indicator. (C++ enum variant: SE_RadioButtonFocusRect = 8)

SERadioButtonClickRect

Clickable area, defaults to SE_RadioButtonFocusRect. (C++ enum variant: SE_RadioButtonClickRect = 9)

SEComboBoxFocusRect

Area for the focus indicator. (C++ enum variant: SE_ComboBoxFocusRect = 10)

SESliderFocusRect

Area for the focus indicator. (C++ enum variant: SE_SliderFocusRect = 11)

SEProgressBarGroove

Area for the groove. (C++ enum variant: SE_ProgressBarGroove = 12)

SEProgressBarContents

Area for the progress indicator. (C++ enum variant: SE_ProgressBarContents = 13)

SEProgressBarLabel

Area for the text label. (C++ enum variant: SE_ProgressBarLabel = 14)

SEToolBoxTabContents

Area for a toolbox tab's icon and label. (C++ enum variant: SE_ToolBoxTabContents = 15)

SEHeaderLabel

Area for the label in a header. (C++ enum variant: SE_HeaderLabel = 16)

SEHeaderArrow

Area for the sort indicator for a header. (C++ enum variant: SE_HeaderArrow = 17)

SETabWidgetTabBar

Area for the tab bar widget in a tab widget. (C++ enum variant: SE_TabWidgetTabBar = 18)

SETabWidgetTabPane

Area for the pane of a tab widget. (C++ enum variant: SE_TabWidgetTabPane = 19)

SETabWidgetTabContents

Area for the contents of the tab widget. (C++ enum variant: SE_TabWidgetTabContents = 20)

SETabWidgetLeftCorner

Area for the left corner widget in a tab widget. (C++ enum variant: SE_TabWidgetLeftCorner = 21)

SETabWidgetRightCorner

Area for the right corner widget in a tab widget. (C++ enum variant: SE_TabWidgetRightCorner = 22)

SEViewItemCheckIndicator

This variant corresponds to multiple C++ enum variants with the same value:

SE_ViewItemCheckIndicator = 23

SE_ItemViewItemCheckIndicator = 23: Area for a view item's check mark.

SETabBarTearIndicator

Area for the tear indicator on a tab bar with scroll arrows. (C++ enum variant: SE_TabBarTearIndicator = 24)

SETreeViewDisclosureItem

Area for the actual disclosure item in a tree branch. (C++ enum variant: SE_TreeViewDisclosureItem = 25)

SELineEditContents

Area for a line edit's contents. (C++ enum variant: SE_LineEditContents = 26)

SEFrameContents

Area for a frame's contents. (C++ enum variant: SE_FrameContents = 27)

SEDockWidgetCloseButton

The close button of a dock widget. (C++ enum variant: SE_DockWidgetCloseButton = 28)

SEDockWidgetFloatButton

The float button of a dock widget. (C++ enum variant: SE_DockWidgetFloatButton = 29)

SEDockWidgetTitleBarText

The text bounds of the dock widgets title. (C++ enum variant: SE_DockWidgetTitleBarText = 30)

SEDockWidgetIcon

The icon of a dock widget. (C++ enum variant: SE_DockWidgetIcon = 31)

SECheckBoxLayoutItem

Area that counts for the parent layout. (C++ enum variant: SE_CheckBoxLayoutItem = 32)

SEComboBoxLayoutItem

Area that counts for the parent layout. (C++ enum variant: SE_ComboBoxLayoutItem = 33)

SEDateTimeEditLayoutItem

Area that counts for the parent layout. (C++ enum variant: SE_DateTimeEditLayoutItem = 34)

SEDialogButtonBoxLayoutItem

Area that counts for the parent layout. (C++ enum variant: SE_DialogButtonBoxLayoutItem = 35)

SELabelLayoutItem

Area that counts for the parent layout. (C++ enum variant: SE_LabelLayoutItem = 36)

SEProgressBarLayoutItem

Area that counts for the parent layout. (C++ enum variant: SE_ProgressBarLayoutItem = 37)

SEPushButtonLayoutItem

Area that counts for the parent layout. (C++ enum variant: SE_PushButtonLayoutItem = 38)

SERadioButtonLayoutItem

Area that counts for the parent layout. (C++ enum variant: SE_RadioButtonLayoutItem = 39)

SESliderLayoutItem

Area that counts for the parent layout. (C++ enum variant: SE_SliderLayoutItem = 40)

SESpinBoxLayoutItem

Area that counts for the parent layout. (C++ enum variant: SE_SpinBoxLayoutItem = 41)

SEToolButtonLayoutItem

Area that counts for the parent layout. (C++ enum variant: SE_ToolButtonLayoutItem = 42)

SEFrameLayoutItem

Area that counts for the parent layout. (C++ enum variant: SE_FrameLayoutItem = 43)

SEGroupBoxLayoutItem

Area that counts for the parent layout. (C++ enum variant: SE_GroupBoxLayoutItem = 44)

SETabWidgetLayoutItem

Area that counts for the parent layout. (C++ enum variant: SE_TabWidgetLayoutItem = 45)

SEItemViewItemDecoration

Area for a view item's decoration (icon). (C++ enum variant: SE_ItemViewItemDecoration = 46)

SEItemViewItemText

Area for a view item's text. (C++ enum variant: SE_ItemViewItemText = 47)

SEItemViewItemFocusRect

Area for a view item's focus rect. (C++ enum variant: SE_ItemViewItemFocusRect = 48)

SETabBarTabLeftButton

Area for a widget on the left side of a tab in a tab bar. (C++ enum variant: SE_TabBarTabLeftButton = 49)

SETabBarTabRightButton

Area for a widget on the right side of a tab in a tab bar. (C++ enum variant: SE_TabBarTabRightButton = 50)

SETabBarTabText

Area for the text on a tab in a tab bar. (C++ enum variant: SE_TabBarTabText = 51)

SEShapedFrameContents

Area for a frame's contents using the shape in QStyleOptionFrameV3; see QFrame (C++ enum variant: SE_ShapedFrameContents = 52)

SEToolBarHandle

Area for the handle of a tool bar. (C++ enum variant: SE_ToolBarHandle = 53)