Enum qt_widgets::style_option::OptionType

pub enum OptionType {
    SODefault,
    SOFocusRect,
    SOButton,
    SOTab,
    SOMenuItem,
    SOFrame,
    SOProgressBar,
    SOToolBox,
    SOHeader,
    SODockWidget,
    SOViewItem,
    SOTabWidgetFrame,
    SOTabBarBase,
    SORubberBand,
    SOToolBar,
    SOGraphicsItem,
    SOCustomBase,
    SOComplex,
    SOSlider,
    SOSpinBox,
    SOToolButton,
    SOComboBox,
    SOTitleBar,
    SOGroupBox,
    SOSizeGrip,
    SOComplexCustomBase,
}

C++ type: QStyleOption::OptionType

C++ documentation:

This enum is used internally by QStyleOption, its subclasses, and qstyleoption_cast() to determine the type of style option. In general you do not need to worry about this unless you want to create your own QStyleOption subclass and your own styles.

The following values are used for custom controls:

See also type.

Variants
SODefault

QStyleOption (C++ enum variant: SO_Default = 0)

SOFocusRect

QStyleOptionFocusRect (C++ enum variant: SO_FocusRect = 1)

SOButton

QStyleOptionButton (C++ enum variant: SO_Button = 2)

SOTab

QStyleOptionTab (C++ enum variant: SO_Tab = 3)

SOMenuItem

QStyleOptionMenuItem (C++ enum variant: SO_MenuItem = 4)

SOFrame

QStyleOptionFrame (C++ enum variant: SO_Frame = 5)

SOProgressBar

QStyleOptionProgressBar (C++ enum variant: SO_ProgressBar = 6)

SOToolBox

QStyleOptionToolBox (C++ enum variant: SO_ToolBox = 7)

SOHeader

QStyleOptionHeader (C++ enum variant: SO_Header = 8)

SODockWidget

QStyleOptionDockWidget (C++ enum variant: SO_DockWidget = 9)

SOViewItem

QStyleOptionViewItem (used in Interviews) (C++ enum variant: SO_ViewItem = 10)

SOTabWidgetFrame

QStyleOptionTabWidgetFrame (C++ enum variant: SO_TabWidgetFrame = 11)

SOTabBarBase

QStyleOptionTabBarBase (C++ enum variant: SO_TabBarBase = 12)

SORubberBand

QStyleOptionRubberBand (C++ enum variant: SO_RubberBand = 13)

SOToolBar

QStyleOptionToolBar (C++ enum variant: SO_ToolBar = 14)

SOGraphicsItem

QStyleOptionGraphicsItem (C++ enum variant: SO_GraphicsItem = 15)

SOCustomBase

Reserved for custom QStyleOptions; all custom controls values must be above this value (C++ enum variant: SO_CustomBase = 3840)

SOComplex

QStyleOptionComplex (C++ enum variant: SO_Complex = 983040)

SOSlider

QStyleOptionSlider (C++ enum variant: SO_Slider = 983041)

SOSpinBox

QStyleOptionSpinBox (C++ enum variant: SO_SpinBox = 983042)

SOToolButton

QStyleOptionToolButton (C++ enum variant: SO_ToolButton = 983043)

SOComboBox

QStyleOptionComboBox (C++ enum variant: SO_ComboBox = 983044)

SOTitleBar

QStyleOptionTitleBar (C++ enum variant: SO_TitleBar = 983045)

SOGroupBox

QStyleOptionGroupBox (C++ enum variant: SO_GroupBox = 983046)

SOSizeGrip

QStyleOptionSizeGrip (C++ enum variant: SO_SizeGrip = 983047)

SOComplexCustomBase

Reserved for custom QStyleOptions; all custom complex controls values must be above this value (C++ enum variant: SO_ComplexCustomBase = 251658240)