Struct qt_widgets::style_option::StyleOption

pub struct StyleOption {
    // some fields omitted
}

C++ type: QStyleOption

C++ documentation:

The QStyleOption class stores the parameters used by QStyle functions.

QStyleOption and its subclasses contain all the information that QStyle functions need to draw a graphical element.

For performance reasons, there are few member functions and the access to the member variables is direct (i.e., using the . or -> operator). This low-level feel makes the structures straightforward to use and emphasizes that these are simply parameters used by the style functions.

The caller of a QStyle function usually creates QStyleOption objects on the stack. This combined with Qt's extensive use of implicit sharing for types such as QString, QPalette, and QColor ensures that no memory allocation needlessly takes place.

The following code snippet shows how to use a specific QStyleOption subclass to paint a push button:

void MyPushButton::paintEvent(QPaintEvent *)
{
    QStyleOptionButton option;
    option.initFrom(this);
    option.state = isDown() ? QStyle::State_Sunken : QStyle::State_Raised;
    if (isDefault())
        option.features |= QStyleOptionButton::DefaultButton;
    option.text = text();
    option.icon = icon();

QPainter painter(this); style()->drawControl(QStyle::CE_PushButton, &option, &painter, this); }
In our example, the control is a QStyle::CE_PushButton, and according to the QStyle::drawControl() documentation the corresponding class is QStyleOptionButton.

When reimplementing QStyle functions that take a QStyleOption parameter, you often need to cast the QStyleOption to a subclass. For safety, you can use qstyleoption_cast() to ensure that the pointer type is correct. For example:

void MyStyle::drawPrimitive(PrimitiveElement element,
                            const QStyleOption option,
                            QPainter painter,
                            const QWidget widget)
{
    if (element == PE_FrameFocusRect) {
        const QStyleOptionFocusRect focusRectOption =
                qstyleoption_cast<const QStyleOptionFocusRect *>(option);
        if (focusRectOption) {
            // ...
        }
    }
    // ...
}

The qstyleoption_cast() function will return 0 if the object to which option points is not of the correct type.

For an example demonstrating how style options can be used, see the Styles example.

Methods

impl StyleOption

fn direction(&self) -> LayoutDirection

C++ method: Qt::LayoutDirection QStyleOption::direction() const

fn font_metrics<'l0>(&'l0 self) -> &'l0 FontMetrics

C++ method: const QFontMetrics& QStyleOption::fontMetrics() const

fn font_metrics_mut<'l0>(&'l0 mut self) -> &'l0 mut FontMetrics

C++ method: QFontMetrics& QStyleOption::fontMetrics_mut()

fn init(&mut self, w: *const Widget)

C++ method: void QStyleOption::init(const QWidget* w)

C++ documentation:

Use initFrom(widget) instead.

fn init_from(&mut self, w: *const Widget)

C++ method: void QStyleOption::initFrom(const QWidget* w)

C++ documentation:

Initializes the state, direction, rect, palette, fontMetrics and styleObject member variables based on the specified widget.

This is a convenience function; the member variables can also be initialized manually.

This function was introduced in Qt 4.1.

See also QWidget::layoutDirection(), QWidget::rect(), QWidget::palette(), and QWidget::fontMetrics().

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyleOptionNewArgs

C++ method: QStyleOption::QStyleOption

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((&::style_option::StyleOption, cpp_utils::AsStruct)) -> ::style_option::StyleOption
2) fn new((&::style_option::StyleOption, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option::StyleOption>
C++ method: [constructor] void QStyleOption::QStyleOption(const QStyleOption& other)

Variant 2

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::style_option::StyleOption
2) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::style_option::StyleOption
3) fn new((libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::style_option::StyleOption
4) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::style_option::StyleOption>
5) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option::StyleOption>
6) fn new((libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option::StyleOption>
C++ method: [constructor] void QStyleOption::QStyleOption(int version = ?, int type = ?)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 StyleOption) -> &'l0 mut StyleOption

C++ method: QStyleOption& QStyleOption::operator=(const QStyleOption& other)

C++ documentation:

Assign other to this QStyleOption.

fn palette<'l0>(&'l0 self) -> &'l0 Palette

C++ method: const QPalette& QStyleOption::palette() const

fn palette_mut<'l0>(&'l0 mut self) -> &'l0 mut Palette

C++ method: QPalette& QStyleOption::palette_mut()

fn rect<'l0>(&'l0 self) -> &'l0 Rect

C++ method: const QRect& QStyleOption::rect() const

fn rect_mut<'l0>(&'l0 mut self) -> &'l0 mut Rect

C++ method: QRect& QStyleOption::rect_mut()

fn set_direction(&mut self, value: LayoutDirection)

C++ method: void QStyleOption::set_direction(Qt::LayoutDirection value)

fn set_font_metrics(&mut self, value: &FontMetrics)

C++ method: void QStyleOption::set_fontMetrics(QFontMetrics value)

fn set_palette(&mut self, value: &Palette)

C++ method: void QStyleOption::set_palette(QPalette value)

fn set_rect(&mut self, value: &Rect)

C++ method: void QStyleOption::set_rect(QRect value)

fn set_state(&mut self, value: Flags<StateFlag>)

C++ method: void QStyleOption::set_state(QFlags<QStyle::StateFlag> value)

fn set_style_object(&mut self, value: *mut Object)

C++ method: void QStyleOption::set_styleObject(QObject* value)

fn set_type(&mut self, value: c_int)

C++ method: void QStyleOption::set_type(int value)

fn set_version(&mut self, value: c_int)

C++ method: void QStyleOption::set_version(int value)

fn state(&self) -> Flags<StateFlag>

C++ method: const QFlags<QStyle::StateFlag>& QStyleOption::state() const

fn style_object(&self) -> *mut Object

C++ method: QObject* QStyleOption::styleObject() const

fn type_(&self) -> c_int

C++ method: int QStyleOption::type() const

fn version(&self) -> c_int

C++ method: int QStyleOption::version() const