Struct qt_widgets::style_option_combo_box::StyleOptionComboBox

pub struct StyleOptionComboBox {
    // some fields omitted
}

C++ type: QStyleOptionComboBox

C++ documentation:

The QStyleOptionComboBox class is used to describe the parameter for drawing a combobox.

QStyleOptionButton contains all the information that QStyle functions need to draw QComboBox.

For performance reasons, the access to the member variables is direct (i.e., using the . or -> operator). This low-level feel makes the structures straightforward to use and emphasizes that these are simply parameters used by the style functions.

For an example demonstrating how style options can be used, see the Styles example.

Methods

impl StyleOptionComboBox

fn current_icon<'l0>(&'l0 self) -> &'l0 Icon

C++ method: const QIcon& QStyleOptionComboBox::currentIcon() const

fn current_icon_mut<'l0>(&'l0 mut self) -> &'l0 mut Icon

C++ method: QIcon& QStyleOptionComboBox::currentIcon_mut()

fn current_text<'l0>(&'l0 self) -> &'l0 String

C++ method: const QString& QStyleOptionComboBox::currentText() const

fn current_text_mut<'l0>(&'l0 mut self) -> &'l0 mut String

C++ method: QString& QStyleOptionComboBox::currentText_mut()

fn editable(&self) -> bool

C++ method: bool QStyleOptionComboBox::editable() const

fn frame(&self) -> bool

C++ method: bool QStyleOptionComboBox::frame() const

fn icon_size<'l0>(&'l0 self) -> &'l0 Size

C++ method: const QSize& QStyleOptionComboBox::iconSize() const

fn icon_size_mut<'l0>(&'l0 mut self) -> &'l0 mut Size

C++ method: QSize& QStyleOptionComboBox::iconSize_mut()

fn init(&mut self, w: *const Widget)

C++ method: void QStyleOptionComboBox::init(const QWidget* w)

C++ documentation:

Use initFrom(widget) instead.

fn init_from(&mut self, w: *const Widget)

C++ method: void QStyleOptionComboBox::initFrom(const QWidget* w)

C++ documentation:

Initializes the state, direction, rect, palette, fontMetrics and styleObject member variables based on the specified widget.

This is a convenience function; the member variables can also be initialized manually.

This function was introduced in Qt 4.1.

See also QWidget::layoutDirection(), QWidget::rect(), QWidget::palette(), and QWidget::fontMetrics().

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyleOptionComboBoxNewArgs

C++ method: QStyleOptionComboBox::QStyleOptionComboBox

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::style_option_combo_box::StyleOptionComboBox
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::style_option_combo_box::StyleOptionComboBox>
C++ method: [constructor] void QStyleOptionComboBox::QStyleOptionComboBox()

Variant 2

Rust arguments:
1) fn new((&::style_option_combo_box::StyleOptionComboBox, cpp_utils::AsStruct)) -> ::style_option_combo_box::StyleOptionComboBox
2) fn new((&::style_option_combo_box::StyleOptionComboBox, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_combo_box::StyleOptionComboBox>
C++ method: [constructor] void QStyleOptionComboBox::QStyleOptionComboBox(const QStyleOptionComboBox& other)

fn popup_rect<'l0>(&'l0 self) -> &'l0 Rect

C++ method: const QRect& QStyleOptionComboBox::popupRect() const

fn popup_rect_mut<'l0>(&'l0 mut self) -> &'l0 mut Rect

C++ method: QRect& QStyleOptionComboBox::popupRect_mut()

fn set_current_icon(&mut self, value: &Icon)

C++ method: void QStyleOptionComboBox::set_currentIcon(QIcon value)

fn set_current_text(&mut self, value: &String)

C++ method: void QStyleOptionComboBox::set_currentText(QString value)

fn set_editable(&mut self, value: bool)

C++ method: void QStyleOptionComboBox::set_editable(bool value)

fn set_frame(&mut self, value: bool)

C++ method: void QStyleOptionComboBox::set_frame(bool value)

fn set_icon_size(&mut self, value: &Size)

C++ method: void QStyleOptionComboBox::set_iconSize(QSize value)

fn set_popup_rect(&mut self, value: &Rect)

C++ method: void QStyleOptionComboBox::set_popupRect(QRect value)