Struct qt_widgets::style_option_complex::StyleOptionComplex

pub struct StyleOptionComplex {
    // some fields omitted
}

C++ type: QStyleOptionComplex

C++ documentation:

The QStyleOptionComplex class is used to hold parameters that are common to all complex controls.

This class is not used on its own. Instead it is used to derive other complex control options, for example QStyleOptionSlider and QStyleOptionSpinBox.

For performance reasons, the access to the member variables is direct (i.e., using the . or -> operator).

For an example demonstrating how style options can be used, see the Styles example.

Methods

impl StyleOptionComplex

fn active_sub_controls(&self) -> Flags<SubControl>

C++ method: const QFlags<QStyle::SubControl>& QStyleOptionComplex::activeSubControls() const

fn init(&mut self, w: *const Widget)

C++ method: void QStyleOptionComplex::init(const QWidget* w)

C++ documentation:

Use initFrom(widget) instead.

fn init_from(&mut self, w: *const Widget)

C++ method: void QStyleOptionComplex::initFrom(const QWidget* w)

C++ documentation:

Initializes the state, direction, rect, palette, fontMetrics and styleObject member variables based on the specified widget.

This is a convenience function; the member variables can also be initialized manually.

This function was introduced in Qt 4.1.

See also QWidget::layoutDirection(), QWidget::rect(), QWidget::palette(), and QWidget::fontMetrics().

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyleOptionComplexNewArgs

C++ method: QStyleOptionComplex::QStyleOptionComplex

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((&::style_option_complex::StyleOptionComplex, cpp_utils::AsStruct)) -> ::style_option_complex::StyleOptionComplex
2) fn new((&::style_option_complex::StyleOptionComplex, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_complex::StyleOptionComplex>
C++ method: [constructor] void QStyleOptionComplex::QStyleOptionComplex(const QStyleOptionComplex& other)

Variant 2

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::style_option_complex::StyleOptionComplex
2) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::style_option_complex::StyleOptionComplex
3) fn new((libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::style_option_complex::StyleOptionComplex
4) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::style_option_complex::StyleOptionComplex>
5) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_complex::StyleOptionComplex>
6) fn new((libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_complex::StyleOptionComplex>
C++ method: [constructor] void QStyleOptionComplex::QStyleOptionComplex(int version = ?, int type = ?)

fn set_active_sub_controls(&mut self, value: Flags<SubControl>)

C++ method: void QStyleOptionComplex::set_activeSubControls(QFlags<QStyle::SubControl> value)

fn set_sub_controls(&mut self, value: Flags<SubControl>)

C++ method: void QStyleOptionComplex::set_subControls(QFlags<QStyle::SubControl> value)

fn sub_controls(&self) -> Flags<SubControl>

C++ method: const QFlags<QStyle::SubControl>& QStyleOptionComplex::subControls() const