Struct qt_widgets::style_option_frame::StyleOptionFrame

pub struct StyleOptionFrame {
    // some fields omitted
}

C++ type: QStyleOptionFrame

C++ documentation:

The QStyleOptionFrame class is used to describe the parameters for drawing a frame.

QStyleOptionFrame is used for drawing several built-in Qt widgets, including QFrame, QGroupBox, QLineEdit, and QMenu.

An instance of the QStyleOptionFrame class has type SO_Frame and version 3.

The type is used internally by QStyleOption, its subclasses, and qstyleoption_cast() to determine the type of style option. In general you do not need to worry about this unless you want to create your own QStyleOption subclass and your own styles. The version is used by QStyleOption subclasses to implement extensions without breaking compatibility. If you use qstyleoption_cast(), you normally do not need to check it.

For an example demonstrating how style options can be used, see the Styles example.

Methods

impl StyleOptionFrame

fn features(&self) -> Flags<FrameFeature>

C++ method: const QFlags<QStyleOptionFrame::FrameFeature>& QStyleOptionFrame::features() const

fn frame_shape(&self) -> Shape

C++ method: QFrame::Shape QStyleOptionFrame::frameShape() const

fn init(&mut self, w: *const Widget)

C++ method: void QStyleOptionFrame::init(const QWidget* w)

C++ documentation:

Use initFrom(widget) instead.

fn init_from(&mut self, w: *const Widget)

C++ method: void QStyleOptionFrame::initFrom(const QWidget* w)

C++ documentation:

Initializes the state, direction, rect, palette, fontMetrics and styleObject member variables based on the specified widget.

This is a convenience function; the member variables can also be initialized manually.

This function was introduced in Qt 4.1.

See also QWidget::layoutDirection(), QWidget::rect(), QWidget::palette(), and QWidget::fontMetrics().

fn line_width(&self) -> c_int

C++ method: int QStyleOptionFrame::lineWidth() const

fn mid_line_width(&self) -> c_int

C++ method: int QStyleOptionFrame::midLineWidth() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyleOptionFrameNewArgs

C++ method: QStyleOptionFrame::QStyleOptionFrame

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::style_option_frame::StyleOptionFrame
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::style_option_frame::StyleOptionFrame>
C++ method: [constructor] void QStyleOptionFrame::QStyleOptionFrame()

Variant 2

Rust arguments:
1) fn new((&::style_option_frame::StyleOptionFrame, cpp_utils::AsStruct)) -> ::style_option_frame::StyleOptionFrame
2) fn new((&::style_option_frame::StyleOptionFrame, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_frame::StyleOptionFrame>
C++ method: [constructor] void QStyleOptionFrame::QStyleOptionFrame(const QStyleOptionFrame& other)

fn set_features(&mut self, value: Flags<FrameFeature>)

C++ method: void QStyleOptionFrame::set_features(QFlags<QStyleOptionFrame::FrameFeature> value)

fn set_frame_shape(&mut self, value: Shape)

C++ method: void QStyleOptionFrame::set_frameShape(QFrame::Shape value)

fn set_line_width(&mut self, value: c_int)

C++ method: void QStyleOptionFrame::set_lineWidth(int value)

fn set_mid_line_width(&mut self, value: c_int)

C++ method: void QStyleOptionFrame::set_midLineWidth(int value)