Struct qt_widgets::style_option_graphics_item::StyleOptionGraphicsItem

pub struct StyleOptionGraphicsItem {
    // some fields omitted
}

C++ type: QStyleOptionGraphicsItem

C++ documentation:

The QStyleOptionGraphicsItem class is used to describe the parameters needed to draw a QGraphicsItem.

For performance reasons, the access to the member variables is direct (i.e., using the . or -> operator). This low-level feel makes the structures straightforward to use and emphasizes that these are simply parameters.

For an example demonstrating how style options can be used, see the Styles example.

Methods

impl StyleOptionGraphicsItem

fn exposed_rect<'l0>(&'l0 self) -> &'l0 RectF

C++ method: const QRectF& QStyleOptionGraphicsItem::exposedRect() const

fn exposed_rect_mut<'l0>(&'l0 mut self) -> &'l0 mut RectF

C++ method: QRectF& QStyleOptionGraphicsItem::exposedRect_mut()

fn init(&mut self, w: *const Widget)

C++ method: void QStyleOptionGraphicsItem::init(const QWidget* w)

C++ documentation:

Use initFrom(widget) instead.

fn init_from(&mut self, w: *const Widget)

C++ method: void QStyleOptionGraphicsItem::initFrom(const QWidget* w)

C++ documentation:

Initializes the state, direction, rect, palette, fontMetrics and styleObject member variables based on the specified widget.

This is a convenience function; the member variables can also be initialized manually.

This function was introduced in Qt 4.1.

See also QWidget::layoutDirection(), QWidget::rect(), QWidget::palette(), and QWidget::fontMetrics().

fn level_of_detail(&self) -> c_double

C++ method: double QStyleOptionGraphicsItem::levelOfDetail() const

fn level_of_detail_from_transform(world_transform: &Transform) -> c_double

C++ method: static double QStyleOptionGraphicsItem::levelOfDetailFromTransform(const QTransform& worldTransform)

C++ documentation:

Returns the level of detail from the worldTransform.

Its value represents the maximum value of the height and width of a unity rectangle, mapped using the worldTransform of the painter used to draw the item. By default, if no transformations are applied, its value is 1. If zoomed out 1:2, the level of detail will be 0.5, and if zoomed in 2:1, its value is 2.

This function was introduced in Qt 4.6.

See also QGraphicsScene::minimumRenderSize().

fn matrix<'l0>(&'l0 self) -> &'l0 Matrix

C++ method: const QMatrix& QStyleOptionGraphicsItem::matrix() const

fn matrix_mut<'l0>(&'l0 mut self) -> &'l0 mut Matrix

C++ method: QMatrix& QStyleOptionGraphicsItem::matrix_mut()

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyleOptionGraphicsItemNewArgs

C++ method: QStyleOptionGraphicsItem::QStyleOptionGraphicsItem

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::style_option_graphics_item::StyleOptionGraphicsItem
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::style_option_graphics_item::StyleOptionGraphicsItem>
C++ method: [constructor] void QStyleOptionGraphicsItem::QStyleOptionGraphicsItem()

Variant 2

Rust arguments:
1) fn new((&::style_option_graphics_item::StyleOptionGraphicsItem, cpp_utils::AsStruct)) -> ::style_option_graphics_item::StyleOptionGraphicsItem
2) fn new((&::style_option_graphics_item::StyleOptionGraphicsItem, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_graphics_item::StyleOptionGraphicsItem>
C++ method: [constructor] void QStyleOptionGraphicsItem::QStyleOptionGraphicsItem(const QStyleOptionGraphicsItem& other)

fn set_exposed_rect(&mut self, value: &RectF)

C++ method: void QStyleOptionGraphicsItem::set_exposedRect(QRectF value)

fn set_level_of_detail(&mut self, value: c_double)

C++ method: void QStyleOptionGraphicsItem::set_levelOfDetail(double value)

fn set_matrix(&mut self, value: &Matrix)

C++ method: void QStyleOptionGraphicsItem::set_matrix(QMatrix value)