Struct qt_widgets::style_option_group_box::StyleOptionGroupBox

pub struct StyleOptionGroupBox {
    // some fields omitted
}

C++ type: QStyleOptionGroupBox

C++ documentation:

The QStyleOptionGroupBox class describes the parameters for drawing a group box.

QStyleOptionButton contains all the information that QStyle functions need the various graphical elements of a group box.

It holds the lineWidth and the midLineWidth for drawing the panel, the group box's title and the title's alignment and color.

For performance reasons, the access to the member variables is direct (i.e., using the . or -> operator). This low-level feel makes the structures straightforward to use and emphasizes that these are simply parameters used by the style functions.

For an example demonstrating how style options can be used, see the Styles example.

Methods

impl StyleOptionGroupBox

fn features(&self) -> Flags<FrameFeature>

C++ method: const QFlags<QStyleOptionFrame::FrameFeature>& QStyleOptionGroupBox::features() const

fn init(&mut self, w: *const Widget)

C++ method: void QStyleOptionGroupBox::init(const QWidget* w)

C++ documentation:

Use initFrom(widget) instead.

fn init_from(&mut self, w: *const Widget)

C++ method: void QStyleOptionGroupBox::initFrom(const QWidget* w)

C++ documentation:

Initializes the state, direction, rect, palette, fontMetrics and styleObject member variables based on the specified widget.

This is a convenience function; the member variables can also be initialized manually.

This function was introduced in Qt 4.1.

See also QWidget::layoutDirection(), QWidget::rect(), QWidget::palette(), and QWidget::fontMetrics().

fn line_width(&self) -> c_int

C++ method: int QStyleOptionGroupBox::lineWidth() const

fn mid_line_width(&self) -> c_int

C++ method: int QStyleOptionGroupBox::midLineWidth() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyleOptionGroupBoxNewArgs

C++ method: QStyleOptionGroupBox::QStyleOptionGroupBox

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::style_option_group_box::StyleOptionGroupBox
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::style_option_group_box::StyleOptionGroupBox>
C++ method: [constructor] void QStyleOptionGroupBox::QStyleOptionGroupBox()

Variant 2

Rust arguments:
1) fn new((&::style_option_group_box::StyleOptionGroupBox, cpp_utils::AsStruct)) -> ::style_option_group_box::StyleOptionGroupBox
2) fn new((&::style_option_group_box::StyleOptionGroupBox, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_group_box::StyleOptionGroupBox>
C++ method: [constructor] void QStyleOptionGroupBox::QStyleOptionGroupBox(const QStyleOptionGroupBox& other)

fn set_features(&mut self, value: Flags<FrameFeature>)

C++ method: void QStyleOptionGroupBox::set_features(QFlags<QStyleOptionFrame::FrameFeature> value)

fn set_line_width(&mut self, value: c_int)

C++ method: void QStyleOptionGroupBox::set_lineWidth(int value)

fn set_mid_line_width(&mut self, value: c_int)

C++ method: void QStyleOptionGroupBox::set_midLineWidth(int value)

fn set_text(&mut self, value: &String)

C++ method: void QStyleOptionGroupBox::set_text(QString value)

fn set_text_alignment(&mut self, value: Flags<AlignmentFlag>)

C++ method: void QStyleOptionGroupBox::set_textAlignment(QFlags<Qt::AlignmentFlag> value)

fn set_text_color(&mut self, value: &Color)

C++ method: void QStyleOptionGroupBox::set_textColor(QColor value)

fn text<'l0>(&'l0 self) -> &'l0 String

C++ method: const QString& QStyleOptionGroupBox::text() const

fn text_alignment(&self) -> Flags<AlignmentFlag>

C++ method: const QFlags<Qt::AlignmentFlag>& QStyleOptionGroupBox::textAlignment() const

fn text_color<'l0>(&'l0 self) -> &'l0 Color

C++ method: const QColor& QStyleOptionGroupBox::textColor() const

fn text_color_mut<'l0>(&'l0 mut self) -> &'l0 mut Color

C++ method: QColor& QStyleOptionGroupBox::textColor_mut()

fn text_mut<'l0>(&'l0 mut self) -> &'l0 mut String

C++ method: QString& QStyleOptionGroupBox::text_mut()