Struct qt_widgets::style_option_header::StyleOptionHeader

pub struct StyleOptionHeader {
    // some fields omitted
}

C++ type: QStyleOptionHeader

C++ documentation:

The QStyleOptionHeader class is used to describe the parameters for drawing a header.

QStyleOptionHeader contains all the information that QStyle functions need to draw the item views' header pane, header sort arrow, and header label.

For performance reasons, the access to the member variables is direct (i.e., using the . or -> operator). This low-level feel makes the structures straightforward to use and emphasizes that these are simply parameters used by the style functions.

For an example demonstrating how style options can be used, see the Styles example.

Methods

impl StyleOptionHeader

fn icon<'l0>(&'l0 self) -> &'l0 Icon

C++ method: const QIcon& QStyleOptionHeader::icon() const

fn icon_alignment(&self) -> Flags<AlignmentFlag>

C++ method: const QFlags<Qt::AlignmentFlag>& QStyleOptionHeader::iconAlignment() const

fn icon_mut<'l0>(&'l0 mut self) -> &'l0 mut Icon

C++ method: QIcon& QStyleOptionHeader::icon_mut()

fn init(&mut self, w: *const Widget)

C++ method: void QStyleOptionHeader::init(const QWidget* w)

C++ documentation:

Use initFrom(widget) instead.

fn init_from(&mut self, w: *const Widget)

C++ method: void QStyleOptionHeader::initFrom(const QWidget* w)

C++ documentation:

Initializes the state, direction, rect, palette, fontMetrics and styleObject member variables based on the specified widget.

This is a convenience function; the member variables can also be initialized manually.

This function was introduced in Qt 4.1.

See also QWidget::layoutDirection(), QWidget::rect(), QWidget::palette(), and QWidget::fontMetrics().

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyleOptionHeaderNewArgs

C++ method: QStyleOptionHeader::QStyleOptionHeader

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::style_option_header::StyleOptionHeader
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::style_option_header::StyleOptionHeader>
C++ method: [constructor] void QStyleOptionHeader::QStyleOptionHeader()

Variant 2

Rust arguments:
1) fn new((&::style_option_header::StyleOptionHeader, cpp_utils::AsStruct)) -> ::style_option_header::StyleOptionHeader
2) fn new((&::style_option_header::StyleOptionHeader, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_header::StyleOptionHeader>
C++ method: [constructor] void QStyleOptionHeader::QStyleOptionHeader(const QStyleOptionHeader& other)

fn orientation(&self) -> Orientation

C++ method: Qt::Orientation QStyleOptionHeader::orientation() const

fn position(&self) -> SectionPosition

C++ method: QStyleOptionHeader::SectionPosition QStyleOptionHeader::position() const

fn section(&self) -> c_int

C++ method: int QStyleOptionHeader::section() const

fn selected_position(&self) -> SelectedPosition

C++ method: QStyleOptionHeader::SelectedPosition QStyleOptionHeader::selectedPosition() const

fn set_icon(&mut self, value: &Icon)

C++ method: void QStyleOptionHeader::set_icon(QIcon value)

fn set_icon_alignment(&mut self, value: Flags<AlignmentFlag>)

C++ method: void QStyleOptionHeader::set_iconAlignment(QFlags<Qt::AlignmentFlag> value)

fn set_orientation(&mut self, value: Orientation)

C++ method: void QStyleOptionHeader::set_orientation(Qt::Orientation value)

fn set_position(&mut self, value: SectionPosition)

C++ method: void QStyleOptionHeader::set_position(QStyleOptionHeader::SectionPosition value)

fn set_section(&mut self, value: c_int)

C++ method: void QStyleOptionHeader::set_section(int value)

fn set_selected_position(&mut self, value: SelectedPosition)

C++ method: void QStyleOptionHeader::set_selectedPosition(QStyleOptionHeader::SelectedPosition value)

fn set_sort_indicator(&mut self, value: SortIndicator)

C++ method: void QStyleOptionHeader::set_sortIndicator(QStyleOptionHeader::SortIndicator value)

fn set_text(&mut self, value: &String)

C++ method: void QStyleOptionHeader::set_text(QString value)

fn set_text_alignment(&mut self, value: Flags<AlignmentFlag>)

C++ method: void QStyleOptionHeader::set_textAlignment(QFlags<Qt::AlignmentFlag> value)

fn sort_indicator(&self) -> SortIndicator

C++ method: QStyleOptionHeader::SortIndicator QStyleOptionHeader::sortIndicator() const

fn text<'l0>(&'l0 self) -> &'l0 String

C++ method: const QString& QStyleOptionHeader::text() const

fn text_alignment(&self) -> Flags<AlignmentFlag>

C++ method: const QFlags<Qt::AlignmentFlag>& QStyleOptionHeader::textAlignment() const

fn text_mut<'l0>(&'l0 mut self) -> &'l0 mut String

C++ method: QString& QStyleOptionHeader::text_mut()