Enum qt_widgets::style_option_menu_item::MenuItemType

pub enum MenuItemType {
    Normal,
    DefaultItem,
    Separator,
    SubMenu,
    Scroller,
    TearOff,
    Margin,
    EmptyArea,
}

C++ type: QStyleOptionMenuItem::MenuItemType

C++ documentation:

This enum indicates the type of menu item that the structure describes.

See also menuItemType.

Variants
Normal

A normal menu item. (C++ enum variant: Normal = 0)

DefaultItem

A menu item that is the default action as specified with QMenu::defaultAction(). (C++ enum variant: DefaultItem = 1)

Separator

A menu separator. (C++ enum variant: Separator = 2)

SubMenu

Indicates the menu item points to a sub-menu. (C++ enum variant: SubMenu = 3)

Scroller

A popup menu scroller (currently only used on Mac OS X). (C++ enum variant: Scroller = 4)

TearOff

A tear-off handle for the menu. (C++ enum variant: TearOff = 5)

Margin

The margin of the menu. (C++ enum variant: Margin = 6)

EmptyArea

The empty area of the menu. (C++ enum variant: EmptyArea = 7)