Struct qt_widgets::style_option_menu_item::StyleOptionMenuItem

pub struct StyleOptionMenuItem {
    // some fields omitted
}

C++ type: QStyleOptionMenuItem

C++ documentation:

The QStyleOptionMenuItem class is used to describe the parameter necessary for drawing a menu item.

QStyleOptionMenuItem contains all the information that QStyle functions need to draw the menu items from QMenu. It is also used for drawing other menu-related widgets.

For performance reasons, the access to the member variables is direct (i.e., using the . or -> operator). This low-level feel makes the structures straightforward to use and emphasizes that these are simply parameters used by the style functions.

For an example demonstrating how style options can be used, see the Styles example.

Methods

impl StyleOptionMenuItem

fn check_type(&self) -> CheckType

C++ method: QStyleOptionMenuItem::CheckType QStyleOptionMenuItem::checkType() const

fn checked(&self) -> bool

C++ method: bool QStyleOptionMenuItem::checked() const

fn font<'l0>(&'l0 self) -> &'l0 Font

C++ method: const QFont& QStyleOptionMenuItem::font() const

fn font_mut<'l0>(&'l0 mut self) -> &'l0 mut Font

C++ method: QFont& QStyleOptionMenuItem::font_mut()

fn icon<'l0>(&'l0 self) -> &'l0 Icon

C++ method: const QIcon& QStyleOptionMenuItem::icon() const

fn icon_mut<'l0>(&'l0 mut self) -> &'l0 mut Icon

C++ method: QIcon& QStyleOptionMenuItem::icon_mut()

fn init(&mut self, w: *const Widget)

C++ method: void QStyleOptionMenuItem::init(const QWidget* w)

C++ documentation:

Use initFrom(widget) instead.

fn init_from(&mut self, w: *const Widget)

C++ method: void QStyleOptionMenuItem::initFrom(const QWidget* w)

C++ documentation:

Initializes the state, direction, rect, palette, fontMetrics and styleObject member variables based on the specified widget.

This is a convenience function; the member variables can also be initialized manually.

This function was introduced in Qt 4.1.

See also QWidget::layoutDirection(), QWidget::rect(), QWidget::palette(), and QWidget::fontMetrics().

fn max_icon_width(&self) -> c_int

C++ method: int QStyleOptionMenuItem::maxIconWidth() const

fn menu_has_checkable_items(&self) -> bool

C++ method: bool QStyleOptionMenuItem::menuHasCheckableItems() const

fn menu_item_type(&self) -> MenuItemType

C++ method: QStyleOptionMenuItem::MenuItemType QStyleOptionMenuItem::menuItemType() const

fn menu_rect<'l0>(&'l0 self) -> &'l0 Rect

C++ method: const QRect& QStyleOptionMenuItem::menuRect() const

fn menu_rect_mut<'l0>(&'l0 mut self) -> &'l0 mut Rect

C++ method: QRect& QStyleOptionMenuItem::menuRect_mut()

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyleOptionMenuItemNewArgs

C++ method: QStyleOptionMenuItem::QStyleOptionMenuItem

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::style_option_menu_item::StyleOptionMenuItem
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::style_option_menu_item::StyleOptionMenuItem>
C++ method: [constructor] void QStyleOptionMenuItem::QStyleOptionMenuItem()

Variant 2

Rust arguments:
1) fn new((&::style_option_menu_item::StyleOptionMenuItem, cpp_utils::AsStruct)) -> ::style_option_menu_item::StyleOptionMenuItem
2) fn new((&::style_option_menu_item::StyleOptionMenuItem, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_menu_item::StyleOptionMenuItem>
C++ method: [constructor] void QStyleOptionMenuItem::QStyleOptionMenuItem(const QStyleOptionMenuItem& other)

fn set_check_type(&mut self, value: CheckType)

C++ method: void QStyleOptionMenuItem::set_checkType(QStyleOptionMenuItem::CheckType value)

fn set_checked(&mut self, value: bool)

C++ method: void QStyleOptionMenuItem::set_checked(bool value)

fn set_font(&mut self, value: &Font)

C++ method: void QStyleOptionMenuItem::set_font(QFont value)

fn set_icon(&mut self, value: &Icon)

C++ method: void QStyleOptionMenuItem::set_icon(QIcon value)

fn set_max_icon_width(&mut self, value: c_int)

C++ method: void QStyleOptionMenuItem::set_maxIconWidth(int value)

fn set_menu_has_checkable_items(&mut self, value: bool)

C++ method: void QStyleOptionMenuItem::set_menuHasCheckableItems(bool value)

fn set_menu_item_type(&mut self, value: MenuItemType)

C++ method: void QStyleOptionMenuItem::set_menuItemType(QStyleOptionMenuItem::MenuItemType value)

fn set_menu_rect(&mut self, value: &Rect)

C++ method: void QStyleOptionMenuItem::set_menuRect(QRect value)

fn set_tab_width(&mut self, value: c_int)

C++ method: void QStyleOptionMenuItem::set_tabWidth(int value)

fn set_text(&mut self, value: &String)

C++ method: void QStyleOptionMenuItem::set_text(QString value)

fn tab_width(&self) -> c_int

C++ method: int QStyleOptionMenuItem::tabWidth() const

fn text<'l0>(&'l0 self) -> &'l0 String

C++ method: const QString& QStyleOptionMenuItem::text() const

fn text_mut<'l0>(&'l0 mut self) -> &'l0 mut String

C++ method: QString& QStyleOptionMenuItem::text_mut()