Struct qt_widgets::style_option_progress_bar::StyleOptionProgressBar

pub struct StyleOptionProgressBar {
    // some fields omitted
}

C++ type: QStyleOptionProgressBar

C++ documentation:

The QStyleOptionProgressBar class is used to describe the parameters necessary for drawing a progress bar.

An instance of the QStyleOptionProgressBar class has type SO_ProgressBar and version 2.

The type is used internally by QStyleOption, its subclasses, and qstyleoption_cast() to determine the type of style option. In general you do not need to worry about this unless you want to create your own QStyleOption subclass and your own styles. The version is used by QStyleOption subclasses to implement extensions without breaking compatibility. If you use qstyleoption_cast(), you normally do not need to check it.

For an example demonstrating how style options can be used, see the Styles example.

Methods

impl StyleOptionProgressBar

fn bottom_to_top(&self) -> bool

C++ method: bool QStyleOptionProgressBar::bottomToTop() const

fn init(&mut self, w: *const Widget)

C++ method: void QStyleOptionProgressBar::init(const QWidget* w)

C++ documentation:

Use initFrom(widget) instead.

fn init_from(&mut self, w: *const Widget)

C++ method: void QStyleOptionProgressBar::initFrom(const QWidget* w)

C++ documentation:

Initializes the state, direction, rect, palette, fontMetrics and styleObject member variables based on the specified widget.

This is a convenience function; the member variables can also be initialized manually.

This function was introduced in Qt 4.1.

See also QWidget::layoutDirection(), QWidget::rect(), QWidget::palette(), and QWidget::fontMetrics().

fn inverted_appearance(&self) -> bool

C++ method: bool QStyleOptionProgressBar::invertedAppearance() const

fn maximum(&self) -> c_int

C++ method: int QStyleOptionProgressBar::maximum() const

fn minimum(&self) -> c_int

C++ method: int QStyleOptionProgressBar::minimum() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyleOptionProgressBarNewArgs

C++ method: QStyleOptionProgressBar::QStyleOptionProgressBar

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::style_option_progress_bar::StyleOptionProgressBar
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::style_option_progress_bar::StyleOptionProgressBar>
C++ method: [constructor] void QStyleOptionProgressBar::QStyleOptionProgressBar()

Variant 2

Rust arguments:
1) fn new((&::style_option_progress_bar::StyleOptionProgressBar, cpp_utils::AsStruct)) -> ::style_option_progress_bar::StyleOptionProgressBar
2) fn new((&::style_option_progress_bar::StyleOptionProgressBar, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_progress_bar::StyleOptionProgressBar>
C++ method: [constructor] void QStyleOptionProgressBar::QStyleOptionProgressBar(const QStyleOptionProgressBar& other)

fn orientation(&self) -> Orientation

C++ method: Qt::Orientation QStyleOptionProgressBar::orientation() const

fn progress(&self) -> c_int

C++ method: int QStyleOptionProgressBar::progress() const

fn set_bottom_to_top(&mut self, value: bool)

C++ method: void QStyleOptionProgressBar::set_bottomToTop(bool value)

fn set_inverted_appearance(&mut self, value: bool)

C++ method: void QStyleOptionProgressBar::set_invertedAppearance(bool value)

fn set_maximum(&mut self, value: c_int)

C++ method: void QStyleOptionProgressBar::set_maximum(int value)

fn set_minimum(&mut self, value: c_int)

C++ method: void QStyleOptionProgressBar::set_minimum(int value)

fn set_orientation(&mut self, value: Orientation)

C++ method: void QStyleOptionProgressBar::set_orientation(Qt::Orientation value)

fn set_progress(&mut self, value: c_int)

C++ method: void QStyleOptionProgressBar::set_progress(int value)

fn set_text(&mut self, value: &String)

C++ method: void QStyleOptionProgressBar::set_text(QString value)

fn set_text_alignment(&mut self, value: Flags<AlignmentFlag>)

C++ method: void QStyleOptionProgressBar::set_textAlignment(QFlags<Qt::AlignmentFlag> value)

fn set_text_visible(&mut self, value: bool)

C++ method: void QStyleOptionProgressBar::set_textVisible(bool value)

fn text<'l0>(&'l0 self) -> &'l0 String

C++ method: const QString& QStyleOptionProgressBar::text() const

fn text_alignment(&self) -> Flags<AlignmentFlag>

C++ method: const QFlags<Qt::AlignmentFlag>& QStyleOptionProgressBar::textAlignment() const

fn text_mut<'l0>(&'l0 mut self) -> &'l0 mut String

C++ method: QString& QStyleOptionProgressBar::text_mut()

fn text_visible(&self) -> bool

C++ method: bool QStyleOptionProgressBar::textVisible() const