Struct qt_widgets::style_option_size_grip::StyleOptionSizeGrip

pub struct StyleOptionSizeGrip {
    // some fields omitted
}

C++ type: QStyleOptionSizeGrip

C++ documentation:

The QStyleOptionSizeGrip class is used to describe the parameter for drawing a size grip.

QStyleOptionButton contains all the information that QStyle functions need to draw QSizeGrip.

For performance reasons, the access to the member variables is direct (i.e., using the . or -> operator). This low-level feel makes the structures straightforward to use and emphasizes that these are simply parameters used by the style functions.

For an example demonstrating how style options can be used, see the Styles example.

Methods

impl StyleOptionSizeGrip

fn corner(&self) -> Corner

C++ method: Qt::Corner QStyleOptionSizeGrip::corner() const

fn init(&mut self, w: *const Widget)

C++ method: void QStyleOptionSizeGrip::init(const QWidget* w)

C++ documentation:

Use initFrom(widget) instead.

fn init_from(&mut self, w: *const Widget)

C++ method: void QStyleOptionSizeGrip::initFrom(const QWidget* w)

C++ documentation:

Initializes the state, direction, rect, palette, fontMetrics and styleObject member variables based on the specified widget.

This is a convenience function; the member variables can also be initialized manually.

This function was introduced in Qt 4.1.

See also QWidget::layoutDirection(), QWidget::rect(), QWidget::palette(), and QWidget::fontMetrics().

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyleOptionSizeGripNewArgs

C++ method: QStyleOptionSizeGrip::QStyleOptionSizeGrip

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::style_option_size_grip::StyleOptionSizeGrip
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::style_option_size_grip::StyleOptionSizeGrip>
C++ method: [constructor] void QStyleOptionSizeGrip::QStyleOptionSizeGrip()

Variant 2

Rust arguments:
1) fn new((&::style_option_size_grip::StyleOptionSizeGrip, cpp_utils::AsStruct)) -> ::style_option_size_grip::StyleOptionSizeGrip
2) fn new((&::style_option_size_grip::StyleOptionSizeGrip, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_size_grip::StyleOptionSizeGrip>
C++ method: [constructor] void QStyleOptionSizeGrip::QStyleOptionSizeGrip(const QStyleOptionSizeGrip& other)

fn set_corner(&mut self, value: Corner)

C++ method: void QStyleOptionSizeGrip::set_corner(Qt::Corner value)