Struct qt_widgets::style_option_slider::StyleOptionSlider

pub struct StyleOptionSlider {
    // some fields omitted
}

C++ type: QStyleOptionSlider

C++ documentation:

The QStyleOptionSlider class is used to describe the parameters needed for drawing a slider.

QStyleOptionSlider contains all the information that QStyle functions need to draw QSlider and QScrollBar.

For performance reasons, the access to the member variables is direct (i.e., using the . or -> operator). This low-level feel makes the structures straightforward to use and emphasizes that these are simply parameters used by the style functions.

For an example demonstrating how style options can be used, see the Styles example.

Methods

impl StyleOptionSlider

fn dial_wrapping(&self) -> bool

C++ method: bool QStyleOptionSlider::dialWrapping() const

fn init(&mut self, w: *const Widget)

C++ method: void QStyleOptionSlider::init(const QWidget* w)

C++ documentation:

Use initFrom(widget) instead.

fn init_from(&mut self, w: *const Widget)

C++ method: void QStyleOptionSlider::initFrom(const QWidget* w)

C++ documentation:

Initializes the state, direction, rect, palette, fontMetrics and styleObject member variables based on the specified widget.

This is a convenience function; the member variables can also be initialized manually.

This function was introduced in Qt 4.1.

See also QWidget::layoutDirection(), QWidget::rect(), QWidget::palette(), and QWidget::fontMetrics().

fn maximum(&self) -> c_int

C++ method: int QStyleOptionSlider::maximum() const

fn minimum(&self) -> c_int

C++ method: int QStyleOptionSlider::minimum() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyleOptionSliderNewArgs

C++ method: QStyleOptionSlider::QStyleOptionSlider

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::style_option_slider::StyleOptionSlider
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::style_option_slider::StyleOptionSlider>
C++ method: [constructor] void QStyleOptionSlider::QStyleOptionSlider()

Variant 2

Rust arguments:
1) fn new((&::style_option_slider::StyleOptionSlider, cpp_utils::AsStruct)) -> ::style_option_slider::StyleOptionSlider
2) fn new((&::style_option_slider::StyleOptionSlider, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_slider::StyleOptionSlider>
C++ method: [constructor] void QStyleOptionSlider::QStyleOptionSlider(const QStyleOptionSlider& other)

fn notch_target(&self) -> c_double

C++ method: double QStyleOptionSlider::notchTarget() const

fn orientation(&self) -> Orientation

C++ method: Qt::Orientation QStyleOptionSlider::orientation() const

fn page_step(&self) -> c_int

C++ method: int QStyleOptionSlider::pageStep() const

fn set_dial_wrapping(&mut self, value: bool)

C++ method: void QStyleOptionSlider::set_dialWrapping(bool value)

fn set_maximum(&mut self, value: c_int)

C++ method: void QStyleOptionSlider::set_maximum(int value)

fn set_minimum(&mut self, value: c_int)

C++ method: void QStyleOptionSlider::set_minimum(int value)

fn set_notch_target(&mut self, value: c_double)

C++ method: void QStyleOptionSlider::set_notchTarget(double value)

fn set_orientation(&mut self, value: Orientation)

C++ method: void QStyleOptionSlider::set_orientation(Qt::Orientation value)

fn set_page_step(&mut self, value: c_int)

C++ method: void QStyleOptionSlider::set_pageStep(int value)

fn set_single_step(&mut self, value: c_int)

C++ method: void QStyleOptionSlider::set_singleStep(int value)

fn set_slider_position(&mut self, value: c_int)

C++ method: void QStyleOptionSlider::set_sliderPosition(int value)

fn set_slider_value(&mut self, value: c_int)

C++ method: void QStyleOptionSlider::set_sliderValue(int value)

fn set_tick_interval(&mut self, value: c_int)

C++ method: void QStyleOptionSlider::set_tickInterval(int value)

fn set_tick_position(&mut self, value: TickPosition)

C++ method: void QStyleOptionSlider::set_tickPosition(QSlider::TickPosition value)

fn set_upside_down(&mut self, value: bool)

C++ method: void QStyleOptionSlider::set_upsideDown(bool value)

fn single_step(&self) -> c_int

C++ method: int QStyleOptionSlider::singleStep() const

fn slider_position(&self) -> c_int

C++ method: int QStyleOptionSlider::sliderPosition() const

fn slider_value(&self) -> c_int

C++ method: int QStyleOptionSlider::sliderValue() const

fn tick_interval(&self) -> c_int

C++ method: int QStyleOptionSlider::tickInterval() const

fn tick_position(&self) -> TickPosition

C++ method: QSlider::TickPosition QStyleOptionSlider::tickPosition() const

fn upside_down(&self) -> bool

C++ method: bool QStyleOptionSlider::upsideDown() const