Struct qt_widgets::style_option_spin_box::StyleOptionSpinBox

pub struct StyleOptionSpinBox {
    // some fields omitted
}

C++ type: QStyleOptionSpinBox

C++ documentation:

The QStyleOptionSpinBox class is used to describe the parameters necessary for drawing a spin box.

QStyleOptionSpinBox contains all the information that QStyle functions need to draw QSpinBox and QDateTimeEdit.

For performance reasons, the access to the member variables is direct (i.e., using the . or -> operator). This low-level feel makes the structures straightforward to use and emphasizes that these are simply parameters used by the style functions.

For an example demonstrating how style options can be used, see the Styles example.

Methods

impl StyleOptionSpinBox

fn button_symbols(&self) -> ButtonSymbols

C++ method: QAbstractSpinBox::ButtonSymbols QStyleOptionSpinBox::buttonSymbols() const

fn frame(&self) -> bool

C++ method: bool QStyleOptionSpinBox::frame() const

fn init(&mut self, w: *const Widget)

C++ method: void QStyleOptionSpinBox::init(const QWidget* w)

C++ documentation:

Use initFrom(widget) instead.

fn init_from(&mut self, w: *const Widget)

C++ method: void QStyleOptionSpinBox::initFrom(const QWidget* w)

C++ documentation:

Initializes the state, direction, rect, palette, fontMetrics and styleObject member variables based on the specified widget.

This is a convenience function; the member variables can also be initialized manually.

This function was introduced in Qt 4.1.

See also QWidget::layoutDirection(), QWidget::rect(), QWidget::palette(), and QWidget::fontMetrics().

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyleOptionSpinBoxNewArgs

C++ method: QStyleOptionSpinBox::QStyleOptionSpinBox

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::style_option_spin_box::StyleOptionSpinBox
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::style_option_spin_box::StyleOptionSpinBox>
C++ method: [constructor] void QStyleOptionSpinBox::QStyleOptionSpinBox()

Variant 2

Rust arguments:
1) fn new((&::style_option_spin_box::StyleOptionSpinBox, cpp_utils::AsStruct)) -> ::style_option_spin_box::StyleOptionSpinBox
2) fn new((&::style_option_spin_box::StyleOptionSpinBox, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_spin_box::StyleOptionSpinBox>
C++ method: [constructor] void QStyleOptionSpinBox::QStyleOptionSpinBox(const QStyleOptionSpinBox& other)

fn set_button_symbols(&mut self, value: ButtonSymbols)

C++ method: void QStyleOptionSpinBox::set_buttonSymbols(QAbstractSpinBox::ButtonSymbols value)

fn set_frame(&mut self, value: bool)

C++ method: void QStyleOptionSpinBox::set_frame(bool value)

fn set_step_enabled(&mut self, value: Flags<StepEnabledFlag>)

C++ method: void QStyleOptionSpinBox::set_stepEnabled(QFlags<QAbstractSpinBox::StepEnabledFlag> value)

fn step_enabled(&self) -> Flags<StepEnabledFlag>

C++ method: const QFlags<QAbstractSpinBox::StepEnabledFlag>& QStyleOptionSpinBox::stepEnabled() const