Struct qt_widgets::style_option_tab::StyleOptionTab

pub struct StyleOptionTab {
    // some fields omitted
}

C++ type: QStyleOptionTab

C++ documentation:

The QStyleOptionTab class is used to describe the parameters for drawing a tab bar.

The QStyleOptionTab class is used for drawing several built-in Qt widgets including QTabBar and the panel for QTabWidget.

An instance of the QStyleOptionTab class has type SO_Tab and version 3. The type is used internally by QStyleOption, its subclasses, and qstyleoption_cast() to determine the type of style option. In general you do not need to worry about this unless you want to create your own QStyleOption subclass and your own styles. The version is used by QStyleOption subclasses to implement extensions without breaking compatibility. If you use qstyleoption_cast(), you normally do not need to check it.

For an example demonstrating how style options can be used, see the Styles example.

Methods

impl StyleOptionTab

fn corner_widgets(&self) -> Flags<CornerWidget>

C++ method: const QFlags<QStyleOptionTab::CornerWidget>& QStyleOptionTab::cornerWidgets() const

fn document_mode(&self) -> bool

C++ method: bool QStyleOptionTab::documentMode() const

fn features(&self) -> Flags<TabFeature>

C++ method: const QFlags<QStyleOptionTab::TabFeature>& QStyleOptionTab::features() const

fn icon<'l0>(&'l0 self) -> &'l0 Icon

C++ method: const QIcon& QStyleOptionTab::icon() const

fn icon_mut<'l0>(&'l0 mut self) -> &'l0 mut Icon

C++ method: QIcon& QStyleOptionTab::icon_mut()

fn icon_size<'l0>(&'l0 self) -> &'l0 Size

C++ method: const QSize& QStyleOptionTab::iconSize() const

fn icon_size_mut<'l0>(&'l0 mut self) -> &'l0 mut Size

C++ method: QSize& QStyleOptionTab::iconSize_mut()

fn init(&mut self, w: *const Widget)

C++ method: void QStyleOptionTab::init(const QWidget* w)

C++ documentation:

Use initFrom(widget) instead.

fn init_from(&mut self, w: *const Widget)

C++ method: void QStyleOptionTab::initFrom(const QWidget* w)

C++ documentation:

Initializes the state, direction, rect, palette, fontMetrics and styleObject member variables based on the specified widget.

This is a convenience function; the member variables can also be initialized manually.

This function was introduced in Qt 4.1.

See also QWidget::layoutDirection(), QWidget::rect(), QWidget::palette(), and QWidget::fontMetrics().

fn left_button_size<'l0>(&'l0 self) -> &'l0 Size

C++ method: const QSize& QStyleOptionTab::leftButtonSize() const

fn left_button_size_mut<'l0>(&'l0 mut self) -> &'l0 mut Size

C++ method: QSize& QStyleOptionTab::leftButtonSize_mut()

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyleOptionTabNewArgs

C++ method: QStyleOptionTab::QStyleOptionTab

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::style_option_tab::StyleOptionTab
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::style_option_tab::StyleOptionTab>
C++ method: [constructor] void QStyleOptionTab::QStyleOptionTab()

Variant 2

Rust arguments:
1) fn new((&::style_option_tab::StyleOptionTab, cpp_utils::AsStruct)) -> ::style_option_tab::StyleOptionTab
2) fn new((&::style_option_tab::StyleOptionTab, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_tab::StyleOptionTab>
C++ method: [constructor] void QStyleOptionTab::QStyleOptionTab(const QStyleOptionTab& other)

fn position(&self) -> TabPosition

C++ method: QStyleOptionTab::TabPosition QStyleOptionTab::position() const

fn right_button_size<'l0>(&'l0 self) -> &'l0 Size

C++ method: const QSize& QStyleOptionTab::rightButtonSize() const

fn right_button_size_mut<'l0>(&'l0 mut self) -> &'l0 mut Size

C++ method: QSize& QStyleOptionTab::rightButtonSize_mut()

fn row(&self) -> c_int

C++ method: int QStyleOptionTab::row() const

fn selected_position(&self) -> SelectedPosition

C++ method: QStyleOptionTab::SelectedPosition QStyleOptionTab::selectedPosition() const

fn set_corner_widgets(&mut self, value: Flags<CornerWidget>)

C++ method: void QStyleOptionTab::set_cornerWidgets(QFlags<QStyleOptionTab::CornerWidget> value)

fn set_document_mode(&mut self, value: bool)

C++ method: void QStyleOptionTab::set_documentMode(bool value)

fn set_features(&mut self, value: Flags<TabFeature>)

C++ method: void QStyleOptionTab::set_features(QFlags<QStyleOptionTab::TabFeature> value)

fn set_icon(&mut self, value: &Icon)

C++ method: void QStyleOptionTab::set_icon(QIcon value)

fn set_icon_size(&mut self, value: &Size)

C++ method: void QStyleOptionTab::set_iconSize(QSize value)

fn set_left_button_size(&mut self, value: &Size)

C++ method: void QStyleOptionTab::set_leftButtonSize(QSize value)

fn set_position(&mut self, value: TabPosition)

C++ method: void QStyleOptionTab::set_position(QStyleOptionTab::TabPosition value)

fn set_right_button_size(&mut self, value: &Size)

C++ method: void QStyleOptionTab::set_rightButtonSize(QSize value)

fn set_row(&mut self, value: c_int)

C++ method: void QStyleOptionTab::set_row(int value)

fn set_selected_position(&mut self, value: SelectedPosition)

C++ method: void QStyleOptionTab::set_selectedPosition(QStyleOptionTab::SelectedPosition value)

fn set_shape(&mut self, value: Shape)

C++ method: void QStyleOptionTab::set_shape(QTabBar::Shape value)

fn set_text(&mut self, value: &String)

C++ method: void QStyleOptionTab::set_text(QString value)

fn shape(&self) -> Shape

C++ method: QTabBar::Shape QStyleOptionTab::shape() const

fn text<'l0>(&'l0 self) -> &'l0 String

C++ method: const QString& QStyleOptionTab::text() const

fn text_mut<'l0>(&'l0 mut self) -> &'l0 mut String

C++ method: QString& QStyleOptionTab::text_mut()