Struct qt_widgets::style_option_tab_bar_base::StyleOptionTabBarBase

pub struct StyleOptionTabBarBase {
    // some fields omitted
}

C++ type: QStyleOptionTabBarBase

C++ documentation:

The QStyleOptionTabBarBase class is used to describe the base of a tab bar, i.e. the part that the tab bar usually overlaps with.

QStyleOptionTabBarBase contains all the information that QStyle functions need to draw the tab bar base. Note that this is only drawn for a standalone QTabBar (one that isn't part of a QTabWidget).

For performance reasons, the access to the member variables is direct (i.e., using the . or -> operator). This low-level feel makes the structures straightforward to use and emphasizes that these are simply parameters used by the style functions.

For an example demonstrating how style options can be used, see the Styles example.

Methods

impl StyleOptionTabBarBase

fn document_mode(&self) -> bool

C++ method: bool QStyleOptionTabBarBase::documentMode() const

fn init(&mut self, w: *const Widget)

C++ method: void QStyleOptionTabBarBase::init(const QWidget* w)

C++ documentation:

Use initFrom(widget) instead.

fn init_from(&mut self, w: *const Widget)

C++ method: void QStyleOptionTabBarBase::initFrom(const QWidget* w)

C++ documentation:

Initializes the state, direction, rect, palette, fontMetrics and styleObject member variables based on the specified widget.

This is a convenience function; the member variables can also be initialized manually.

This function was introduced in Qt 4.1.

See also QWidget::layoutDirection(), QWidget::rect(), QWidget::palette(), and QWidget::fontMetrics().

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyleOptionTabBarBaseNewArgs

C++ method: QStyleOptionTabBarBase::QStyleOptionTabBarBase

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::style_option_tab_bar_base::StyleOptionTabBarBase
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::style_option_tab_bar_base::StyleOptionTabBarBase>
C++ method: [constructor] void QStyleOptionTabBarBase::QStyleOptionTabBarBase()

Variant 2

Rust arguments:
1) fn new((&::style_option_tab_bar_base::StyleOptionTabBarBase, cpp_utils::AsStruct)) -> ::style_option_tab_bar_base::StyleOptionTabBarBase
2) fn new((&::style_option_tab_bar_base::StyleOptionTabBarBase, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_tab_bar_base::StyleOptionTabBarBase>
C++ method: [constructor] void QStyleOptionTabBarBase::QStyleOptionTabBarBase(const QStyleOptionTabBarBase& other)

fn selected_tab_rect<'l0>(&'l0 self) -> &'l0 Rect

C++ method: const QRect& QStyleOptionTabBarBase::selectedTabRect() const

fn selected_tab_rect_mut<'l0>(&'l0 mut self) -> &'l0 mut Rect

C++ method: QRect& QStyleOptionTabBarBase::selectedTabRect_mut()

fn set_document_mode(&mut self, value: bool)

C++ method: void QStyleOptionTabBarBase::set_documentMode(bool value)

fn set_selected_tab_rect(&mut self, value: &Rect)

C++ method: void QStyleOptionTabBarBase::set_selectedTabRect(QRect value)

fn set_shape(&mut self, value: Shape)

C++ method: void QStyleOptionTabBarBase::set_shape(QTabBar::Shape value)

fn set_tab_bar_rect(&mut self, value: &Rect)

C++ method: void QStyleOptionTabBarBase::set_tabBarRect(QRect value)

fn shape(&self) -> Shape

C++ method: QTabBar::Shape QStyleOptionTabBarBase::shape() const

fn tab_bar_rect<'l0>(&'l0 self) -> &'l0 Rect

C++ method: const QRect& QStyleOptionTabBarBase::tabBarRect() const

fn tab_bar_rect_mut<'l0>(&'l0 mut self) -> &'l0 mut Rect

C++ method: QRect& QStyleOptionTabBarBase::tabBarRect_mut()