Struct qt_widgets::style_option_tab_widget_frame::StyleOptionTabWidgetFrame

pub struct StyleOptionTabWidgetFrame {
    // some fields omitted
}

C++ type: QStyleOptionTabWidgetFrame

C++ documentation:

The QStyleOptionTabWidgetFrame class is used to describe the parameters for drawing the frame around a tab widget.

QStyleOptionTabWidgetFrame contains all the information that QStyle functions need to draw the frame around QTabWidget.

For performance reasons, the access to the member variables is direct (i.e., using the . or -> operator). This low-level feel makes the structures straightforward to use and emphasizes that these are simply parameters used by the style functions.

For an example demonstrating how style options can be used, see the Styles example.

Methods

impl StyleOptionTabWidgetFrame

fn init(&mut self, w: *const Widget)

C++ method: void QStyleOptionTabWidgetFrame::init(const QWidget* w)

C++ documentation:

Use initFrom(widget) instead.

fn init_from(&mut self, w: *const Widget)

C++ method: void QStyleOptionTabWidgetFrame::initFrom(const QWidget* w)

C++ documentation:

Initializes the state, direction, rect, palette, fontMetrics and styleObject member variables based on the specified widget.

This is a convenience function; the member variables can also be initialized manually.

This function was introduced in Qt 4.1.

See also QWidget::layoutDirection(), QWidget::rect(), QWidget::palette(), and QWidget::fontMetrics().

fn left_corner_widget_size<'l0>(&'l0 self) -> &'l0 Size

C++ method: const QSize& QStyleOptionTabWidgetFrame::leftCornerWidgetSize() const

fn left_corner_widget_size_mut<'l0>(&'l0 mut self) -> &'l0 mut Size

C++ method: QSize& QStyleOptionTabWidgetFrame::leftCornerWidgetSize_mut()

fn line_width(&self) -> c_int

C++ method: int QStyleOptionTabWidgetFrame::lineWidth() const

fn mid_line_width(&self) -> c_int

C++ method: int QStyleOptionTabWidgetFrame::midLineWidth() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyleOptionTabWidgetFrameNewArgs

C++ method: QStyleOptionTabWidgetFrame::QStyleOptionTabWidgetFrame

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::style_option_tab_widget_frame::StyleOptionTabWidgetFrame
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::style_option_tab_widget_frame::StyleOptionTabWidgetFrame>
C++ method: [constructor] void QStyleOptionTabWidgetFrame::QStyleOptionTabWidgetFrame()

Variant 2

Rust arguments:
1) fn new((&::style_option_tab_widget_frame::StyleOptionTabWidgetFrame, cpp_utils::AsStruct)) -> ::style_option_tab_widget_frame::StyleOptionTabWidgetFrame
2) fn new((&::style_option_tab_widget_frame::StyleOptionTabWidgetFrame, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_tab_widget_frame::StyleOptionTabWidgetFrame>
C++ method: [constructor] void QStyleOptionTabWidgetFrame::QStyleOptionTabWidgetFrame(const QStyleOptionTabWidgetFrame& other)

fn right_corner_widget_size<'l0>(&'l0 self) -> &'l0 Size

C++ method: const QSize& QStyleOptionTabWidgetFrame::rightCornerWidgetSize() const

fn right_corner_widget_size_mut<'l0>(&'l0 mut self) -> &'l0 mut Size

C++ method: QSize& QStyleOptionTabWidgetFrame::rightCornerWidgetSize_mut()

fn selected_tab_rect<'l0>(&'l0 self) -> &'l0 Rect

C++ method: const QRect& QStyleOptionTabWidgetFrame::selectedTabRect() const

fn selected_tab_rect_mut<'l0>(&'l0 mut self) -> &'l0 mut Rect

C++ method: QRect& QStyleOptionTabWidgetFrame::selectedTabRect_mut()

fn set_left_corner_widget_size(&mut self, value: &Size)

C++ method: void QStyleOptionTabWidgetFrame::set_leftCornerWidgetSize(QSize value)

fn set_line_width(&mut self, value: c_int)

C++ method: void QStyleOptionTabWidgetFrame::set_lineWidth(int value)

fn set_mid_line_width(&mut self, value: c_int)

C++ method: void QStyleOptionTabWidgetFrame::set_midLineWidth(int value)

fn set_right_corner_widget_size(&mut self, value: &Size)

C++ method: void QStyleOptionTabWidgetFrame::set_rightCornerWidgetSize(QSize value)

fn set_selected_tab_rect(&mut self, value: &Rect)

C++ method: void QStyleOptionTabWidgetFrame::set_selectedTabRect(QRect value)

fn set_shape(&mut self, value: Shape)

C++ method: void QStyleOptionTabWidgetFrame::set_shape(QTabBar::Shape value)

fn set_tab_bar_rect(&mut self, value: &Rect)

C++ method: void QStyleOptionTabWidgetFrame::set_tabBarRect(QRect value)

fn set_tab_bar_size(&mut self, value: &Size)

C++ method: void QStyleOptionTabWidgetFrame::set_tabBarSize(QSize value)

fn shape(&self) -> Shape

C++ method: QTabBar::Shape QStyleOptionTabWidgetFrame::shape() const

fn tab_bar_rect<'l0>(&'l0 self) -> &'l0 Rect

C++ method: const QRect& QStyleOptionTabWidgetFrame::tabBarRect() const

fn tab_bar_rect_mut<'l0>(&'l0 mut self) -> &'l0 mut Rect

C++ method: QRect& QStyleOptionTabWidgetFrame::tabBarRect_mut()

fn tab_bar_size<'l0>(&'l0 self) -> &'l0 Size

C++ method: const QSize& QStyleOptionTabWidgetFrame::tabBarSize() const

fn tab_bar_size_mut<'l0>(&'l0 mut self) -> &'l0 mut Size

C++ method: QSize& QStyleOptionTabWidgetFrame::tabBarSize_mut()