Enum qt_widgets::style_option_tool_button::ToolButtonFeature

pub enum ToolButtonFeature {
    None,
    Arrow,
    Menu,
    PopupDelay,
    HasMenu,
}

C++ type: QStyleOptionToolButton::ToolButtonFeature

C++ documentation:

Describes the various features that a tool button can have.

The ToolButtonFeatures type is a typedef for QFlags<ToolButtonFeature>. It stores an OR combination of ToolButtonFeature values.

See also features, QToolButton::toolButtonStyle(), and QToolButton::popupMode().

Variants
None

A normal tool button. (C++ enum variant: None = 0)

Arrow

The tool button is an arrow. (C++ enum variant: Arrow = 1)

Menu

This variant corresponds to multiple C++ enum variants with the same value:

Menu = 4: The tool button has a menu.

MenuButtonPopup = 4: The button should display an arrow to indicate that a menu is present.

PopupDelay

There is a delay to showing the menu. (C++ enum variant: PopupDelay = 8)

HasMenu

The button has a popup menu. (C++ enum variant: HasMenu = 16)