Struct qt_widgets::style_option_tool_button::StyleOptionToolButton

pub struct StyleOptionToolButton {
    // some fields omitted
}

C++ type: QStyleOptionToolButton

C++ documentation:

The QStyleOptionToolButton class is used to describe the parameters for drawing a tool button.

QStyleOptionToolButton contains all the information that QStyle functions need to draw QToolButton.

For performance reasons, the access to the member variables is direct (i.e., using the . or -> operator). This low-level feel makes the structures straightforward to use and emphasizes that these are simply parameters used by the style functions.

For an example demonstrating how style options can be used, see the Styles example.

Methods

impl StyleOptionToolButton

fn arrow_type(&self) -> ArrowType

C++ method: Qt::ArrowType QStyleOptionToolButton::arrowType() const

fn features(&self) -> Flags<ToolButtonFeature>

C++ method: const QFlags<QStyleOptionToolButton::ToolButtonFeature>& QStyleOptionToolButton::features() const

fn font<'l0>(&'l0 self) -> &'l0 Font

C++ method: const QFont& QStyleOptionToolButton::font() const

fn font_mut<'l0>(&'l0 mut self) -> &'l0 mut Font

C++ method: QFont& QStyleOptionToolButton::font_mut()

fn icon<'l0>(&'l0 self) -> &'l0 Icon

C++ method: const QIcon& QStyleOptionToolButton::icon() const

fn icon_mut<'l0>(&'l0 mut self) -> &'l0 mut Icon

C++ method: QIcon& QStyleOptionToolButton::icon_mut()

fn icon_size<'l0>(&'l0 self) -> &'l0 Size

C++ method: const QSize& QStyleOptionToolButton::iconSize() const

fn icon_size_mut<'l0>(&'l0 mut self) -> &'l0 mut Size

C++ method: QSize& QStyleOptionToolButton::iconSize_mut()

fn init(&mut self, w: *const Widget)

C++ method: void QStyleOptionToolButton::init(const QWidget* w)

C++ documentation:

Use initFrom(widget) instead.

fn init_from(&mut self, w: *const Widget)

C++ method: void QStyleOptionToolButton::initFrom(const QWidget* w)

C++ documentation:

Initializes the state, direction, rect, palette, fontMetrics and styleObject member variables based on the specified widget.

This is a convenience function; the member variables can also be initialized manually.

This function was introduced in Qt 4.1.

See also QWidget::layoutDirection(), QWidget::rect(), QWidget::palette(), and QWidget::fontMetrics().

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyleOptionToolButtonNewArgs

C++ method: QStyleOptionToolButton::QStyleOptionToolButton

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::style_option_tool_button::StyleOptionToolButton
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::style_option_tool_button::StyleOptionToolButton>
C++ method: [constructor] void QStyleOptionToolButton::QStyleOptionToolButton()

Variant 2

Rust arguments:
1) fn new((&::style_option_tool_button::StyleOptionToolButton, cpp_utils::AsStruct)) -> ::style_option_tool_button::StyleOptionToolButton
2) fn new((&::style_option_tool_button::StyleOptionToolButton, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_tool_button::StyleOptionToolButton>
C++ method: [constructor] void QStyleOptionToolButton::QStyleOptionToolButton(const QStyleOptionToolButton& other)

fn pos<'l0>(&'l0 self) -> &'l0 Point

C++ method: const QPoint& QStyleOptionToolButton::pos() const

fn pos_mut<'l0>(&'l0 mut self) -> &'l0 mut Point

C++ method: QPoint& QStyleOptionToolButton::pos_mut()

fn set_arrow_type(&mut self, value: ArrowType)

C++ method: void QStyleOptionToolButton::set_arrowType(Qt::ArrowType value)

fn set_features(&mut self, value: Flags<ToolButtonFeature>)

C++ method: void QStyleOptionToolButton::set_features(QFlags<QStyleOptionToolButton::ToolButtonFeature> value)

fn set_font(&mut self, value: &Font)

C++ method: void QStyleOptionToolButton::set_font(QFont value)

fn set_icon(&mut self, value: &Icon)

C++ method: void QStyleOptionToolButton::set_icon(QIcon value)

fn set_icon_size(&mut self, value: &Size)

C++ method: void QStyleOptionToolButton::set_iconSize(QSize value)

fn set_pos(&mut self, value: &Point)

C++ method: void QStyleOptionToolButton::set_pos(QPoint value)

fn set_text(&mut self, value: &String)

C++ method: void QStyleOptionToolButton::set_text(QString value)

fn set_tool_button_style(&mut self, value: ToolButtonStyle)

C++ method: void QStyleOptionToolButton::set_toolButtonStyle(Qt::ToolButtonStyle value)

fn text<'l0>(&'l0 self) -> &'l0 String

C++ method: const QString& QStyleOptionToolButton::text() const

fn text_mut<'l0>(&'l0 mut self) -> &'l0 mut String

C++ method: QString& QStyleOptionToolButton::text_mut()

fn tool_button_style(&self) -> ToolButtonStyle

C++ method: Qt::ToolButtonStyle QStyleOptionToolButton::toolButtonStyle() const