Struct qt_widgets::style_option_view_item::StyleOptionViewItem

pub struct StyleOptionViewItem {
    // some fields omitted
}

C++ type: QStyleOptionViewItem

C++ documentation:

The QStyleOptionViewItem class is used to describe the parameters used to draw an item in a view widget.

QStyleOptionViewItem contains all the information that QStyle functions need to draw the items for Qt's model/view classes.

For performance reasons, the access to the member variables is direct (i.e., using the . or -> operator). This low-level feel makes the structures straightforward to use and emphasizes that these are simply parameters used by the style functions.

For an example demonstrating how style options can be used, see the Styles example.

Methods

impl StyleOptionViewItem

fn background_brush<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QStyleOptionViewItem::backgroundBrush() const

fn background_brush_mut<'l0>(&'l0 mut self) -> &'l0 mut Brush

C++ method: QBrush& QStyleOptionViewItem::backgroundBrush_mut()

fn check_state(&self) -> CheckState

C++ method: Qt::CheckState QStyleOptionViewItem::checkState() const

fn decoration_alignment(&self) -> Flags<AlignmentFlag>

C++ method: const QFlags<Qt::AlignmentFlag>& QStyleOptionViewItem::decorationAlignment() const

fn decoration_position(&self) -> Position

C++ method: QStyleOptionViewItem::Position QStyleOptionViewItem::decorationPosition() const

fn decoration_size<'l0>(&'l0 self) -> &'l0 Size

C++ method: const QSize& QStyleOptionViewItem::decorationSize() const

fn decoration_size_mut<'l0>(&'l0 mut self) -> &'l0 mut Size

C++ method: QSize& QStyleOptionViewItem::decorationSize_mut()

fn display_alignment(&self) -> Flags<AlignmentFlag>

C++ method: const QFlags<Qt::AlignmentFlag>& QStyleOptionViewItem::displayAlignment() const

fn features(&self) -> Flags<ViewItemFeature>

C++ method: const QFlags<QStyleOptionViewItem::ViewItemFeature>& QStyleOptionViewItem::features() const

fn font<'l0>(&'l0 self) -> &'l0 Font

C++ method: const QFont& QStyleOptionViewItem::font() const

fn font_mut<'l0>(&'l0 mut self) -> &'l0 mut Font

C++ method: QFont& QStyleOptionViewItem::font_mut()

fn icon<'l0>(&'l0 self) -> &'l0 Icon

C++ method: const QIcon& QStyleOptionViewItem::icon() const

fn icon_mut<'l0>(&'l0 mut self) -> &'l0 mut Icon

C++ method: QIcon& QStyleOptionViewItem::icon_mut()

fn index<'l0>(&'l0 self) -> &'l0 ModelIndex

C++ method: const QModelIndex& QStyleOptionViewItem::index() const

fn index_mut<'l0>(&'l0 mut self) -> &'l0 mut ModelIndex

C++ method: QModelIndex& QStyleOptionViewItem::index_mut()

fn init(&mut self, w: *const Widget)

C++ method: void QStyleOptionViewItem::init(const QWidget* w)

C++ documentation:

Use initFrom(widget) instead.

fn init_from(&mut self, w: *const Widget)

C++ method: void QStyleOptionViewItem::initFrom(const QWidget* w)

C++ documentation:

Initializes the state, direction, rect, palette, fontMetrics and styleObject member variables based on the specified widget.

This is a convenience function; the member variables can also be initialized manually.

This function was introduced in Qt 4.1.

See also QWidget::layoutDirection(), QWidget::rect(), QWidget::palette(), and QWidget::fontMetrics().

fn locale<'l0>(&'l0 self) -> &'l0 Locale

C++ method: const QLocale& QStyleOptionViewItem::locale() const

fn locale_mut<'l0>(&'l0 mut self) -> &'l0 mut Locale

C++ method: QLocale& QStyleOptionViewItem::locale_mut()

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyleOptionViewItemNewArgs

C++ method: QStyleOptionViewItem::QStyleOptionViewItem

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::style_option_view_item::StyleOptionViewItem
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::style_option_view_item::StyleOptionViewItem>
C++ method: [constructor] void QStyleOptionViewItem::QStyleOptionViewItem()

Variant 2

Rust arguments:
1) fn new((&::style_option_view_item::StyleOptionViewItem, cpp_utils::AsStruct)) -> ::style_option_view_item::StyleOptionViewItem
2) fn new((&::style_option_view_item::StyleOptionViewItem, cpp_utils::AsBox)) -> cpp_utils::CppBox<::style_option_view_item::StyleOptionViewItem>
C++ method: [constructor] void QStyleOptionViewItem::QStyleOptionViewItem(const QStyleOptionViewItem& other)

fn set_background_brush(&mut self, value: &Brush)

C++ method: void QStyleOptionViewItem::set_backgroundBrush(QBrush value)

fn set_check_state(&mut self, value: CheckState)

C++ method: void QStyleOptionViewItem::set_checkState(Qt::CheckState value)

fn set_decoration_alignment(&mut self, value: Flags<AlignmentFlag>)

C++ method: void QStyleOptionViewItem::set_decorationAlignment(QFlags<Qt::AlignmentFlag> value)

fn set_decoration_position(&mut self, value: Position)

C++ method: void QStyleOptionViewItem::set_decorationPosition(QStyleOptionViewItem::Position value)

fn set_decoration_size(&mut self, value: &Size)

C++ method: void QStyleOptionViewItem::set_decorationSize(QSize value)

fn set_display_alignment(&mut self, value: Flags<AlignmentFlag>)

C++ method: void QStyleOptionViewItem::set_displayAlignment(QFlags<Qt::AlignmentFlag> value)

fn set_features(&mut self, value: Flags<ViewItemFeature>)

C++ method: void QStyleOptionViewItem::set_features(QFlags<QStyleOptionViewItem::ViewItemFeature> value)

fn set_font(&mut self, value: &Font)

C++ method: void QStyleOptionViewItem::set_font(QFont value)

fn set_icon(&mut self, value: &Icon)

C++ method: void QStyleOptionViewItem::set_icon(QIcon value)

fn set_index(&mut self, value: &ModelIndex)

C++ method: void QStyleOptionViewItem::set_index(QModelIndex value)

fn set_locale(&mut self, value: &Locale)

C++ method: void QStyleOptionViewItem::set_locale(QLocale value)

fn set_show_decoration_selected(&mut self, value: bool)

C++ method: void QStyleOptionViewItem::set_showDecorationSelected(bool value)

fn set_text(&mut self, value: &String)

C++ method: void QStyleOptionViewItem::set_text(QString value)

fn set_text_elide_mode(&mut self, value: TextElideMode)

C++ method: void QStyleOptionViewItem::set_textElideMode(Qt::TextElideMode value)

fn set_view_item_position(&mut self, value: ViewItemPosition)

C++ method: void QStyleOptionViewItem::set_viewItemPosition(QStyleOptionViewItem::ViewItemPosition value)

fn set_widget(&mut self, value: *const Widget)

C++ method: void QStyleOptionViewItem::set_widget(const QWidget* value)

fn show_decoration_selected(&self) -> bool

C++ method: bool QStyleOptionViewItem::showDecorationSelected() const

fn text<'l0>(&'l0 self) -> &'l0 String

C++ method: const QString& QStyleOptionViewItem::text() const

fn text_elide_mode(&self) -> TextElideMode

C++ method: Qt::TextElideMode QStyleOptionViewItem::textElideMode() const

fn text_mut<'l0>(&'l0 mut self) -> &'l0 mut String

C++ method: QString& QStyleOptionViewItem::text_mut()

fn view_item_position(&self) -> ViewItemPosition

C++ method: QStyleOptionViewItem::ViewItemPosition QStyleOptionViewItem::viewItemPosition() const

fn widget(&self) -> *const Widget

C++ method: const QWidget* QStyleOptionViewItem::widget() const