Module qt_widgets::style_painter::overloading

Traits

StylePainterBeginArgs
This trait represents a set of arguments accepted by StylePainter::begin method.

StylePainterBoundingRectArgs
This trait represents a set of arguments accepted by StylePainter::bounding_rect method.

StylePainterBrushOriginArgs
This trait represents a set of arguments accepted by StylePainter::brush_origin method.

StylePainterClipBoundingRectArgs
This trait represents a set of arguments accepted by StylePainter::clip_bounding_rect method.

StylePainterClipPathArgs
This trait represents a set of arguments accepted by StylePainter::clip_path method.

StylePainterClipRegionArgs
This trait represents a set of arguments accepted by StylePainter::clip_region method.

StylePainterCombinedMatrixArgs
This trait represents a set of arguments accepted by StylePainter::combined_matrix method.

StylePainterCombinedTransformArgs
This trait represents a set of arguments accepted by StylePainter::combined_transform method.

StylePainterDrawArcArgs
This trait represents a set of arguments accepted by StylePainter::draw_arc method.

StylePainterDrawChordArgs
This trait represents a set of arguments accepted by StylePainter::draw_chord method.

StylePainterDrawConvexPolygonArgs
This trait represents a set of arguments accepted by StylePainter::draw_convex_polygon method.

StylePainterDrawEllipseArgs
This trait represents a set of arguments accepted by StylePainter::draw_ellipse method.

StylePainterDrawImageArgs
This trait represents a set of arguments accepted by StylePainter::draw_image method.

StylePainterDrawItemTextArgs
This trait represents a set of arguments accepted by StylePainter::draw_item_text method.

StylePainterDrawLineArgs
This trait represents a set of arguments accepted by StylePainter::draw_line method.

StylePainterDrawLinesArgs
This trait represents a set of arguments accepted by StylePainter::draw_lines method.

StylePainterDrawPictureArgs
This trait represents a set of arguments accepted by StylePainter::draw_picture method.

StylePainterDrawPieArgs
This trait represents a set of arguments accepted by StylePainter::draw_pie method.

StylePainterDrawPixmapArgs
This trait represents a set of arguments accepted by StylePainter::draw_pixmap method.

StylePainterDrawPixmapFragmentsArgs
This trait represents a set of arguments accepted by StylePainter::draw_pixmap_fragments method.

StylePainterDrawPointArgs
This trait represents a set of arguments accepted by StylePainter::draw_point method.

StylePainterDrawPointsArgs
This trait represents a set of arguments accepted by StylePainter::draw_points method.

StylePainterDrawPolygonArgs
This trait represents a set of arguments accepted by StylePainter::draw_polygon method.

StylePainterDrawPolylineArgs
This trait represents a set of arguments accepted by StylePainter::draw_polyline method.

StylePainterDrawRectArgs
This trait represents a set of arguments accepted by StylePainter::draw_rect method.

StylePainterDrawRectsArgs
This trait represents a set of arguments accepted by StylePainter::draw_rects method.

StylePainterDrawRoundRectArgs
This trait represents a set of arguments accepted by StylePainter::draw_round_rect method.

StylePainterDrawRoundedRectArgs
This trait represents a set of arguments accepted by StylePainter::draw_rounded_rect method.

StylePainterDrawStaticTextArgs
This trait represents a set of arguments accepted by StylePainter::draw_static_text method.

StylePainterDrawTextArgs
This trait represents a set of arguments accepted by StylePainter::draw_text method.

StylePainterDrawTextItemArgs
This trait represents a set of arguments accepted by StylePainter::draw_text_item method.

StylePainterDrawTiledPixmapArgs
This trait represents a set of arguments accepted by StylePainter::draw_tiled_pixmap method.

StylePainterEraseRectArgs
This trait represents a set of arguments accepted by StylePainter::erase_rect method.

StylePainterFillRectArgs
This trait represents a set of arguments accepted by StylePainter::fill_rect method.

StylePainterFontInfoArgs
This trait represents a set of arguments accepted by StylePainter::font_info method.

StylePainterFontMetricsArgs
This trait represents a set of arguments accepted by StylePainter::font_metrics method.

StylePainterNewArgs
This trait represents a set of arguments accepted by StylePainter::new method.

StylePainterRedirectedArgs
This trait represents a set of arguments accepted by StylePainter::redirected method.

StylePainterSetBrushArgs
This trait represents a set of arguments accepted by StylePainter::set_brush method.

StylePainterSetBrushOriginArgs
This trait represents a set of arguments accepted by StylePainter::set_brush_origin method.

StylePainterSetClipPathArgs
This trait represents a set of arguments accepted by StylePainter::set_clip_path method.

StylePainterSetClipRectArgs
This trait represents a set of arguments accepted by StylePainter::set_clip_rect method.

StylePainterSetClipRegionArgs
This trait represents a set of arguments accepted by StylePainter::set_clip_region method.

StylePainterSetMatrixArgs
This trait represents a set of arguments accepted by StylePainter::set_matrix method.

StylePainterSetPenArgs
This trait represents a set of arguments accepted by StylePainter::set_pen method.

StylePainterSetRedirectedArgs
This trait represents a set of arguments accepted by StylePainter::set_redirected method.

StylePainterSetRenderHintArgs
This trait represents a set of arguments accepted by StylePainter::set_render_hint method.

StylePainterSetRenderHintsArgs
This trait represents a set of arguments accepted by StylePainter::set_render_hints method.

StylePainterSetTransformArgs
This trait represents a set of arguments accepted by StylePainter::set_transform method.

StylePainterSetViewportArgs
This trait represents a set of arguments accepted by StylePainter::set_viewport method.

StylePainterSetWindowArgs
This trait represents a set of arguments accepted by StylePainter::set_window method.

StylePainterSetWorldMatrixArgs
This trait represents a set of arguments accepted by StylePainter::set_world_matrix method.

StylePainterSetWorldTransformArgs
This trait represents a set of arguments accepted by StylePainter::set_world_transform method.

StylePainterTranslateArgs
This trait represents a set of arguments accepted by StylePainter::translate method.

StylePainterViewportArgs
This trait represents a set of arguments accepted by StylePainter::viewport method.

StylePainterWindowArgs
This trait represents a set of arguments accepted by StylePainter::window method.