Module qt_widgets::style_plugin::overloading

Traits

StylePluginConnectArgs
This trait represents a set of arguments accepted by StylePlugin::connect method.

StylePluginConnectStaticArgs
This trait represents a set of arguments accepted by StylePlugin::connect_static method.

StylePluginDestroyedArgs
This trait represents a set of arguments accepted by StylePlugin::destroyed method.

StylePluginDisconnectArgs
This trait represents a set of arguments accepted by StylePlugin::disconnect method.

StylePluginDisconnectStaticArgs
This trait represents a set of arguments accepted by StylePlugin::disconnect_static method.

StylePluginDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by StylePlugin::dynamic_property_names method.

StylePluginObjectNameArgs
This trait represents a set of arguments accepted by StylePlugin::object_name method.

StylePluginPropertyArgs
This trait represents a set of arguments accepted by StylePlugin::property method.

StylePluginStartTimerArgs
This trait represents a set of arguments accepted by StylePlugin::start_timer method.

StylePluginTrArgs
This trait represents a set of arguments accepted by StylePlugin::tr method.

StylePluginTrUtf8Args
This trait represents a set of arguments accepted by StylePlugin::tr_utf8 method.