Module qt_widgets::styled_item_delegate::overloading

Traits

StyledItemDelegateCloseEditorArgs
This trait represents a set of arguments accepted by StyledItemDelegate::close_editor method.

StyledItemDelegateConnectArgs
This trait represents a set of arguments accepted by StyledItemDelegate::connect method.

StyledItemDelegateConnectStaticArgs
This trait represents a set of arguments accepted by StyledItemDelegate::connect_static method.

StyledItemDelegateDestroyedArgs
This trait represents a set of arguments accepted by StyledItemDelegate::destroyed method.

StyledItemDelegateDisconnectArgs
This trait represents a set of arguments accepted by StyledItemDelegate::disconnect method.

StyledItemDelegateDisconnectStaticArgs
This trait represents a set of arguments accepted by StyledItemDelegate::disconnect_static method.

StyledItemDelegateDisplayTextArgs
This trait represents a set of arguments accepted by StyledItemDelegate::display_text method.

StyledItemDelegateDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by StyledItemDelegate::dynamic_property_names method.

StyledItemDelegateElidedTextArgs
This trait represents a set of arguments accepted by StyledItemDelegate::elided_text method.

StyledItemDelegateNewArgs
This trait represents a set of arguments accepted by StyledItemDelegate::new method.

StyledItemDelegateObjectNameArgs
This trait represents a set of arguments accepted by StyledItemDelegate::object_name method.

StyledItemDelegatePaintingRolesArgs
This trait represents a set of arguments accepted by StyledItemDelegate::painting_roles method.

StyledItemDelegatePropertyArgs
This trait represents a set of arguments accepted by StyledItemDelegate::property method.

StyledItemDelegateSizeHintArgs
This trait represents a set of arguments accepted by StyledItemDelegate::size_hint method.

StyledItemDelegateStartTimerArgs
This trait represents a set of arguments accepted by StyledItemDelegate::start_timer method.

StyledItemDelegateTrArgs
This trait represents a set of arguments accepted by StyledItemDelegate::tr method.

StyledItemDelegateTrUtf8Args
This trait represents a set of arguments accepted by StyledItemDelegate::tr_utf8 method.