Struct qt_widgets::styled_item_delegate::StyledItemDelegate

pub struct StyledItemDelegate {
    // some fields omitted
}

C++ type: QStyledItemDelegate

C++ documentation:

The QStyledItemDelegate class provides display and editing facilities for data items from a model.

When displaying data from models in Qt item views, e.g., a QTableView, the individual items are drawn by a delegate. Also, when an item is edited, it provides an editor widget, which is placed on top of the item view while editing takes place. QStyledItemDelegate is the default delegate for all Qt item views, and is installed upon them when they are created.

The QStyledItemDelegate class is one of the Model/View Classes and is part of Qt's model/view framework. The delegate allows the display and editing of items to be developed independently from the model and view.

The data of items in models are assigned an ItemDataRole; each item can store a QVariant for each role. QStyledItemDelegate implements display and editing for the most common datatypes expected by users, including booleans, integers, and strings.

The data will be drawn differently depending on which role they have in the model. The following table describes the roles and the data types the delegate can handle for each of them. It is often sufficient to ensure that the model returns appropriate data for each of the roles to determine the appearance of items in views.

Role	Accepted Types

Qt::BackgroundRole	QBrush

Qt::BackgroundColorRole	QColor (obsolete; use Qt::BackgroundRole instead)

Qt::CheckStateRole	Qt::CheckState

Qt::DecorationRole	QIcon, QPixmap, QImage and QColor

Qt::DisplayRole	QString and types with a string representation

Qt::EditRole	See QItemEditorFactory for details

Qt::FontRole	QFont

Qt::SizeHintRole	QSize

Qt::TextAlignmentRole	Qt::Alignment

Qt::ForegroundRole	QBrush

Qt::TextColorRole	QColor (obsolete; use Qt::ForegroundRole instead)

Editors are created with a QItemEditorFactory; a default static instance provided by QItemEditorFactory is installed on all item delegates. You can set a custom factory using setItemEditorFactory() or set a new default factory with QItemEditorFactory::setDefaultFactory(). It is the data stored in the item model with the EditRole that is edited. See the QItemEditorFactory class for a more high-level introduction to item editor factories. The Color Editor Factory example shows how to create custom editors with a factory.

Methods

impl StyledItemDelegate

fn block_signals(&mut self, b: bool) -> bool

C++ method: bool QStyledItemDelegate::blockSignals(bool b)

C++ documentation:

If block is true, signals emitted by this object are blocked (i.e., emitting a signal will not invoke anything connected to it). If block is false, no such blocking will occur.

The return value is the previous value of signalsBlocked().

Note that the destroyed() signal will be emitted even if the signals for this object have been blocked.

See also signalsBlocked().

fn children<'l0>(&'l0 self) -> &'l0 ListObjectMutPtr

C++ method: const QList<QObject*>& QStyledItemDelegate::children() const

C++ documentation:

Returns a list of child objects. The QObjectList class is defined in the <QObject> header file as the following:

typedef QList<QObject*> QObjectList;

The first child added is the first object in the list and the last child added is the last object in the list, i.e. new children are appended at the end.

Note that the list order changes when QWidget children are raised or lowered. A widget that is raised becomes the last object in the list, and a widget that is lowered becomes the first object in the list.

See also findChild(), findChildren(), parent(), and setParent().

fn close_editor<'largs, Args>(&'largs mut self, args: Args) where Args: StyledItemDelegateCloseEditorArgs<'largs>

C++ method: QStyledItemDelegate::closeEditor

This is an overloaded function. Available variants:

Rust arguments:
1) fn close_editor(&mut self, *mut ::widget::Widget) -> ()
2) fn close_editor(&mut self, (*mut ::widget::Widget, ::abstract_item_delegate::EndEditHint)) -> ()
C++ method: void QStyledItemDelegate::closeEditor(QWidget* editor, QAbstractItemDelegate::EndEditHint hint = ?)

C++ documentation:

This signal is emitted when the user has finished editing an item using the specified editor.

The hint provides a way for the delegate to influence how the model and view behave after editing is completed. It indicates to these components what action should be performed next to provide a comfortable editing experience for the user. For example, if EditNextItem is specified, the view should use a delegate to open an editor on the next item in the model.

See also EndEditHint.

fn commit_data(&mut self, editor: *mut Widget)

C++ method: void QStyledItemDelegate::commitData(QWidget* editor)

C++ documentation:

This signal must be emitted when the editor widget has completed editing the data, and wants to write it back into the model.

fn connect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StyledItemDelegateConnectArgs<'largs>

C++ method: QStyledItemDelegate::connect

This is an overloaded function. Available variants:

Rust arguments:
1) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: QMetaObject::Connection QStyledItemDelegate::connect(const QObject* sender, const char* signal, const char* member, Qt::ConnectionType type = ?) const

C++ documentation:

This function overloads connect().

Connects signal from the sender object to this object's method.

Equivalent to connect(sender, signal, this, method, type).

Every connection you make emits a signal, so duplicate connections emit two signals. You can break a connection using disconnect().

Note: This function is thread-safe.

See also disconnect().

fn connect_static<Args>(args: Args) -> Args::ReturnType where Args: StyledItemDelegateConnectStaticArgs

C++ method: QStyledItemDelegate::connect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: static QMetaObject::Connection QStyledItemDelegate::connect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& method, Qt::ConnectionType type = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

The Connection handle will be invalid if it cannot create the connection, for example, the parameters were invalid. You can check if the QMetaObject::Connection is valid by casting it to a bool.

This function works in the same way as connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type) but it uses QMetaMethod to specify signal and method.

This function was introduced in Qt 4.8.

See also connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type).

Variant 2

Rust arguments:
1) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: static QMetaObject::Connection QStyledItemDelegate::connect(const QObject* sender, const char* signal, const QObject* receiver, const char* member, Qt::ConnectionType arg5 = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

You must use the SIGNAL() and SLOT() macros when specifying the signal and the method, for example:

QLabel label = new QLabel;
QScrollBar scrollBar = new QScrollBar;
QObject::connect(scrollBar, SIGNAL(valueChanged(int)),
                 label,  SLOT(setNum(int)));

This example ensures that the label always displays the current scroll bar value. Note that the signal and slots parameters must not contain any variable names, only the type. E.g. the following would not work and return false:

// WRONG
QObject::connect(scrollBar, SIGNAL(valueChanged(int value)),
                 label, SLOT(setNum(int value)));

A signal can also be connected to another signal:

class MyWidget : public QWidget
{
    Q_OBJECT

public: MyWidget();

signals: void buttonClicked();

private: QPushButton *myButton; };

MyWidget::MyWidget() { myButton = new QPushButton(this); connect(myButton, SIGNAL(clicked()), this, SIGNAL(buttonClicked())); }
In this example, the MyWidget constructor relays a signal from a private member variable, and makes it available under a name that relates to MyWidget.

A signal can be connected to many slots and signals. Many signals can be connected to one slot.

If a signal is connected to several slots, the slots are activated in the same order in which the connections were made, when the signal is emitted.

The function returns a QMetaObject::Connection that represents a handle to a connection if it successfully connects the signal to the slot. The connection handle will be invalid if it cannot create the connection, for example, if QObject is unable to verify the existence of either signal or method, or if their signatures aren't compatible. You can check if the handle is valid by casting it to a bool.

By default, a signal is emitted for every connection you make; two signals are emitted for duplicate connections. You can break all of these connections with a single disconnect() call. If you pass the Qt::UniqueConnection type, the connection will only be made if it is not a duplicate. If there is already a duplicate (exact same signal to the exact same slot on the same objects), the connection will fail and connect will return an invalid QMetaObject::Connection.

The optional type parameter describes the type of connection to establish. In particular, it determines whether a particular signal is delivered to a slot immediately or queued for delivery at a later time. If the signal is queued, the parameters must be of types that are known to Qt's meta-object system, because Qt needs to copy the arguments to store them in an event behind the scenes. If you try to use a queued connection and get the error message

QObject::connect: Cannot queue arguments of type 'MyType'
(Make sure 'MyType' is registered using qRegisterMetaType().)

call qRegisterMetaType() to register the data type before you establish the connection.

Note: This function is thread-safe.

See also disconnect(), sender(), qRegisterMetaType(), and Q_DECLARE_METATYPE().

fn create_editor(&self, parent: *mut Widget, option: &StyleOptionViewItem, index: &ModelIndex) -> *mut Widget

C++ method: virtual QWidget* QStyledItemDelegate::createEditor(QWidget* parent, const QStyleOptionViewItem& option, const QModelIndex& index) const

C++ documentation:

Reimplemented from QAbstractItemDelegate::createEditor().

Returns the widget used to edit the item specified by index for editing. The parent widget and style option are used to control how the editor widget appears.

See also QAbstractItemDelegate::createEditor().

fn delete_later(&mut self)

C++ method: void QStyledItemDelegate::deleteLater()

C++ documentation:

Schedules this object for deletion.

The object will be deleted when control returns to the event loop. If the event loop is not running when this function is called (e.g. deleteLater() is called on an object before QCoreApplication::exec()), the object will be deleted once the event loop is started. If deleteLater() is called after the main event loop has stopped, the object will not be deleted. Since Qt 4.8, if deleteLater() is called on an object that lives in a thread with no running event loop, the object will be destroyed when the thread finishes.

Note that entering and leaving a new event loop (e.g., by opening a modal dialog) will not perform the deferred deletion; for the object to be deleted, the control must return to the event loop from which deleteLater() was called.

Note: It is safe to call this function more than once; when the first deferred deletion event is delivered, any pending events for the object are removed from the event queue.

See also destroyed() and QPointer.

fn destroy_editor(&self, editor: *mut Widget, index: &ModelIndex)

C++ method: virtual void QStyledItemDelegate::destroyEditor(QWidget* editor, const QModelIndex& index) const

C++ documentation:

Called when the editor is no longer needed for editing the data item with the given index and should be destroyed. The default behavior is a call to deleteLater on the editor. It is possible e.g. to avoid this delete by reimplementing this function.

This function was introduced in Qt 5.0.

See also createEditor().

fn destroyed<'largs, Args>(&'largs mut self, args: Args) where Args: StyledItemDelegateDestroyedArgs<'largs>

C++ method: QStyledItemDelegate::destroyed

This is an overloaded function. Available variants:

Rust arguments:
1) fn destroyed(&mut self, *mut qt_core::object::Object) -> ()
2) fn destroyed(&mut self, ()) -> ()
C++ method: void QStyledItemDelegate::destroyed(QObject* arg1 = ?)

C++ documentation:

This signal is emitted immediately before the object obj is destroyed, and can not be blocked.

All the objects's children are destroyed immediately after this signal is emitted.

See also deleteLater() and QPointer.

fn disconnect<'largs, Args>(&'largs self, args: Args) -> bool where Args: StyledItemDelegateDisconnectArgs<'largs>

C++ method: QStyledItemDelegate::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn disconnect(&self, *const qt_core::object::Object) -> bool
2) fn disconnect(&self, (*const qt_core::object::Object, *const libc::c_char)) -> bool
C++ method: bool QStyledItemDelegate::disconnect(const QObject* receiver, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects all signals in this object from receiver's method.

A signal-slot connection is removed when either of the objects involved are destroyed.

Variant 2

Rust arguments:
1) fn disconnect(&self, *const libc::c_char) -> bool
2) fn disconnect(&self, (*const libc::c_char, *const qt_core::object::Object)) -> bool
3) fn disconnect(&self, (*const libc::c_char, *const qt_core::object::Object, *const libc::c_char)) -> bool
4) fn disconnect(&self, ()) -> bool
C++ method: bool QStyledItemDelegate::disconnect(const char* signal = ?, const QObject* receiver = ?, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects signal from method of receiver.

A signal-slot connection is removed when either of the objects involved are destroyed.

Note: This function is thread-safe.

fn disconnect_static<Args>(args: Args) -> bool where Args: StyledItemDelegateDisconnectStaticArgs

C++ method: QStyledItemDelegate::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn disconnect_static(&qt_core::meta_object::Connection) -> bool
C++ method: static bool QStyledItemDelegate::disconnect(const QMetaObject::Connection& arg1)

C++ documentation:

Disconnect a connection.

If the connection is invalid or has already been disconnected, do nothing and return false.

See also connect().

Variant 2

Rust arguments: fn disconnect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod)) -> bool
C++ method: static bool QStyledItemDelegate::disconnect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

This function provides the same possibilities like disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method) but uses QMetaMethod to represent the signal and the method to be disconnected.

Additionally this function returnsfalse and no signals and slots disconnected if:

signal is not a member of sender class or one of its parent classes.

method is not a member of receiver class or one of its parent classes.

signal instance represents not a signal.

QMetaMethod() may be used as wildcard in the meaning "any signal" or "any slot in receiving object". In the same way 0 can be used for receiver in the meaning "any receiving object". In this case method should also be QMetaMethod(). sender parameter should be never 0.

This function was introduced in Qt 4.8.

See also disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method).

Variant 3

Rust arguments: fn disconnect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char)) -> bool
C++ method: static bool QStyledItemDelegate::disconnect(const QObject* sender, const char* signal, const QObject* receiver, const char* member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

A signal-slot connection is removed when either of the objects involved are destroyed.

disconnect() is typically used in three ways, as the following examples demonstrate.

Disconnect everything connected to an object's signals:
disconnect(myObject, 0, 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect();

Disconnect everything connected to a specific signal:
disconnect(myObject, SIGNAL(mySignal()), 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect(SIGNAL(mySignal()));

Disconnect a specific receiver:
disconnect(myObject, 0, myReceiver, 0);

equivalent to the non-static overloaded function

myObject->disconnect(myReceiver);

0 may be used as a wildcard, meaning "any signal", "any receiving object", or "any slot in the receiving object", respectively.

The sender may never be 0. (You cannot disconnect signals from more than one object in a single call.)

If signal is 0, it disconnects receiver and method from any signal. If not, only the specified signal is disconnected.

If receiver is 0, it disconnects anything connected to signal. If not, slots in objects other than receiver are not disconnected.

If method is 0, it disconnects anything that is connected to receiver. If not, only slots named method will be disconnected, and all other slots are left alone. The method must be 0 if receiver is left out, so you cannot disconnect a specifically-named slot on all objects.

Note: This function is thread-safe.

See also connect().

fn display_text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StyledItemDelegateDisplayTextArgs<'largs>

C++ method: QStyledItemDelegate::displayText

This is an overloaded function. Available variants:

Rust arguments:
1) fn display_text(&self, (&qt_core::variant::Variant, &qt_core::locale::Locale, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn display_text(&self, (&qt_core::variant::Variant, &qt_core::locale::Locale, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QStyledItemDelegate::displayText(const QVariant& value, const QLocale& locale) const

C++ documentation:

This function returns the string that the delegate will use to display the Qt::DisplayRole of the model in locale. value is the value of the Qt::DisplayRole provided by the model.

The default implementation uses the QLocale::toString to convert value into a QString.

This function is not called for empty model indices, i.e., indices for which the model returns an invalid QVariant.

See also QAbstractItemModel::data().

fn dump_object_info(&mut self)

C++ method: void QStyledItemDelegate::dumpObjectInfo()

C++ documentation:

Dumps information about signal connections, etc. for this object to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectTree().

fn dump_object_tree(&mut self)

C++ method: void QStyledItemDelegate::dumpObjectTree()

C++ documentation:

Dumps a tree of children to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectInfo().

fn dynamic_property_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StyledItemDelegateDynamicPropertyNamesArgs<'largs>

C++ method: QStyledItemDelegate::dynamicPropertyNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn dynamic_property_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn dynamic_property_names(&self, cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: QList<QByteArray> QStyledItemDelegate::dynamicPropertyNames() const

C++ documentation:

Returns the names of all properties that were dynamically added to the object using setProperty().

This function was introduced in Qt 4.2.

fn elided_text<Args>(args: Args) -> Args::ReturnType where Args: StyledItemDelegateElidedTextArgs

C++ method: QStyledItemDelegate::elidedText

This is an overloaded function. Available variants:

Rust arguments:
1) fn elided_text((&qt_gui::font_metrics::FontMetrics, libc::c_int, qt_core::qt::TextElideMode, &qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn elided_text((&qt_gui::font_metrics::FontMetrics, libc::c_int, qt_core::qt::TextElideMode, &qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QStyledItemDelegate::elidedText(const QFontMetrics& fontMetrics, int width, Qt::TextElideMode mode, const QString& text)

C++ documentation:

Use QFontMetrics::elidedText() instead.

For example, if you have code like

QFontMetrics fm = ...
QString str = QAbstractItemDelegate::elidedText(fm, width, mode, text);

you can rewrite it as

QFontMetrics fm = ...
QString str = fm.elidedText(text, mode, width);

fn event(&mut self, arg1: *mut Event) -> bool

C++ method: virtual bool QStyledItemDelegate::event(QEvent* arg1)

C++ documentation:

This virtual function receives events to an object and should return true if the event e was recognized and processed.

The event() function can be reimplemented to customize the behavior of an object.

See also installEventFilter(), timerEvent(), QCoreApplication::sendEvent(), and QCoreApplication::postEvent().

fn help_event(&mut self, event: *mut HelpEvent, view: *mut AbstractItemView, option: &StyleOptionViewItem, index: &ModelIndex) -> bool

C++ method: virtual bool QStyledItemDelegate::helpEvent(QHelpEvent* event, QAbstractItemView* view, const QStyleOptionViewItem& option, const QModelIndex& index)

C++ documentation:

Whenever a help event occurs, this function is called with the event view option and the index that corresponds to the item where the event occurs.

Returns true if the delegate can handle the event; otherwise returns false. A return value of true indicates that the data obtained using the index had the required role.

For QEvent::ToolTip and QEvent::WhatsThis events that were handled successfully, the relevant popup may be shown depending on the user's system configuration.

This function was introduced in Qt 4.3.

See also QHelpEvent.

fn inherits(&self, classname: *const c_char) -> bool

C++ method: bool QStyledItemDelegate::inherits(const char* classname) const

C++ documentation:

Returns true if this object is an instance of a class that inherits className or a QObject subclass that inherits className; otherwise returns false.

A class is considered to inherit itself.

Example:

QTimer *timer = new QTimer;         // QTimer inherits QObject
timer->inherits("QTimer");          // returns true
timer->inherits("QObject");         // returns true
timer->inherits("QAbstractButton"); // returns false

// QVBoxLayout inherits QObject and QLayoutItem QVBoxLayout *layout = new QVBoxLayout; layout->inherits("QObject"); // returns true layout->inherits("QLayoutItem"); // returns true (even though QLayoutItem is not a QObject)
If you need to determine whether an object is an instance of a particular class for the purpose of casting it, consider using qobject_cast<Type *>(object) instead.

See also metaObject() and qobject_cast().

fn install_event_filter(&mut self, arg1: *mut Object)

C++ method: void QStyledItemDelegate::installEventFilter(QObject* arg1)

C++ documentation:

Installs an event filter filterObj on this object. For example:

monitoredObj->installEventFilter(filterObj);

An event filter is an object that receives all events that are sent to this object. The filter can either stop the event or forward it to this object. The event filter filterObj receives events via its eventFilter() function. The eventFilter() function must return true if the event should be filtered, (i.e. stopped); otherwise it must return false.

If multiple event filters are installed on a single object, the filter that was installed last is activated first.

Here's a KeyPressEater class that eats the key presses of its monitored objects:

class KeyPressEater : public QObject
{
    Q_OBJECT
    ...

protected: bool eventFilter(QObject obj, QEvent event); };

bool KeyPressEater::eventFilter(QObject obj, QEvent event) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent >(event); qDebug("Ate key press %d", keyEvent->key()); return true; } else { // standard event processing return QObject::eventFilter(obj, event); } }
And here's how to install it on two widgets:

KeyPressEater keyPressEater = new KeyPressEater(this);
QPushButton pushButton = new QPushButton(this);
QListView *listView = new QListView(this);

pushButton->installEventFilter(keyPressEater); listView->installEventFilter(keyPressEater);
The QShortcut class, for example, uses this technique to intercept shortcut key presses.

Warning: If you delete the receiver object in your eventFilter() function, be sure to return true. If you return false, Qt sends the event to the deleted object and the program will crash.

Note that the filtering object must be in the same thread as this object. If filterObj is in a different thread, this function does nothing. If either filterObj or this object are moved to a different thread after calling this function, the event filter will not be called until both objects have the same thread affinity again (it is not removed).

See also removeEventFilter(), eventFilter(), and event().

fn is_widget_type(&self) -> bool

C++ method: bool QStyledItemDelegate::isWidgetType() const

C++ documentation:

Returns true if the object is a widget; otherwise returns false.

Calling this function is equivalent to calling inherits("QWidget"), except that it is much faster.

fn is_window_type(&self) -> bool

C++ method: bool QStyledItemDelegate::isWindowType() const

C++ documentation:

Returns true if the object is a window; otherwise returns false.

Calling this function is equivalent to calling inherits("QWindow"), except that it is much faster.

fn item_editor_factory(&self) -> *mut ItemEditorFactory

C++ method: QItemEditorFactory* QStyledItemDelegate::itemEditorFactory() const

C++ documentation:

Returns the editor factory used by the item delegate. If no editor factory is set, the function will return null.

See also setItemEditorFactory().

fn kill_timer(&mut self, id: c_int)

C++ method: void QStyledItemDelegate::killTimer(int id)

C++ documentation:

Kills the timer with timer identifier, id.

The timer identifier is returned by startTimer() when a timer event is started.

See also timerEvent() and startTimer().

fn meta_object(&self) -> *const MetaObject

C++ method: virtual const QMetaObject* QStyledItemDelegate::metaObject() const

fn move_to_thread(&mut self, thread: *mut Thread)

C++ method: void QStyledItemDelegate::moveToThread(QThread* thread)

C++ documentation:

Changes the thread affinity for this object and its children. The object cannot be moved if it has a parent. Event processing will continue in the targetThread.

To move an object to the main thread, use QApplication::instance() to retrieve a pointer to the current application, and then use QApplication::thread() to retrieve the thread in which the application lives. For example:

myObject->moveToThread(QApplication::instance()->thread());

If targetThread is zero, all event processing for this object and its children stops.

Note that all active timers for the object will be reset. The timers are first stopped in the current thread and restarted (with the same interval) in the targetThread. As a result, constantly moving an object between threads can postpone timer events indefinitely.

A QEvent::ThreadChange event is sent to this object just before the thread affinity is changed. You can handle this event to perform any special processing. Note that any new events that are posted to this object will be handled in the targetThread.

Warning: This function is not thread-safe; the current thread must be same as the current thread affinity. In other words, this function can only "push" an object from the current thread to another thread, it cannot "pull" an object from any arbitrary thread to the current thread.

See also thread().

fn new<Args>(args: Args) -> Args::ReturnType where Args: StyledItemDelegateNewArgs

C++ method: QStyledItemDelegate::QStyledItemDelegate

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::styled_item_delegate::StyledItemDelegate
2) fn new((*mut qt_core::object::Object, cpp_utils::AsStruct)) -> ::styled_item_delegate::StyledItemDelegate
3) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::styled_item_delegate::StyledItemDelegate>
4) fn new((*mut qt_core::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::styled_item_delegate::StyledItemDelegate>
C++ method: [constructor] void QStyledItemDelegate::QStyledItemDelegate(QObject* parent = ?)

fn object_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StyledItemDelegateObjectNameArgs<'largs>

C++ method: QStyledItemDelegate::objectName

This is an overloaded function. Available variants:

Rust arguments:
1) fn object_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn object_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QStyledItemDelegate::objectName() const

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn paint(&self, painter: *mut Painter, option: &StyleOptionViewItem, index: &ModelIndex)

C++ method: virtual void QStyledItemDelegate::paint(QPainter* painter, const QStyleOptionViewItem& option, const QModelIndex& index) const

C++ documentation:

Reimplemented from QAbstractItemDelegate::paint().

Renders the delegate using the given painter and style option for the item specified by index.

This function paints the item using the view's QStyle.

When reimplementing paint in a subclass. Use the initStyleOption() to set up the option in the same way as the QStyledItemDelegate.

Whenever possible, use the option while painting. Especially its rect variable to decide where to draw and its state to determine if it is enabled or selected.

After painting, you should ensure that the painter is returned to its the state it was supplied in when this function was called. For example, it may be useful to call QPainter::save() before painting and QPainter::restore() afterwards.

See also QItemDelegate::paint(), QStyle::drawControl(), and QStyle::CE_ItemViewItem.

fn painting_roles<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StyledItemDelegatePaintingRolesArgs<'largs>

C++ method: QStyledItemDelegate::paintingRoles

This is an overloaded function. Available variants:

Rust arguments:
1) fn painting_roles(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::vector::VectorCInt>
2) fn painting_roles(&self, cpp_utils::AsStruct) -> qt_core::vector::VectorCInt
C++ method: virtual QVector<int> QStyledItemDelegate::paintingRoles() const

fn parent(&self) -> *mut Object

C++ method: QObject* QStyledItemDelegate::parent() const

C++ documentation:

Returns a pointer to the parent object.

See also setParent() and children().

fn property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StyledItemDelegatePropertyArgs<'largs>

C++ method: QStyledItemDelegate::property

This is an overloaded function. Available variants:

Rust arguments:
1) fn property(&self, (*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn property(&self, (*const libc::c_char, cpp_utils::AsStruct)) -> qt_core::variant::Variant
C++ method: QVariant QStyledItemDelegate::property(const char* name) const

C++ documentation:

Returns the value of the object's name property.

If no such property exists, the returned variant is invalid.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

See also setProperty(), QVariant::isValid(), metaObject(), and dynamicPropertyNames().

fn qt_metacall(&mut self, arg1: Call, arg2: c_int, arg3: *mut *mut c_void) -> c_int

C++ method: virtual int QStyledItemDelegate::qt_metacall(QMetaObject::Call arg1, int arg2, void** arg3)

fn qt_metacast(&mut self, arg1: *const c_char) -> *mut c_void

C++ method: virtual void* QStyledItemDelegate::qt_metacast(const char* arg1)

fn register_user_data() -> c_uint

C++ method: static unsigned int QStyledItemDelegate::registerUserData()

fn remove_event_filter(&mut self, arg1: *mut Object)

C++ method: void QStyledItemDelegate::removeEventFilter(QObject* arg1)

C++ documentation:

Removes an event filter object obj from this object. The request is ignored if such an event filter has not been installed.

All event filters for this object are automatically removed when this object is destroyed.

It is always safe to remove an event filter, even during event filter activation (i.e. from the eventFilter() function).

See also installEventFilter(), eventFilter(), and event().

fn set_editor_data(&self, editor: *mut Widget, index: &ModelIndex)

C++ method: virtual void QStyledItemDelegate::setEditorData(QWidget* editor, const QModelIndex& index) const

C++ documentation:

Reimplemented from QAbstractItemDelegate::setEditorData().

Sets the data to be displayed and edited by the editor from the data model item specified by the model index.

The default implementation stores the data in the editor widget's user property.

See also QMetaProperty::isUser().

fn set_item_editor_factory(&mut self, factory: *mut ItemEditorFactory)

C++ method: void QStyledItemDelegate::setItemEditorFactory(QItemEditorFactory* factory)

C++ documentation:

Sets the editor factory to be used by the item delegate to be the factory specified. If no editor factory is set, the item delegate will use the default editor factory.

See also itemEditorFactory().

fn set_model_data(&self, editor: *mut Widget, model: *mut AbstractItemModel, index: &ModelIndex)

C++ method: virtual void QStyledItemDelegate::setModelData(QWidget* editor, QAbstractItemModel* model, const QModelIndex& index) const

C++ documentation:

Reimplemented from QAbstractItemDelegate::setModelData().

Gets data from the editor widget and stores it in the specified model at the item index.

The default implementation gets the value to be stored in the data model from the editor widget's user property.

See also QMetaProperty::isUser().

fn set_object_name(&mut self, name: &String)

C++ method: void QStyledItemDelegate::setObjectName(const QString& name)

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn set_parent(&mut self, arg1: *mut Object)

C++ method: void QStyledItemDelegate::setParent(QObject* arg1)

C++ documentation:

Makes the object a child of parent.

See also parent() and children().

fn set_property(&mut self, name: *const c_char, value: &Variant) -> bool

C++ method: bool QStyledItemDelegate::setProperty(const char* name, const QVariant& value)

C++ documentation:

Sets the value of the object's name property to value.

If the property is defined in the class using Q_PROPERTY then true is returned on success and false otherwise. If the property is not defined using Q_PROPERTY, and therefore not listed in the meta-object, it is added as a dynamic property and false is returned.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

Dynamic properties can be queried again using property() and can be removed by setting the property value to an invalid QVariant. Changing the value of a dynamic property causes a QDynamicPropertyChangeEvent to be sent to the object.

Note: Dynamic properties starting with "_q_" are reserved for internal purposes.

See also property(), metaObject(), and dynamicPropertyNames().

fn set_user_data(&mut self, id: c_uint, data: *mut ObjectUserData)

C++ method: void QStyledItemDelegate::setUserData(unsigned int id, QObjectUserData* data)

fn signals_blocked(&self) -> bool

C++ method: bool QStyledItemDelegate::signalsBlocked() const

C++ documentation:

Returns true if signals are blocked; otherwise returns false.

Signals are not blocked by default.

See also blockSignals().

fn size_hint<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StyledItemDelegateSizeHintArgs<'largs>

C++ method: QStyledItemDelegate::sizeHint

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_hint(&self, (&::style_option_view_item::StyleOptionViewItem, &qt_core::model_index::ModelIndex, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size_hint(&self, (&::style_option_view_item::StyleOptionViewItem, &qt_core::model_index::ModelIndex, cpp_utils::AsStruct)) -> qt_core::size::Size
C++ method: virtual QSize QStyledItemDelegate::sizeHint(const QStyleOptionViewItem& option, const QModelIndex& index) const

C++ documentation:

Reimplemented from QAbstractItemDelegate::sizeHint().

Returns the size needed by the delegate to display the item specified by index, taking into account the style information provided by option.

This function uses the view's QStyle to determine the size of the item.

See also QStyle::sizeFromContents() and QStyle::CT_ItemViewItem.

fn size_hint_changed(&mut self, arg1: &ModelIndex)

C++ method: void QStyledItemDelegate::sizeHintChanged(const QModelIndex& arg1)

C++ documentation:

This signal must be emitted when the sizeHint() of index changed.

Views automatically connect to this signal and relayout items as necessary.

This function was introduced in Qt 4.4.

fn start_timer<'largs, Args>(&'largs mut self, args: Args) -> c_int where Args: StyledItemDelegateStartTimerArgs<'largs>

C++ method: QStyledItemDelegate::startTimer

This is an overloaded function. Available variants:

Rust arguments:
1) fn start_timer(&mut self, libc::c_int) -> libc::c_int
2) fn start_timer(&mut self, (libc::c_int, qt_core::qt::TimerType)) -> libc::c_int
C++ method: int QStyledItemDelegate::startTimer(int interval, Qt::TimerType timerType = ?)

C++ documentation:

Starts a timer and returns a timer identifier, or returns zero if it could not start a timer.

A timer event will occur every interval milliseconds until killTimer() is called. If interval is 0, then the timer event occurs once every time there are no more window system events to process.

The virtual timerEvent() function is called with the QTimerEvent event parameter class when a timer event occurs. Reimplement this function to get timer events.

If multiple timers are running, the QTimerEvent::timerId() can be used to find out which timer was activated.

Example:

class MyObject : public QObject
{
    Q_OBJECT

public: MyObject(QObject *parent = 0);

protected: void timerEvent(QTimerEvent *event); };

MyObject::MyObject(QObject *parent) : QObject(parent) { startTimer(50); // 50-millisecond timer startTimer(1000); // 1-second timer startTimer(60000); // 1-minute timer }

void MyObject::timerEvent(QTimerEvent *event) { qDebug() << "Timer ID:" << event->timerId(); }
Note that QTimer's accuracy depends on the underlying operating system and hardware. The timerType argument allows you to customize the accuracy of the timer. See Qt::TimerType for information on the different timer types. Most platforms support an accuracy of 20 milliseconds; some provide more. If Qt is unable to deliver the requested number of timer events, it will silently discard some.

The QTimer class provides a high-level programming interface with single-shot timers and timer signals instead of events. There is also a QBasicTimer class that is more lightweight than QTimer and less clumsy than using timer IDs directly.

See also timerEvent(), killTimer(), and QTimer::singleShot().

fn thread(&self) -> *mut Thread

C++ method: QThread* QStyledItemDelegate::thread() const

C++ documentation:

Returns the thread in which the object lives.

See also moveToThread().

fn tr<Args>(args: Args) -> Args::ReturnType where Args: StyledItemDelegateTrArgs

C++ method: QStyledItemDelegate::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QStyledItemDelegate::tr(const char* s, const char* c, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: StyledItemDelegateTrUtf8Args

C++ method: QStyledItemDelegate::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QStyledItemDelegate::trUtf8(const char* s, const char* c, int n)

fn update_editor_geometry(&self, editor: *mut Widget, option: &StyleOptionViewItem, index: &ModelIndex)

C++ method: virtual void QStyledItemDelegate::updateEditorGeometry(QWidget* editor, const QStyleOptionViewItem& option, const QModelIndex& index) const

C++ documentation:

Reimplemented from QAbstractItemDelegate::updateEditorGeometry().

Updates the editor for the item specified by index according to the style option given.

fn user_data(&self, id: c_uint) -> *mut ObjectUserData

C++ method: QObjectUserData* QStyledItemDelegate::userData(unsigned int id) const