Module qt_widgets::swipe_gesture::overloading

Traits

SwipeGestureConnectArgs
This trait represents a set of arguments accepted by SwipeGesture::connect method.

SwipeGestureConnectStaticArgs
This trait represents a set of arguments accepted by SwipeGesture::connect_static method.

SwipeGestureDestroyedArgs
This trait represents a set of arguments accepted by SwipeGesture::destroyed method.

SwipeGestureDisconnectArgs
This trait represents a set of arguments accepted by SwipeGesture::disconnect method.

SwipeGestureDisconnectStaticArgs
This trait represents a set of arguments accepted by SwipeGesture::disconnect_static method.

SwipeGestureDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by SwipeGesture::dynamic_property_names method.

SwipeGestureHotSpotArgs
This trait represents a set of arguments accepted by SwipeGesture::hot_spot method.

SwipeGestureNewArgs
This trait represents a set of arguments accepted by SwipeGesture::new method.

SwipeGestureObjectNameArgs
This trait represents a set of arguments accepted by SwipeGesture::object_name method.

SwipeGesturePropertyArgs
This trait represents a set of arguments accepted by SwipeGesture::property method.

SwipeGestureStartTimerArgs
This trait represents a set of arguments accepted by SwipeGesture::start_timer method.

SwipeGestureTrArgs
This trait represents a set of arguments accepted by SwipeGesture::tr method.

SwipeGestureTrUtf8Args
This trait represents a set of arguments accepted by SwipeGesture::tr_utf8 method.