Module qt_widgets::system_tray_icon::overloading

Traits

SystemTrayIconConnectArgs
This trait represents a set of arguments accepted by SystemTrayIcon::connect method.

SystemTrayIconConnectStaticArgs
This trait represents a set of arguments accepted by SystemTrayIcon::connect_static method.

SystemTrayIconDestroyedArgs
This trait represents a set of arguments accepted by SystemTrayIcon::destroyed method.

SystemTrayIconDisconnectArgs
This trait represents a set of arguments accepted by SystemTrayIcon::disconnect method.

SystemTrayIconDisconnectStaticArgs
This trait represents a set of arguments accepted by SystemTrayIcon::disconnect_static method.

SystemTrayIconDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by SystemTrayIcon::dynamic_property_names method.

SystemTrayIconGeometryArgs
This trait represents a set of arguments accepted by SystemTrayIcon::geometry method.

SystemTrayIconIconArgs
This trait represents a set of arguments accepted by SystemTrayIcon::icon method.

SystemTrayIconNewArgs
This trait represents a set of arguments accepted by SystemTrayIcon::new method.

SystemTrayIconObjectNameArgs
This trait represents a set of arguments accepted by SystemTrayIcon::object_name method.

SystemTrayIconPropertyArgs
This trait represents a set of arguments accepted by SystemTrayIcon::property method.

SystemTrayIconShowMessageArgs
This trait represents a set of arguments accepted by SystemTrayIcon::show_message method.

SystemTrayIconStartTimerArgs
This trait represents a set of arguments accepted by SystemTrayIcon::start_timer method.

SystemTrayIconToolTipArgs
This trait represents a set of arguments accepted by SystemTrayIcon::tool_tip method.

SystemTrayIconTrArgs
This trait represents a set of arguments accepted by SystemTrayIcon::tr method.

SystemTrayIconTrUtf8Args
This trait represents a set of arguments accepted by SystemTrayIcon::tr_utf8 method.