Module qt_widgets::tab_bar::overloading

Traits

TabBarAccessibleDescriptionArgs
This trait represents a set of arguments accepted by TabBar::accessible_description method.

TabBarAccessibleNameArgs
This trait represents a set of arguments accepted by TabBar::accessible_name method.

TabBarActionsArgs
This trait represents a set of arguments accepted by TabBar::actions method.

TabBarAddTabArgs
This trait represents a set of arguments accepted by TabBar::add_tab method.

TabBarBaseSizeArgs
This trait represents a set of arguments accepted by TabBar::base_size method.

TabBarChildAtArgs
This trait represents a set of arguments accepted by TabBar::child_at method.

TabBarChildrenRectArgs
This trait represents a set of arguments accepted by TabBar::children_rect method.

TabBarChildrenRegionArgs
This trait represents a set of arguments accepted by TabBar::children_region method.

TabBarConnectArgs
This trait represents a set of arguments accepted by TabBar::connect method.

TabBarConnectStaticArgs
This trait represents a set of arguments accepted by TabBar::connect_static method.

TabBarContentsMarginsArgs
This trait represents a set of arguments accepted by TabBar::contents_margins method.

TabBarContentsRectArgs
This trait represents a set of arguments accepted by TabBar::contents_rect method.

TabBarCreateWindowContainerArgs
This trait represents a set of arguments accepted by TabBar::create_window_container method.

TabBarCursorArgs
This trait represents a set of arguments accepted by TabBar::cursor method.

TabBarDestroyedArgs
This trait represents a set of arguments accepted by TabBar::destroyed method.

TabBarDisconnectArgs
This trait represents a set of arguments accepted by TabBar::disconnect method.

TabBarDisconnectStaticArgs
This trait represents a set of arguments accepted by TabBar::disconnect_static method.

TabBarDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by TabBar::dynamic_property_names method.

TabBarFontInfoArgs
This trait represents a set of arguments accepted by TabBar::font_info method.

TabBarFontMetricsArgs
This trait represents a set of arguments accepted by TabBar::font_metrics method.

TabBarFrameGeometryArgs
This trait represents a set of arguments accepted by TabBar::frame_geometry method.

TabBarFrameSizeArgs
This trait represents a set of arguments accepted by TabBar::frame_size method.

TabBarGrabArgs
This trait represents a set of arguments accepted by TabBar::grab method.

TabBarGrabGestureArgs
This trait represents a set of arguments accepted by TabBar::grab_gesture method.

TabBarGrabMouseArgs
This trait represents a set of arguments accepted by TabBar::grab_mouse method.

TabBarGrabShortcutArgs
This trait represents a set of arguments accepted by TabBar::grab_shortcut method.

TabBarIconSizeArgs
This trait represents a set of arguments accepted by TabBar::icon_size method.

TabBarInputMethodQueryArgs
This trait represents a set of arguments accepted by TabBar::input_method_query method.

TabBarInsertTabArgs
This trait represents a set of arguments accepted by TabBar::insert_tab method.

TabBarLocaleArgs
This trait represents a set of arguments accepted by TabBar::locale method.

TabBarMapFromArgs
This trait represents a set of arguments accepted by TabBar::map_from method.

TabBarMapFromGlobalArgs
This trait represents a set of arguments accepted by TabBar::map_from_global method.

TabBarMapFromParentArgs
This trait represents a set of arguments accepted by TabBar::map_from_parent method.

TabBarMapToArgs
This trait represents a set of arguments accepted by TabBar::map_to method.

TabBarMapToGlobalArgs
This trait represents a set of arguments accepted by TabBar::map_to_global method.

TabBarMapToParentArgs
This trait represents a set of arguments accepted by TabBar::map_to_parent method.

TabBarMaskArgs
This trait represents a set of arguments accepted by TabBar::mask method.

TabBarMaximumSizeArgs
This trait represents a set of arguments accepted by TabBar::maximum_size method.

TabBarMinimumSizeArgs
This trait represents a set of arguments accepted by TabBar::minimum_size method.

TabBarMinimumSizeHintArgs
This trait represents a set of arguments accepted by TabBar::minimum_size_hint method.

TabBarMoveArgs
This trait represents a set of arguments accepted by TabBar::move_ method.

TabBarNewArgs
This trait represents a set of arguments accepted by TabBar::new method.

TabBarNormalGeometryArgs
This trait represents a set of arguments accepted by TabBar::normal_geometry method.

TabBarObjectNameArgs
This trait represents a set of arguments accepted by TabBar::object_name method.

TabBarPosArgs
This trait represents a set of arguments accepted by TabBar::pos method.

TabBarPropertyArgs
This trait represents a set of arguments accepted by TabBar::property method.

TabBarRectArgs
This trait represents a set of arguments accepted by TabBar::rect method.

TabBarRenderArgs
This trait represents a set of arguments accepted by TabBar::render method.

TabBarRepaintArgs
This trait represents a set of arguments accepted by TabBar::repaint method.

TabBarResizeArgs
This trait represents a set of arguments accepted by TabBar::resize method.

TabBarSaveGeometryArgs
This trait represents a set of arguments accepted by TabBar::save_geometry method.

TabBarScrollArgs
This trait represents a set of arguments accepted by TabBar::scroll method.

TabBarSetAttributeArgs
This trait represents a set of arguments accepted by TabBar::set_attribute method.

TabBarSetBaseSizeArgs
This trait represents a set of arguments accepted by TabBar::set_base_size method.

TabBarSetContentsMarginsArgs
This trait represents a set of arguments accepted by TabBar::set_contents_margins method.

TabBarSetFixedSizeArgs
This trait represents a set of arguments accepted by TabBar::set_fixed_size method.

TabBarSetFocusArgs
This trait represents a set of arguments accepted by TabBar::set_focus method.

TabBarSetGeometryArgs
This trait represents a set of arguments accepted by TabBar::set_geometry method.

TabBarSetMaskArgs
This trait represents a set of arguments accepted by TabBar::set_mask method.

TabBarSetMaximumSizeArgs
This trait represents a set of arguments accepted by TabBar::set_maximum_size method.

TabBarSetMinimumSizeArgs
This trait represents a set of arguments accepted by TabBar::set_minimum_size method.

TabBarSetParentArgs
This trait represents a set of arguments accepted by TabBar::set_parent method.

TabBarSetShortcutAutoRepeatArgs
This trait represents a set of arguments accepted by TabBar::set_shortcut_auto_repeat method.

TabBarSetShortcutEnabledArgs
This trait represents a set of arguments accepted by TabBar::set_shortcut_enabled method.

TabBarSetSizeIncrementArgs
This trait represents a set of arguments accepted by TabBar::set_size_increment method.

TabBarSetSizePolicyArgs
This trait represents a set of arguments accepted by TabBar::set_size_policy method.

TabBarSizeArgs
This trait represents a set of arguments accepted by TabBar::size method.

TabBarSizeHintArgs
This trait represents a set of arguments accepted by TabBar::size_hint method.

TabBarSizeIncrementArgs
This trait represents a set of arguments accepted by TabBar::size_increment method.

TabBarSizePolicyArgs
This trait represents a set of arguments accepted by TabBar::size_policy method.

TabBarStartTimerArgs
This trait represents a set of arguments accepted by TabBar::start_timer method.

TabBarStatusTipArgs
This trait represents a set of arguments accepted by TabBar::status_tip method.

TabBarStyleSheetArgs
This trait represents a set of arguments accepted by TabBar::style_sheet method.

TabBarTabDataArgs
This trait represents a set of arguments accepted by TabBar::tab_data method.

TabBarTabIconArgs
This trait represents a set of arguments accepted by TabBar::tab_icon method.

TabBarTabRectArgs
This trait represents a set of arguments accepted by TabBar::tab_rect method.

TabBarTabTextArgs
This trait represents a set of arguments accepted by TabBar::tab_text method.

TabBarTabTextColorArgs
This trait represents a set of arguments accepted by TabBar::tab_text_color method.

TabBarTabToolTipArgs
This trait represents a set of arguments accepted by TabBar::tab_tool_tip method.

TabBarTabWhatsThisArgs
This trait represents a set of arguments accepted by TabBar::tab_whats_this method.

TabBarToolTipArgs
This trait represents a set of arguments accepted by TabBar::tool_tip method.

TabBarTrArgs
This trait represents a set of arguments accepted by TabBar::tr method.

TabBarTrUtf8Args
This trait represents a set of arguments accepted by TabBar::tr_utf8 method.

TabBarUpdateArgs
This trait represents a set of arguments accepted by TabBar::update method.

TabBarVisibleRegionArgs
This trait represents a set of arguments accepted by TabBar::visible_region method.

TabBarWhatsThisArgs
This trait represents a set of arguments accepted by TabBar::whats_this method.

TabBarWindowFilePathArgs
This trait represents a set of arguments accepted by TabBar::window_file_path method.

TabBarWindowIconArgs
This trait represents a set of arguments accepted by TabBar::window_icon method.

TabBarWindowIconTextArgs
This trait represents a set of arguments accepted by TabBar::window_icon_text method.

TabBarWindowRoleArgs
This trait represents a set of arguments accepted by TabBar::window_role method.

TabBarWindowTitleArgs
This trait represents a set of arguments accepted by TabBar::window_title method.