Module qt_widgets::tab_widget::overloading

Traits

TabWidgetAccessibleDescriptionArgs
This trait represents a set of arguments accepted by TabWidget::accessible_description method.

TabWidgetAccessibleNameArgs
This trait represents a set of arguments accepted by TabWidget::accessible_name method.

TabWidgetActionsArgs
This trait represents a set of arguments accepted by TabWidget::actions method.

TabWidgetAddTabArgs
This trait represents a set of arguments accepted by TabWidget::add_tab method.

TabWidgetBaseSizeArgs
This trait represents a set of arguments accepted by TabWidget::base_size method.

TabWidgetChildAtArgs
This trait represents a set of arguments accepted by TabWidget::child_at method.

TabWidgetChildrenRectArgs
This trait represents a set of arguments accepted by TabWidget::children_rect method.

TabWidgetChildrenRegionArgs
This trait represents a set of arguments accepted by TabWidget::children_region method.

TabWidgetConnectArgs
This trait represents a set of arguments accepted by TabWidget::connect method.

TabWidgetConnectStaticArgs
This trait represents a set of arguments accepted by TabWidget::connect_static method.

TabWidgetContentsMarginsArgs
This trait represents a set of arguments accepted by TabWidget::contents_margins method.

TabWidgetContentsRectArgs
This trait represents a set of arguments accepted by TabWidget::contents_rect method.

TabWidgetCornerWidgetArgs
This trait represents a set of arguments accepted by TabWidget::corner_widget method.

TabWidgetCreateWindowContainerArgs
This trait represents a set of arguments accepted by TabWidget::create_window_container method.

TabWidgetCursorArgs
This trait represents a set of arguments accepted by TabWidget::cursor method.

TabWidgetDestroyedArgs
This trait represents a set of arguments accepted by TabWidget::destroyed method.

TabWidgetDisconnectArgs
This trait represents a set of arguments accepted by TabWidget::disconnect method.

TabWidgetDisconnectStaticArgs
This trait represents a set of arguments accepted by TabWidget::disconnect_static method.

TabWidgetDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by TabWidget::dynamic_property_names method.

TabWidgetFontInfoArgs
This trait represents a set of arguments accepted by TabWidget::font_info method.

TabWidgetFontMetricsArgs
This trait represents a set of arguments accepted by TabWidget::font_metrics method.

TabWidgetFrameGeometryArgs
This trait represents a set of arguments accepted by TabWidget::frame_geometry method.

TabWidgetFrameSizeArgs
This trait represents a set of arguments accepted by TabWidget::frame_size method.

TabWidgetGrabArgs
This trait represents a set of arguments accepted by TabWidget::grab method.

TabWidgetGrabGestureArgs
This trait represents a set of arguments accepted by TabWidget::grab_gesture method.

TabWidgetGrabMouseArgs
This trait represents a set of arguments accepted by TabWidget::grab_mouse method.

TabWidgetGrabShortcutArgs
This trait represents a set of arguments accepted by TabWidget::grab_shortcut method.

TabWidgetIconSizeArgs
This trait represents a set of arguments accepted by TabWidget::icon_size method.

TabWidgetInputMethodQueryArgs
This trait represents a set of arguments accepted by TabWidget::input_method_query method.

TabWidgetInsertTabArgs
This trait represents a set of arguments accepted by TabWidget::insert_tab method.

TabWidgetLocaleArgs
This trait represents a set of arguments accepted by TabWidget::locale method.

TabWidgetMapFromArgs
This trait represents a set of arguments accepted by TabWidget::map_from method.

TabWidgetMapFromGlobalArgs
This trait represents a set of arguments accepted by TabWidget::map_from_global method.

TabWidgetMapFromParentArgs
This trait represents a set of arguments accepted by TabWidget::map_from_parent method.

TabWidgetMapToArgs
This trait represents a set of arguments accepted by TabWidget::map_to method.

TabWidgetMapToGlobalArgs
This trait represents a set of arguments accepted by TabWidget::map_to_global method.

TabWidgetMapToParentArgs
This trait represents a set of arguments accepted by TabWidget::map_to_parent method.

TabWidgetMaskArgs
This trait represents a set of arguments accepted by TabWidget::mask method.

TabWidgetMaximumSizeArgs
This trait represents a set of arguments accepted by TabWidget::maximum_size method.

TabWidgetMinimumSizeArgs
This trait represents a set of arguments accepted by TabWidget::minimum_size method.

TabWidgetMinimumSizeHintArgs
This trait represents a set of arguments accepted by TabWidget::minimum_size_hint method.

TabWidgetMoveArgs
This trait represents a set of arguments accepted by TabWidget::move_ method.

TabWidgetNewArgs
This trait represents a set of arguments accepted by TabWidget::new method.

TabWidgetNormalGeometryArgs
This trait represents a set of arguments accepted by TabWidget::normal_geometry method.

TabWidgetObjectNameArgs
This trait represents a set of arguments accepted by TabWidget::object_name method.

TabWidgetPosArgs
This trait represents a set of arguments accepted by TabWidget::pos method.

TabWidgetPropertyArgs
This trait represents a set of arguments accepted by TabWidget::property method.

TabWidgetRectArgs
This trait represents a set of arguments accepted by TabWidget::rect method.

TabWidgetRenderArgs
This trait represents a set of arguments accepted by TabWidget::render method.

TabWidgetRepaintArgs
This trait represents a set of arguments accepted by TabWidget::repaint method.

TabWidgetResizeArgs
This trait represents a set of arguments accepted by TabWidget::resize method.

TabWidgetSaveGeometryArgs
This trait represents a set of arguments accepted by TabWidget::save_geometry method.

TabWidgetScrollArgs
This trait represents a set of arguments accepted by TabWidget::scroll method.

TabWidgetSetAttributeArgs
This trait represents a set of arguments accepted by TabWidget::set_attribute method.

TabWidgetSetBaseSizeArgs
This trait represents a set of arguments accepted by TabWidget::set_base_size method.

TabWidgetSetContentsMarginsArgs
This trait represents a set of arguments accepted by TabWidget::set_contents_margins method.

TabWidgetSetCornerWidgetArgs
This trait represents a set of arguments accepted by TabWidget::set_corner_widget method.

TabWidgetSetFixedSizeArgs
This trait represents a set of arguments accepted by TabWidget::set_fixed_size method.

TabWidgetSetFocusArgs
This trait represents a set of arguments accepted by TabWidget::set_focus method.

TabWidgetSetGeometryArgs
This trait represents a set of arguments accepted by TabWidget::set_geometry method.

TabWidgetSetMaskArgs
This trait represents a set of arguments accepted by TabWidget::set_mask method.

TabWidgetSetMaximumSizeArgs
This trait represents a set of arguments accepted by TabWidget::set_maximum_size method.

TabWidgetSetMinimumSizeArgs
This trait represents a set of arguments accepted by TabWidget::set_minimum_size method.

TabWidgetSetParentArgs
This trait represents a set of arguments accepted by TabWidget::set_parent method.

TabWidgetSetShortcutAutoRepeatArgs
This trait represents a set of arguments accepted by TabWidget::set_shortcut_auto_repeat method.

TabWidgetSetShortcutEnabledArgs
This trait represents a set of arguments accepted by TabWidget::set_shortcut_enabled method.

TabWidgetSetSizeIncrementArgs
This trait represents a set of arguments accepted by TabWidget::set_size_increment method.

TabWidgetSetSizePolicyArgs
This trait represents a set of arguments accepted by TabWidget::set_size_policy method.

TabWidgetSizeArgs
This trait represents a set of arguments accepted by TabWidget::size method.

TabWidgetSizeHintArgs
This trait represents a set of arguments accepted by TabWidget::size_hint method.

TabWidgetSizeIncrementArgs
This trait represents a set of arguments accepted by TabWidget::size_increment method.

TabWidgetSizePolicyArgs
This trait represents a set of arguments accepted by TabWidget::size_policy method.

TabWidgetStartTimerArgs
This trait represents a set of arguments accepted by TabWidget::start_timer method.

TabWidgetStatusTipArgs
This trait represents a set of arguments accepted by TabWidget::status_tip method.

TabWidgetStyleSheetArgs
This trait represents a set of arguments accepted by TabWidget::style_sheet method.

TabWidgetTabIconArgs
This trait represents a set of arguments accepted by TabWidget::tab_icon method.

TabWidgetTabTextArgs
This trait represents a set of arguments accepted by TabWidget::tab_text method.

TabWidgetTabToolTipArgs
This trait represents a set of arguments accepted by TabWidget::tab_tool_tip method.

TabWidgetTabWhatsThisArgs
This trait represents a set of arguments accepted by TabWidget::tab_whats_this method.

TabWidgetToolTipArgs
This trait represents a set of arguments accepted by TabWidget::tool_tip method.

TabWidgetTrArgs
This trait represents a set of arguments accepted by TabWidget::tr method.

TabWidgetTrUtf8Args
This trait represents a set of arguments accepted by TabWidget::tr_utf8 method.

TabWidgetUpdateArgs
This trait represents a set of arguments accepted by TabWidget::update method.

TabWidgetVisibleRegionArgs
This trait represents a set of arguments accepted by TabWidget::visible_region method.

TabWidgetWhatsThisArgs
This trait represents a set of arguments accepted by TabWidget::whats_this method.

TabWidgetWindowFilePathArgs
This trait represents a set of arguments accepted by TabWidget::window_file_path method.

TabWidgetWindowIconArgs
This trait represents a set of arguments accepted by TabWidget::window_icon method.

TabWidgetWindowIconTextArgs
This trait represents a set of arguments accepted by TabWidget::window_icon_text method.

TabWidgetWindowRoleArgs
This trait represents a set of arguments accepted by TabWidget::window_role method.

TabWidgetWindowTitleArgs
This trait represents a set of arguments accepted by TabWidget::window_title method.