Module qt_widgets::table_widget_item::overloading

Traits

TableWidgetItemBackgroundArgs
This trait represents a set of arguments accepted by TableWidgetItem::background method.

TableWidgetItemBackgroundColorArgs
This trait represents a set of arguments accepted by TableWidgetItem::background_color method.

TableWidgetItemDataArgs
This trait represents a set of arguments accepted by TableWidgetItem::data method.

TableWidgetItemFontArgs
This trait represents a set of arguments accepted by TableWidgetItem::font method.

TableWidgetItemForegroundArgs
This trait represents a set of arguments accepted by TableWidgetItem::foreground method.

TableWidgetItemIconArgs
This trait represents a set of arguments accepted by TableWidgetItem::icon method.

TableWidgetItemNewArgs
This trait represents a set of arguments accepted by TableWidgetItem::new method.

TableWidgetItemSizeHintArgs
This trait represents a set of arguments accepted by TableWidgetItem::size_hint method.

TableWidgetItemStatusTipArgs
This trait represents a set of arguments accepted by TableWidgetItem::status_tip method.

TableWidgetItemTextArgs
This trait represents a set of arguments accepted by TableWidgetItem::text method.

TableWidgetItemTextColorArgs
This trait represents a set of arguments accepted by TableWidgetItem::text_color method.

TableWidgetItemToolTipArgs
This trait represents a set of arguments accepted by TableWidgetItem::tool_tip method.

TableWidgetItemWhatsThisArgs
This trait represents a set of arguments accepted by TableWidgetItem::whats_this method.