Struct qt_widgets::table_widget_item::TableWidgetItem

pub struct TableWidgetItem {
    // some fields omitted
}

C++ type: QTableWidgetItem

C++ documentation:

The QTableWidgetItem class provides an item for use with the QTableWidget class.

Table items are used to hold pieces of information for table widgets. Items usually contain text, icons, or checkboxes

The QTableWidgetItem class is a convenience class that replaces the QTableItem class in Qt 3. It provides an item for use with the QTableWidget class.

Top-level items are constructed without a parent then inserted at the position specified by a pair of row and column numbers:

    QTableWidgetItem *newItem = new QTableWidgetItem(tr("%1").arg(
        pow(row, column+1)));
    tableWidget->setItem(row, column, newItem);

Each item can have its own background brush which is set with the setBackground() function. The current background brush can be found with background(). The text label for each item can be rendered with its own font and brush. These are specified with the setFont() and setForeground() functions, and read with font() and foreground().

By default, items are enabled, editable, selectable, checkable, and can be used both as the source of a drag and drop operation and as a drop target. Each item's flags can be changed by calling setFlags() with the appropriate value (see Qt::ItemFlags). Checkable items can be checked and unchecked with the setCheckState() function. The corresponding checkState() function indicates whether the item is currently checked.

Methods

impl TableWidgetItem

fn background<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TableWidgetItemBackgroundArgs<'largs>

C++ method: QTableWidgetItem::background

This is an overloaded function. Available variants:

Rust arguments:
1) fn background(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::brush::Brush>
2) fn background(&self, cpp_utils::AsStruct) -> qt_gui::brush::Brush
C++ method: QBrush QTableWidgetItem::background() const

C++ documentation:

Returns the brush used to render the item's background.

This function was introduced in Qt 4.2.

See also setBackground() and foreground().

fn background_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TableWidgetItemBackgroundColorArgs<'largs>

C++ method: QTableWidgetItem::backgroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn background_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn background_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: QColor QTableWidgetItem::backgroundColor() const

C++ documentation:

This function is deprecated. Use background() instead.

See also setBackgroundColor().

fn check_state(&self) -> CheckState

C++ method: Qt::CheckState QTableWidgetItem::checkState() const

C++ documentation:

Returns the checked state of the table item.

See also setCheckState() and flags().

fn clone(&self) -> *mut TableWidgetItem

C++ method: virtual QTableWidgetItem* QTableWidgetItem::clone() const

C++ documentation:

Creates a copy of the item.

fn column(&self) -> c_int

C++ method: int QTableWidgetItem::column() const

C++ documentation:

Returns the column of the item in the table. If the item is not in a table, this function will return -1.

This function was introduced in Qt 4.2.

See also row().

fn data<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TableWidgetItemDataArgs<'largs>

C++ method: QTableWidgetItem::data

This is an overloaded function. Available variants:

Rust arguments:
1) fn data(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn data(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::variant::Variant
C++ method: virtual QVariant QTableWidgetItem::data(int role) const

C++ documentation:

Returns the item's data for the given role.

See also setData().

fn flags(&self) -> Flags<ItemFlag>

C++ method: QFlags<Qt::ItemFlag> QTableWidgetItem::flags() const

C++ documentation:

Returns the flags used to describe the item. These determine whether the item can be checked, edited, and selected.

See also setFlags().

fn font<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TableWidgetItemFontArgs<'largs>

C++ method: QTableWidgetItem::font

This is an overloaded function. Available variants:

Rust arguments:
1) fn font(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::font::Font>
2) fn font(&self, cpp_utils::AsStruct) -> qt_gui::font::Font
C++ method: QFont QTableWidgetItem::font() const

C++ documentation:

Returns the font used to render the item's text.

See also setFont().

fn foreground<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TableWidgetItemForegroundArgs<'largs>

C++ method: QTableWidgetItem::foreground

This is an overloaded function. Available variants:

Rust arguments:
1) fn foreground(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::brush::Brush>
2) fn foreground(&self, cpp_utils::AsStruct) -> qt_gui::brush::Brush
C++ method: QBrush QTableWidgetItem::foreground() const

C++ documentation:

Returns the brush used to render the item's foreground (e.g. text).

This function was introduced in Qt 4.2.

See also setForeground() and background().

fn icon<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TableWidgetItemIconArgs<'largs>

C++ method: QTableWidgetItem::icon

This is an overloaded function. Available variants:

Rust arguments:
1) fn icon(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::icon::Icon>
2) fn icon(&self, cpp_utils::AsStruct) -> qt_gui::icon::Icon
C++ method: QIcon QTableWidgetItem::icon() const

C++ documentation:

Returns the item's icon.

See also setIcon() and iconSize.

fn is_selected(&self) -> bool

C++ method: bool QTableWidgetItem::isSelected() const

C++ documentation:

Returns true if the item is selected, otherwise returns false.

This function was introduced in Qt 4.2.

See also setSelected().

fn new<Args>(args: Args) -> Args::ReturnType where Args: TableWidgetItemNewArgs

C++ method: QTableWidgetItem::QTableWidgetItem

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((&qt_gui::icon::Icon, &qt_core::string::String, cpp_utils::AsStruct)) -> ::table_widget_item::TableWidgetItem
2) fn new((&qt_gui::icon::Icon, &qt_core::string::String, libc::c_int, cpp_utils::AsStruct)) -> ::table_widget_item::TableWidgetItem
3) fn new((&qt_gui::icon::Icon, &qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::table_widget_item::TableWidgetItem>
4) fn new((&qt_gui::icon::Icon, &qt_core::string::String, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::table_widget_item::TableWidgetItem>
C++ method: [constructor] void QTableWidgetItem::QTableWidgetItem(const QIcon& icon, const QString& text, int type = ?)

Variant 2

Rust arguments:
1) fn new((&qt_core::string::String, cpp_utils::AsStruct)) -> ::table_widget_item::TableWidgetItem
2) fn new((&qt_core::string::String, libc::c_int, cpp_utils::AsStruct)) -> ::table_widget_item::TableWidgetItem
3) fn new((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::table_widget_item::TableWidgetItem>
4) fn new((&qt_core::string::String, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::table_widget_item::TableWidgetItem>
C++ method: [constructor] void QTableWidgetItem::QTableWidgetItem(const QString& text, int type = ?)

Variant 3

Rust arguments:
1) fn new((&::table_widget_item::TableWidgetItem, cpp_utils::AsStruct)) -> ::table_widget_item::TableWidgetItem
2) fn new((&::table_widget_item::TableWidgetItem, cpp_utils::AsBox)) -> cpp_utils::CppBox<::table_widget_item::TableWidgetItem>
C++ method: [constructor] void QTableWidgetItem::QTableWidgetItem(const QTableWidgetItem& other)

Variant 4

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::table_widget_item::TableWidgetItem
2) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::table_widget_item::TableWidgetItem
3) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::table_widget_item::TableWidgetItem>
4) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::table_widget_item::TableWidgetItem>
C++ method: [constructor] void QTableWidgetItem::QTableWidgetItem(int type = ?)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 TableWidgetItem) -> &'l0 mut TableWidgetItem

C++ method: QTableWidgetItem& QTableWidgetItem::operator=(const QTableWidgetItem& other)

C++ documentation:

Assigns other's data and flags to this item. Note that type() and tableWidget() are not copied.

This function is useful when reimplementing clone().

See also data() and flags().

fn op_lt(&self, other: &TableWidgetItem) -> bool

C++ method: virtual bool QTableWidgetItem::operator<(const QTableWidgetItem& other) const

C++ documentation:

Returns true if the item is less than the other item; otherwise returns false.

fn read(&mut self, in_: &mut DataStream)

C++ method: virtual void QTableWidgetItem::read(QDataStream& in)

C++ documentation:

Reads the item from stream in.

See also write().

fn row(&self) -> c_int

C++ method: int QTableWidgetItem::row() const

C++ documentation:

Returns the row of the item in the table. If the item is not in a table, this function will return -1.

This function was introduced in Qt 4.2.

See also column().

fn set_background(&mut self, brush: &Brush)

C++ method: void QTableWidgetItem::setBackground(const QBrush& brush)

C++ documentation:

Sets the item's background brush to the specified brush.

This function was introduced in Qt 4.2.

See also background() and setForeground().

fn set_background_color(&mut self, color: &Color)

C++ method: void QTableWidgetItem::setBackgroundColor(const QColor& color)

C++ documentation:

This function is deprecated. Use setBackground() instead.

See also backgroundColor().

fn set_check_state(&mut self, state: CheckState)

C++ method: void QTableWidgetItem::setCheckState(Qt::CheckState state)

C++ documentation:

Sets the check state of the table item to be state.

See also checkState().

fn set_data(&mut self, role: c_int, value: &Variant)

C++ method: virtual void QTableWidgetItem::setData(int role, const QVariant& value)

C++ documentation:

Sets the item's data for the given role to the specified value.

See also Qt::ItemDataRole and data().

fn set_flags(&mut self, flags: Flags<ItemFlag>)

C++ method: void QTableWidgetItem::setFlags(QFlags<Qt::ItemFlag> flags)

C++ documentation:

Sets the flags for the item to the given flags. These determine whether the item can be selected or modified.

See also flags().

fn set_font(&mut self, font: &Font)

C++ method: void QTableWidgetItem::setFont(const QFont& font)

C++ documentation:

Sets the font used to display the item's text to the given font.

See also font(), setText(), and setForeground().

fn set_foreground(&mut self, brush: &Brush)

C++ method: void QTableWidgetItem::setForeground(const QBrush& brush)

C++ documentation:

Sets the item's foreground brush to the specified brush.

This function was introduced in Qt 4.2.

See also foreground() and setBackground().

fn set_icon(&mut self, icon: &Icon)

C++ method: void QTableWidgetItem::setIcon(const QIcon& icon)

C++ documentation:

Sets the item's icon to the icon specified.

See also icon(), setText(), and iconSize.

fn set_selected(&mut self, select: bool)

C++ method: void QTableWidgetItem::setSelected(bool select)

C++ documentation:

Sets the selected state of the item to select.

This function was introduced in Qt 4.2.

See also isSelected().

fn set_size_hint(&mut self, size: &Size)

C++ method: void QTableWidgetItem::setSizeHint(const QSize& size)

C++ documentation:

Sets the size hint for the table item to be size. If no size hint is set, the item delegate will compute the size hint based on the item data.

This function was introduced in Qt 4.1.

See also sizeHint().

fn set_status_tip(&mut self, status_tip: &String)

C++ method: void QTableWidgetItem::setStatusTip(const QString& statusTip)

C++ documentation:

Sets the status tip for the table item to the text specified by statusTip. QTableWidget mouse tracking needs to be enabled for this feature to work.

See also statusTip(), setToolTip(), and setWhatsThis().

fn set_text(&mut self, text: &String)

C++ method: void QTableWidgetItem::setText(const QString& text)

C++ documentation:

Sets the item's text to the text specified.

See also text(), setFont(), and setForeground().

fn set_text_alignment(&mut self, alignment: c_int)

C++ method: void QTableWidgetItem::setTextAlignment(int alignment)

C++ documentation:

Sets the text alignment for the item's text to the alignment specified.

See also textAlignment() and Qt::Alignment.

fn set_text_color(&mut self, color: &Color)

C++ method: void QTableWidgetItem::setTextColor(const QColor& color)

C++ documentation:

This function is deprecated. Use setForeground() instead.

See also textColor().

fn set_tool_tip(&mut self, tool_tip: &String)

C++ method: void QTableWidgetItem::setToolTip(const QString& toolTip)

C++ documentation:

Sets the item's tooltip to the string specified by toolTip.

See also toolTip(), setStatusTip(), and setWhatsThis().

fn set_whats_this(&mut self, whats_this: &String)

C++ method: void QTableWidgetItem::setWhatsThis(const QString& whatsThis)

C++ documentation:

Sets the item's "What's This?" help to the string specified by whatsThis.

See also whatsThis(), setStatusTip(), and setToolTip().

fn size_hint<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TableWidgetItemSizeHintArgs<'largs>

C++ method: QTableWidgetItem::sizeHint

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_hint(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size_hint(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: QSize QTableWidgetItem::sizeHint() const

C++ documentation:

Returns the size hint set for the table item.

This function was introduced in Qt 4.1.

See also setSizeHint().

fn status_tip<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TableWidgetItemStatusTipArgs<'largs>

C++ method: QTableWidgetItem::statusTip

This is an overloaded function. Available variants:

Rust arguments:
1) fn status_tip(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn status_tip(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTableWidgetItem::statusTip() const

C++ documentation:

Returns the item's status tip.

See also setStatusTip().

fn table_widget(&self) -> *mut TableWidget

C++ method: QTableWidget* QTableWidgetItem::tableWidget() const

C++ documentation:

Returns the table widget that contains the item.

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TableWidgetItemTextArgs<'largs>

C++ method: QTableWidgetItem::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTableWidgetItem::text() const

C++ documentation:

Returns the item's text.

See also setText().

fn text_alignment(&self) -> c_int

C++ method: int QTableWidgetItem::textAlignment() const

C++ documentation:

Returns the text alignment for the item's text.

See also setTextAlignment() and Qt::Alignment.

fn text_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TableWidgetItemTextColorArgs<'largs>

C++ method: QTableWidgetItem::textColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn text_color(&self, cpp_utils::AsStruct) -> qt_gui::color::Color
C++ method: QColor QTableWidgetItem::textColor() const

C++ documentation:

This function is deprecated. Use foreground() instead.

See also setTextColor().

fn tool_tip<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TableWidgetItemToolTipArgs<'largs>

C++ method: QTableWidgetItem::toolTip

This is an overloaded function. Available variants:

Rust arguments:
1) fn tool_tip(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tool_tip(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTableWidgetItem::toolTip() const

C++ documentation:

Returns the item's tooltip.

See also setToolTip().

fn type_(&self) -> c_int

C++ method: int QTableWidgetItem::type() const

C++ documentation:

Returns the type passed to the QTableWidgetItem constructor.

fn whats_this<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TableWidgetItemWhatsThisArgs<'largs>

C++ method: QTableWidgetItem::whatsThis

This is an overloaded function. Available variants:

Rust arguments:
1) fn whats_this(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn whats_this(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTableWidgetItem::whatsThis() const

C++ documentation:

Returns the item's "What's This?" help.

See also setWhatsThis().

fn write(&self, out: &mut DataStream)

C++ method: virtual void QTableWidgetItem::write(QDataStream& out) const

C++ documentation:

Writes the item to stream out.

See also read().