Module qt_widgets::tap_and_hold_gesture::overloading

Traits

TapAndHoldGestureConnectArgs
This trait represents a set of arguments accepted by TapAndHoldGesture::connect method.

TapAndHoldGestureConnectStaticArgs
This trait represents a set of arguments accepted by TapAndHoldGesture::connect_static method.

TapAndHoldGestureDestroyedArgs
This trait represents a set of arguments accepted by TapAndHoldGesture::destroyed method.

TapAndHoldGestureDisconnectArgs
This trait represents a set of arguments accepted by TapAndHoldGesture::disconnect method.

TapAndHoldGestureDisconnectStaticArgs
This trait represents a set of arguments accepted by TapAndHoldGesture::disconnect_static method.

TapAndHoldGestureDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by TapAndHoldGesture::dynamic_property_names method.

TapAndHoldGestureHotSpotArgs
This trait represents a set of arguments accepted by TapAndHoldGesture::hot_spot method.

TapAndHoldGestureNewArgs
This trait represents a set of arguments accepted by TapAndHoldGesture::new method.

TapAndHoldGestureObjectNameArgs
This trait represents a set of arguments accepted by TapAndHoldGesture::object_name method.

TapAndHoldGesturePositionArgs
This trait represents a set of arguments accepted by TapAndHoldGesture::position method.

TapAndHoldGesturePropertyArgs
This trait represents a set of arguments accepted by TapAndHoldGesture::property method.

TapAndHoldGestureStartTimerArgs
This trait represents a set of arguments accepted by TapAndHoldGesture::start_timer method.

TapAndHoldGestureTrArgs
This trait represents a set of arguments accepted by TapAndHoldGesture::tr method.

TapAndHoldGestureTrUtf8Args
This trait represents a set of arguments accepted by TapAndHoldGesture::tr_utf8 method.