Struct qt_widgets::text_edit::ExtraSelection

pub struct ExtraSelection {
    // some fields omitted
}

C++ type: QTextEdit::ExtraSelection

C++ documentation:

The QTextEdit::ExtraSelection structure provides a way of specifying a character format for a given selection in a document

Methods

impl ExtraSelection

fn cursor<'l0>(&'l0 self) -> &'l0 TextCursor

C++ method: const QTextCursor& QTextEdit::ExtraSelection::cursor() const

fn cursor_mut<'l0>(&'l0 mut self) -> &'l0 mut TextCursor

C++ method: QTextCursor& QTextEdit::ExtraSelection::cursor_mut()

fn format<'l0>(&'l0 self) -> &'l0 TextCharFormat

C++ method: const QTextCharFormat& QTextEdit::ExtraSelection::format() const

fn format_mut<'l0>(&'l0 mut self) -> &'l0 mut TextCharFormat

C++ method: QTextCharFormat& QTextEdit::ExtraSelection::format_mut()

fn set_cursor(&mut self, value: &TextCursor)

C++ method: void QTextEdit::ExtraSelection::set_cursor(QTextCursor value)

fn set_format(&mut self, value: &TextCharFormat)

C++ method: void QTextEdit::ExtraSelection::set_format(QTextCharFormat value)