Function qt_widgets::tile_rules::draw_shade_line

pub fn draw_shade_line<Args>(args: Args) where Args: DrawShadeLineArgs

C++ method: qDrawShadeLine

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn draw_shade_line((*mut qt_gui::painter::Painter, &qt_core::point::Point, &qt_core::point::Point, &qt_gui::palette::Palette)) -> ()
2) fn draw_shade_line((*mut qt_gui::painter::Painter, &qt_core::point::Point, &qt_core::point::Point, &qt_gui::palette::Palette, bool)) -> ()
3) fn draw_shade_line((*mut qt_gui::painter::Painter, &qt_core::point::Point, &qt_core::point::Point, &qt_gui::palette::Palette, bool, libc::c_int)) -> ()
4) fn draw_shade_line((*mut qt_gui::painter::Painter, &qt_core::point::Point, &qt_core::point::Point, &qt_gui::palette::Palette, bool, libc::c_int, libc::c_int)) -> ()
C++ method: void qDrawShadeLine(QPainter* p, const QPoint& p1, const QPoint& p2, const QPalette& pal, bool sunken = ?, int lineWidth = ?, int midLineWidth = ?)

C++ documentation:

This is an overloaded function.

Draws a horizontal or vertical shaded line between p1 and p2 using the given painter. Note that nothing is drawn if the line between the points would be neither horizontal nor vertical.

The provided palette specifies the shading colors (light, dark and middle colors). The given lineWidth specifies the line width for each of the lines; it is not the total line width. The given midLineWidth specifies the width of a middle line drawn in the QPalette::mid() color.

The line appears sunken if sunken is true, otherwise raised.

Warning: This function does not look at QWidget::style() or QApplication::style(). Use the drawing functions in QStyle to make widgets that follow the current GUI style.

Alternatively you can use a QFrame widget and apply the QFrame::setFrameStyle() function to display a shaded line:

QFrame frame:
frame.setFrameStyle(QFrame::HLine | QFrame::Sunken);

See also qDrawShadeRect(), qDrawShadePanel(), and QStyle.

Variant 2

Rust arguments:
1) fn draw_shade_line((*mut qt_gui::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, &qt_gui::palette::Palette)) -> ()
2) fn draw_shade_line((*mut qt_gui::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, &qt_gui::palette::Palette, bool)) -> ()
3) fn draw_shade_line((*mut qt_gui::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, &qt_gui::palette::Palette, bool, libc::c_int)) -> ()
4) fn draw_shade_line((*mut qt_gui::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, &qt_gui::palette::Palette, bool, libc::c_int, libc::c_int)) -> ()
C++ method: void qDrawShadeLine(QPainter* p, int x1, int y1, int x2, int y2, const QPalette& pal, bool sunken = ?, int lineWidth = ?, int midLineWidth = ?)

C++ documentation:

Draws a horizontal (y1 == y2) or vertical (x1 == x2) shaded line using the given painter. Note that nothing is drawn if y1 != y2 and x1 != x2 (i.e. the line is neither horizontal nor vertical).

The provided palette specifies the shading colors (light, dark and middle colors). The given lineWidth specifies the line width for each of the lines; it is not the total line width. The given midLineWidth specifies the width of a middle line drawn in the QPalette::mid() color.

The line appears sunken if sunken is true, otherwise raised.

Warning: This function does not look at QWidget::style() or QApplication::style(). Use the drawing functions in QStyle to make widgets that follow the current GUI style.

Alternatively you can use a QFrame widget and apply the QFrame::setFrameStyle() function to display a shaded line:

QFrame frame:
frame.setFrameStyle(QFrame::HLine | QFrame::Sunken);

See also qDrawShadeRect(), qDrawShadePanel(), and QStyle.