Function qt_widgets::tile_rules::draw_shade_rect

pub fn draw_shade_rect<Args>(args: Args) where Args: DrawShadeRectArgs

C++ method: qDrawShadeRect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn draw_shade_rect((*mut qt_gui::painter::Painter, &qt_core::rect::Rect, &qt_gui::palette::Palette)) -> ()
2) fn draw_shade_rect((*mut qt_gui::painter::Painter, &qt_core::rect::Rect, &qt_gui::palette::Palette, bool)) -> ()
3) fn draw_shade_rect((*mut qt_gui::painter::Painter, &qt_core::rect::Rect, &qt_gui::palette::Palette, bool, libc::c_int)) -> ()
4) fn draw_shade_rect((*mut qt_gui::painter::Painter, &qt_core::rect::Rect, &qt_gui::palette::Palette, bool, libc::c_int, libc::c_int)) -> ()
5) fn draw_shade_rect((*mut qt_gui::painter::Painter, &qt_core::rect::Rect, &qt_gui::palette::Palette, bool, libc::c_int, libc::c_int, *const qt_gui::brush::Brush)) -> ()
C++ method: void qDrawShadeRect(QPainter* p, const QRect& r, const QPalette& pal, bool sunken = ?, int lineWidth = ?, int midLineWidth = ?, const QBrush* fill = ?)

C++ documentation:

This is an overloaded function.

Draws the shaded rectangle specified by rect using the given painter.

The provide palette specifies the shading colors (light, dark and middle colors. The given lineWidth specifies the line width for each of the lines; it is not the total line width. The midLineWidth specifies the width of a middle line drawn in the QPalette::mid() color. The rectangle's interior is filled with the fill brush unless fill is 0.

The rectangle appears sunken if sunken is true, otherwise raised.

Warning: This function does not look at QWidget::style() or QApplication::style(). Use the drawing functions in QStyle to make widgets that follow the current GUI style.

Alternatively you can use a QFrame widget and apply the QFrame::setFrameStyle() function to display a shaded rectangle:

QFrame frame:
frame.setFrameStyle(QFrame::Box | QFrame::Raised);

See also qDrawShadeLine(), qDrawShadePanel(), qDrawPlainRect(), and QStyle.

Variant 2

Rust arguments:
1) fn draw_shade_rect((*mut qt_gui::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, &qt_gui::palette::Palette)) -> ()
2) fn draw_shade_rect((*mut qt_gui::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, &qt_gui::palette::Palette, bool)) -> ()
3) fn draw_shade_rect((*mut qt_gui::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, &qt_gui::palette::Palette, bool, libc::c_int)) -> ()
4) fn draw_shade_rect((*mut qt_gui::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, &qt_gui::palette::Palette, bool, libc::c_int, libc::c_int)) -> ()
5) fn draw_shade_rect((*mut qt_gui::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, &qt_gui::palette::Palette, bool, libc::c_int, libc::c_int, *const qt_gui::brush::Brush)) -> ()
C++ method: void qDrawShadeRect(QPainter* p, int x, int y, int w, int h, const QPalette& pal, bool sunken = ?, int lineWidth = ?, int midLineWidth = ?, const QBrush* fill = ?)

C++ documentation:

Draws the shaded rectangle beginning at (x, y) with the given width and height using the provided painter.

The provide palette specifies the shading colors (light, dark and middle colors. The given lineWidth specifies the line width for each of the lines; it is not the total line width. The midLineWidth specifies the width of a middle line drawn in the QPalette::mid() color. The rectangle's interior is filled with the fill brush unless fill is 0.

The rectangle appears sunken if sunken is true, otherwise raised.

Warning: This function does not look at QWidget::style() or QApplication::style(). Use the drawing functions in QStyle to make widgets that follow the current GUI style.

Alternatively you can use a QFrame widget and apply the QFrame::setFrameStyle() function to display a shaded rectangle:

QFrame frame:
frame.setFrameStyle(QFrame::Box | QFrame::Raised);

See also qDrawShadeLine(), qDrawShadePanel(), qDrawPlainRect(), and QStyle.