Function qt_widgets::tile_rules::draw_win_button

pub fn draw_win_button<Args>(args: Args) where Args: DrawWinButtonArgs

C++ method: qDrawWinButton

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn draw_win_button((*mut qt_gui::painter::Painter, &qt_core::rect::Rect, &qt_gui::palette::Palette)) -> ()
2) fn draw_win_button((*mut qt_gui::painter::Painter, &qt_core::rect::Rect, &qt_gui::palette::Palette, bool)) -> ()
3) fn draw_win_button((*mut qt_gui::painter::Painter, &qt_core::rect::Rect, &qt_gui::palette::Palette, bool, *const qt_gui::brush::Brush)) -> ()
C++ method: void qDrawWinButton(QPainter* p, const QRect& r, const QPalette& pal, bool sunken = ?, const QBrush* fill = ?)

C++ documentation:

This is an overloaded function.

Draws the Windows-style button at the rectangle specified by rect using the given painter with a line width of 2 pixels. The button's interior is filled with the fill brush unless fill is 0.

The given palette specifies the shading colors (light, dark and middle colors).

The button appears sunken if sunken is true, otherwise raised.

Warning: This function does not look at QWidget::style() or QApplication::style()-> Use the drawing functions in QStyle to make widgets that follow the current GUI style.

See also qDrawWinPanel() and QStyle.

Variant 2

Rust arguments:
1) fn draw_win_button((*mut qt_gui::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, &qt_gui::palette::Palette)) -> ()
2) fn draw_win_button((*mut qt_gui::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, &qt_gui::palette::Palette, bool)) -> ()
3) fn draw_win_button((*mut qt_gui::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, &qt_gui::palette::Palette, bool, *const qt_gui::brush::Brush)) -> ()
C++ method: void qDrawWinButton(QPainter* p, int x, int y, int w, int h, const QPalette& pal, bool sunken = ?, const QBrush* fill = ?)

C++ documentation:

Draws the Windows-style button specified by the given point (x, y}, width and height using the provided painter with a line width of 2 pixels. The button's interior is filled with the fill brush unless fill is 0.

The given palette specifies the shading colors (light, dark and middle colors).

The button appears sunken if sunken is true, otherwise raised.

Warning: This function does not look at QWidget::style() or QApplication::style()-> Use the drawing functions in QStyle to make widgets that follow the current GUI style.

See also qDrawWinPanel() and QStyle.