Function qt_widgets::tile_rules::draw_win_panel

pub fn draw_win_panel<Args>(args: Args) where Args: DrawWinPanelArgs

C++ method: qDrawWinPanel

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn draw_win_panel((*mut qt_gui::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, &qt_gui::palette::Palette)) -> ()
2) fn draw_win_panel((*mut qt_gui::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, &qt_gui::palette::Palette, bool)) -> ()
3) fn draw_win_panel((*mut qt_gui::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, &qt_gui::palette::Palette, bool, *const qt_gui::brush::Brush)) -> ()
C++ method: void qDrawWinPanel(QPainter* p, int x, int y, int w, int h, const QPalette& pal, bool sunken = ?, const QBrush* fill = ?)

C++ documentation:

Draws the Windows-style panel specified by the given point(x, y), width and height using the provided painter with a line width of 2 pixels. The button's interior is filled with the fill brush unless fill is 0.

The given palette specifies the shading colors. The panel appears sunken if sunken is true, otherwise raised.

Warning: This function does not look at QWidget::style() or QApplication::style(). Use the drawing functions in QStyle to make widgets that follow the current GUI style.

Alternatively you can use a QFrame widget and apply the QFrame::setFrameStyle() function to display a shaded panel:

QFrame frame:
frame.setFrameStyle(QFrame::WinPanel | QFrame::Raised);

See also qDrawShadePanel(), qDrawWinButton(), and QStyle.

Variant 2

Rust arguments:
1) fn draw_win_panel((*mut qt_gui::painter::Painter, &qt_core::rect::Rect, &qt_gui::palette::Palette)) -> ()
2) fn draw_win_panel((*mut qt_gui::painter::Painter, &qt_core::rect::Rect, &qt_gui::palette::Palette, bool)) -> ()
3) fn draw_win_panel((*mut qt_gui::painter::Painter, &qt_core::rect::Rect, &qt_gui::palette::Palette, bool, *const qt_gui::brush::Brush)) -> ()
C++ method: void qDrawWinPanel(QPainter* p, const QRect& r, const QPalette& pal, bool sunken = ?, const QBrush* fill = ?)