Module qt_widgets::tile_rules::overloading

Traits

DrawBorderPixmapArgs
This trait represents a set of arguments accepted by draw_border_pixmap method.

DrawPlainRectArgs
This trait represents a set of arguments accepted by draw_plain_rect method.

DrawShadeLineArgs
This trait represents a set of arguments accepted by draw_shade_line method.

DrawShadePanelArgs
This trait represents a set of arguments accepted by draw_shade_panel method.

DrawShadeRectArgs
This trait represents a set of arguments accepted by draw_shade_rect method.

DrawWinButtonArgs
This trait represents a set of arguments accepted by draw_win_button method.

DrawWinPanelArgs
This trait represents a set of arguments accepted by draw_win_panel method.

TileRulesNewArgs
This trait represents a set of arguments accepted by TileRules::new method.