Struct qt_widgets::tool_tip::ToolTip

pub struct ToolTip {
    // some fields omitted
}

C++ type: QToolTip

C++ documentation:

The QToolTip class provides tool tips (balloon help) for any widget.

The tip is a short piece of text reminding the user of the widget's function. It is drawn immediately below the given position in a distinctive black-on-yellow color combination. The tip can be any rich text formatted string.

Rich text displayed in a tool tip is implicitly word-wrapped unless specified differently with <p style='white-space:pre'>.

The simplest and most common way to set a widget's tool tip is by calling its QWidget::setToolTip() function.

It is also possible to show different tool tips for different regions of a widget, by using a QHelpEvent of type QEvent::ToolTip. Intercept the help event in your widget's event() function and call QToolTip::showText() with the text you want to display. The Tooltips example illustrates this technique.

If you are calling QToolTip::hideText(), or QToolTip::showText() with an empty string, as a result of a ToolTip-event you should also call ignore() on the event, to signal that you don't want to start any tooltip specific modes.

Note that, if you want to show tooltips in an item view, the model/view architecture provides functionality to set an item's tool tip; e.g., the QTableWidgetItem::setToolTip() function. However, if you want to provide custom tool tips in an item view, you must intercept the help event in the QAbstractItemView::viewportEvent() function and handle it yourself.

The default tool tip color and font can be customized with setPalette() and setFont(). When a tooltip is currently on display, isVisible() returns true and text() the currently visible text.

Note: Tool tips use the inactive color group of QPalette, because tool tips are not active windows.

Methods

impl ToolTip

fn font<Args>(args: Args) -> Args::ReturnType where Args: ToolTipFontArgs

C++ method: QToolTip::font

This is an overloaded function. Available variants:

Rust arguments:
1) fn font(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::font::Font>
2) fn font(cpp_utils::AsStruct) -> qt_gui::font::Font
C++ method: static QFont QToolTip::font()

C++ documentation:

Returns the font used to render tooltips.

This function was introduced in Qt 4.2.

See also setFont().

fn hide_text()

C++ method: static void QToolTip::hideText()

C++ documentation:

Hides the tool tip. This is the same as calling showText() with an empty string.

This function was introduced in Qt 4.2.

See also showText().

fn is_visible() -> bool

C++ method: static bool QToolTip::isVisible()

C++ documentation:

Returns true if this tooltip is currently shown.

This function was introduced in Qt 4.4.

See also showText().

fn palette<Args>(args: Args) -> Args::ReturnType where Args: ToolTipPaletteArgs

C++ method: QToolTip::palette

This is an overloaded function. Available variants:

Rust arguments:
1) fn palette(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_gui::palette::Palette>
2) fn palette(cpp_utils::AsStruct) -> qt_gui::palette::Palette
C++ method: static QPalette QToolTip::palette()

C++ documentation:

Returns the palette used to render tooltips.

Note: Tool tips use the inactive color group of QPalette, because tool tips are not active windows.

See also setPalette().

fn set_font(arg1: &Font)

C++ method: static void QToolTip::setFont(const QFont& arg1)

C++ documentation:

Sets the font used to render tooltips.

This function was introduced in Qt 4.2.

See also font().

fn set_palette(arg1: &Palette)

C++ method: static void QToolTip::setPalette(const QPalette& arg1)

C++ documentation:

Sets the palette used to render tooltips.

Note: Tool tips use the inactive color group of QPalette, because tool tips are not active windows.

This function was introduced in Qt 4.2.

See also palette().

fn show_text<Args>(args: Args) where Args: ToolTipShowTextArgs

C++ method: QToolTip::showText

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn show_text((&qt_core::point::Point, &qt_core::string::String)) -> ()
2) fn show_text((&qt_core::point::Point, &qt_core::string::String, *mut ::widget::Widget)) -> ()
C++ method: static void QToolTip::showText(const QPoint& pos, const QString& text, QWidget* w = ?)

C++ documentation:

This is an overloaded function.

This is analogous to calling QToolTip::showText(pos, text, w, QRect())

Variant 2

Rust arguments: fn show_text((&qt_core::point::Point, &qt_core::string::String, *mut ::widget::Widget, &qt_core::rect::Rect)) -> ()
C++ method: static void QToolTip::showText(const QPoint& pos, const QString& text, QWidget* w, const QRect& rect)

C++ documentation:

Shows text as a tool tip, with the global position pos as the point of interest. The tool tip will be shown with a platform specific offset from this point of interest.

If you specify a non-empty rect the tip will be hidden as soon as you move your cursor out of this area.

The rect is in the coordinates of the widget you specify with w. If the rect is not empty you must specify a widget. Otherwise this argument can be 0 but it is used to determine the appropriate screen on multi-head systems.

If text is empty the tool tip is hidden. If the text is the same as the currently shown tooltip, the tip will not move. You can force moving by first hiding the tip with an empty text, and then showing the new tip at the new position.

Variant 3

Rust arguments: fn show_text((&qt_core::point::Point, &qt_core::string::String, *mut ::widget::Widget, &qt_core::rect::Rect, libc::c_int)) -> ()
C++ method: static void QToolTip::showText(const QPoint& pos, const QString& text, QWidget* w, const QRect& rect, int msecShowTime)

C++ documentation:

This is an overloaded function.

This is similar to QToolTip::showText(pos, text, w, rect) but with an extra parameter msecDisplayTime that specifies how long the tool tip will be displayed, in milliseconds.

This function was introduced in Qt 5.2.

fn text<Args>(args: Args) -> Args::ReturnType where Args: ToolTipTextArgs

C++ method: QToolTip::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QToolTip::text()

C++ documentation:

Returns the tooltip text, if a tooltip is visible, or an empty string if a tooltip is not visible.

This function was introduced in Qt 4.4.