Module qt_widgets::tree_widget_item::overloading

Traits

TreeWidgetItemBackgroundArgs
This trait represents a set of arguments accepted by TreeWidgetItem::background method.

TreeWidgetItemBackgroundColorArgs
This trait represents a set of arguments accepted by TreeWidgetItem::background_color method.

TreeWidgetItemDataArgs
This trait represents a set of arguments accepted by TreeWidgetItem::data method.

TreeWidgetItemFontArgs
This trait represents a set of arguments accepted by TreeWidgetItem::font method.

TreeWidgetItemForegroundArgs
This trait represents a set of arguments accepted by TreeWidgetItem::foreground method.

TreeWidgetItemIconArgs
This trait represents a set of arguments accepted by TreeWidgetItem::icon method.

TreeWidgetItemNewArgs
This trait represents a set of arguments accepted by TreeWidgetItem::new method.

TreeWidgetItemSizeHintArgs
This trait represents a set of arguments accepted by TreeWidgetItem::size_hint method.

TreeWidgetItemStatusTipArgs
This trait represents a set of arguments accepted by TreeWidgetItem::status_tip method.

TreeWidgetItemTakeChildrenArgs
This trait represents a set of arguments accepted by TreeWidgetItem::take_children method.

TreeWidgetItemTextArgs
This trait represents a set of arguments accepted by TreeWidgetItem::text method.

TreeWidgetItemTextColorArgs
This trait represents a set of arguments accepted by TreeWidgetItem::text_color method.

TreeWidgetItemToolTipArgs
This trait represents a set of arguments accepted by TreeWidgetItem::tool_tip method.

TreeWidgetItemWhatsThisArgs
This trait represents a set of arguments accepted by TreeWidgetItem::whats_this method.