Struct qt_widgets::tree_widget_item::TreeWidgetItem

pub struct TreeWidgetItem {
    // some fields omitted
}

C++ type: QTreeWidgetItem

C++ documentation:

The QTreeWidgetItem class provides an item for use with the QTreeWidget convenience class.

Tree widget items are used to hold rows of information for tree widgets. Rows usually contain several columns of data, each of which can contain a text label and an icon.

The QTreeWidgetItem class is a convenience class that replaces the QListViewItem class in Qt 3. It provides an item for use with the QTreeWidget class.

Items are usually constructed with a parent that is either a QTreeWidget (for top-level items) or a QTreeWidgetItem (for items on lower levels of the tree). For example, the following code constructs a top-level item to represent cities of the world, and adds a entry for Oslo as a child item:

    QTreeWidgetItem cities = new QTreeWidgetItem(treeWidget);
    cities->setText(0, tr("Cities"));
    QTreeWidgetItem osloItem = new QTreeWidgetItem(cities);
    osloItem->setText(0, tr("Oslo"));
    osloItem->setText(1, tr("Yes"));

Items can be added in a particular order by specifying the item they follow when they are constructed:

    QTreeWidgetItem *planets = new QTreeWidgetItem(treeWidget, cities);
    planets->setText(0, tr("Planets"));

Each column in an item can have its own background brush which is set with the setBackground() function. The current background brush can be found with background(). The text label for each column can be rendered with its own font and brush. These are specified with the setFont() and setForeground() functions, and read with font() and foreground().

The main difference between top-level items and those in lower levels of the tree is that a top-level item has no parent(). This information can be used to tell the difference between items, and is useful to know when inserting and removing items from the tree. Children of an item can be removed with takeChild() and inserted at a given index in the list of children with the insertChild() function.

By default, items are enabled, selectable, checkable, and can be the source of a drag and drop operation. Each item's flags can be changed by calling setFlags() with the appropriate value (see Qt::ItemFlags). Checkable items can be checked and unchecked with the setCheckState() function. The corresponding checkState() function indicates whether the item is currently checked.

Methods

impl TreeWidgetItem

fn add_child(&mut self, child: *mut TreeWidgetItem)

C++ method: void QTreeWidgetItem::addChild(QTreeWidgetItem* child)

C++ documentation:

Appends the child item to the list of children.

See also insertChild() and takeChild().

fn add_children(&mut self, children: &ListTreeWidgetItemMutPtr)

C++ method: void QTreeWidgetItem::addChildren(const QList<QTreeWidgetItem*>& children)

C++ documentation:

Appends the given list of children to the item.

This function was introduced in Qt 4.1.

See also insertChildren() and takeChildren().

fn background<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TreeWidgetItemBackgroundArgs<'largs>

C++ method: QTreeWidgetItem::background

This is an overloaded function. Available variants:

Rust arguments:
1) fn background(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_gui::brush::Brush>
2) fn background(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_gui::brush::Brush
C++ method: QBrush QTreeWidgetItem::background(int column) const

C++ documentation:

Returns the brush used to render the background of the specified column.

This function was introduced in Qt 4.2.

See also setBackground() and foreground().

fn background_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TreeWidgetItemBackgroundColorArgs<'largs>

C++ method: QTreeWidgetItem::backgroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn background_color(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn background_color(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_gui::color::Color
C++ method: QColor QTreeWidgetItem::backgroundColor(int column) const

C++ documentation:

This function is deprecated. Use background() instead.

See also setBackgroundColor().

fn check_state(&self, column: c_int) -> CheckState

C++ method: Qt::CheckState QTreeWidgetItem::checkState(int column) const

C++ documentation:

Returns the check state of the label in the given column.

See also setCheckState() and Qt::CheckState.

fn child(&self, index: c_int) -> *mut TreeWidgetItem

C++ method: QTreeWidgetItem* QTreeWidgetItem::child(int index) const

C++ documentation:

Returns the item at the given index in the list of the item's children.

See also parent().

fn child_count(&self) -> c_int

C++ method: int QTreeWidgetItem::childCount() const

C++ documentation:

Returns the number of child items.

fn child_indicator_policy(&self) -> ChildIndicatorPolicy

C++ method: QTreeWidgetItem::ChildIndicatorPolicy QTreeWidgetItem::childIndicatorPolicy() const

C++ documentation:

Returns the item indicator policy. This policy decides when the tree branch expand/collapse indicator is shown.

See also setChildIndicatorPolicy().

fn clone(&self) -> *mut TreeWidgetItem

C++ method: virtual QTreeWidgetItem* QTreeWidgetItem::clone() const

C++ documentation:

Creates a deep copy of the item and of its children.

fn column_count(&self) -> c_int

C++ method: int QTreeWidgetItem::columnCount() const

C++ documentation:

Returns the number of columns in the item.

fn data<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TreeWidgetItemDataArgs<'largs>

C++ method: QTreeWidgetItem::data

This is an overloaded function. Available variants:

Rust arguments:
1) fn data(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn data(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> qt_core::variant::Variant
C++ method: virtual QVariant QTreeWidgetItem::data(int column, int role) const

C++ documentation:

Returns the value for the item's column and role.

See also setData().

fn flags(&self) -> Flags<ItemFlag>

C++ method: QFlags<Qt::ItemFlag> QTreeWidgetItem::flags() const

C++ documentation:

Returns the flags used to describe the item. These determine whether the item can be checked, edited, and selected.

The default value for flags is Qt::ItemIsSelectable | Qt::ItemIsUserCheckable | Qt::ItemIsEnabled | Qt::ItemIsDragEnabled | Qt::ItemIsDropEnabled.

See also setFlags().

fn font<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TreeWidgetItemFontArgs<'largs>

C++ method: QTreeWidgetItem::font

This is an overloaded function. Available variants:

Rust arguments:
1) fn font(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_gui::font::Font>
2) fn font(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_gui::font::Font
C++ method: QFont QTreeWidgetItem::font(int column) const

C++ documentation:

Returns the font used to render the text in the specified column.

See also setFont().

fn foreground<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TreeWidgetItemForegroundArgs<'largs>

C++ method: QTreeWidgetItem::foreground

This is an overloaded function. Available variants:

Rust arguments:
1) fn foreground(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_gui::brush::Brush>
2) fn foreground(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_gui::brush::Brush
C++ method: QBrush QTreeWidgetItem::foreground(int column) const

C++ documentation:

Returns the brush used to render the foreground (e.g. text) of the specified column.

This function was introduced in Qt 4.2.

See also setForeground() and background().

fn icon<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TreeWidgetItemIconArgs<'largs>

C++ method: QTreeWidgetItem::icon

This is an overloaded function. Available variants:

Rust arguments:
1) fn icon(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_gui::icon::Icon>
2) fn icon(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_gui::icon::Icon
C++ method: QIcon QTreeWidgetItem::icon(int column) const

C++ documentation:

Returns the icon that is displayed in the specified column.

See also setIcon() and iconSize.

fn index_of_child(&self, child: *mut TreeWidgetItem) -> c_int

C++ method: int QTreeWidgetItem::indexOfChild(QTreeWidgetItem* child) const

C++ documentation:

Returns the index of the given child in the item's list of children.

fn insert_child(&mut self, index: c_int, child: *mut TreeWidgetItem)

C++ method: void QTreeWidgetItem::insertChild(int index, QTreeWidgetItem* child)

C++ documentation:

Inserts the child item at index in the list of children.

If the child has already been inserted somewhere else it won't be inserted again.

fn insert_children(&mut self, index: c_int, children: &ListTreeWidgetItemMutPtr)

C++ method: void QTreeWidgetItem::insertChildren(int index, const QList<QTreeWidgetItem*>& children)

C++ documentation:

Inserts the given list of children into the list of the item children at index .

Children that have already been inserted somewhere else won't be inserted.

This function was introduced in Qt 4.1.

fn is_disabled(&self) -> bool

C++ method: bool QTreeWidgetItem::isDisabled() const

C++ documentation:

Returns true if the item is disabled; otherwise returns false.

This function was introduced in Qt 4.3.

See also setFlags().

fn is_expanded(&self) -> bool

C++ method: bool QTreeWidgetItem::isExpanded() const

C++ documentation:

Returns true if the item is expanded, otherwise returns false.

This function was introduced in Qt 4.2.

See also setExpanded().

fn is_first_column_spanned(&self) -> bool

C++ method: bool QTreeWidgetItem::isFirstColumnSpanned() const

C++ documentation:

Returns true if the item is spanning all the columns in a row; otherwise returns false.

This function was introduced in Qt 4.3.

See also setFirstColumnSpanned().

fn is_hidden(&self) -> bool

C++ method: bool QTreeWidgetItem::isHidden() const

C++ documentation:

Returns true if the item is hidden, otherwise returns false.

This function was introduced in Qt 4.2.

See also setHidden().

fn is_selected(&self) -> bool

C++ method: bool QTreeWidgetItem::isSelected() const

C++ documentation:

Returns true if the item is selected, otherwise returns false.

This function was introduced in Qt 4.2.

See also setSelected().

fn new<Args>(args: Args) -> Args::ReturnType where Args: TreeWidgetItemNewArgs

C++ method: QTreeWidgetItem::QTreeWidgetItem

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((*mut ::tree_widget::TreeWidget, *mut ::tree_widget_item::TreeWidgetItem, cpp_utils::AsStruct)) -> ::tree_widget_item::TreeWidgetItem
2) fn new((*mut ::tree_widget::TreeWidget, *mut ::tree_widget_item::TreeWidgetItem, libc::c_int, cpp_utils::AsStruct)) -> ::tree_widget_item::TreeWidgetItem
3) fn new((*mut ::tree_widget::TreeWidget, *mut ::tree_widget_item::TreeWidgetItem, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item::TreeWidgetItem>
4) fn new((*mut ::tree_widget::TreeWidget, *mut ::tree_widget_item::TreeWidgetItem, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item::TreeWidgetItem>
C++ method: [constructor] void QTreeWidgetItem::QTreeWidgetItem(QTreeWidget* view, QTreeWidgetItem* after, int type = ?)

Variant 2

Rust arguments:
1) fn new((*mut ::tree_widget::TreeWidget, &qt_core::string_list::StringList, cpp_utils::AsStruct)) -> ::tree_widget_item::TreeWidgetItem
2) fn new((*mut ::tree_widget::TreeWidget, &qt_core::string_list::StringList, libc::c_int, cpp_utils::AsStruct)) -> ::tree_widget_item::TreeWidgetItem
3) fn new((*mut ::tree_widget::TreeWidget, &qt_core::string_list::StringList, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item::TreeWidgetItem>
4) fn new((*mut ::tree_widget::TreeWidget, &qt_core::string_list::StringList, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item::TreeWidgetItem>
C++ method: [constructor] void QTreeWidgetItem::QTreeWidgetItem(QTreeWidget* view, const QStringList& strings, int type = ?)

Variant 3

Rust arguments:
1) fn new((*mut ::tree_widget::TreeWidget, cpp_utils::AsStruct)) -> ::tree_widget_item::TreeWidgetItem
2) fn new((*mut ::tree_widget::TreeWidget, libc::c_int, cpp_utils::AsStruct)) -> ::tree_widget_item::TreeWidgetItem
3) fn new((*mut ::tree_widget::TreeWidget, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item::TreeWidgetItem>
4) fn new((*mut ::tree_widget::TreeWidget, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item::TreeWidgetItem>
C++ method: [constructor] void QTreeWidgetItem::QTreeWidgetItem(QTreeWidget* view, int type = ?)

Variant 4

Rust arguments:
1) fn new((*mut ::tree_widget_item::TreeWidgetItem, *mut ::tree_widget_item::TreeWidgetItem, cpp_utils::AsStruct)) -> ::tree_widget_item::TreeWidgetItem
2) fn new((*mut ::tree_widget_item::TreeWidgetItem, *mut ::tree_widget_item::TreeWidgetItem, libc::c_int, cpp_utils::AsStruct)) -> ::tree_widget_item::TreeWidgetItem
3) fn new((*mut ::tree_widget_item::TreeWidgetItem, *mut ::tree_widget_item::TreeWidgetItem, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item::TreeWidgetItem>
4) fn new((*mut ::tree_widget_item::TreeWidgetItem, *mut ::tree_widget_item::TreeWidgetItem, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item::TreeWidgetItem>
C++ method: [constructor] void QTreeWidgetItem::QTreeWidgetItem(QTreeWidgetItem* parent, QTreeWidgetItem* after, int type = ?)

Variant 5

Rust arguments:
1) fn new((*mut ::tree_widget_item::TreeWidgetItem, &qt_core::string_list::StringList, cpp_utils::AsStruct)) -> ::tree_widget_item::TreeWidgetItem
2) fn new((*mut ::tree_widget_item::TreeWidgetItem, &qt_core::string_list::StringList, libc::c_int, cpp_utils::AsStruct)) -> ::tree_widget_item::TreeWidgetItem
3) fn new((*mut ::tree_widget_item::TreeWidgetItem, &qt_core::string_list::StringList, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item::TreeWidgetItem>
4) fn new((*mut ::tree_widget_item::TreeWidgetItem, &qt_core::string_list::StringList, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item::TreeWidgetItem>
C++ method: [constructor] void QTreeWidgetItem::QTreeWidgetItem(QTreeWidgetItem* parent, const QStringList& strings, int type = ?)

Variant 6

Rust arguments:
1) fn new((*mut ::tree_widget_item::TreeWidgetItem, cpp_utils::AsStruct)) -> ::tree_widget_item::TreeWidgetItem
2) fn new((*mut ::tree_widget_item::TreeWidgetItem, libc::c_int, cpp_utils::AsStruct)) -> ::tree_widget_item::TreeWidgetItem
3) fn new((*mut ::tree_widget_item::TreeWidgetItem, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item::TreeWidgetItem>
4) fn new((*mut ::tree_widget_item::TreeWidgetItem, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item::TreeWidgetItem>
C++ method: [constructor] void QTreeWidgetItem::QTreeWidgetItem(QTreeWidgetItem* parent, int type = ?)

Variant 7

Rust arguments:
1) fn new((&qt_core::string_list::StringList, cpp_utils::AsStruct)) -> ::tree_widget_item::TreeWidgetItem
2) fn new((&qt_core::string_list::StringList, libc::c_int, cpp_utils::AsStruct)) -> ::tree_widget_item::TreeWidgetItem
3) fn new((&qt_core::string_list::StringList, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item::TreeWidgetItem>
4) fn new((&qt_core::string_list::StringList, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item::TreeWidgetItem>
C++ method: [constructor] void QTreeWidgetItem::QTreeWidgetItem(const QStringList& strings, int type = ?)

Variant 8

Rust arguments:
1) fn new((&::tree_widget_item::TreeWidgetItem, cpp_utils::AsStruct)) -> ::tree_widget_item::TreeWidgetItem
2) fn new((&::tree_widget_item::TreeWidgetItem, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item::TreeWidgetItem>
C++ method: [constructor] void QTreeWidgetItem::QTreeWidgetItem(const QTreeWidgetItem& other)

Variant 9

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::tree_widget_item::TreeWidgetItem
2) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::tree_widget_item::TreeWidgetItem
3) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::tree_widget_item::TreeWidgetItem>
4) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item::TreeWidgetItem>
C++ method: [constructor] void QTreeWidgetItem::QTreeWidgetItem(int type = ?)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 TreeWidgetItem) -> &'l0 mut TreeWidgetItem

C++ method: QTreeWidgetItem& QTreeWidgetItem::operator=(const QTreeWidgetItem& other)

C++ documentation:

Assigns other's data and flags to this item. Note that type() and treeWidget() are not copied.

This function is useful when reimplementing clone().

See also data() and flags().

fn op_lt(&self, other: &TreeWidgetItem) -> bool

C++ method: virtual bool QTreeWidgetItem::operator<(const QTreeWidgetItem& other) const

C++ documentation:

Returns true if the text in the item is less than the text in the other item, otherwise returns false.

fn parent(&self) -> *mut TreeWidgetItem

C++ method: QTreeWidgetItem* QTreeWidgetItem::parent() const

C++ documentation:

Returns the item's parent.

See also child().

fn read(&mut self, in_: &mut DataStream)

C++ method: virtual void QTreeWidgetItem::read(QDataStream& in)

C++ documentation:

Reads the item from stream in. This only reads data into a single item.

See also write().

fn remove_child(&mut self, child: *mut TreeWidgetItem)

C++ method: void QTreeWidgetItem::removeChild(QTreeWidgetItem* child)

C++ documentation:

Removes the given item indicated by child. The removed item will not be deleted.

fn set_background(&mut self, column: c_int, brush: &Brush)

C++ method: void QTreeWidgetItem::setBackground(int column, const QBrush& brush)

C++ documentation:

Sets the background brush of the label in the given column to the specified brush.

This function was introduced in Qt 4.2.

See also background() and setForeground().

fn set_background_color(&mut self, column: c_int, color: &Color)

C++ method: void QTreeWidgetItem::setBackgroundColor(int column, const QColor& color)

C++ documentation:

This function is deprecated. Use setBackground() instead.

See also backgroundColor().

fn set_check_state(&mut self, column: c_int, state: CheckState)

C++ method: void QTreeWidgetItem::setCheckState(int column, Qt::CheckState state)

C++ documentation:

Sets the item in the given column check state to be state.

See also checkState().

fn set_child_indicator_policy(&mut self, policy: ChildIndicatorPolicy)

C++ method: void QTreeWidgetItem::setChildIndicatorPolicy(QTreeWidgetItem::ChildIndicatorPolicy policy)

C++ documentation:

Sets the item indicator policy. This policy decides when the tree branch expand/collapse indicator is shown. The default value is ShowForChildren.

See also childIndicatorPolicy().

fn set_data(&mut self, column: c_int, role: c_int, value: &Variant)

C++ method: virtual void QTreeWidgetItem::setData(int column, int role, const QVariant& value)

C++ documentation:

Sets the value for the item's column and role to the given value.

The role describes the type of data specified by value, and is defined by the Qt::ItemDataRole enum.

See also data().

fn set_disabled(&mut self, disabled: bool)

C++ method: void QTreeWidgetItem::setDisabled(bool disabled)

C++ documentation:

Disables the item if disabled is true; otherwise enables the item.

This function was introduced in Qt 4.3.

See also isDisabled() and setFlags().

fn set_expanded(&mut self, expand: bool)

C++ method: void QTreeWidgetItem::setExpanded(bool expand)

C++ documentation:

Expands the item if expand is true, otherwise collapses the item.

Warning: The QTreeWidgetItem must be added to the QTreeWidget before calling this function.

This function was introduced in Qt 4.2.

See also isExpanded().

fn set_first_column_spanned(&mut self, span: bool)

C++ method: void QTreeWidgetItem::setFirstColumnSpanned(bool span)

C++ documentation:

Sets the first section to span all columns if span is true; otherwise all item sections are shown.

This function was introduced in Qt 4.3.

See also isFirstColumnSpanned().

fn set_flags(&mut self, flags: Flags<ItemFlag>)

C++ method: void QTreeWidgetItem::setFlags(QFlags<Qt::ItemFlag> flags)

C++ documentation:

Sets the flags for the item to the given flags. These determine whether the item can be selected or modified. This is often used to disable an item.

See also flags().

fn set_font(&mut self, column: c_int, font: &Font)

C++ method: void QTreeWidgetItem::setFont(int column, const QFont& font)

C++ documentation:

Sets the font used to display the text in the given column to the given font.

See also font(), setText(), and setForeground().

fn set_foreground(&mut self, column: c_int, brush: &Brush)

C++ method: void QTreeWidgetItem::setForeground(int column, const QBrush& brush)

C++ documentation:

Sets the foreground brush of the label in the given column to the specified brush.

This function was introduced in Qt 4.2.

See also foreground() and setBackground().

fn set_hidden(&mut self, hide: bool)

C++ method: void QTreeWidgetItem::setHidden(bool hide)

C++ documentation:

Hides the item if hide is true, otherwise shows the item.

Note: A call to this function has no effect if the item is not currently in a view. In particular, calling setHidden(true) on an item and only then adding it to a view will result in a visible item.

This function was introduced in Qt 4.2.

See also isHidden().

fn set_icon(&mut self, column: c_int, icon: &Icon)

C++ method: void QTreeWidgetItem::setIcon(int column, const QIcon& icon)

C++ documentation:

Sets the icon to be displayed in the given column to icon.

See also icon(), setText(), and iconSize.

fn set_selected(&mut self, select: bool)

C++ method: void QTreeWidgetItem::setSelected(bool select)

C++ documentation:

Sets the selected state of the item to select.

This function was introduced in Qt 4.2.

See also isSelected().

fn set_size_hint(&mut self, column: c_int, size: &Size)

C++ method: void QTreeWidgetItem::setSizeHint(int column, const QSize& size)

C++ documentation:

Sets the size hint for the tree item in the given column to be size. If no size hint is set, the item delegate will compute the size hint based on the item data.

This function was introduced in Qt 4.1.

See also sizeHint().

fn set_status_tip(&mut self, column: c_int, status_tip: &String)

C++ method: void QTreeWidgetItem::setStatusTip(int column, const QString& statusTip)

C++ documentation:

Sets the status tip for the given column to the given statusTip. QTreeWidget mouse tracking needs to be enabled for this feature to work.

See also statusTip(), setToolTip(), and setWhatsThis().

fn set_text(&mut self, column: c_int, text: &String)

C++ method: void QTreeWidgetItem::setText(int column, const QString& text)

C++ documentation:

Sets the text to be displayed in the given column to the given text.

See also text(), setFont(), and setForeground().

fn set_text_alignment(&mut self, column: c_int, alignment: c_int)

C++ method: void QTreeWidgetItem::setTextAlignment(int column, int alignment)

C++ documentation:

Sets the text alignment for the label in the given column to the alignment specified (see Qt::AlignmentFlag).

See also textAlignment().

fn set_text_color(&mut self, column: c_int, color: &Color)

C++ method: void QTreeWidgetItem::setTextColor(int column, const QColor& color)

C++ documentation:

This function is deprecated. Use setForeground() instead.

See also textColor().

fn set_tool_tip(&mut self, column: c_int, tool_tip: &String)

C++ method: void QTreeWidgetItem::setToolTip(int column, const QString& toolTip)

C++ documentation:

Sets the tooltip for the given column to toolTip.

See also toolTip(), setStatusTip(), and setWhatsThis().

fn set_whats_this(&mut self, column: c_int, whats_this: &String)

C++ method: void QTreeWidgetItem::setWhatsThis(int column, const QString& whatsThis)

C++ documentation:

Sets the "What's This?" help for the given column to whatsThis.

See also whatsThis(), setStatusTip(), and setToolTip().

fn size_hint<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TreeWidgetItemSizeHintArgs<'largs>

C++ method: QTreeWidgetItem::sizeHint

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_hint(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size_hint(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::size::Size
C++ method: QSize QTreeWidgetItem::sizeHint(int column) const

C++ documentation:

Returns the size hint set for the tree item in the given column (see QSize).

This function was introduced in Qt 4.1.

See also setSizeHint().

fn sort_children(&mut self, column: c_int, order: SortOrder)

C++ method: void QTreeWidgetItem::sortChildren(int column, Qt::SortOrder order)

C++ documentation:

Sorts the children of the item using the given order, by the values in the given column.

Note: This function does nothing if the item is not associated with a QTreeWidget.

This function was introduced in Qt 4.2.

fn status_tip<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TreeWidgetItemStatusTipArgs<'largs>

C++ method: QTreeWidgetItem::statusTip

This is an overloaded function. Available variants:

Rust arguments:
1) fn status_tip(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn status_tip(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: QString QTreeWidgetItem::statusTip(int column) const

C++ documentation:

Returns the status tip for the contents of the given column.

See also setStatusTip().

fn take_child(&mut self, index: c_int) -> *mut TreeWidgetItem

C++ method: QTreeWidgetItem* QTreeWidgetItem::takeChild(int index)

C++ documentation:

Removes the item at index and returns it, otherwise return 0.

fn take_children<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: TreeWidgetItemTakeChildrenArgs<'largs>

C++ method: QTreeWidgetItem::takeChildren

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_children(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListTreeWidgetItemMutPtr>
2) fn take_children(&mut self, cpp_utils::AsStruct) -> ::list::ListTreeWidgetItemMutPtr
C++ method: QList<QTreeWidgetItem*> QTreeWidgetItem::takeChildren()

C++ documentation:

Removes the list of children and returns it, otherwise returns an empty list.

This function was introduced in Qt 4.1.

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TreeWidgetItemTextArgs<'largs>

C++ method: QTreeWidgetItem::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: QString QTreeWidgetItem::text(int column) const

C++ documentation:

Returns the text in the specified column.

See also setText().

fn text_alignment(&self, column: c_int) -> c_int

C++ method: int QTreeWidgetItem::textAlignment(int column) const

C++ documentation:

Returns the text alignment for the label in the given column (see Qt::AlignmentFlag).

See also setTextAlignment().

fn text_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TreeWidgetItemTextColorArgs<'largs>

C++ method: QTreeWidgetItem::textColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_color(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_gui::color::Color>
2) fn text_color(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_gui::color::Color
C++ method: QColor QTreeWidgetItem::textColor(int column) const

C++ documentation:

This function is deprecated. Use foreground() instead.

See also setTextColor().

fn tool_tip<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TreeWidgetItemToolTipArgs<'largs>

C++ method: QTreeWidgetItem::toolTip

This is an overloaded function. Available variants:

Rust arguments:
1) fn tool_tip(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tool_tip(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: QString QTreeWidgetItem::toolTip(int column) const

C++ documentation:

Returns the tool tip for the given column.

See also setToolTip().

fn tree_widget(&self) -> *mut TreeWidget

C++ method: QTreeWidget* QTreeWidgetItem::treeWidget() const

C++ documentation:

Returns the tree widget that contains the item.

fn type_(&self) -> c_int

C++ method: int QTreeWidgetItem::type() const

C++ documentation:

Returns the type passed to the QTreeWidgetItem constructor.

fn whats_this<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TreeWidgetItemWhatsThisArgs<'largs>

C++ method: QTreeWidgetItem::whatsThis

This is an overloaded function. Available variants:

Rust arguments:
1) fn whats_this(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn whats_this(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: QString QTreeWidgetItem::whatsThis(int column) const

C++ documentation:

Returns the "What's This?" help for the contents of the given column.

See also setWhatsThis().

fn write(&self, out: &mut DataStream)

C++ method: virtual void QTreeWidgetItem::write(QDataStream& out) const

C++ documentation:

Writes the item to stream out. This only writes data from one single item.

See also read().