Struct qt_widgets::tree_widget_item_iterator::TreeWidgetItemIterator

pub struct TreeWidgetItemIterator {
    // some fields omitted
}

C++ type: QTreeWidgetItemIterator

C++ documentation:

The QTreeWidgetItemIterator class provides a way to iterate over the items in a QTreeWidget instance.

The iterator will walk the items in a pre-order traversal order, thus visiting the parent node before it continues to the child nodes.

For example, the following code examples each item in a tree, checking the text in the first column against a user-specified search string:

    QTreeWidgetItemIterator it(treeWidget);
    while (it) {
        if ((it)->text(0) == itemText)
            (*it)->setSelected(true);
        ++it;
    }

It is also possible to filter out certain types of node by passing certain flags to the constructor of QTreeWidgetItemIterator.

Methods

impl TreeWidgetItemIterator

fn new<Args>(args: Args) -> Args::ReturnType where Args: TreeWidgetItemIteratorNewArgs

C++ method: QTreeWidgetItemIterator::QTreeWidgetItemIterator

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((*mut ::tree_widget::TreeWidget, cpp_utils::AsStruct)) -> ::tree_widget_item_iterator::TreeWidgetItemIterator
2) fn new((*mut ::tree_widget::TreeWidget, qt_core::flags::Flags<::tree_widget_item_iterator::IteratorFlag>, cpp_utils::AsStruct)) -> ::tree_widget_item_iterator::TreeWidgetItemIterator
3) fn new((*mut ::tree_widget::TreeWidget, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item_iterator::TreeWidgetItemIterator>
4) fn new((*mut ::tree_widget::TreeWidget, qt_core::flags::Flags<::tree_widget_item_iterator::IteratorFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item_iterator::TreeWidgetItemIterator>
C++ method: [constructor] void QTreeWidgetItemIterator::QTreeWidgetItemIterator(QTreeWidget* widget, QFlags<QTreeWidgetItemIterator::IteratorFlag> flags = ?)

Variant 2

Rust arguments:
1) fn new((*mut ::tree_widget_item::TreeWidgetItem, cpp_utils::AsStruct)) -> ::tree_widget_item_iterator::TreeWidgetItemIterator
2) fn new((*mut ::tree_widget_item::TreeWidgetItem, qt_core::flags::Flags<::tree_widget_item_iterator::IteratorFlag>, cpp_utils::AsStruct)) -> ::tree_widget_item_iterator::TreeWidgetItemIterator
3) fn new((*mut ::tree_widget_item::TreeWidgetItem, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item_iterator::TreeWidgetItemIterator>
4) fn new((*mut ::tree_widget_item::TreeWidgetItem, qt_core::flags::Flags<::tree_widget_item_iterator::IteratorFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item_iterator::TreeWidgetItemIterator>
C++ method: [constructor] void QTreeWidgetItemIterator::QTreeWidgetItemIterator(QTreeWidgetItem* item, QFlags<QTreeWidgetItemIterator::IteratorFlag> flags = ?)

Variant 3

Rust arguments:
1) fn new((&::tree_widget_item_iterator::TreeWidgetItemIterator, cpp_utils::AsStruct)) -> ::tree_widget_item_iterator::TreeWidgetItemIterator
2) fn new((&::tree_widget_item_iterator::TreeWidgetItemIterator, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item_iterator::TreeWidgetItemIterator>
C++ method: [constructor] void QTreeWidgetItemIterator::QTreeWidgetItemIterator(const QTreeWidgetItemIterator& it)

fn op_add_assign<'l0>(&'l0 mut self, n: c_int) -> &'l0 mut TreeWidgetItemIterator

C++ method: QTreeWidgetItemIterator& QTreeWidgetItemIterator::operator+=(int n)

C++ documentation:

Makes the iterator go forward by n matching items. (If n is negative, the iterator goes backward.)

If the current item is beyond the last item, the current item pointer is set to 0. Returns the resulting iterator.

fn op_assign<'l0, 'l1>(&'l0 mut self, it: &'l1 TreeWidgetItemIterator) -> &'l0 mut TreeWidgetItemIterator

C++ method: QTreeWidgetItemIterator& QTreeWidgetItemIterator::operator=(const QTreeWidgetItemIterator& it)

C++ documentation:

Assignment. Makes a copy of it and returns a reference to its iterator.

fn op_dec<'l0>(&'l0 mut self) -> &'l0 mut TreeWidgetItemIterator

C++ method: QTreeWidgetItemIterator& QTreeWidgetItemIterator::operator--()

C++ documentation:

The prefix -- operator (--it) advances the iterator to the previous matching item and returns a reference to the resulting iterator. Sets the current pointer to 0 if the current item is the first matching item.

fn op_dec_postfix<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: TreeWidgetItemIteratorOpDecPostfixArgs<'largs>

C++ method: QTreeWidgetItemIterator::operator--

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_dec_postfix(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item_iterator::TreeWidgetItemIterator>
2) fn op_dec_postfix(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::tree_widget_item_iterator::TreeWidgetItemIterator
C++ method: const QTreeWidgetItemIterator QTreeWidgetItemIterator::operator--(int arg1)

C++ documentation:

The postfix -- operator (it--) makes the preceding matching item current and returns an iterator to the previously current item.

fn op_inc<'l0>(&'l0 mut self) -> &'l0 mut TreeWidgetItemIterator

C++ method: QTreeWidgetItemIterator& QTreeWidgetItemIterator::operator++()

C++ documentation:

The prefix ++ operator (++it) advances the iterator to the next matching item and returns a reference to the resulting iterator. Sets the current pointer to 0 if the current item is the last matching item.

fn op_inc_postfix<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: TreeWidgetItemIteratorOpIncPostfixArgs<'largs>

C++ method: QTreeWidgetItemIterator::operator++

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_inc_postfix(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::tree_widget_item_iterator::TreeWidgetItemIterator>
2) fn op_inc_postfix(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::tree_widget_item_iterator::TreeWidgetItemIterator
C++ method: const QTreeWidgetItemIterator QTreeWidgetItemIterator::operator++(int arg1)

C++ documentation:

The postfix ++ operator (it++) advances the iterator to the next matching item and returns an iterator to the previously current item.

fn op_indirection(&self) -> *mut TreeWidgetItem

C++ method: QTreeWidgetItem* QTreeWidgetItemIterator::operator*() const

C++ documentation:

Dereference operator. Returns a pointer to the current item.

fn op_sub_assign<'l0>(&'l0 mut self, n: c_int) -> &'l0 mut TreeWidgetItemIterator

C++ method: QTreeWidgetItemIterator& QTreeWidgetItemIterator::operator-=(int n)

C++ documentation:

Makes the iterator go backward by n matching items. (If n is negative, the iterator goes forward.)

If the current item is ahead of the last item, the current item pointer is set to 0. Returns the resulting iterator.