Module qt_widgets::undo_group::overloading

Traits

UndoGroupConnectArgs
This trait represents a set of arguments accepted by UndoGroup::connect method.

UndoGroupConnectStaticArgs
This trait represents a set of arguments accepted by UndoGroup::connect_static method.

UndoGroupCreateRedoActionArgs
This trait represents a set of arguments accepted by UndoGroup::create_redo_action method.

UndoGroupCreateUndoActionArgs
This trait represents a set of arguments accepted by UndoGroup::create_undo_action method.

UndoGroupDestroyedArgs
This trait represents a set of arguments accepted by UndoGroup::destroyed method.

UndoGroupDisconnectArgs
This trait represents a set of arguments accepted by UndoGroup::disconnect method.

UndoGroupDisconnectStaticArgs
This trait represents a set of arguments accepted by UndoGroup::disconnect_static method.

UndoGroupDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by UndoGroup::dynamic_property_names method.

UndoGroupNewArgs
This trait represents a set of arguments accepted by UndoGroup::new method.

UndoGroupObjectNameArgs
This trait represents a set of arguments accepted by UndoGroup::object_name method.

UndoGroupPropertyArgs
This trait represents a set of arguments accepted by UndoGroup::property method.

UndoGroupRedoTextArgs
This trait represents a set of arguments accepted by UndoGroup::redo_text method.

UndoGroupStacksArgs
This trait represents a set of arguments accepted by UndoGroup::stacks method.

UndoGroupStartTimerArgs
This trait represents a set of arguments accepted by UndoGroup::start_timer method.

UndoGroupTrArgs
This trait represents a set of arguments accepted by UndoGroup::tr method.

UndoGroupTrUtf8Args
This trait represents a set of arguments accepted by UndoGroup::tr_utf8 method.

UndoGroupUndoTextArgs
This trait represents a set of arguments accepted by UndoGroup::undo_text method.