Module qt_widgets::undo_stack::overloading

Traits

UndoStackConnectArgs
This trait represents a set of arguments accepted by UndoStack::connect method.

UndoStackConnectStaticArgs
This trait represents a set of arguments accepted by UndoStack::connect_static method.

UndoStackCreateRedoActionArgs
This trait represents a set of arguments accepted by UndoStack::create_redo_action method.

UndoStackCreateUndoActionArgs
This trait represents a set of arguments accepted by UndoStack::create_undo_action method.

UndoStackDestroyedArgs
This trait represents a set of arguments accepted by UndoStack::destroyed method.

UndoStackDisconnectArgs
This trait represents a set of arguments accepted by UndoStack::disconnect method.

UndoStackDisconnectStaticArgs
This trait represents a set of arguments accepted by UndoStack::disconnect_static method.

UndoStackDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by UndoStack::dynamic_property_names method.

UndoStackNewArgs
This trait represents a set of arguments accepted by UndoStack::new method.

UndoStackObjectNameArgs
This trait represents a set of arguments accepted by UndoStack::object_name method.

UndoStackPropertyArgs
This trait represents a set of arguments accepted by UndoStack::property method.

UndoStackRedoTextArgs
This trait represents a set of arguments accepted by UndoStack::redo_text method.

UndoStackSetActiveArgs
This trait represents a set of arguments accepted by UndoStack::set_active method.

UndoStackStartTimerArgs
This trait represents a set of arguments accepted by UndoStack::start_timer method.

UndoStackTextArgs
This trait represents a set of arguments accepted by UndoStack::text method.

UndoStackTrArgs
This trait represents a set of arguments accepted by UndoStack::tr method.

UndoStackTrUtf8Args
This trait represents a set of arguments accepted by UndoStack::tr_utf8 method.

UndoStackUndoTextArgs
This trait represents a set of arguments accepted by UndoStack::undo_text method.