Module qt_widgets::widget::overloading

Traits

OpBitOrArgs
This trait represents a set of arguments accepted by op_bit_or method.

WidgetAccessibleDescriptionArgs
This trait represents a set of arguments accepted by Widget::accessible_description method.

WidgetAccessibleNameArgs
This trait represents a set of arguments accepted by Widget::accessible_name method.

WidgetActionsArgs
This trait represents a set of arguments accepted by Widget::actions method.

WidgetBaseSizeArgs
This trait represents a set of arguments accepted by Widget::base_size method.

WidgetChildAtArgs
This trait represents a set of arguments accepted by Widget::child_at method.

WidgetChildrenRectArgs
This trait represents a set of arguments accepted by Widget::children_rect method.

WidgetChildrenRegionArgs
This trait represents a set of arguments accepted by Widget::children_region method.

WidgetConnectArgs
This trait represents a set of arguments accepted by Widget::connect method.

WidgetConnectStaticArgs
This trait represents a set of arguments accepted by Widget::connect_static method.

WidgetContentsMarginsArgs
This trait represents a set of arguments accepted by Widget::contents_margins method.

WidgetContentsRectArgs
This trait represents a set of arguments accepted by Widget::contents_rect method.

WidgetCreateWindowContainerArgs
This trait represents a set of arguments accepted by Widget::create_window_container method.

WidgetCursorArgs
This trait represents a set of arguments accepted by Widget::cursor method.

WidgetDestroyedArgs
This trait represents a set of arguments accepted by Widget::destroyed method.

WidgetDisconnectArgs
This trait represents a set of arguments accepted by Widget::disconnect method.

WidgetDisconnectStaticArgs
This trait represents a set of arguments accepted by Widget::disconnect_static method.

WidgetDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Widget::dynamic_property_names method.

WidgetFontInfoArgs
This trait represents a set of arguments accepted by Widget::font_info method.

WidgetFontMetricsArgs
This trait represents a set of arguments accepted by Widget::font_metrics method.

WidgetFrameGeometryArgs
This trait represents a set of arguments accepted by Widget::frame_geometry method.

WidgetFrameSizeArgs
This trait represents a set of arguments accepted by Widget::frame_size method.

WidgetGrabArgs
This trait represents a set of arguments accepted by Widget::grab method.

WidgetGrabGestureArgs
This trait represents a set of arguments accepted by Widget::grab_gesture method.

WidgetGrabMouseArgs
This trait represents a set of arguments accepted by Widget::grab_mouse method.

WidgetGrabShortcutArgs
This trait represents a set of arguments accepted by Widget::grab_shortcut method.

WidgetInputMethodQueryArgs
This trait represents a set of arguments accepted by Widget::input_method_query method.

WidgetLocaleArgs
This trait represents a set of arguments accepted by Widget::locale method.

WidgetMapFromArgs
This trait represents a set of arguments accepted by Widget::map_from method.

WidgetMapFromGlobalArgs
This trait represents a set of arguments accepted by Widget::map_from_global method.

WidgetMapFromParentArgs
This trait represents a set of arguments accepted by Widget::map_from_parent method.

WidgetMapToArgs
This trait represents a set of arguments accepted by Widget::map_to method.

WidgetMapToGlobalArgs
This trait represents a set of arguments accepted by Widget::map_to_global method.

WidgetMapToParentArgs
This trait represents a set of arguments accepted by Widget::map_to_parent method.

WidgetMaskArgs
This trait represents a set of arguments accepted by Widget::mask method.

WidgetMaximumSizeArgs
This trait represents a set of arguments accepted by Widget::maximum_size method.

WidgetMinimumSizeArgs
This trait represents a set of arguments accepted by Widget::minimum_size method.

WidgetMinimumSizeHintArgs
This trait represents a set of arguments accepted by Widget::minimum_size_hint method.

WidgetMoveArgs
This trait represents a set of arguments accepted by Widget::move_ method.

WidgetNewArgs
This trait represents a set of arguments accepted by Widget::new method.

WidgetNormalGeometryArgs
This trait represents a set of arguments accepted by Widget::normal_geometry method.

WidgetObjectNameArgs
This trait represents a set of arguments accepted by Widget::object_name method.

WidgetPosArgs
This trait represents a set of arguments accepted by Widget::pos method.

WidgetPropertyArgs
This trait represents a set of arguments accepted by Widget::property method.

WidgetRectArgs
This trait represents a set of arguments accepted by Widget::rect method.

WidgetRenderArgs
This trait represents a set of arguments accepted by Widget::render method.

WidgetRepaintArgs
This trait represents a set of arguments accepted by Widget::repaint method.

WidgetResizeArgs
This trait represents a set of arguments accepted by Widget::resize method.

WidgetSaveGeometryArgs
This trait represents a set of arguments accepted by Widget::save_geometry method.

WidgetScrollArgs
This trait represents a set of arguments accepted by Widget::scroll method.

WidgetSetAttributeArgs
This trait represents a set of arguments accepted by Widget::set_attribute method.

WidgetSetBaseSizeArgs
This trait represents a set of arguments accepted by Widget::set_base_size method.

WidgetSetContentsMarginsArgs
This trait represents a set of arguments accepted by Widget::set_contents_margins method.

WidgetSetFixedSizeArgs
This trait represents a set of arguments accepted by Widget::set_fixed_size method.

WidgetSetFocusArgs
This trait represents a set of arguments accepted by Widget::set_focus method.

WidgetSetGeometryArgs
This trait represents a set of arguments accepted by Widget::set_geometry method.

WidgetSetMaskArgs
This trait represents a set of arguments accepted by Widget::set_mask method.

WidgetSetMaximumSizeArgs
This trait represents a set of arguments accepted by Widget::set_maximum_size method.

WidgetSetMinimumSizeArgs
This trait represents a set of arguments accepted by Widget::set_minimum_size method.

WidgetSetParentArgs
This trait represents a set of arguments accepted by Widget::set_parent method.

WidgetSetShortcutAutoRepeatArgs
This trait represents a set of arguments accepted by Widget::set_shortcut_auto_repeat method.

WidgetSetShortcutEnabledArgs
This trait represents a set of arguments accepted by Widget::set_shortcut_enabled method.

WidgetSetSizeIncrementArgs
This trait represents a set of arguments accepted by Widget::set_size_increment method.

WidgetSetSizePolicyArgs
This trait represents a set of arguments accepted by Widget::set_size_policy method.

WidgetSizeArgs
This trait represents a set of arguments accepted by Widget::size method.

WidgetSizeHintArgs
This trait represents a set of arguments accepted by Widget::size_hint method.

WidgetSizeIncrementArgs
This trait represents a set of arguments accepted by Widget::size_increment method.

WidgetSizePolicyArgs
This trait represents a set of arguments accepted by Widget::size_policy method.

WidgetStartTimerArgs
This trait represents a set of arguments accepted by Widget::start_timer method.

WidgetStatusTipArgs
This trait represents a set of arguments accepted by Widget::status_tip method.

WidgetStyleSheetArgs
This trait represents a set of arguments accepted by Widget::style_sheet method.

WidgetToolTipArgs
This trait represents a set of arguments accepted by Widget::tool_tip method.

WidgetTrArgs
This trait represents a set of arguments accepted by Widget::tr method.

WidgetTrUtf8Args
This trait represents a set of arguments accepted by Widget::tr_utf8 method.

WidgetUpdateArgs
This trait represents a set of arguments accepted by Widget::update method.

WidgetVisibleRegionArgs
This trait represents a set of arguments accepted by Widget::visible_region method.

WidgetWhatsThisArgs
This trait represents a set of arguments accepted by Widget::whats_this method.

WidgetWindowFilePathArgs
This trait represents a set of arguments accepted by Widget::window_file_path method.

WidgetWindowIconArgs
This trait represents a set of arguments accepted by Widget::window_icon method.

WidgetWindowIconTextArgs
This trait represents a set of arguments accepted by Widget::window_icon_text method.

WidgetWindowRoleArgs
This trait represents a set of arguments accepted by Widget::window_role method.

WidgetWindowTitleArgs
This trait represents a set of arguments accepted by Widget::window_title method.