Module qt_widgets::widget_action::overloading

Traits

WidgetActionAssociatedGraphicsWidgetsArgs
This trait represents a set of arguments accepted by WidgetAction::associated_graphics_widgets method.

WidgetActionAssociatedWidgetsArgs
This trait represents a set of arguments accepted by WidgetAction::associated_widgets method.

WidgetActionConnectArgs
This trait represents a set of arguments accepted by WidgetAction::connect method.

WidgetActionConnectStaticArgs
This trait represents a set of arguments accepted by WidgetAction::connect_static method.

WidgetActionDataArgs
This trait represents a set of arguments accepted by WidgetAction::data method.

WidgetActionDestroyedArgs
This trait represents a set of arguments accepted by WidgetAction::destroyed method.

WidgetActionDisconnectArgs
This trait represents a set of arguments accepted by WidgetAction::disconnect method.

WidgetActionDisconnectStaticArgs
This trait represents a set of arguments accepted by WidgetAction::disconnect_static method.

WidgetActionDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by WidgetAction::dynamic_property_names method.

WidgetActionFontArgs
This trait represents a set of arguments accepted by WidgetAction::font method.

WidgetActionIconArgs
This trait represents a set of arguments accepted by WidgetAction::icon method.

WidgetActionIconTextArgs
This trait represents a set of arguments accepted by WidgetAction::icon_text method.

WidgetActionNewArgs
This trait represents a set of arguments accepted by WidgetAction::new method.

WidgetActionObjectNameArgs
This trait represents a set of arguments accepted by WidgetAction::object_name method.

WidgetActionPropertyArgs
This trait represents a set of arguments accepted by WidgetAction::property method.

WidgetActionSetShortcutsArgs
This trait represents a set of arguments accepted by WidgetAction::set_shortcuts method.

WidgetActionShortcutArgs
This trait represents a set of arguments accepted by WidgetAction::shortcut method.

WidgetActionShortcutsArgs
This trait represents a set of arguments accepted by WidgetAction::shortcuts method.

WidgetActionShowStatusTextArgs
This trait represents a set of arguments accepted by WidgetAction::show_status_text method.

WidgetActionStartTimerArgs
This trait represents a set of arguments accepted by WidgetAction::start_timer method.

WidgetActionStatusTipArgs
This trait represents a set of arguments accepted by WidgetAction::status_tip method.

WidgetActionTextArgs
This trait represents a set of arguments accepted by WidgetAction::text method.

WidgetActionToolTipArgs
This trait represents a set of arguments accepted by WidgetAction::tool_tip method.

WidgetActionTrArgs
This trait represents a set of arguments accepted by WidgetAction::tr method.

WidgetActionTrUtf8Args
This trait represents a set of arguments accepted by WidgetAction::tr_utf8 method.

WidgetActionTriggeredArgs
This trait represents a set of arguments accepted by WidgetAction::triggered method.

WidgetActionWhatsThisArgs
This trait represents a set of arguments accepted by WidgetAction::whats_this method.