Struct qt_widgets::widget_item::WidgetItem

pub struct WidgetItem {
    // some fields omitted
}

C++ type: QWidgetItem

C++ documentation:

The QWidgetItem class is a layout item that represents a widget.

Normally, you don't need to use this class directly. Qt's built-in layout managers provide the following functions for manipulating widgets in layouts:

Class	Functions

QBoxLayout	addWidget(), insertWidget(), setStretchFactor()

QGridLayout	addWidget()

QStackedLayout	addWidget(), insertWidget(), currentWidget(), setCurrentWidget(), widget()

Methods

impl WidgetItem

fn alignment(&self) -> Flags<AlignmentFlag>

C++ method: QFlags<Qt::AlignmentFlag> QWidgetItem::alignment() const

C++ documentation:

Returns the alignment of this item.

See also setAlignment().

fn control_types(&self) -> Flags<ControlType>

C++ method: virtual QFlags<QSizePolicy::ControlType> QWidgetItem::controlTypes() const

C++ documentation:

Reimplemented from QLayoutItem::controlTypes().

Returns the control type associated with the widget for which this size policy applies.

See also QSizePolicy::controlType().

fn expanding_directions(&self) -> Flags<Orientation>

C++ method: virtual QFlags<Qt::Orientation> QWidgetItem::expandingDirections() const

C++ documentation:

Reimplemented from QLayoutItem::expandingDirections().

fn geometry<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: WidgetItemGeometryArgs<'largs>

C++ method: QWidgetItem::geometry

This is an overloaded function. Available variants:

Rust arguments:
1) fn geometry(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn geometry(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: virtual QRect QWidgetItem::geometry() const

C++ documentation:

Reimplemented from QLayoutItem::geometry().

See also setGeometry().

fn has_height_for_width(&self) -> bool

C++ method: virtual bool QWidgetItem::hasHeightForWidth() const

C++ documentation:

Reimplemented from QLayoutItem::hasHeightForWidth().

fn height_for_width(&self, arg1: c_int) -> c_int

C++ method: virtual int QWidgetItem::heightForWidth(int arg1) const

C++ documentation:

Reimplemented from QLayoutItem::heightForWidth().

fn invalidate(&mut self)

C++ method: virtual void QWidgetItem::invalidate()

C++ documentation:

Invalidates any cached information in this layout item.

fn is_empty(&self) -> bool

C++ method: virtual bool QWidgetItem::isEmpty() const

C++ documentation:

Reimplemented from QLayoutItem::isEmpty().

Returns true if the widget is hidden; otherwise returns false.

See also QWidget::isHidden().

fn layout(&mut self) -> *mut Layout

C++ method: virtual QLayout* QWidgetItem::layout()

C++ documentation:

If this item is a QLayout, it is returned as a QLayout; otherwise 0 is returned. This function provides type-safe casting.

fn maximum_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: WidgetItemMaximumSizeArgs<'largs>

C++ method: QWidgetItem::maximumSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn maximum_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn maximum_size(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: virtual QSize QWidgetItem::maximumSize() const

C++ documentation:

Reimplemented from QLayoutItem::maximumSize().

fn minimum_height_for_width(&self, arg1: c_int) -> c_int

C++ method: virtual int QWidgetItem::minimumHeightForWidth(int arg1) const

C++ documentation:

Returns the minimum height this widget needs for the given width, w. The default implementation simply returns heightForWidth(w).

fn minimum_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: WidgetItemMinimumSizeArgs<'largs>

C++ method: QWidgetItem::minimumSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn minimum_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn minimum_size(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: virtual QSize QWidgetItem::minimumSize() const

C++ documentation:

Reimplemented from QLayoutItem::minimumSize().

fn new<Args>(args: Args) -> Args::ReturnType where Args: WidgetItemNewArgs

C++ method: QWidgetItem::QWidgetItem

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((*mut ::widget::Widget, cpp_utils::AsStruct)) -> ::widget_item::WidgetItem
2) fn new((*mut ::widget::Widget, cpp_utils::AsBox)) -> cpp_utils::CppBox<::widget_item::WidgetItem>
C++ method: [constructor] void QWidgetItem::QWidgetItem(QWidget* w)

fn set_alignment(&mut self, a: Flags<AlignmentFlag>)

C++ method: void QWidgetItem::setAlignment(QFlags<Qt::AlignmentFlag> a)

C++ documentation:

Sets the alignment of this item to alignment.

Note: Item alignment is only supported by QLayoutItem subclasses where it would have a visual effect. Except for QSpacerItem, which provides blank space for layouts, all public Qt classes that inherit QLayoutItem support item alignment.

See also alignment().

fn set_geometry(&mut self, arg1: &Rect)

C++ method: virtual void QWidgetItem::setGeometry(const QRect& arg1)

C++ documentation:

Reimplemented from QLayoutItem::setGeometry().

See also geometry().

fn size_hint<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: WidgetItemSizeHintArgs<'largs>

C++ method: QWidgetItem::sizeHint

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_hint(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size_hint(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: virtual QSize QWidgetItem::sizeHint() const

C++ documentation:

Reimplemented from QLayoutItem::sizeHint().

fn spacer_item(&mut self) -> *mut SpacerItem

C++ method: virtual QSpacerItem* QWidgetItem::spacerItem()

C++ documentation:

If this item is a QSpacerItem, it is returned as a QSpacerItem; otherwise 0 is returned. This function provides type-safe casting.

fn widget(&mut self) -> *mut Widget

C++ method: virtual QWidget* QWidgetItem::widget()

C++ documentation:

Reimplemented from QLayoutItem::widget().

Returns the widget managed by this item.