Module qt_widgets::widget_item_v2::overloading

Traits

WidgetItemV2GeometryArgs
This trait represents a set of arguments accepted by WidgetItemV2::geometry method.

WidgetItemV2MaximumSizeArgs
This trait represents a set of arguments accepted by WidgetItemV2::maximum_size method.

WidgetItemV2MinimumSizeArgs
This trait represents a set of arguments accepted by WidgetItemV2::minimum_size method.

WidgetItemV2NewArgs
This trait represents a set of arguments accepted by WidgetItemV2::new method.

WidgetItemV2SizeHintArgs
This trait represents a set of arguments accepted by WidgetItemV2::size_hint method.