Enum qt_widgets::wizard::WizardOption

pub enum WizardOption {
    IndependentPages,
    IgnoreSubTitles,
    ExtendedWatermarkPixmap,
    NoDefaultButton,
    NoBackButtonOnStartPage,
    NoBackButtonOnLastPage,
    DisabledBackButtonOnLastPage,
    HaveNextButtonOnLastPage,
    HaveFinishButtonOnEarlyPages,
    NoCancelButton,
    CancelButtonOnLeft,
    HaveHelpButton,
    HelpButtonOnRight,
    HaveCustomButton1,
    HaveCustomButton2,
    HaveCustomButton3,
    NoCancelButtonOnLastPage,
}

C++ type: QWizard::WizardOption

C++ documentation:

This enum specifies various options that affect the look and feel of a wizard.

The WizardOptions type is a typedef for QFlags<WizardOption>. It stores an OR combination of WizardOption values.

See also setOptions(), setOption(), and testOption().

Variants
IndependentPages

The pages are independent of each other (i.e., they don't derive values from each other). (C++ enum variant: IndependentPages = 1)

IgnoreSubTitles

Don't show any subtitles, even if they are set. (C++ enum variant: IgnoreSubTitles = 2)

ExtendedWatermarkPixmap

Extend any WatermarkPixmap all the way down to the window's edge. (C++ enum variant: ExtendedWatermarkPixmap = 4)

NoDefaultButton

Don't make the Next or Finish button the dialog's default button. (C++ enum variant: NoDefaultButton = 8)

NoBackButtonOnStartPage

Don't show the Back button on the start page. (C++ enum variant: NoBackButtonOnStartPage = 16)

NoBackButtonOnLastPage

Don't show the Back button on the last page. (C++ enum variant: NoBackButtonOnLastPage = 32)

DisabledBackButtonOnLastPage

Disable the Back button on the last page. (C++ enum variant: DisabledBackButtonOnLastPage = 64)

HaveNextButtonOnLastPage

Show the (disabled) Next button on the last page. (C++ enum variant: HaveNextButtonOnLastPage = 128)

HaveFinishButtonOnEarlyPages

Show the (disabled) Finish button on non-final pages. (C++ enum variant: HaveFinishButtonOnEarlyPages = 256)

NoCancelButton

Don't show the Cancel button. (C++ enum variant: NoCancelButton = 512)

CancelButtonOnLeft

Put the Cancel button on the left of Back (rather than on the right of Finish or Next). (C++ enum variant: CancelButtonOnLeft = 1024)

HaveHelpButton

Show the Help button. (C++ enum variant: HaveHelpButton = 2048)

HelpButtonOnRight

Put the Help button on the far right of the button layout (rather than on the far left). (C++ enum variant: HelpButtonOnRight = 4096)

HaveCustomButton1

Show the first user-defined button (CustomButton1). (C++ enum variant: HaveCustomButton1 = 8192)

HaveCustomButton2

Show the second user-defined button (CustomButton2). (C++ enum variant: HaveCustomButton2 = 16384)

HaveCustomButton3

Show the third user-defined button (CustomButton3). (C++ enum variant: HaveCustomButton3 = 32768)

NoCancelButtonOnLastPage

Don't show the Cancel button on the last page. (C++ enum variant: NoCancelButtonOnLastPage = 65536)